ATK {RGtk2}	R Documentation

ATK

Description

ATK is the Accessibility Toolkit. It provides a set of generic interfaces allowing accessibility technologies to interact with a graphical user interface. For example, a screen reader uses ATK to discover the text in an interface and read it to blind users. GTK+ widgets have built-in support for accessibility using the ATK framework.

Details

The RGtk binding to the ATK library consists of the following components:

AtkAction

The ATK interface provided by UI components which the user can activate/interact with,

AtkComponent

The ATK interface provided by UI components which occupy a physical area on the screen.

AtkDocument

The ATK interface which represents the toplevel container for document content.

AtkEditableText

The ATK interface implemented by components containing user-editable text content.

AtkGObjectAccessible

This object class is derived from AtkObject and can be used as a basis implementing accessible objects.

AtkHyperlink

An ATK object which encapsulates a link or set of links in a hypertext document.

AtkHypertext

The ATK interface which provides standard mechanism for manipulating hyperlinks.

AtkImage

The ATK Interface implemented by components which expose image or pixmap content on-screen.

atk-AtkMisc

undocumented

AtkNoOpObject

An AtkObject which purports to implement all ATK interfaces.

AtkNoOpObjectFactory

The AtkObjectFactory which creates an AtkNoOpObject.

AtkObject

The base object class for the Accessibility Toolkit API.

AtkObjectFactory

The base object class for a factory used to create accessible objects for objects of a specific GType.

AtkRegistry

An object used to store the GType of the factories used to create an accessible object for an object of a particular GType.

AtkRelation

An object used to describe a relation between a object and one or more other objects.

AtkRelationSet

A set of AtkRelations, normally the set of AtkRelations which an AtkObject has.

AtkSelection

The ATK interface implemented by container objects whose children can be selected.

atk-AtkState

An AtkState describes a component's particular state.

AtkStateSet

An AtkStateSet determines a component's state set.

AtkStreamableContent

The ATK interface which provides access to streamable content.

AtkTable

The ATK interface implemented for UI components which contain tabular or row/column information.

AtkText

The ATK interface implemented by components with text content.

AtkUtil

A set of ATK utility functions for event and toolkit support.

AtkValue

The ATK interface implemented by valuators and components which display or select a value from a bounded range of values.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk

[Package RGtk2 version 2.20.31 Index]