AtkComponent {RGtk2}	R Documentation

AtkComponent

Description

The ATK interface provided by UI components which occupy a physical area on the screen.

Methods and Functions

atkComponentAddFocusHandler(object, handler)
atkComponentContains(object, x, y, coord.type)
atkComponentGetExtents(object, coord.type)
atkComponentGetLayer(object)
atkComponentGetMdiZorder(object)
atkComponentGetPosition(object, coord.type)
atkComponentGetSize(object)
atkComponentGrabFocus(object)
atkComponentRefAccessibleAtPoint(object, x, y, coord.type)
atkComponentRemoveFocusHandler(object, handler.id)
atkComponentSetExtents(object, x, y, width, height, coord.type)
atkComponentSetPosition(object, x, y, coord.type)
atkComponentSetSize(object, width, height)
atkComponentGetAlpha(object)

Hierarchy

GInterface
   +----AtkComponent

Implementations

AtkComponent is implemented by AtkNoOpObject.

Detailed Description

AtkComponent should be implemented by most if not all UI elements with an actual on-screen presence, i.e. components which can be said to have a screen-coordinate bounding box. Virtually all widgets will need to have AtkComponent implementations provided for their corresponding AtkObject class. In short, only UI elements which are *not* GUI elements will omit this ATK interface.

A possible exception might be textual information with a transparent background, in which case text glyph bounding box information is provided by AtkText.

Structures

AtkComponent

The AtkComponent structure does not contain any fields.

Signals

bounds-changed(atkcomponent, arg1, user.data)

The 'bounds-changed" signal is emitted when the bposition or size of the a component changes.

atkcomponent

[AtkComponent] the object which received the signal.

arg1

[AtkRectangle] The AtkRectangle giving the new position and size.

user.data

[R object] user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkComponent.html

[Package RGtk2 version 2.20.31 Index]