AtkDocument {RGtk2}	R Documentation

AtkDocument

Description

The ATK interface which represents the toplevel container for document content.

Methods and Functions

atkDocumentGetDocumentType(object)
atkDocumentGetDocument(object)
atkDocumentGetAttributeValue(object, attribute.name)
atkDocumentSetAttributeValue(object, attribute.name, attribute.value)
atkDocumentGetAttributes(object)
atkDocumentGetLocale(object)

Hierarchy

GInterface
   +----AtkDocument

Implementations

AtkDocument is implemented by AtkNoOpObject.

Detailed Description

The AtkDocument interface should be supported by any object whose content is a representation or view of a document. The AtkDocument interface should appear on the toplevel container for the document content; however AtkDocument instances may be nested (i.e. an AtkDocument may be a descendant of another AtkDocument) in those cases where one document contains "embedded content" which can reasonably be considered a document in its own right.

Structures

AtkDocument

The AtkDocument structure does not contain any fields.

Signals

load-complete(atkdocument, user.data)

The 'load-complete' signal is emitted when a pending load of a static document has completed. This signal is to be expected by ATK clients if and when AtkDocument implementors expose ATK_STATE_BUSY. If the state of an AtkObject which implements AtkDocument does not include ATK_STATE_BUSY, it should be safe for clients to assume that the AtkDocument's static contents are fully loaded into the container. (Dynamic document contents should be exposed via other signals.)

atkdocument

[AtkDocument] the object which received the signal.

user.data

[R object] user data set when the signal handler was connected.

load-stopped(atkdocument, user.data)

The 'load-stopped' signal is emitted when a pending load of document contents is cancelled, paused, or otherwise interrupted by the user or application logic. It should not however be emitted while waiting for a resource (for instance while blocking on a file or network read) unless a user-significant timeout has occurred.

atkdocument

[AtkDocument] the object which received the signal.

user.data

[R object] user data set when the signal handler was connected.

reload(atkdocument, user.data)

The 'reload' signal is emitted when the contents of a document is refreshed from its source. Once 'reload' has been emitted, a matching 'load-complete' or 'load-stopped' signal should follow, which clients may await before interrogating ATK for the latest document content.

atkdocument

[AtkDocument] the object which received the signal.

user.data

[R object] user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkDocument.html

[Package RGtk2 version 2.20.31 Index]