AtkEditableText {RGtk2}	R Documentation

AtkEditableText

Description

The ATK interface implemented by components containing user-editable text content.

Methods and Functions

atkEditableTextSetRunAttributes(object, attrib.set, start.offset, end.offset)
atkEditableTextSetTextContents(object, string)
atkEditableTextInsertText(object, string, position)
atkEditableTextCopyText(object, start.pos, end.pos)
atkEditableTextCutText(object, start.pos, end.pos)
atkEditableTextDeleteText(object, start.pos, end.pos)
atkEditableTextPasteText(object, position)

Hierarchy

GInterface
   +----AtkEditableText

Implementations

AtkEditableText is implemented by AtkNoOpObject.

Detailed Description

AtkEditableText should be implemented by UI components which contain text which the user can edit, via the AtkObject corresponding to that component (see AtkObject). AtkEditableText is a subclass of AtkText, and as such, an object which implements AtkEditableText is by definition an AtkText implementor as well.

Structures

AtkEditableText

The AtkEditableText structure does not contain any fields.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkEditableText.html

See Also

AtkText

[Package RGtk2 version 2.20.31 Index]