AtkGObjectAccessible {RGtk2}	R Documentation

AtkGObjectAccessible

Description

This object class is derived from AtkObject and can be used as a basis implementing accessible objects.

Methods and Functions

atkGObjectAccessibleForObject(obj)
atkGObjectAccessibleGetObject(object)

Hierarchy

GObject
   +----AtkObject
         +----AtkGObjectAccessible

Detailed Description

This object class is derived from AtkObject. It can be used as a basis for implementing accessible objects for GObjects which are not derived from GtkWidget. One example of its use is in providing an accessible object for GnomeCanvasItem in the GAIL library.

Structures

AtkGObjectAccessible

The AtkGObjectAccessible structure should not be accessed directly.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkGObjectAccessible.html

[Package RGtk2 version 2.20.31 Index]