AtkHyperlink {RGtk2}	R Documentation

AtkHyperlink

Description

An ATK object which encapsulates a link or set of links in a hypertext document.

Methods and Functions

atkHyperlinkGetUri(object, i)
atkHyperlinkGetObject(object, i)
atkHyperlinkGetEndIndex(object)
atkHyperlinkGetStartIndex(object)
atkHyperlinkIsValid(object)
atkHyperlinkIsInline(object)
atkHyperlinkGetNAnchors(object)
atkHyperlinkIsSelectedLink(object)
atkHyperlinkIsSelectedLink(object)

Hierarchy

GObject
   +----AtkHyperlink

Interfaces

AtkHyperlink implements AtkAction.

Detailed Description

An ATK object which encapsulates a link or set of links (for instance in the case of client-side image maps) in a hypertext document. It may implement the AtkAction interface. AtkHyperlink may also be used to refer to inline embedded content, since it allows specification of a start and end offset within the host AtkHypertext object.

Structures

AtkHyperlink

The AtkHyperlink structure should not be accessed directly.

Signals

link-activated(atkhyperlink, user.data)

The signal link-activated is emitted when a link is activated.

atkhyperlink

[AtkHyperlink] the object which received the signal.

user.data

[R object] user data set when the signal handler was connected.

Properties

end-index [integer : Read]

The end index of the AtkHyperlink object. Allowed values: >= 0 Default value: 0

number-of-anchors [integer : Read]

The number of anchors associated with the AtkHyperlink object. Allowed values: >= 0 Default value: 0

selected-link [logical : Read]

Specifies whether the AtkHyperlink object is selected. Default value: FALSE

start-index [integer : Read]

The start index of the AtkHyperlink object. Allowed values: >= 0 Default value: 0

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkHyperlink.html

[Package RGtk2 version 2.20.31 Index]