AtkHypertext {RGtk2}	R Documentation

AtkHypertext

Description

The ATK interface which provides standard mechanism for manipulating hyperlinks.

Methods and Functions

atkHypertextGetLink(object, link.index)
atkHypertextGetNLinks(object)
atkHypertextGetLinkIndex(object, char.index)

Hierarchy

GInterface
   +----AtkHypertext

Implementations

AtkHypertext is implemented by AtkNoOpObject.

Detailed Description

An interface used for objects which implement linking between multiple resource or content locations, or multiple 'markers' within a single document. A Hypertext instance is associated with one or more Hyperlinks, which are associated with particular offsets within the Hypertext's included content. While this interface is derived from Text, there is no requirement that Hypertext instances have textual content; they may implement Image as well, and Hyperlinks need not have non-zero text offsets.

Structures

AtkHypertext

The AtkHypertext structure does not contain any fields.

Signals

link-selected(atkhypertext, arg1, user.data)

The "link-selected" signal is emitted by an AtkHyperText object when one of the hyperlinks associated with the object is selected.

atkhypertext

[AtkHypertext] the object which received the signal.

arg1

[integer] the index of the hyperlink which is selected

user.data

[R object] user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkHypertext.html

[Package RGtk2 version 2.20.31 Index]