AtkImage {RGtk2}	R Documentation

AtkImage

Description

The ATK Interface implemented by components which expose image or pixmap content on-screen.

Methods and Functions

atkImageGetImagePosition(object, coord.type)
atkImageGetImageDescription(object)
atkImageSetImageDescription(object, description)
atkImageGetImageSize(object)
atkImageGetImageLocale(object)

Hierarchy

GInterface
   +----AtkImage

Implementations

AtkImage is implemented by AtkNoOpObject.

Detailed Description

AtkImage should be implemented by AtkObject subtypes on behalf of components which display image/pixmap information onscreen, and which provide information (other than just widget borders, etc.) via that image content. For instance, icons, buttons with icons, toolbar elements, and image viewing panes typically should implement AtkImage. AtkImage primarily provides two types of information: coordinate information (useful for screen review mode of screenreaders, and for use by onscreen magnifiers), and descriptive information. The descriptive information is provided for alternative, text-only presentation of the most significant information present in the image.

Structures

AtkImage

The AtkImage structure does not contain any fields.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkImage.html

[Package RGtk2 version 2.20.31 Index]