AtkNoOpObject {RGtk2}	R Documentation

AtkNoOpObject

Description

An AtkObject which purports to implement all ATK interfaces.

Methods and Functions

atkNoOpObjectNew(obj)
atkNoOpObject(obj)

Hierarchy

GObject
   +----AtkObject
         +----AtkNoOpObject

Interfaces

AtkNoOpObject implements AtkComponent, AtkAction, AtkEditableText, AtkImage, AtkSelection, AtkTable, AtkText, AtkHypertext, AtkValue and AtkDocument.

Detailed Description

An AtkNoOpObject is an AtkObject which purports to implement all ATK interfaces. It is the type of AtkObject which is created if an accessible object is requested for an object type for which no factory type is specified.

Structures

AtkNoOpObject

The AtkNoOpObject structure should not be accessed directly.

Convenient Construction

atkNoOpObject is the equivalent of atkNoOpObjectNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkNoOpObject.html

[Package RGtk2 version 2.20.31 Index]