AtkNoOpObjectFactory {RGtk2}	R Documentation

AtkNoOpObjectFactory

Description

The AtkObjectFactory which creates an AtkNoOpObject.

Methods and Functions

atkNoOpObjectFactoryNew()
atkNoOpObjectFactory()

Hierarchy

GObject
   +----AtkObjectFactory
         +----AtkNoOpObjectFactory

Detailed Description

The AtkObjectFactory which creates an AtkNoOpObject. An instance of this is created by an AtkRegistry if no factory type has not been specified to create an accessible object of a particular type.

Structures

AtkNoOpObjectFactory

The AtkNoOpObjectFactory structure should not be accessed directly.

Convenient Construction

atkNoOpObjectFactory is the equivalent of atkNoOpObjectFactoryNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkNoOpObjectFactory.html

[Package RGtk2 version 2.20.31 Index]