AtkObject {RGtk2}	R Documentation

AtkObject

Description

The base object class for the Accessibility Toolkit API.

Methods and Functions

atkRoleRegister(name)
atkImplementorRefAccessible(object)
atkObjectGetName(object)
atkObjectGetDescription(object)
atkObjectGetParent(object)
atkObjectGetNAccessibleChildren(object)
atkObjectRefAccessibleChild(object, i)
atkObjectRefRelationSet(object)
atkObjectGetLayer(object)
atkObjectGetMdiZorder(object)
atkObjectGetRole(object)
atkObjectRefStateSet(object)
atkObjectGetIndexInParent(object)
atkObjectSetName(object, name)
atkObjectSetDescription(object, description)
atkObjectSetParent(object, parent)
atkObjectSetRole(object, role)
atkObjectConnectPropertyChangeHandler(object, handler)
atkObjectRemovePropertyChangeHandler(object, handler.id)
atkObjectNotifyStateChange(object, state, value)
atkObjectInitialize(object, data)
atkObjectAddRelationship(object, relationship, target)
atkObjectRemoveRelationship(object, relationship, target)
atkObjectGetAttributes(object)
atkRoleGetName(role)
atkRoleGetLocalizedName(role)
atkRoleForName(name)

Hierarchy

GObject
   +----AtkObject
         +----AtkGObjectAccessible
         +----AtkNoOpObject

Detailed Description

This class is the primary class for accessibility support via the Accessibility ToolKit (ATK). Objects which are instances of AtkObject (or instances of AtkObject-derived types) are queried for properties which relate basic (and generic) properties of a UI component such as name and description. Instances of AtkObject may also be queried as to whether they implement other ATK interfaces (e.g. AtkAction, AtkComponent, etc.), as appropriate to the role which a given UI component plays in a user interface. All UI components in an application which provide useful information or services to the user must provide corresponding AtkObject instances on request (in GTK+, for instance, usually on a call to #gtkWidgetGetAccessible), either via ATK support built into the toolkit for the widget class or ancestor class, or in the case of custom widgets, if the inherited AtkObject implementation is insufficient, via instances of a new AtkObject subclass.

Structures

AtkObject

The AtkObject structure should not be accessed directly.

AtkImplementor

The AtkImplementor interface is implemented by objects for which AtkObject peers may be obtained via calls to iface->(ref_accessible)(implementor);

Enums and Flags

AtkRole

Describes the role of an object

These are the built-in enumerated roles that UI components can have in ATK. Other roles may be added at runtime, so an AtkRole >= ATK_ROLE_LAST_DEFINED is not necessarily an error.

invalid

Invalid role

accel-label

A label which represents an accelerator

alert

An object which is an alert to the user. Assistive Technologies typically respond to ATK_ROLE_ALERT by reading the entire onscreen contents of containers advertising this role. Should be used for warning dialogs, etc.

animation

An object which is an animated image

arrow

An arrow in one of the four cardinal directions

calendar

An object that displays a calendar and allows the user to select a date

canvas

An object that can be drawn into and is used to trap events

check-box

A choice that can be checked or unchecked and provides a separate indicator for the current state

check-menu-item

A menu item with a check box

color-chooser

A specialized dialog that lets the user choose a color

column-header

The header for a column of data

combo-box

A list of choices the user can select from

date-editor

An object whose purpose is to allow a user to edit a date

desktop-icon

An inconifed internal frame within a DESKTOP_PANE

desktop-frame

A pane that supports internal frames and iconified versions of those internal frames

dial

An object whose purpose is to allow a user to set a value

dialog

A top level window with title bar and a border

directory-pane

A pane that allows the user to navigate through and select the contents of a directory

drawing-area

An object used for drawing custom user interface elements

file-chooser

A specialized dialog that lets the user choose a file

filler

A object that fills up space in a user interface

font-chooser

A specialized dialog that lets the user choose a font

frame

A top level window with a title bar, border, menubar, etc.

glass-pane

A pane that is guaranteed to be painted on top of all panes beneath it

html-container

A document container for HTML, whose children represent the document content

icon

A small fixed size picture, typically used to decorate components

image

An object whose primary purpose is to display an image

internal-frame

A frame-like object that is clipped by a desktop pane

label

An object used to present an icon or short string in an interface

layered-pane

A specialized pane that allows its children to be drawn in layers, providing a form of stacking order

list

An object that presents a list of objects to the user and allows the user to select one or more of them

list-item

An object that represents an element of a list

menu

An object usually found inside a menu bar that contains a list of actions the user can choose from

menu-bar

An object usually drawn at the top of the primary dialog box of an application that contains a list of menus the user can choose from

menu-item

An object usually contained in a menu that presents an action the user can choose

option-pane

A specialized pane whose primary use is inside a DIALOG

page-tab

An object that is a child of a page tab list

page-tab-list

An object that presents a series of panels (or page tabs), one at a time, through some mechanism provided by the object

panel

A generic container that is often used to group objects

password-text

A text object uses for passwords, or other places where the text content is not shown visibly to the user

popup-menu

A temporary window that is usually used to offer the user a list of choices, and then hides when the user selects one of those choices

progress-bar

An object used to indicate how much of a task has been completed

push-button

An object the user can manipulate to tell the application to do something

radio-button

A specialized check box that will cause other radio buttons in the same group to become unchecked when this one is checked

radio-menu-item

A check menu item which belongs to a group. At each instant exactly one of the radio menu items from a group is selected

root-pane

A specialized pane that has a glass pane and a layered pane as its children

row-header

The header for a row of data

scroll-bar

An object usually used to allow a user to incrementally view a large amount of data.

scroll-pane

An object that allows a user to incrementally view a large amount of information

separator

An object usually contained in a menu to provide a visible and logical separation of the contents in a menu

slider

An object that allows the user to select from a bounded range

split-pane

A specialized panel that presents two other panels at the same time

spin-button

An object used to get an integer or floating point number from the user

statusbar

An object which reports messages of minor importance to the user

table

An object used to represent information in terms of rows and columns

table-cell

A cell in a table

table-column-header

The header for a column of a table

table-row-header

The header for a row of a table

tear-off-menu-item

A menu item used to tear off and reattach its menu

terminal

An object that represents an accessible terminal. Since: ATK-0.6

text

An object that presents text to the user

toggle-button

A specialized push button that can be checked or unchecked, but does not provide a separate indicator for the current state

tool-bar

A bar or palette usually composed of push buttons or toggle buttons

tool-tip

An object that provides information about another object

tree

An object used to represent hierarchical information to the user

tree-table

An object capable of expanding and collapsing rows as well as showing multiple columns of data. Since: ATK-0.7

unknown

The object contains some Accessible information, but its role is not known

viewport

An object usually used in a scroll pane

window

A top level window with no title or border.

header

An object that serves as a document header. Since: ATK-1.1.1

footer

An object that serves as a document footer. Since: ATK-1.1.1

paragraph

An object which is contains a paragraph of text content. Since: ATK-1.1.1

ruler

An object which describes margins and tab stops, etc. for text objects which it controls (should have CONTROLLER_FOR relation to such). Since: ATK-1.1.1

application

The object is an application object, which may contain ATK.ROLE.FRAME objects or other types of accessibles. The root accessible of any application's ATK hierarchy should have ATK_ROLE_APPLICATION. Since: ATK-1.1.4

autocomplete

The object is a dialog or list containing items for insertion into an entry widget, for instance a list of words for completion of a text entry. Since: ATK-1.3

editbar

The object is an editable text object in a toolbar. Since: ATK-1.5

embedded

The object is an embedded container within a document or panel. This role is a grouping "hint" indicating that the contained objects share a context. Since: ATK-1.7.2

form

The object is a component whose textual content may be entered or modified by the user, provided ATK.STATE.EDITABLE is present. Since: ATK-1.11

last-defined

The object is a graphical depiction of quantitative data. It may contain multiple subelements whose attributes and/or description may be queried to obtain both the quantitative data and information about how the data is being presented. The LABELLED_BY relation is particularly important in interpreting objects of this type, as is the accessible-description property. Since: ATK-1.11

AtkLayer

Describes the layer of a component

These enumerated "layer values" are used when determining which UI rendering layer a component is drawn into, which can help in making determinations of when components occlude one another.

invalid

The object does not have a layer

background

This layer is reserved for the desktop background

canvas

This layer is used for Canvas components

widget

This layer is normally used for components

mdi

This layer is used for layered components

popup

This layer is used for popup components, such as menus

overlay

This layer is reserved for future use.

window

This layer is used for toplevel windows.

User Functions

AtkFunction(data)

An AtkFunction is a function definition used for padding which has been added to class and interface structures to allow for expansion in the future.

data

[R object] a gpointer to parameter data.

Returns: [logical] Nothing useful, this is only a dummy prototype.

AtkPropertyChangeHandler(Param1, Param2)

An AtkPropertyChangeHandler is a function which is executed when an AtkObject's property changes value. It is specified in a call to atkObjectConnectPropertyChangeHandler.

Param1

[AtkObject] an AtkObject

Param2

[AtkPropertyValues] an AtkPropertyValues

Signals

active-descendant-changed(atkobject, arg1, user.data)

The "active-descendant-changed" signal is emitted by an object which has the state ATK_STATE_MANAGES_DESCENDANTS when the focus object in the object changes. For instance, a table will emit the signal when the cell in the table which has focus changes.

atkobject

[AtkObject] the object which received the signal.

arg1

[R object] the newly focused object.

user.data

[R object] user data set when the signal handler was connected.

children-changed(atkobject, arg1, arg2, user.data)

The signal "children-changed" is emitted when a child is added or removed form an object. It supports two details: "add" and "remove"

atkobject

[AtkObject] the object which received the signal.

arg1

[numeric] The index of the added or removed child

arg2

[R object] A gpointer to the child AtkObject which was added or removed

user.data

[R object] user data set when the signal handler was connected.

focus-event(atkobject, arg1, user.data)

The signal "focus-event" is emitted when an object gains or loses focus.

atkobject

[AtkObject] the object which received the signal.

arg1

[logical] A boolean value which indicates whether the object gained or lost focus.

user.data

[R object] user data set when the signal handler was connected.

property-change(atkobject, arg1, user.data)

The signal "property-change" is emitted when an object's property value changes. The detail identifies the name of the property whose value has changed.

atkobject

[AtkObject] the object which received the signal.

arg1

[R object] The new value of the property which changed.

user.data

[R object] user data set when the signal handler was connected.

state-change(atkobject, arg1, arg2, user.data)

The "state-change" signal is emitted when an object's state changes. The detail value identifies the state type which has changed.

atkobject

[AtkObject] the object which received the signal.

arg1

[character] The name of the state which has changed

arg2

[logical] A boolean which indicates whether the state has been set or unset.

user.data

[R object] user data set when the signal handler was connected.

visible-data-changed(atkobject, user.data)

The "visible-data-changed" signal is emitted when the visual appearance of the object changed.

atkobject

[AtkObject] the object which received the signal.

user.data

[R object] user data set when the signal handler was connected.

Properties

accessible-component-layer [integer : Read]

The accessible layer of this object. Allowed values: >= 0 Default value: 0

accessible-component-mdi-zorder [integer : Read]

The accessible MDI value of this object. Default value: -2147483648

accessible-description [character : * : Read / Write]

Description of an object, formatted for assistive technology access. Default value: NULL

accessible-hypertext-nlinks [integer : Read]

The number of links which the current AtkHypertext has. Allowed values: >= 0 Default value: 0

accessible-name [character : * : Read / Write]

Object instance's name formatted for assistive technology access. Default value: NULL

accessible-parent [AtkObject : * : Read / Write]

Is used to notify that the parent has changed.

accessible-role [integer : Read / Write]

The accessible role of this object. Allowed values: >= 0 Default value: 0

accessible-table-caption [character : * : Read / Write]

Is used to notify that the table caption has changed; this property should not be used. accessible-table-caption-object should be used instead. Default value: NULL

accessible-table-caption-object [AtkObject : * : Read / Write]

Is used to notify that the table caption has changed.

accessible-table-column-description [character : * : Read / Write]

Is used to notify that the table column description has changed. Default value: NULL

accessible-table-column-header [AtkObject : * : Read / Write]

Is used to notify that the table column header has changed.

accessible-table-row-description [character : * : Read / Write]

Is used to notify that the table row description has changed. Default value: NULL

accessible-table-row-header [AtkObject : * : Read / Write]

Is used to notify that the table row header has changed.

accessible-table-summary [AtkObject : * : Read / Write]

Is used to notify that the table summary has changed.

accessible-value [numeric : Read / Write]

Is used to notify that the value has changed. Allowed values: >= 0 Default value: 0

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkObject.html

See Also

AtkObjectFactory AtkRegistry GtkAccessible

[Package RGtk2 version 2.20.31 Index]