AtkObjectFactory {RGtk2}	R Documentation

AtkObjectFactory

Description

The base object class for a factory used to create accessible objects for objects of a specific GType.

Methods and Functions

atkObjectFactoryCreateAccessible(object, obj)
atkObjectFactoryGetAccessibleType(object)
atkObjectFactoryInvalidate(object)

Hierarchy

GObject
   +----AtkObjectFactory
         +----AtkNoOpObjectFactory

Detailed Description

This class is the base object class for a factory used to create an accessible object for a specific GType. The function atkRegistrySetFactoryType is normally called to store in the registry the factory type to be used to create an accessible of a particular GType.

Structures

AtkObjectFactory

The AtkObjectFactory structure should not be accessed directly.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkObjectFactory.html

[Package RGtk2 version 2.20.31 Index]