AtkRegistry {RGtk2}	R Documentation

AtkRegistry

Description

An object used to store the GType of the factories used to create an accessible object for an object of a particular GType.

Methods and Functions

atkRegistrySetFactoryType(object, type, factory.type)
atkRegistryGetFactoryType(object, type)
atkRegistryGetFactory(object, type)
atkGetDefaultRegistry()

Hierarchy

GObject
   +----AtkRegistry

Detailed Description

The AtkRegistry is normally used to create appropriate ATK "peers" for user interface components. Application developers usually need only interact with the AtkRegistry by associating appropriate ATK implementation classes with GObject classes via the atk_registry_set_factory_type call, passing the appropriate GType for application custom widget classes.

Structures

AtkRegistry

The AtkRegistry structure should not be accessed directly.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkRegistry.html

[Package RGtk2 version 2.20.31 Index]