AtkRelation {RGtk2}	R Documentation

AtkRelation

Description

An object used to describe a relation between a object and one or more other objects.

Methods and Functions

atkRelationTypeRegister(name)
atkRelationTypeGetName(type)
atkRelationTypeForName(name)
atkRelationNew(targets, relationship)
atkRelationGetRelationType(object)
atkRelationGetTarget(object)
atkRelationAddTarget(object, target)
atkRelation(targets, relationship)

Hierarchy

GObject
   +----AtkRelation

Detailed Description

An AtkRelation describes a relation between an object and one or more other objects. The actual relations that an object has with other objects are defined as an AtkRelationSet, which is a set of AtkRelations.

Structures

AtkRelation

The AtkRelation structure should not be accessed directly.

Convenient Construction

atkRelation is the equivalent of atkRelationNew.

Enums and Flags

AtkRelationType

Describes the type of the relation

null

Not used, represens "no relationship" or an error condition.

controlled-by

Indicates an object controlled by one or more target objects.

controller-for

Indicates an object is an controller for one or more target objects.

label-for

Indicates an object is a label for one or more target objects.

labelled-by

Indicates an object is labelled by one or more target objects.

member-of

Indicates an object is a member of a group of one or more target objects.

node-child-of

Indicates an object is a cell in a treetable which is displayed because a cell in the same column is expanded and identifies that cell.

flows-to

Indicates that the object has content that flows logically to another AtkObject in a sequential way, (for instance text-flow).

flows-from

Indicates that the object has content that flows logically from another AtkObject in a sequential way, (for instance text-flow).

subwindow-of

Indicates a subwindow attached to a component but otherwise has no connection in the UI heirarchy to that component.

embeds

Indicates that the object visually embeds another object's content, i.e. this object's content flows around another's content.

embedded-by

Inverse of ATK_RELATION_EMBEDS, indicates that this object's content is visualy embedded in another object.

popup-for

Indicates that an object is a popup for another object.

parent-window-of

Indicates that an object is a parent window of another object.

description-for

Indicates that another object provides descriptive information about this object; more verbose than ATK_RELATION_LABELLED_BY.

described-by

Indicates that an object provides descriptive information about another object; more verbose than ATK_RELATION_LABEL_FOR.

last-defined

Indicates an object is a cell in a treetable and is expanded to display other cells in the same column.

Properties

relation-type [AtkRelationType : Read / Write]

The type of the relation. Default value: ATK_RELATION_NULL

target [GValueArray : * : Read / Write]

An list of the targets for the relation.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkRelation.html

[Package RGtk2 version 2.20.31 Index]