AtkRelationSet {RGtk2}	R Documentation

AtkRelationSet

Description

A set of AtkRelations, normally the set of AtkRelations which an AtkObject has.

Methods and Functions

atkRelationSetNew()
atkRelationSetContains(object, relationship)
atkRelationSetRemove(object, relation)
atkRelationSetAdd(object, relation)
atkRelationSetGetNRelations(object)
atkRelationSetGetRelation(object, i)
atkRelationSetGetRelationByType(object, relationship)
atkRelationSetAddRelationByType(object, relationship, target)
atkRelationSet()

Hierarchy

GObject
   +----AtkRelationSet

Detailed Description

The AtkRelationSet held by an object establishes its relationships with objects beyond the normal "parent/child" hierarchical relationships that all user interface objects have. AtkRelationSets establish whether objects are labelled or controlled by other components, share group membership with other components (for instance within a radio-button group), or share content which "flows" between them, among other types of possible relationships.

Structures

AtkRelationSet

The AtkRelationSet structure should not be accessed directly.

Convenient Construction

atkRelationSet is the equivalent of atkRelationSetNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkRelationSet.html

[Package RGtk2 version 2.20.31 Index]