AtkSelection {RGtk2}	R Documentation

AtkSelection

Description

The ATK interface implemented by container objects whose children can be selected.

Methods and Functions

atkSelectionAddSelection(object, i)
atkSelectionClearSelection(object)
atkSelectionRefSelection(object, i)
atkSelectionGetSelectionCount(object)
atkSelectionIsChildSelected(object, i)
atkSelectionRemoveSelection(object, i)
atkSelectionSelectAllSelection(object)

Hierarchy

GInterface
   +----AtkSelection

Implementations

AtkSelection is implemented by AtkNoOpObject.

Detailed Description

AtkSelection should be implemented by UI components with children which are exposed by atk_object_ref_child and atk_object_get_n_children, if the use of the parent UI component ordinarily involves selection of one or more of the objects corresponding to those AtkObject children - for example, selectable lists.

Note that other types of "selection" (for instance text selection) are accomplished a other ATK interfaces - AtkSelection is limited to the selection/deselection of children.

Structures

AtkSelection

The AtkAction structure does not contain any fields.

Signals

selection-changed(atkselection, user.data)

The "selection-changed" signal is emitted by an object which implements AtkSelection interface when the selection changes.

atkselection

[AtkSelection] the object which received the signal.

user.data

[R object] user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkSelection.html

See Also

AtkText

[Package RGtk2 version 2.20.31 Index]