AtkStreamableContent {RGtk2}	R Documentation

AtkStreamableContent

Description

The ATK interface which provides access to streamable content.

Methods and Functions

atkStreamableContentGetNMimeTypes(object)
atkStreamableContentGetMimeType(object, i)
atkStreamableContentGetStream(object, mime.type)

Hierarchy

GInterface
   +----AtkStreamableContent

Detailed Description

An interface whereby an object allows its backing content to be streamed to clients. Typical implementors would be images or icons, HTML content, or multimedia display/rendering widgets.

Negotiation of content type is allowed. Clients may examine the backing data and transform, convert, or parse the content in order to present it in an alternate form to end-users.

The AtkStreamableContent interface is particularly useful for saving, printing, or post-processing entire documents, or for persisting alternate views of a document. If document content itself is being serialized, stored, or converted, then use of the AtkStreamableContent interface can help address performance issues. Unlike most ATK interfaces, this interface is not strongly tied to the current user-agent view of the a particular document, but may in some cases give access to the underlying model data.

Structures

AtkStreamableContent

The AtkStreamableContent structure does not contain any fields.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkStreamableContent.html

[Package RGtk2 version 2.20.31 Index]