AtkTable {RGtk2}	R Documentation

AtkTable

Description

The ATK interface implemented for UI components which contain tabular or row/column information.

Methods and Functions

atkTableRefAt(object, row, column)
atkTableGetIndexAt(object, row, column)
atkTableGetColumnAtIndex(object, index)
atkTableGetRowAtIndex(object, index)
atkTableGetNColumns(object)
atkTableGetNRows(object)
atkTableGetColumnExtentAt(object, row, column)
atkTableGetRowExtentAt(object, row, column)
atkTableGetCaption(object)
atkTableGetColumnDescription(object, column)
atkTableGetRowDescription(object, row)
atkTableGetColumnHeader(object, column)
atkTableGetRowHeader(object, row)
atkTableGetSummary(object)
atkTableSetCaption(object, caption)
atkTableSetRowDescription(object, row, description)
atkTableSetColumnDescription(object, column, description)
atkTableSetRowHeader(object, row, header)
atkTableSetColumnHeader(object, column, header)
atkTableSetSummary(object, accessible)
atkTableGetSelectedColumns(object)
atkTableGetSelectedRows(object)
atkTableIsColumnSelected(object, column)
atkTableIsRowSelected(object, row)
atkTableIsSelected(object, row, column)
atkTableAddColumnSelection(object, column)
atkTableAddRowSelection(object, row)
atkTableRemoveColumnSelection(object, column)
atkTableRemoveRowSelection(object, row)

Hierarchy

GInterface
   +----AtkTable

Implementations

AtkTable is implemented by AtkNoOpObject.

Detailed Description

AtkTable should be implemented by components which present elements ordered via rows and columns. It may also be used to present tree-structured information if the nodes of the trees can be said to contain multiple "columns". Individual elements of an AtkTable are typically referred to as "cells", and these cells are exposed by AtkTable as child AtkObjects of the AtkTable. Both row/column and child-index-based access to these children is provided.

Children of AtkTable are frequently "lightweight" objects, that is, they may not have backing widgets in the host UI toolkit. They are therefore often transient.

Since tables are often very complex, AtkTable includes provision for offering simplified summary information, as well as row and column headers and captions. Headers and captions are AtkObjects which may implement other interfaces (AtkText, AtkImage, etc.) as appropriate. AtkTable summaries may themselves be (simplified) AtkTables, etc.

Structures

AtkTable

The AtkTable structure does not contain any fields.

Signals

column-deleted(atktable, arg1, arg2, user.data)

The "column-deleted" signal is emitted by an object which implements the AtkTable interface when a column is deleted.

atktable

[AtkTable] the object which received the signal.

arg1

[integer] The index of the first column deleted.

arg2

[integer] The number of columns deleted.

user.data

[R object] user data set when the signal handler was connected.

column-inserted(atktable, arg1, arg2, user.data)

The "column-inserted" signal is emitted by an object which implements the AtkTable interface when a column is inserted.

atktable

[AtkTable] the object which received the signal.

arg1

[integer] The index of the column inserted.

arg2

[integer] The number of colums inserted.

user.data

[R object] user data set when the signal handler was connected.

column-reordered(atktable, user.data)

The "column-reordered" signal is emitted by an object which implements the AtkTable interface when the columns are reordered.

atktable

[AtkTable] the object which received the signal.

user.data

[R object] user data set when the signal handler was connected.

model-changed(atktable, user.data)

The "model-changed" signal is emitted by an object which implements the AtkTable interface when the model displayed by the table changes.

atktable

[AtkTable] the object which received the signal.

user.data

[R object] user data set when the signal handler was connected.

row-deleted(atktable, arg1, arg2, user.data)

The "row-deleted" signal is emitted by an object which implements the AtkTable interface when a column is inserted.

atktable

[AtkTable] the object which received the signal.

arg1

[integer] The index of the first row deleted.

arg2

[integer] The number of rows deleted.

user.data

[R object] user data set when the signal handler was connected.

row-inserted(atktable, arg1, arg2, user.data)

The "row-inserted" signal is emitted by an object which implements the AtkTable interface when a column is inserted.

atktable

[AtkTable] the object which received the signal.

arg1

[integer] The index of the first row deleted.

arg2

[integer] The number of rows deleted.

user.data

[R object] user data set when the signal handler was connected.

row-reordered(atktable, user.data)

The "row-reordered" signal is emitted by an object which implements the AtkTable interface when the columns are reordered.

atktable

[AtkTable] the object which received the signal.

user.data

[R object] user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkTable.html

See Also

AtkObject

[Package RGtk2 version 2.20.31 Index]