AtkValue {RGtk2}	R Documentation

AtkValue

Description

The ATK interface implemented by valuators and components which display or select a value from a bounded range of values.

Methods and Functions

atkValueGetCurrentValue(object)
atkValueGetMaximumValue(object)
atkValueGetMinimumValue(object)
atkValueSetCurrentValue(object, value)

Hierarchy

GInterface
   +----AtkValue

Implementations

AtkValue is implemented by AtkNoOpObject.

Detailed Description

AtkValue should be implemented for components which either display a value from a bounded range, or which allow the user to specify a value from a bounded range, or both. For instance, most sliders and range controls, as well as dials, should have AtkObject representations which implement AtkValue on the component's behalf. AtKValues may be read-only, in which case attempts to alter the value return FALSE to indicate failure.

Structures

AtkValue

The AtkValue structure does not contain any fields.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//atk/AtkValue.html

[Package RGtk2 version 2.20.31 Index]