CAIRO {RGtk2}	R Documentation

CAIRO

Description

Cairo is a 2D graphics library with support for multiple output devices. Currently supported output targets include the X Window System, win32, and image buffers.

Details

The RGtk binding to the CAIRO library consists of the following components:

cairo-font-face

Base class for font faces

cairo-font-options

How a font should be rendered

cairo-image-surface

Rendering to memory buffers

cairo-matrix

Generic matrix operations

cairo-paths

Creating paths and manipulating path data

cairo-pattern

Sources for drawing

cairo-pdf-surface

Rendering PDF documents

cairo-png-functions

Reading and writing PNG images

cairo-ps-surface

Rendering PostScript documents

cairo-scaled-font

Font face at particular size and options

cairo-error-status

Decoding cairo's status

cairo-surface

Base class for surfaces

cairo-svg-surface

Rendering SVG documents

cairo-text

Rendering text and glyphs

cairo-transformations

Manipulating the current transformation matrix

cairo-types

Generic data types

cairo-user-font

Font support with font data provided by the user

cairo-version-info

Compile-time and run-time version checks.

cairo-context

The cairo drawing context

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://www.cairographics.org/manual

[Package RGtk2 version 2.20.31 Index]