GAsyncInitable {RGtk2}	R Documentation

GAsyncInitable

Description

Asynchronously failable object initialization interface

Methods and Functions

gAsyncInitableInitAsync(object, io.priority, cancellable = NULL, callback, user.data = NULL)
gAsyncInitableInitFinish(object, res, .errwarn = TRUE)
gAsyncInitableNewAsync(object.type, io.priority, cancellable, callback, user.data, ...)
gAsyncInitableNewFinish(object, res, .errwarn = TRUE)

Hierarchy

GInterface
   +----GAsyncInitable

Detailed Description

This is the asynchronous version of GInitable, it behaves the same in all ways except that initialization is asynchronous. For more details see the descriptions on GInitable.

A class may implement both the GInitable and GAsyncInitable interfaces.

Users of objects implementing this are not intended to use the interface method directly, instead it will be used automatically in various ways. For C applications you generally just call gAsyncInitableNewAsync directly, or indirectly via a fooThingNewAsync() wrapper. This will call gAsyncInitableInitAsync under the cover, calling back with NULL and a set GError on failure.

Structures

GAsyncInitable

Interface for asynchronously initializable objects. Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GAsyncInitable.html

[Package RGtk2 version 2.20.31 Index]