GAsyncResult {RGtk2}	R Documentation

GAsyncResult

Description

Asynchronous Function Results

Methods and Functions

gAsyncResultGetUserData(object)
gAsyncResultGetSourceObject(object)

Hierarchy

GInterface
   +----GAsyncResult

Implementations

GAsyncResult is implemented by GSimpleAsyncResult.

Detailed Description

Provides a base class for implementing asynchronous function results.

Asynchronous operations are broken up into two separate operations which are chained together by a GAsyncReadyCallback. To begin an asynchronous operation, provide a GAsyncReadyCallback to the asynchronous function. This callback will be triggered when the operation has completed, and will be passed a GAsyncResult instance filled with the details of the operation's success or failure, the object the asynchronous function was started for and any error codes returned. The asynchronous callback function is then expected to call the corresponding "finish()" function with the object the function was called for, and the GAsyncResult instance, and optionally, an error to grab any error conditions that may have occurred.

The purpose of the "finish()" function is to take the generic result of type GAsyncResult and return the specific result that the operation in question yields (e.g. a GFileEnumerator for a "enumerate children" operation). If the result or error status of the operation is not needed, there is no need to call the "finish()" function, GIO will take care of cleaning up the result and error information after the GAsyncReadyCallback returns. It is also allowed to take a reference to the GAsyncResult and call "finish()" later.

Example of a typical asynchronous operation flow:

frobnitz_result_func <- function(source_object, res, user_data)
{
  success <- _theoretical_frobnitz_finish (source_object, res, NULL)

  if (success)
    message("Hurray!")
  else 
    message("Uh oh!")

  ## ....
}

_theoretical_frobnitz_async (theoretical_data, 
                             NULL, 
                             frobnitz_result_func, 
                             NULL)

The callback for an asynchronous operation is called only once, and is always called, even in the case of a cancelled operation. On cancellation the result is a G_IO_ERROR_CANCELLED error.

Some ascynchronous operations are implemented using synchronous calls. These are run in a separate thread, if GThread has been initialized, but otherwise they are sent to the Main Event Loop and processed in an idle function. So, if you truly need asynchronous operations, make sure to initialize GThread.

Structures

GAsyncResult

Holds results information for an asynchronous operation, usually passed directly to a asynchronous finish() operation.

User Functions

GAsyncReadyCallback(source.object, res, user.data)

Type definition for a function that will be called back when an asynchronous operation within GIO has been completed.

source.object

the object the asynchronous operation was started with.

res

a GAsyncResult.

user.data

user data passed to the callback.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GAsyncResult.html

[Package RGtk2 version 2.20.31 Index]