GBufferedInputStream {RGtk2}	R Documentation

GBufferedInputStream

Description

Buffered Input Stream

Methods and Functions

gBufferedInputStreamNew(base.stream = NULL)
gBufferedInputStreamNewSized(base.stream, size)
gBufferedInputStreamGetBufferSize(object)
gBufferedInputStreamSetBufferSize(object, size)
gBufferedInputStreamGetAvailable(object)
gBufferedInputStreamPeekBuffer(object)
gBufferedInputStreamFill(object, count, cancellable = NULL, .errwarn = TRUE)
gBufferedInputStreamFillAsync(object, count, io.priority = 0, cancellable = NULL, callback, user.data = NULL)
gBufferedInputStreamFillFinish(object, result, .errwarn = TRUE)
gBufferedInputStreamReadByte(object, cancellable = NULL, .errwarn = TRUE)
gBufferedInputStream(base.stream, size)

Hierarchy

GObject
   +----GInputStream
         +----GFilterInputStream
               +----GBufferedInputStream
                     +----GDataInputStream

Detailed Description

Buffered input stream implements GFilterInputStream and provides for buffered reads.

By default, GBufferedInputStream's buffer size is set at 4 kilobytes.

To create a buffered input stream, use gBufferedInputStreamNew, or gBufferedInputStreamNewSized to specify the buffer's size at construction.

To get the size of a buffer within a buffered input stream, use gBufferedInputStreamGetBufferSize. To change the size of a buffered input stream's buffer, use gBufferedInputStreamSetBufferSize. Note that the buffer's size cannot be reduced below the size of the data within the buffer.

Structures

GBufferedInputStream

Implements GFilterInputStream with a sized input buffer.

Convenient Construction

gBufferedInputStream is the result of collapsing the constructors of GBufferedInputStream (gBufferedInputStreamNew, gBufferedInputStreamNewSized) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Properties

buffer-size [numeric : Read / Write / Construct]

The size of the backend buffer. Allowed values: >= 1 Default value: 4096

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GBufferedInputStream.html

[Package RGtk2 version 2.20.31 Index]