GBufferedOutputStream {RGtk2}	R Documentation

GBufferedOutputStream

Description

Buffered Output Stream

Methods and Functions

gBufferedOutputStreamNew(base.stream = NULL)
gBufferedOutputStreamNewSized(base.stream, size)
gBufferedOutputStreamGetBufferSize(object)
gBufferedOutputStreamSetBufferSize(object, size)
gBufferedOutputStreamGetAutoGrow(object)
gBufferedOutputStreamSetAutoGrow(object, auto.grow)
gBufferedOutputStream(base.stream, size)

Hierarchy

GObject
   +----GOutputStream
         +----GFilterOutputStream
               +----GBufferedOutputStream

Detailed Description

Buffered output stream implements GFilterOutputStream and provides for buffered writes.

By default, GBufferedOutputStream's buffer size is set at 4 kilobytes.

To create a buffered output stream, use gBufferedOutputStreamNew, or gBufferedOutputStreamNewSized to specify the buffer's size at construction.

To get the size of a buffer within a buffered input stream, use gBufferedOutputStreamGetBufferSize. To change the size of a buffered output stream's buffer, use gBufferedOutputStreamSetBufferSize. Note that the buffer's size cannot be reduced below the size of the data within the buffer.

Structures

GBufferedOutputStream

An implementation of GFilterOutputStream with a sized buffer.

Convenient Construction

gBufferedOutputStream is the result of collapsing the constructors of GBufferedOutputStream (gBufferedOutputStreamNew, gBufferedOutputStreamNewSized) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Properties

auto-grow [logical : Read / Write]

Whether the buffer should automatically grow. Default value: FALSE

buffer-size [numeric : Read / Write / Construct]

The size of the backend buffer. Allowed values: >= 1 Default value: 4096

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GBufferedOutputStream.html

[Package RGtk2 version 2.20.31 Index]