GDK {RGtk2}	R Documentation

GDK

Description

GDK is the abstraction layer that allows GTK+ to support multiple windowing systems. GDK provides drawing and window system facilities on X11, Windows, and the Linux framebuffer device.

Details

The RGtk binding to the GDK library consists of the following components:

gdk-Cairo-Interaction

Functions to support using Cairo

gdk-Colormaps-and-Colors

Manipulation of colors and colormaps

gdk-Cursors

Standard and pixmap cursors

gdk-Drag-and-Drop

Functions for controlling drag and drop handling

gdk-Drawing-Primitives

Functions for drawing points, lines, arcs, and text

gdk-Event-Structures

Data structures specific to each type of event

gdk-Events

Functions for handling events from the window system

gdk-Fonts

Loading and manipulating fonts

gdk-Graphics-Contexts

Objects to encapsulate drawing properties

gdk-Application-launching

Startup notification for applications

GdkDisplay

Controls the keyboard/mouse pointer grabs and a set of s

GdkDisplayManager

Maintains a list of all open s

GdkScreen

Object representing a physical screen

gdk-Testing

Test utilities

gdk-General

Library initialization and miscellaneous functions

gdk-Images

A client-side area for bit-mapped graphics

gdk-Input-Devices

Functions for handling extended input devices

gdk-Keyboard-Handling

Functions for manipulating keyboard codes

gdk-Pango-Interaction

Using Pango in GDK

gdk-Pixbufs

Functions for rendering pixbufs on drawables

gdk-Bitmaps-and-Pixmaps

Offscreen drawables

gdk-Properties-and-Atoms

Functions to manipulate properties on windows

gdk-Points-Rectangles-and-Regions

Simple graphical data types

gdk-GdkRGB

Renders RGB, grayscale, or indexed image data to a GdkDrawable

gdk-Visuals

Low-level display hardware information

gdk-Windows

Onscreen display areas in the target window system

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk

[Package RGtk2 version 2.20.31 Index]