GDK-Pixbuf {RGtk2}	R Documentation

GDK-Pixbuf

Description

This is a small library which allows you to create GdkPixbuf ('pixel buffer') objects from image data or image files. Use a GdkPixbuf in combination with GtkImage to display images.

Details

The RGtk binding to the GDK-Pixbuf library consists of the following components:

gdk-pixbuf-animation

Animated images.

gdk-pixbuf-creating

Creating a pixbuf from image data that is already in memory.

gdk-pixbuf-File-Loading

Loading a pixbuf from a file.

gdk-pixbuf-File-saving

Saving a pixbuf to a file.

GdkPixbufLoader

Application-driven progressive image loading.

gdk-pixbuf-gdk-pixbuf

Information that describes an image.

gdk-pixbuf-Versioning

Library version numbers.

gdk-pixbuf-Module-Interface

Extending

gdk-pixbuf-scaling

Scaling pixbufs and scaling and compositing pixbufs

gdk-pixbuf-util

Utility and miscellaneous convenience functions.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk-pixbuf

[Package RGtk2 version 2.20.31 Index]