GDataInputStream {RGtk2}	R Documentation

GDataInputStream

Description

Data Input Stream

Methods and Functions

gDataInputStreamNew(base.stream = NULL)
gDataInputStreamSetByteOrder(object, order)
gDataInputStreamGetByteOrder(object)
gDataInputStreamSetNewlineType(object, type)
gDataInputStreamGetNewlineType(object)
gDataInputStreamReadByte(object, cancellable = NULL, .errwarn = TRUE)
gDataInputStreamReadInt16(object, cancellable = NULL, .errwarn = TRUE)
gDataInputStreamReadUint16(object, cancellable = NULL, .errwarn = TRUE)
gDataInputStreamReadInt32(object, cancellable = NULL, .errwarn = TRUE)
gDataInputStreamReadUint32(object, cancellable = NULL, .errwarn = TRUE)
gDataInputStreamReadInt64(object, cancellable = NULL, .errwarn = TRUE)
gDataInputStreamReadUint64(object, cancellable = NULL, .errwarn = TRUE)
gDataInputStreamReadLine(object, cancellable = NULL, .errwarn = TRUE)
gDataInputStreamReadLineAsync(object, io.priority, cancellable = NULL, callback, user.data = NULL)
gDataInputStreamReadLineFinish(object, result, .errwarn = TRUE)
gDataInputStreamReadUntil(object, stop.chars, cancellable = NULL, .errwarn = TRUE)
gDataInputStreamReadUntilAsync(object, stop.chars, io.priority, cancellable = NULL, callback, user.data = NULL)
gDataInputStreamReadUntilFinish(object, result, length, .errwarn = TRUE)
gDataInputStream(base.stream = NULL)

Hierarchy

GObject
   +----GInputStream
         +----GFilterInputStream
               +----GBufferedInputStream
                     +----GDataInputStream
GEnum
   +----GDataStreamByteOrder
GEnum
   +----GDataStreamNewlineType

Detailed Description

Data input stream implements GInputStream and includes functions for reading structured data directly from a binary input stream.

Structures

GDataInputStream

An implementation of GBufferedInputStream that allows for high-level data manipulation of arbitrary data (including binary operations).

Convenient Construction

gDataInputStream is the equivalent of gDataInputStreamNew.

Enums and Flags

GDataStreamByteOrder

GDataStreamByteOrder is used to ensure proper endianness of streaming data sources across various machine architectures.

big-endian

Selects Big Endian byte order.

little-endian

Selects Little Endian byte order.

host-endian

Selects endianness based on host machine's architecture.

GDataStreamNewlineType

GDataStreamNewlineType is used when checking for or setting the line endings for a given file.

lf

Selects "LF" line endings, common on most modern UNIX platforms.

cr

Selects "CR" line endings.

cr-lf

Selects "CR, LF" line ending, common on Microsoft Windows.

any

Automatically try to handle any line ending type.

Properties

byte-order [GDataStreamByteOrder : Read / Write]

The byte order. Default value: G_DATA_STREAM_BYTE_ORDER_BIG_ENDIAN

newline-type [GDataStreamNewlineType : Read / Write]

The accepted types of line ending. Default value: G_DATA_STREAM_NEWLINE_TYPE_LF

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GDataInputStream.html

[Package RGtk2 version 2.20.31 Index]