GDataOutputStream {RGtk2}	R Documentation

GDataOutputStream

Description

Data Output Stream

Methods and Functions

gDataOutputStreamNew(base.stream = NULL)
gDataOutputStreamSetByteOrder(object, order)
gDataOutputStreamGetByteOrder(object)
gDataOutputStreamPutByte(object, data, cancellable = NULL, .errwarn = TRUE)
gDataOutputStreamPutInt16(object, data, cancellable = NULL, .errwarn = TRUE)
gDataOutputStreamPutUint16(object, data, cancellable = NULL, .errwarn = TRUE)
gDataOutputStreamPutInt32(object, data, cancellable = NULL, .errwarn = TRUE)
gDataOutputStreamPutUint32(object, data, cancellable = NULL, .errwarn = TRUE)
gDataOutputStreamPutInt64(object, data, cancellable = NULL, .errwarn = TRUE)
gDataOutputStreamPutUint64(object, data, cancellable = NULL, .errwarn = TRUE)
gDataOutputStreamPutString(object, str, cancellable = NULL, .errwarn = TRUE)
gDataOutputStream(base.stream = NULL)

Hierarchy

GObject
   +----GOutputStream
         +----GFilterOutputStream
               +----GDataOutputStream

Detailed Description

Data output stream implements GOutputStream and includes functions for writing data directly to an output stream.

Structures

GDataOutputStream

An implementation of GBufferedOutputStream that allows for high-level data manipulation of arbitrary data (including binary operations).

Convenient Construction

gDataOutputStream is the equivalent of gDataOutputStreamNew.

Properties

byte-order [GDataStreamByteOrder : Read / Write]

Determines the byte ordering that is used when writing multi-byte entities (such as integers) to the stream. Default value: G_DATA_STREAM_BYTE_ORDER_BIG_ENDIAN

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GDataOutputStream.html

[Package RGtk2 version 2.20.31 Index]