GEmblem {RGtk2}	R Documentation

GEmblem

Description

An object for emblems

Methods and Functions

gEmblemNew(icon = NULL, origin = NULL)
gEmblemNewWithOrigin(icon, origin)
gEmblemGetIcon(object)
gEmblemGetOrigin(object)
gEmblem(icon, origin)

Hierarchy

GObject
   +----GEmblem
GEnum
   +----GEmblemOrigin

Interfaces

GEmblem implements GIcon.

Detailed Description

GEmblem is an implementation of GIcon that supports having an emblem, which is an icon with additional properties. It can than be added to a GEmblemedIcon.

Currently, only metainformation about the emblem's origin is supported. More may be added in the future.

Structures

GEmblem

An object for Emblems

Convenient Construction

gEmblem is the result of collapsing the constructors of GEmblem (gEmblemNew, gEmblemNewWithOrigin) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Enums and Flags

GEmblemOrigin

GEmblemOrigin is used to add information about the origin of the emblem to GEmblem. Since 2.18

unknown

Emblem of unknown origin

device

Emblem adds device-specific information

livemetadata

Emblem depicts live metadata, such as "readonly"

tag

Emblem comes from a user-defined tag, e.g. set by nautilus (in the future)

Properties

icon [GObject : * : Read / Write / Construct Only]

The actual icon of the emblem.

origin [GEmblemOrigin : Read / Write / Construct Only]

Tells which origin the emblem is derived from. Default value: G_EMBLEM_ORIGIN_UNKNOWN

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GEmblem.html

[Package RGtk2 version 2.20.31 Index]