GEmblemedIcon {RGtk2}	R Documentation

GEmblemedIcon

Description

Icon with emblems

Methods and Functions

gEmblemedIconNew(icon, emblem)
gEmblemedIconGetIcon(object)
gEmblemedIconGetEmblems(object)
gEmblemedIconAddEmblem(object, emblem)
gEmblemedIcon(icon, emblem)

Hierarchy

GObject
   +----GEmblemedIcon

Interfaces

GEmblemedIcon implements GIcon.

Detailed Description

GEmblemedIcon is an implementation of GIcon that supports adding an emblem to an icon. Adding multiple emblems to an icon is ensured via gEmblemedIconAddEmblem.

Note that GEmblemedIcon allows no control over the position of the emblems. See also GEmblem for more information.

Structures

GEmblemedIcon

An implementation of GIcon for icons with emblems.

Convenient Construction

gEmblemedIcon is the equivalent of gEmblemedIconNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GEmblemedIcon.html

[Package RGtk2 version 2.20.31 Index]