GError {RGtk2}	R Documentation

The GError object

Description

GLib provides a standard method of reporting errors from a called function to the calling code. (This is the same problem solved by exceptions in other languages.)

Details

As a transparent-type, a GError is returned to the user as a list with the following elements:

domain

a numeric vector with one named element that identifies the error domain

code

an integer code (often an enum value) identifying the specific error type within the domain

message

a descriptive error message

The domain is usually retrived via a function of the form NAME\_OF\_DOMAIN\_ERROR(). For example, for GdkPixbuf, the error domain is provided by GDK_PIXBUF_ERROR().

References

http://developer.gnome.org/doc/API/2.0/glib/glib-Error-Reporting.html

See Also

GdkPixbufError GtkFileChooserError GtkPrintError GtkRecentChooserError GtkRecentManagerError

[Package RGtk2 version 2.20.31 Index]