GFileAttribute {RGtk2}	R Documentation

File attribute keys

Description

File attributes in GIO consist of a list of key-value pairs, organized by namespace. Each of the GFileAttribute* objects is a character vector of keys belonging to the same namespace. For example, GFileAttributeStandard["name"] returns the key for the file name in the standard namespace. See gio-GFileAttribute for more information.

Author(s)

Michael Lawrence

[Package RGtk2 version 2.20.31 Index]