GFileEnumerator {RGtk2}	R Documentation

GFileEnumerator

Description

Enumerated Files Routines

Methods and Functions

gFileEnumeratorNextFile(object, cancellable = NULL, .errwarn = TRUE)
gFileEnumeratorClose(object, cancellable = NULL, .errwarn = TRUE)
gFileEnumeratorNextFilesAsync(object, num.files, io.priority = 0, cancellable = NULL, callback, user.data = NULL)
gFileEnumeratorNextFilesFinish(object, result, .errwarn = TRUE)
gFileEnumeratorCloseAsync(object, io.priority = 0, cancellable = NULL, callback, user.data = NULL)
gFileEnumeratorCloseFinish(object, result, .errwarn = TRUE)
gFileEnumeratorIsClosed(object)
gFileEnumeratorHasPending(object)
gFileEnumeratorSetPending(object, pending)
gFileEnumeratorGetContainer(object)

Hierarchy

GObject
   +----GFileEnumerator

Detailed Description

GFileEnumerator allows you to operate on a set of GFiles, returning a GFileInfo structure for each file enumerated (e.g. gFileEnumerateChildren will return a GFileEnumerator for each of the children within a directory).

To get the next file's information from a GFileEnumerator, use gFileEnumeratorNextFile or its asynchronous version, gFileEnumeratorNextFilesAsync. Note that the asynchronous version will return a list of GFileInfos, whereas the synchronous will only return the next file in the enumerator.

To close a GFileEnumerator, use gFileEnumeratorClose, or its asynchronous version, gFileEnumeratorCloseAsync. Once a GFileEnumerator is closed, no further actions may be performed on it, and it should be freed with gObjectUnref().

Structures

GFileEnumerator

A per matched file iterator.

Properties

container [GFile : * : Write / Construct Only]

The container that is being enumerated.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GFileEnumerator.html

[Package RGtk2 version 2.20.31 Index]