GFileIOStream {RGtk2}	R Documentation

GFileIOStream

Description

File read and write streaming operations

Methods and Functions

gFileIOStreamGetEtag(object)
gFileIOStreamQueryInfo(object, attributes, cancellable = NULL, .errwarn = TRUE)
gFileIOStreamQueryInfoAsync(object, attributes, io.priority, cancellable = NULL, callback, user.data = NULL)
gFileIOStreamQueryInfoFinish(object, result, .errwarn = TRUE)

Hierarchy

GObject
   +----GIOStream
         +----GFileIOStream

Interfaces

GFileIOStream implements GSeekable.

Detailed Description

GFileIOStream provides io streams that both read and write to the same file handle.

GFileIOStream implements GSeekable, which allows the io stream to jump to arbitrary positions in the file and to truncate the file, provided the filesystem of the file supports these operations.

To find the position of a file io stream, use gSeekableTell.

To find out if a file io stream supports seeking, use gSeekableCanSeek. To position a file io stream, use gSeekableSeek. To find out if a file io stream supports truncating, use gSeekableCanTruncate. To truncate a file io stream, use gSeekableTruncate.

The default implementation of all the GFileIOStream operations and the implementation of GSeekable just call into the same operations on the output stream.

Structures

GFileIOStream

A subclass of GIOStream for opened files. This adds a few file-specific operations and seeking and truncating. GFileIOStream implements GSeekable.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GFileIOStream.html

[Package RGtk2 version 2.20.31 Index]