GFileInfo {RGtk2}	R Documentation

GFileInfo

Description

File Information and Attributes

Methods and Functions

gFileInfoNew()
gFileInfoDup(object)
gFileInfoCopyInto(object, dest.info)
gFileInfoHasAttribute(object, attribute)
gFileInfoHasNamespace(object, name.space)
gFileInfoListAttributes(object, name.space)
gFileInfoGetAttributeType(object, attribute)
gFileInfoRemoveAttribute(object, attribute)
gFileInfoGetAttributeAsString(object, attribute)
gFileInfoGetAttributeData(object, attribute, type, status)
gFileInfoGetAttributeStatus(object, attribute)
gFileInfoGetAttributeString(object, attribute)
gFileInfoGetAttributeStringv(object, attribute)
gFileInfoGetAttributeByteString(object, attribute)
gFileInfoGetAttributeBoolean(object, attribute)
gFileInfoGetAttributeUint32(object, attribute)
gFileInfoGetAttributeInt32(object, attribute)
gFileInfoGetAttributeUint64(object, attribute)
gFileInfoGetAttributeInt64(object, attribute)
gFileInfoGetAttributeObject(object, attribute)
gFileInfoSetAttribute(object, attribute, type, value.p)
gFileInfoSetAttributeStatus(object, attribute, status)
gFileInfoSetAttributeString(object, attribute, attr.value)
gFileInfoSetAttributeStringv(object, attribute, attr.value)
gFileInfoSetAttributeByteString(object, attribute, attr.value)
gFileInfoSetAttributeBoolean(object, attribute, attr.value)
gFileInfoSetAttributeUint32(object, attribute, attr.value)
gFileInfoSetAttributeInt32(object, attribute, attr.value)
gFileInfoSetAttributeUint64(object, attribute, attr.value)
gFileInfoSetAttributeInt64(object, attribute, attr.value)
gFileInfoSetAttributeObject(object, attribute, attr.value)
gFileInfoClearStatus(object)
gFileInfoGetFileType(object)
gFileInfoGetIsHidden(object)
gFileInfoGetIsBackup(object)
gFileInfoGetIsSymlink(object)
gFileInfoGetName(object)
gFileInfoGetDisplayName(object)
gFileInfoGetEditName(object)
gFileInfoGetIcon(object)
gFileInfoGetContentType(object)
gFileInfoGetSize(object)
gFileInfoGetModificationTime(object, result)
gFileInfoGetSymlinkTarget(object)
gFileInfoGetEtag(object)
gFileInfoGetSortOrder(object)
gFileInfoSetAttributeMask(object, mask)
gFileInfoUnsetAttributeMask(object)
gFileInfoSetFileType(object, type)
gFileInfoSetIsHidden(object, is.hidden)
gFileInfoSetIsSymlink(object, is.symlink)
gFileInfoSetName(object, name)
gFileInfoSetDisplayName(object, display.name)
gFileInfoSetEditName(object, edit.name)
gFileInfoSetIcon(object, icon)
gFileInfoSetContentType(object, content.type)
gFileInfoSetSize(object, size)
gFileInfoSetModificationTime(object, mtime)
gFileInfoSetSymlinkTarget(object, symlink.target)
gFileInfoSetSortOrder(object, sort.order)
gFileAttributeMatcherNew(attributes)
gFileAttributeMatcherMatches(object, attribute)
gFileAttributeMatcherMatchesOnly(object, attribute)
gFileAttributeMatcherEnumerateNamespace(object, ns)
gFileAttributeMatcherEnumerateNext(object)
gFileInfo()

Hierarchy

GEnum
   +----GFileType
GObject
   +----GFileInfo

Detailed Description

Functionality for manipulating basic metadata for files. GFileInfo implements methods for getting information that all files should contain, and allows for manipulation of extended attributes.

See GFileAttribute for more information on how GIO handles file attributes.

To obtain a GFileInfo for a GFile, use gFileQueryInfo (or its async variant). To obtain a GFileInfo for a file input or output stream, use gFileInputStreamQueryInfo or gFileOutputStreamQueryInfo (or their async variants).

To change the actual attributes of a file, you should then set the attribute in the GFileInfo and call gFileSetAttributesFromInfo or gFileSetAttributesAsync on a GFile.

However, not all attributes can be changed in the file. For instance, the actual size of a file cannot be changed via gFileInfoSetSize. You may call gFileQuerySettableAttributes and gFileQueryWritableNamespaces to discover the settable attributes of a particular file at runtime. GFileAttributeMatcher allows for searching through a GFileInfo for attributes.

Structures

GFileAttributeMatcher

Determines if a string matches a file attribute.

GFileInfo

Stores information about a file system object referenced by a GFile.

Convenient Construction

gFileInfo is the equivalent of gFileInfoNew.

Enums and Flags

GFileType

Indicates the file's on-disk type.

unknown

File's type is unknown.

regular

File handle represents a regular file.

directory

File handle represents a directory.

symbolic-link

File handle represents a symbolic link (Unix systems).

special

File is a "special" file, such as a socket, fifo, block device, or character device.

shortcut

File is a shortcut (Windows systems).

mountable

File is a mountable location.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GFileInfo.html

[Package RGtk2 version 2.20.31 Index]