GFileInputStream {RGtk2}	R Documentation

GFileInputStream

Description

File input streaming operations

Methods and Functions

gFileInputStreamQueryInfo(object, attributes, cancellable = NULL, .errwarn = TRUE)
gFileInputStreamQueryInfoAsync(object, attributes, io.priority = 0, cancellable = NULL, callback, user.data = NULL)
gFileInputStreamQueryInfoFinish(object, result, .errwarn = TRUE)

Hierarchy

GObject
   +----GInputStream
         +----GFileInputStream

Interfaces

GFileInputStream implements GSeekable.

Detailed Description

GFileInputStream provides input streams that take their content from a file.

GFileInputStream implements GSeekable, which allows the input stream to jump to arbitrary positions in the file, provided the filesystem of the file allows it. To find the position of a file input stream, use gSeekableTell. To find out if a file input stream supports seeking, use gSeekableStreamCanSeek(). To position a file input stream, use gSeekableSeek.

Structures

GFileInputStream

A subclass of GInputStream for opened files. This adds a few file-specific operations and seeking. GFileInputStream implements GSeekable.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GFileInputStream.html

[Package RGtk2 version 2.20.31 Index]