GFileMonitor-concept {RGtk2}	R Documentation

GFileMonitor

Description

File Monitor

Methods and Functions

gFileMonitorCancel(object)
gFileMonitorIsCancelled(object)
gFileMonitorSetRateLimit(object, limit.msecs)
gFileMonitorEmitEvent(object, file, other.file, event.type)

Hierarchy

GEnum
   +----GFileMonitorEvent
GObject
   +----GFileMonitor

Detailed Description

Monitors a file or directory for changes.

To obtain a GFileMonitor for a file or directory, use gFileMonitor, gFileMonitorFile, or gFileMonitorDirectory.

To get informed about changes to the file or directory you are monitoring, connect to the "changed" signal. The signal will be emitted in the thread-default main context of the thread that the monitor was created in (though if the global default main context is blocked, this may cause notifications to be blocked even if the thread-default context is still running).

Structures

GFileMonitor

Watches for changes to a file.

Enums and Flags

GFileMonitorEvent

Specifies what type of event a monitor event is.

changed

a file changed.

changes-done-hint

a hint that this was probably the last change in a set of changes.

deleted

a file was deleted.

created

a file was created.

attribute-changed

a file attribute was changed.

pre-unmount

the file location will soon be unmounted.

unmounted

the file location was unmounted.

Signals

changed(monitor, file, other.file, event.type, user.data)

Emitted when a file has been changed.

monitor

a GFileMonitor.

file

a GFile.

other.file

a GFile.

event.type

a GFileMonitorEvent.

user.data

user data set when the signal handler was connected.

Properties

cancelled [logical : Read]

Whether the monitor has been cancelled. Default value: FALSE

rate-limit [integer : Read / Write]

The limit of the monitor to watch for changes, in milliseconds. Allowed values: >= 0 Default value: 800

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GFileMonitor.html

[Package RGtk2 version 2.20.31 Index]