GFileOutputStream {RGtk2}	R Documentation

GFileOutputStream

Description

File output streaming operations

Methods and Functions

gFileOutputStreamQueryInfo(object, attributes, cancellable = NULL, .errwarn = TRUE)
gFileOutputStreamQueryInfoAsync(object, attributes, io.priority = 0, cancellable = NULL, callback, user.data = NULL)
gFileOutputStreamQueryInfoFinish(object, result, .errwarn = TRUE)
gFileOutputStreamGetEtag(object)

Hierarchy

GObject
   +----GOutputStream
         +----GFileOutputStream

Interfaces

GFileOutputStream implements GSeekable.

Detailed Description

GFileOutputStream provides output streams that write their content to a file.

GFileOutputStream implements GSeekable, which allows the output stream to jump to arbitrary positions in the file and to truncate the file, provided the filesystem of the file supports these operations.

To find the position of a file output stream, use gSeekableTell. To find out if a file output stream supports seeking, use gSeekableCanSeek.To position a file output stream, use gSeekableSeek. To find out if a file output stream supports truncating, use gSeekableCanTruncate. To truncate a file output stream, use gSeekableTruncate.

Structures

GFileOutputStream

A subclass of GOutputStream for opened files. This adds a few file-specific operations and seeking and truncating. GFileOutputStream implements GSeekable.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GFileOutputStream.html

[Package RGtk2 version 2.20.31 Index]