GFilenameCompleter {RGtk2}	R Documentation

GFilenameCompleter

Description

Filename Completer

Methods and Functions

gFilenameCompleterNew()
gFilenameCompleterGetCompletionSuffix(object, initial.text)
gFilenameCompleterGetCompletions(object, initial.text)
gFilenameCompleterSetDirsOnly(object, dirs.only)
gFilenameCompleter()

Hierarchy

GObject
   +----GFilenameCompleter

Detailed Description

Completes partial file and directory names given a partial string by looking in the file system for clues. Can return a list of possible completion strings for widget implementations.

Structures

GFilenameCompleter

Completes filenames based on files that exist within the file system.

Convenient Construction

gFilenameCompleter is the equivalent of gFilenameCompleterNew.

Signals

got-completion-data(user.data)

Emitted when the file name completion information comes available.

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GFilenameCompleter.html

[Package RGtk2 version 2.20.31 Index]