GFilterOutputStream {RGtk2}	R Documentation

GFilterOutputStream

Description

Filter Output Stream

Methods and Functions

gFilterOutputStreamGetBaseStream(object)
gFilterOutputStreamGetCloseBaseStream(object)
gFilterOutputStreamSetCloseBaseStream(object, close.base)

Hierarchy

GObject
   +----GOutputStream
         +----GFilterOutputStream
               +----GBufferedOutputStream
               +----GDataOutputStream

Structures

GFilterOutputStream

A base class for all output streams that work on an underlying stream.

Properties

base-stream [GOutputStream : * : Read / Write / Construct Only]

The underlying base stream on which the io ops will be done.

close-base-stream [logical : Read / Write / Construct Only]

If the base stream should be closed when the filter stream is closed. Default value: TRUE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GFilterOutputStream.html

[Package RGtk2 version 2.20.31 Index]