GIOModule {RGtk2}	R Documentation

GIOModule

Description

Loadable GIO Modules

Methods and Functions

gIOModuleNew(filename)
gIOModulesLoadAllInDirectory(dirname)

Hierarchy

GObject
   +----GTypeModule
         +----GIOModule

Interfaces

GIOModule implements GTypePlugin.

Detailed Description

Provides an interface and default functions for loading and unloading modules. This is used internally to make GIO extensible, but can also be used by others to implement module loading.

Structures

GIOModule

Opaque module base class for extending GIO.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GIOModule.html

[Package RGtk2 version 2.20.31 Index]