GIcon {RGtk2}	R Documentation

GIcon

Description

Interface for icons

Methods and Functions

gIconHash(icon)
gIconEqual(object, icon2)
gIconToString(object)
gIconNewForString(str, .errwarn = TRUE)

Hierarchy

GInterface
   +----GIcon

Implementations

GIcon is implemented by GEmblem, GEmblemedIcon, GFileIcon and GThemedIcon.

Interface Derivations

GIcon is required by GLoadableIcon.

Detailed Description

GIcon is a very minimal interface for icons. It provides functions for checking the equality of two icons, hashing of icons and serializing an icon to and from strings. GIcon does not provide the actual pixmap for the icon as this is out of GIO's scope, however implementations of GIcon may contain the name of an icon (see GThemedIcon), or the path to an icon (see GLoadableIcon).

To obtain a hash of a GIcon, see gIconHash.

To check if two GIcons are equal, see gIconEqual.

For serializing a GIcon, use gIconToString and gIconNewForString.

If your application or library provides one or more GIcon implementations you need to ensure that each GType is registered with the type system prior to calling gIconNewForString.

Structures

GIcon

An abstract type that specifies an icon.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GIcon.html

[Package RGtk2 version 2.20.31 Index]