GInetAddress {RGtk2}	R Documentation

GInetAddress

Description

An IPv4/IPv6 address

Methods and Functions

gInetAddressNewFromString(string)
gInetAddressNewFromBytes(bytes, family)
gInetAddressNewAny(family)
gInetAddressNewLoopback(family)
gInetAddressToBytes(object)
gInetAddressGetNativeSize(object)
gInetAddressToString(object)
gInetAddressGetFamily(object)
gInetAddressGetIsAny(object)
gInetAddressGetIsLoopback(object)
gInetAddressGetIsLinkLocal(object)
gInetAddressGetIsSiteLocal(object)
gInetAddressGetIsMulticast(object)
gInetAddressGetIsMcLinkLocal(object)
gInetAddressGetIsMcNodeLocal(object)
gInetAddressGetIsMcSiteLocal(object)
gInetAddressGetIsMcOrgLocal(object)
gInetAddressGetIsMcGlobal(object)

Hierarchy

GObject
   +----GInetAddress

Detailed Description

GInetAddress represents an IPv4 or IPv6 internet address. Use gResolverLookupByName or gResolverLookupByNameAsync to look up the GInetAddress for a hostname. Use gResolverLookupByAddress or gResolverLookupByAddressAsync to look up the hostname for a GInetAddress.

To actually connect to a remote host, you will need a GInetSocketAddress (which includes a GInetAddress as well as a port number).

Structures

GInetAddress

An IPv4 or IPv6 internet address.

Properties

bytes [R object : Read / Write / Construct Only]

The raw address data.

family [GSocketFamily : Read / Write / Construct Only]

The address family (IPv4 or IPv6). Default value: G_SOCKET_FAMILY_INVALID

is-any [logical : Read]

Whether this is the "any" address for its family. See gInetAddressGetIsAny. Default value: FALSE Since 2.22

is-link-local [logical : Read]

Whether this is a link-local address. See gInetAddressGetIsLinkLocal. Default value: FALSE Since 2.22

is-loopback [logical : Read]

Whether this is the loopback address for its family. See gInetAddressGetIsLoopback. Default value: FALSE Since 2.22

is-mc-global [logical : Read]

Whether this is a global multicast address. See gInetAddressGetIsMcGlobal. Default value: FALSE Since 2.22

is-mc-link-local [logical : Read]

Whether this is a link-local multicast address. See gInetAddressGetIsMcLinkLocal. Default value: FALSE Since 2.22

is-mc-node-local [logical : Read]

Whether this is a node-local multicast address. See gInetAddressGetIsMcNodeLocal. Default value: FALSE Since 2.22

is-mc-org-local [logical : Read]

Whether this is an organization-local multicast address. See gInetAddressGetIsMcOrgLocal. Default value: FALSE Since 2.22

is-mc-site-local [logical : Read]

Whether this is a site-local multicast address. See gInetAddressGetIsMcSiteLocal. Default value: FALSE Since 2.22

is-multicast [logical : Read]

Whether this is a multicast address. See gInetAddressGetIsMulticast. Default value: FALSE Since 2.22

is-site-local [logical : Read]

Whether this is a site-local address. See gInetAddressGetIsLoopback. Default value: FALSE Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GInetAddress.html

[Package RGtk2 version 2.20.31 Index]