GInputStream {RGtk2}	R Documentation

GInputStream

Description

Base class for implementing streaming input

Methods and Functions

gInputStreamRead(object, count, cancellable = NULL, .errwarn = TRUE)
gInputStreamReadAll(object, count, cancellable = NULL, .errwarn = TRUE)
gInputStreamSkip(object, count, cancellable = NULL, .errwarn = TRUE)
gInputStreamClose(object, cancellable = NULL, .errwarn = TRUE)
gInputStreamReadAsync(object, count, io.priority = 0, cancellable = NULL, callback, user.data = NULL)
gInputStreamReadFinish(object, result, .errwarn = TRUE)
gInputStreamSkipAsync(object, count, io.priority = 0, cancellable = NULL, callback, user.data = NULL)
gInputStreamSkipFinish(object, result, .errwarn = TRUE)
gInputStreamCloseAsync(object, io.priority = 0, cancellable = NULL, callback, user.data = NULL)
gInputStreamCloseFinish(object, result, .errwarn = TRUE)
gInputStreamIsClosed(object)
gInputStreamHasPending(object)
gInputStreamSetPending(object, .errwarn = TRUE)
gInputStreamClearPending(object)

Hierarchy

GObject
   +----GInputStream
         +----GFilterInputStream
         +----GFileInputStream
         +----GMemoryInputStream
         +----GUnixInputStream

Detailed Description

GInputStream has functions to read from a stream (gInputStreamRead), to close a stream (gInputStreamClose) and to skip some content (gInputStreamSkip).

To copy the content of an input stream to an output stream without manually handling the reads and writes, use gOutputStreamSplice.

All of these functions have async variants too.

Structures

GInputStream

Base class for streaming input operations.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GInputStream.html

[Package RGtk2 version 2.20.31 Index]