GLibIOEnums {RGtk2}	R Documentation

GLib IO Symbols

Description

Enumerations from GLib used by GIO

Details

GSeekType

An enumeration specifying the base position for a seek operation

cur

The current position in the file.

set

The start of the file.

end

The end of the file.

GIOCondition

A bitwise combination representing a condition to watch for on an event source.

in

There is data to read.

out

Data can be written (unblocked).

pri

There is urgent data to read.

err

Error condition.

hup

Hung up (the connection has been broken, usually for pipes and sockets).

nval

Invalid request. The file descriptor is not open.

GPriority

Convenient constants for indicating the priority of I/O operations.

high

Use this for high priority event sources. It is not used within GLib or GTK+.

default

Use this for default priority event sources. In GLib this priority is used when adding timeout functions with g_timeout_add(). In GDK this priority is used for events from the X server.

high-idle

Use this for high priority idle functions. GTK+ uses G_PRIORITY_HIGH_IDLE + 10 for resizing operations, and G_PRIORITY_HIGH_IDLE + 20 for redrawing operations. (This is done to ensure that any pending resizes are processed before any pending redraws, so that widgets are not redrawn twice unnecessarily.)

default-idle

Use this for default priority idle functions. In GLib this priority is used when adding idle functions with g_idle_add().

low

Use this for very low priority background tasks. It is not used within GLib or GTK+.

Author(s)

Michael Lawrence

[Package RGtk2 version 2.20.31 Index]