GLoadableIcon {RGtk2}	R Documentation

GLoadableIcon

Description

Loadable Icons

Methods and Functions

gLoadableIconLoad(object, size, cancellable = NULL, .errwarn = TRUE)
gLoadableIconLoadAsync(object, size, cancellable = NULL, callback, user.data = NULL)
gLoadableIconLoadFinish(object, res, type, .errwarn = TRUE)

Hierarchy

GInterface
   +----GLoadableIcon

Implementations

GLoadableIcon is implemented by GFileIcon.

Detailed Description

Extends the GIcon interface and adds the ability to load icons from streams.

Structures

GLoadableIcon

Generic type for all kinds of icons that can be loaded as a stream.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GLoadableIcon.html

[Package RGtk2 version 2.20.31 Index]