GMemoryInputStream {RGtk2}	R Documentation

GMemoryInputStream

Description

Streaming input operations on memory chunks

Methods and Functions

gMemoryInputStreamNew()
gMemoryInputStreamNewFromData(data)
gMemoryInputStreamAddData(object, data)
gMemoryInputStream(data)

Hierarchy

GObject
   +----GInputStream
         +----GMemoryInputStream

Interfaces

GMemoryInputStream implements GSeekable.

Detailed Description

GMemoryInputStream is a class for using arbitrary memory chunks as input for GIO streaming input operations.

Structures

GMemoryInputStream

Implements GInputStream for arbitrary memory chunks.

Convenient Construction

gMemoryInputStream is the result of collapsing the constructors of GMemoryInputStream (gMemoryInputStreamNew, gMemoryInputStreamNewFromData) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GMemoryInputStream.html

[Package RGtk2 version 2.20.31 Index]