GMemoryOutputStream {RGtk2}	R Documentation

GMemoryOutputStream

Description

Streaming output operations on memory chunks

Methods and Functions

gMemoryOutputStreamNew(len)
gMemoryOutputStreamGetData(object)
gMemoryOutputStreamGetSize(object)
gMemoryOutputStreamGetDataSize(object)
gMemoryOutputStream(len)

Hierarchy

GObject
   +----GOutputStream
         +----GMemoryOutputStream

Interfaces

GMemoryOutputStream implements GSeekable.

Detailed Description

GMemoryOutputStream is a class for using arbitrary memory chunks as output for GIO streaming output operations.

Structures

GMemoryOutputStream

Implements GOutputStream for arbitrary memory chunks.

Convenient Construction

gMemoryOutputStream is the equivalent of gMemoryOutputStreamNew.

User Functions

GReallocFunc(data, size)

Changes the size of the memory block pointed to by data to size bytes.

The function should have the same semantics as realloc().

data

memory block to reallocate

size

size to reallocate data to

Returns: [R object] a pointer to the reallocated memory

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GMemoryOutputStream.html

[Package RGtk2 version 2.20.31 Index]