GMount {RGtk2}	R Documentation

GMount

Description

Mount management

Methods and Functions

gMountGetName(object)
gMountGetUuid(object)
gMountGetIcon(object)
gMountGetDrive(object)
gMountGetRoot(object)
gMountGetVolume(object)
gMountCanUnmount(object)
gMountUnmount(object, flags = "G_MOUNT_UNMOUNT_NONE", cancellable = NULL, callback, user.data = NULL)
gMountUnmountFinish(object, result, .errwarn = TRUE)
gMountUnmountWithOperation(object, flags, mount.operation, cancellable = NULL, callback, user.data = NULL)
gMountUnmountWithOperationFinish(object, result, .errwarn = TRUE)
gMountRemount(object, flags, mount.operation, cancellable = NULL, callback, user.data = NULL)
gMountRemountFinish(object, result, .errwarn = TRUE)
gMountCanEject(object)
gMountEject(object, flags = "G_MOUNT_UNMOUNT_NONE", cancellable = NULL, callback, user.data = NULL)
gMountEjectFinish(object, result, .errwarn = TRUE)
gMountEjectWithOperation(object, flags, mount.operation, cancellable = NULL, callback, user.data = NULL)
gMountEjectWithOperationFinish(object, result, .errwarn = TRUE)
gMountGuessContentType(object, force.rescan, cancellable = NULL, callback, user.data = NULL)
gMountGuessContentTypeFinish(object, result, .errwarn = TRUE)
gMountGuessContentTypeSync(object, force.rescan, cancellable = NULL, .errwarn = TRUE)
gMountIsShadowed(object)
gMountShadow(object)
gMountUnshadow(object)

Hierarchy

GInterface
   +----GMount
GEnum
   +----GMountMountFlags
GFlags
   +----GMountUnmountFlags

Detailed Description

The GMount interface represents user-visible mounts. Note, when porting from GnomeVFS, GMount is the moral equivalent of GnomeVFSVolume. GMount is a "mounted" filesystem that you can access. Mounted is in quotes because it's not the same as a unix mount, it might be a gvfs mount, but you can still access the files on it if you use GIO. Might or might not be related to a volume object.

Unmounting a GMount instance is an asynchronous operation. For more information about asynchronous operations, see GAsyncReady and GSimpleAsyncReady. To unmount a GMount instance, first call gMountUnmountWithOperation with (at least) the GMount instance and a GAsyncReadyCallback. The callback will be fired when the operation has resolved (either with success or failure), and a GAsyncReady structure will be passed to the callback. That callback should then call gMountUnmountWithOperationFinish with the GMount and the GAsyncReady data to see if the operation was completed successfully. If an error is present when gMountUnmountWithOperationFinish is called, then it will be filled with any error information.

Structures

GMount

A handle to an object implementing the GMountIface interface.

Enums and Flags

GMountMountFlags

Flags used when mounting a mount.

none

No flags set.

GMountUnmountFlags

Flags used when an unmounting a mount.

none

No flags set.

force

Unmount even if there are outstanding file operations on the mount.

Signals

changed(mount, user.data)

Emitted when the mount has been changed.

mount

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

pre-unmount(mount, user.data)

This signal is emitted when the GMount is about to be unmounted. Since 2.22

mount

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

unmounted(mount, user.data)

This signal is emitted when the GMount have been unmounted. If the recipient is holding references to the object they should release them so the object can be finalized.

mount

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GMount.html

[Package RGtk2 version 2.20.31 Index]