GNetworkAddress {RGtk2}	R Documentation

GNetworkAddress

Description

A GSocketConnectable for resolving hostnames

Methods and Functions

gNetworkAddressNew(hostname, port)
gNetworkAddressGetHostname(object)
gNetworkAddressGetPort(object)
gNetworkAddressParse(host.and.port, default.port, .errwarn = TRUE)
gNetworkAddress(hostname, port)

Hierarchy

GObject
   +----GNetworkAddress

Interfaces

GNetworkAddress implements GSocketConnectable.

Detailed Description

GNetworkAddress provides an easy way to resolve a hostname and then attempt to connect to that host, handling the possibility of multiple IP addresses and multiple address families.

See GSocketConnectable for and example of using the connectable interface.

Structures

GNetworkAddress

A GSocketConnectable for resolving a hostname and connecting to that host.

Convenient Construction

gNetworkAddress is the equivalent of gNetworkAddressNew.

Properties

hostname [character : * : Read / Write / Construct Only]

Hostname to resolve. Default value: NULL

port [numeric : Read / Write / Construct Only]

Network port. Allowed values: <= 65535 Default value: 0

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GNetworkAddress.html

[Package RGtk2 version 2.20.31 Index]