GNetworkService {RGtk2}	R Documentation

GNetworkService

Description

A GSocketConnectable for resolving SRV records

Methods and Functions

gNetworkServiceNew(service, protocol, domain)
gNetworkServiceGetService(object)
gNetworkServiceGetProtocol(object)
gNetworkServiceGetDomain(object)
gNetworkService(service, protocol, domain)

Hierarchy

GObject
   +----GNetworkService

Interfaces

GNetworkService implements GSocketConnectable.

Detailed Description

Like GNetworkAddress does with hostnames, GNetworkService provides an easy way to resolve a SRV record, and then attempt to connect to one of the hosts that implements that service, handling service priority/weighting, multiple IP addresses, and multiple address families.

See GSrvTarget for more information about SRV records, and see GSocketConnectable for and example of using the connectable interface.

Structures

GNetworkService

A GSocketConnectable for resolving a SRV record and connecting to that service.

Convenient Construction

gNetworkService is the equivalent of gNetworkServiceNew.

Properties

domain [character : * : Read / Write / Construct Only]

Network domain, eg, "example.com". Default value: NULL

protocol [character : * : Read / Write / Construct Only]

Network protocol, eg "tcp". Default value: NULL

service [character : * : Read / Write / Construct Only]

Service name, eg "ldap". Default value: NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GNetworkService.html

[Package RGtk2 version 2.20.31 Index]