GOutputStream {RGtk2}	R Documentation

GOutputStream

Description

Base class for implementing streaming output

Methods and Functions

gOutputStreamWrite(object, buffer, cancellable = NULL, .errwarn = TRUE)
gOutputStreamWriteAll(object, buffer, bytes.written, cancellable = NULL, .errwarn = TRUE)
gOutputStreamSplice(object, source, flags = "G_OUTPUT_STREAM_SPLICE_NONE", cancellable = NULL, .errwarn = TRUE)
gOutputStreamFlush(object, cancellable = NULL, .errwarn = TRUE)
gOutputStreamClose(object, cancellable = NULL, .errwarn = TRUE)
gOutputStreamWriteAsync(object, buffer, io.priority = 0, cancellable = NULL, callback, user.data = NULL)
gOutputStreamWriteFinish(object, result, .errwarn = TRUE)
gOutputStreamSpliceAsync(object, source, flags = "G_OUTPUT_STREAM_SPLICE_NONE", io.priority = 0, cancellable = NULL, callback, user.data = NULL)
gOutputStreamSpliceFinish(object, result, .errwarn = TRUE)
gOutputStreamFlushAsync(object, io.priority = 0, cancellable = NULL, callback, user.data = NULL)
gOutputStreamFlushFinish(object, result, .errwarn = TRUE)
gOutputStreamCloseAsync(object, io.priority = 0, cancellable = NULL, callback, user.data = NULL)
gOutputStreamCloseFinish(object, result, .errwarn = TRUE)
gOutputStreamIsClosed(object)
gOutputStreamHasPending(object)
gOutputStreamSetPending(object, .errwarn = TRUE)
gOutputStreamClearPending(object)

Hierarchy

GFlags
   +----GOutputStreamSpliceFlags
GObject
   +----GOutputStream
         +----GFilterOutputStream
         +----GFileOutputStream
         +----GMemoryOutputStream
         +----GUnixOutputStream

Detailed Description

GOutputStream has functions to write to a stream (gOutputStreamWrite), to close a stream (gOutputStreamClose) and to flush pending writes (gOutputStreamFlush).

To copy the content of an input stream to an output stream without manually handling the reads and writes, use gOutputStreamSplice.

All of these functions have async variants too.

Structures

GOutputStream

Base class for writing output.

All classes derived from GOutputStream should implement synchronous writing, splicing, flushing and closing streams, but may implement asynchronous versions.

Enums and Flags

GOutputStreamSpliceFlags

GOutputStreamSpliceFlags determine how streams should be spliced.

none

Do not close either stream.

close-source

Close the source stream after the splice.

close-target

Close the target stream after the splice.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GOutputStream.html

[Package RGtk2 version 2.20.31 Index]