GParamSpec {RGtk2}	R Documentation

Parameter metadata in GObject

Description

GParamSpec is an object structure that encapsulates the metadata required to specify parameters, such as e.g. GObject properties.

Usage

gParamSpec(type, name, nick = NULL, blurb = NULL, flags = NULL, ...)
as.GParamSpec(x)

Arguments

type
a string naming the R type (ie 'character', 'numeric', ...) corresponding to the desired GParamSpec (property) type

name
the name of the GParamSpec (property)

nick
the nickname of the GParamSpec (property)

blurb
a short string description of the GParamSpec (property)

flags
a vector of values from the GParamFlags enumeration - please see the GObject documentation for more information

...
named arguments specific to the type of GParamSpec - please see the GObject documentation for more information

x
A list corresponding to a GParamSpec - the class of the list should be the name of the GParamSpec type

Details

As a transparent type, the various GParamSpec structures should be returned to R as corresponding lists, and as.GParamSpec coerces a list to one understandable by the C wrappers, assuming that the class of the list is the name of the GParamSpec type. gParamSpec is a more user-friendly wrapper to as.GParamSpec that constructs the correctly-classed list on the fly from its arguments.

Value

A list representing a GParamSpec, ready to be passed to the underlying C libraries

Author(s)

Michael Lawrence

References

http://developer.gnome.org/doc/API/2.0/gobject/gobject-GParamSpec.html

[Package RGtk2 version 2.20.31 Index]