GResolver {RGtk2}	R Documentation

GResolver

Description

Asynchronous and cancellable DNS resolver

Methods and Functions

gResolverGetDefault()
gResolverSetDefault(object)
gResolverLookupByName(object, hostname, cancellable = NULL, .errwarn = TRUE)
gResolverLookupByNameAsync(object, hostname, cancellable = NULL, callback, user.data = NULL)
gResolverLookupByNameFinish(object, result, .errwarn = TRUE)
gResolverFreeAddresses(addresses)
gResolverLookupByAddress(object, address, cancellable = NULL, .errwarn = TRUE)
gResolverLookupByAddressAsync(object, address, cancellable = NULL, callback, user.data = NULL)
gResolverLookupByAddressFinish(object, result, .errwarn = TRUE)
gResolverLookupService(object, service, protocol, domain, cancellable = NULL, .errwarn = TRUE)
gResolverLookupServiceAsync(object, service, protocol, domain, cancellable = NULL, callback, user.data = NULL)
gResolverLookupServiceFinish(object, result, .errwarn = TRUE)
gResolverFreeTargets(targets)

Hierarchy

GObject
   +----GResolver
GEnum
   +----GResolverError

Detailed Description

GResolver provides cancellable synchronous and asynchronous DNS resolution, for hostnames (gResolverLookupByAddress, gResolverLookupByName and their async variants) and SRV (service) records (gResolverLookupService). GNetworkAddress and GNetworkService provide wrappers around GResolver functionality that also implement GSocketConnectable, making it easy to connect to a remote host/service.

Structures

GResolver

The object that handles DNS resolution. Use gResolverGetDefault to get the default resolver.

Signals

reload(resolver, user.data)

Emitted when the resolver notices that the system resolver configuration has changed.

resolver

a GResolver

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GResolver.html

[Package RGtk2 version 2.20.31 Index]