GSocketAddress {RGtk2}	R Documentation

GSocketAddress

Description

Abstract base class representing endpoints for socket communication

Methods and Functions

gSocketAddressNewFromNative(native, len)
gSocketAddressGetFamily(object)
gSocketAddressToNative(object, dest, destlen, .errwarn = TRUE)
gSocketAddressGetNativeSize(object)

Hierarchy

GObject
   +----GSocketAddress
         +----GInetSocketAddress
         +----GUnixSocketAddress
GEnum
   +----GSocketFamily

Interfaces

GSocketAddress implements GSocketConnectable.

Detailed Description

GSocketAddress is the equivalent of structsockaddr in the BSD sockets API. This is an abstract class; use GInetSocketAddress for internet sockets, or GUnixSocketAddress for UNIX domain sockets.

Structures

GSocketAddress

A socket endpoint address, corresponding to structsockaddr or one of its subtypes.

Enums and Flags

GSocketFamily

The protocol family of a GSocketAddress. (These values are identical to the system defines AF_INET, AF_INET6 and AF_UNIX, if available.) Since 2.22

invalid

no address family

unix

the UNIX domain family

ipv4

the IPv4 family

ipv6

the IPv6 family

Properties

family [GSocketFamily : Read]

The family of the socket address. Default value: G_SOCKET_FAMILY_INVALID

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GSocketAddress.html

[Package RGtk2 version 2.20.31 Index]