GSocketClient {RGtk2}	R Documentation

GSocketClient

Description

Helper for connecting to a network service

Methods and Functions

gSocketClientNew()
gSocketClientConnect(object, connectable, cancellable = NULL, .errwarn = TRUE)
gSocketClientConnectAsync(object, connectable, cancellable = NULL, callback, user.data = NULL)
gSocketClientConnectFinish(object, result, .errwarn = TRUE)
gSocketClientConnectToHost(object, host.and.port, default.port, cancellable = NULL, .errwarn = TRUE)
gSocketClientConnectToHostAsync(object, host.and.port, default.port, cancellable = NULL, callback, user.data = NULL)
gSocketClientConnectToHostFinish(object, result, .errwarn = TRUE)
gSocketClientConnectToService(object, domain, service, cancellable = NULL, .errwarn = TRUE)
gSocketClientConnectToServiceAsync(object, domain, service, cancellable = NULL, callback, user.data = NULL)
gSocketClientConnectToServiceFinish(object, result, .errwarn = TRUE)
gSocketClientSetFamily(object, family)
gSocketClientSetLocalAddress(object, address)
gSocketClientSetProtocol(object, protocol)
gSocketClientSetSocketType(object, type)
gSocketClientGetFamily(object)
gSocketClientGetLocalAddress(object)
gSocketClientGetProtocol(object)
gSocketClientGetSocketType(object)
gSocketClient()

Hierarchy

GObject
   +----GSocketClient

Detailed Description

GSocketClient is a high-level utility class for connecting to a network host using a connection oriented socket type.

You create a GSocketClient object, set any options you want, then call a sync or async connect operation, which returns a GSocketConnection subclass on success.

The type of the GSocketConnection object returned depends on the type of the underlying socket that is in use. For instance, for a TCP/IP connection it will be a GTcpConnection.

Structures

GSocketClient

A helper class for network servers to listen for and accept connections. Since 2.22

Convenient Construction

gSocketClient is the equivalent of gSocketClientNew.

Properties

family [GSocketFamily : Read / Write / Construct]

The sockets address family to use for socket construction. Default value: G_SOCKET_FAMILY_INVALID

local-address [GSocketAddress : * : Read / Write / Construct]

The local address constructed sockets will be bound to.

protocol [GSocketProtocol : Read / Write / Construct]

The protocol to use for socket construction, or 0 for default. Default value: G_SOCKET_PROTOCOL_DEFAULT

type [GSocketType : Read / Write / Construct]

The sockets type to use for socket construction. Default value: G_SOCKET_TYPE_STREAM

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GSocketClient.html

[Package RGtk2 version 2.20.31 Index]