GSocketConnection {RGtk2}	R Documentation

GSocketConnection

Description

A socket connection

Methods and Functions

gSocketConnectionGetLocalAddress(object, .errwarn = TRUE)
gSocketConnectionGetRemoteAddress(object, .errwarn = TRUE)
gSocketConnectionGetSocket(object)
gTcpConnectionSetGracefulDisconnect(object, graceful.disconnect)
gTcpConnectionGetGracefulDisconnect(object)
gSocketConnectionFactoryCreateConnection(object)
gSocketConnectionFactoryLookupType(family, type, protocol.id)
gSocketConnectionFactoryRegisterType(g.type, family, type, protocol)

Hierarchy

GObject
   +----GIOStream
         +----GSocketConnection
               +----GTcpConnection
               +----GUnixConnection
GObject
   +----GIOStream
         +----GSocketConnection
               +----GTcpConnection
GObject
   +----GIOStream
         +----GSocketConnection
               +----GUnixConnection

Detailed Description

GSocketConnection is a GIOStream for a connected socket. They can be created either by GSocketClient when connecting to a host, or by GSocketListener when accepting a new client.

The type of the GSocketConnection object returned from these calls depends on the type of the underlying socket that is in use. For instance, for a TCP/IP connection it will be a GTcpConnection.

Chosing what type of object to construct is done with the socket connection factory, and it is possible for 3rd parties to register custom socket connection types for specific combination of socket family/type/protocol using gSocketConnectionFactoryRegisterType.

Structures

GSocketConnection

A socket connection GIOStream object for connection-oriented sockets. Since 2.22

GTcpConnection

A GSocketConnection for UNIX domain socket connections. Since 2.22

Properties

socket [GSocket : * : Read / Write / Construct Only]

The underlying GSocket.

graceful-disconnect [logical : Read / Write]

Whether or not close does a graceful disconnect. Default value: FALSE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GSocketConnection.html

[Package RGtk2 version 2.20.31 Index]