GSocketControlMessage {RGtk2}	R Documentation

GSocketControlMessage

Description

A GSocket control message

Methods and Functions

gSocketControlMessageDeserialize(level, type, size, data)
gSocketControlMessageGetLevel(object)
gSocketControlMessageGetMsgType(object)
gSocketControlMessageGetSize(object)
gSocketControlMessageSerialize(object, data)

Hierarchy

GObject
   +----GSocketControlMessage
         +----GUnixFDMessage

Detailed Description

A GSocketControlMessage is a special-purpose utility message that can be sent to or received from a GSocket. These types of messages are often called "ancillary data".

The message can represent some sort of special instruction to or information from the socket or can represent a special kind of transfer to the peer (for example, sending a file description over a UNIX socket).

These messages are sent with gSocketSendMessage and received with gSocketReceiveMessage.

To extend the set of control message that can be sent, subclass this class and override the get_size, get_level, get_type and serialize methods.

To extend the set of control messages that can be received, subclass this class and implement the deserialize method. Also, make sure your class is registered with the GType typesystem before calling gSocketReceiveMessage to read such a message.

Structures

GSocketControlMessage

Base class for socket-type specific control messages that can be sent and received over GSocket.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GSocketControlMessage.html

[Package RGtk2 version 2.20.31 Index]