GSocketListener {RGtk2}	R Documentation

GSocketListener

Description

Helper for accepting network client connections

Methods and Functions

gSocketListenerNew()
gSocketListenerAddSocket(object, socket, source.object, .errwarn = TRUE)
gSocketListenerAddAddress(object, address, type, protocol, source.object, .errwarn = TRUE)
gSocketListenerAddInetPort(object, port, source.object, .errwarn = TRUE)
gSocketListenerAccept(object, cancellable = NULL, .errwarn = TRUE)
gSocketListenerAcceptAsync(object, cancellable = NULL, callback, user.data = NULL)
gSocketListenerAcceptFinish(object, result, .errwarn = TRUE)
gSocketListenerAcceptSocket(object, cancellable = NULL, .errwarn = TRUE)
gSocketListenerAcceptSocketAsync(object, cancellable = NULL, callback, user.data = NULL)
gSocketListenerAcceptSocketFinish(object, result, .errwarn = TRUE)
gSocketListenerClose(object)
gSocketListenerSetBacklog(object, listen.backlog)
gSocketListener()

Hierarchy

GObject
   +----GSocketListener
         +----GSocketService

Detailed Description

A GSocketListener is an object that keeps track of a set of server sockets and helps you accept sockets from any of the socket, either sync or async.

If you want to implement a network server, also look at GSocketService and GThreadedSocketService which are subclass of GSocketListener that makes this even easier.

Structures

GSocketListener

undocumented

Convenient Construction

gSocketListener is the equivalent of gSocketListenerNew.

Properties

listen-backlog [integer : Read / Write / Construct]

outstanding connections in the listen queue. Allowed values: [0,2000] Default value: 10

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GSocketListener.html

[Package RGtk2 version 2.20.31 Index]