GSrvTarget {RGtk2}	R Documentation

GSrvTarget

Description

DNS SRV record target

Methods and Functions

gSrvTargetNew(hostname, port, priority, weight)
gSrvTargetCopy(object)
gSrvTargetFree(object)
gSrvTargetGetHostname(object)
gSrvTargetGetPort(object)
gSrvTargetGetPriority(object)
gSrvTargetGetWeight(object)
gSrvTargetListSort(targets)

Hierarchy

GBoxed
   +----GSrvTarget

Detailed Description

SRV (service) records are used by some network protocols to provide service-specific aliasing and load-balancing. For example, XMPP (Jabber) uses SRV records to locate the XMPP server for a domain; rather than connecting directly to "example.com" or assuming a specific server hostname like "xmpp.example.com", an XMPP client would look up the "xmpp-client" SRV record for "example.com", and then connect to whatever host was pointed to by that record.

You can use gResolverLookupService or gResolverLookupServiceAsync to find the GSrvTargets for a given service. However, if you are simply planning to connect to the remote service, you can use GNetworkService's GSocketConnectable interface and not need to worry about GSrvTarget at all.

Structures

GSrvTarget

A single target host/port that a network service is running on.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GSrvTarget.html

[Package RGtk2 version 2.20.31 Index]