GTK {RGtk2}	R Documentation

GTK

Description

The GTK+ library itself contains widgets, that is, GUI components such as GtkButton or GtkTextView.

Details

The RGtk binding to the GTK library consists of the following components:

chap-drawing-model

The GTK+ drawing model in detail

gtk-Filesystem-utilities

Functions for working with GIO

GtkAboutDialog

Display information about an application

gtk-Keyboard-Accelerators

Groups of global keyboard accelerators for an entire GtkWindow

GtkAccelLabel

A label which displays an accelerator key on the right of the text

gtk-Accelerator-Maps

Loadable keyboard accelerator specifications

GtkAccessible

Accessibility support for widgets

GtkAction

An action which can be triggered by a menu or toolbar item

GtkActionGroup

A group of actions

GtkActivatable

An interface for activatable widgets

GtkAdjustment

A GtkObject representing an adjustable bounded value

GtkAlignment

A widget which controls the alignment and size of its child

GtkArrow

Displays an arrow

GtkAspectFrame

A frame that constrains its child to a particular aspect ratio

GtkAssistant

A widget used to guide users through multi-step operations

GtkButtonBox

Base class for GtkHButtonBox and GtkVButtonBox

GtkBin

A container with just one child

GtkBox

Base class for box containers

gtk-gtkbuildable

Interface for objects that can be built by GtkBuilder

GtkBuilder

Build an interface from an XML UI definition

GtkButton

A widget that creates a signal when clicked on

GtkCalendar

Displays a calendar and allows the user to select a date

GtkCellEditable

Interface for widgets which can are used for editing cells

GtkCellLayout

An interface for packing cells

GtkCellRenderer

An object for rendering a single cell on a GdkDrawable

GtkCellRendererAccel

Renders a keyboard accelerator in a cell

GtkCellRendererCombo

Renders a combobox in a cell

GtkCellRendererPixbuf

Renders a pixbuf in a cell

GtkCellRendererProgress

Renders numbers as progress bars

GtkCellRendererSpin

Renders a spin button in a cell

GtkCellRendererSpinner

Renders a spinning animation in a cell

GtkCellRendererText

Renders text in a cell

GtkCellRendererToggle

Renders a toggle button in a cell

GtkCellView

A widget displaying a single row of a GtkTreeModel

GtkCheckButton

Create widgets with a discrete toggle button

gtk-gtkcheckmenuitem

A menu item with a check box

gtk-Clipboards

Storing data on clipboards

GtkCList

A multi-columned scrolling list widget

GtkColorButton

A button to launch a color selection dialog

GtkColorSelection

A widget used to select a color

GtkColorSelectionDialog

A standard dialog box for selecting a color

GtkCombo

A text entry field with a dropdown list

GtkComboBox

A widget used to choose from a list of items

GtkComboBoxEntry

A text entry field with a dropdown list

GtkContainer

Base class for widgets which contain other widgets

GtkCTree

A widget displaying a hierarchical tree

GtkCurve

Allows direct editing of a curve

GtkDialog

Create popup windows

gtk-Drag-and-Drop

Functions for controlling drag and drop handling

GtkDrawingArea

A widget for custom user interface elements

GtkEditable

Interface for text-editing widgets

GtkEntry

A single line text entry field

GtkEntryBuffer

Text buffer for GtkEntry

GtkEntryCompletion

Completion functionality for GtkEntry

gtk-Standard-Enumerations

Public enumerated types used throughout GTK+

GtkEventBox

A widget used to catch events for widgets which do not have their own window

GtkExpander

A container which can hide its child

GtkFileChooser

File chooser interface used by GtkFileChooserWidget and GtkFileChooserDialog

GtkFileChooserButton

A button to launch a file selection dialog

GtkFileChooserDialog

A file chooser dialog, suitable for "File/Open" or "File/Save" commands

GtkFileChooserWidget

File chooser widget that can be embedded in other widgets

gtk-gtkfilefilter

A filter for selecting a file subset

GtkFileSelection

Prompt the user for a file or directory name

GtkFixed

A container which allows you to position widgets at fixed coordinates

GtkFontButton

A button to launch a font selection dialog

GtkFontSelection

A widget for selecting fonts

GtkFontSelectionDialog

A dialog box for selecting fonts

GtkFrame

A bin with a decorative frame and optional label

GtkGammaCurve

A subclass of GtkCurve for editing gamma curves

gtk-Graphics-Contexts

A shared pool of GdkGC objects

GtkHandleBox

a widget for detachable window portions

GtkHButtonBox

A container for arranging buttons horizontally

GtkHBox

A horizontal container box

GtkHPaned

A container with two panes arranged horizontally

GtkHRuler

A horizontal ruler

GtkHScale

A horizontal slider widget for selecting a value from a range

GtkHScrollbar

A horizontal scrollbar

GtkHSeparator

A horizontal separator

GtkHSV

A 'color wheel' widget

gtk-Themeable-Stock-Images

Manipulating stock icons

GtkIconTheme

Looking up icons by name

GtkIconView

A widget which displays a list of icons in a grid

GtkImage

A widget displaying an image

GtkImageMenuItem

A menu item with an icon

GtkIMContext

Base class for input method contexts

GtkIMContextSimple

An input method context supporting table-based input methods

GtkIMMulticontext

An input method context supporting multiple, loadable input methods

GtkInfoBar

Report important messages to the user

GtkInputDialog

Configure devices for the XInput extension

GtkInvisible

A widget which is not displayed

GtkItem

Abstract base class for GtkMenuItem, GtkListItem and GtkTreeItem

GtkItemFactory

A factory for menus

GtkLabel

A widget that displays a small to medium amount of text

GtkLayout

Infinite scrollable area containing child widgets and/or custom drawing

GtkLinkButton

Create buttons bound to a URL

GtkList

Widget for packing a list of selectable items

GtkListItem

An item in a GtkList

GtkListStore

A list-like data structure that can be used with the GtkTreeView

gtk-General

Library initialization, main event loop, and events

GtkMenu

A menu widget

GtkMenuBar

A subclass widget for GtkMenuShell which holds GtkMenuItem widgets

GtkMenuItem

The widget used for item in menus

GtkMenuShell

A base class for menu objects

GtkMenuToolButton

A GtkToolItem containing a button with an additional dropdown menu

GtkMessageDialog

A convenient message window

GtkMisc

Base class for widgets with alignments and padding

GtkNotebook

A tabbed notebook container

GtkOffscreenWindow

A toplevel container widget used to manage offscreen rendering of child widgets.

GtkOldEditable

Base class for text-editing widgets

GtkOptionMenu

A widget used to choose from a list of valid choices

gtk-Orientable

An interface for flippable widgets

GtkPageSetup

Stores page setup information

GtkPaned

Base class for widgets with two adjustable panes

GtkPaperSize

Support for named paper sizes

GtkPixmap

A widget displaying a graphical image or icon

GtkPlug

Toplevel for embedding into other processes

GtkPreview

A widget to display RGB or grayscale data

GtkPrintContext

Encapsulates context for drawing pages

gtk-High-level-Printing-API

High-level Printing API

GtkPrintSettings

Stores print settings

GtkProgress

Base class for GtkProgressBar

GtkProgressBar

A widget which indicates progress visually

GtkRadioAction

An action of which only one in a group can be active

GtkRadioButton

A choice from multiple check buttons

GtkRadioMenuItem

A choice from multiple check menu items

GtkRadioToolButton

A toolbar item that contains a radio button

GtkRange

Base class for widgets which visualize an adjustment

gtk-Resource-Files

Routines for handling resource files

GtkRecentAction

An action of which represents a list of recently used files

GtkRecentChooser

Interface implemented by widgets displaying recently used files

GtkRecentChooserDialog

Displays recently used files in a dialog

GtkRecentChooserMenu

Displays recently used files in a menu

GtkRecentChooserWidget

Displays recently used files

GtkRecentFilter

A filter for selecting a subset of recently used files

GtkRecentManager

Managing Recently Used Files

GtkRuler

Base class for horizontal or vertical rulers

GtkScale

Base class for GtkHScale and GtkVScale

GtkScaleButton

A button which pops up a scale

GtkScrollbar

Base class for GtkHScrollbar and GtkVScrollbar

GtkScrolledWindow

Adds scrollbars to its child widget

gtk-Selections

Functions for handling inter-process communication via selections

GtkSeparator

Base class for GtkHSeparator and GtkVSeparator

GtkSeparatorMenuItem

A separator used in menus

GtkSeparatorToolItem

A toolbar item that separates groups of other toolbar items

GtkSettings

Sharing settings between applications

GtkSizeGroup

Grouping widgets so they request the same size

GtkSocket

Container for widgets from other processes

GtkSpinButton

Retrieve an integer or floating-point number from the user

GtkSpinner

Show a spinner animation

GtkStatusbar

Report messages of minor importance to the user

GtkStatusIcon

Display an icon in the system tray

gtk-Stock-Items

Prebuilt common menu/toolbar items and corresponding icons

GtkStyle

Functions for drawing widget parts

GtkTable

Pack widgets in regular patterns

GtkTearoffMenuItem

A menu item used to tear off and reattach its menu

gtk-Testing

Utilities for testing GTK+ applications

GtkTextBuffer

Stores attributed text for display in a GtkTextView

GtkTextIter

Text buffer iterator

GtkTextMark

A position in the buffer preserved across buffer modifications

GtkTextTag

A tag that can be applied to text in a GtkTextBuffer

GtkTextTagTable

Collection of tags that can be used together

GtkTextView

Widget that displays a GtkTextBuffer

GtkTipsQuery

Displays help about widgets in the user interface

GtkToggleAction

An action which can be toggled between two states

GtkToggleButton

Create buttons which retain their state

GtkToggleToolButton

A GtkToolItem containing a toggle button

GtkToolbar

Create bars of buttons and other widgets

GtkToolButton

A GtkToolItem subclass that displays buttons

GtkToolItem

The base class of widgets that can be added to GtkToolShell

GtkToolItemGroup

A sub container used in a tool palette

GtkToolPalette

A tool palette with categories

GtkToolShell

Interface for containers containing GtkToolItem widgets

GtkTooltip

Add tips to your widgets

GtkTooltips

Add tips to your widgets

gtk-GtkTreeView-drag-and-drop

Interfaces for drag-and-drop support in GtkTreeView

GtkTreeModel

The tree interface used by GtkTreeView

GtkTreeModelFilter

A GtkTreeModel which hides parts of an underlying tree model

GtkTreeModelSort

A GtkTreeModel which makes an underlying tree model sortable

GtkTreeSelection

The selection object for GtkTreeView

GtkTreeSortable

The interface for sortable models used by GtkTreeView

GtkTreeStore

A tree-like data structure that can be used with the GtkTreeView

GtkTreeView

A widget for displaying both trees and lists

GtkTreeViewColumn

A visible column in a GtkTreeView widget

GtkUIManager

Constructing menus and toolbars from an XML description

GtkVButtonBox

A container for arranging buttons vertically

GtkVBox

A vertical container box

GtkViewport

An adapter which makes widgets scrollable

GtkVolumeButton

A button which pops up a volume control

GtkVPaned

A container with two panes arranged vertically

GtkVRuler

A vertical ruler

GtkVScale

A vertical slider widget for selecting a value from a range

GtkVScrollbar

A vertical scrollbar

GtkVSeparator

A vertical separator

GtkWidget

Base class for all widgets

GtkWindow

Toplevel which can contain other widgets

GtkWindowGroup

Limit the effect of grabs

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk

[Package RGtk2 version 2.20.31 Index]