GThemedIcon {RGtk2}	R Documentation

GThemedIcon

Description

Icon theming support

Methods and Functions

gThemedIconNew(iconname = NULL)
gThemedIconNewFromNames(iconnames, len)
gThemedIconNewWithDefaultFallbacks(iconname)
gThemedIconPrependName(object, iconname)
gThemedIconAppendName(object, iconname)
gThemedIconGetNames(object)
gThemedIcon(iconname, iconnames, len)

Hierarchy

GObject
   +----GThemedIcon

Interfaces

GThemedIcon implements GIcon.

Detailed Description

GThemedIcon is an implementation of GIcon that supports icon themes. GThemedIcon contains a list of all of the icons present in an icon theme, so that icons can be looked up quickly. GThemedIcon does not provide actual pixmaps for icons, just the icon names. Ideally something like gtkIconThemeChooseIcon should be used to resolve the list of names so that fallback icons work nicely with themes that inherit other themes.

Structures

GThemedIcon

An implementation of GIcon for themed icons.

Convenient Construction

gThemedIcon is the result of collapsing the constructors of GThemedIcon (gThemedIconNew, gThemedIconNewWithDefaultFallbacks, gThemedIconNewFromNames) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Properties

name [character : * : Write / Construct Only]

The icon name. Default value: NULL

names [character list : Read / Write / Construct Only]

A list of icon names.

use-default-fallbacks [logical : Read / Write / Construct Only]

Whether to use the default fallbacks found by shortening the icon name at '-' characters. If the "names" list has more than one element, ignores any past the first.

For example, if the icon name was "gnome-dev-cdrom-audio", the list would become

c("gnome-dev-cdrom-audio", "gnome-dev-cdrom", "gnome-dev", "gnome")

Default value: FALSE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GThemedIcon.html

[Package RGtk2 version 2.20.31 Index]