GVolumeMonitor {RGtk2}	R Documentation

GVolumeMonitor

Description

Volume Monitor

Methods and Functions

gVolumeMonitorGet()
gVolumeMonitorGetConnectedDrives(object)
gVolumeMonitorGetVolumes(object)
gVolumeMonitorGetMounts(object)
gVolumeMonitorAdoptOrphanMount(mount)
gVolumeMonitorGetMountForUuid(object, uuid)
gVolumeMonitorGetVolumeForUuid(object, uuid)

Hierarchy

GObject
   +----GVolumeMonitor

Detailed Description

GVolumeMonitor is for listing the user interesting devices and volumes on the computer. In other words, what a file selector or file manager would show in a sidebar. GVolumeMonitor is not thread-default-context aware, and so should not be used other than from the main thread, with no thread-default-context active.

Structures

GVolumeMonitor

A Volume Monitor that watches for volume events.

Signals

drive-changed(volume.monitor, drive, user.data)

Emitted when a drive changes.

volume.monitor

The volume monitor emitting the signal.

drive

the drive that changed

user.data

user data set when the signal handler was connected.

drive-connected(volume.monitor, drive, user.data)

Emitted when a drive is connected to the system.

volume.monitor

The volume monitor emitting the signal.

drive

a GDrive that was connected.

user.data

user data set when the signal handler was connected.

drive-disconnected(volume.monitor, drive, user.data)

Emitted when a drive is disconnected from the system.

volume.monitor

The volume monitor emitting the signal.

drive

a GDrive that was disconnected.

user.data

user data set when the signal handler was connected.

drive-eject-button(volume.monitor, drive, user.data)

Emitted when the eject button is pressed on drive. Since 2.18

volume.monitor

The volume monitor emitting the signal.

drive

the drive where the eject button was pressed

user.data

user data set when the signal handler was connected.

drive-stop-button(volume.monitor, drive, user.data)

Emitted when the stop button is pressed on drive. Since 2.22

volume.monitor

The volume monitor emitting the signal.

drive

the drive where the stop button was pressed

user.data

user data set when the signal handler was connected.

mount-added(volume.monitor, mount, user.data)

Emitted when a mount is added.

volume.monitor

The volume monitor emitting the signal.

mount

a GMount that was added.

user.data

user data set when the signal handler was connected.

mount-changed(volume.monitor, mount, user.data)

Emitted when a mount changes.

volume.monitor

The volume monitor emitting the signal.

mount

a GMount that changed.

user.data

user data set when the signal handler was connected.

mount-pre-unmount(volume.monitor, mount, user.data)

Emitted when a mount is about to be removed.

volume.monitor

The volume monitor emitting the signal.

mount

a GMount that is being unmounted.

user.data

user data set when the signal handler was connected.

mount-removed(volume.monitor, mount, user.data)

Emitted when a mount is removed.

volume.monitor

The volume monitor emitting the signal.

mount

a GMount that was removed.

user.data

user data set when the signal handler was connected.

volume-added(volume.monitor, volume, user.data)

Emitted when a mountable volume is added to the system.

volume.monitor

The volume monitor emitting the signal.

volume

a GVolume that was added.

user.data

user data set when the signal handler was connected.

volume-changed(volume.monitor, volume, user.data)

Emitted when mountable volume is changed.

volume.monitor

The volume monitor emitting the signal.

volume

a GVolume that changed.

user.data

user data set when the signal handler was connected.

volume-removed(volume.monitor, volume, user.data)

Emitted when a mountable volume is removed from the system.

volume.monitor

The volume monitor emitting the signal.

volume

a GVolume that was removed.

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/GVolumeMonitor.html

[Package RGtk2 version 2.20.31 Index]