GdkDisplay {RGtk2}	R Documentation

GdkDisplay

Description

Controls the keyboard/mouse pointer grabs and a set of s

Methods and Functions

gdkDisplayOpen(display.name)
gdkDisplayGetDefault()
gdkDisplayGetName(object)
gdkDisplayGetNScreens(object)
gdkDisplayGetScreen(object, screen.num)
gdkDisplayGetDefaultScreen(object)
gdkDisplayPointerUngrab(object, time. = "GDK_CURRENT_TIME")
gdkDisplayKeyboardUngrab(object, time. = "GDK_CURRENT_TIME")
gdkDisplayPointerIsGrabbed(object)
gdkDisplayBeep(object)
gdkDisplaySync(object)
gdkDisplayFlush(object)
gdkDisplayClose(object)
gdkDisplayListDevices(object)
gdkDisplayGetEvent(object)
gdkDisplayPeekEvent(object)
gdkDisplayPutEvent(object, event)
gdkDisplayAddClientMessageFilter(object, message.type, func, data)
gdkDisplaySetDoubleClickTime(object, msec)
gdkDisplaySetDoubleClickDistance(object, distance)
gdkDisplayGetPointer(object)
gdkDisplayGetWindowAtPointer(object)
gdkDisplaySetPointerHooks(object, new.hooks)
gdkDisplayWarpPointer(object, screen, x, y)
gdkDisplaySupportsCursorColor(object)
gdkDisplaySupportsCursorAlpha(object)
gdkDisplayGetDefaultCursorSize(object)
gdkDisplayGetMaximalCursorSize(object)
gdkDisplayGetDefaultGroup(object)
gdkDisplaySupportsSelectionNotification(object)
gdkDisplayRequestSelectionNotification(object, selection)
gdkDisplaySupportsClipboardPersistence(object)
gdkDisplayStoreClipboard(object, clipboard.window, targets)
gdkDisplaySupportsShapes(object)
gdkDisplaySupportsInputShapes(object)
gdkDisplaySupportsComposite(object)
gdkDisplay(display.name)

Hierarchy

GObject
   +----GdkDisplay

Detailed Description

GdkDisplay objects purpose are two fold:

To grab/ungrab keyboard focus and mouse pointer

To manage and provide information about the GdkScreen(s) available for this GdkDisplay

GdkDisplay objects are the GDK representation of the X Display which can be described as a workstation consisting of a keyboard a pointing device (such as a mouse) and one or more screens. It is used to open and keep track of various GdkScreen objects currently instanciated by the application. It is also used to grab and release the keyboard and the mouse pointer.

Structures

GdkDisplay

The GdkDisplay struct is the GDK representation of an X display. All its fields are private and should not be accessed directly. Since 2.2

Convenient Construction

gdkDisplay is the equivalent of gdkDisplayOpen.

Signals

closed(display, is.error, user.data)

The ::closed signal is emitted when the connection to the windowing system for display is closed. Since 2.2

display

the object on which the signal is emitted

is.error

TRUE if the display was closed due to an error

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/GdkDisplay.html

[Package RGtk2 version 2.20.31 Index]