GdkDisplayManager {RGtk2}	R Documentation

GdkDisplayManager

Description

Maintains a list of all open s

Methods and Functions

gdkDisplayManagerGet()
gdkDisplayManagerGetDefaultDisplay(object)
gdkDisplayManagerSetDefaultDisplay(object, display)
gdkDisplayManagerListDisplays(object)
gdkDisplayGetCorePointer(object)

Hierarchy

GObject
   +----GdkDisplayManager

Detailed Description

The purpose of the GdkDisplayManager singleton object is to offer notification when displays appear or disappear or the default display changes.

Structures

GdkDisplayManager

The GdkDisplayManager struct has no interesting fields. Since 2.2

Signals

display-opened(display.manager, display, user.data)

The ::display_opened signal is emitted when a display is opened. Since 2.2

display.manager

the object on which the signal is emitted

display

the opened display

user.data

user data set when the signal handler was connected.

Properties

default-display [GdkDisplay : * : Read / Write]

The default display for GDK.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/GdkDisplayManager.html

[Package RGtk2 version 2.20.31 Index]