GdkEvent {RGtk2}	R Documentation

The GdkEvent Framework

Description

A GdkEvent represents a hardware-level event in GDK.

Details

In RGtk2, R objects corresponding to GdkEvents inherit from the GdkEvent class. A GdkEvent is not a GObject, but there are many specific event types that extend from the GdkEvent base S3 class. Specific event fields may be accessed in the same way as the fields of any external object in RGtk2 (via the [[ operator).

Note

While functions exist to create GdkEvents, there is currently no way in RGtk2 to set the event fields, so the functions are useless. If the user has some strange motivation to create hardware events in R, please contact the author.

Author(s)

Michael Lawrence

See Also

gdk-Event-Structures gdk-Events

[Package RGtk2 version 2.20.31 Index]