GdkPixbufLoader {RGtk2}	R Documentation

GdkPixbufLoader

Description

Application-driven progressive image loading.

Methods and Functions

gdkPixbufLoaderNew()
gdkPixbufLoaderNewWithType(image.type, .errwarn = TRUE)
gdkPixbufLoaderNewWithMimeType(mime.type, .errwarn = TRUE)
gdkPixbufLoaderGetFormat(object)
gdkPixbufLoaderWrite(object, buf, .errwarn = TRUE)
gdkPixbufLoaderSetSize(object, width, height)
gdkPixbufLoaderGetPixbuf(object)
gdkPixbufLoaderGetAnimation(object)
gdkPixbufLoaderClose(object, .errwarn = TRUE)
gdkPixbufLoader(image.type, mime.type, .errwarn = TRUE)

Hierarchy

GObject
   +----GdkPixbufLoader

Detailed Description

GdkPixbufLoader provides a way for applications to drive the process of loading an image, by letting them send the image data directly to the loader instead of having the loader read the data from a file. Applications can use this functionality instead of gdkPixbufNewFromFile or gdkPixbufAnimationNewFromFile when they need to parse image data in small chunks. For example, it should be used when reading an image from a (potentially) slow network connection, or when loading an extremely large file.

To use GdkPixbufLoader to load an image, just create a new one, and call gdkPixbufLoaderWrite to send the data to it. When done, gdkPixbufLoaderClose should be called to end the stream and finalize everything. The loader will emit three important signals throughout the process. The first, "size_prepared", will be called as soon as the image has enough information to determine the size of the image to be used. If you want to scale the image while loading it, you can call gdkPixbufLoaderSetSize in response to this signal. The second signal, "area_prepared", will be called as soon as the pixbuf of the desired has been allocated. You can obtain it by calling gdkPixbufLoaderGetPixbuf. If you want to use it, simply ref it. In addition, no actual information will be passed in yet, so the pixbuf can be safely filled with any temporary graphics (or an initial color) as needed. You can also call gdkPixbufLoaderGetPixbuf later and get the same pixbuf.

The last signal, "area_updated" gets called every time a region is updated. This way you can update a partially completed image. Note that you do not know anything about the completeness of an image from the area updated. For example, in an interlaced image, you need to make several passes before the image is done loading.

Loading an animation

Loading an animation is almost as easy as loading an image. Once the first "area_prepared" signal has been emitted, you can call gdkPixbufLoaderGetAnimation to get the GdkPixbufAnimation struct and gdkPixbufAnimationGetIter to get an GdkPixbufAnimationIter for displaying it.

Structures

GdkPixbufLoader

The GdkPixbufLoader struct contains only private fields.

Convenient Construction

gdkPixbufLoader is the result of collapsing the constructors of GdkPixbufLoader (gdkPixbufLoaderNew, gdkPixbufLoaderNewWithType, gdkPixbufLoaderNewWithMimeType) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Signals

area-prepared(loader, user.data)

This signal is emitted when the pixbuf loader has allocated the pixbuf in the desired size. After this signal is emitted, applications can call gdkPixbufLoaderGetPixbuf to fetch the partially-loaded pixbuf.

loader

the object which received the signal.

user.data

user data set when the signal handler was connected.

area-updated(loader, x, y, width, height, user.data)

This signal is emitted when a significant area of the image being loaded has been updated. Normally it means that a complete scanline has been read in, but it could be a different area as well. Applications can use this signal to know when to repaint areas of an image that is being loaded.

loader

the object which received the signal.

x

X offset of upper-left corner of the updated area.

y

Y offset of upper-left corner of the updated area.

width

Width of updated area.

height

Height of updated area.

user.data

user data set when the signal handler was connected.

closed(loader, user.data)

This signal is emitted when gdkPixbufLoaderClose is called. It can be used by different parts of an application to receive notification when an image loader is closed by the code that drives it.

loader

the object which received the signal.

user.data

user data set when the signal handler was connected.

size-prepared(loader, width, height, user.data)

This signal is emitted when the pixbuf loader has been fed the initial amount of data that is required to figure out the size of the image that it will create. Applications can call gdkPixbufLoaderSetSize in response to this signal to set the desired size to which the image should be scaled.

loader

the object which received the signal.

width

the original width of the image

height

the original height of the image

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk-pixbuf/GdkPixbufLoader.html

[Package RGtk2 version 2.20.31 Index]