GdkScreen {RGtk2}	R Documentation

GdkScreen

Description

Object representing a physical screen

Methods and Functions

gdkScreenGetDefault()
gdkScreenGetDefaultColormap(object)
gdkScreenSetDefaultColormap(object, colormap)
gdkScreenGetSystemColormap(object)
gdkScreenGetSystemVisual(object)
gdkScreenGetRgbColormap(object)
gdkScreenGetRgbVisual(object)
gdkScreenGetRgbaColormap(object)
gdkScreenGetRgbaVisual(object)
gdkScreenIsComposited(object)
gdkScreenGetRootWindow(object)
gdkScreenGetDisplay(object)
gdkScreenGetNumber(object)
gdkScreenGetWidth(object)
gdkScreenGetHeight(object)
gdkScreenGetWidthMm(object)
gdkScreenGetHeightMm(object)
gdkScreenListVisuals(object)
gdkScreenGetToplevelWindows(object)
gdkScreenMakeDisplayName(object)
gdkScreenGetNMonitors(object)
gdkScreenGetPrimaryMonitor(object)
gdkScreenGetMonitorGeometry(object, monitor.num)
gdkScreenGetMonitorAtPoint(object, x, y)
gdkScreenGetMonitorAtWindow(object, window)
gdkScreenGetMonitorHeightMm(object, monitor.num)
gdkScreenGetMonitorWidthMm(object, monitor.num)
gdkScreenGetMonitorPlugName(object, monitor.num)
gdkScreenBroadcastClientMessage(object, event)
gdkScreenGetSetting(object, name)
gdkScreenGetFontOptions(object)
gdkScreenSetFontOptions(object, options)
gdkScreenGetResolution(object)
gdkScreenSetResolution(object, dpi)
gdkScreenGetActiveWindow(object)
gdkScreenGetWindowStack(object)
gdkSpawnCommandLineOnScreen(screen, command.line, .errwarn = TRUE)

Hierarchy

GObject
   +----GdkScreen

Detailed Description

GdkScreen objects are the GDK representation of a physical screen. It is used throughout GDK and GTK+ to specify which screen the top level windows are to be displayed on. It is also used to query the screen specification and default settings such as the default colormap (gdkScreenGetDefaultColormap), the screen width (gdkScreenGetWidth), etc. Note that a screen may consist of multiple monitors which are merged to form a large screen area.

Structures

GdkScreen

This is a currently just a placeholder typedef for the first argument of the window.at.pointer function in GdkPointerHooks. It will be used when GDK gets multihead support. Since 2.2

Signals

composited-changed(screen, user.data)

The ::composited-changed signal is emitted when the composited status of the screen changes Since 2.10

screen

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

monitors-changed(screen, user.data)

The ::monitors-changed signal is emitted when the number, size or position of the monitors attached to the screen change.

Only for X11 and OS X for now. A future implementation for Win32 may be a possibility. Since 2.14

screen

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

size-changed(screen, user.data)

The ::size-changed signal is emitted when the pixel width or height of a screen changes. Since 2.2

screen

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

Properties

font-options [R object : Read / Write]

The default font options for the screen.

resolution [numeric : Read / Write]

The resolution for fonts on the screen. Default value: -1

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/GdkScreen.html

[Package RGtk2 version 2.20.31 Index]