GtkAdjustment {RGtk2}	R Documentation

GtkAdjustment

Description

A GtkObject representing an adjustable bounded value

Methods and Functions

gtkAdjustmentNew(value = NULL, lower = NULL, upper = NULL, step.incr = NULL, page.incr = NULL, page.size = NULL)
gtkAdjustmentGetValue(object)
gtkAdjustmentSetValue(object, value)
gtkAdjustmentClampPage(object, lower, upper)
gtkAdjustmentChanged(object)
gtkAdjustmentValueChanged(object)
gtkAdjustmentConfigure(object, value, lower, upper, step.increment, page.increment, page.size)
gtkAdjustmentGetLower(object)
gtkAdjustmentGetPageIncrement(object)
gtkAdjustmentGetPageSize(object)
gtkAdjustmentGetStepIncrement(object)
gtkAdjustmentGetUpper(object)
gtkAdjustmentSetLower(object, lower)
gtkAdjustmentSetPageIncrement(object, page.increment)
gtkAdjustmentSetPageSize(object, page.size)
gtkAdjustmentSetStepIncrement(object, step.increment)
gtkAdjustmentSetUpper(object, upper)
gtkAdjustment(value = NULL, lower = NULL, upper = NULL, step.incr = NULL, page.incr = NULL, page.size = NULL)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkAdjustment

Detailed Description

The GtkAdjustment object represents a value which has an associated lower and upper bound, together with step and page increments, and a page size. It is used within several GTK+ widgets, including GtkSpinButton, GtkViewport, and GtkRange (which is a base class for GtkHScrollbar, GtkVScrollbar, GtkHScale, and GtkVScale).

The GtkAdjustment object does not update the value itself. Instead it is left up to the owner of the GtkAdjustment to control the value.

The owner of the GtkAdjustment typically calls the gtkAdjustmentValueChanged and gtkAdjustmentChanged functions after changing the value and its bounds. This results in the emission of the "value_changed" or "changed" signal respectively.

Structures

GtkAdjustment

The GtkAdjustment struct contains the following fields.

numeric lower; 	 the minimum value.

numeric upper; 	 the maximum value.

numeric value; 	 the current value.

numeric step_increment; 	 the increment to use to make minor changes to the value. In a GtkScrollbar this increment is used when the mouse is clicked on the arrows at the top and bottom of the scrollbar, to scroll by a small amount.

numeric page_increment; 	 the increment to use to make major changes to the value. In a GtkScrollbar this increment is used when the mouse is clicked in the trough, to scroll by a large amount.

numeric page_size; 	 the page size. In a GtkScrollbar this is the size of the area which is currently visible.

Convenient Construction

gtkAdjustment is the equivalent of gtkAdjustmentNew.

Signals

changed(adjustment, user.data)

Emitted when one or more of the GtkAdjustment fields have been changed, other than the value field.

adjustment

the object which received the signal.

user.data

user data set when the signal handler was connected.

value-changed(adjustment, user.data)

Emitted when the GtkAdjustment value field has been changed.

adjustment

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

lower [numeric : Read / Write]

The minimum value of the adjustment. Default value: 0 Since 2.4

page-increment [numeric : Read / Write]

The page increment of the adjustment. Default value: 0 Since 2.4

page-size [numeric : Read / Write]

The page size of the adjustment. Note that the page-size is irrelevant and should be set to zero if the adjustment is used for a simple scalar value, e.g. in a GtkSpinButton. Default value: 0 Since 2.4

step-increment [numeric : Read / Write]

The step increment of the adjustment. Default value: 0 Since 2.4

upper [numeric : Read / Write]

The maximum value of the adjustment. Note that values will be restricted by upper - page-size if the page-size property is nonzero. Default value: 0 Since 2.4

value [numeric : Read / Write]

The value of the adjustment. Default value: 0 Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkAdjustment.html

[Package RGtk2 version 2.20.31 Index]