GtkArrow {RGtk2}	R Documentation

GtkArrow

Description

Displays an arrow

Methods and Functions

gtkArrowNew(arrow.type = NULL, shadow.type = NULL, show = TRUE)
gtkArrowSet(object, arrow.type, shadow.type)
gtkArrow(arrow.type = NULL, shadow.type = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkMisc
                           +----GtkArrow

Interfaces

GtkArrow implements AtkImplementorIface and GtkBuildable.

Detailed Description

GtkArrow should be used to draw simple arrows that need to point in one of the four cardinal directions (up, down, left, or right). The style of the arrow can be one of shadow in, shadow out, etched in, or etched out. Note that these directions and style types may be ammended in versions of Gtk to come.

GtkArrow will fill any space alloted to it, but since it is inherited from GtkMisc, it can be padded and/or aligned, to fill exactly the space the programmer desires.

Arrows are created with a call to gtkArrowNew. The direction or style of an arrow can be changed after creation by using gtkArrowSet.

Structures

GtkArrow

The GtkArrow struct containes the following fields. (These fields should be considered read-only. They should never be set by an application.)

Convenient Construction

gtkArrow is the equivalent of gtkArrowNew.

Properties

arrow-type [GtkArrowType : Read / Write]

The direction the arrow should point. Default value: GTK_ARROW_RIGHT

shadow-type [GtkShadowType : Read / Write]

Appearance of the shadow surrounding the arrow. Default value: GTK_SHADOW_OUT

Style Properties

arrow-scaling [numeric : Read]

Amount of space used up by arrow. Allowed values: [0,1] Default value: 0.7

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkArrow.html

[Package RGtk2 version 2.20.31 Index]