GtkAspectFrame {RGtk2}	R Documentation

GtkAspectFrame

Description

A frame that constrains its child to a particular aspect ratio

Methods and Functions

gtkAspectFrameNew(label = NULL, xalign = NULL, yalign = NULL, ratio = NULL, obey.child = NULL, show = TRUE)
gtkAspectFrameSet(object, xalign = 0, yalign = 0, ratio = 1, obey.child = 1)
gtkAspectFrame(label = NULL, xalign = NULL, yalign = NULL, ratio = NULL, obey.child = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkFrame
                                       +----GtkAspectFrame

Interfaces

GtkAspectFrame implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkAspectFrame is useful when you want pack a widget so that it can resize but always retains the same aspect ratio. For instance, one might be drawing a small preview of a larger image. GtkAspectFrame derives from GtkFrame, so it can draw a label and a frame around the child. The frame will be "shrink-wrapped" to the size of the child.

Structures

GtkAspectFrame

undocumented

Convenient Construction

gtkAspectFrame is the equivalent of gtkAspectFrameNew.

Properties

obey-child [logical : Read / Write]

Force aspect ratio to match that of the frame's child. Default value: TRUE

ratio [numeric : Read / Write]

Aspect ratio if obey_child is FALSE. Allowed values: [0.0001,10000] Default value: 1

xalign [numeric : Read / Write]

X alignment of the child. Allowed values: [0,1] Default value: 0.5

yalign [numeric : Read / Write]

Y alignment of the child. Allowed values: [0,1] Default value: 0.5

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkAspectFrame.html

[Package RGtk2 version 2.20.31 Index]