GtkBin {RGtk2}	R Documentation

GtkBin

Description

A container with just one child

Methods and Functions

gtkBinGetChild(object)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkWindow
                                 +----GtkAlignment
                                 +----GtkFrame
                                 +----GtkButton
                                 +----GtkItem
                                 +----GtkComboBox
                                 +----GtkEventBox
                                 +----GtkExpander
                                 +----GtkHandleBox
                                 +----GtkToolItem
                                 +----GtkScrolledWindow
                                 +----GtkViewport

Interfaces

GtkBin implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkBin widget is a container with just one child. It is not very useful itself, but it is useful for deriving subclasses, since it provides common code needed for handling a single child widget.

Many GTK+ widgets are subclasses of GtkBin, including GtkWindow, GtkButton, GtkFrame, GtkHandleBox, and GtkScrolledWindow.

Structures

GtkBin

The GtkBin struct contains the following fields. (These fields should be considered read-only. They should never be set by an application.)

GtkWidget *child; 	 the child widget.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkBin.html

[Package RGtk2 version 2.20.31 Index]