GtkButtonBox {RGtk2}	R Documentation

GtkButtonBox

Description

Base class for GtkHButtonBox and GtkVButtonBox

Methods and Functions

gtkButtonBoxGetLayout(object)
gtkButtonBoxGetChildSize(object)
gtkButtonBoxGetChildIpadding(object)
gtkButtonBoxGetChildSecondary(object, child)
gtkButtonBoxSetLayout(object, layout.style)
gtkButtonBoxSetChildSize(object, min.width, min.height)
gtkButtonBoxSetChildIpadding(object, ipad.x, ipad.y)
gtkButtonBoxSetChildSecondary(object, child, is.secondary)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBox
                                 +----GtkButtonBox
                                       +----GtkHButtonBox
                                       +----GtkVButtonBox

Interfaces

GtkButtonBox implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

The primary purpose of this class is to keep track of the various properties of GtkHButtonBox and GtkVButtonBox widgets. gtkButtonBoxGetChildSize retrieves the minimum width and height for widgets in a given button box. gtkButtonBoxSetChildSize allows those properties to be changed.

The internal padding of buttons can be retrieved and changed per button box using gtkButtonBoxGetChildIpadding and gtkButtonBoxSetChildIpadding respectively. gtkButtonBoxGetSpacing() and gtkButtonBoxSetSpacing() retrieve and change default number of pixels between buttons, respectively. gtkButtonBoxGetLayout and gtkButtonBoxSetLayout retrieve and alter the method used to spread the buttons in a button box across the container, respectively.

The main purpose of GtkButtonBox is to make sure the children have all the same size. Therefore it ignores the homogeneous property which it inherited from GtkBox, and always behaves as if homogeneous was TRUE.

Structures

GtkButtonBox

This is a read-only struct; no members should be modified directly.

Properties

layout-style [GtkButtonBoxStyle : Read / Write]

How to layout the buttons in the box. Possible values are default, spread, edge, start and end. Default value: GTK_BUTTONBOX_DEFAULT_STYLE

Style Properties

child-internal-pad-x [integer : Read]

Amount to increase child's size on either side. Allowed values: >= 0 Default value: 4

child-internal-pad-y [integer : Read]

Amount to increase child's size on the top and bottom. Allowed values: >= 0 Default value: 0

child-min-height [integer : Read]

Minimum height of buttons inside the box. Allowed values: >= 0 Default value: 27

child-min-width [integer : Read]

Minimum width of buttons inside the box. Allowed values: >= 0 Default value: 85

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkButtonBox.html

[Package RGtk2 version 2.20.31 Index]