GtkCList {RGtk2}	R Documentation

GtkCList

Description

A multi-columned scrolling list widget

Methods and Functions

gtkCListNew(columns = 1, show = TRUE)
gtkCListNewWithTitles(columns = 1, titles, show = TRUE)
gtkCListSetShadowType(object, type)
gtkCListSetSelectionMode(object, mode)
gtkCListFreeze(object)
gtkCListThaw(object)
gtkCListColumnTitlesShow(object)
gtkCListColumnTitlesHide(object)
gtkCListColumnTitleActive(object, column)
gtkCListColumnTitlePassive(object, column)
gtkCListColumnTitlesActive(object)
gtkCListColumnTitlesPassive(object)
gtkCListSetColumnTitle(object, column, title)
gtkCListSetColumnWidget(object, column, widget)
gtkCListSetColumnJustification(object, column, justification)
gtkCListSetColumnVisibility(object, column, visible)
gtkCListSetColumnResizeable(object, column, resizeable)
gtkCListSetColumnAutoResize(object, column, auto.resize)
gtkCListOptimalColumnWidth(object, column)
gtkCListSetColumnWidth(object, column, width)
gtkCListSetColumnMinWidth(object, column, min.width)
gtkCListSetColumnMaxWidth(object, column, max.width)
gtkCListSetRowHeight(object, height)
gtkCListMoveto(object, row, column, row.align, col.align)
gtkCListRowIsVisible(object, row)
gtkCListGetCellType(object, row, column)
gtkCListSetText(w, row, cols, values, zeroBased = TRUE)
gtkCListGetText(w, row, cols, zeroBased = TRUE)
gtkCListSetPixmap(object, row, column, pixmap, mask = NULL)
gtkCListGetPixmap(object, row, column)
gtkCListSetPixtext(object, row, column, text, spacing, pixmap, mask)
gtkCListGetPixtext(object, row, column)
gtkCListSetForeground(object, row, color)
gtkCListSetBackground(object, row, color)
gtkCListSetCellStyle(object, row, column, style)
gtkCListGetCellStyle(object, row, column)
gtkCListSetRowStyle(object, row, style)
gtkCListGetRowStyle(object, row)
gtkCListSetShift(object, row, column, vertical, horizontal)
gtkCListSetSelectable(object, row, selectable)
gtkCListGetSelectable(object, row)
gtkCListPrepend(object, text)
gtkCListAppend(object, text)
gtkCListInsert(object, row, text)
gtkCListRemove(object, row)
gtkCListSetRowData(object, row, data = NULL)
gtkCListSetRowDataFull(object, row, data = NULL)
gtkCListGetRowData(object, row)
gtkCListFindRowFromData(object, data)
gtkCListSelectRow(object, row, column)
gtkCListUnselectRow(object, row, column)
gtkCListUndoSelection(object)
gtkCListClear(object)
gtkCListGetSelectionInfo(object, x, y)
gtkCListSelectAll(object)
gtkCListUnselectAll(object)
gtkCListSwapRows(object, row1, row2)
gtkCListSetCompareFunc(object, cmp.func)
gtkCListSetSortColumn(object, column)
gtkCListSetSortType(object, sort.type)
gtkCListSort(object)
gtkCListSetAutoSort(object, auto.sort)
gtkCListColumnsAutosize(object)
gtkCListGetColumnTitle(object, column)
gtkCListGetColumnWidget(object, column)
gtkCListGetHadjustment(object)
gtkCListGetVadjustment(object)
gtkCListRowMove(object, source.row, dest.row)
gtkCListSetButtonActions(object, button, button.actions)
gtkCListSetHadjustment(object, adjustment)
gtkCListSetReorderable(object, reorderable)
gtkCListSetUseDragIcons(object, use.icons)
gtkCListSetVadjustment(object, adjustment)
gtkCList(columns = 1, titles, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkCList
                                 +----GtkCTree

Interfaces

GtkCList implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkCList widget is a very useful multi-columned scrolling list. It can display data in nicely aligned vertical columns, with titles at the top of the list.

GtkCList has been deprecated since GTK+ 2.0 and should not be used in newly written code. Use GtkTreeView instead.

Structures

GtkCList

WARNING: GtkCList is deprecated and should not be used in newly-written code. This is the embodiment of the GtkCList widget. This structure contains only private data, and should be accessed only via the CList API.

GtkCListRow

WARNING: GtkCListRow is deprecated and should not be used in newly-written code. A structure that the GtkCList widget uses to keep track of information about its rows.

cell

[GtkCell]

state

[GtkStateType]

foreground

[GdkColor]

background

[GdkColor]

style

[GtkStyle]

data

[Robject]

destroy

[GtkDestroyNotify]

fgSet

[numeric]

bgSet

[numeric]

selectable

[numeric]

Convenient Construction

gtkCList is the result of collapsing the constructors of GtkCList (gtkCListNew, gtkCListNewWithTitles) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Enums and Flags

GtkCellType

WARNING: GtkCellType is deprecated and should not be used in newly-written code. Identifies the type of element in the current cell of the CList. Cells can contain text, pixmaps, or both. Unfortunately support for GTK_CELL_WIDGET was never completed.

empty

undocumented

text

undocumented

pixmap

undocumented

pixtext

undocumented

widget

undocumented

GtkButtonAction

WARNING: GtkButtonAction is deprecated and should not be used in newly-written code. Values for specifying what mouse button events a CList will react to.

ignored

undocumented

selects

undocumented

drags

undocumented

expands

undocumented

GtkCListDragPos

WARNING: GtkCListDragPos is deprecated and should not be used in newly-written code. An enumeration for drag operations.

none

undocumented

before

undocumented

into

undocumented

after

undocumented

User Functions

GtkCListCompareFunc(clist, ptr1, ptr2)

Function prototype for the compare function callback.

clist

The GtkCList that is affected.

ptr1

A R object to the first node to compare.

ptr2

A R object to the second node to compare.

Returns: [integer] 0 if the nodes are equal, less than 0 if the first node should come before the second, and greater than 1 if the second come before the first.

Signals

abort-column-resize(clist, user.data)

This signal is emitted when a column resize is aborted.

clist

the object which received the signal.

user.data

user data set when the signal handler was connected.

click-column(clist, column, user.data)

This signal is emitted when a column title is clicked.

clist

The object which received the signal.

column

The number of the column.

user.data

user data set when the signal handler was connected.

end-selection(clist, user.data)

This signal is emitted when a selection ends in a multiple selection CList.

clist

the object which received the signal.

user.data

user data set when the signal handler was connected.

extend-selection(clist, scroll.type, position, auto.start.selection, user.data)

This signal is emitted when the selection is extended.

clist

the object which received the signal.

scroll.type

A GtkScrollType value of any scrolling operation the occured during the selection.

position

A value between 0.0 and 1.0.

auto.start.selection

TRUE or FALSE.

user.data

user data set when the signal handler was connected.

resize-column(clist, column, width, user.data)

This signal is emitted when a column is resized.

clist

The object which received the signal.

column

The number of the column

width

The new width of the column.

user.data

user data set when the signal handler was connected.

row-move(clist, arg1, arg2, user.data)

This signal is emitted when a row is moved.

clist

The object which received the signal.

arg1

The source position of the row.

arg2

The destination position of the row.

user.data

user data set when the signal handler was connected.

scroll-horizontal(clist, scroll.type, position, user.data)

This signal is emitted when the CList is scrolled horizontally.

clist

the object which received the signal.

scroll.type

A GtkScrollType value of how the scroll operation occured.

position

a value between 0.0 and 1.0.

user.data

user data set when the signal handler was connected.

scroll-vertical(clist, scroll.type, position, user.data)

This signal is emitted when the CList is scrolled vertically.

clist

the object which received the signal.

scroll.type

A GtkScrollType value of how the scroll operation occured.

position

A value between 0.0 and 1.0.

user.data

user data set when the signal handler was connected.

select-all(clist, user.data)

This signal is emitted when all the rows are selected in a CList.

clist

the object which received the signal.

user.data

user data set when the signal handler was connected.

select-row(clist, row, column, event, user.data)

This signal is emitted when the user selects a row in the list. It is emitted for every row that is selected in a multi-selection or by calling gtkCListSelectAll.

clist

The object which received the signal.

row

The row selected.

column

The column where the selection occured.

event

A GdkEvent structure for the selection.

user.data

user data set when the signal handler was connected.

set-scroll-adjustments(clist, user.data)

undocumented

clist

the object which received the signal.

user.data

user data set when the signal handler was connected.

start-selection(clist, user.data)

This signal is emitted when a drag-selection is started in a multiple-selection CList.

clist

the object which received the signal.

user.data

user data set when the signal handler was connected.

toggle-add-mode(clist, user.data)

This signal is emitted when "add mode" is toggled.

clist

the object which received the signal.

user.data

user data set when the signal handler was connected.

toggle-focus-row(clist, user.data)

undocumented

clist

The object which received the signal.

user.data

user data set when the signal handler was connected.

undo-selection(clist, user.data)

This signal is emitted when an undo selection occurs in the CList, probably via calling gtkCListUndoSelection.

clist

the object which received the signal.

user.data

user data set when the signal handler was connected.

unselect-all(clist, user.data)

This signal is emitted when all rows are unselected in a CList.

clist

the object which received the signal.

user.data

user data set when the signal handler was connected.

unselect-row(clist, row, column, user.data)

This signal is emitted when the user unselects a row in the list. It is emitted for every row that is unselected in a multi-selection or by calling gtkCListUnselectAll. It is also emitted for the previously selected row in a "single" or "browse" mode CList.

clist

The object which received the signal.

row

The selected row

column

The column where the selection occured.

user.data

user data set when the signal handler was connected.

Properties

n-columns [numeric : Read / Write / Construct Only]

An integer value for a column. Default value: 0

reorderable [logical : Read / Write]

A boolean value for determining if the user can re-order the CList's columns. Default value: FALSE

row-height [numeric : Read / Write]

An integer value representing the height of a row in pixels. Default value: 0

selection-mode [GtkSelectionMode : Read / Write]

Sets the type of selection mode for the CList. Default value: GTK_SELECTION_NONE

shadow-type [GtkShadowType : Read / Write]

Sets the shadowing for the CList. Default value: GTK_SHADOW_NONE

sort-type [GtkSortType : Read / Write]

Default value: GTK_SORT_ASCENDING

titles-active [logical : Read / Write]

A boolean value for setting whether the column titles can be clicked. Default value: FALSE

use-drag-icons [logical : Read / Write]

A boolean value for setting whether to use icons during drag operations. Default value: FALSE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCList.html

[Package RGtk2 version 2.20.31 Index]