GtkCalendar {RGtk2}	R Documentation

GtkCalendar

Description

Displays a calendar and allows the user to select a date

Methods and Functions

gtkCalendarNew(show = TRUE)
gtkCalendarSelectMonth(object, month, year)
gtkCalendarSelectDay(object, day)
gtkCalendarMarkDay(object, day)
gtkCalendarUnmarkDay(object, day)
gtkCalendarClearMarks(object)
gtkCalendarGetDisplayOptions(object)
gtkCalendarSetDisplayOptions(object, flags)
gtkCalendarGetDate(object)
gtkCalendarSetDetailFunc(object, func, data)
gtkCalendarGetDetailWidthChars(object)
gtkCalendarSetDetailWidthChars(object, chars)
gtkCalendarGetDetailHeightRows(object)
gtkCalendarSetDetailHeightRows(object, rows)
gtkCalendarDisplayOptions(object, flags)
gtkCalendarFreeze(object)
gtkCalendarThaw(object)
gtkCalendar(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkCalendar

Interfaces

GtkCalendar implements AtkImplementorIface and GtkBuildable.

Detailed Description

GtkCalendar is a widget that displays a calendar, one month at a time. It can be created with gtkCalendarNew.

The month and year currently displayed can be altered with gtkCalendarSelectMonth. The exact day can be selected from the displayed month using gtkCalendarSelectDay.

To place a visual marker on a particular day, use gtkCalendarMarkDay and to remove the marker, gtkCalendarUnmarkDay. Alternative, all marks can be cleared with gtkCalendarClearMarks.

The way in which the calendar itself is displayed can be altered using gtkCalendarSetDisplayOptions.

The selected date can be retrieved from a GtkCalendar using gtkCalendarGetDate.

Structures

GtkCalendar

num_marked_dates is an integer containing the number of days that have a mark over them. marked_date is a list containing the day numbers that currently have a mark over them. month, year, and selected_day contain the currently visible month, year, and selected day respectively.

All of these fields should be considered read only, and everything in this struct should only be modified using the functions provided below.

Convenient Construction

gtkCalendar is the equivalent of gtkCalendarNew.

Enums and Flags

GtkCalendarDisplayOptions

These options can be used to influence the display and behaviour of a GtkCalendar.

show-heading

Specifies that the month and year should be displayed.

show-day-names

Specifies that three letter day descriptions should be present.

no-month-change

Prevents the user from switching months with the calendar.

show-week-numbers

Displays each week numbers of the current year, down the left side of the calendar.

week-start-monday

Since GTK+ 2.4, this option is deprecated and ignored by GTK+. The information on which day the calendar week starts is derived from the locale.

User Functions

GtkCalendarDetailFunc(calendar, year, month, day, user.data)

This kind of functions provide Pango markup with detail information for the specified day. Examples for such details are holidays or appointments. The function returns NULL when no information is available. Since 2.14

calendar

a GtkCalendar.

year

the year for which details are needed.

month

the month for which details are needed.

day

the day of month for which details are needed.

user.data

the data passed with gtkCalendarSetDetailFunc.

Returns: [character] Newly allocated string with Pango markup with details for the specified day, or NULL.

Signals

day-selected(calendar, user.data)

Emitted when the user selects a day.

calendar

the object which received the signal.

user.data

user data set when the signal handler was connected.

day-selected-double-click(calendar, user.data)

undocumented

calendar

the object which received the signal.

user.data

user data set when the signal handler was connected.

month-changed(calendar, user.data)

Emitted when the user clicks a button to change the selected month on a calendar.

calendar

the object which received the signal.

user.data

user data set when the signal handler was connected.

next-month(calendar, user.data)

undocumented

calendar

the object which received the signal.

user.data

user data set when the signal handler was connected.

next-year(calendar, user.data)

undocumented

calendar

the object which received the signal.

user.data

user data set when the signal handler was connected.

prev-month(calendar, user.data)

undocumented

calendar

the object which received the signal.

user.data

user data set when the signal handler was connected.

prev-year(calendar, user.data)

undocumented

calendar

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

day [integer : Read / Write]

The selected day (as a number between 1 and 31, or 0 to unselect the currently selected day). This property gets initially set to the current day. Allowed values: [0,31] Default value: 0

detail-height-rows [integer : Read / Write]

Height of a detail cell, in rows. A value of 0 allows any width. See gtkCalendarSetDetailFunc. Allowed values: [0,127] Default value: 0 Since 2.14

detail-width-chars [integer : Read / Write]

Width of a detail cell, in characters. A value of 0 allows any width. See gtkCalendarSetDetailFunc. Allowed values: [0,127] Default value: 0 Since 2.14

month [integer : Read / Write]

The selected month (as a number between 0 and 11). This property gets initially set to the current month. Allowed values: [0,11] Default value: 0

no-month-change [logical : Read / Write]

Determines whether the selected month can be changed. Default value: FALSE Since 2.4

show-day-names [logical : Read / Write]

Determines whether day names are displayed. Default value: TRUE Since 2.4

show-details [logical : Read / Write]

Determines whether details are shown directly in the widget, or if they are available only as tooltip. When this property is set days with details are marked. Default value: TRUE Since 2.14

show-heading [logical : Read / Write]

Determines whether a heading is displayed. Default value: TRUE Since 2.4

show-week-numbers [logical : Read / Write]

Determines whether week numbers are displayed. Default value: FALSE Since 2.4

year [integer : Read / Write]

The selected year. This property gets initially set to the current year. Allowed values: [0,4194303] Default value: 0

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCalendar.html

[Package RGtk2 version 2.20.31 Index]