GtkCellEditable {RGtk2}	R Documentation

GtkCellEditable

Description

Interface for widgets which can are used for editing cells

Methods and Functions

gtkCellEditableStartEditing(object, event = NULL)
gtkCellEditableEditingDone(object)
gtkCellEditableRemoveWidget(object)

Hierarchy

GInterface
   +----GtkCellEditable

Implementations

GtkCellEditable is implemented by GtkComboBox, GtkComboBoxEntry, GtkEntry and GtkSpinButton.

Detailed Description

The GtkCellEditable interface must be implemented for widgets to be usable when editing the contents of a GtkTreeView cell.

Structures

GtkCellEditable

undocumented

Signals

editing-done(cell.editable, user.data)

This signal is a sign for the cell renderer to update its value from the cell.editable.

Implementations of GtkCellEditable are responsible for emitting this signal when they are done editing, e.g. GtkEntry is emitting it when the user presses Enter. gtkCellEditableEditingDone is a convenience method for emitting ::editing-done.

cell.editable

the object on which the signal was emitted

user.data

user data set when the signal handler was connected.

remove-widget(cell.editable, user.data)

This signal is meant to indicate that the cell is finished editing, and the widget may now be destroyed.

Implementations of GtkCellEditable are responsible for emitting this signal when they are done editing. It must be emitted after the gtkCellEditableEditingDone signal, to give the cell renderer a chance to update the cell's value before the widget is removed. gtkCellEditableRemoveWidget is a convenience method for emitting ::remove-widget.

cell.editable

the object on which the signal was emitted

user.data

user data set when the signal handler was connected.

Properties

editing-canceled [logical : Read / Write]

Indicates whether editing on the cell has been canceled. Default value: FALSE Since 2.20

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCellEditable.html

[Package RGtk2 version 2.20.31 Index]