GtkCellRendererAccel {RGtk2}	R Documentation

GtkCellRendererAccel

Description

Renders a keyboard accelerator in a cell

Methods and Functions

gtkCellRendererAccelNew()
gtkCellRendererAccel()

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkCellRenderer
                     +----GtkCellRendererText
                           +----GtkCellRendererAccel

Detailed Description

GtkCellRendererAccel displays a keyboard accelerator (i.e. a key combination like <Control>-a). If the cell renderer is editable, the accelerator can be changed by simply typing the new combination.

The GtkCellRendererAccel cell renderer was added in GTK+ 2.10.

Structures

GtkCellRendererAccel

undocumented

Convenient Construction

gtkCellRendererAccel is the equivalent of gtkCellRendererAccelNew.

Enums and Flags

GtkCellRendererAccelMode

undocumented

gtk

undocumented

other

undocumented

Signals

accel-cleared(accel, path.string, user.data)

Gets emitted when the user has removed the accelerator. Since 2.10

accel

the object reveiving the signal

path.string

the path identifying the row of the edited cell

user.data

user data set when the signal handler was connected.

accel-edited(accel, path.string, accel.key, accel.mods, hardware.keycode, user.data)

Gets emitted when the user has selected a new accelerator. Since 2.10

accel

the object reveiving the signal

path.string

the path identifying the row of the edited cell

accel.key

the new accelerator keyval

accel.mods

the new acclerator modifier mask

hardware.keycode

the keycode of the new accelerator

user.data

user data set when the signal handler was connected.

Properties

accel-key [numeric : Read / Write]

The keyval of the accelerator. Allowed values: <= G_MAXINT Default value: 0 Since 2.10

accel-mode [GtkCellRendererAccelMode : Read / Write]

Determines if the edited accelerators are GTK+ accelerators. If they are, consumed modifiers are suppressed, only accelerators accepted by GTK+ are allowed, and the accelerators are rendered in the same way as they are in menus. Default value: GTK_CELL_RENDERER_ACCEL_MODE_GTK Since 2.10

accel-mods [GdkModifierType : Read / Write]

The modifier mask of the accelerator. Since 2.10

keycode [numeric : Read / Write]

The hardware keycode of the accelerator. Note that the hardware keycode is only relevant if the key does not have a keyval. Normally, the keyboard configuration should assign keyvals to all keys. Allowed values: <= G_MAXINT Default value: 0 Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCellRendererAccel.html

[Package RGtk2 version 2.20.31 Index]