GtkCellRendererCombo {RGtk2}	R Documentation

GtkCellRendererCombo

Description

Renders a combobox in a cell

Methods and Functions

gtkCellRendererComboNew()
gtkCellRendererCombo()

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkCellRenderer
                     +----GtkCellRendererText
                           +----GtkCellRendererCombo

Detailed Description

GtkCellRendererCombo renders text in a cell like GtkCellRendererText from which it is derived. But while GtkCellRendererText offers a simple entry to edit the text, GtkCellRendererCombo offers a GtkComboBox or GtkComboBoxEntry widget to edit the text. The values to display in the combo box are taken from the tree model specified in the model property.

The combo cell renderer takes care of adding a text cell renderer to the combo box and sets it to display the column specified by its text-column property. Further properties of the comnbo box can be set in a handler for the editing-started signal.

The GtkCellRendererCombo cell renderer was added in GTK+ 2.6.

Structures

GtkCellRendererCombo

undocumented

Convenient Construction

gtkCellRendererCombo is the equivalent of gtkCellRendererComboNew.

Signals

changed(combo, path.string, new.iter, user.data)

This signal is emitted each time after the user selected an item in the combo box, either by using the mouse or the arrow keys. Contrary to GtkComboBox, GtkCellRendererCombo::changed is not emitted for changes made to a selected item in the entry. The argument new.iter corresponds to the newly selected item in the combo box and it is relative to the GtkTreeModel set via the model property on GtkCellRendererCombo.

Note that as soon as you change the model displayed in the tree view, the tree view will immediately cease the editing operating. This means that you most probably want to refrain from changing the model until the combo cell renderer emits the edited or editing_canceled signal. Since 2.14

combo

the object on which the signal is emitted

path.string

a string of the path identifying the edited cell (relative to the tree view model)

new.iter

the new iter selected in the combo box (relative to the combo box model)

user.data

user data set when the signal handler was connected.

Properties

has-entry [logical : Read / Write]

If TRUE, the cell renderer will include an entry and allow to enter values other than the ones in the popup list. Default value: TRUE Since 2.6

model [GtkTreeModel : * : Read / Write]

Holds a tree model containing the possible values for the combo box. Use the text_column property to specify the column holding the values. Since 2.6

text-column [integer : Read / Write]

Specifies the model column which holds the possible values for the combo box.

Note that this refers to the model specified in the model property, not the model backing the tree view to which this cell renderer is attached. GtkCellRendererCombo automatically adds a text cell renderer for this column to its combo box. Allowed values: >= -1 Default value: -1 Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCellRendererCombo.html

[Package RGtk2 version 2.20.31 Index]