GtkCellRendererPixbuf {RGtk2}	R Documentation

GtkCellRendererPixbuf

Description

Renders a pixbuf in a cell

Methods and Functions

gtkCellRendererPixbufNew()
gtkCellRendererPixbuf()

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkCellRenderer
                     +----GtkCellRendererPixbuf

Detailed Description

A GtkCellRendererPixbuf can be used to render an image in a cell. It allows to render either a given GdkPixbuf (set via the pixbuf property) or a stock icon (set via the stock-id property).

To support the tree view, GtkCellRendererPixbuf also supports rendering two alternative pixbufs, when the is-expander property is TRUE. If the is-expanded property is TRUE and the pixbuf-expander-open property is set to a pixbuf, it renders that pixbuf, if the is-expanded property is FALSE and the pixbuf-expander-closed property is set to a pixbuf, it renders that one.

Structures

GtkCellRendererPixbuf

undocumented

Convenient Construction

gtkCellRendererPixbuf is the equivalent of gtkCellRendererPixbufNew.

Properties

follow-state [logical : Read / Write]

Specifies whether the rendered pixbuf should be colorized according to the GtkCellRendererState. Default value: FALSE Since 2.8

gicon [GIcon : * : Read / Write]

The GIcon representing the icon to display. If the icon theme is changed, the image will be updated automatically. Since 2.14

icon-name [character : * : Read / Write]

The name of the themed icon to display. This property only has an effect if not overridden by "stock_id" or "pixbuf" properties. Default value: NULL Since 2.8

pixbuf [GdkPixbuf : * : Read / Write]

The pixbuf to render.

pixbuf-expander-closed [GdkPixbuf : * : Read / Write]

Pixbuf for closed expander.

pixbuf-expander-open [GdkPixbuf : * : Read / Write]

Pixbuf for open expander.

stock-detail [character : * : Read / Write]

Render detail to pass to the theme engine. Default value: NULL

stock-id [character : * : Read / Write]

The stock ID of the stock icon to render. Default value: NULL

stock-size [numeric : Read / Write]

The GtkIconSize value that specifies the size of the rendered icon. Default value: 1

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCellRendererPixbuf.html

[Package RGtk2 version 2.20.31 Index]