GtkCellRendererSpin {RGtk2}	R Documentation

GtkCellRendererSpin

Description

Renders a spin button in a cell

Methods and Functions

gtkCellRendererSpinNew()
gtkCellRendererSpin()

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkCellRenderer
                     +----GtkCellRendererText
                           +----GtkCellRendererSpin

Detailed Description

GtkCellRendererSpin renders text in a cell like GtkCellRendererText from which it is derived. But while GtkCellRendererText offers a simple entry to edit the text, GtkCellRendererSpin offers a GtkSpinButton widget. Of course, that means that the text has to be parseable as a floating point number.

The range of the spinbutton is taken from the adjustment property of the cell renderer, which can be set explicitly or mapped to a column in the tree model, like all properties of cell renders. GtkCellRendererSpin also has properties for the climb rate and the number of digits to display. Other GtkSpinButton properties can be set in a handler for the start-editing signal.

The GtkCellRendererSpin cell renderer was added in GTK+ 2.10.

Structures

GtkCellRendererSpin

undocumented

Convenient Construction

gtkCellRendererSpin is the equivalent of gtkCellRendererSpinNew.

Properties

adjustment [GtkAdjustment : * : Read / Write]

The adjustment that holds the value of the spinbutton. This must be non-NULL for the cell renderer to be editable. Since 2.10

climb-rate [numeric : Read / Write]

The acceleration rate when you hold down a button. Allowed values: >= 0 Default value: 0 Since 2.10

digits [numeric : Read / Write]

The number of decimal places to display. Allowed values: <= 20 Default value: 0 Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCellRendererSpin.html

See Also

GtkCellRendererText GtkSpinButton

[Package RGtk2 version 2.20.31 Index]