GtkCellRendererSpinner {RGtk2}	R Documentation

GtkCellRendererSpinner

Description

Renders a spinning animation in a cell

Methods and Functions

gtkCellRendererSpinnerNew()
gtkCellRendererSpinner()

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkCellRenderer
                     +----GtkCellRendererSpinner

Detailed Description

GtkCellRendererSpinner renders a spinning animation in a cell, very similar to GtkSpinner. It can often be used as an alternative to a GtkCellRendererProgress for displaying indefinite activity, instead of actual progress.

To start the animation in a cell, set the "active" property to TRUE and increment the "pulse" property at regular intervals. The usual way to set the cell renderer properties for each cell is to bind them to columns in your tree model using e.g. gtkTreeViewColumnAddAttribute.

Structures

GtkCellRendererSpinner

undocumented

Convenient Construction

gtkCellRendererSpinner is the equivalent of gtkCellRendererSpinnerNew.

Properties

active [logical : Read / Write]

Whether the spinner is active (ie. shown) in the cell. Default value: FALSE

pulse [numeric : Read / Write]

Pulse of the spinner. Increment this value to draw the next frame of the spinner animation. Usually, you would update this value in a timeout.

The GtkSpinner widget draws one full cycle of the animation per second by default. You can learn about the number of frames used by the theme by looking at the "num-steps" style property and the duration of the cycle by looking at "cycle-duration". Default value: 0 Since 2.20

size [GtkIconSize : Read / Write]

The GtkIconSize value that specifies the size of the rendered spinner. Default value: GTK_ICON_SIZE_MENU Since 2.20

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCellRendererSpinner.html

[Package RGtk2 version 2.20.31 Index]