GtkCellRendererText {RGtk2}	R Documentation

GtkCellRendererText

Description

Renders text in a cell

Methods and Functions

gtkCellRendererTextNew()
gtkCellRendererTextSetFixedHeightFromFont(object, number.of.rows)
gtkCellRendererText()

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkCellRenderer
                     +----GtkCellRendererText
                           +----GtkCellRendererAccel
                           +----GtkCellRendererCombo
                           +----GtkCellRendererSpin

Detailed Description

A GtkCellRendererText renders a given text in its cell, using the font, color and style information provided by its properties. The text will be ellipsized if it is too long and the ellipsize property allows it.

If the mode is GTK_CELL_RENDERER_MODE_EDITABLE, the GtkCellRendererText allows to edit its text using an entry.

Structures

GtkCellRendererText

undocumented

Convenient Construction

gtkCellRendererText is the equivalent of gtkCellRendererTextNew.

Signals

edited(renderer, path, new.text, user.data)

This signal is emitted after renderer has been edited.

It is the responsibility of the application to update the model and store new.text at the position indicated by path.

renderer

the object which received the signal

path

the path identifying the edited cell

new.text

the new text

user.data

user data set when the signal handler was connected.

Properties

align-set [logical : Read / Write]

Whether this tag affects the alignment mode. Default value: FALSE

alignment [PangoAlignment : Read / Write]

Specifies how to align the lines of text with respect to each other.

Note that this property describes how to align the lines of text in case there are several of them. The "xalign" property of GtkCellRenderer, on the other hand, sets the horizontal alignment of the whole text. Default value: PANGO_ALIGN_LEFT Since 2.10

attributes [PangoAttrList : * : Read / Write]

A list of style attributes to apply to the text of the renderer.

background [character : * : Write]

Background color as a string. Default value: NULL

background-gdk [GdkColor : * : Read / Write]

Background color as a GdkColor.

background-set [logical : Read / Write]

Whether this tag affects the background color. Default value: FALSE

editable [logical : Read / Write]

Whether the text can be modified by the user. Default value: FALSE

editable-set [logical : Read / Write]

Whether this tag affects text editability. Default value: FALSE

ellipsize [PangoEllipsizeMode : Read / Write]

Specifies the preferred place to ellipsize the string, if the cell renderer does not have enough room to display the entire string. Setting it to PANGO_ELLIPSIZE_NONE turns off ellipsizing. See the wrap-width property for another way of making the text fit into a given width. Default value: PANGO_ELLIPSIZE_NONE Since 2.6

ellipsize-set [logical : Read / Write]

Whether this tag affects the ellipsize mode. Default value: FALSE

family [character : * : Read / Write]

Name of the font family, e.g. Sans, Helvetica, Times, Monospace. Default value: NULL

family-set [logical : Read / Write]

Whether this tag affects the font family. Default value: FALSE

font [character : * : Read / Write]

Font description as a string, e.g. "Sans Italic 12". Default value: NULL

font-desc [PangoFontDescription : * : Read / Write]

Font description as a PangoFontDescription struct.

foreground [character : * : Write]

Foreground color as a string. Default value: NULL

foreground-gdk [GdkColor : * : Read / Write]

Foreground color as a GdkColor.

foreground-set [logical : Read / Write]

Whether this tag affects the foreground color. Default value: FALSE

language [character : * : Read / Write]

The language this text is in, as an ISO code. Pango can use this as a hint when rendering the text. If you don't understand this parameter, you probably don't need it. Default value: NULL

language-set [logical : Read / Write]

Whether this tag affects the language the text is rendered as. Default value: FALSE

markup [character : * : Write]

Marked up text to render. Default value: NULL

rise [integer : Read / Write]

Offset of text above the baseline (below the baseline if rise is negative). Allowed values: >= -2147483647 Default value: 0

rise-set [logical : Read / Write]

Whether this tag affects the rise. Default value: FALSE

scale [numeric : Read / Write]

Font scaling factor. Allowed values: >= 0 Default value: 1

scale-set [logical : Read / Write]

Whether this tag scales the font size by a factor. Default value: FALSE

single-paragraph-mode [logical : Read / Write]

Whether or not to keep all text in a single paragraph. Default value: FALSE

size [integer : Read / Write]

Font size. Allowed values: >= 0 Default value: 0

size-points [numeric : Read / Write]

Font size in points. Allowed values: >= 0 Default value: 0

size-set [logical : Read / Write]

Whether this tag affects the font size. Default value: FALSE

stretch [PangoStretch : Read / Write]

Font stretch. Default value: PANGO_STRETCH_NORMAL

stretch-set [logical : Read / Write]

Whether this tag affects the font stretch. Default value: FALSE

strikethrough [logical : Read / Write]

Whether to strike through the text. Default value: FALSE

strikethrough-set [logical : Read / Write]

Whether this tag affects strikethrough. Default value: FALSE

style [PangoStyle : Read / Write]

Font style. Default value: PANGO_STYLE_NORMAL

style-set [logical : Read / Write]

Whether this tag affects the font style. Default value: FALSE

text [character : * : Read / Write]

Text to render. Default value: NULL

underline [PangoUnderline : Read / Write]

Style of underline for this text. Default value: PANGO_UNDERLINE_NONE

underline-set [logical : Read / Write]

Whether this tag affects underlining. Default value: FALSE

variant [PangoVariant : Read / Write]

Font variant. Default value: PANGO_VARIANT_NORMAL

variant-set [logical : Read / Write]

Whether this tag affects the font variant. Default value: FALSE

weight [integer : Read / Write]

Font weight. Allowed values: >= 0 Default value: 400

weight-set [logical : Read / Write]

Whether this tag affects the font weight. Default value: FALSE

width-chars [integer : Read / Write]

The desired width of the cell, in characters. If this property is set to -1, the width will be calculated automatically, otherwise the cell will request either 3 characters or the property value, whichever is greater. Allowed values: >= -1 Default value: -1 Since 2.6

wrap-mode [PangoWrapMode : Read / Write]

Specifies how to break the string into multiple lines, if the cell renderer does not have enough room to display the entire string. This property has no effect unless the wrap-width property is set. Default value: PANGO_WRAP_CHAR Since 2.8

wrap-width [integer : Read / Write]

Specifies the width at which the text is wrapped. The wrap-mode property can be used to influence at what character positions the line breaks can be placed. Setting wrap-width to -1 turns wrapping off. Allowed values: >= -1 Default value: -1 Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCellRendererText.html

[Package RGtk2 version 2.20.31 Index]