GtkCellRendererToggle {RGtk2}	R Documentation

GtkCellRendererToggle

Description

Renders a toggle button in a cell

Methods and Functions

gtkCellRendererToggleNew()
gtkCellRendererToggleGetRadio(object)
gtkCellRendererToggleSetRadio(object, radio)
gtkCellRendererToggleGetActive(object)
gtkCellRendererToggleSetActive(object, setting)
gtkCellRendererToggleGetActivatable(object)
gtkCellRendererToggleSetActivatable(object, setting)
gtkCellRendererToggle()

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkCellRenderer
                     +----GtkCellRendererToggle

Detailed Description

GtkCellRendererToggle renders a toggle button in a cell. The button is drawn as a radio- or checkbutton, depending on the radio property. When activated, it emits the toggled signal.

Structures

GtkCellRendererToggle

undocumented

Convenient Construction

gtkCellRendererToggle is the equivalent of gtkCellRendererToggleNew.

Signals

toggled(cell.renderer, path, user.data)

The ::toggled signal is emitted when the cell is toggled.

cell.renderer

the object which received the signal

path

string representation of GtkTreePath describing the event location

user.data

user data set when the signal handler was connected.

Properties

activatable [logical : Read / Write]

The toggle button can be activated. Default value: TRUE

active [logical : Read / Write]

The toggle state of the button. Default value: FALSE

inconsistent [logical : Read / Write]

The inconsistent state of the button. Default value: FALSE

indicator-size [integer : Read / Write]

Size of check or radio indicator. Allowed values: >= 0 Default value: 13

radio [logical : Read / Write]

Draw the toggle button as a radio button. Default value: FALSE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCellRendererToggle.html

[Package RGtk2 version 2.20.31 Index]