GtkCellView {RGtk2}	R Documentation

GtkCellView

Description

A widget displaying a single row of a GtkTreeModel

Methods and Functions

gtkCellViewNew(show = TRUE)
gtkCellViewNewWithText(text)
gtkCellViewNewWithMarkup(markup)
gtkCellViewNewWithPixbuf(pixbuf)
gtkCellViewSetModel(object, model = NULL)
gtkCellViewGetModel(object)
gtkCellViewSetDisplayedRow(object, path = NULL)
gtkCellViewGetDisplayedRow(object)
gtkCellViewGetSizeOfRow(object, path)
gtkCellViewSetBackgroundColor(object, color)
gtkCellViewGetCellRenderers(object)
gtkCellViewGetCellRenderers(object)
gtkCellView(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkCellView

Interfaces

GtkCellView implements AtkImplementorIface, GtkBuildable and GtkCellLayout.

Detailed Description

A GtkCellView displays a single row of a GtkTreeModel, using cell renderers just like GtkTreeView. GtkCellView doesn't support some of the more complex features of GtkTreeView, like cell editing and drag and drop.

Structures

GtkCellView

undocumented

Convenient Construction

gtkCellView is the equivalent of gtkCellViewNew.

Properties

background [character : * : Write]

Background color as a string. Default value: NULL

background-gdk [GdkColor : * : Read / Write]

Background color as a GdkColor.

background-set [logical : Read / Write]

Whether this tag affects the background color. Default value: FALSE

model [GtkTreeModel : * : Read / Write]

The model for cell view

since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCellView.html

[Package RGtk2 version 2.20.31 Index]