GtkCheckButton {RGtk2}	R Documentation

GtkCheckButton

Description

Create widgets with a discrete toggle button

Methods and Functions

gtkCheckButtonNew(show = TRUE)
gtkCheckButtonNewWithLabel(label, show = TRUE)
gtkCheckButtonNewWithMnemonic(label, show = TRUE)
gtkCheckButton(label, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkButton
                                       +----GtkToggleButton
                                             +----GtkCheckButton
                                                   +----GtkRadioButton

Interfaces

GtkCheckButton implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

A GtkCheckButton places a discrete GtkToggleButton next to a widget, (usually a GtkLabel). See the section on GtkToggleButton widgets for more information about toggle/check buttons.

The important signal ('toggled') is also inherited from GtkToggleButton.

Structures

GtkCheckButton

toggle_button is a GtkToggleButton representing the actual toggle button that composes the check button.

Convenient Construction

gtkCheckButton is the result of collapsing the constructors of GtkCheckButton (gtkCheckButtonNew, gtkCheckButtonNewWithLabel, gtkCheckButtonNewWithMnemonic) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Style Properties

indicator-size [integer : Read]

Size of check or radio indicator. Allowed values: >= 0 Default value: 13

indicator-spacing [integer : Read]

Spacing around check or radio indicator. Allowed values: >= 0 Default value: 2

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCheckButton.html

[Package RGtk2 version 2.20.31 Index]