GtkColorSelection {RGtk2}	R Documentation

GtkColorSelection

Description

A widget used to select a color

Methods and Functions

gtkColorSelectionNew(show = TRUE)
gtkColorSelectionSetUpdatePolicy(object, policy)
gtkColorSelectionSetHasOpacityControl(object, has.opacity)
gtkColorSelectionGetHasOpacityControl(object)
gtkColorSelectionSetHasPalette(object, has.palette)
gtkColorSelectionGetHasPalette(object)
gtkColorSelectionGetCurrentAlpha(object)
gtkColorSelectionSetCurrentAlpha(object, alpha)
gtkColorSelectionGetCurrentColor(object)
gtkColorSelectionSetCurrentColor(object, color)
gtkColorSelectionGetPreviousAlpha(object)
gtkColorSelectionSetPreviousAlpha(object, alpha)
gtkColorSelectionGetPreviousColor(object, color)
gtkColorSelectionSetPreviousColor(object, color)
gtkColorSelectionIsAdjusting(object)
gtkColorSelectionPaletteFromString(str)
gtkColorSelectionPaletteToString(colors)
gtkColorSelectionSetChangePaletteHook(func)
gtkColorSelectionSetChangePaletteWithScreenHook(func)
gtkColorSelectionSetColor(object, color)
gtkColorSelectionGetColor(object)
gtkColorSelection(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBox
                                 +----GtkVBox
                                       +----GtkColorSelection

Interfaces

GtkColorSelection implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

The GtkColorSelection is a widget that is used to select a color. It consists of a color wheel and number of sliders and entry boxes for color parameters such as hue, saturation, value, red, green, blue, and opacity. It is found on the standard color selection dialog box GtkColorSelectionDialog.

Structures

GtkColorSelection

The GtkColorSelection struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkColorSelection is the equivalent of gtkColorSelectionNew.

User Functions

GtkColorSelectionChangePaletteFunc()

undocumented

GtkColorSelectionChangePaletteWithScreenFunc()

Since 2.2

Signals

color-changed(colorselection, user.data)

This signal is emitted when the color changes in the GtkColorSelection according to its update policy.

colorselection

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

current-alpha [numeric : Read / Write]

The current opacity value (0 fully transparent, 65535 fully opaque). Allowed values: <= 65535 Default value: 65535

current-color [GdkColor : * : Read / Write]

The current color.

has-opacity-control [logical : Read / Write]

Whether the color selector should allow setting opacity. Default value: FALSE

has-palette [logical : Read / Write]

Whether a palette should be used. Default value: FALSE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkColorSelection.html

[Package RGtk2 version 2.20.31 Index]