GtkColorSelectionDialog {RGtk2}	R Documentation

GtkColorSelectionDialog

Description

A standard dialog box for selecting a color

Methods and Functions

gtkColorSelectionDialogNew(title = NULL, show = TRUE)
gtkColorSelectionDialogGetColorSelection(object)
gtkColorSelectionDialog(title = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkWindow
                                       +----GtkDialog
                                             +----GtkColorSelectionDialog

Interfaces

GtkColorSelectionDialog implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkColorSelectionDialog provides a standard dialog which allows the user to select a color much like the GtkFileSelection provides a standard dialog for file selection.

GtkColorSelectionDialog as GtkBuildable

The GtkColorSelectionDialog implementation of the GtkBuildable interface exposes the embedded GtkColorSelection as internal child with the name "color_selection". It also exposes the buttons with the names "ok_button", "cancel_button" and "help_button".

Structures

GtkColorSelectionDialog

The GtkColorSelectionDialog struct contains the following fields. (These fields should be considered read-only. They should never be set by an application.)

GtkWidget *colorsel; 	 The GtkColorSelection widget contained within the dialog. Use this widget and its gtkColorSelectionGetCurrentColor function to gain access to the selected color. Connect a handler for this widget's color_changed signal to be notified when the color changes.

GtkWidget *ok_button; 	 The OK button widget contained within the dialog. Connect a handler for the clicked event.

GtkWidget *cancel_button; 	 The cancel button widget contained within the dialog. Connect a handler for the clicked event.

GtkWidget *help_button; 	 The help button widget contained within the dialog. Connect a handler for the clicked event.

Convenient Construction

gtkColorSelectionDialog is the equivalent of gtkColorSelectionDialogNew.

Properties

cancel-button [GtkWidget : * : Read]

The cancel button of the dialog.

color-selection [GtkWidget : * : Read]

The color selection embedded in the dialog.

help-button [GtkWidget : * : Read]

The help button of the dialog.

ok-button [GtkWidget : * : Read]

The OK button of the dialog.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkColorSelectionDialog.html

[Package RGtk2 version 2.20.31 Index]