GtkCurve {RGtk2}	R Documentation

GtkCurve

Description

Allows direct editing of a curve

Methods and Functions

gtkCurveNew(show = TRUE)
gtkCurveReset(object)
gtkCurveSetGamma(object, gamma)
gtkCurveSetRange(object, min.x, max.x, min.y, max.y)
gtkCurveGetVector(object, veclen)
gtkCurveSetVector(object, vector)
gtkCurveSetCurveType(object, type)
gtkCurve(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkDrawingArea
                           +----GtkCurve

Interfaces

GtkCurve implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkCurve widget allows the user to edit a curve covering a range of values. It is typically used to fine-tune color balances in graphics applications like the Gimp.

The GtkCurve widget has 3 modes of operation - spline, linear and free. In spline mode the user places points on the curve which are automatically connected together into a smooth curve. In linear mode the user places points on the curve which are connected by straight lines. In free mode the user can draw the points of the curve freely, and they are not connected at all.

As of GTK+ 2.20, GtkCurve has been deprecated since it is too specialized.

Structures

GtkCurve

WARNING: GtkCurve is deprecated and should not be used in newly-written code. The GtkCurve struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkCurve is the equivalent of gtkCurveNew.

Signals

curve-type-changed(curve, user.data)

Emitted when the curve type has been changed. The curve type can be changed explicitly with a call to gtkCurveSetCurveType. It is also changed as a side-effect of calling gtkCurveReset or gtkCurveSetGamma.

curve

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

curve-type [GtkCurveType : Read / Write]

Is this curve linear, spline interpolated, or free-form. Default value: GTK_CURVE_TYPE_SPLINE

max-x [numeric : Read / Write]

Maximum possible X value. Default value: 1

max-y [numeric : Read / Write]

Maximum possible value for Y. Default value: 1

min-x [numeric : Read / Write]

Minimum possible value for X. Default value: 0

min-y [numeric : Read / Write]

Minimum possible value for Y. Default value: 0

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkCurve.html

[Package RGtk2 version 2.20.31 Index]