GtkDialog {RGtk2}	R Documentation

GtkDialog

Description

Create popup windows

Methods and Functions

gtkDialogNew(show = TRUE)
gtkDialogNewWithButtons(title = NULL, parent = NULL, flags = 0, ..., show = TRUE)
gtkDialogRun(object)
gtkDialogResponse(object, response.id)
gtkDialogAddButton(object, button.text, response.id)
gtkDialogAddButtons(object, ...)
gtkDialogAddActionWidget(object, child, response.id)
gtkDialogGetHasSeparator(object)
gtkDialogSetDefaultResponse(object, response.id)
gtkDialogSetHasSeparator(object, setting)
gtkDialogSetResponseSensitive(object, response.id, setting)
gtkDialogGetResponseForWidget(object, widget)
gtkDialogGetWidgetForResponse(object, response.id)
gtkDialogGetActionArea(object)
gtkDialogGetContentArea(object)
gtkAlternativeDialogButtonOrder(object)
gtkDialogSetAlternativeButtonOrder(object, ...)
gtkDialogSetAlternativeButtonOrderFromArray(object, new.order)
gtkDialog(title = NULL, parent = NULL, flags = 0, ..., show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkWindow
                                       +----GtkDialog
                                             +----GtkAboutDialog
                                             +----GtkColorSelectionDialog
                                             +----GtkFileChooserDialog
                                             +----GtkFileSelection
                                             +----GtkFontSelectionDialog
                                             +----GtkInputDialog
                                             +----GtkMessageDialog
                                             +----GtkPageSetupUnixDialog
                                             +----GtkPrintUnixDialog
                                             +----GtkRecentChooserDialog

Interfaces

GtkDialog implements AtkImplementorIface and GtkBuildable.

Detailed Description

Dialog boxes are a convenient way to prompt the user for a small amount of input, e.g. to display a message, ask a question, or anything else that does not require extensive effort on the user's part.

GTK+ treats a dialog as a window split vertically. The top section is a GtkVBox, and is where widgets such as a GtkLabel or a GtkEntry should be packed. The bottom area is known as the action_area. This is generally used for packing buttons into the dialog which may perform functions such as cancel, ok, or apply. The two areas are separated by a GtkHSeparator. GtkDialog boxes are created with a call to gtkDialogNew or gtkDialogNewWithButtons. gtkDialogNewWithButtons is recommended; it allows you to set the dialog title, some convenient flags, and add simple buttons.

If 'dialog' is a newly created dialog, the two primary areas of the window can be accessed through gtkDialogGetContentArea and gtkDialogGetActionArea, as can be seen from the example, below.

A 'modal' dialog (that is, one which freezes the rest of the application from user input), can be created by calling gtkWindowSetModal on the dialog. Use the gtkWindow() function to cast the widget returned from gtkDialogNew into a GtkWindow. When using gtkDialogNewWithButtons you can also pass the GTK_DIALOG_MODAL flag to make a dialog modal.

If you add buttons to GtkDialog using gtkDialogNewWithButtons, gtkDialogAddButton, gtkDialogAddButtons, or gtkDialogAddActionWidget, clicking the button will emit a signal called "response" with a response ID that you specified. GTK+ will never assign a meaning to positive response IDs; these are entirely user-defined. But for convenience, you can use the response IDs in the GtkResponseType enumeration (these all have values less than zero). If a dialog receives a delete event, the "response" signal will be emitted with a response ID of GTK_RESPONSE_DELETE_EVENT.

If you want to block waiting for a dialog to return before returning control flow to your code, you can call gtkDialogRun. This function enters a recursive main loop and waits for the user to respond to the dialog, returning the response ID corresponding to the button the user clicked.

For the simple dialog in the following example, in reality you'd probably use GtkMessageDialog to save yourself some effort. But you'd need to create the dialog contents manually if you had more than a simple message in the dialog.

Simple GtkDialog usage.

# Function to open a dialog box displaying the message provided.
quick_message <- function(message) {
  ## Create the widgets 
   
  dialog <- gtkDialog("Message", NULL, "destroy-with-parent",
                      "gtk-ok", GtkResponseType["none"],
                      show = FALSE)
  label <- gtkLabel(message)
   
  ## Ensure that the dialog box is destroyed when the user responds.
   
  gSignalConnect(dialog, "response", gtkWidgetDestroy)

  ## Add the label, and show everything we've added to the dialog.

  dialog[["vbox"]]$add(label)
  dialog$showAll()
}

GtkDialog as GtkBuildable

The GtkDialog implementation of the GtkBuildable interface exposes the vbox and action.area as internal children with the names "vbox" and "action_area".

GtkDialog supports a custom <action-widgets> element, which can contain multiple <action-widget> elements. The "response" attribute specifies a numeric response, and the content of the element is the id of widget (which should be a child of the dialogs action.area).

A GtkDialog UI definition fragment.

<object class="GtkDialog" id="dialog1">
  <child internal-child="vbox">"
    <object class="GtkVBox" id="vbox">
      <child internal-child="action_area">
        <object class="GtkHButtonBox" id="button_box">
          <child>
            <object class="GtkButton" id="button_cancel"/>
          </child>
          <child>
            <object class="GtkButton" id="button_ok"/>
          </child>
        </object>
      </child>
    </object>
  </child>
  <action-widgets>
    <action-widget response="3">button_ok</action-widget>
    <action-widget response="-5">button_cancel</action-widget>
  </action-widgets>
</object>

Structures

GtkDialog

vbox is a GtkVBox - the main part of the dialog box. action_area is a GtkHButtonBox packed below the dividing GtkHSeparator in the dialog. It is treated exactly the same as any other GtkHButtonBox.

vbox

[GtkWidget]

actionArea

[GtkWidget]

Convenient Construction

gtkDialog is the result of collapsing the constructors of GtkDialog (gtkDialogNew, gtkDialogNewWithButtons) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Enums and Flags

GtkDialogFlags

Flags used to influence dialog construction.

modal

Make the constructed dialog modal, see gtkWindowSetModal.

destroy-with-parent

Destroy the dialog when its parent is destroyed, see gtkWindowSetDestroyWithParent.

no-separator

Don't put a separator between the action area and the dialog content.

GtkResponseType

Predefined values for use as response ids in gtkDialogAddButton. All predefined values are negative, GTK+ leaves positive values for application-defined response ids.

none

Returned if an action widget has no response id, or if the dialog gets programmatically hidden or destroyed.

reject

Generic response id, not used by GTK+ dialogs.

accept

Generic response id, not used by GTK+ dialogs.

delete-event

Returned if the dialog is deleted.

ok

Returned by OK buttons in GTK+ dialogs.

cancel

Returned by Cancel buttons in GTK+ dialogs.

close

Returned by Close buttons in GTK+ dialogs.

yes

Returned by Yes buttons in GTK+ dialogs.

no

Returned by No buttons in GTK+ dialogs.

apply

Returned by Apply buttons in GTK+ dialogs.

help

Returned by Help buttons in GTK+ dialogs.

Signals

close(user.data)

The ::close signal is a keybinding signal which gets emitted when the user uses a keybinding to close the dialog.

The default binding for this signal is the Escape key.

user.data

user data set when the signal handler was connected.

response(dialog, response.id, user.data)

Emitted when an action widget is clicked, the dialog receives a delete event, or the application programmer calls gtkDialogResponse. On a delete event, the response ID is GTK_RESPONSE_DELETE_EVENT. Otherwise, it depends on which action widget was clicked.

dialog

the object on which the signal is emitted

response.id

the response ID

user.data

user data set when the signal handler was connected.

Properties

has-separator [logical : Read / Write]

When TRUE, the dialog has a separator bar above its buttons. Default value: TRUE

Style Properties

action-area-border [integer : Read]

Width of border around the button area at the bottom of the dialog. Allowed values: >= 0 Default value: 5

button-spacing [integer : Read]

Spacing between buttons. Allowed values: >= 0 Default value: 6

content-area-border [integer : Read]

Width of border around the main dialog area. Allowed values: >= 0 Default value: 2

content-area-spacing [integer : Read]

The default spacing used between elements of the content area of the dialog, as returned by gtkDialogGetContentArea, unless gtkBoxSetSpacing was called on that widget directly. Allowed values: >= 0 Default value: 0 Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkDialog.html

[Package RGtk2 version 2.20.31 Index]