GtkDrawingArea {RGtk2}	R Documentation

GtkDrawingArea

Description

A widget for custom user interface elements

Methods and Functions

gtkDrawingAreaNew(show = TRUE)
gtkDrawingAreaSize(object, width, height)
gtkDrawingArea(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkDrawingArea
                           +----GtkCurve
                           +----GtkSpinner

Interfaces

GtkDrawingArea implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkDrawingArea widget is used for creating custom user interface elements. It's essentially a blank widget; you can draw on widget->window. After creating a drawing area, the application may want to connect to:

Mouse and button press signals to respond to input from the user. (Use gtkWidgetAddEvents to enable events you wish to receive.)

The "realize" signal to take any necessary actions when the widget is instantiated on a particular display. (Create GDK resources in response to this signal.)

The "configure_event" signal to take any necessary actions when the widget changes size.

The "expose_event" signal to handle redrawing the contents of the widget.

The following code portion demonstrates using a drawing area to display a circle in the normal widget foreground color. Note that GDK automatically clears the exposed area to the background color before sending the expose event, and that drawing is implicitly clipped to the exposed area.

Simple GtkDrawingArea usage.

expose_event_callback <- function(widget, event, data) {
  gdkDrawArc(widget[["window"]],
             widget[["style"]][["fgGc"]][[widget[["state"]]+1]],
             TRUE, 0, 0, widget[["allocation"]]$width,
             widget[["allocation"]]$height, 0, 64 * 360)
  return(TRUE)
}
[...]
  
drawing_area = gtkDrawingArea()
drawing_area$setSizeRequest(100, 100)
gSignalConnect(drawing_area, "expose_event", expose_event_callback)

Expose events are normally delivered when a drawing area first comes onscreen, or when it's covered by another window and then uncovered (exposed). You can also force an expose event by adding to the "damage region" of the drawing area's window; gtkWidgetQueueDrawArea and gdkWindowInvalidateRect are equally good ways to do this. You'll then get an expose event for the invalid region.

The available routines for drawing are documented on the GDK Drawing Primitives page. See also gdkDrawPixbuf for drawing a GdkPixbuf.

To receive mouse events on a drawing area, you will need to enable them with gtkWidgetAddEvents. To receive keyboard events, you will need to set the GTK_CAN_FOCUS flag on the drawing area, and should probably draw some user-visible indication that the drawing area is focused. Use the gtkHasFocus() function in your expose event handler to decide whether to draw the focus indicator. See gtkPaintFocus for one way to draw focus.

Structures

GtkDrawingArea

The GtkDrawingArea struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkDrawingArea is the equivalent of gtkDrawingAreaNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkDrawingArea.html

See Also

GtkImage GdkPixmap GtkImage GdkPixmap GtkImage

[Package RGtk2 version 2.20.31 Index]