GtkEntryBuffer {RGtk2}	R Documentation

GtkEntryBuffer

Description

Text buffer for GtkEntry

Methods and Functions

gtkEntryBufferNew(initial.chars = NULL, n.initial.chars = -1)
gtkEntryBufferGetText(object)
gtkEntryBufferSetText(object, chars, n.chars)
gtkEntryBufferGetBytes(object)
gtkEntryBufferGetLength(object)
gtkEntryBufferGetMaxLength(object)
gtkEntryBufferSetMaxLength(object, max.length)
gtkEntryBufferInsertText(object, position, chars, n.chars)
gtkEntryBufferDeleteText(object, position, n.chars)
gtkEntryBufferEmitDeletedText(object, position, n.chars)
gtkEntryBufferEmitInsertedText(object, position, chars, n.chars)
gtkEntryBuffer(initial.chars = NULL, n.initial.chars = -1)

Hierarchy

GObject
   +----GtkEntryBuffer

Detailed Description

The GtkEntryBuffer class contains the actual text displayed in a GtkEntry widget.

A single GtkEntryBuffer object can be shared by multiple GtkEntry widgets which will then share the same text content, but not the cursor position, visibility attributes, icon etc. GtkEntryBuffer may be derived from. Such a derived class might allow text to be stored in an alternate location, such as non-pageable memory, useful in the case of important passwords. Or a derived class could integrate with an application's concept of undo/redo.

Structures

GtkEntryBuffer

undocumented

Convenient Construction

gtkEntryBuffer is the equivalent of gtkEntryBufferNew.

Signals

deleted-text(entrybuffer, user.data)

undocumented

entrybuffer

the object which received the signal.

user.data

user data set when the signal handler was connected.

inserted-text(entrybuffer, user.data)

undocumented

entrybuffer

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

length [numeric : Read]

The length (in characters) of the text in buffer. Allowed values: <= 65535 Default value: 0 Since 2.18

max-length [integer : Read / Write]

The maximum length (in characters) of the text in the buffer. Allowed values: [0,65535] Default value: 0 Since 2.18

text [character : * : Read / Write]

The contents of the buffer. Default value: "" Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkEntryBuffer.html

[Package RGtk2 version 2.20.31 Index]