GtkEntryCompletion {RGtk2}	R Documentation

GtkEntryCompletion

Description

Completion functionality for GtkEntry

Methods and Functions

gtkEntryCompletionNew()
gtkEntryCompletionGetEntry(object)
gtkEntryCompletionSetModel(object, model = NULL)
gtkEntryCompletionGetModel(object)
gtkEntryCompletionSetMatchFunc(object, func, func.data = NULL)
gtkEntryCompletionSetMinimumKeyLength(object, length)
gtkEntryCompletionGetMinimumKeyLength(object)
gtkEntryCompletionComplete(object)
gtkEntryCompletionGetCompletionPrefix(object)
gtkEntryCompletionInsertPrefix(object)
gtkEntryCompletionInsertActionText(object, index, text)
gtkEntryCompletionInsertActionMarkup(object, index, markup)
gtkEntryCompletionDeleteAction(object, index)
gtkEntryCompletionSetTextColumn(object, column)
gtkEntryCompletionGetTextColumn(object)
gtkEntryCompletionSetInlineCompletion(object, inline.completion)
gtkEntryCompletionGetInlineCompletion(object)
gtkEntryCompletionSetInlineSelection(object, inline.selection)
gtkEntryCompletionGetInlineSelection(object)
gtkEntryCompletionSetPopupCompletion(object, popup.completion)
gtkEntryCompletionGetPopupCompletion(object)
gtkEntryCompletionSetPopupSetWidth(object, popup.set.width)
gtkEntryCompletionGetPopupSetWidth(object)
gtkEntryCompletionSetPopupSingleMatch(object, popup.single.match)
gtkEntryCompletionGetPopupSingleMatch(object)
gtkEntryCompletion()

Hierarchy

GObject
   +----GtkEntryCompletion

Interfaces

GtkEntryCompletion implements GtkCellLayout and GtkBuildable.

Detailed Description

GtkEntryCompletion is an auxiliary object to be used in conjunction with GtkEntry to provide the completion functionality. It implements the GtkCellLayout interface, to allow the user to add extra cells to the GtkTreeView with completion matches.

"Completion functionality" means that when the user modifies the text in the entry, GtkEntryCompletion checks which rows in the model match the current content of the entry, and displays a list of matches. By default, the matching is done by comparing the entry text case-insensitively against the text column of the model (see gtkEntryCompletionSetTextColumn), but this can be overridden with a custom match function (see gtkEntryCompletionSetMatchFunc).

When the user selects a completion, the content of the entry is updated. By default, the content of the entry is replaced by the text column of the model, but this can be overridden by connecting to the ::match-selected signal and updating the entry in the signal handler. Note that you should return TRUE from the signal handler to suppress the default behaviour.

To add completion functionality to an entry, use gtkEntrySetCompletion.

In addition to regular completion matches, which will be inserted into the entry when they are selected, GtkEntryCompletion also allows to display "actions" in the popup window. Their appearance is similar to menuitems, to differentiate them clearly from completion strings. When an action is selected, the ::action-activated signal is emitted.

Structures

GtkEntryCompletion

The GtkEntryCompletion struct contains only private data.

Convenient Construction

gtkEntryCompletion is the equivalent of gtkEntryCompletionNew.

User Functions

GtkEntryCompletionMatchFunc(completion, key, iter, user.data)

A function which decides whether the row indicated by iter matches a given key, and should be displayed as a possible completion for key. Note that key is normalized and case-folded (see gUtf8Normalize() and gUtf8Casefold()). If this is not appropriate, match functions have access to the unmodified key via gtk_entry_get_text (GTK_ENTRY (gtk_entry_completion_get_entry ( ))).

completion

the GtkEntryCompletion

key

the string to match, normalized and case-folded

iter

a GtkTreeIter indicating the row to match

user.data

user data given to gtkEntryCompletionSetMatchFunc

Returns: [logical] TRUE if iter should be displayed as a possible completion for key

Signals

action-activated(widget, index, user.data)

Gets emitted when an action is activated. Since 2.4

widget

the object which received the signal

index

the index of the activated action

user.data

user data set when the signal handler was connected.

cursor-on-match(widget, model, iter, user.data)

Gets emitted when a match from the cursor is on a match of the list.The default behaviour is to replace the contents of the entry with the contents of the text column in the row pointed to by iter. Since 2.12

widget

the object which received the signal

model

the GtkTreeModel containing the matches

iter

a GtkTreeIter positioned at the selected match

user.data

user data set when the signal handler was connected.

Returns: [logical] TRUE if the signal has been handled

insert-prefix(widget, prefix, user.data)

Gets emitted when the inline autocompletion is triggered. The default behaviour is to make the entry display the whole prefix and select the newly inserted part.

Applications may connect to this signal in order to insert only a smaller part of the prefix into the entry - e.g. the entry used in the GtkFileChooser inserts only the part of the prefix up to the next '/'. Since 2.6

widget

the object which received the signal

prefix

the common prefix of all possible completions

user.data

user data set when the signal handler was connected.

Returns: [logical] TRUE if the signal has been handled

match-selected(widget, model, iter, user.data)

Gets emitted when a match from the list is selected. The default behaviour is to replace the contents of the entry with the contents of the text column in the row pointed to by iter. Since 2.4

widget

the object which received the signal

model

the GtkTreeModel containing the matches

iter

a GtkTreeIter positioned at the selected match

user.data

user data set when the signal handler was connected.

Returns: [logical] TRUE if the signal has been handled

Properties

inline-completion [logical : Read / Write]

Determines whether the common prefix of the possible completions should be inserted automatically in the entry. Note that this requires text-column to be set, even if you are using a custom match function. Default value: FALSE Since 2.6

inline-selection [logical : Read / Write]

Determines whether the possible completions on the popup will appear in the entry as you navigate through them. Default value: FALSE Since 2.12

minimum-key-length [integer : Read / Write]

Minimum length of the search key in order to look up matches. Allowed values: >= 0 Default value: 1

model [GtkTreeModel : * : Read / Write]

The model to find matches in.

popup-completion [logical : Read / Write]

Determines whether the possible completions should be shown in a popup window. Default value: TRUE Since 2.6

popup-set-width [logical : Read / Write]

Determines whether the completions popup window will be resized to the width of the entry. Default value: TRUE Since 2.8

popup-single-match [logical : Read / Write]

Determines whether the completions popup window will shown for a single possible completion. You probably want to set this to FALSE if you are using inline completion. Default value: TRUE Since 2.8

text-column [integer : Read / Write]

The column of the model containing the strings. Note that the strings must be UTF-8. Allowed values: >= -1 Default value: -1 Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkEntryCompletion.html

[Package RGtk2 version 2.20.31 Index]