GtkFileChooser {RGtk2}	R Documentation

GtkFileChooser

Description

File chooser interface used by GtkFileChooserWidget and GtkFileChooserDialog

Methods and Functions

gtkFileChooserSetAction(object, action)
gtkFileChooserGetAction(object)
gtkFileChooserSetLocalOnly(object, local.only)
gtkFileChooserGetLocalOnly(object)
gtkFileChooserSetSelectMultiple(object, select.multiple)
gtkFileChooserGetSelectMultiple(object)
gtkFileChooserSetShowHidden(object, show.hidden)
gtkFileChooserGetShowHidden(object)
gtkFileChooserSetDoOverwriteConfirmation(object, do.overwrite.confirmation)
gtkFileChooserGetDoOverwriteConfirmation(object)
gtkFileChooserSetCreateFolders(object, create.folders)
gtkFileChooserGetCreateFolders(object)
gtkFileChooserSetCurrentName(object, name)
gtkFileChooserSelectAll(object)
gtkFileChooserUnselectAll(object)
gtkFileChooserGetUri(object)
gtkFileChooserSetUri(object, uri)
gtkFileChooserSelectUri(object, uri)
gtkFileChooserUnselectUri(object, uri)
gtkFileChooserGetUris(object)
gtkFileChooserSetCurrentFolderUri(object, uri)
gtkFileChooserGetCurrentFolderUri(object)
gtkFileChooserSetPreviewWidget(object, preview.widget)
gtkFileChooserGetPreviewWidget(object)
gtkFileChooserSetPreviewWidgetActive(object, active)
gtkFileChooserGetPreviewWidgetActive(object)
gtkFileChooserSetUsePreviewLabel(object, use.label)
gtkFileChooserGetUsePreviewLabel(object)
gtkFileChooserGetPreviewUri(object)
gtkFileChooserSetExtraWidget(object, extra.widget)
gtkFileChooserGetExtraWidget(object)
gtkFileChooserAddFilter(object, filter)
gtkFileChooserRemoveFilter(object, filter)
gtkFileChooserListFilters(object)
gtkFileChooserSetFilter(object, filter)
gtkFileChooserGetFilter(object)
gtkFileChooserAddShortcutFolderUri(object, uri, .errwarn = TRUE)
gtkFileChooserRemoveShortcutFolderUri(object, uri, .errwarn = TRUE)
gtkFileChooserListShortcutFolderUris(object)
gtkFileChooserGetCurrentFolderFile(object)
gtkFileChooserGetFile(object)
gtkFileChooserGetFiles(object)
gtkFileChooserGetPreviewFile(object)
gtkFileChooserSelectFile(object, file, .errwarn = TRUE)
gtkFileChooserSetCurrentFolderFile(object, file, .errwarn = TRUE)
gtkFileChooserSetFile(object, file, .errwarn = TRUE)
gtkFileChooserUnselectFile(object, file)

Hierarchy

GInterface
   +----GtkFileChooser

Implementations

GtkFileChooser is implemented by GtkFileChooserButton, GtkFileChooserDialog and GtkFileChooserWidget.

Detailed Description

GtkFileChooser is an interface that can be implemented by file selection widgets. In GTK+, the main objects that implement this interface are GtkFileChooserWidget, GtkFileChooserDialog, and GtkFileChooserButton. You do not need to write an object that implements the GtkFileChooser interface unless you are trying to adapt an existing file selector to expose a standard programming interface. GtkFileChooser allows for shortcuts to various places in the filesystem. In the default implementation these are displayed in the left pane. It may be a bit confusing at first that these shortcuts come from various sources and in various flavours, so lets explain the terminology here:

Bookmarks

undocumented

Shortcuts

undocumented

Volumes

undocumented

File Names and Encodings When the user is finished selecting files in a GtkFileChooser, your program can get the selected names either as filenames or as URIs. For URIs, the normal escaping rules are applied if the URI contains non-ASCII characters. However, filenames are always returned in the character set specified by the G_FILENAME_ENCODING environment variable. Please see the Glib documentation for more details about this variable. PLEASE NOTE: This means that while you can pass the result of gtkFileChooserGetFilename to open(2) or fopen(3) , you may not be able to directly set it as the text of a GtkLabel widget unless you convert it first to UTF-8, which all GTK+ widgets expect. You should use gFilenameToUtf8() to convert filenames into strings that can be passed to GTK+ widgets.

Adding a Preview Widget You can add a custom preview widget to a file chooser and then get notification about when the preview needs to be updated. To install a preview widget, use gtkFileChooserSetPreviewWidget. Then, connect to the "update-preview" signal to get notified when you need to update the contents of the preview.

Your callback should use gtkFileChooserGetPreviewFilename to see what needs previewing. Once you have generated the preview for the corresponding file, you must call gtkFileChooserSetPreviewWidgetActive with a boolean flag that indicates whether your callback could successfully generate a preview.

Sample Usage

update_preview_cb <- function(file_chooser, preview)
{
  filename <- file_chooser$getPreviewFilename()

  pixbuf <- gdkPixbuf(file=filename, w=128, h=128)[[1]]
  have_preview <- !is.null(pixbuf)

  preview$setFromPixbuf(pixbuf)

  file_chooser$setPreviewWidgetActive(have_preview)
}

preview <- gtkImage()
my_file_chooser$setPreviewWidget(preview)
gSignalConnect(my_file_chooser, "update-preview", update_preview_cb, preview)

Adding Extra Widgets You can add extra widgets to a file chooser to provide options that are not present in the default design. For example, you can add a toggle button to give the user the option to open a file in read-only mode. You can use gtkFileChooserSetExtraWidget to insert additional widgets in a file chooser.

Sample Usage

toggle <- gtkCheckButton("Open file read-only")
my_file_chooser$setExtraWidget(toggle)

PLEASE NOTE: If you want to set more than one extra widget in the file chooser, you can a container such as a GtkVBox or a GtkTable and include your widgets in it. Then, set the container as the whole extra widget. Key Bindings Internally, GTK+ implements a file chooser's graphical user interface with the private GtkFileChooserDefaultClass. This widget has several key bindings and their associated signals. This section describes the available key binding signals.

GtkFileChooser key binding example The default keys that activate the key-binding signals in GtkFileChooserDefaultClass are as follows:

Signal name 	 Default key combinations

location-popup 	 Control-L (empty path); / (path of "/") PLEASE NOTE: Both the individual / key and the numeric keypad's "divide" key are supported. ; ~ (path of "~")

up-folder 	 Alt-Up PLEASE NOTE: Both the individual Up key and the numeric keypad's Up key are supported. ; Backspace

down-folder 	 Alt-Down

home-folder 	 Alt-Home

desktop-folder 	 Alt-D

quick-bookmark 	 Alt-1 through Alt-0

You can change these defaults to something else. For example, to add a Shift modifier to a few of the default bindings, you can include the following fragment in your ‘.gtkrc-2.0’ file:

binding "my-own-gtkfilechooser-bindings" {
	bind "<Alt><Shift>Up" {
		"up-folder" ()
	}
	bind "<Alt><Shift>Down" {
		"down-folder" ()
	}
	bind "<Alt><Shift>Home" {
		"home-folder" ()
	}
}

class "GtkFileChooserDefault" binding "my-own-gtkfilechooser-bindings"

The "GtkFileChooserDefault::location-popup" signal

   void user_function (GtkFileChooserDefault *chooser,
                       const char            *path,
  gpointer        user_data);

This is used to make the file chooser show a "Location" dialog which the user can use to manually type the name of the file he wishes to select. The path argument is a string that gets put in the text entry for the file name. By default this is bound to Control-L with a path string of "" (the empty string). It is also bound to / with a path string of "/" (a slash): this lets you type / and immediately type a path name. On Unix systems, this is bound to ~ (tilde) with a path string of "~" itself for access to home directories.

chooser :

the object which received the signal.

path :

default contents for the text entry for the file name

user.data :

user data set when the signal handler was connected.

PLEASE NOTE: You can create your own bindings for the GtkFileChooserDefault::location-popup signal with custom path strings, and have a crude form of easily-to-type bookmarks. For example, say you access the path ‘/home/username/misc’ very frequently. You could then create an Alt-M shortcut by including the following in your ‘.gtkrc-2.0’ :

   binding "misc-shortcut" {
      bind "<Alt>M" {
         "location-popup" ("/home/username/misc")
	 }
   }

   class "GtkFileChooserDefault" binding "misc-shortcut"
   

The "GtkFileChooserDefault::up-folder" signal

          void user_function (GtkFileChooserDefault *chooser,
                                gpointer   user_data);

This is used to make the file chooser go to the parent of the current folder in the file hierarchy. By default this is bound to Backspace and Alt-Up (the Up key in the numeric keypad also works).

chooser :

the object which received the signal.

user.data :

user data set when the signal handler was connected.

The "GtkFileChooserDefault::down-folder" signal

          void user_function (GtkFileChooserDefault *chooser,
                                gpointer   user_data);

This is used to make the file chooser go to a child of the current folder in the file hierarchy. The subfolder that will be used is displayed in the path bar widget of the file chooser. For example, if the path bar is showing "/foo/bar/baz", then this will cause the file chooser to switch to the "baz" subfolder. By default this is bound to Alt-Down (the Down key in the numeric keypad also works).

chooser :

the object which received the signal.

user.data :

user data set when the signal handler was connected.

The "GtkFileChooserDefault::home-folder" signal

          void user_function (GtkFileChooserDefault *chooser,
                                gpointer   user_data);

This is used to make the file chooser show the user's home folder in the file list. By default this is bound to Alt-Home (the Home key in the numeric keypad also works).

chooser :

the object which received the signal.

user.data :

user data set when the signal handler was connected.

The "GtkFileChooserDefault::desktop-folder" signal

          void user_function (GtkFileChooserDefault *chooser,
                                gpointer   user_data);

This is used to make the file chooser show the user's Desktop folder in the file list. By default this is bound to Alt-D.

chooser :

the object which received the signal.

user.data :

user data set when the signal handler was connected.

The "GtkFileChooserDefault::quick-bookmark" signal

          void user_function (GtkFileChooserDefault *chooser,
                              gint bookmark_index,
                                gpointer   user_data);

This is used to make the file chooser switch to the bookmark specified in the bookmark.index parameter. For example, if you have three bookmarks, you can pass 0, 1, 2 to this signal to switch to each of them, respectively. By default this is bound to Alt-1, Alt-2, etc. until Alt-0. Note that in the default binding, that Alt-1 is actually defined to switch to the bookmark at index 0, and so on successively; Alt-0 is defined to switch to the bookmark at index 10.

chooser :

the object which received the signal.

bookmark.indes :

index of the bookmark to switch to; the indices start at 0.

user.data :

user data set when the signal handler was connected.

Structures

GtkFileChooser

undocumented

Enums and Flags

GtkFileChooserAction

Describes whether a GtkFileChooser is being used to open existing files or to save to a possibly new file.

open

Indicates open mode. The file chooser will only let the user pick an existing file.

save

Indicates save mode. The file chooser will let the user pick an existing file, or type in a new filename.

select-folder

Indicates an Open mode for selecting folders. The file chooser will let the user pick an existing folder.

create-folder

Indicates a mode for creating a new folder. The file chooser will let the user name an existing or new folder.

GtkFileChooserConfirmation

Used as a return value of handlers for the "confirm-overwrite" signal of a GtkFileChooser. This value determines whether the file chooser will present the stock confirmation dialog, accept the user's choice of a filename, or let the user choose another filename. Since 2.8

confirm

The file chooser will present its stock dialog to confirm about overwriting an existing file.

accept-filename

The file chooser will terminate and accept the user's choice of a file name.

select-again

The file chooser will continue running, so as to let the user select another file name.

GtkFileChooserError

These identify the various errors that can occur while calling GtkFileChooser functions.

nonexistent

Indicates that a file does not exist.

bad-filename

Indicates a malformed filename.

Signals

confirm-overwrite(chooser, user.data)

This signal gets emitted whenever it is appropriate to present a confirmation dialog when the user has selected a file name that already exists. The signal only gets emitted when the file chooser is in GTK_FILE_CHOOSER_ACTION_SAVE mode.

Most applications just need to turn on the "do-overwrite-confirmation" property (or call the gtkFileChooserSetDoOverwriteConfirmation function), and they will automatically get a stock confirmation dialog. Applications which need to customize this behavior should do that, and also connect to the "confirm-overwrite" signal.

A signal handler for this signal must return a GtkFileChooserConfirmation value, which indicates the action to take. If the handler determines that the user wants to select a different filename, it should return GTK_FILE_CHOOSER_CONFIRMATION_SELECT_AGAIN. If it determines that the user is satisfied with his choice of file name, it should return GTK_FILE_CHOOSER_CONFIRMATION_ACCEPT_FILENAME. On the other hand, if it determines that the stock confirmation dialog should be used, it should return GTK_FILE_CHOOSER_CONFIRMATION_CONFIRM. The following example illustrates this.

Custom confirmation

static GtkFileChooserConfirmation
confirm_overwrite_callback (GtkFileChooser *chooser, gpointer data)
{
  char *uri;

  uri = gtk_file_chooser_get_uri (chooser);

  if (is_uri_read_only (uri))
    {
      if (user_wants_to_replace_read_only_file (uri))
        return GTK_FILE_CHOOSER_CONFIRMATION_ACCEPT_FILENAME;
      else
        return GTK_FILE_CHOOSER_CONFIRMATION_SELECT_AGAIN;
    } else
      return GTK_FILE_CHOOSER_CONFIRMATION_CONFIRM; // fall back to the default dialog
}

...

chooser = gtk_file_chooser_dialog_new (...);

gtk_file_chooser_set_do_overwrite_confirmation (GTK_FILE_CHOOSER (dialog), TRUE);
g_signal_connect (chooser, "confirm-overwrite",
                  G_CALLBACK (confirm_overwrite_callback), NULL);

if (gtk_dialog_run (chooser) == GTK_RESPONSE_ACCEPT)
        save_to_file (gtk_file_chooser_get_filename (GTK_FILE_CHOOSER (chooser));

gtk_widget_destroy (chooser);

Since 2.8

chooser

the object which received the signal.

user.data

user data set when the signal handler was connected.

Returns: [GtkFileChooserConfirmation] a GtkFileChooserConfirmation value that indicates which action to take after emitting the signal.

current-folder-changed(chooser, user.data)

This signal is emitted when the current folder in a GtkFileChooser changes. This can happen due to the user performing some action that changes folders, such as selecting a bookmark or visiting a folder on the file list. It can also happen as a result of calling a function to explicitly change the current folder in a file chooser.

Normally you do not need to connect to this signal, unless you need to keep track of which folder a file chooser is showing.

See also: gtkFileChooserSetCurrentFolder, gtkFileChooserGetCurrentFolder, gtkFileChooserSetCurrentFolderUri, gtkFileChooserGetCurrentFolderUri.

chooser

the object which received the signal.

user.data

user data set when the signal handler was connected.

file-activated(chooser, user.data)

This signal is emitted when the user "activates" a file in the file chooser. This can happen by double-clicking on a file in the file list, or by pressing Enter.

Normally you do not need to connect to this signal. It is used internally by GtkFileChooserDialog to know when to activate the default button in the dialog.

See also: gtkFileChooserGetFilename, gtkFileChooserGetFilenames, gtkFileChooserGetUri, gtkFileChooserGetUris.

chooser

the object which received the signal.

user.data

user data set when the signal handler was connected.

selection-changed(chooser, user.data)

This signal is emitted when there is a change in the set of selected files in a GtkFileChooser. This can happen when the user modifies the selection with the mouse or the keyboard, or when explicitly calling functions to change the selection.

Normally you do not need to connect to this signal, as it is easier to wait for the file chooser to finish running, and then to get the list of selected files using the functions mentioned below.

See also: gtkFileChooserSelectFilename, gtkFileChooserUnselectFilename, gtkFileChooserGetFilename, gtkFileChooserGetFilenames, gtkFileChooserSelectUri, gtkFileChooserUnselectUri, gtkFileChooserGetUri, gtkFileChooserGetUris.

chooser

the object which received the signal.

user.data

user data set when the signal handler was connected.

update-preview(chooser, user.data)

This signal is emitted when the preview in a file chooser should be regenerated. For example, this can happen when the currently selected file changes. You should use this signal if you want your file chooser to have a preview widget.

Once you have installed a preview widget with gtkFileChooserSetPreviewWidget, you should update it when this signal is emitted. You can use the functions gtkFileChooserGetPreviewFilename or gtkFileChooserGetPreviewUri to get the name of the file to preview. Your widget may not be able to preview all kinds of files; your callback must call gtkFileChooserSetPreviewWidgetActive to inform the file chooser about whether the preview was generated successfully or not.

Please see the example code in .

See also: gtkFileChooserSetPreviewWidget, gtkFileChooserSetPreviewWidgetActive, gtkFileChooserSetUsePreviewLabel, gtkFileChooserGetPreviewFilename, gtkFileChooserGetPreviewUri.

chooser

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

action [GtkFileChooserAction : Read / Write]

The type of operation that the file selector is performing. Default value: GTK_FILE_CHOOSER_ACTION_OPEN

create-folders [logical : Read / Write]

Whether a file chooser not in GTK_FILE_CHOOSER_ACTION_OPEN mode will offer the user to create new folders. Default value: TRUE Since 2.18

do-overwrite-confirmation [logical : Read / Write]

Whether a file chooser in GTK_FILE_CHOOSER_ACTION_SAVE mode will present an overwrite confirmation dialog if the user selects a file name that already exists. Default value: FALSE Since 2.8

extra-widget [GtkWidget : * : Read / Write]

Application supplied widget for extra options.

file-system-backend [character : * : Write / Construct Only]

Name of file system backend to use. Default value: NULL

filter [GtkFileFilter : * : Read / Write]

The current filter for selecting which files are displayed.

local-only [logical : Read / Write]

Whether the selected file(s) should be limited to local file: URLs. Default value: TRUE

preview-widget [GtkWidget : * : Read / Write]

Application supplied widget for custom previews.

preview-widget-active [logical : Read / Write]

Whether the application supplied widget for custom previews should be shown. Default value: TRUE

select-multiple [logical : Read / Write]

Whether to allow multiple files to be selected. Default value: FALSE

show-hidden [logical : Read / Write]

Whether the hidden files and folders should be displayed. Default value: FALSE

use-preview-label [logical : Read / Write]

Whether to display a stock label with the name of the previewed file. Default value: TRUE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkFileChooser.html

[Package RGtk2 version 2.20.31 Index]