GtkFileChooserWidget {RGtk2}	R Documentation

GtkFileChooserWidget

Description

File chooser widget that can be embedded in other widgets

Methods and Functions

gtkFileChooserWidgetNew(action, show = TRUE)
gtkFileChooserWidgetNewWithBackend(action, backend, show = TRUE)
gtkFileChooserWidgetNewWithBackend(action, backend, show = TRUE)
gtkFileChooserWidget(action, backend, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBox
                                 +----GtkVBox
                                       +----GtkFileChooserWidget

Interfaces

GtkFileChooserWidget implements AtkImplementorIface, GtkBuildable, GtkOrientable, GtkFileChooser and GtkFileChooserEmbed.

Detailed Description

GtkFileChooserWidget is a widget suitable for selecting files. It is the main building block of a GtkFileChooserDialog. Most applications will only need to use the latter; you can use GtkFileChooserWidget as part of a larger window if you have special needs.

Note that GtkFileChooserWidget does not have any methods of its own. Instead, you should use the functions that work on a GtkFileChooser.

Structures

GtkFileChooserWidget

undocumented

Convenient Construction

gtkFileChooserWidget is the result of collapsing the constructors of GtkFileChooserWidget (gtkFileChooserWidgetNew, gtkFileChooserWidgetNewWithBackend) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkFileChooserWidget.html

See Also

GtkFileChooser GtkFileChooserDialog

[Package RGtk2 version 2.20.31 Index]