GtkFixed {RGtk2}	R Documentation

GtkFixed

Description

A container which allows you to position widgets at fixed coordinates

Methods and Functions

gtkFixedNew(show = TRUE)
gtkFixedPut(object, widget, x, y)
gtkFixedMove(object, widget, x, y)
gtkFixedGetHasWindow(object)
gtkFixedSetHasWindow(object, has.window)
gtkFixed(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkFixed

Interfaces

GtkFixed implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkFixed widget is a container which can place child widgets at fixed positions and with fixed sizes, given in pixels. GtkFixed performs no automatic layout management.

For most applications, you should not use this container! It keeps you from having to learn about the other GTK+ containers, but it results in broken applications. With GtkFixed, the following things will result in truncated text, overlapping widgets, and other display bugs:

Themes, which may change widget sizes.

Fonts other than the one you used to write the app will of course change the size of widgets containing text; keep in mind that users may use a larger font because of difficulty reading the default, or they may be using Windows or the framebuffer port of GTK+, where different fonts are available.

Translation of text into other languages changes its size. Also, display of non-English text will use a different font in many cases.

In addition, the fixed widget can't properly be mirrored in right-to-left languages such as Hebrew and Arabic. i.e. normally GTK+ will flip the interface to put labels to the right of the thing they label, but it can't do that with GtkFixed. So your application will not be usable in right-to-left languages.

Finally, fixed positioning makes it kind of annoying to add/remove GUI elements, since you have to reposition all the other elements. This is a long-term maintenance problem for your application.

If you know none of these things are an issue for your application, and prefer the simplicity of GtkFixed, by all means use the widget. But you should be aware of the tradeoffs.

Structures

GtkFixed

The GtkFixed struct contains the following fields. (These fields should be considered read-only. They should never be set by an application.)

list *children; 	 a list of GtkFixedChild elements, containing the child widgets and their positions.

Convenient Construction

gtkFixed is the equivalent of gtkFixedNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkFixed.html

[Package RGtk2 version 2.20.31 Index]