GtkFontButton {RGtk2}	R Documentation

GtkFontButton

Description

A button to launch a font selection dialog

Methods and Functions

gtkFontButtonNew(show = TRUE)
gtkFontButtonNewWithFont(fontname)
gtkFontButtonSetFontName(object, fontname)
gtkFontButtonGetFontName(object)
gtkFontButtonSetShowStyle(object, show.style)
gtkFontButtonGetShowStyle(object)
gtkFontButtonSetShowSize(object, show.size)
gtkFontButtonGetShowSize(object)
gtkFontButtonSetUseFont(object, use.font)
gtkFontButtonGetUseFont(object)
gtkFontButtonSetUseSize(object, use.size)
gtkFontButtonGetUseSize(object)
gtkFontButtonSetTitle(object, title)
gtkFontButtonGetTitle(object)
gtkFontButton(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkButton
                                       +----GtkFontButton

Interfaces

GtkFontButton implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

The GtkFontButton is a button which displays the currently selected font an allows to open a font selection dialog to change the font. It is suitable widget for selecting a font in a preference dialog.

Structures

GtkFontButton

The GtkFontButton struct has only private members and should not be used directly.

Convenient Construction

gtkFontButton is the equivalent of gtkFontButtonNew.

Signals

font-set(widget, user.data)

The ::font-set signal is emitted when the user selects a font. When handling this signal, use gtkFontButtonGetFontName to find out which font was just selected.

Note that this signal is only emitted when the user changes the font. If you need to react to programmatic font changes as well, use the notify::font-name signal. Since 2.4

widget

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

font-name [character : * : Read / Write]

The name of the currently selected font. Default value: "Sans 12" Since 2.4

show-size [logical : Read / Write]

If this property is set to TRUE, the selected font size will be shown in the label. For a more WYSIWYG way to show the selected size, see the ::use-size property. Default value: TRUE Since 2.4

show-style [logical : Read / Write]

If this property is set to TRUE, the name of the selected font style will be shown in the label. For a more WYSIWYG way to show the selected style, see the ::use-font property. Default value: TRUE Since 2.4

title [character : * : Read / Write]

The title of the font selection dialog. Default value: "Pick a Font" Since 2.4

use-font [logical : Read / Write]

If this property is set to TRUE, the label will be drawn in the selected font. Default value: FALSE Since 2.4

use-size [logical : Read / Write]

If this property is set to TRUE, the label will be drawn with the selected font size. Default value: FALSE Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkFontButton.html

See Also

GtkFontSelectionDialog GtkColorButton

[Package RGtk2 version 2.20.31 Index]