GtkFontSelection {RGtk2}	R Documentation

GtkFontSelection

Description

A widget for selecting fonts

Methods and Functions

gtkFontSelectionNew(show = TRUE)
gtkFontSelectionGetFont(object)
gtkFontSelectionGetFontName(object)
gtkFontSelectionSetFontName(object, fontname)
gtkFontSelectionGetPreviewText(object)
gtkFontSelectionSetPreviewText(object, text)
gtkFontSelectionGetFace(object)
gtkFontSelectionGetFaceList(object)
gtkFontSelectionGetFamily(object)
gtkFontSelectionGetSize(object)
gtkFontSelectionGetFamilyList(object)
gtkFontSelectionGetPreviewEntry(object)
gtkFontSelectionGetSizeEntry(object)
gtkFontSelectionGetSizeList(object)
gtkFontSelection(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBox
                                 +----GtkVBox
                                       +----GtkFontSelection

Interfaces

GtkFontSelection implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

The GtkFontSelection widget lists the available fonts, styles and sizes, allowing the user to select a font. It is used in the GtkFontSelectionDialog widget to provide a dialog box for selecting fonts.

To set the font which is initially selected, use gtkFontSelectionSetFontName.

To get the selected font use gtkFontSelectionGetFontName.

To change the text which is shown in the preview area, use gtkFontSelectionSetPreviewText.

Structures

GtkFontSelection

The GtkFontSelection struct contains private data only, and should only be accessed using the functions below.

Convenient Construction

gtkFontSelection is the equivalent of gtkFontSelectionNew.

Properties

font [GdkFont : * : Read]

The GdkFont that is currently selected.

font-name [character : * : Read / Write]

The string that represents this font. Default value: "Sans 10"

preview-text [character : * : Read / Write]

The text to display in order to demonstrate the selected font. Default value: "abcdefghijk ABCDEFGHIJK"

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkFontSelection.html

[Package RGtk2 version 2.20.31 Index]