GtkFontSelectionDialog {RGtk2}	R Documentation

GtkFontSelectionDialog

Description

A dialog box for selecting fonts

Methods and Functions

gtkFontSelectionDialogNew(title = NULL, show = TRUE)
gtkFontSelectionDialogGetFont(object)
gtkFontSelectionDialogGetFontName(object)
gtkFontSelectionDialogSetFontName(object, fontname)
gtkFontSelectionDialogGetPreviewText(object)
gtkFontSelectionDialogSetPreviewText(object, text)
gtkFontSelectionDialogGetApplyButton(object)
gtkFontSelectionDialogGetApplyButton(object)
gtkFontSelectionDialogGetCancelButton(object)
gtkFontSelectionDialogGetOkButton(object)
gtkFontSelectionDialog(title = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkWindow
                                       +----GtkDialog
                                             +----GtkFontSelectionDialog

Interfaces

GtkFontSelectionDialog implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkFontSelectionDialog widget is a dialog box for selecting a font.

To set the font which is initially selected, use gtkFontSelectionDialogSetFontName.

To get the selected font use gtkFontSelectionDialogGetFontName.

To change the text which is shown in the preview area, use gtkFontSelectionDialogSetPreviewText.

GtkFontSelectionDialog as GtkBuildable

The GtkFontSelectionDialog implementation of the GtkBuildable interface exposes the embedded GtkFontSelection as internal child with the name "font_selection". It also exposes the buttons with the names "ok_button", "cancel_button" and "apply_button".

Structures

GtkFontSelectionDialog

undocumented

okButton

[GtkWidget]

applyButton

[GtkWidget]

cancelButton

[GtkWidget]

Convenient Construction

gtkFontSelectionDialog is the equivalent of gtkFontSelectionDialogNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkFontSelectionDialog.html

[Package RGtk2 version 2.20.31 Index]