GtkFrame {RGtk2}	R Documentation

GtkFrame

Description

A bin with a decorative frame and optional label

Methods and Functions

gtkFrameNew(label = NULL, show = TRUE)
gtkFrameSetLabel(object, label = NULL)
gtkFrameSetLabelWidget(object, label.widget)
gtkFrameSetLabelAlign(object, xalign, yalign)
gtkFrameSetShadowType(object, type)
gtkFrameGetLabel(object)
gtkFrameGetLabelAlign(object)
gtkFrameGetLabelWidget(object)
gtkFrameGetShadowType(object)
gtkFrame(label = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkFrame
                                       +----GtkAspectFrame

Interfaces

GtkFrame implements AtkImplementorIface and GtkBuildable.

Detailed Description

The frame widget is a Bin that surrounds its child with a decorative frame and an optional label. If present, the label is drawn in a gap in the top side of the frame. The position of the label can be controlled with gtkFrameSetLabelAlign.

GtkFrame as GtkBuildable

The GtkFrame implementation of the GtkBuildable interface supports placing a child in the label position by specifying "label" as the "type" attribute of a <child> element. A normal content child can be specified without specifying a <child> type attribute.

A UI definition fragment with GtkFrame

<object class="GtkFrame">
  <child type="label">
    <object class="GtkLabel" id="frame-label"/>
  </child>
  <child>
    <object class="GtkEntry" id="frame-content"/>
  </child>
</object>

Structures

GtkFrame

undocumented

Convenient Construction

gtkFrame is the equivalent of gtkFrameNew.

Properties

label [character : * : Read / Write]

Text of the frame's label. Default value: NULL

label-widget [GtkWidget : * : Read / Write]

A widget to display in place of the usual frame label.

label-xalign [numeric : Read / Write]

The horizontal alignment of the label. Allowed values: [0,1] Default value: 0

label-yalign [numeric : Read / Write]

The vertical alignment of the label. Allowed values: [0,1] Default value: 0.5

shadow [GtkShadowType : Read / Write]

Deprecated property, use shadow_type instead. Default value: GTK_SHADOW_ETCHED_IN

shadow-type [GtkShadowType : Read / Write]

Appearance of the frame border. Default value: GTK_SHADOW_ETCHED_IN

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkFrame.html

[Package RGtk2 version 2.20.31 Index]