GtkHBox {RGtk2}	R Documentation

GtkHBox

Description

A horizontal container box

Methods and Functions

gtkHBoxNew(homogeneous = NULL, spacing = NULL, show = TRUE)
gtkHBox(homogeneous = NULL, spacing = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBox
                                 +----GtkHBox
                                       +----GtkCombo
                                       +----GtkFileChooserButton
                                       +----GtkInfoBar
                                       +----GtkStatusbar

Interfaces

GtkHBox implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

GtkHBox is a container that organizes child widgets into a single row.

Use the GtkBox packing interface to determine the arrangement, spacing, width, and alignment of GtkHBox children.

All children are allocated the same height.

Structures

GtkHBox

undocumented

Convenient Construction

gtkHBox is the equivalent of gtkHBoxNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkHBox.html

[Package RGtk2 version 2.20.31 Index]