GtkHPaned {RGtk2}	R Documentation

GtkHPaned

Description

A container with two panes arranged horizontally

Methods and Functions

gtkHPanedNew(show = TRUE)
gtkHPaned(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkPaned
                                 +----GtkHPaned

Interfaces

GtkHPaned implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

The HPaned widget is a container widget with two children arranged horizontally. The division between the two panes is adjustable by the user by dragging a handle. See GtkPaned for details.

Structures

GtkHPaned

undocumented

Convenient Construction

gtkHPaned is the equivalent of gtkHPanedNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkHPaned.html

[Package RGtk2 version 2.20.31 Index]