GtkHSV {RGtk2}	R Documentation

GtkHSV

Description

A 'color wheel' widget

Methods and Functions

gtkHSVNew()
gtkHSVSetColor(object, h, s, v)
gtkHSVGetColor(object, h, s, v)
gtkHSVSetMetrics(object, size, ring.width)
gtkHSVGetMetrics(object, size, ring.width)
gtkHSVIsAdjusting(object)
gtkHSVToRgb(h, s, r, g, b)
gtkRgbToHsv(r, g, h, s, v)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkHSV

Interfaces

GtkHSV implements AtkImplementorIface and GtkBuildable.

Detailed Description

GtkHSV is the 'color wheel' part of a complete color selector widget. It allows to select a color by determining its HSV components in an intuitive way. Moving the selection around the outer ring changes the hue, and moving the selection point inside the inner triangle changes value and saturation.

Structures

GtkHSV

undocumented

Signals

changed(hsv, user.data)

undocumented

hsv

the object which received the signal.

user.data

user data set when the signal handler was connected.

move(hsv, user.data)

undocumented

hsv

the object which received the signal.

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkHSV.html

[Package RGtk2 version 2.20.31 Index]