GtkHScale {RGtk2}	R Documentation

GtkHScale

Description

A horizontal slider widget for selecting a value from a range

Methods and Functions

gtkHScaleNew(adjustment = NULL, show = TRUE)
gtkHScaleNewWithRange(min, max, step, show = TRUE)
gtkHScale(adjustment = NULL, min, max, step, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkRange
                           +----GtkScale
                                 +----GtkHScale

Interfaces

GtkHScale implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

The GtkHScale widget is used to allow the user to select a value using a horizontal slider. To create one, use gtkHScaleNewWithRange.

The position to show the current value, and the number of decimal places shown can be set using the parent GtkScale class's functions.

Structures

GtkHScale

The GtkHScale struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkHScale is the result of collapsing the constructors of GtkHScale (gtkHScaleNew, gtkHScaleNewWithRange) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkHScale.html

[Package RGtk2 version 2.20.31 Index]