GtkHScrollbar {RGtk2}	R Documentation

GtkHScrollbar

Description

A horizontal scrollbar

Methods and Functions

gtkHScrollbarNew(adjustment = NULL, show = TRUE)
gtkHScrollbar(adjustment = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkRange
                           +----GtkScrollbar
                                 +----GtkHScrollbar

Interfaces

GtkHScrollbar implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

The GtkHScrollbar widget is a widget arranged horizontally creating a scrollbar. See GtkScrollbar for details on scrollbars. GtkAdjustment pointers may be added to handle the adjustment of the scrollbar or it may be left NULL in which case one will be created for you. See GtkScrollbar for a description of what the fields in an adjustment represent for a scrollbar.

Structures

GtkHScrollbar

The GtkHScrollbar struct contains private data and should be accessed using the functions below.

Convenient Construction

gtkHScrollbar is the equivalent of gtkHScrollbarNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkHScrollbar.html

See Also

GtkScrollbar GtkScrolledWindow

[Package RGtk2 version 2.20.31 Index]