GtkHSeparator {RGtk2}	R Documentation

GtkHSeparator

Description

A horizontal separator

Methods and Functions

gtkHSeparatorNew(show = TRUE)
gtkHSeparator(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkSeparator
                           +----GtkHSeparator

Interfaces

GtkHSeparator implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

The GtkHSeparator widget is a horizontal separator, used to group the widgets within a window. It displays a horizontal line with a shadow to make it appear sunken into the interface. PLEASE NOTE: The GtkHSeparator widget is not used as a separator within menus. To create a separator in a menu create an empty GtkSeparatorMenuItem widget using gtkSeparatorMenuItemNew and add it to the menu with gtkMenuShellAppend.

Structures

GtkHSeparator

The GtkHSeparator struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkHSeparator is the equivalent of gtkHSeparatorNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkHSeparator.html

[Package RGtk2 version 2.20.31 Index]