GtkHandleBox {RGtk2}	R Documentation

GtkHandleBox

Description

a widget for detachable window portions

Methods and Functions

gtkHandleBoxNew(show = TRUE)
gtkHandleBoxSetShadowType(object, type)
gtkHandleBoxSetHandlePosition(object, position)
gtkHandleBoxSetSnapEdge(object, edge)
gtkHandleBoxGetHandlePosition(object)
gtkHandleBoxGetShadowType(object)
gtkHandleBoxGetSnapEdge(object)
gtkHandleBoxGetChildDetached(object)
gtkHandleBox(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkHandleBox

Interfaces

GtkHandleBox implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkHandleBox widget allows a portion of a window to be "torn off". It is a bin widget which displays its child and a handle that the user can drag to tear off a separate window (the float window) containing the child widget. A thin ghost is drawn in the original location of the handlebox. By dragging the separate window back to its original location, it can be reattached.

When reattaching, the ghost and float window, must be aligned along one of the edges, the snap edge. This either can be specified by the application programmer explicitely, or GTK+ will pick a reasonable default based on the handle position.

To make detaching and reattaching the handlebox as minimally confusing as possible to the user, it is important to set the snap edge so that the snap edge does not move when the handlebox is deattached. For instance, if the handlebox is packed at the bottom of a VBox, then when the handlebox is detached, the bottom edge of the handlebox's allocation will remain fixed as the height of the handlebox shrinks, so the snap edge should be set to GTK_POS_BOTTOM.

Structures

GtkHandleBox

The GtkHandleBox struct contains the following fields. (These fields should be considered read-only. They should never be set by an application.)

GtkShadowType shadow_type; 	 The shadow type for the entry. (See gtkHandleBoxSetShadowType ).

GtkPositionType handle_position; 	 The position of the handlebox's handle with respect to the child. (See gtkHandleBoxSetHandlePosition )

gint snap_edge; 	 A value of type GtkPosition type indicating snap edge for the widget. (See gtk_handle_box_set_snap_edge). The value of -1 indicates that this value has not been set.

logical child_detached; 	 A boolean value indicating whether the handlebox's child is attached or detached.

Convenient Construction

gtkHandleBox is the equivalent of gtkHandleBoxNew.

Signals

child-attached(handlebox, widget, user.data)

This signal is emitted when the contents of the handlebox are reattached to the main window.

handlebox

the object which received the signal.

widget

the child widget of the handlebox. (this argument provides no extra information and is here only for backwards-compatibility)

user.data

user data set when the signal handler was connected.

child-detached(handlebox, widget, user.data)

This signal is emitted when the contents of the handlebox are detached from the main window.

handlebox

the object which received the signal.

widget

the child widget of the handlebox. (this argument provides no extra information and is here only for backwards-compatibility)

user.data

user data set when the signal handler was connected.

Properties

child-detached [logical : Read]

A boolean value indicating whether the handlebox's child is attached or detached. Default value: FALSE

handle-position [GtkPositionType : Read / Write]

Position of the handle relative to the child widget. Default value: GTK_POS_LEFT

shadow [GtkShadowType : Read / Write]

Deprecated property, use shadow_type instead. Default value: GTK_SHADOW_OUT

shadow-type [GtkShadowType : Read / Write]

Appearance of the shadow that surrounds the container. Default value: GTK_SHADOW_OUT

snap-edge [GtkPositionType : Read / Write]

Side of the handlebox that's lined up with the docking point to dock the handlebox. Default value: GTK_POS_TOP

snap-edge-set [logical : Read / Write]

Whether to use the value from the snap_edge property or a value derived from handle_position. Default value: FALSE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkHandleBox.html

[Package RGtk2 version 2.20.31 Index]