GtkIMContext {RGtk2}	R Documentation

GtkIMContext

Description

Base class for input method contexts

Methods and Functions

gtkIMContextSetClientWindow(object, window)
gtkIMContextGetPreeditString(object)
gtkIMContextFilterKeypress(object, event)
gtkIMContextFocusIn(object)
gtkIMContextFocusOut(object)
gtkIMContextReset(object)
gtkIMContextSetCursorLocation(object, area)
gtkIMContextSetUsePreedit(object, use.preedit)
gtkIMContextSetSurrounding(object, text, cursor.index)
gtkIMContextGetSurrounding(object)
gtkIMContextDeleteSurrounding(object, offset, n.chars)

Hierarchy

GObject
   +----GtkIMContext
         +----GtkIMContextSimple
         +----GtkIMMulticontext

Detailed Description

GtkIMContext defines the interface for GTK+ input methods. An input method is used by GTK+ text input widgets like GtkEntry to map from key events to Unicode character strings.

The user may change the current input method via a context menu, unless the "gtk-show-input-method-menu" GtkSettings property is set to FALSE. The default input method can be set programmatically via the "gtk-im-module" GtkSettings property. Alternatively, you may set the GTK_IM_MODULE environment variable as documented in gtk-running.

The GtkEntry "im-module" and GtkTextView "im-module" properties may also be used to set input methods for specific widget instances. For instance, a certain entry widget might be expected to contain certain characters which would be easier to input with a certain input method.

An input method may consume multiple key events in sequence and finally output the composed result. This is called preediting, and an input method may provide feedback about this process by displaying the intermediate composition states as preedit text. For instance, the default GTK+ input method implements the input of arbitrary Unicode code points by holding down the Control and Shift keys and then typing "U" followed by the hexadecimal digits of the code point. When releasing the Control and Shift keys, preediting ends and the character is inserted as text. Ctrl+Shift+u20AC for example results in the € sign.

Additional input methods can be made available for use by GTK+ widgets as loadable modules. An input method module is a small shared library which implements a subclass of GtkIMContext or GtkIMContextSimple and exports these four functions:

void im_module_init(  GTypeModule   *module);

This function should register the GType of the GtkIMContext subclass which implements the input method by means of gTypeModuleRegisterType(). Note that gTypeRegisterStatic() cannot be used as the type needs to be registered dynamically.

void im_module_exit(void);

Here goes any cleanup code your input method might require on module unload.

void im_module_list(const   GtkIMContextInfo   ***contexts, int *n_contexts)
{
  *contexts = info_list;
  *n_contexts = G_N_ELEMENTS (info_list);
}

This function returns the list of input methods provided by the module. The example implementation above shows a common solution and simply returns a pointer to statically defined list of GtkIMContextInfo items for each provided input method.

GtkIMContext   * im_module_create(const   gchar   *context_id);

This function should return a pointer to a newly created instance of the GtkIMContext subclass identified by context.id. The context ID is the same as specified in the GtkIMContextInfo list returned by imModuleList().

After a new loadable input method module has been installed on the system, the configuration file ‘gtk.immodules’ needs to be regenerated by gtk-query-immodules-2.0, in order for the new input method to become available to GTK+ applications.

Structures

GtkIMContext

undocumented

Signals

commit(context, str, user.data)

The ::commit signal is emitted when a complete input sequence has been entered by the user. This can be a single character immediately after a key press or the final result of preediting.

context

the object on which the signal is emitted

str

the completed character(s) entered by the user

user.data

user data set when the signal handler was connected.

delete-surrounding(context, offset, n.chars, user.data)

The ::delete-surrounding signal is emitted when the input method needs to delete all or part of the context surrounding the cursor.

context

the object on which the signal is emitted

offset

the character offset from the cursor position of the text to be deleted. A negative value indicates a position before the cursor.

n.chars

the number of characters to be deleted

user.data

user data set when the signal handler was connected.

Returns: [logical] TRUE if the signal was handled.

preedit-changed(context, user.data)

The ::preedit-changed signal is emitted whenever the preedit sequence currently being entered has changed. It is also emitted at the end of a preedit sequence, in which case gtkIMContextGetPreeditString returns the empty string.

context

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

preedit-end(context, user.data)

The ::preedit-end signal is emitted when a preediting sequence has been completed or canceled.

context

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

preedit-start(context, user.data)

The ::preedit-start signal is emitted when a new preediting sequence starts.

context

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

retrieve-surrounding(context, user.data)

The ::retrieve-surrounding signal is emitted when the input method requires the context surrounding the cursor. The callback should set the input method surrounding context by calling the gtkIMContextSetSurrounding method.

context

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

Returns: [logical] TRUE if the signal was handled.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkIMContext.html

[Package RGtk2 version 2.20.31 Index]