GtkIconTheme {RGtk2}	R Documentation

GtkIconTheme

Description

Looking up icons by name

Methods and Functions

gtkIconThemeNew()
gtkIconThemeGetDefault()
gtkIconThemeGetForScreen(screen)
gtkIconThemeSetScreen(object, screen)
gtkIconThemeSetCustomTheme(object, theme.name)
gtkIconThemeHasIcon(object, icon.name)
gtkIconThemeLookupIcon(object, icon.name, size, flags)
gtkIconThemeChooseIcon(object, icon.names, size, flags)
gtkIconThemeLookupByGicon(object, icon, size, flags)
gtkIconThemeLoadIcon(object, icon.name, size, flags, .errwarn = TRUE)
gtkIconThemeListContexts(object)
gtkIconThemeListIcons(object, context = NULL)
gtkIconThemeGetIconSizes(object, icon.name)
gtkIconThemeGetExampleIconName(object)
gtkIconThemeRescanIfNeeded(object)
gtkIconThemeAddBuiltinIcon(icon.name, size, pixbuf)
gtkIconInfoCopy(object)
gtkIconInfoNewForPixbuf(icon.theme, pixbuf)
gtkIconInfoGetBaseSize(object)
gtkIconInfoGetBuiltinPixbuf(object)
gtkIconInfoLoadIcon(object, .errwarn = TRUE)
gtkIconInfoSetRawCoordinates(object, raw.coordinates)
gtkIconInfoGetEmbeddedRect(object)
gtkIconInfoGetAttachPoints(object)
gtkIconInfoGetDisplayName(object)
gtkIconTheme()

Hierarchy

GObject
   +----GtkIconTheme

Detailed Description

GtkIconTheme provides a facility for looking up icons by name and size. The main reason for using a name rather than simply providing a filename is to allow different icons to be used depending on what icon theme is selected by the user. The operation of icon themes on Linux and Unix follows the Icon Theme Specification (http://www.freedesktop.org/Standards/icon-theme-spec). There is a default icon theme, named hicolor where applications should install their icons, but more additional application themes can be installed as operating system vendors and users choose.

Named icons are similar to the facility, and the distinction between the two may be a bit confusing. A few things to keep in mind:

Stock images usually are used in conjunction with ., such as GTK_STOCK_OK or GTK_STOCK_OPEN. Named icons are easier to set up and therefore are more useful for new icons that an application wants to add, such as application icons or window icons.

Stock images can only be loaded at the symbolic sizes defined by the GtkIconSize enumeration, or by custom sizes defined by gtkIconSizeRegister, while named icons are more flexible and any pixel size can be specified.

Because stock images are closely tied to stock items, and thus to actions in the user interface, stock images may come in multiple variants for different widget states or writing directions.

A good rule of thumb is that if there is a stock image for what you want to use, use it, otherwise use a named icon. It turns out that internally stock images are generally defined in terms of one or more named icons. (An example of the more than one case is icons that depend on writing direction; GTK_STOCK_GO_FORWARD uses the two themed icons "gtk-stock-go-forward-ltr" and "gtk-stock-go-forward-rtl".)

In many cases, named themes are used indirectly, via GtkImage or stock items, rather than directly, but looking up icons directly is also simple. The GtkIconTheme object acts as a database of all the icons in the current theme. You can create new GtkIconTheme objects, but its much more efficient to use the standard icon theme for the GdkScreen so that the icon information is shared with other people looking up icons. In the case where the default screen is being used, looking up an icon can be as simple as:

icon_theme <- gtkIconThemeGetDefault()
result <- icon_theme$loadIcon("my-icon-name", 48, 0)
if (!result[[1]]) {
  warning("Couldn't load icon: ", result$error$message)
} else {
  pixbuf <- result[[1]]
  ## Use the pixbuf
}

Structures

GtkIconInfo

Contains information found when looking up an icon in an icon theme.

GtkIconTheme

Acts as a database of information about an icon theme. Normally, you retrieve the icon theme for a particular screen using gtkIconThemeGetForScreen and it will contain information about current icon theme for that screen, but you can also create a new GtkIconTheme object and set the icon theme name explicitely using gtkIconThemeSetCustomTheme.

Convenient Construction

gtkIconTheme is the equivalent of gtkIconThemeNew.

Enums and Flags

GtkIconLookupFlags

undocumented

no-svg

Never return SVG icons, even if gdk-pixbuf supports them. Cannot be used together with GTK_ICON_LOOKUP_FORCE_SVG.

force-svg

Return SVG icons, even if gdk-pixbuf doesn't support them. Cannot be used together with GTK_ICON_LOOKUP_NO_SVG.

use-builtin

When passed to gtkIconThemeLookupIcon includes builtin icons as well as files. For a builtin icon, gtkIconInfoGetFilename

GtkIconThemeError

Error codes for GtkIconTheme operations.

not-found

The icon specified does not exist in the theme

failed

An unspecified error occurred.

Signals

changed(icon.theme, user.data)

Emitted when the current icon theme is switched or GTK+ detects that a change has occurred in the contents of the current icon theme.

icon.theme

the icon theme

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkIconTheme.html

[Package RGtk2 version 2.20.31 Index]