GtkIconView {RGtk2}	R Documentation

GtkIconView

Description

A widget which displays a list of icons in a grid

Methods and Functions

gtkIconViewNew(show = TRUE)
gtkIconViewNewWithModel(model = NULL, show = TRUE)
gtkIconViewSetModel(object, model = NULL)
gtkIconViewGetModel(object)
gtkIconViewSetTextColumn(object, column)
gtkIconViewGetTextColumn(object)
gtkIconViewSetMarkupColumn(object, column)
gtkIconViewGetMarkupColumn(object)
gtkIconViewSetPixbufColumn(object, column)
gtkIconViewGetPixbufColumn(object)
gtkIconViewGetPathAtPos(object, x, y)
gtkIconViewGetItemAtPos(object, x, y)
gtkIconViewConvertWidgetToBinWindowCoords(object, wx, wy)
gtkIconViewSetCursor(object, path, cell, start.editing)
gtkIconViewGetCursor(object)
gtkIconViewSelectedForeach(object, func, data = NULL)
gtkIconViewSetSelectionMode(object, mode)
gtkIconViewGetSelectionMode(object)
gtkIconViewSetOrientation(object, orientation)
gtkIconViewGetOrientation(object)
gtkIconViewSetColumns(object, columns)
gtkIconViewGetColumns(object)
gtkIconViewSetItemWidth(object, item.width)
gtkIconViewGetItemWidth(object)
gtkIconViewSetSpacing(object, spacing)
gtkIconViewGetSpacing(object)
gtkIconViewSetRowSpacing(object, row.spacing)
gtkIconViewGetRowSpacing(object)
gtkIconViewSetColumnSpacing(object, column.spacing)
gtkIconViewGetColumnSpacing(object)
gtkIconViewSetMargin(object, margin)
gtkIconViewGetMargin(object)
gtkIconViewSetItemPadding(object, item.padding)
gtkIconViewGetItemPadding(object)
gtkIconViewSelectPath(object, path)
gtkIconViewUnselectPath(object, path)
gtkIconViewPathIsSelected(object, path)
gtkIconViewGetSelectedItems(object)
gtkIconViewSelectAll(object)
gtkIconViewUnselectAll(object)
gtkIconViewItemActivated(object, path)
gtkIconViewScrollToPath(object, path, use.align, row.align, col.align)
gtkIconViewGetVisibleRange(object)
gtkIconViewSetTooltipItem(object, tooltip, path)
gtkIconViewSetTooltipCell(object, tooltip, path, cell)
gtkIconViewGetTooltipContext(object, x, y, keyboard.tip)
gtkIconViewSetTooltipColumn(object, column)
gtkIconViewGetTooltipColumn(object)
gtkIconViewEnableModelDragSource(object, start.button.mask, targets, actions)
gtkIconViewEnableModelDragDest(object, targets, actions)
gtkIconViewUnsetModelDragSource(object)
gtkIconViewUnsetModelDragDest(object)
gtkIconViewSetReorderable(object, reorderable)
gtkIconViewGetReorderable(object)
gtkIconViewSetDragDestItem(object, path, pos)
gtkIconViewGetDragDestItem(object)
gtkIconViewGetDestItemAtPos(object, drag.x, drag.y)
gtkIconViewCreateDragIcon(object, path)
gtkIconView(model = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkIconView

Interfaces

GtkIconView implements AtkImplementorIface, GtkBuildable and GtkCellLayout.

Detailed Description

GtkIconView provides an alternative view on a list model. It displays the model as a grid of icons with labels. Like GtkTreeView, it allows to select one or multiple items (depending on the selection mode, see gtkIconViewSetSelectionMode). In addition to selection with the arrow keys, GtkIconView supports rubberband selection, which is controlled by dragging the pointer.

Structures

GtkIconView

The GtkIconView struct contains only private fields and should not be directly accessed.

Convenient Construction

gtkIconView is the result of collapsing the constructors of GtkIconView (gtkIconViewNew, gtkIconViewNewWithModel) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Enums and Flags

GtkIconViewDropPosition

An enum for determining where a dropped item goes.

no-drop

no drop possible

drop-into

dropped item replaces the item

drop-left

droppped item is inserted to the left

drop-right

dropped item is inserted to the right

drop-above

dropped item is inserted above

drop-below

dropped item is inserted below

User Functions

GtkIconViewForeachFunc(icon.view, path, data)

A function used by gtkIconViewSelectedForeach to map all selected rows. It will be called on every selected row in the view.

icon.view

a GtkIconView

path

The GtkTreePath of a selected row

data

user data

Signals

activate-cursor-item(iconview, user.data)

A keybinding signal which gets emitted when the user activates the currently focused item.

Applications should not connect to it, but may emit it with gSignalEmitByName() if they need to control activation programmatically.

The default bindings for this signal are Space, Return and Enter.

iconview

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

item-activated(iconview, path, user.data)

The ::item-activated signal is emitted when the method gtkIconViewItemActivated is called or the user double clicks an item. It is also emitted when a non-editable item is selected and one of the keys: Space, Return or Enter is pressed.

iconview

the object on which the signal is emitted

path

the GtkTreePath for the activated item

user.data

user data set when the signal handler was connected.

move-cursor(iconview, step, count, user.data)

The ::move-cursor signal is a keybinding signal which gets emitted when the user initiates a cursor movement.

Applications should not connect to it, but may emit it with gSignalEmitByName() if they need to control the cursor programmatically.

The default bindings for this signal include

All of these will extend the selection when combined with the Shift modifier.

iconview

the object which received the signal

step

the granularity of the move, as a GtkMovementStep

count

the number of step units to move

user.data

user data set when the signal handler was connected.

select-all(iconview, user.data)

A keybinding signal which gets emitted when the user selects all items.

Applications should not connect to it, but may emit it with gSignalEmitByName() if they need to control selection programmatically.

The default binding for this signal is Ctrl-a.

iconview

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

select-cursor-item(iconview, user.data)

A keybinding signal which gets emitted when the user selects the item that is currently focused.

Applications should not connect to it, but may emit it with gSignalEmitByName() if they need to control selection programmatically.

There is no default binding for this signal.

iconview

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

selection-changed(iconview, user.data)

The ::selection-changed signal is emitted when the selection (i.e. the set of selected items) changes.

iconview

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

set-scroll-adjustments(horizontal, vertical, user.data)

Set the scroll adjustments for the icon view. Usually scrolled containers like GtkScrolledWindow will emit this signal to connect two instances of GtkScrollbar to the scroll directions of the GtkIconView.

horizontal

the horizontal GtkAdjustment

vertical

the vertical GtkAdjustment

user.data

user data set when the signal handler was connected.

toggle-cursor-item(iconview, user.data)

A keybinding signal which gets emitted when the user toggles whether the currently focused item is selected or not. The exact effect of this depend on the selection mode.

Applications should not connect to it, but may emit it with gSignalEmitByName() if they need to control selection programmatically.

There is no default binding for this signal is Ctrl-Space.

iconview

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

unselect-all(iconview, user.data)

A keybinding signal which gets emitted when the user unselects all items.

Applications should not connect to it, but may emit it with gSignalEmitByName() if they need to control selection programmatically.

The default binding for this signal is Ctrl-Shift-a.

iconview

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

Properties

column-spacing [integer : Read / Write]

The column-spacing property specifies the space which is inserted between the columns of the icon view. Allowed values: >= 0 Default value: 6 Since 2.6

columns [integer : Read / Write]

The columns property contains the number of the columns in which the items should be displayed. If it is -1, the number of columns will be chosen automatically to fill the available area. Allowed values: >= -1 Default value: -1 Since 2.6

item-padding [integer : Read / Write]

The item-padding property specifies the padding around each of the icon view's item. Allowed values: >= 0 Default value: 6 Since 2.18

item-width [integer : Read / Write]

The item-width property specifies the width to use for each item. If it is set to -1, the icon view will automatically determine a suitable item size. Allowed values: >= -1 Default value: -1 Since 2.6

margin [integer : Read / Write]

The margin property specifies the space which is inserted at the edges of the icon view. Allowed values: >= 0 Default value: 6 Since 2.6

markup-column [integer : Read / Write]

The ::markup-column property contains the number of the model column containing markup information to be displayed. The markup column must be of type G_TYPE_STRING. If this property and the :text-column property are both set to column numbers, it overrides the text column. If both are set to -1, no texts are displayed. Allowed values: >= -1 Default value: -1 Since 2.6

model [GtkTreeModel : * : Read / Write]

The model for the icon view.

orientation [GtkOrientation : Read / Write]

The orientation property specifies how the cells (i.e. the icon and the text) of the item are positioned relative to each other. Default value: GTK_ORIENTATION_VERTICAL Since 2.6

pixbuf-column [integer : Read / Write]

The ::pixbuf-column property contains the number of the model column containing the pixbufs which are displayed. The pixbuf column must be of type GDK_TYPE_PIXBUF. Setting this property to -1 turns off the display of pixbufs. Allowed values: >= -1 Default value: -1 Since 2.6

reorderable [logical : Read / Write]

The reorderable property specifies if the items can be reordered by DND. Default value: FALSE Since 2.8

row-spacing [integer : Read / Write]

The row-spacing property specifies the space which is inserted between the rows of the icon view. Allowed values: >= 0 Default value: 6 Since 2.6

selection-mode [GtkSelectionMode : Read / Write]

The ::selection-mode property specifies the selection mode of icon view. If the mode is GTK_SELECTION_MULTIPLE, rubberband selection is enabled, for the other modes, only keyboard selection is possible. Default value: GTK_SELECTION_SINGLE Since 2.6

spacing [integer : Read / Write]

The spacing property specifies the space which is inserted between the cells (i.e. the icon and the text) of an item. Allowed values: >= 0 Default value: 0 Since 2.6

text-column [integer : Read / Write]

The ::text-column property contains the number of the model column containing the texts which are displayed. The text column must be of type G_TYPE_STRING. If this property and the :markup-column property are both set to -1, no texts are displayed. Allowed values: >= -1 Default value: -1 Since 2.6

tooltip-column [integer : Read / Write]

The column in the model containing the tooltip texts for the items. Allowed values: >= -1 Default value: -1

Style Properties

selection-box-alpha [raw : Read]

Opacity of the selection box. Default value: 64

selection-box-color [GdkColor : * : Read]

Color of the selection box.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkIconView.html

[Package RGtk2 version 2.20.31 Index]