GtkImage {RGtk2}	R Documentation

GtkImage

Description

A widget displaying an image

Methods and Functions

gtkImageGetIconSet(object)
gtkImageGetImage(object)
gtkImageGetPixbuf(object)
gtkImageGetPixmap(object)
gtkImageGetStock(object)
gtkImageGetAnimation(object)
gtkImageGetIconName(object)
gtkImageGetGicon(object)
gtkImageGetStorageType(object)
gtkImageNewFromIconSet(icon.set, size, show = TRUE)
gtkImageNewFromImage(image = NULL, mask = NULL, show = TRUE)
gtkImageNewFromPixbuf(pixbuf = NULL, show = TRUE)
gtkImageNewFromPixmap(pixmap = NULL, mask = NULL, show = TRUE)
gtkImageNewFromStock(stock.id, size, show = TRUE)
gtkImageNewFromAnimation(animation, show = TRUE)
gtkImageNewFromIconName(icon.name, size)
gtkImageNewFromGicon(icon, size, show = TRUE)
gtkImageSetFromIconSet(object, icon.set, size)
gtkImageSetFromImage(object, gdk.image = NULL, mask = NULL)
gtkImageSetFromPixbuf(object, pixbuf = NULL)
gtkImageSetFromPixmap(object, pixmap, mask = NULL)
gtkImageSetFromStock(object, stock.id, size)
gtkImageSetFromAnimation(object, animation)
gtkImageSetFromIconName(object, icon.name, size)
gtkImageSetFromGicon(object, icon, size)
gtkImageClear(object)
gtkImageNew(show = TRUE)
gtkImageSet(object, val, mask)
gtkImageGet(object)
gtkImageSetPixelSize(object, pixel.size)
gtkImageGetPixelSize(object)
gtkImage(size, mask = NULL, pixmap = NULL, image = NULL, filename, pixbuf = NULL, stock.id, icon.set, animation, icon, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkMisc
                           +----GtkImage

Interfaces

GtkImage implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkImage widget displays an image. Various kinds of object can be displayed as an image; most typically, you would load a GdkPixbuf ("pixel buffer") from a file, and then display that. There's a convenience function to do this, gtkImageNewFromFile, used as follows:

image <- gtkImageNewFromFile("myfile.png")
# or, perhaps more conveniently
image <- gtkImage(file="myfile.png")

If the file isn't loaded successfully, the image will contain a "broken image" icon similar to that used in many web browsers. If you want to handle errors in loading the file yourself, for example by displaying an error message, then load the image with gdkPixbufNewFromFile, then create the GtkImage with gtkImageNewFromPixbuf.

The image file may contain an animation, if so the GtkImage will display an animation (GdkPixbufAnimation) instead of a static image. GtkImage is a subclass of GtkMisc, which implies that you can align it (center, left, right) and add padding to it, using GtkMisc methods. GtkImage is a "no window" widget (has no GdkWindow of its own), so by default does not receive events. If you want to receive events on the image, such as button clicks, place the image inside a GtkEventBox, then connect to the event signals on the event box.

Handling button press events on a GtkImage .

# Handling button-press events on a GtkImage
button_press_callback <- function(event_box, event, data) {
  print(paste("Event box clicked at coordinates ", event[["x"]], ",",
              event[["y"]], sep=""))

  ## Returning TRUE means we handled the event, so the signal 
  ## emission should be stopped (don't call any further 
  ## callbacks that may be connected). Return FALSE
  ## to continue invoking callbacks.
  
  return(TRUE)
}

create_image <- function() {
  image <- gtkImage(file="myfile.png")
  
  event_box <- gtkEventBox()
  event_box$add(image)
  
  
  gSignalConnect(event_box, "button_press_event", button_press_callback, image)
  
  return(image)
}

When handling events on the event box, keep in mind that coordinates in the image may be different from event box coordinates due to the alignment and padding settings on the image (see GtkMisc). The simplest way to solve this is to set the alignment to 0.0 (left/top), and set the padding to zero. Then the origin of the image will be the same as the origin of the event box.

Sometimes an application will want to avoid depending on external data files, such as image files. GTK+ comes with a program to avoid this, called gdk-pixbuf-csource. This program allows you to convert an image into a C variable declaration, which can then be loaded into a GdkPixbuf using gdkPixbufNewFromInline().

Structures

GtkImage

This struct contain private data only and should be accessed by the functions below.

Convenient Construction

gtkImage is the result of collapsing the constructors of GtkImage (gtkImageNew, gtkImageNewFromPixmap, gtkImageNewFromImage, gtkImageNewFromFile, gtkImageNewFromPixbuf, gtkImageNewFromStock, gtkImageNewFromIconSet, gtkImageNewFromAnimation, gtkImageNewFromGicon) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Enums and Flags

GtkImageType

Describes the image data representation used by a GtkImage. If you want to get the image from the widget, you can only get the currently-stored representation. e.g. if the gtkImageGetStorageType returns GTK_IMAGE_PIXBUF, then you can call gtkImageGetPixbuf but not gtkImageGetStock. For empty images, you can request any storage type (call any of the "get" functions), but they will all return NULL values.

empty

there is no image displayed by the widget

pixmap

the widget contains a GdkPixmap

image

the widget contains a GdkImage

pixbuf

the widget contains a GdkPixbuf

stock

the widget contains a stock icon name (see )

icon-set

the widget contains a GtkIconSet

animation

the widget contains a GdkPixbufAnimation

Properties

file [character : * : Read / Write]

Filename to load and display. Default value: NULL

gicon [GIcon : * : Read / Write]

The GIcon displayed in the GtkImage. For themed icons, If the icon theme is changed, the image will be updated automatically. Since 2.14

icon-name [character : * : Read / Write]

The name of the icon in the icon theme. If the icon theme is changed, the image will be updated automatically. Default value: NULL Since 2.6

icon-set [GtkIconSet : * : Read / Write]

Icon set to display.

icon-size [integer : Read / Write]

Symbolic size to use for stock icon, icon set or named icon. Allowed values: >= 0 Default value: 4

image [GdkImage : * : Read / Write]

A GdkImage to display.

mask [GdkPixmap : * : Read / Write]

Mask bitmap to use with GdkImage or GdkPixmap.

pixbuf [GdkPixbuf : * : Read / Write]

A GdkPixbuf to display.

pixbuf-animation [GdkPixbufAnimation : * : Read / Write]

GdkPixbufAnimation to display.

pixel-size [integer : Read / Write]

The "pixel-size" property can be used to specify a fixed size overriding the "icon-size" property for images of type GTK_IMAGE_ICON_NAME. Allowed values: >= -1 Default value: -1 Since 2.6

pixmap [GdkPixmap : * : Read / Write]

A GdkPixmap to display.

stock [character : * : Read / Write]

Stock ID for a stock image to display. Default value: NULL

storage-type [GtkImageType : Read]

The representation being used for image data. Default value: GTK_IMAGE_EMPTY

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkImage.html

[Package RGtk2 version 2.20.31 Index]