GtkImageMenuItem {RGtk2}	R Documentation

GtkImageMenuItem

Description

A menu item with an icon

Methods and Functions

gtkImageMenuItemSetImage(object, image = NULL)
gtkImageMenuItemGetImage(object)
gtkImageMenuItemNew(show = TRUE)
gtkImageMenuItemNewFromStock(stock.id, accel.group, show = TRUE)
gtkImageMenuItemNewWithLabel(label, show = TRUE)
gtkImageMenuItemNewWithMnemonic(label, show = TRUE)
gtkImageMenuItemGetUseStock(object)
gtkImageMenuItemSetUseStock(object, use.stock)
gtkImageMenuItemGetAlwaysShowImage(object)
gtkImageMenuItemSetAlwaysShowImage(object, always.show)
gtkImageMenuItemSetAccelGroup(object, accel.group)
gtkImageMenuItem(label, stock.id, accel.group, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkItem
                                       +----GtkMenuItem
                                             +----GtkImageMenuItem

Interfaces

GtkImageMenuItem implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

A GtkImageMenuItem is a menu item which has an icon next to the text label.

Note that the user can disable display of menu icons, so make sure to still fill in the text label.

Structures

GtkImageMenuItem

undocumented

Convenient Construction

gtkImageMenuItem is the result of collapsing the constructors of GtkImageMenuItem (gtkImageMenuItemNew, gtkImageMenuItemNewWithLabel, gtkImageMenuItemNewWithMnemonic, gtkImageMenuItemNewFromStock) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Properties

accel-group [GtkAccelGroup : * : Write]

The Accel Group to use for stock accelerator keys Since 2.16

always-show-image [logical : Read / Write / Construct]

If TRUE, the menu item will ignore the "gtk-menu-images" setting and always show the image, if available.

Use this property if the menuitem would be useless or hard to use without the image. Default value: FALSE Since 2.16

image [GtkWidget : * : Read / Write]

Child widget to appear next to the menu text.

use-stock [logical : Read / Write / Construct]

If TRUE, the label set in the menuitem is used as a stock id to select the stock item for the item. Default value: FALSE Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkImageMenuItem.html

[Package RGtk2 version 2.20.31 Index]