GtkInputDialog {RGtk2}	R Documentation

GtkInputDialog

Description

Configure devices for the XInput extension

Methods and Functions

gtkInputDialogNew(show = TRUE)
gtkInputDialog(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkWindow
                                       +----GtkDialog
                                             +----GtkInputDialog

Interfaces

GtkInputDialog implements AtkImplementorIface and GtkBuildable.

Detailed Description

GtkInputDialog displays a dialog which allows the user to configure XInput extension devices. For each device, they can control the mode of the device (disabled, screen-relative, or window-relative), the mapping of axes to coordinates, and the mapping of the devices function keys to key press events. GtkInputDialog contains two buttons to which the application can connect; one for closing the dialog, and one for saving the changes. No actions are bound to these by default. The changes that the user makes take effect immediately.

As of GTK+ 2.20, GtkInputDialog has been deprecated since it is too specialized.

Structures

GtkInputDialog

WARNING: GtkInputDialog is deprecated and should not be used in newly-written code. undocumented

Convenient Construction

gtkInputDialog is the equivalent of gtkInputDialogNew.

Signals

disable-device(inputdialog, deviceid, user.data)

This signal is emitted when the user changes the mode of a device from a GDK_MODE_SCREEN or GDK_MODE_WINDOW to GDK_MODE_ENABLED.

inputdialog

the object which received the signal.

deviceid

The ID of the newly disabled device.

user.data

user data set when the signal handler was connected.

enable-device(inputdialog, deviceid, user.data)

This signal is emitted when the user changes the mode of a device from GDK_MODE_DISABLED to a GDK_MODE_SCREEN or GDK_MODE_WINDOW.

inputdialog

the object which received the signal.

deviceid

The ID of the newly enabled device.

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkInputDialog.html

[Package RGtk2 version 2.20.31 Index]