GtkInvisible {RGtk2}	R Documentation

GtkInvisible

Description

A widget which is not displayed

Methods and Functions

gtkInvisibleNew(show = TRUE)
gtkInvisibleNewForScreen(screen, show = TRUE)
gtkInvisibleSetScreen(object, screen)
gtkInvisibleGetScreen(object)
gtkInvisible(screen, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkInvisible

Interfaces

GtkInvisible implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkInvisible widget is used internally in GTK+, and is probably not very useful for application developers.

It is used for reliable pointer grabs and selection handling in the code for drag-and-drop.

Structures

GtkInvisible

The GtkInvisible struct contains no public fields.

Convenient Construction

gtkInvisible is the result of collapsing the constructors of GtkInvisible (gtkInvisibleNew, gtkInvisibleNewForScreen) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Properties

screen [GdkScreen : * : Read / Write]

The screen where this window will be displayed.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkInvisible.html

[Package RGtk2 version 2.20.31 Index]