GtkItem {RGtk2}	R Documentation

GtkItem

Description

Abstract base class for GtkMenuItem, GtkListItem and GtkTreeItem

Methods and Functions

gtkItemSelect(object)
gtkItemDeselect(object)
gtkItemToggle(object)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkItem
                                       +----GtkMenuItem
                                       +----GtkListItem
                                       +----GtkTreeItem

Interfaces

GtkItem implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkItem widget is an abstract base class for GtkMenuItem, GtkListItem and GtkTreeItem.

Structures

GtkItem

The GtkItem struct contains private data only, and should be accessed using the functions below.

Signals

deselect(item, user.data)

Emitted when the item is deselected.

item

the object which received the signal.

user.data

user data set when the signal handler was connected.

select(item, user.data)

Emitted when the item is selected.

item

the object which received the signal.

user.data

user data set when the signal handler was connected.

toggle(item, user.data)

Emitted when the item is toggled.

item

the object which received the signal.

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkItem.html

[Package RGtk2 version 2.20.31 Index]