GtkItemFactory {RGtk2}	R Documentation

GtkItemFactory

Description

A factory for menus

Methods and Functions

gtkItemFactoryNew(container.type, path, accel.group = NULL)
gtkItemFactoryConstruct(object, container.type, path, accel.group)
gtkItemFactoryAddForeign(accel.widget, full.path, accel.group, keyval, modifiers)
gtkItemFactoryFromWidget(widget)
gtkItemFactoryPathFromWidget(widget)
gtkItemFactoryGetItem(object, path)
gtkItemFactoryGetWidget(object, path)
gtkItemFactoryGetWidgetByAction(object, action)
gtkItemFactoryGetItemByAction(object, action)
gtkItemFactoryCreateItem(object, entry, callback.data = NULL, callback.type)
gtkItemFactoryCreateItems(object, entries, callback.data = NULL)
gtkItemFactoryCreateItemsAc(object, entries, callback.data, callback.type)
gtkItemFactoryDeleteItem(object, path)
gtkItemFactoryDeleteEntry(object, entry)
gtkItemFactoryDeleteEntries(object, entries)
gtkItemFactoryPopup(object, x, y, mouse.button, time = "GDK_CURRENT_TIME")
gtkItemFactoryPopupWithData(object, popup.data, x, y, mouse.button, time = "GDK_CURRENT_TIME")
gtkItemFactoryPopupData(object)
gtkItemFactoryPopupDataFromWidget(widget)
gtkItemFactoryFromPath(path)
gtkItemFactoryCreateMenuEntries(entries)
gtkItemFactoriesPathDelete(ifactory.path, path)
gtkItemFactorySetTranslateFunc(object, func, data = NULL)
gtkItemFactory(container.type, path, accel.group = NULL)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkItemFactory

Detailed Description

As of GTK+ 2.4, GtkItemFactory has been deprecated in favour of GtkUIManager.

Structures

GtkItemFactory

WARNING: GtkItemFactory is deprecated and should not be used in newly-written code. undocumented

GtkItemFactoryEntry

WARNING: GtkItemFactoryEntry is deprecated and should not be used in newly-written code. undocumented GtkItemFactoryEntry is a transparent-type.

Convenient Construction

gtkItemFactory is the equivalent of gtkItemFactoryNew.

User Functions

GtkPrintFunc()

undocumented

GtkTranslateFunc(path, func.data)

The function used to translate messages in e.g. GtkIconFactory and GtkActionGroup.

path

The id of the message. In GtkItemFactory this will be a path from a GtkItemFactoryEntry, in GtkActionGroup, it will be a label or tooltip from a GtkActionEntry.

func.data

user data passed in when registering the function

Returns: [character] the translated message

GtkItemFactoryCallback()

undocumented

GtkItemFactoryCallback1()

undocumented

GtkItemFactoryCallback2()

undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkItemFactory.html

[Package RGtk2 version 2.20.31 Index]