GtkLayout {RGtk2}	R Documentation

GtkLayout

Description

Infinite scrollable area containing child widgets and/or custom drawing

Methods and Functions

gtkLayoutNew(hadjustment = NULL, vadjustment = NULL, show = TRUE)
gtkLayoutPut(object, child.widget, x, y)
gtkLayoutMove(object, child.widget, x, y)
gtkLayoutSetSize(object, width, height)
gtkLayoutGetSize(object)
gtkLayoutFreeze(object)
gtkLayoutThaw(object)
gtkLayoutGetHadjustment(object)
gtkLayoutGetVadjustment(object)
gtkLayoutSetHadjustment(object, adjustment = NULL)
gtkLayoutSetVadjustment(object, adjustment = NULL)
gtkLayoutGetBinWindow(object)
gtkLayout(hadjustment = NULL, vadjustment = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkLayout

Interfaces

GtkLayout implements AtkImplementorIface and GtkBuildable.

Detailed Description

GtkLayout is similar to GtkDrawingArea in that it's a "blank slate" and doesn't do anything but paint a blank background by default. It's different in that it supports scrolling natively (you can add it to a GtkScrolledWindow), and it can contain child widgets, since it's a GtkContainer. However if you're just going to draw, a GtkDrawingArea is a better choice since it has lower overhead.

When handling expose events on a GtkLayout, you must draw to GTK_LAYOUT (layout)->bin_window, rather than to GTK_WIDGET (layout)->window, as you would for a drawing area.

Structures

GtkLayout

undocumented

binWindow

[GdkWindow]

Convenient Construction

gtkLayout is the equivalent of gtkLayoutNew.

Signals

set-scroll-adjustments(horizontal, vertical, user.data)

Set the scroll adjustments for the layout. Usually scrolled containers like GtkScrolledWindow will emit this signal to connect two instances of GtkScrollbar to the scroll directions of the GtkLayout.

horizontal

the horizontal GtkAdjustment

vertical

the vertical GtkAdjustment

user.data

user data set when the signal handler was connected.

Properties

hadjustment [GtkAdjustment : * : Read / Write]

The GtkAdjustment for the horizontal position.

height [numeric : Read / Write]

The height of the layout. Allowed values: <= G_MAXINT Default value: 100

vadjustment [GtkAdjustment : * : Read / Write]

The GtkAdjustment for the vertical position.

width [numeric : Read / Write]

The width of the layout. Allowed values: <= G_MAXINT Default value: 100

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkLayout.html

See Also

GtkDrawingArea GtkScrolledWindow

[Package RGtk2 version 2.20.31 Index]