GtkLinkButton {RGtk2}	R Documentation

GtkLinkButton

Description

Create buttons bound to a URL

Methods and Functions

gtkLinkButtonNew(uri)
gtkLinkButtonNewWithLabel(uri, label = NULL, show = TRUE)
gtkLinkButtonGetUri(object)
gtkLinkButtonSetUri(object, uri)
gtkLinkButtonSetUriHook(func, data)
gtkLinkButtonGetVisited(object)
gtkLinkButtonSetVisited(object, visited)
gtkLinkButton(uri, label = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkButton
                                       +----GtkLinkButton

Interfaces

GtkLinkButton implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

A GtkLinkButton is a GtkButton with a hyperlink, similar to the one used by web browsers, which triggers an action when clicked. It is useful to show quick links to resources.

A link button is created by calling either gtkLinkButtonNew or gtkLinkButtonNewWithLabel. If using the former, the URI you pass to the constructor is used as a label for the widget.

The URI bound to a GtkLinkButton can be set specifically using gtkLinkButtonSetUri, and retrieved using gtkLinkButtonGetUri. GtkLinkButton offers a global hook, which is called when the used clicks on it: see gtkLinkButtonSetUriHook. GtkLinkButton was added in GTK+ 2.10.

Structures

GtkLinkButton

The GtkLinkButton struct contains private data only, and should be manipulated using the functions below.

Convenient Construction

gtkLinkButton is the equivalent of gtkLinkButtonNewWithLabel.

User Functions

GtkLinkButtonUriFunc(button, link., user.data)

The type of a function which is called when the GtkLinkButton is clicked.

button

the GtkLinkButton which was clicked

link.

the URI to which the clicked GtkLinkButton points

user.data

user data that was passed when the function was registered with gtkLinkButtonSetUriHook

Properties

uri [character : * : Read / Write]

The URI bound to this button. Default value: NULL Since 2.10

visited [logical : Read / Write]

The 'visited' state of this button. A visited link is drawn in a different color. Default value: FALSE Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkLinkButton.html

See Also

GtkButton

[Package RGtk2 version 2.20.31 Index]