GtkList {RGtk2}	R Documentation

GtkList

Description

Widget for packing a list of selectable items

Methods and Functions

gtkListNew(show = TRUE)
gtkListInsertItems(object, items, position)
gtkListAppendItems(object, items)
gtkListPrependItems(object, items)
gtkListRemoveItems(object, items)
gtkListClearItems(object, start, end)
gtkListSelectItem(object, item)
gtkListUnselectItem(object, item)
gtkListSelectChild(object, child)
gtkListUnselectChild(object, child)
gtkListChildPosition(object, child)
gtkListSetSelectionMode(object, mode)
gtkListExtendSelection(object, scroll.type, position, auto.start.selection)
gtkListStartSelection(object)
gtkListEndSelection(object)
gtkListSelectAll(object)
gtkListUnselectAll(object)
gtkListScrollHorizontal(object, scroll.type, position)
gtkListScrollVertical(object, scroll.type, position)
gtkListToggleAddMode(object)
gtkListToggleFocusRow(object)
gtkListToggleRow(object, item)
gtkListUndoSelection(object)
gtkListEndDragSelection(object)
gtkList(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkList

Interfaces

GtkList implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkList widget is a container whose children are displayed vertically in order, and can be selected. The list has many selection modes, which are programmer selective and depend on how many elements are able to be selected at the same time.

GtkList has been deprecated since GTK+ 2.0 and should not be used in newly written code. Use GtkTreeView instead.

Structures

GtkList

WARNING: GtkList is deprecated and should not be used in newly-written code. undocumented

Convenient Construction

gtkList is the equivalent of gtkListNew.

Signals

select-child(list, widget, user.data)

The child widget has just been selected.

list

the object which received the signal.

widget

the newly selected child.

user.data

user data set when the signal handler was connected.

selection-changed(list, user.data)

The selection of the widget has just changed.

list

the object which received the signal.

user.data

user data set when the signal handler was connected.

unselect-child(list, widget, user.data)

The child widget has just been unselected.

list

the object which received the signal.

widget

the newly unselected child.

user.data

user data set when the signal handler was connected.

Properties

selection-mode [GtkSelectionMode : Read / Write]

Default value: GTK_SELECTION_NONE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkList.html

[Package RGtk2 version 2.20.31 Index]