GtkListItem {RGtk2}	R Documentation

GtkListItem

Description

An item in a GtkList

Methods and Functions

gtkListItemNew(show = TRUE)
gtkListItemNewWithLabel(label, show = TRUE)
gtkListItemSelect(object)
gtkListItemDeselect(object)
gtkListItem(label, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkItem
                                       +----GtkListItem

Interfaces

GtkListItem implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkListItem widget is used for each item in a GtkList.

GtkList has has been deprecated since GTK+ 2.0 and should not be used in newly written code. Use GtkTreeView instead.

Structures

GtkListItem

WARNING: GtkListItem is deprecated and should not be used in newly-written code. The GtkListItem struct contains private data only, and should only be accessed using the functions below.

Convenient Construction

gtkListItem is the result of collapsing the constructors of GtkListItem (gtkListItemNew, gtkListItemNewWithLabel) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Signals

end-selection(listitem, user.data)

undocumented

listitem

the object which received the signal.

user.data

user data set when the signal handler was connected.

extend-selection(listitem, user.data)

undocumented

listitem

the object which received the signal.

user.data

user data set when the signal handler was connected.

scroll-horizontal(listitem, user.data)

undocumented

listitem

the object which received the signal.

user.data

user data set when the signal handler was connected.

scroll-vertical(listitem, user.data)

undocumented

listitem

the object which received the signal.

user.data

user data set when the signal handler was connected.

select-all(listitem, user.data)

undocumented

listitem

the object which received the signal.

user.data

user data set when the signal handler was connected.

start-selection(listitem, user.data)

undocumented

listitem

the object which received the signal.

user.data

user data set when the signal handler was connected.

toggle-add-mode(listitem, user.data)

undocumented

listitem

the object which received the signal.

user.data

user data set when the signal handler was connected.

toggle-focus-row(listitem, user.data)

undocumented

listitem

the object which received the signal.

user.data

user data set when the signal handler was connected.

undo-selection(listitem, user.data)

undocumented

listitem

the object which received the signal.

user.data

user data set when the signal handler was connected.

unselect-all(listitem, user.data)

undocumented

listitem

the object which received the signal.

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkListItem.html

[Package RGtk2 version 2.20.31 Index]