GtkListStore {RGtk2}	R Documentation

GtkListStore

Description

A list-like data structure that can be used with the GtkTreeView

Methods and Functions

gtkListStoreNew(...)
gtkListStoreNewv(value)
gtkListStoreSetColumnTypes(object, types)
gtkListStoreSet(object, iter, ...)
gtkListStoreSetValue(object, iter, column, value)
gtkListStoreSetValuesv(object, iter, columns, values)
gtkListStoreRemove(object, iter)
gtkListStoreInsert(object, position)
gtkListStoreInsertBefore(object, sibling)
gtkListStoreInsertAfter(object, sibling)
gtkListStoreInsertWithValues(object, position, ...)
gtkListStoreInsertWithValuesv(object, position, columns, values)
gtkListStorePrepend(object, iter)
gtkListStoreAppend(object)
gtkListStoreClear(object)
gtkListStoreIterIsValid(object, iter)
gtkListStoreReorder(object, new.order)
gtkListStoreSwap(object, a, b)
gtkListStoreMoveBefore(object, iter, position = NULL)
gtkListStoreMoveAfter(object, iter, position = NULL)
gtkListStore(..., value)

Hierarchy

GObject
   +----GtkListStore

Interfaces

GtkListStore implements GtkTreeModel, GtkTreeDragSource, GtkTreeDragDest, GtkTreeSortable and GtkBuildable.

Detailed Description

The GtkListStore object is a list model for use with a GtkTreeView widget. It implements the GtkTreeModel interface, and consequentialy, can use all of the methods available there. It also implements the GtkTreeSortable interface so it can be sorted by the view. Finally, it also implements the tree drag and drop interfaces.

The GtkListStore can accept most GObject types as a column type, though it can't accept all custom types. Internally, it will keep a copy of data passed in (such as a string or a boxed pointer). Columns that accept GObjects are handled a little differently. The GtkListStore will keep a reference to the object instead of copying the value. As a result, if the object is modified, it is up to the application writer to call gtk.tree.model.row.changed to emit the "row_changed" signal. This most commonly affects lists with GdkPixbufs stored.

Creating a simple list store.

list_store <- gtk_list_store_new ("character", "integer", "logical")

sapply(character_vector,
       function(string) {
         ## Add a new row to the model
         iter <- list_store$append(iter)$iter
         list_store$set(iter, 0, string, 1, i, 2,  FALSE)
       })

## Modify a particular row
path <- gtkTreePathNewFromString("4")
iter <- list_store$getIter(path)$iter
list_store$set(iter, 2, TRUE)

Performance Considerations

Internally, the GtkListStore was implemented with a linked list with a tail pointer prior to GTK+ 2.6. As a result, it was fast at data insertion and deletion, and not fast at random data access. The GtkListStore sets the GTK_TREE_MODEL_ITERS_PERSIST flag, which means that GtkTreeIters can be cached while the row exists. Thus, if access to a particular row is needed often and your code is expected to run on older versions of GTK+, it is worth keeping the iter around.

It is important to note that only the methods gtkListStoreInsertWithValues and gtkListStoreInsertWithValuesv are atomic, in the sense that the row is being appended to the store and the values filled in in a single operation with regard to GtkTreeModel signaling. In contrast, using e.g. gtkListStoreAppend and then gtkListStoreSet will first create a row, which triggers the gtkTreeModelRowInserted signal on GtkListStore. The row, however, is still empty, and any signal handler connecting to "row-inserted" on this particular store should be prepared for the situation that the row might be empty. This is especially important if you are wrapping the GtkListStore inside a GtkTreeModelFilter and are using a GtkTreeModelFilterVisibleFunc. Using any of the non-atomic operations to append rows to the GtkListStore will cause the GtkTreeModelFilterVisibleFunc to be visited with an empty row first; the function must be prepared for that.

GtkListStore as GtkBuildable

The GtkListStore implementation of the GtkBuildable interface allows to specify the model columns with a <columns> element that may contain multiple <column> elements, each specifying one model column. The "type" attribute specifies the data type for the column.

Additionally, it is possible to specify content for the list store in the UI definition, with the <data> element. It can contain multiple <row> elements, each specifying to content for one row of the list model. Inside a <row>, the <col> elements specify the content for individual cells.

Note that it is probably more common to define your models in the code, and one might consider it a layering violation to specify the content of a list store in a UI definition, data, not presentation, and common wisdom is to separate the two, as far as possible.

A UI Definition fragment for a list store

<object class="GtkListStore">
  <columns>
    <column type="gchararray"/>
    <column type="gchararray"/>
    <column type="gint"/>
  </columns>
  <data>
    <row>
      <col id="0">John</col>
      <col id="1">Doe</col>
      <col id="2">25</col>
    </row>
    <row>
      <col id="0">Johan</col>
      <col id="1">Dahlin</col>
      <col id="2">50</col>
    </row>
  </data>
</object>

Structures

GtkListStore

undocumented

Convenient Construction

gtkListStore is the result of collapsing the constructors of GtkListStore (gtkListStoreNew, gtkListStoreNewv) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkListStore.html

See Also

GtkTreeModel GtkTreeStore

[Package RGtk2 version 2.20.31 Index]