GtkMenuBar {RGtk2}	R Documentation

GtkMenuBar

Description

A subclass widget for GtkMenuShell which holds GtkMenuItem widgets

Methods and Functions

gtkMenuBarNew(show = TRUE)
gtkMenuBarSetPackDirection(object, pack.dir)
gtkMenuBarGetPackDirection(object)
gtkMenuBarSetChildPackDirection(object, child.pack.dir)
gtkMenuBarGetChildPackDirection(object)
gtkMenuBar(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkMenuShell
                                 +----GtkMenuBar

Interfaces

GtkMenuBar implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkMenuBar is a subclass of GtkMenuShell which contains one to many GtkMenuItem. The result is a standard menu bar which can hold many menu items. GtkMenuBar allows for a shadow type to be set for aesthetic purposes. The shadow types are defined in the gtk_menu_bar_set_shadow_type function.

Structures

GtkMenuBar

The GtkMenuBar struct contains the following fields. (These fields should be considered read-only. They should never be set by an application.)

Convenient Construction

gtkMenuBar is the equivalent of gtkMenuBarNew.

Enums and Flags

GtkPackDirection

Determines how widgets should be packed insided menubars and menuitems contained in menubars.

ltr

Widgets are packed left-to-right.

rtl

Widgets are packed right-to-left.

ttb

Widgets are packed top-to-bottom.

btt

Widgets are packed bottom-to-top.

Properties

child-pack-direction [GtkPackDirection : Read / Write]

The child pack direction of the menubar. It determines how the widgets contained in child menuitems are arranged. Default value: GTK_PACK_DIRECTION_LTR Since 2.8

pack-direction [GtkPackDirection : Read / Write]

The pack direction of the menubar. It determines how menuitems are arranged in the menubar. Default value: GTK_PACK_DIRECTION_LTR Since 2.8

Style Properties

internal-padding [integer : Read]

Amount of border space between the menubar shadow and the menu items. Allowed values: >= 0 Default value: 1

shadow-type [GtkShadowType : Read]

Style of bevel around the menubar. Default value: GTK_SHADOW_OUT

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkMenuBar.html

See Also

GtkMenuShell GtkMenu GtkMenuItem

[Package RGtk2 version 2.20.31 Index]