GtkMenuShell {RGtk2}	R Documentation

GtkMenuShell

Description

A base class for menu objects

Methods and Functions

gtkMenuShellAppend(object, child)
gtkMenuShellPrepend(object, child)
gtkMenuShellInsert(object, child, position)
gtkMenuShellDeactivate(object)
gtkMenuShellSelectItem(object, menu.item)
gtkMenuShellSelectFirst(object, search.sensitive)
gtkMenuShellDeselect(object)
gtkMenuShellActivateItem(object, menu.item, force.deactivate)
gtkMenuShellCancel(object)
gtkMenuShellSetTakeFocus(object, take.focus)
gtkMenuShellGetTakeFocus(object)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkMenuShell
                                 +----GtkMenuBar
                                 +----GtkMenu

Interfaces

GtkMenuShell implements AtkImplementorIface and GtkBuildable.

Detailed Description

A GtkMenuShell is the abstract base class used to derive the GtkMenu and GtkMenuBar subclasses.

A GtkMenuShell is a container of GtkMenuItem objects arranged in a list which can be navigated, selected, and activated by the user to perform application functions. A GtkMenuItem can have a submenu associated with it, allowing for nested hierarchical menus.

Structures

GtkMenuShell

The GtkMenuShell struct contains the following fields. (These fields should be considered read-only. They should never be set by an application.)

list *children; 	 The list of GtkMenuItem objects contained by this GtkMenuShell .

Enums and Flags

GtkMenuDirectionType

An enumeration representing directional movements within a menu.

parent

To the parent menu shell.

child

To the submenu, if any, associated with the item.

next

To the next menu item.

prev

To the previous menu item.

Signals

activate-current(menushell, force.hide, user.data)

An action signal that activates the current menu item within the menu shell.

menushell

the object which received the signal.

force.hide

if TRUE, hide the menu after activating the menu item.

user.data

user data set when the signal handler was connected.

cancel(menushell, user.data)

An action signal which cancels the selection within the menu shell. Causes the GtkMenuShell::selection-done signal to be emitted.

menushell

the object which received the signal.

user.data

user data set when the signal handler was connected.

cycle-focus(menushell, user.data)

undocumented

menushell

the object which received the signal.

user.data

user data set when the signal handler was connected.

deactivate(menushell, user.data)

This signal is emitted when a menu shell is deactivated.

menushell

the object which received the signal.

user.data

user data set when the signal handler was connected.

move-current(menushell, direction, user.data)

An action signal which moves the current menu item in the direction specified by direction.

menushell

the object which received the signal.

direction

the direction to move.

user.data

user data set when the signal handler was connected.

move-selected(menu.shell, distance, user.data)

The ::move-selected signal is emitted to move the selection to another item. Since 2.12

menu.shell

the object on which the signal is emitted

distance

+1 to move to the next item, -1 to move to the previous

user.data

user data set when the signal handler was connected.

Returns: [logical] TRUE to stop the signal emission, FALSE to continue

selection-done(menushell, user.data)

This signal is emitted when a selection has been completed within a menu shell.

menushell

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

take-focus [logical : Read / Write]

A boolean that determines whether the menu and its submenus grab the keyboard focus. See gtkMenuShellSetTakeFocus and gtkMenuShellGetTakeFocus. Default value: TRUE Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkMenuShell.html

[Package RGtk2 version 2.20.31 Index]