GtkMenuToolButton {RGtk2}	R Documentation

GtkMenuToolButton

Description

A GtkToolItem containing a button with an additional dropdown menu

Methods and Functions

gtkMenuToolButtonNew(icon.widget, label, show = TRUE)
gtkMenuToolButtonNewFromStock(stock.id)
gtkMenuToolButtonSetMenu(object, menu)
gtkMenuToolButtonGetMenu(object)
gtkMenuToolButtonSetArrowTooltip(object, tooltips, tip.text = NULL, tip.private = NULL)
gtkMenuToolButtonSetArrowTooltip(object, tooltips, tip.text = NULL, tip.private = NULL)
gtkMenuToolButtonSetArrowTooltipText(object, text)
gtkMenuToolButtonSetArrowTooltipMarkup(object, markup)
gtkMenuToolButton(icon.widget, label, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkToolItem
                                       +----GtkToolButton
                                             +----GtkMenuToolButton

Interfaces

GtkMenuToolButton implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

A GtkMenuToolButton is a GtkToolItem that contains a button and a small additional button with an arrow. When clicked, the arrow button pops up a dropdown menu.

Use gtkMenuToolButtonNew to create a new GtkMenuToolButton. Use gtkMenuToolButtonNewFromStock to create a new GtkMenuToolButton containing a stock item.

Structures

GtkMenuToolButton

The GtkMenuToolButton struct contains only private data and should only be accessed through the functions described below.

Convenient Construction

gtkMenuToolButton is the equivalent of gtkMenuToolButtonNew.

Signals

show-menu(button, user.data)

The ::show-menu signal is emitted before the menu is shown.

It can be used to populate the menu on demand, using gtkMenuToolButtonGetMenu.

Note that even if you populate the menu dynamically in this way, you must set an empty menu on the GtkMenuToolButton beforehand, since the arrow is made insensitive if the menu is not set.

button

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

Properties

menu [GtkMenu : * : Read / Write]

The dropdown menu.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkMenuToolButton.html

[Package RGtk2 version 2.20.31 Index]