GtkMisc {RGtk2}	R Documentation

GtkMisc

Description

Base class for widgets with alignments and padding

Methods and Functions

gtkMiscSetAlignment(object, xalign, yalign)
gtkMiscSetPadding(object, xpad, ypad)
gtkMiscGetAlignment(object)
gtkMiscGetPadding(object)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkMisc
                           +----GtkLabel
                           +----GtkArrow
                           +----GtkImage
                           +----GtkPixmap

Interfaces

GtkMisc implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkMisc widget is an abstract widget which is not useful itself, but is used to derive subclasses which have alignment and padding attributes.

The horizontal and vertical padding attributes allows extra space to be added around the widget.

The horizontal and vertical alignment attributes enable the widget to be positioned within its allocated area. Note that if the widget is added to a container in such a way that it expands automatically to fill its allocated area, the alignment settings will not alter the widgets position.

Structures

GtkMisc

The GtkMisc struct contains the following fields. (These fields should be considered read-only. They should never be set by an application.)

numeric xalign ; 	 the horizontal alignment, from 0 (left) to 1 (right).

numeric yalign ; 	 the vertical alignment, from 0 (top) to 1 (bottom).

integer xpad ; 	 the amount of space to add on the left and right of the widget, in pixels.

integer ypad ; 	 the amount of space to add on the top and bottom of the widget, in pixels.

Properties

xalign [numeric : Read / Write]

The horizontal alignment, from 0 (left) to 1 (right). Reversed for RTL layouts. Allowed values: [0,1] Default value: 0.5

xpad [integer : Read / Write]

The amount of space to add on the left and right of the widget, in pixels. Allowed values: >= 0 Default value: 0

yalign [numeric : Read / Write]

The vertical alignment, from 0 (top) to 1 (bottom). Allowed values: [0,1] Default value: 0.5

ypad [integer : Read / Write]

The amount of space to add on the top and bottom of the widget, in pixels. Allowed values: >= 0 Default value: 0

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkMisc.html

[Package RGtk2 version 2.20.31 Index]