GtkOffscreenWindow {RGtk2}	R Documentation

GtkOffscreenWindow

Description

A toplevel container widget used to manage offscreen rendering of child widgets.

Methods and Functions

gtkOffscreenWindowNew(show = TRUE)
gtkOffscreenWindowGetPixmap(object)
gtkOffscreenWindowGetPixbuf(object)
gtkOffscreenWindow(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkWindow
                                       +----GtkOffscreenWindow

Interfaces

GtkOffscreenWindow implements AtkImplementorIface and GtkBuildable.

Detailed Description

GtkOffscreenWindow is strictly intended to be used for obtaining snapshots of widgets that are not part of a normal widget hierarchy. It differs from gtkWidgetGetSnapshot in that the widget you want to get a snapshot of need not be displayed on the user's screen as a part of a widget hierarchy. However, since GtkOffscreenWindow is a toplevel widget you cannot obtain snapshots of a full window with it since you cannot pack a toplevel widget in another toplevel.

The idea is to take a widget and manually set the state of it, add it to a GtkOffscreenWindow and then retrieve the snapshot as a GdkPixmap or GdkPixbuf. GtkOffscreenWindow derives from GtkWindow only as an implementation detail. Applications should not use any API specific to GtkWindow to operate on this object. It should be treated as a GtkBin that has no parent widget.

When contained offscreen widgets are redrawn, GtkOffscreenWindow will emit a "damage-event" signal.

Structures

GtkOffscreenWindow

undocumented

Convenient Construction

gtkOffscreenWindow is the equivalent of gtkOffscreenWindowNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkOffscreenWindow.html

[Package RGtk2 version 2.20.31 Index]