GtkOldEditable {RGtk2}	R Documentation

GtkOldEditable

Description

Base class for text-editing widgets

Methods and Functions

gtkOldEditableClaimSelection(object, claim, time = "GDK_CURRENT_TIME")
gtkOldEditableChanged(object)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkOldEditable
                           +----GtkText

Interfaces

GtkOldEditable implements AtkImplementorIface, GtkBuildable and GtkEditable.

Detailed Description

GtkOldEditable has been deprecated since GTK+ 2.0 and should not be used in newly written code. Use the GtkEditable interface instead.

Structures

GtkOldEditable

WARNING: GtkOldEditable is deprecated and should not be used in newly-written code. undocumented

User Functions

GtkTextFunction()

undocumented

Signals

activate(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

copy-clipboard(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

cut-clipboard(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

kill-char(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

kill-line(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

kill-word(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

move-cursor(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

move-page(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

move-to-column(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

move-to-row(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

move-word(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

paste-clipboard(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

set-editable(oldeditable, user.data)

undocumented

oldeditable

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

editable [logical : Read / Write]

Default value: FALSE

text-position [integer : Read / Write]

Default value: 0

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkOldEditable.html

[Package RGtk2 version 2.20.31 Index]