GtkOptionMenu {RGtk2}	R Documentation

GtkOptionMenu

Description

A widget used to choose from a list of valid choices

Methods and Functions

gtkOptionMenuNew(show = TRUE)
gtkOptionMenuGetMenu(object)
gtkOptionMenuSetMenu(object, menu)
gtkOptionMenuRemoveMenu(object)
gtkOptionMenuSetHistory(object, index)
gtkOptionMenuGetHistory(object)
gtkOptionMenu(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkButton
                                       +----GtkOptionMenu

Interfaces

GtkOptionMenu implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

A GtkOptionMenu is a widget that allows the user to choose from a list of valid choices. The GtkOptionMenu displays the selected choice. When activated the GtkOptionMenu displays a popup GtkMenu which allows the user to make a new choice.

Using a GtkOptionMenu is simple; build a GtkMenu, by calling gtkMenuNew, then appending menu items to it with gtkMenuShellAppend. Set that menu on the option menu with gtkOptionMenuSetMenu. Set the selected menu item with gtkOptionMenuSetHistory; connect to the "changed" signal on the option menu; in the "changed" signal, check the new selected menu item with gtkOptionMenuGetHistory.

As of GTK+ 2.4, GtkOptionMenu has been deprecated in favor of GtkComboBox.

Structures

GtkOptionMenu

WARNING: GtkOptionMenu has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkComboBox instead. The GtkOptionMenu struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkOptionMenu is the equivalent of gtkOptionMenuNew.

Signals

changed(optionmenu, user.data)

undocumented

optionmenu

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

menu [GtkMenu : * : Read / Write]

The menu of options.

Style Properties

indicator-size [GtkRequisition : * : Read]

Size of dropdown indicator.

indicator-spacing [GtkBorder : * : Read]

Spacing around indicator.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkOptionMenu.html

[Package RGtk2 version 2.20.31 Index]