GtkPaned {RGtk2}	R Documentation

GtkPaned

Description

Base class for widgets with two adjustable panes

Methods and Functions

gtkPanedAdd1(object, child)
gtkPanedAdd2(object, child)
gtkPanedPack1(object, child, resize = FALSE, shrink = TRUE)
gtkPanedPack2(object, child, resize = TRUE, shrink = TRUE)
gtkPanedGetChild1(object)
gtkPanedGetChild2(object)
gtkPanedSetPosition(object, position)
gtkPanedGetPosition(object)
gtkPanedGetHandleWindow(object)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkPaned
                                 +----GtkHPaned
                                 +----GtkVPaned

Interfaces

GtkPaned implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

GtkPaned is the base class for widgets with two panes, arranged either horizontally (GtkHPaned) or vertically (GtkVPaned). Child widgets are added to the panes of the widget with gtkPanedPack1 and gtkPanedPack2. The division beween the two children is set by default from the size requests of the children, but it can be adjusted by the user.

A paned widget draws a separator between the two child widgets and a small handle that the user can drag to adjust the division. It does not draw any relief around the children or around the separator. (The space in which the separator is called the gutter.) Often, it is useful to put each child inside a GtkFrame with the shadow type set to GTK_SHADOW_IN so that the gutter appears as a ridge. No separator is drawn if one of the children is missing.

Each child has two options that can be set, resize and shrink. If resize is true, then when the GtkPaned is resized, that child will expand or shrink along with the paned widget. If shrink is true, then when that child can be made smaller than its requisition by the user. Setting shrink to FALSE allows the application to set a minimum size. If resize is false for both children, then this is treated as if resize is true for both children.

The application can set the position of the slider as if it were set by the user, by calling gtkPanedSetPosition.

Creating a paned widget with minimum sizes.

hpaned <- gtkHPaned()
frame1 <- gtkFrame()
frame2 <- gtkFrame()
frame1$setShadowType("in")
frame2$setShadowType("in")

hpaned$setSizeRequest(200 + hpaned$styleGet("handle-size"), -1)

hpaned$pack1(frame1, TRUE, FALSE)
frame1$setSizeRequest(50, -1)

hpaned$pack2(frame2, FALSE, FALSE)
frame2$setSizeRequest(50, -1)

Structures

GtkPaned

undocumented

Signals

accept-position(widget, user.data)

The ::accept-position signal is a keybinding signal which gets emitted to accept the current position of the handle when moving it using key bindings.

The default binding for this signal is Return or Space. Since 2.0

widget

the object that received the signal

user.data

user data set when the signal handler was connected.

cancel-position(widget, user.data)

The ::cancel-position signal is a keybinding signal which gets emitted to cancel moving the position of the handle using key bindings. The position of the handle will be reset to the value prior to moving it.

The default binding for this signal is Escape. Since 2.0

widget

the object that received the signal

user.data

user data set when the signal handler was connected.

cycle-child-focus(widget, reversed, user.data)

The ::cycle-child-focus signal is a keybinding signal which gets emitted to cycle the focus between the children of the paned.

The default binding is f6. Since 2.0

widget

the object that received the signal

reversed

whether cycling backward or forward

user.data

user data set when the signal handler was connected.

cycle-handle-focus(widget, reversed, user.data)

The ::cycle-handle-focus signal is a keybinding signal which gets emitted to cycle whether the paned should grab focus to allow the user to change position of the handle by using key bindings.

The default binding for this signal is f8. Since 2.0

widget

the object that received the signal

reversed

whether cycling backward or forward

user.data

user data set when the signal handler was connected.

move-handle(widget, scroll.type, user.data)

The ::move-handle signal is a keybinding signal which gets emitted to move the handle when the user is using key bindings to move it. Since 2.0

widget

the object that received the signal

scroll.type

a GtkScrollType

user.data

user data set when the signal handler was connected.

toggle-handle-focus(widget, user.data)

The ::toggle-handle-focus is a keybinding signal which gets emitted to accept the current position of the handle and then move focus to the next widget in the focus chain.

The default binding is Tab. Since 2.0

widget

the object that received the signal

user.data

user data set when the signal handler was connected.

Properties

max-position [integer : Read]

The largest possible value for the position property. This property is derived from the size and shrinkability of the widget's children. Allowed values: >= 0 Default value: 2147483647 Since 2.4

min-position [integer : Read]

The smallest possible value for the position property. This property is derived from the size and shrinkability of the widget's children. Allowed values: >= 0 Default value: 0 Since 2.4

position [integer : Read / Write]

Position of paned separator in pixels (0 means all the way to the left/top). Allowed values: >= 0 Default value: 0

position-set [logical : Read / Write]

TRUE if the Position property should be used. Default value: FALSE

Style Properties

handle-size [integer : Read]

Width of handle. Allowed values: >= 0 Default value: 5

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkPaned.html

[Package RGtk2 version 2.20.31 Index]