GtkPaperSize {RGtk2}	R Documentation

GtkPaperSize

Description

Support for named paper sizes

Methods and Functions

gtkPaperSizeNew(name = NULL)
gtkPaperSizeNewFromPpd(ppd.name, ppd.display.name, width, height)
gtkPaperSizeNewCustom(name, display.name, width, height, unit)
gtkPaperSizeCopy(object)
gtkPaperSizeIsEqual(object, size2)
gtkPaperSizeGetPaperSizes(include.custom)
gtkPaperSizeGetName(object)
gtkPaperSizeGetDisplayName(object)
gtkPaperSizeGetPpdName(object)
gtkPaperSizeGetWidth(object, unit)
gtkPaperSizeGetHeight(object, unit)
gtkPaperSizeIsCustom(object)
gtkPaperSizeSetSize(object, width, height, unit)
gtkPaperSizeGetDefaultTopMargin(object, unit)
gtkPaperSizeGetDefaultBottomMargin(object, unit)
gtkPaperSizeGetDefaultLeftMargin(object, unit)
gtkPaperSizeGetDefaultRightMargin(object, unit)
gtkPaperSizeGetDefault()
gtkPaperSizeNewFromKeyFile(key.file, group.name, .errwarn = TRUE)
gtkPaperSizeToKeyFile(object, key.file, group.name)

Hierarchy

GBoxed
   +----GtkPaperSize

Detailed Description

GtkPaperSize handles paper sizes. It uses the standard called "PWG 5101.1-2002 PWG: Standard for Media Standardized Names"

to name the paper sizes (and to get the data for the page sizes). In addition to standard paper sizes, GtkPaperSize allows to construct custom paper sizes with arbitrary dimensions.

The GtkPaperSize object stores not only the dimensions (width and height) of a paper size and its name, it also provides default print margins.

Printing support has been added in GTK+ 2.10.

Structures

GtkPaperSize

undocumented

Enums and Flags

GtkUnit

undocumented

pixel

undocumented

points

undocumented

inch

undocumented

mm

undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkPaperSize.html

See Also

GtkPageSetup

[Package RGtk2 version 2.20.31 Index]