GtkPixmap {RGtk2}	R Documentation

GtkPixmap

Description

A widget displaying a graphical image or icon

Methods and Functions

gtkPixmapNew(pixmap, mask = NULL, show = TRUE)
gtkPixmapSet(object, val, mask = NULL)
gtkPixmapGet(object)
gtkPixmapSetBuildInsensitive(object, build)
gtkPixmap(pixmap, mask = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkMisc
                           +----GtkPixmap

Interfaces

GtkPixmap implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkPixmap widget displays a graphical image or icon. The icon is typically created using gdkPixmapColormapCreateFromXpm or gdkPixmapColormapCreateFromXpmD.

The pixels in a GtkPixmap cannot be manipulated by the application after creation, since under the X Window system the pixel data is stored on the X server and so is not available to the client application. If you want to create graphical images which can be manipulated by the application, look at GtkImage and GdkRGB.

GtkPixmap has been deprecated since GTK+ 2.0 and should not be used in newly written code. Use GtkImage instead.

Structures

GtkPixmap

WARNING: GtkPixmap is deprecated and should not be used in newly-written code. The GtkPixmap struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkPixmap is the equivalent of gtkPixmapNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkPixmap.html

[Package RGtk2 version 2.20.31 Index]