GtkPlug {RGtk2}	R Documentation

GtkPlug

Description

Toplevel for embedding into other processes

Methods and Functions

gtkPlugConstruct(object, socket.id)
gtkPlugConstructForDisplay(object, display, socket.id)
gtkPlugNew(socket.id, show = TRUE)
gtkPlugNewForDisplay(display, socket.id)
gtkPlugGetId(object)
gtkPlugGetEmbedded(object)
gtkPlugGetSocketWindow(object)
gtkPlug(socket.id, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkWindow
                                       +----GtkPlug

Interfaces

GtkPlug implements AtkImplementorIface and GtkBuildable.

Detailed Description

Together with GtkSocket, GtkPlug provides the ability to embed widgets from one process into another process in a fashion that is transparent to the user. One process creates a GtkSocket widget and passes the ID of that widget's window to the other process, which then creates a GtkPlug with that window ID. Any widgets contained in the GtkPlug then will appear inside the first application's window. PLEASE NOTE: The GtkPlug and GtkSocket widgets are currently not available on all platforms supported by GTK+.

Structures

GtkPlug

undocumented

Convenient Construction

gtkPlug is the equivalent of gtkPlugNew.

Signals

embedded(plug, user.data)

Gets emitted when the plug becomes embedded in a socket.

plug

the object on which the signal was emitted

user.data

user data set when the signal handler was connected.

Properties

embedded [logical : Read]

TRUE if the plug is embedded in a socket. Default value: FALSE Since 2.12

socket-window [GdkWindow : * : Read]

The window of the socket the plug is embedded in. Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkPlug.html

[Package RGtk2 version 2.20.31 Index]