GtkPreview {RGtk2}	R Documentation

GtkPreview

Description

A widget to display RGB or grayscale data

Methods and Functions

gtkPreviewUninit()
gtkPreviewNew(type, show = TRUE)
gtkPreviewSize(object, width, height)
gtkPreviewPut(object, window, gc, srcx, srcy, destx, desty, width, height)
gtkPreviewDrawRow(object, data, y, w)
gtkPreviewSetExpand(object, expand)
gtkPreviewSetGamma(gamma)
gtkPreviewSetColorCube(nred.shades, ngreen.shades, nblue.shades, ngray.shades)
gtkPreviewSetInstallCmap(install.cmap)
gtkPreviewSetReserved(nreserved)
gtkPreviewSetDither(object, dither)
gtkPreviewGetVisual()
gtkPreviewGetCmap()
gtkPreviewGetInfo()
gtkPreviewReset()
gtkPreview(type, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkPreview

Interfaces

GtkPreview implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkPreview widget provides a simple interface used to display images as RGB or grayscale data. It's deprecated; just use a GdkPixbuf displayed by a GtkImage, or perhaps a GtkDrawingArea. GtkPreview has no advantage over those approaches.

Structures

GtkPreview

WARNING: GtkPreview is deprecated and should not be used in newly-written code. The GtkPreview struct contains private data only, and should be accessed using the functions below.

GtkPreviewInfo

WARNING: GtkPreviewInfo is deprecated and should not be used in newly-written code. Contains information about global properties of preview widgets. The GtkPreviewInfo struct contains the following fields. (These fields should be considered read-only. They should never be set by an application.)

GdkVisual *visual; 	 the visual used by all previews.

GdkColormap *cmap; 	 the colormap used by all previews.

gdouble gamma; 	 the gamma correction value used by all previews (See gtkPreviewSetGamma ).

Convenient Construction

gtkPreview is the equivalent of gtkPreviewNew.

Properties

expand [logical : Read / Write]

Whether the preview widget should take up the entire space it is allocated. Default value: FALSE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkPreview.html

[Package RGtk2 version 2.20.31 Index]