GtkPrintContext {RGtk2}	R Documentation

GtkPrintContext

Description

Encapsulates context for drawing pages

Methods and Functions

gtkPrintContextGetCairoContext(object)
gtkPrintContextSetCairoContext(object, cr, dpi.x, dpi.y)
gtkPrintContextGetPageSetup(object)
gtkPrintContextGetWidth(object)
gtkPrintContextGetHeight(object)
gtkPrintContextGetDpiX(object)
gtkPrintContextGetDpiY(object)
gtkPrintContextGetPangoFontmap(object)
gtkPrintContextCreatePangoContext(object)
gtkPrintContextCreatePangoLayout(object)
gtkPrintContextGetHardMargins(object)

Hierarchy

GObject
   +----GtkPrintContext

Detailed Description

A GtkPrintContext encapsulates context information that is required when drawing pages for printing, such as the cairo context and important parameters like page size and resolution. It also lets you easily create PangoLayout and PangoContext objects that match the font metrics of the cairo surface.

GtkPrintContext objects gets passed to the ::begin-print, ::end-print, ::request-page-setup and ::draw-page signals on the GtkPrintOperation.

Using GtkPrintContext in a ::draw-page callback

draw_page <- function(operation, context, page_nr)
{
  cr <- context$getCairoContext()

  # Draw a red rectangle, as wide as the paper (inside the margins)
  cr$setSourceRgb(1.0, 0, 0)
  cr$rectangle(0, 0, context$getWidth(), 50)
  
  cr$fill()

  # Draw some lines
  cr$moveTo(20, 10)
  cr$lineTo(40, 20)
  cr$arc(60, 60, 20, 0, pi)
  cr$lineTo(80, 20)
  
  cr$setSourceRgb(0, 0, 0)
  cr$setLineWidth(5)
  cr$setLineCap("round")
  cr$setLineJoin("round")
  
  cr$stroke()

  # Draw some text
  
  layout <- context$createLayout()
  layout$setText("Hello World! Printing is easy")
  desc <- pangoFontDescriptionFromString("sans 28")
  layout$setFontDescription(desc)
  
  cr$moveTo(30, 20)
  cr$layoutPath(layout)

  # Font Outline
  cr$setSourceRgb(0.93, 1.0, 0.47)
  cr$setLineWidth(0.5)
  cr$strokePreserve()

  # Font Fill
  cr$setSourceRgb(0, 0.0, 1.0)
  cr$fill()
}

Printing support was added in GTK+ 2.10.

Structures

GtkPrintContext

undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkPrintContext.html

[Package RGtk2 version 2.20.31 Index]