GtkPrintSettings {RGtk2}	R Documentation

GtkPrintSettings

Description

Stores print settings

Methods and Functions

gtkPrintSettingsNew()
gtkPrintSettingsCopy(object)
gtkPrintSettingsHasKey(object, key)
gtkPrintSettingsGet(object, key)
gtkPrintSettingsSet(object, key, value)
gtkPrintSettingsUnset(object, key)
gtkPrintSettingsForeach(object, func, user.data = NULL)
gtkPrintSettingsGetBool(object, key)
gtkPrintSettingsSetBool(object, key, value)
gtkPrintSettingsGetDouble(object, key)
gtkPrintSettingsGetDoubleWithDefault(object, key, def)
gtkPrintSettingsSetDouble(object, key, value)
gtkPrintSettingsGetLength(object, key, unit)
gtkPrintSettingsSetLength(object, key, value, unit)
gtkPrintSettingsGetInt(object, key)
gtkPrintSettingsGetIntWithDefault(object, key, def)
gtkPrintSettingsSetInt(object, key, value)
gtkPrintSettingsGetPrinter(object)
gtkPrintSettingsSetPrinter(object, printer)
gtkPrintSettingsGetOrientation(object)
gtkPrintSettingsSetOrientation(object, orientation)
gtkPrintSettingsGetPaperSize(object)
gtkPrintSettingsSetPaperSize(object, paper.size)
gtkPrintSettingsGetPaperWidth(object, unit)
gtkPrintSettingsSetPaperWidth(object, width, unit)
gtkPrintSettingsGetPaperHeight(object, unit)
gtkPrintSettingsSetPaperHeight(object, height, unit)
gtkPrintSettingsGetUseColor(object)
gtkPrintSettingsSetUseColor(object, use.color)
gtkPrintSettingsGetCollate(object)
gtkPrintSettingsSetCollate(object, collate)
gtkPrintSettingsGetReverse(object)
gtkPrintSettingsSetReverse(object, reverse)
gtkPrintSettingsGetDuplex(object)
gtkPrintSettingsSetDuplex(object, duplex)
gtkPrintSettingsGetQuality(object)
gtkPrintSettingsSetQuality(object, quality)
gtkPrintSettingsGetNCopies(object)
gtkPrintSettingsSetNCopies(object, num.copies)
gtkPrintSettingsGetNumberUp(object)
gtkPrintSettingsSetNumberUp(object, number.up)
gtkPrintSettingsGetNumberUpLayout(object)
gtkPrintSettingsSetNumberUpLayout(object, number.up.layout)
gtkPrintSettingsGetResolution(object)
gtkPrintSettingsSetResolution(object, resolution)
gtkPrintSettingsSetResolutionXy(object, resolution.x, resolution.y)
gtkPrintSettingsGetResolutionX(object)
gtkPrintSettingsGetResolutionY(object)
gtkPrintSettingsGetPrinterLpi(object)
gtkPrintSettingsSetPrinterLpi(object, lpi)
gtkPrintSettingsGetScale(object)
gtkPrintSettingsSetScale(object, scale)
gtkPrintSettingsGetPrintPages(object)
gtkPrintSettingsSetPrintPages(object, pages)
gtkPrintSettingsGetPageRanges(object, num.ranges)
gtkPrintSettingsSetPageRanges(object, page.ranges, num.ranges)
gtkPrintSettingsGetPageSet(object)
gtkPrintSettingsSetPageSet(object, page.set)
gtkPrintSettingsGetDefaultSource(object)
gtkPrintSettingsSetDefaultSource(object, default.source)
gtkPrintSettingsGetMediaType(object)
gtkPrintSettingsSetMediaType(object, media.type)
gtkPrintSettingsGetDither(object)
gtkPrintSettingsSetDither(object, dither)
gtkPrintSettingsGetFinishings(object)
gtkPrintSettingsSetFinishings(object, finishings)
gtkPrintSettingsGetOutputBin(object)
gtkPrintSettingsSetOutputBin(object, output.bin)
gtkPrintSettingsNewFromFile(file.name, .errwarn = TRUE)
gtkPrintSettingsNewFromKeyFile(key.file, group.name, .errwarn = TRUE)
gtkPrintSettingsLoadFile(object, file.name, .errwarn = TRUE)
gtkPrintSettingsLoadKeyFile(object, key.file, group.name, .errwarn = TRUE)
gtkPrintSettingsToFile(object, file.name, .errwarn = TRUE)
gtkPrintSettingsToKeyFile(object, key.file, group.name)
gtkPrintSettings()

Hierarchy

GObject
   +----GtkPrintSettings

Detailed Description

A GtkPrintSettings object represents the settings of a print dialog in a system-independent way. The main use for this object is that once you've printed you can get a settings object that represents the settings the user chose, and the next time you print you can pass that object in so that the user doesn't have to re-set all his settings.

Its also possible to enumerate the settings so that you can easily save the settings for the next time your app runs, or even store them in a document. The predefined keys try to use shared values as much as possible so that moving such a document between systems still works.

Printing support was added in GTK+ 2.10.

Structures

GtkPrintSettings

undocumented

GtkPageRange

undocumented GtkPageRange is a transparent-type.

start

[integer]

end

[integer]

Convenient Construction

gtkPrintSettings is the equivalent of gtkPrintSettingsNew.

Enums and Flags

GtkPageOrientation

undocumented

portrait

undocumented

landscape

undocumented

reverse-portrait

undocumented

reverse-landscape

undocumented

GtkPrintDuplex

undocumented

simplex

undocumented

horizontal

undocumented

vertical

undocumented

GtkPrintQuality

undocumented

low

undocumented

normal

undocumented

high

undocumented

draft

undocumented

GtkNumberUpLayout

Used to determine the layout of pages on a sheet when printing multiple pages per sheet.

left-to-right-top-to-bottom

undocumented

left-to-right-bottom-to-top

undocumented

right-to-left-top-to-bottom

undocumented

right-to-left-bottom-to-top

undocumented

top-to-bottom-left-to-right

undocumented

top-to-bottom-right-to-left

undocumented

bottom-to-top-left-to-right

undocumented

bottom-to-top-right-to-left

undocumented

GtkPrintPages

undocumented

all

undocumented

current

undocumented

ranges

undocumented

GtkPageSet

undocumented

all

undocumented

even

undocumented

odd

undocumented

User Functions

GtkPrintSettingsFunc()

undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkPrintSettings.html

[Package RGtk2 version 2.20.31 Index]