GtkProgress {RGtk2}	R Documentation

GtkProgress

Description

Base class for GtkProgressBar

Methods and Functions

gtkProgressSetShowText(object, show.text)
gtkProgressSetTextAlignment(object, x.align, y.align)
gtkProgressSetFormatString(object, format)
gtkProgressSetAdjustment(object, adjustment)
gtkProgressSetPercentage(object, percentage)
gtkProgressSetValue(object, value)
gtkProgressGetValue(object)
gtkProgressSetActivityMode(object, activity.mode)
gtkProgressGetCurrentText(object)
gtkProgressGetTextFromValue(object, value)
gtkProgressGetCurrentPercentage(object)
gtkProgressGetPercentageFromValue(object, value)
gtkProgressConfigure(object, value, min, max)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkProgress
                           +----GtkProgressBar

Interfaces

GtkProgress implements AtkImplementorIface and GtkBuildable.

Detailed Description

A GtkProgress is the abstract base class used to derive a GtkProgressBar which provides a visual representation of the progress of a long running operation.

Structures

GtkProgress

The GtkProgress struct contains private data only. and should be accessed using the functions below.

Properties

activity-mode [logical : Read / Write]

If TRUE, the GtkProgress is in activity mode, meaning that it signals something is happening, but not how much of the activity is finished. This is used when you're doing something but don't know how long it will take. Default value: FALSE

show-text [logical : Read / Write]

Whether the progress is shown as text. Default value: FALSE

text-xalign [numeric : Read / Write]

The horizontal text alignment, from 0 (left) to 1 (right). Reversed for RTL layouts. Allowed values: [0,1] Default value: 0.5

text-yalign [numeric : Read / Write]

The vertical text alignment, from 0 (top) to 1 (bottom). Allowed values: [0,1] Default value: 0.5

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkProgress.html

[Package RGtk2 version 2.20.31 Index]