GtkProgressBar {RGtk2}	R Documentation

GtkProgressBar

Description

A widget which indicates progress visually

Methods and Functions

gtkProgressBarNew(show = TRUE)
gtkProgressBarPulse(object)
gtkProgressBarSetText(object, text)
gtkProgressBarSetFraction(object, fraction)
gtkProgressBarSetPulseStep(object, fraction)
gtkProgressBarSetOrientation(object, orientation)
gtkProgressBarSetEllipsize(object, mode)
gtkProgressBarGetText(object)
gtkProgressBarGetFraction(object)
gtkProgressBarGetPulseStep(object)
gtkProgressBarGetOrientation(object)
gtkProgressBarGetEllipsize(object)
gtkProgressBarNewWithAdjustment(adjustment = NULL, show = TRUE)
gtkProgressBarSetBarStyle(object, style)
gtkProgressBarSetDiscreteBlocks(object, blocks)
gtkProgressBarSetActivityStep(object, step)
gtkProgressBarSetActivityBlocks(object, blocks)
gtkProgressBarUpdate(object, percentage)
gtkProgressBar(adjustment = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkProgress
                           +----GtkProgressBar

Interfaces

GtkProgressBar implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkProgressBar is typically used to display the progress of a long running operation. It provides a visual clue that processing is underway. The GtkProgressBar can be used in two different modes: percentage mode and activity mode.

When an application can determine how much work needs to take place (e.g. read a fixed number of bytes from a file) and can monitor its progress, it can use the GtkProgressBar in percentage mode and the user sees a growing bar indicating the percentage of the work that has been completed. In this mode, the application is required to call gtkProgressBarSetFraction periodically to update the progress bar.

When an application has no accurate way of knowing the amount of work to do, it can use the GtkProgressBar in activity mode, which shows activity by a block moving back and forth within the progress area. In this mode, the application is required to call gtkProgressBarPulse perodically to update the progress bar.

There is quite a bit of flexibility provided to control the appearance of the GtkProgressBar. Functions are provided to control the orientation of the bar, optional text can be displayed along with the bar, and the step size used in activity mode can be set. PLEASE NOTE: The GtkProgressBar/GtkProgress API in GTK 1.2 was bloated, needlessly complex and hard to use properly. Therefore GtkProgress has been deprecated completely and the GtkProgressBar API has been reduced to the following 10 functions: gtkProgressBarNew, gtkProgressBarPulse, gtkProgressBarSetText, gtkProgressBarSetFraction, gtkProgressBarSetPulseStep, gtkProgressBarSetOrientation, gtkProgressBarGetText, gtkProgressBarGetFraction, gtkProgressBarGetPulseStep, gtkProgressBarGetOrientation. These have been grouped at the beginning of this section, followed by a large chunk of deprecated 1.2 compatibility functions.

Structures

GtkProgressBar

The GtkProgressBar struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkProgressBar is the result of collapsing the constructors of GtkProgressBar (gtkProgressBarNew, gtkProgressBarNewWithAdjustment) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Enums and Flags

GtkProgressBarOrientation

An enumeration representing possible orientations and growth directions for the visible progress bar.

left-to-right

A horizontal progress bar growing from left to right.

right-to-left

A horizontal progress bar growing from right to left.

bottom-to-top

A vertical progress bar growing from bottom to top.

top-to-bottom

A vertical progress bar growing from top to bottom.

GtkProgressBarStyle

An enumeration representing the styles for drawing the progress bar.

continuous

The progress bar grows in a smooth, continuous manner.

discrete

The progress bar grows in discrete, visible blocks.

Properties

activity-blocks [numeric : Read / Write]

The number of blocks which can fit in the progress bar area in activity mode (Deprecated). Allowed values: >= 2 Default value: 5

activity-step [numeric : Read / Write]

The increment used for each iteration in activity mode (Deprecated). Default value: 3

adjustment [GtkAdjustment : * : Read / Write]

The GtkAdjustment connected to the progress bar (Deprecated).

bar-style [GtkProgressBarStyle : Read / Write]

Specifies the visual style of the bar in percentage mode (Deprecated). Default value: GTK_PROGRESS_CONTINUOUS

discrete-blocks [numeric : Read / Write]

The number of discrete blocks in a progress bar (when shown in the discrete style). Allowed values: >= 2 Default value: 10

ellipsize [PangoEllipsizeMode : Read / Write]

The preferred place to ellipsize the string, if the progressbar does not have enough room to display the entire string, specified as a PangoEllisizeMode.

Note that setting this property to a value other than PANGO_ELLIPSIZE_NONE has the side-effect that the progressbar requests only enough space to display the ellipsis "...". Another means to set a progressbar's width is gtkWidgetSetSizeRequest. Default value: PANGO_ELLIPSIZE_NONE Since 2.6

fraction [numeric : Read / Write]

The fraction of total work that has been completed. Allowed values: [0,1] Default value: 0

orientation [GtkProgressBarOrientation : Read / Write]

Orientation and growth direction of the progress bar. Default value: GTK_PROGRESS_LEFT_TO_RIGHT

pulse-step [numeric : Read / Write]

The fraction of total progress to move the bouncing block when pulsed. Allowed values: [0,1] Default value: 0.1

text [character : * : Read / Write]

Text to be displayed in the progress bar. Default value: NULL

Style Properties

min-horizontal-bar-height [integer : Read / Write]

Minimum horizontal height of the progress bar. Allowed values: >= 1 Default value: 20 Since 2.14

min-horizontal-bar-width [integer : Read / Write]

The minimum horizontal width of the progress bar. Allowed values: >= 1 Default value: 150 Since 2.14

min-vertical-bar-height [integer : Read / Write]

The minimum vertical height of the progress bar. Allowed values: >= 1 Default value: 80 Since 2.14

min-vertical-bar-width [integer : Read / Write]

The minimum vertical width of the progress bar. Allowed values: >= 1 Default value: 22 Since 2.14

xspacing [integer : Read / Write]

Extra spacing applied to the width of a progress bar. Allowed values: >= 0 Default value: 7

yspacing [integer : Read / Write]

Extra spacing applied to the height of a progress bar. Allowed values: >= 0 Default value: 7

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkProgressBar.html

[Package RGtk2 version 2.20.31 Index]