GtkRadioAction {RGtk2}	R Documentation

GtkRadioAction

Description

An action of which only one in a group can be active

Methods and Functions

gtkRadioActionNew(name = NULL, label = NULL, tooltip = NULL, stock.id = NULL, value = NULL)
gtkRadioActionGetGroup(object)
gtkRadioActionSetGroup(object, group)
gtkRadioActionGetCurrentValue(object)
gtkRadioActionSetCurrentValue(object, current.value)
gtkRadioAction(name = NULL, label = NULL, tooltip = NULL, stock.id = NULL, value = NULL)

Hierarchy

GObject
   +----GtkAction
         +----GtkToggleAction
               +----GtkRadioAction

Interfaces

GtkRadioAction implements GtkBuildable.

Detailed Description

A GtkRadioAction is similar to GtkRadioMenuItem. A number of radio actions can be linked together so that only one may be active at any one time.

Structures

GtkRadioAction

The GtkRadioAction struct contains only private members and should not be accessed directly.

Convenient Construction

gtkRadioAction is the equivalent of gtkRadioActionNew.

Signals

changed(action, current, user.data)

The ::changed signal is emitted on every member of a radio group when the active member is changed. The signal gets emitted after the ::activate signals for the previous and current active members. Since 2.4

action

the action on which the signal is emitted

current

the member of actions group which has just been activated

user.data

user data set when the signal handler was connected.

Properties

current-value [integer : Read / Write]

The value property of the currently active member of the group to which this action belongs. Default value: 0 Since 2.10

group [GtkRadioAction : * : Write]

Sets a new group for a radio action. Since 2.4

value [integer : Read / Write]

The value is an arbitrary integer which can be used as a convenient way to determine which action in the group is currently active in an ::activate or ::changed signal handler. See gtkRadioActionGetCurrentValue and GtkRadioActionEntry for convenient ways to get and set this property. Default value: 0 Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkRadioAction.html

[Package RGtk2 version 2.20.31 Index]