GtkRadioMenuItem {RGtk2}	R Documentation

GtkRadioMenuItem

Description

A choice from multiple check menu items

Methods and Functions

gtkRadioMenuItemNew(group = NULL, show = TRUE)
gtkRadioMenuItemNewWithLabel(group = NULL, label, show = TRUE)
gtkRadioMenuItemNewWithMnemonic(group = NULL, label, show = TRUE)
gtkRadioMenuItemNewFromWidget(group = NULL, show = TRUE)
gtkRadioMenuItemNewWithLabelFromWidget(group = NULL, label, show = TRUE)
gtkRadioMenuItemNewWithMnemonicFromWidget(group = NULL, label, show = TRUE)
gtkRadioMenuItemSetGroup(object, group)
gtkRadioMenuItemGetGroup(object)
gtkRadioMenuItem(group = NULL, label, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkItem
                                       +----GtkMenuItem
                                             +----GtkCheckMenuItem
                                                   +----GtkRadioMenuItem

Interfaces

GtkRadioMenuItem implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

A radio menu item is a check menu item that belongs to a group. At each instant exactly one of the radio menu items from a group is selected.

The group list does not need to be freed, as each GtkRadioMenuItem will remove itself and its list item when it is destroyed.

The correct way to create a group of radio menu items is approximatively this:

How to create a group of radio menu items.

group <- NULL
for (i in 1:5) {
  item <- gtkRadioMenuItem(group, "This is an example")
  group <- item$getGroup()
  if (i == 1)
    item$setActive(TRUE)
}

Structures

GtkRadioMenuItem

The structure contains only private data that must be accessed through the interface functions.

Convenient Construction

gtkRadioMenuItem is the result of collapsing the constructors of GtkRadioMenuItem (gtkRadioMenuItemNew, gtkRadioMenuItemNewWithLabel, gtkRadioMenuItemNewWithMnemonic, gtkRadioMenuItemNewFromWidget, gtkRadioMenuItemNewWithMnemonicFromWidget, gtkRadioMenuItemNewWithLabelFromWidget) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Signals

group-changed(radiomenuitem, user.data)

undocumented

radiomenuitem

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

group [GtkRadioMenuItem : * : Write]

The radio menu item whose group this widget belongs to. Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkRadioMenuItem.html

[Package RGtk2 version 2.20.31 Index]