GtkRadioToolButton {RGtk2}	R Documentation

GtkRadioToolButton

Description

A toolbar item that contains a radio button

Methods and Functions

gtkRadioToolButtonNew(group = NULL, show = TRUE)
gtkRadioToolButtonNewFromStock(group = NULL, stock.id, show = TRUE)
gtkRadioToolButtonNewFromWidget(group = NULL, show = TRUE)
gtkRadioToolButtonNewWithStockFromWidget(group = NULL, stock.id, show = TRUE)
gtkRadioToolButtonGetGroup(object)
gtkRadioToolButtonSetGroup(object, group)
gtkRadioToolButton(group = NULL, stock.id, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkToolItem
                                       +----GtkToolButton
                                             +----GtkToggleToolButton
                                                   +----GtkRadioToolButton

Interfaces

GtkRadioToolButton implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

A GtkRadioToolButton is a GtkToolItem that contains a radio button, that is, a button that is part of a group of toggle buttons where only one button can be active at a time.

Use gtkRadioToolButtonNew to create a new GtkRadioToolButton. use gtkRadioToolButtonNewFromWidget to create a new GtkRadioToolButton that is part of the same group as an existing GtkRadioToolButton. Use gtkRadioToolButtonNewFromStock or gtkRadioToolButtonNewFromWidgetWithStock() to create a new GtkRAdioToolButton containing a stock item.

Structures

GtkRadioToolButton

The GtkRadioToolButton contains only private data and should only be accessed through the functions described below.

Convenient Construction

gtkRadioToolButton is the result of collapsing the constructors of GtkRadioToolButton (gtkRadioToolButtonNew, gtkRadioToolButtonNewFromStock, gtkRadioToolButtonNewFromWidget, gtkRadioToolButtonNewWithStockFromWidget) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Properties

group [GtkRadioToolButton : * : Write]

Sets a new group for a radio tool button. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkRadioToolButton.html

[Package RGtk2 version 2.20.31 Index]