GtkRange {RGtk2}	R Documentation

GtkRange

Description

Base class for widgets which visualize an adjustment

Methods and Functions

gtkRangeGetFillLevel(object)
gtkRangeGetRestrictToFillLevel(object)
gtkRangeGetShowFillLevel(object)
gtkRangeSetFillLevel(object, fill.level)
gtkRangeSetRestrictToFillLevel(object, restrict.to.fill.level)
gtkRangeSetShowFillLevel(object, show.fill.level)
gtkRangeGetAdjustment(object)
gtkRangeSetUpdatePolicy(object, policy)
gtkRangeSetAdjustment(object, adjustment)
gtkRangeGetInverted(object)
gtkRangeSetInverted(object, setting)
gtkRangeGetUpdatePolicy(object)
gtkRangeGetValue(object)
gtkRangeSetIncrements(object, step, page)
gtkRangeSetRange(object, min, max)
gtkRangeSetValue(object, value)
gtkRangeSetLowerStepperSensitivity(object, sensitivity)
gtkRangeGetLowerStepperSensitivity(object)
gtkRangeSetUpperStepperSensitivity(object, sensitivity)
gtkRangeGetUpperStepperSensitivity(object)
gtkRangeGetFlippable(object)
gtkRangeSetFlippable(object, flippable)
gtkRangeGetMinSliderSize(object)
gtkRangeGetRangeRect(object)
gtkRangeGetSliderRange(object)
gtkRangeGetSliderSizeFixed(object)
gtkRangeSetMinSliderSize(object, min.size)
gtkRangeSetSliderSizeFixed(object, size.fixed)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkRange
                           +----GtkScale
                           +----GtkScrollbar

Interfaces

GtkRange implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

GtkRange is the common base class for widgets which visualize an adjustment, e.g scales or scrollbars.

Apart from signals for monitoring the parameters of the adjustment, GtkRange provides properties and methods for influencing the sensitivity of the "steppers". It also provides properties and methods for setting a "fill level" on range widgets. See gtkRangeSetFillLevel.

Structures

GtkRange

undocumented

Enums and Flags

GtkSensitivityType

Determines how GTK+ handles the sensitivity of stepper arrows at the end of range widgets.

auto

The arrow is made insensitive if the thumb is at the end

on

The arrow is always sensitive

off

The arrow is always insensitive

Signals

adjust-bounds(range, user.data)

undocumented

range

the object which received the signal.

user.data

user data set when the signal handler was connected.

change-value(range, scroll, value, returns, user.data)

The ::change-value signal is emitted when a scroll action is performed on a range. It allows an application to determine the type of scroll event that occurred and the resultant new value. The application can handle the event itself and return TRUE to prevent further processing. Or, by returning FALSE, it can pass the event to other handlers until the default GTK+ handler is reached.

The value parameter is unrounded. An application that overrides the ::change-value signal is responsible for clamping the value to the desired number of decimal digits; the default GTK+ handler clamps the value based on range->round.digits.

It is not possible to use delayed update policies in an overridden ::change-value handler. Since 2.6

range

the range that received the signal

scroll

the type of scroll action that was performed

value

the new value resulting from the scroll action

returns

TRUE to prevent other handlers from being invoked for the signal, FALSE to propagate the signal further

user.data

user data set when the signal handler was connected.

move-slider(range, step, user.data)

Virtual function that moves the slider. Used for keybindings.

range

the GtkRange

step

how to move the slider

user.data

user data set when the signal handler was connected.

value-changed(range, user.data)

Emitted when the range value changes.

range

the GtkRange

user.data

user data set when the signal handler was connected.

Properties

adjustment [GtkAdjustment : * : Read / Write / Construct]

The GtkAdjustment that contains the current value of this range object.

fill-level [numeric : Read / Write]

The fill level (e.g. prebuffering of a network stream). See gtkRangeSetFillLevel. Default value: 1.79769e+308 Since 2.12

inverted [logical : Read / Write]

Invert direction slider moves to increase range value. Default value: FALSE

lower-stepper-sensitivity [GtkSensitivityType : Read / Write]

The sensitivity policy for the stepper that points to the adjustment's lower side. Default value: GTK_SENSITIVITY_AUTO

restrict-to-fill-level [logical : Read / Write]

The restrict-to-fill-level property controls whether slider movement is restricted to an upper boundary set by the fill level. See gtkRangeSetRestrictToFillLevel. Default value: TRUE Since 2.12

show-fill-level [logical : Read / Write]

The show-fill-level property controls whether fill level indicator graphics are displayed on the trough. See gtkRangeSetShowFillLevel. Default value: FALSE Since 2.12

update-policy [GtkUpdateType : Read / Write]

How the range should be updated on the screen. Default value: GTK_UPDATE_CONTINUOUS

upper-stepper-sensitivity [GtkSensitivityType : Read / Write]

The sensitivity policy for the stepper that points to the adjustment's upper side. Default value: GTK_SENSITIVITY_AUTO

Style Properties

activate-slider [logical : Read]

With this option set to TRUE, sliders will be drawn ACTIVE and with shadow IN while they are dragged. Default value: FALSE

arrow-displacement-x [integer : Read]

How far in the x direction to move the arrow when the button is depressed. Default value: 0

arrow-displacement-y [integer : Read]

How far in the y direction to move the arrow when the button is depressed. Default value: 0

arrow-scaling [numeric : Read]

The arrow size proportion relative to the scroll button size. Allowed values: [0,1] Default value: 0.5 Since 2.14

slider-width [integer : Read]

Width of scrollbar or scale thumb. Allowed values: >= 0 Default value: 14

stepper-size [integer : Read]

Length of step buttons at ends. Allowed values: >= 0 Default value: 14

stepper-spacing [integer : Read]

The spacing between the stepper buttons and thumb. Note that setting this value to anything > 0 will automatically set the trough-under-steppers style property to TRUE as well. Also, stepper-spacing won't have any effect if there are no steppers. Allowed values: >= 0 Default value: 0

trough-border [integer : Read]

Spacing between thumb/steppers and outer trough bevel. Allowed values: >= 0 Default value: 1

trough-side-details [logical : Read]

When TRUE, the parts of the trough on the two sides of the slider are drawn with different details. Default value: FALSE Since 2.10

trough-under-steppers [logical : Read]

Whether to draw the trough across the full length of the range or to exclude the steppers and their spacing. Note that setting the "stepper-spacing" style property to any value > 0 will automatically enable trough-under-steppers too. Default value: TRUE Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkRange.html

[Package RGtk2 version 2.20.31 Index]