GtkRecentAction {RGtk2}	R Documentation

GtkRecentAction

Description

An action of which represents a list of recently used files

Methods and Functions

gtkRecentActionNew(name, label, tooltip, stock.id)
gtkRecentActionNewForManager(name, label, tooltip, stock.id, manager)
gtkRecentActionGetShowNumbers(object)
gtkRecentActionSetShowNumbers(object, show.numbers)
gtkRecentAction(name, label, tooltip, stock.id)

Hierarchy

GObject
   +----GtkAction
         +----GtkRecentAction

Interfaces

GtkRecentAction implements GtkBuildable and GtkRecentChooser.

Detailed Description

A GtkRecentAction represents a list of recently used files, which can be shown by widgets such as GtkRecentChooserDialog or GtkRecentChooserMenu.

To construct a submenu showing recently used files, use a GtkRecentAction as the action for a <menuitem>. To construct a menu toolbutton showing the recently used files in the popup menu, use a GtkRecentAction as the action for a <toolitem> element.

Structures

GtkRecentAction

undocumented

Convenient Construction

gtkRecentAction is the equivalent of gtkRecentActionNew.

Properties

show-numbers [logical : Read / Write]

Whether the items should be displayed with a number. Default value: FALSE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkRecentAction.html

[Package RGtk2 version 2.20.31 Index]