GtkRecentChooser {RGtk2}	R Documentation

GtkRecentChooser

Description

Interface implemented by widgets displaying recently used files

Methods and Functions

gtkRecentChooserSetShowPrivate(object, show.private)
gtkRecentChooserGetShowPrivate(object)
gtkRecentChooserSetShowNotFound(object, show.not.found)
gtkRecentChooserGetShowNotFound(object)
gtkRecentChooserSetShowIcons(object, show.icons)
gtkRecentChooserGetShowIcons(object)
gtkRecentChooserSetSelectMultiple(object, select.multiple)
gtkRecentChooserGetSelectMultiple(object)
gtkRecentChooserSetLocalOnly(object, local.only)
gtkRecentChooserGetLocalOnly(object)
gtkRecentChooserSetLimit(object, limit)
gtkRecentChooserGetLimit(object)
gtkRecentChooserSetShowTips(object, show.tips)
gtkRecentChooserGetShowTips(object)
gtkRecentChooserSetSortType(object, sort.type)
gtkRecentChooserGetSortType(object)
gtkRecentChooserSetSortFunc(object, sort.func, sort.data)
gtkRecentChooserSetCurrentUri(object, uri, .errwarn = TRUE)
gtkRecentChooserGetCurrentUri(object)
gtkRecentChooserGetCurrentItem(object)
gtkRecentChooserSelectUri(object, uri, .errwarn = TRUE)
gtkRecentChooserUnselectUri(object, uri)
gtkRecentChooserSelectAll(object)
gtkRecentChooserUnselectAll(object)
gtkRecentChooserGetItems(object)
gtkRecentChooserGetUris(object)
gtkRecentChooserAddFilter(object, filter)
gtkRecentChooserRemoveFilter(object, filter)
gtkRecentChooserListFilters(object)
gtkRecentChooserSetFilter(object, filter)
gtkRecentChooserGetFilter(object)

Hierarchy

GInterface
   +----GtkRecentChooser

Implementations

GtkRecentChooser is implemented by GtkRecentAction, GtkRecentChooserDialog, GtkRecentChooserMenu and GtkRecentChooserWidget.

Detailed Description

GtkRecentChooser is an interface that can be implemented by widgets displaying the list of recently used files. In GTK+, the main objects that implement this interface are GtkRecentChooserWidget, GtkRecentChooserDialog and GtkRecentChooserMenu.

Recently used files are supported since GTK+ 2.10.

Structures

GtkRecentChooser

undocumented

Enums and Flags

GtkRecentChooserError

These identify the various errors that can occur while calling GtkRecentChooser functions.

not-found

Indicates that a file does not exist

invalid-uri

Indicates a malformed URI

GtkRecentSortType

Used to specify the sorting method to be applyed to the recently used resource list.

none

Do not sort the returned list of recently used resources.

mru

Sort the returned list with the most recently used items first.

lru

Sort the returned list with the least recently used items first.

custom

Sort the returned list using a custom sorting function passed using gtkRecentManagerSetSortFunc().

User Functions

GtkRecentSortFunc()

undocumented

Signals

item-activated(chooser, user.data)

This signal is emitted when the user "activates" a recent item in the recent chooser. This can happen by double-clicking on an item in the recently used resources list, or by pressing Enter. Since 2.10

chooser

the object which received the signal

user.data

user data set when the signal handler was connected.

selection-changed(chooser, user.data)

This signal is emitted when there is a change in the set of selected recently used resources. This can happen when a user modifies the selection with the mouse or the keyboard, or when explicitely calling functions to change the selection. Since 2.10

chooser

the object which received the signal

user.data

user data set when the signal handler was connected.

Properties

filter [GtkRecentFilter : * : Read / Write]

The GtkRecentFilter object to be used when displaying the recently used resources. Since 2.10

limit [integer : Read / Write]

The maximum number of recently used resources to be displayed, or -1 to display all items. By default, the GtkSetting:gtk-recent-files-limit setting is respected: you can override that limit on a particular instance of GtkRecentChooser by setting this property. Allowed values: >= -1 Default value: -1 Since 2.10

local-only [logical : Read / Write]

Whether this GtkRecentChooser should display only local (file:) resources. Default value: TRUE Since 2.10

recent-manager [GtkRecentManager : * : Write / Construct Only]

The GtkRecentManager instance used by the GtkRecentChooser to display the list of recently used resources. Since 2.10

select-multiple [logical : Read / Write]

Allow the user to select multiple resources. Default value: FALSE Since 2.10

show-icons [logical : Read / Write]

Whether this GtkRecentChooser should display an icon near the item. Default value: TRUE Since 2.10

show-not-found [logical : Read / Write]

Whether this GtkRecentChooser should display the recently used resources even if not present anymore. Setting this to FALSE will perform a potentially expensive check on every local resource (every remote resource will always be displayed). Default value: TRUE Since 2.10

show-private [logical : Read / Write]

Whether the private items should be displayed. Default value: FALSE

show-tips [logical : Read / Write]

Whether this GtkRecentChooser should display a tooltip containing the full path of the recently used resources. Default value: FALSE Since 2.10

sort-type [GtkRecentSortType : Read / Write]

Sorting order to be used when displaying the recently used resources. Default value: GTK_RECENT_SORT_NONE Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkRecentChooser.html

See Also

GtkRecentManager GtkRecentChooserDialog GtkRecentChooserWidget GtkRecentChooserMenu

[Package RGtk2 version 2.20.31 Index]