GtkRecentChooserDialog {RGtk2}	R Documentation

GtkRecentChooserDialog

Description

Displays recently used files in a dialog

Methods and Functions

gtkRecentChooserDialogNew(title = NULL, parent = NULL, ..., show = TRUE)
gtkRecentChooserDialogNewForManager(title = NULL, parent = NULL, manager, ..., show = TRUE)
gtkRecentChooserDialog(title = NULL, parent = NULL, ..., show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkWindow
                                       +----GtkDialog
                                             +----GtkRecentChooserDialog

Interfaces

GtkRecentChooserDialog implements AtkImplementorIface, GtkBuildable and GtkRecentChooser.

Detailed Description

GtkRecentChooserDialog is a dialog box suitable for displaying the recently used documents. This widgets works by putting a GtkRecentChooserWidget inside a GtkDialog. It exposes the GtkRecentChooserIface interface, so you can use all the GtkRecentChooser functions on the recent chooser dialog as well as those for GtkDialog.

Note that GtkRecentChooserDialog does not have any methods of its own. Instead, you should use the functions that work on a GtkRecentChooser.

Typical usage

dialog <- gtkRecentChooserDialog("Recent Documents", parent_window,
                                 "gtk-cancel", GtkResponseType["cancel"],
                                 "gtk-open", GtkResponseType["accept"])
					
if (dialog$run() == GtkResponseType["accept"])
{
  info <- dialog$getCurrentItem()
  open_file(info$getUri())
}

Recently used files are supported since GTK+ 2.10.

Structures

GtkRecentChooserDialog

undocumented

Convenient Construction

gtkRecentChooserDialog is the equivalent of gtkRecentChooserDialogNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkRecentChooserDialog.html

See Also

GtkRecentChooser GtkDialog

[Package RGtk2 version 2.20.31 Index]