GtkRecentChooserMenu {RGtk2}	R Documentation

GtkRecentChooserMenu

Description

Displays recently used files in a menu

Methods and Functions

gtkRecentChooserMenuNew()
gtkRecentChooserMenuNewForManager(manager = NULL, show = TRUE)
gtkRecentChooserMenuGetShowNumbers(object)
gtkRecentChooserMenuSetShowNumbers(object, show.numbers)
gtkRecentChooserMenu(manager = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkMenuShell
                                 +----GtkMenu
                                       +----GtkRecentChooserMenu

Interfaces

GtkRecentChooserMenu implements AtkImplementorIface, GtkBuildable, GtkRecentChooser and GtkActivatable.

Detailed Description

GtkRecentChooserMenu is a widget suitable for displaying recently used files inside a menu. It can be used to set a sub-menu of a GtkMenuItem using gtkMenuItemSetSubmenu, or as the menu of a GtkMenuToolButton.

Note that GtkRecentChooserMenu does not have any methods of its own. Instead, you should use the functions that work on a GtkRecentChooser.

Note also that GtkRecentChooserMenu does not support multiple filters, as it has no way to let the user choose between them as the GtkRecentChooserWidget and GtkRecentChooserDialog widgets do. Thus using gtkRecentChooserAddFilter on a GtkRecentChooserMenu widget will yield the same effects as using gtkRecentChooserSetFilter, replacing any currently set filter with the supplied filter; gtkRecentChooserRemoveFilter will remove any currently set GtkRecentFilter object and will unset the current filter; gtkRecentChooserListFilters will return a list containing a single GtkRecentFilter object.

Recently used files are supported since GTK+ 2.10.

Structures

GtkRecentChooserMenu

undocumented

Convenient Construction

gtkRecentChooserMenu is the equivalent of gtkRecentChooserMenuNewForManager.

Properties

show-numbers [logical : Read / Write]

Whether the first ten items in the menu should be prepended by a number acting as a unique mnemonic. Default value: FALSE Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkRecentChooserMenu.html

See Also

GtkRecentChooser

[Package RGtk2 version 2.20.31 Index]