GtkRecentChooserWidget {RGtk2}	R Documentation

GtkRecentChooserWidget

Description

Displays recently used files

Methods and Functions

gtkRecentChooserWidgetNew()
gtkRecentChooserWidgetNewForManager(manager = NULL, show = TRUE)
gtkRecentChooserWidget(manager = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBox
                                 +----GtkVBox
                                       +----GtkRecentChooserWidget

Interfaces

GtkRecentChooserWidget implements AtkImplementorIface, GtkBuildable, GtkOrientable and GtkRecentChooser.

Detailed Description

GtkRecentChooserWidget is a widget suitable for selecting recently used files. It is the main building block of a GtkRecentChooserDialog. Most applications will only need to use the latter; you can use GtkRecentChooserWidget as part of a larger window if you have special needs.

Note that GtkRecentChooserWidget does not have any methods of its own. Instead, you should use the functions that work on a GtkRecentChooser.

Recently used files are supported since GTK+ 2.10.

Structures

GtkRecentChooserWidget

undocumented

Convenient Construction

gtkRecentChooserWidget is the equivalent of gtkRecentChooserWidgetNewForManager.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkRecentChooserWidget.html

See Also

GtkRecentChooser GtkRecentChooserDialog

[Package RGtk2 version 2.20.31 Index]