GtkRecentManager {RGtk2}	R Documentation

GtkRecentManager

Description

Managing Recently Used Files

Methods and Functions

gtkRecentManagerNew()
gtkRecentManagerGetDefault()
gtkRecentManagerGetForScreen(screen)
gtkRecentManagerGetForScreen(screen)
gtkRecentManagerSetScreen(object, screen)
gtkRecentManagerSetScreen(object, screen)
gtkRecentManagerAddItem(object, uri)
gtkRecentManagerAddFull(object, uri, recent.data)
gtkRecentManagerRemoveItem(object, uri, .errwarn = TRUE)
gtkRecentManagerLookupItem(object, uri, .errwarn = TRUE)
gtkRecentManagerHasItem(object, uri)
gtkRecentManagerMoveItem(object, uri, new.uri, .errwarn = TRUE)
gtkRecentManagerGetLimit(object)
gtkRecentManagerSetLimit(object, limit)
gtkRecentManagerGetItems(object)
gtkRecentManagerPurgeItems(object, .errwarn = TRUE)
gtkRecentInfoRef(object)
gtkRecentInfoUnref(object)
gtkRecentInfoGetUri(object)
gtkRecentInfoGetDisplayName(object)
gtkRecentInfoGetDescription(object)
gtkRecentInfoGetMimeType(object)
gtkRecentInfoGetAdded(object)
gtkRecentInfoGetModified(object)
gtkRecentInfoGetVisited(object)
gtkRecentInfoGetPrivateHint(object)
gtkRecentInfoGetApplicationInfo(object, app.name)
gtkRecentInfoGetApplications(object, length)
gtkRecentInfoLastApplication(object)
gtkRecentInfoGetGroups(object)
gtkRecentInfoHasGroup(object, group.name)
gtkRecentInfoHasApplication(object, app.name)
gtkRecentInfoGetIcon(object, size)
gtkRecentInfoGetShortName(object)
gtkRecentInfoGetUriDisplay(object)
gtkRecentInfoGetAge(object)
gtkRecentInfoIsLocal(object)
gtkRecentInfoExists(object)
gtkRecentInfoMatch(object, info.b)
gtkRecentManager()

Hierarchy

GObject
   +----GtkRecentManager

Detailed Description

GtkRecentManager provides a facility for adding, removing and looking up recently used files. Each recently used file is identified by its URI, and has meta-data associated to it, like the names and command lines of the applications that have registered it, the number of time each application has registered the same file, the mime type of the file and whether the file should be displayed only by the applications that have registered it.

The GtkRecentManager acts like a database of all the recently used files. You can create new GtkRecentManager objects, but it is more efficient to use the standard recent manager for the GdkScreen so that informations about the recently used files is shared with other people using them. In case the default screen is being used, adding a new recently used file is as simple as:

manager <- gtkRecentManagerGetDefault()
manager$addItem(file_uri)

manager <- gtkRecentManagerGetDefault()
lookup <- manager$lookupItem(file_uri)
if (lookup$error)
  warning("Could not find the file:", lookup$error$message)
else use_info_object(lookup$retval)

Recently used files are supported since GTK+ 2.10.

Structures

GtkRecentManager

Acts as a database of information about the list of recently used files. Normally, you retrieve the recent manager for a particular screen using gtkRecentManagerGetForScreen and it will contain information about current recent manager for that screen.

GtkRecentInfo

Contains informations found when looking up an entry of the recently used files list.

GtkRecentData

Meta-data to be passed to gtkRecentManagerAddFull when registering a recently used resource. GtkRecentData is a transparent-type.

displayName

[character] a UTF-8 encoded string, containing the name of the recently used resource to be displayed, or NULL;

description

[character] a UTF-8 encoded string, containing a short description of the resource, or NULL;

mimeType

[character] the MIME type of the resource;

appName

[character] the name of the application that is registering this recently used resource;

appExec

[character] command line used to launch this resource; may contain the "%f" and "%u" escape characters which will be expanded to the resource file path and URI respectively when the command line is retrieved;

groups

[character] a vector of strings containing groups names;

isPrivate

[logical] whether this resource should be displayed only by the applications that have registered it or not.

Convenient Construction

gtkRecentManager is the equivalent of gtkRecentManagerNew.

Enums and Flags

GtkRecentManagerError

Error codes for GtkRecentManager operations

not-found

the URI specified does not exists in the recently used resources list.

invalid-uri

the URI specified is not valid.

invalid-encoding

the supplied string is not UTF-8 encoded.

not-registered

no application has registered the specified item.

read

failure while reading the recently used resources file.

write

failure while writing the recently used resources file.

unknown

unspecified error.

Signals

changed(recent.manager, user.data)

Emitted when the current recently used resources manager changes its contents. Since 2.10

recent.manager

the recent manager

user.data

user data set when the signal handler was connected.

Properties

filename [character : * : Read / Write / Construct Only]

The full path to the file to be used to store and read the recently used resources list Default value: NULL Since 2.10

limit [integer : Read / Write]

The maximum number of items to be returned by the gtkRecentManagerGetItems function. Allowed values: >= -1 Default value: -1 Since 2.10

size [integer : Read]

The size of the recently used resources list. Allowed values: >= -1 Default value: 0 Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkRecentManager.html

[Package RGtk2 version 2.20.31 Index]