GtkRuler {RGtk2}	R Documentation

GtkRuler

Description

Base class for horizontal or vertical rulers

Methods and Functions

gtkRulerSetMetric(object, metric)
gtkRulerSetRange(object, lower, upper, position, max.size)
gtkRulerGetMetric(object)
gtkRulerGetRange(object)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkRuler
                           +----GtkHRuler
                           +----GtkVRuler

Interfaces

GtkRuler implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

PLEASE NOTE: This widget is considered too specialized/little-used for GTK+, and will in the future be moved to some other package. If your application needs this widget, feel free to use it, as the widget does work and is useful in some applications; it's just not of general interest. However, we are not accepting new features for the widget, and it will eventually move out of the GTK+ distribution. The GTKRuler widget is a base class for horizontal and vertical rulers. Rulers are used to show the mouse pointer's location in a window. The ruler can either be horizontal or vertical on the window. Within the ruler a small triangle indicates the location of the mouse relative to the horizontal or vertical ruler. See GtkHRuler to learn how to create a new horizontal ruler. See GtkVRuler to learn how to create a new vertical ruler.

Structures

GtkRuler

All distances are in 1/72nd's of an inch. (According to Adobe thats a point, but points are really 1/72.27 in.)

Properties

lower [numeric : Read / Write]

Lower limit of ruler. Default value: 0

max-size [numeric : Read / Write]

Maximum size of the ruler. Default value: 0

metric [GtkMetricType : Read / Write]

The metric used for the ruler. Default value: GTK_PIXELS Since 2.8

position [numeric : Read / Write]

Position of mark on the ruler. Default value: 0

upper [numeric : Read / Write]

Upper limit of ruler. Default value: 0

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkRuler.html

See Also

GtkHRuler GtkVRuler

[Package RGtk2 version 2.20.31 Index]