GtkScaleButton {RGtk2}	R Documentation

GtkScaleButton

Description

A button which pops up a scale

Methods and Functions

gtkScaleButtonNew(size, min, max, step, icons, show = TRUE)
gtkScaleButtonSetAdjustment(object, adjustment)
gtkScaleButtonSetIcons(object, icons)
gtkScaleButtonSetValue(object, value)
gtkScaleButtonGetAdjustment(object)
gtkScaleButtonGetValue(object)
gtkScaleButtonGetPopup(object)
gtkScaleButtonGetPlusButton(object)
gtkScaleButtonGetMinusButton(object)
gtkScaleButtonSetOrientation(object, orientation)
gtkScaleButtonSetOrientation(object, orientation)
gtkScaleButtonGetOrientation(object)
gtkScaleButtonGetOrientation(object)
gtkScaleButton(size, min, max, step, icons, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkButton
                                       +----GtkScaleButton
                                             +----GtkVolumeButton

Interfaces

GtkScaleButton implements AtkImplementorIface, GtkBuildable, GtkActivatable and GtkOrientable.

Detailed Description

GtkScaleButton provides a button which pops up a scale widget. This kind of widget is commonly used for volume controls in multimedia applications, and GTK+ provides a GtkVolumeButton subclass that is tailored for this use case.

Structures

GtkScaleButton

undocumented

Convenient Construction

gtkScaleButton is the equivalent of gtkScaleButtonNew.

Signals

popdown(button, user.data)

The ::popdown signal is a keybinding signal which gets emitted to popdown the scale widget.

The default binding for this signal is Escape. Since 2.12

button

the object which received the signal

user.data

user data set when the signal handler was connected.

popup(button, user.data)

The ::popup signal is a keybinding signal which gets emitted to popup the scale widget.

The default bindings for this signal are Space, Enter and Return. Since 2.12

button

the object which received the signal

user.data

user data set when the signal handler was connected.

value-changed(button, value, user.data)

The ::value-changed signal is emitted when the value field has changed. Since 2.12

button

the object which received the signal

value

the new value

user.data

user data set when the signal handler was connected.

Properties

adjustment [GtkAdjustment : * : Read / Write]

The GtkAdjustment that contains the current value of this scale button object.

icons [character list : Read / Write]

The names of the icons to be used by the scale button. The first item in the list will be used in the button when the current value is the lowest value, the second item for the highest value. All the subsequent icons will be used for all the other values, spread evenly over the range of values.

If there's only one icon name in the icons list, it will be used for all the values. If only two icon names are in the icons list, the first one will be used for the bottom 50% of the scale, and the second one for the top 50%.

It is recommended to use at least 3 icons so that the GtkScaleButton reflects the current value of the scale better for the users. Since 2.12

size [GtkIconSize : Read / Write]

The icon size. Default value: GTK_ICON_SIZE_SMALL_TOOLBAR

value [numeric : Read / Write]

The value of the scale. Default value: 0

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkScaleButton.html

[Package RGtk2 version 2.20.31 Index]