GtkScrollbar {RGtk2}	R Documentation

GtkScrollbar

Description

Base class for GtkHScrollbar and GtkVScrollbar

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkRange
                           +----GtkScrollbar
                                 +----GtkHScrollbar
                                 +----GtkVScrollbar

Interfaces

GtkScrollbar implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

The GtkScrollbar widget is an abstract base class for GtkHScrollbar and GtkVScrollbar. It is not very useful in itself.

The position of the thumb in a scrollbar is controlled by the scroll adjustments. See GtkAdjustment for the fields in an adjustment - for GtkScrollbar, the "value" field represents the position of the scrollbar, which must be between the "lower" field and "upper - page_size." The "page_size" field represents the size of the visible scrollable area. The "step_increment" and "page_increment" fields are used when the user asks to step down (using the small stepper arrows) or page down (using for example the PageDown key).

Structures

GtkScrollbar

The GtkScrollbar struct does not contain any public data.

Style Properties

fixed-slider-length [logical : Read]

Don't change slider size, just lock it to the minimum length. Default value: FALSE

has-backward-stepper [logical : Read]

Display the standard backward arrow button. Default value: TRUE

has-forward-stepper [logical : Read]

Display the standard forward arrow button. Default value: TRUE

has-secondary-backward-stepper [logical : Read]

Display a second backward arrow button on the opposite end of the scrollbar. Default value: FALSE

has-secondary-forward-stepper [logical : Read]

Display a second forward arrow button on the opposite end of the scrollbar. Default value: FALSE

min-slider-length [integer : Read]

Minimum length of scrollbar slider. Allowed values: >= 0 Default value: 21

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkScrollbar.html

[Package RGtk2 version 2.20.31 Index]