GtkSeparatorMenuItem {RGtk2}	R Documentation

GtkSeparatorMenuItem

Description

A separator used in menus

Methods and Functions

gtkSeparatorMenuItemNew(show = TRUE)
gtkSeparatorMenuItem(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkItem
                                       +----GtkMenuItem
                                             +----GtkSeparatorMenuItem

Interfaces

GtkSeparatorMenuItem implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

The GtkSeparatorMenuItem is a separator used to group items within a menu. It displays a horizontal line with a shadow to make it appear sunken into the interface.

Structures

GtkSeparatorMenuItem

The GtkSeparatorMenuItem struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkSeparatorMenuItem is the equivalent of gtkSeparatorMenuItemNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkSeparatorMenuItem.html

[Package RGtk2 version 2.20.31 Index]