GtkSeparatorToolItem {RGtk2}	R Documentation

GtkSeparatorToolItem

Description

A toolbar item that separates groups of other toolbar items

Methods and Functions

gtkSeparatorToolItemNew(show = TRUE)
gtkSeparatorToolItemSetDraw(object, draw)
gtkSeparatorToolItemGetDraw(object)
gtkSeparatorToolItem(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkToolItem
                                       +----GtkSeparatorToolItem

Interfaces

GtkSeparatorToolItem implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

A GtkSeparatorItem is a GtkToolItem that separates groups of other GtkToolItems. Depending on the theme, a GtkSeparatorToolItem will often look like a vertical line on horizontally docked toolbars.

If the property "expand" is TRUE and the property "draw" is FALSE, a GtkSeparatorToolItem will act as a "spring" that forces other items to the ends of the toolbar.

Use gtkSeparatorToolItemNew to create a new GtkSeparatorToolItem.

Structures

GtkSeparatorToolItem

The GtkSeparatorToolItem struct contains only private data and should only be accessed through the functions described below.

Convenient Construction

gtkSeparatorToolItem is the equivalent of gtkSeparatorToolItemNew.

Properties

draw [logical : Read / Write]

Whether the separator is drawn, or just blank. Default value: TRUE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkSeparatorToolItem.html

[Package RGtk2 version 2.20.31 Index]