GtkSizeGroup {RGtk2}	R Documentation

GtkSizeGroup

Description

Grouping widgets so they request the same size

Methods and Functions

gtkSizeGroupNew(mode = NULL)
gtkSizeGroupSetMode(object, mode)
gtkSizeGroupGetMode(object)
gtkSizeGroupSetIgnoreHidden(object, ignore.hidden)
gtkSizeGroupGetIgnoreHidden(object)
gtkSizeGroupAddWidget(object, widget)
gtkSizeGroupRemoveWidget(object, widget)
gtkSizeGroupGetWidgets(object)
gtkSizeGroup(mode = NULL)

Hierarchy

GObject
   +----GtkSizeGroup

Interfaces

GtkSizeGroup implements GtkBuildable.

Detailed Description

GtkSizeGroup provides a mechanism for grouping a number of widgets together so they all request the same amount of space. This is typically useful when you want a column of widgets to have the same size, but you can't use a GtkTable widget.

In detail, the size requested for each widget in a GtkSizeGroup is the maximum of the sizes that would have been requested for each widget in the size group if they were not in the size group. The mode of the size group (see gtkSizeGroupSetMode) determines whether this applies to the horizontal size, the vertical size, or both sizes.

Note that size groups only affect the amount of space requested, not the size that the widgets finally receive. If you want the widgets in a GtkSizeGroup to actually be the same size, you need to pack them in such a way that they get the size they request and not more. For example, if you are packing your widgets into a table, you would not include the GTK_FILL flag. GtkSizeGroup objects are referenced by each widget in the size group, so once you have added all widgets to a GtkSizeGroup, you can drop the initial reference to the size group with gObjectUnref(). If the widgets in the size group are subsequently destroyed, then they will be removed from the size group and drop their references on the size group; when all widgets have been removed, the size group will be freed.

Widgets can be part of multiple size groups; GTK+ will compute the horizontal size of a widget from the horizontal requisition of all widgets that can be reached from the widget by a chain of size groups of type GTK_SIZE_GROUP_HORIZONTAL or GTK_SIZE_GROUP_BOTH, and the vertical size from the vertical requisition of all widgets that can be reached from the widget by a chain of size groups of type GTK_SIZE_GROUP_VERTICAL or GTK_SIZE_GROUP_BOTH.

GtkSizeGroup as GtkBuildable

Size groups can be specified in a UI definition by placing an <object> element with class="GtkSizeGroup" somewhere in the UI definition. The widgets that belong to the size group are specified by a <widgets> element that may contain multiple <widget> elements, one for each member of the size group. The name attribute gives the id of the widget.

A UI definition fragment with GtkSizeGroup

<object class="GtkSizeGroup">
  <property name="mode">GTK_SIZE_GROUP_HORIZONTAL</property>
  <widgets>
    <widget name="radio1"/>
    <widget name="radio2"/>
  </widgets>
</object>

Structures

GtkSizeGroup

undocumented

Convenient Construction

gtkSizeGroup is the equivalent of gtkSizeGroupNew.

Enums and Flags

GtkSizeGroupMode

The mode of the size group determines the directions in which the size group affects the requested sizes of its component widgets.

none

group has no effect

horizontal

group affects horizontal requisition

vertical

group affects vertical requisition

both

group affects both horizontal and vertical requisition

Properties

ignore-hidden [logical : Read / Write]

If TRUE, unmapped widgets are ignored when determining the size of the group. Default value: FALSE Since 2.8

mode [GtkSizeGroupMode : Read / Write]

The directions in which the size group affects the requested sizes of its component widgets. Default value: GTK_SIZE_GROUP_HORIZONTAL

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkSizeGroup.html

[Package RGtk2 version 2.20.31 Index]