GtkSocket {RGtk2}	R Documentation

GtkSocket

Description

Container for widgets from other processes

Methods and Functions

gtkSocketNew(show = TRUE)
gtkSocketSteal(object, wid)
gtkSocketAddId(object, window.id)
gtkSocketGetId(object)
gtkSocketGetPlugWindow(object)
gtkSocket(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkSocket

Interfaces

GtkSocket implements AtkImplementorIface and GtkBuildable.

Detailed Description

Together with GtkPlug, GtkSocket provides the ability to embed widgets from one process into another process in a fashion that is transparent to the user. One process creates a GtkSocket widget and, passes the that widget's window ID to the other process, which then creates a GtkPlug with that window ID. Any widgets contained in the GtkPlug then will appear inside the first applications window.

The socket's window ID is obtained by using gtkSocketGetId. Before using this function, the socket must have been realized, and for hence, have been added to its parent.

Obtaining the window ID of a socket.

socket <- gtkSocket()
parent$add(socket)

## The following call is only necessary if one of
## the ancestors of the socket is not yet visible.
socket$realize()
print(paste("The ID of the sockets window is", socket$getId()))

Note that if you pass the window ID of the socket to another process that will create a plug in the socket, you must make sure that the socket widget is not destroyed until that plug is created. Violating this rule will cause unpredictable consequences, the most likely consequence being that the plug will appear as a separate toplevel window. You can check if the plug has been created by examining the plug_window field of the GtkSocket structure. If this field is non-NULL, then the plug has been successfully created inside of the socket.

When GTK+ is notified that the embedded window has been destroyed, then it will destroy the socket as well. You should always, therefore, be prepared for your sockets to be destroyed at any time when the main event loop is running. To prevent this from happening, you can connect to the "plug-removed" signal.

The communication between a GtkSocket and a GtkPlug follows the XEmbed (http://www.freedesktop.org/standards/xembed-spec) protocol. This protocol has also been implemented in other toolkits, e.g. Qt, allowing the same level of integration when embedding a Qt widget in GTK or vice versa. A socket can also be used to swallow arbitrary pre-existing top-level windows using gtkSocketSteal, though the integration when this is done will not be as close as between a GtkPlug and a GtkSocket. PLEASE NOTE: The GtkPlug and GtkSocket widgets are currently not available on all platforms supported by GTK+.

Structures

GtkSocket

The GtkSocket structure contains the plug_window field. (This field should be considered read-only. It should never be set by an application.)

Convenient Construction

gtkSocket is the equivalent of gtkSocketNew.

Signals

plug-added(socket., user.data)

This signal is emitted when a client is successfully added to the socket.

socket.

the object which received the signal

user.data

user data set when the signal handler was connected.

plug-removed(socket., user.data)

This signal is emitted when a client is removed from the socket. The default action is to destroy the GtkSocket widget, so if you want to reuse it you must add a signal handler that returns TRUE.

socket.

the object which received the signal

user.data

user data set when the signal handler was connected.

Returns: [logical] TRUE to stop other handlers from being invoked.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkSocket.html

[Package RGtk2 version 2.20.31 Index]