GtkSpinner {RGtk2}	R Documentation

GtkSpinner

Description

Show a spinner animation

Methods and Functions

gtkSpinnerNew(show = TRUE)
gtkSpinnerStart(object)
gtkSpinnerStop(object)
gtkSpinner(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkDrawingArea
                           +----GtkSpinner

Interfaces

GtkSpinner implements AtkImplementorIface and GtkBuildable.

Detailed Description

A GtkSpinner widget displays an icon-size spinning animation. It is often used as an alternative to a GtkProgressBar for displaying indefinite activity, instead of actual progress.

To start the animation, use gtkSpinnerStart, to stop it use gtkSpinnerStop.

Structures

GtkSpinner

undocumented

Convenient Construction

gtkSpinner is the equivalent of gtkSpinnerNew.

Properties

active [logical : Read / Write]

Whether the spinner is active. Default value: FALSE

Style Properties

cycle-duration [numeric : Read]

The duration in milliseconds for the spinner to complete a full cycle. Allowed values: >= 500 Default value: 1000 Since 2.20

num-steps [numeric : Read]

The number of steps for the spinner to complete a full loop. The animation will complete a full cycle in one second by default (see the "cycle-duration" style property). Allowed values: >= 1 Default value: 12 Since 2.20

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkSpinner.html

[Package RGtk2 version 2.20.31 Index]