GtkStatusbar {RGtk2}	R Documentation

GtkStatusbar

Description

Report messages of minor importance to the user

Methods and Functions

gtkStatusbarNew(show = TRUE)
gtkStatusbarGetContextId(object, context.description)
gtkStatusbarPush(object, context.id, text)
gtkStatusbarPop(object, context.id)
gtkStatusbarRemove(object, context.id, message.id)
gtkStatusbarSetHasResizeGrip(object, setting)
gtkStatusbarGetHasResizeGrip(object)
gtkStatusbarGetMessageArea(object)
gtkStatusbar(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBox
                                 +----GtkHBox
                                       +----GtkStatusbar

Interfaces

GtkStatusbar implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

A GtkStatusbar is usually placed along the bottom of an application's main GtkWindow. It may provide a regular commentary of the application's status (as is usually the case in a web browser, for example), or may be used to simply output a message when the status changes, (when an upload is complete in an FTP client, for example). It may also have a resize grip (a triangular area in the lower right corner) which can be clicked on to resize the window containing the statusbar.

Status bars in GTK+ maintain a stack of messages. The message at the top of the each bar's stack is the one that will currently be displayed.

Any messages added to a statusbar's stack must specify a context id that is used to uniquely identify the source of a message. This context id can be generated by gtkStatusbarGetContextId, given a message and the statusbar that it will be added to. Note that messages are stored in a stack, and when choosing which message to display, the stack structure is adhered to, regardless of the context identifier of a message.

One could say that a statusbar maintains one stack of messages for display purposes, but allows multiple message producers to maintain sub-stacks of the messages they produced (via context ids).

Status bars are created using gtkStatusbarNew.

Messages are added to the bar's stack with gtkStatusbarPush.

The message at the top of the stack can be removed using gtkStatusbarPop. A message can be removed from anywhere in the stack if its message_id was recorded at the time it was added. This is done using gtkStatusbarRemove.

Structures

GtkStatusbar

Contains private data that should be modified with the functions described below.

Convenient Construction

gtkStatusbar is the equivalent of gtkStatusbarNew.

Signals

text-popped(statusbar, context.id, text, user.data)

Is emitted whenever a new message is popped off a statusbar's stack.

statusbar

the object which received the signal.

context.id

the context id of the relevant message/statusbar.

text

the message that was just popped.

user.data

user data set when the signal handler was connected.

text-pushed(statusbar, context.id, text, user.data)

Is emitted whenever a new message gets pushed onto a statusbar's stack.

statusbar

the object which received the signal.

context.id

the context id of the relevant message/statusbar.

text

the message that was pushed.

user.data

user data set when the signal handler was connected.

Properties

has-resize-grip [logical : Read / Write]

Whether the statusbar has a grip for resizing the toplevel window. Default value: TRUE Since 2.4

Style Properties

shadow-type [GtkShadowType : Read]

Style of bevel around the statusbar text. Default value: GTK_SHADOW_IN

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkStatusbar.html

[Package RGtk2 version 2.20.31 Index]