GtkStyle {RGtk2}	R Documentation

Styles

Description

Functions for drawing widget parts

Methods and Functions

gtkStyleNew()
gtkStyleCopy(object)
gtkStyleAttach(object, window)
gtkStyleDetach(object)
gtkStyleSetBackground(object, window, state.type)
gtkStyleApplyDefaultBackground(object, window, set.bg, state.type, area = NULL, x, y, width, height)
gtkStyleLookupColor(object, color.name)
gtkStyleLookupIconSet(object, stock.id)
gtkStyleRenderIcon(object, source, direction, state, size, widget = NULL, detail = NULL)
gtkStyleGetFont(object)
gtkStyleSetFont(object, font)
gtkStyleGetStyleProperty(object, widget.type, property.name)
gtkStyleGet(object, widget.type, first.property.name, ...)
gtkDrawHline(object, window, state.type, x1, x2, y)
gtkDrawVline(object, window, state.type, y1, y2, x)
gtkDrawShadow(object, window, state.type, shadow.type, x, y, width, height)
gtkDrawPolygon(object, window, state.type, shadow.type, points, fill)
gtkDrawArrow(object, window, state.type, shadow.type, arrow.type, fill, x, y, width, height)
gtkDrawDiamond(object, window, state.type, shadow.type, x, y, width, height)
gtkDrawString(object, window, state.type, x, y, string)
gtkDrawBox(object, window, state.type, shadow.type, x, y, width, height)
gtkDrawBoxGap(object, window, state.type, shadow.type, x, y, width, height, gap.side, gap.x, gap.width)
gtkDrawCheck(object, window, state.type, shadow.type, x, y, width, height)
gtkDrawExtension(object, window, state.type, shadow.type, x, y, width, height, gap.side)
gtkDrawFlatBox(object, window, state.type, shadow.type, x, y, width, height)
gtkDrawFocus(object, window, x, y, width, height)
gtkDrawHandle(object, window, state.type, shadow.type, x, y, width, height, orientation)
gtkDrawOption(object, window, state.type, shadow.type, x, y, width, height)
gtkDrawShadowGap(object, window, state.type, shadow.type, x, y, width, height, gap.side, gap.x, gap.width)
gtkDrawSlider(object, window, state.type, shadow.type, x, y, width, height, orientation)
gtkDrawTab(object, window, state.type, shadow.type, x, y, width, height)
gtkDrawExpander(object, window, state.type, x, y, is.open)
gtkDrawLayout(object, window, state.type, use.text, x, y, layout)
gtkDrawResizeGrip(object, window, state.type, edge, x, y, width, height)
gtkPaintArrow(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, arrow.type, fill, x, y, width, height)
gtkPaintBox(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height)
gtkPaintBoxGap(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height, gap.side, gap.x, gap.width)
gtkPaintCheck(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height)
gtkPaintDiamond(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height)
gtkPaintExtension(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height, gap.side)
gtkPaintFlatBox(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height)
gtkPaintFocus(object, window, state.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height)
gtkPaintHandle(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height, orientation)
gtkPaintHline(object, window, state.type, area = NULL, widget = NULL, detail = NULL, x1, x2, y)
gtkPaintOption(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height)
gtkPaintPolygon(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, points, fill)
gtkPaintShadow(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height)
gtkPaintShadowGap(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height, gap.side, gap.x, gap.width)
gtkPaintSlider(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height, orientation)
gtkPaintSpinner(object, window, state.type, area, widget, detail, step, x, y, width, height)
gtkPaintString(object, window, state.type, area = NULL, widget = NULL, detail = NULL, x, y, string)
gtkPaintTab(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height)
gtkPaintVline(object, window, state.type, area = NULL, widget = NULL, detail = NULL, y1, y2, x)
gtkPaintExpander(object, window, state.type, area = NULL, widget = NULL, detail = NULL, x, y, expander.style)
gtkPaintLayout(object, window, state.type, use.text, area = NULL, widget = NULL, detail = NULL, x, y, layout)
gtkPaintResizeGrip(object, window, state.type, area = NULL, widget = NULL, detail = NULL, edge, x, y, width, height)
gtkDrawInsertionCursor(widget, drawable, area = NULL, location, is.primary, direction, draw.arrow)
gtkBorderNew()
gtkBorderCopy(object)
gtkStyle()

Hierarchy

GObject
   +----GtkStyle
GBoxed
   +----GtkBorder

Structures

GtkStyle

undocumented

fg

[GdkColor]

bg

[GdkColor]

light

[GdkColor]

dark

[GdkColor]

mid

[GdkColor]

text

[GdkColor]

base

[GdkColor]

textAa

[GdkColor]

white

[GdkColor]

black

[GdkColor]

fontDesc

[PangoFontDescription]

xthickness

[integer]

ythickness

[integer]

fgGc

[GdkGC]

bgGc

[GdkGC]

lightGc

[GdkGC]

darkGc

[GdkGC]

midGc

[GdkGC]

textGc

[GdkGC]

baseGc

[GdkGC]

textAaGc

[GdkGC]

whiteGc

[GdkGC]

blackGc

[GdkGC]

bgPixmap

[GdkPixmap]

GtkBorder

A struct that specifies a border around a rectangular area that can be of different width on each side.

Convenient Construction

gtkStyle is the equivalent of gtkStyleNew.

User Functions

GtkRcPropertyParser()

undocumented

Signals

realize(style, user.data)

Emitted when the style has been initialized for a particular colormap and depth. Connecting to this signal is probably seldom useful since most of the time applications and widgets only deal with styles that have been already realized. Since 2.4

style

the object which received the signal

user.data

user data set when the signal handler was connected.

unrealize(style, user.data)

Emitted when the aspects of the style specific to a particular colormap and depth are being cleaned up. A connection to this signal can be useful if a widget wants to cache objects like a GdkGC as object data on GtkStyle. This signal provides a convenient place to free such cached objects. Since 2.4

style

the object which received the signal

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkStyle.html

[Package RGtk2 version 2.20.31 Index]