GtkTable {RGtk2}	R Documentation

GtkTable

Description

Pack widgets in regular patterns

Methods and Functions

gtkTableNew(rows = NULL, columns = NULL, homogeneous = NULL, show = TRUE)
gtkTableResize(object, rows, columns)
gtkTableAttach(object, child, left.attach, right.attach, top.attach, bottom.attach, xoptions = 5, yoptions = 5, xpadding = 0, ypadding = 0)
gtkTableAttachDefaults(object, widget, left.attach, right.attach, top.attach, bottom.attach)
gtkTableSetRowSpacing(object, row, spacing)
gtkTableSetColSpacing(object, column, spacing)
gtkTableSetRowSpacings(object, spacing)
gtkTableSetColSpacings(object, spacing)
gtkTableSetHomogeneous(object, homogeneous)
gtkTableGetDefaultRowSpacing(object)
gtkTableGetHomogeneous(object)
gtkTableGetRowSpacing(object, row)
gtkTableGetColSpacing(object, column)
gtkTableGetDefaultColSpacing(object)
gtkTable(rows = NULL, columns = NULL, homogeneous = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkTable

Interfaces

GtkTable implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkTable functions allow the programmer to arrange widgets in rows and columns, making it easy to align many widgets next to each other, horizontally and vertically.

Tables are created with a call to gtkTableNew, the size of which can later be changed with gtkTableResize.

Widgets can be added to a table using gtkTableAttach or the more convenient (but slightly less flexible) gtkTableAttachDefaults.

To alter the space next to a specific row, use gtkTableSetRowSpacing, and for a column, gtkTableSetColSpacing. The gaps between all rows or columns can be changed by calling gtkTableSetRowSpacings or gtkTableSetColSpacings respectively. gtkTableSetHomogeneous, can be used to set whether all cells in the table will resize themselves to the size of the largest widget in the table.

Structures

GtkTable

The GtkTable structure holds the data for the actual table itself. children is a list of all the widgets the table contains. rows and columns are pointers to GtkTableRowCol structures, which contain the default spacing and expansion details for the GtkTable's rows and columns, respectively. nrows and ncols are 16bit integers storing the number of rows and columns the table has.

children

[list]

rows

[GtkTableRowCol]

cols

[GtkTableRowCol]

nrows

[integer]

ncols

[integer]

Convenient Construction

gtkTable is the equivalent of gtkTableNew.

Properties

column-spacing [numeric : Read / Write]

The amount of space between two consecutive columns. Allowed values: <= 65535 Default value: 0

homogeneous [logical : Read / Write]

If TRUE, the table cells are all the same width/height. Default value: FALSE

n-columns [numeric : Read / Write]

The number of columns in the table. Allowed values: [1,65535] Default value: 1

n-rows [numeric : Read / Write]

The number of rows in the table. Allowed values: [1,65535] Default value: 1

row-spacing [numeric : Read / Write]

The amount of space between two consecutive rows. Allowed values: <= 65535 Default value: 0

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkTable.html

[Package RGtk2 version 2.20.31 Index]