GtkTearoffMenuItem {RGtk2}	R Documentation

GtkTearoffMenuItem

Description

A menu item used to tear off and reattach its menu

Methods and Functions

gtkTearoffMenuItemNew(show = TRUE)
gtkTearoffMenuItem(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkItem
                                       +----GtkMenuItem
                                             +----GtkTearoffMenuItem

Interfaces

GtkTearoffMenuItem implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

A GtkTearoffMenuItem is a special GtkMenuItem which is used to tear off and reattach its menu.

When its menu is shown normally, the GtkTearoffMenuItem is drawn as a dotted line indicating that the menu can be torn off. Activating it causes its menu to be torn off and displayed in its own window as a tearoff menu.

When its menu is shown as a tearoff menu, the GtkTearoffMenuItem is drawn as a dotted line which has a left pointing arrow graphic indicating that the tearoff menu can be reattached. Activating it will erase the tearoff menu window.

Structures

GtkTearoffMenuItem

The GtkTearoffMenuItem struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkTearoffMenuItem is the equivalent of gtkTearoffMenuItemNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkTearoffMenuItem.html

[Package RGtk2 version 2.20.31 Index]