GtkTextIter {RGtk2}	R Documentation

GtkTextIter

Description

Text buffer iterator

Methods and Functions

gtkTextIterGetBuffer(object)
gtkTextIterCopy(object)
gtkTextIterGetOffset(object)
gtkTextIterGetLine(object)
gtkTextIterGetLineOffset(object)
gtkTextIterGetLineIndex(object)
gtkTextIterGetVisibleLineIndex(object)
gtkTextIterGetVisibleLineOffset(object)
gtkTextIterGetChar(object)
gtkTextIterGetSlice(object, end)
gtkTextIterGetText(object, end)
gtkTextIterGetVisibleSlice(object, end)
gtkTextIterGetVisibleText(object, end)
gtkTextIterGetPixbuf(object)
gtkTextIterGetMarks(object)
gtkTextIterGetToggledTags(object, toggled.on)
gtkTextIterGetChildAnchor(object)
gtkTextIterBeginsTag(object, tag = NULL)
gtkTextIterEndsTag(object, tag = NULL)
gtkTextIterTogglesTag(object, tag = NULL)
gtkTextIterHasTag(object, tag)
gtkTextIterGetTags(object)
gtkTextIterEditable(object, default.setting)
gtkTextIterCanInsert(object, default.editability)
gtkTextIterStartsWord(object)
gtkTextIterEndsWord(object)
gtkTextIterInsideWord(object)
gtkTextIterStartsLine(object)
gtkTextIterEndsLine(object)
gtkTextIterStartsSentence(object)
gtkTextIterEndsSentence(object)
gtkTextIterInsideSentence(object)
gtkTextIterIsCursorPosition(object)
gtkTextIterGetCharsInLine(object)
gtkTextIterGetBytesInLine(object)
gtkTextIterGetAttributes(object, values)
gtkTextIterGetLanguage(object)
gtkTextIterIsEnd(object)
gtkTextIterIsStart(object)
gtkTextIterForwardChar(object)
gtkTextIterBackwardChar(object)
gtkTextIterForwardChars(object, count)
gtkTextIterBackwardChars(object, count)
gtkTextIterForwardLine(object)
gtkTextIterBackwardLine(object)
gtkTextIterForwardLines(object, count)
gtkTextIterBackwardLines(object, count)
gtkTextIterForwardWordEnds(object, count)
gtkTextIterBackwardWordStarts(object, count)
gtkTextIterForwardWordEnd(object)
gtkTextIterBackwardWordStart(object)
gtkTextIterForwardCursorPosition(object)
gtkTextIterBackwardCursorPosition(object)
gtkTextIterForwardCursorPositions(object, count)
gtkTextIterBackwardCursorPositions(object, count)
gtkTextIterBackwardSentenceStart(object)
gtkTextIterBackwardSentenceStarts(object, count)
gtkTextIterForwardSentenceEnd(object)
gtkTextIterForwardSentenceEnds(object, count)
gtkTextIterForwardVisibleWordEnds(object, count)
gtkTextIterBackwardVisibleWordStarts(object, count)
gtkTextIterForwardVisibleWordEnd(object)
gtkTextIterBackwardVisibleWordStart(object)
gtkTextIterForwardVisibleCursorPosition(object)
gtkTextIterBackwardVisibleCursorPosition(object)
gtkTextIterForwardVisibleCursorPositions(object, count)
gtkTextIterBackwardVisibleCursorPositions(object, count)
gtkTextIterForwardVisibleLine(object)
gtkTextIterBackwardVisibleLine(object)
gtkTextIterForwardVisibleLines(object, count)
gtkTextIterBackwardVisibleLines(object, count)
gtkTextIterSetOffset(object, char.offset)
gtkTextIterSetLine(object, line.number)
gtkTextIterSetLineOffset(object, char.on.line)
gtkTextIterSetLineIndex(object, byte.on.line)
gtkTextIterSetVisibleLineIndex(object, byte.on.line)
gtkTextIterSetVisibleLineOffset(object, char.on.line)
gtkTextIterForwardToEnd(object)
gtkTextIterForwardToLineEnd(object)
gtkTextIterForwardToTagToggle(object, tag = NULL)
gtkTextIterBackwardToTagToggle(object, tag = NULL)
gtkTextIterForwardFindChar(object, pred, user.data = NULL, limit)
gtkTextIterBackwardFindChar(object, pred, user.data = NULL, limit)
gtkTextIterForwardSearch(object, str, flags, limit = NULL)
gtkTextIterBackwardSearch(object, str, flags, limit = NULL)
gtkTextIterEqual(object, rhs)
gtkTextIterCompare(object, rhs)
gtkTextIterInRange(object, start, end)
gtkTextIterOrder(object, second)

Hierarchy

GBoxed
   +----GtkTextIter

Detailed Description

You may wish to begin by reading the text widget conceptual overview which gives an overview of all the objects and data types related to the text widget and how they work together.

Structures

GtkTextIter

undocumented

Enums and Flags

GtkTextSearchFlags

undocumented

visible-only

undocumented

text-only

undocumented

User Functions

GtkTextCharPredicate()

undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkTextIter.html

[Package RGtk2 version 2.20.31 Index]