GtkTextMark {RGtk2}	R Documentation

GtkTextMark

Description

A position in the buffer preserved across buffer modifications

Methods and Functions

gtkTextMarkNew(name, left.gravity)
gtkTextMarkSetVisible(object, setting)
gtkTextMarkGetVisible(object)
gtkTextMarkGetDeleted(object)
gtkTextMarkGetName(object)
gtkTextMarkGetBuffer(object)
gtkTextMarkGetLeftGravity(object)
gtkTextMark(name, left.gravity)

Hierarchy

GObject
   +----GtkTextMark

Detailed Description

You may wish to begin by reading the text widget conceptual overview which gives an overview of all the objects and data types related to the text widget and how they work together.

A GtkTextMark is like a bookmark in a text buffer; it preserves a position in the text. You can convert the mark to an iterator using gtkTextBufferGetIterAtMark. Unlike iterators, marks remain valid across buffer mutations, because their behavior is defined when text is inserted or deleted. When text containing a mark is deleted, the mark remains in the position originally occupied by the deleted text. When text is inserted at a mark, a mark with left gravity will be moved to the beginning of the newly-inserted text, and a mark with right gravity will be moved to the end. PLEASE NOTE: "left" and "right" here refer to logical direction (left is the toward the start of the buffer); in some languages such as Hebrew the logically-leftmost text is not actually on the left when displayed.

Marks are reference counted, but the reference count only controls the validity of the memory; marks can be deleted from the buffer at any time with gtkTextBufferDeleteMark. Once deleted from the buffer, a mark is essentially useless.

Marks optionally have names; these can be convenient to avoid passing the GtkTextMark object around.

Marks are typically created using the gtkTextBufferCreateMark function.

Structures

GtkTextMark

undocumented

Convenient Construction

gtkTextMark is the equivalent of gtkTextMarkNew.

Properties

left-gravity [logical : Read / Write / Construct Only]

Whether the mark has left gravity. Default value: FALSE

name [character : * : Read / Write / Construct Only]

Mark name. Default value: NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkTextMark.html

[Package RGtk2 version 2.20.31 Index]