GtkTextTagTable {RGtk2}	R Documentation

GtkTextTagTable

Description

Collection of tags that can be used together

Methods and Functions

gtkTextTagTableNew()
gtkTextTagTableAdd(object, tag)
gtkTextTagTableRemove(object, tag)
gtkTextTagTableLookup(object, name)
gtkTextTagTableForeach(object, func, data = NULL)
gtkTextTagTableGetSize(object)
gtkTextTagTable()

Hierarchy

GObject
   +----GtkTextTagTable

Detailed Description

You may wish to begin by reading the text widget conceptual overview which gives an overview of all the objects and data types related to the text widget and how they work together.

Structures

GtkTextTagTable

undocumented

Convenient Construction

gtkTextTagTable is the equivalent of gtkTextTagTableNew.

User Functions

GtkTextTagTableForeach()

undocumented

Signals

tag-added(texttagtable, user.data)

undocumented

texttagtable

the object which received the signal.

user.data

user data set when the signal handler was connected.

tag-changed(texttagtable, user.data)

undocumented

texttagtable

the object which received the signal.

user.data

user data set when the signal handler was connected.

tag-removed(texttagtable, user.data)

undocumented

texttagtable

the object which received the signal.

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkTextTagTable.html

[Package RGtk2 version 2.20.31 Index]