GtkTipsQuery {RGtk2}	R Documentation

GtkTipsQuery

Description

Displays help about widgets in the user interface

Methods and Functions

gtkTipsQueryNew(show = TRUE)
gtkTipsQueryStartQuery(object)
gtkTipsQueryStopQuery(object)
gtkTipsQuerySetCaller(object, caller)
gtkTipsQuerySetLabels(object, label.inactive, label.no.tip)
gtkTipsQuery(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkMisc
                           +----GtkLabel
                                 +----GtkTipsQuery

Interfaces

GtkTipsQuery implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkTipsQuery widget is a subclass of GtkLabel which is used to display help about widgets in a user interface.

A query is started with a call to gtkTipsQueryStartQuery, usually when some kind of 'Help' button is pressed. The GtkTipsQuery then grabs all events, stopping the user interface from functioning normally. Then as the user moves the mouse over the widgets, the GtkTipsQuery displays each widget's tooltip text.

By connecting to the "widget-entered" or "widget-selected" signals, it is possible to customize the GtkTipsQuery to perform other actions when widgets are entered or selected. For example, a help browser could be opened with documentation on the widget selected.

At some point a call to gtkTipsQueryStopQuery must be made in order to stop the query and return the interface to its normal state. The gtkTipsQuerySetCaller function can be used to specify a widget which the user can select to stop the query (often the same button used to start the query).

Structures

GtkTipsQuery

WARNING: GtkTipsQuery is deprecated and should not be used in newly-written code. The GtkTipsQuery struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkTipsQuery is the equivalent of gtkTipsQueryNew.

Signals

start-query(tipsquery, user.data)

Emitted when the query is started.

tipsquery

the object which received the signal.

user.data

user data set when the signal handler was connected.

stop-query(tipsquery, user.data)

Emitted when the query is stopped.

tipsquery

the object which received the signal.

user.data

user data set when the signal handler was connected.

widget-entered(tipsquery, widget, tip.text, tip.private, user.data)

Emitted when a widget is entered by the pointer while the query is in effect.

tipsquery

the object which received the signal.

widget

the widget that was entered by the pointer.

tip.text

the widget's tooltip.

tip.private

the widget's private tooltip (see gtkTooltipsSetTip).

user.data

user data set when the signal handler was connected.

widget-selected(tipsquery, widget, tip.text, tip.private, event, user.data)

Emitted when a widget is selected during a query.

tipsquery

the object which received the signal.

widget

the widget that was selected.

tip.text

the widget's tooltip.

tip.private

the widget's private tooltip (see gtkTooltipsSetTip).

event

the button press or button release event.

user.data

user data set when the signal handler was connected.

Returns: [logical] TRUE if the query should be stopped.

Properties

caller [GtkWidget : * : Read / Write]

The widget that starts the tips query, usually a button. If it is selected while the query is in effect the query is automatically stopped.

emit-always [logical : Read / Write]

TRUE if the widget-entered and widget-selected signals are emitted even when the widget has no tooltip set. Default value: FALSE

label-inactive [character : * : Read / Write]

The text to display in the GtkTipsQuery widget when the query is not in effect. Default value: NULL

label-no-tip [character : * : Read / Write]

The text to display in the GtkTipsQuery widget when the query is running and the widget that the pointer is over has no tooltip. Default value: NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkTipsQuery.html

[Package RGtk2 version 2.20.31 Index]