GtkToggleAction {RGtk2}	R Documentation

GtkToggleAction

Description

An action which can be toggled between two states

Methods and Functions

gtkToggleActionNew(name = NULL, label = NULL, tooltip = NULL, stock.id = NULL)
gtkToggleActionToggled(object)
gtkToggleActionSetActive(object, is.active)
gtkToggleActionGetActive(object)
gtkToggleActionSetDrawAsRadio(object, draw.as.radio)
gtkToggleActionGetDrawAsRadio(object)
gtkToggleAction(name = NULL, label = NULL, tooltip = NULL, stock.id = NULL)

Hierarchy

GObject
   +----GtkAction
         +----GtkToggleAction
               +----GtkRadioAction

Interfaces

GtkToggleAction implements GtkBuildable.

Detailed Description

A GtkToggleAction corresponds roughly to a GtkCheckMenuItem. It has an "active" state specifying whether the action has been checked or not.

Structures

GtkToggleAction

The GtkToggleAction struct contains only private members and should not be accessed directly.

Convenient Construction

gtkToggleAction is the equivalent of gtkToggleActionNew.

Signals

toggled(toggleaction, user.data)

undocumented

toggleaction

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

active [logical : Read / Write]

If the toggle action should be active in or not. Default value: FALSE Since 2.10

draw-as-radio [logical : Read / Write]

Whether the proxies for this action look like radio action proxies.

This is an appearance property and thus only applies if "use-action-appearance" is TRUE. Default value: FALSE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkToggleAction.html

[Package RGtk2 version 2.20.31 Index]