GtkToggleToolButton {RGtk2}	R Documentation

GtkToggleToolButton

Description

A GtkToolItem containing a toggle button

Methods and Functions

gtkToggleToolButtonNew(show = TRUE)
gtkToggleToolButtonNewFromStock(stock.id)
gtkToggleToolButtonSetActive(object, is.active)
gtkToggleToolButtonGetActive(object)
gtkToggleToolButton(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkToolItem
                                       +----GtkToolButton
                                             +----GtkToggleToolButton
                                                   +----GtkRadioToolButton

Interfaces

GtkToggleToolButton implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

A GtkToggleToolButton is a GtkToolItem that contains a toggle button.

Use gtkToggleToolButtonNew to create a new GtkToggleToolButton. Use gtkToggleToolButtonNewFromStock to create a new GtkToggleToolButton containing a stock item.

Structures

GtkToggleToolButton

The GtkToggleToolButton struct contains only private data and should only be accessed through the functions described below.

Convenient Construction

gtkToggleToolButton is the equivalent of gtkToggleToolButtonNew.

Signals

toggled(toggle.tool.button, user.data)

Emitted whenever the toggle tool button changes state.

toggle.tool.button

the object that emitted the signal

user.data

user data set when the signal handler was connected.

Properties

active [logical : Read / Write]

If the toggle tool button should be pressed in or not. Default value: FALSE Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkToggleToolButton.html

[Package RGtk2 version 2.20.31 Index]