GtkToolButton {RGtk2}	R Documentation

GtkToolButton

Description

A GtkToolItem subclass that displays buttons

Methods and Functions

gtkToolButtonNew(icon.widget = NULL, label = NULL, show = TRUE)
gtkToolButtonNewFromStock(stock.id, show = TRUE)
gtkToolButtonSetLabel(object, label = NULL)
gtkToolButtonGetLabel(object)
gtkToolButtonSetUseUnderline(object, use.underline)
gtkToolButtonGetUseUnderline(object)
gtkToolButtonSetStockId(object, stock.id = NULL)
gtkToolButtonGetStockId(object)
gtkToolButtonSetIconName(object, icon.name)
gtkToolButtonGetIconName(object)
gtkToolButtonSetIconWidget(object, icon.widget = NULL)
gtkToolButtonGetIconWidget(object)
gtkToolButtonSetLabelWidget(object, label.widget = NULL)
gtkToolButtonGetLabelWidget(object)
gtkToolButton(icon.widget = NULL, label = NULL, stock.id, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkToolItem
                                       +----GtkToolButton
                                             +----GtkMenuToolButton
                                             +----GtkToggleToolButton

Interfaces

GtkToolButton implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

GtkToolButtons are GtkToolItems containing buttons.

Use gtkToolButtonNew to create a new GtkToolButton. Use gtkToolButtonNewWithStock() to create a GtkToolButton containing a stock item.

The label of a GtkToolButton is determined by the properties "label-widget", "label", and "stock-id". If "label-widget" is non-NULL, then that widget is used as the label. Otherwise, if "label" is non-NULL, that string is used as the label. Otherwise, if "stock-id" is non-NULL, the label is determined by the stock item. Otherwise, the button does not have a label.

The icon of a GtkToolButton is determined by the properties "icon-widget" and "stock-id". If "icon-widget" is non-NULL, then that widget is used as the icon. Otherwise, if "stock-id" is non-NULL, the icon is determined by the stock item. Otherwise, the button does not have a icon.

Structures

GtkToolButton

The GtkToolButton struct contains only private. It should only be accessed with the function described below.

Convenient Construction

gtkToolButton is the result of collapsing the constructors of GtkToolButton (gtkToolButtonNew, gtkToolButtonNewFromStock) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Signals

clicked(toolbutton, user.data)

This signal is emitted when the tool button is clicked with the mouse or activated with the keyboard.

toolbutton

the object that emitted the signal

user.data

user data set when the signal handler was connected.

Properties

icon-name [character : * : Read / Write]

The name of the themed icon displayed on the item. This property only has an effect if not overridden by "label", "icon_widget" or "stock_id" properties. Default value: NULL Since 2.8

icon-widget [GtkWidget : * : Read / Write]

Icon widget to display in the item.

label [character : * : Read / Write]

Text to show in the item. Default value: NULL

label-widget [GtkWidget : * : Read / Write]

Widget to use as the item label.

stock-id [character : * : Read / Write]

The stock icon displayed on the item. Default value: NULL

use-underline [logical : Read / Write]

If set, an underline in the label property indicates that the next character should be used for the mnemonic accelerator key in the overflow menu. Default value: FALSE

Style Properties

icon-spacing [integer : Read / Write]

Spacing in pixels between the icon and label. Allowed values: >= 0 Default value: 3

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkToolButton.html

[Package RGtk2 version 2.20.31 Index]