GtkToolItem {RGtk2}	R Documentation

GtkToolItem

Description

The base class of widgets that can be added to GtkToolShell

Methods and Functions

gtkToolItemNew(show = TRUE)
gtkToolItemSetHomogeneous(object, homogeneous)
gtkToolItemGetHomogeneous(object)
gtkToolItemSetExpand(object, expand)
gtkToolItemGetExpand(object)
gtkToolItemSetTooltip(object, tooltips, tip.text = NULL, tip.private = NULL)
gtkToolItemSetTooltip(object, tooltips, tip.text = NULL, tip.private = NULL)
gtkToolItemSetTooltipText(object, text)
gtkToolItemSetTooltipMarkup(object, markup)
gtkToolItemSetUseDragWindow(object, use.drag.window)
gtkToolItemGetUseDragWindow(object)
gtkToolItemSetVisibleHorizontal(object, visible.horizontal)
gtkToolItemGetVisibleHorizontal(object)
gtkToolItemSetVisibleVertical(object, visible.vertical)
gtkToolItemGetVisibleVertical(object)
gtkToolItemSetIsImportant(object, is.important)
gtkToolItemGetIsImportant(object)
gtkToolItemGetEllipsizeMode(object)
gtkToolItemGetIconSize(object)
gtkToolItemGetOrientation(object)
gtkToolItemGetToolbarStyle(object)
gtkToolItemGetReliefStyle(object)
gtkToolItemGetTextAlignment(object)
gtkToolItemGetTextOrientation(object)
gtkToolItemRetrieveProxyMenuItem(object)
gtkToolItemGetProxyMenuItem(object, menu.item.id)
gtkToolItemSetProxyMenuItem(object, menu.item.id, menu.item = NULL)
gtkToolItemRebuildMenu(object)
gtkToolItemToolbarReconfigured(object)
gtkToolItemGetTextSizeGroup(object)
gtkToolItem(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkToolItem
                                       +----GtkToolButton
                                       +----GtkSeparatorToolItem

Interfaces

GtkToolItem implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

GtkToolItems are widgets that can appear on a toolbar. To create a toolbar item that contain something else than a button, use gtkToolItemNew. Use gtkContainerAdd to add a child widget to the tool item.

For toolbar items that contain buttons, see the GtkToolButton, GtkToggleToolButton and GtkRadioToolButton classes.

See the GtkToolbar class for a description of the toolbar widget, and GtkToolShell for a description of the tool shell interface.

Structures

GtkToolItem

The GtkToolItem struct contains only private data. It should only be accessed through the functions described below.

Convenient Construction

gtkToolItem is the equivalent of gtkToolItemNew.

Signals

create-menu-proxy(tool.item, user.data)

This signal is emitted when the toolbar needs information from tool.item about whether the item should appear in the toolbar overflow menu. In response the tool item should either

The toolbar may cache the result of this signal. When the tool item changes how it will respond to this signal it must call gtkToolItemRebuildMenu to invalidate the cache and ensure that the toolbar rebuilds its overflow menu.

tool.item

the object the signal was emitted on

user.data

user data set when the signal handler was connected.

Returns: [logical] TRUE if the signal was handled, FALSE if not

set-tooltip(tool.item, tooltips, tip.text, tip.private, user.data)

This signal is emitted when the toolitem's tooltip changes. Application developers can use gtkToolItemSetTooltip to set the item's tooltip.

tool.item

the object the signal was emitted on

tooltips

the GtkTooltips

tip.text

the tooltip text

tip.private

the tooltip private text

user.data

user data set when the signal handler was connected.

Returns: [logical] TRUE if the signal was handled, FALSE if not

toolbar-reconfigured(tool.item, user.data)

This signal is emitted when some property of the toolbar that the item is a child of changes. For custom subclasses of GtkToolItem, the default handler of this signal use the functions

to find out what the toolbar should look like and change themselves accordingly.

tool.item

the object the signal was emitted on

user.data

user data set when the signal handler was connected.

Properties

is-important [logical : Read / Write]

Whether the toolbar item is considered important. When TRUE, toolbar buttons show text in GTK_TOOLBAR_BOTH_HORIZ mode. Default value: FALSE

visible-horizontal [logical : Read / Write]

Whether the toolbar item is visible when the toolbar is in a horizontal orientation. Default value: TRUE

visible-vertical [logical : Read / Write]

Whether the toolbar item is visible when the toolbar is in a vertical orientation. Default value: TRUE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkToolItem.html

[Package RGtk2 version 2.20.31 Index]