GtkToolItemGroup {RGtk2}	R Documentation

GtkToolItemGroup

Description

A sub container used in a tool palette

Methods and Functions

gtkToolItemGroupGetCollapsed(object)
gtkToolItemGroupGetDropItem(object, x, y)
gtkToolItemGroupGetEllipsize(object)
gtkToolItemGroupGetItemPosition(object, item)
gtkToolItemGroupGetNItems(object)
gtkToolItemGroupGetLabel(object)
gtkToolItemGroupGetLabelWidget(object)
gtkToolItemGroupGetNthItem(object, index)
gtkToolItemGroupGetHeaderRelief(object)
gtkToolItemGroupInsert(object, item, position)
gtkToolItemGroupNew(label, show = TRUE)
gtkToolItemGroupSetCollapsed(object, collapsed)
gtkToolItemGroupSetEllipsize(object, ellipsize)
gtkToolItemGroupSetItemPosition(object, item, position)
gtkToolItemGroupSetLabel(object, label)
gtkToolItemGroupSetLabelWidget(object, label.widget)
gtkToolItemGroupSetHeaderRelief(object, style)
gtkToolItemGroup(label, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkToolItemGroup

Interfaces

GtkToolItemGroup implements AtkImplementorIface, GtkBuildable and GtkToolShell.

Detailed Description

A GtkToolItemGroup is used together with GtkToolPalette to add GtkToolItems to a palette like container with different categories and drag and drop support.

Structures

GtkToolItemGroup

This should not be accessed directly. Use the accessor functions below.

Convenient Construction

gtkToolItemGroup is the equivalent of gtkToolItemGroupNew.

Properties

collapsed [logical : Read / Write]

Wether the group has been collapsed and items are hidden. Default value: FALSE

ellipsize [PangoEllipsizeMode : Read / Write]

Ellipsize for item group headers. Default value: PANGO_ELLIPSIZE_NONE

header-relief [GtkReliefStyle : Read / Write]

Relief of the group header button. Default value: GTK_RELIEF_NORMAL

label [character : * : Read / Write]

The human-readable title of this item group. Default value: ""

label-widget [GtkWidget : * : Read / Write]

A widget to display in place of the usual label.

Style Properties

expander-size [integer : Read]

Size of the expander arrow. Allowed values: >= 0 Default value: 16

header-spacing [integer : Read]

Spacing between expander arrow and caption. Allowed values: >= 0 Default value: 2

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkToolItemGroup.html

[Package RGtk2 version 2.20.31 Index]