GtkTooltip {RGtk2}	R Documentation

GtkTooltip

Description

Add tips to your widgets

Methods and Functions

gtkTooltipSetMarkup(object, markup)
gtkTooltipSetText(object, text)
gtkTooltipSetIcon(object, pixbuf)
gtkTooltipSetIconFromStock(object, stock.id, size)
gtkTooltipSetIconFromIconName(object, icon.name = NULL, size)
gtkTooltipSetIconFromGicon(object, gicon, size)
gtkTooltipSetCustom(object, custom.widget)
gtkTooltipTriggerTooltipQuery(display)
gtkTooltipSetTipArea(object, area)

Hierarchy

GObject
   +----GtkTooltip

Detailed Description

GtkTooltip belongs to the new tooltips API that was introduced in GTK+ 2.12 and which deprecates the old GtkTooltips API.

Basic tooltips can be realized simply by using gtkWidgetSetTooltipText or gtkWidgetSetTooltipMarkup without any explicit tooltip object.

When you need a tooltip with a little more fancy contents, like adding an image, or you want the tooltip to have different contents per GtkTreeView row or cell, you will have to do a little more work:

Set the "has-tooltip" property to TRUE, this will make GTK+ monitor the widget for motion and related events which are needed to determine when and where to show a tooltip.

Connect to the "query-tooltip" signal. This signal will be emitted when a tooltip is supposed to be shown. One of the arguments passed to the signal handler is a GtkTooltip object. This is the object that we are about to display as a tooltip, and can be manipulated in your callback using functions like gtkTooltipSetIcon. There are functions for setting the tooltip's markup, setting an image from a stock icon, or even putting in a custom widget.

Return TRUE from your query-tooltip handler. This causes the tooltip to be show. If you return FALSE, it will not be shown.

In the probably rare case where you want to have even more control over the tooltip that is about to be shown, you can set your own GtkWindow which will be used as tooltip window. This works as follows:

Set "has-tooltip" and connect to "query-tooltip" as before.

Use gtkWidgetSetTooltipWindow to set a GtkWindow created by you as tooltip window.

In the ::query-tooltip callback you can access your window using gtkWidgetGetTooltipWindow and manipulate as you wish. The semantics of the return value are exactly as before, return TRUE to show the window, FALSE to not show it.

Structures

GtkTooltip

undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkTooltip.html

[Package RGtk2 version 2.20.31 Index]