GtkTreeModel {RGtk2}	R Documentation

GtkTreeModel

Description

The tree interface used by GtkTreeView

Methods and Functions

gtkTreePathNew()
gtkTreePathNewFromString(path)
gtkTreePathNewFromIndices(...)
gtkTreePathToString(object)
gtkTreePathNewFirst()
gtkTreePathAppendIndex(object, index)
gtkTreePathPrependIndex(object, index)
gtkTreePathGetDepth(object)
gtkTreePathGetIndices(object)
gtkTreePathCopy(object)
gtkTreePathCompare(object, b)
gtkTreePathNext(object)
gtkTreePathPrev(object)
gtkTreePathUp(object)
gtkTreePathDown(object)
gtkTreePathIsAncestor(object, descendant)
gtkTreePathIsDescendant(object, ancestor)
gtkTreeRowReferenceNew(model, path)
gtkTreeRowReferenceNewProxy(proxy, model, path)
gtkTreeRowReferenceGetModel(object)
gtkTreeRowReferenceGetPath(object)
gtkTreeRowReferenceValid(object)
gtkTreeRowReferenceCopy(object)
gtkTreeRowReferenceInserted(proxy, path)
gtkTreeRowReferenceDeleted(proxy, path)
gtkTreeRowReferenceReordered(proxy, path, iter, new.order)
gtkTreeIterCopy(object)
gtkTreeModelGetFlags(object)
gtkTreeModelGetNColumns(object)
gtkTreeModelGetColumnType(object, index)
gtkTreeModelGetIter(object, path)
gtkTreeModelGetIterFromString(object, path.string)
gtkTreeModelGetIterFirst(object)
gtkTreeModelGetPath(object, iter)
gtkTreeModelGetValue(object, iter, column)
gtkTreeModelIterNext(object, iter)
gtkTreeModelIterChildren(object, parent = NULL)
gtkTreeModelIterHasChild(object, iter)
gtkTreeModelIterNChildren(object, iter = NULL)
gtkTreeModelIterNthChild(object, parent = NULL, n)
gtkTreeModelIterParent(object, child)
gtkTreeModelGetStringFromIter(object, iter)
gtkTreeModelRefNode(object, iter)
gtkTreeModelUnrefNode(object, iter)
gtkTreeModelGet(object, iter, ...)
gtkTreeModelForeach(object, func, user.data = NULL)
gtkTreeModelRowChanged(object, path, iter)
gtkTreeModelRowInserted(object, path, iter)
gtkTreeModelRowHasChildToggled(object, path, iter)
gtkTreeModelRowDeleted(object, path)
gtkTreeModelRowsReordered(object, path, iter, new.order)

Hierarchy

GInterface
   +----GtkTreeModel
GBoxed
   +----GtkTreeIter
GBoxed
   +----GtkTreePath

Implementations

GtkTreeModel is implemented by GtkListStore, GtkTreeModelFilter, GtkTreeModelSort and GtkTreeStore.

Interface Derivations

GtkTreeModel is required by GtkTreeSortable.

Detailed Description

The GtkTreeModel interface defines a generic tree interface for use by the GtkTreeView widget. It is an abstract interface, and is designed to be usable with any appropriate data structure. The programmer just has to implement this interface on their own data type for it to be viewable by a GtkTreeView widget.

The model is represented as a hierarchical tree of strongly-typed, columned data. In other words, the model can be seen as a tree where every node has different values depending on which column is being queried. The type of data found in a column is determined by using the GType system (ie. G_TYPE_INT, GTK_TYPE_BUTTON, G_TYPE_POINTER, etc.). The types are homogeneous per column across all nodes. It is important to note that this interface only provides a way of examining a model and observing changes. The implementation of each individual model decides how and if changes are made.

In order to make life simpler for programmers who do not need to write their own specialized model, two generic models are provided – the GtkTreeStore and the GtkListStore. To use these, the developer simply pushes data into these models as necessary. These models provide the data structure as well as all appropriate tree interfaces. As a result, implementing drag and drop, sorting, and storing data is trivial. For the vast majority of trees and lists, these two models are sufficient.

Models are accessed on a node/column level of granularity. One can query for the value of a model at a certain node and a certain column on that node. There are two structures used to reference a particular node in a model. They are the GtkTreePath and the GtkTreeIter PLEASE NOTE: Here, iter is short for “iterator” Most of the interface consists of operations on a GtkTreeIter.

A path is essentially a potential node. It is a location on a model that may or may not actually correspond to a node on a specific model. The GtkTreePath struct can be converted into either a list of unsigned integers or a string. The string form is a list of numbers separated by a colon. Each number refers to the offset at that level. Thus, the path “0” refers to the root node and the path “2:4” refers to the fifth child of the third node.

By contrast, a GtkTreeIter is a reference to a specific node on a specific model. It is a generic struct with an integer and three generic pointers. These are filled in by the model in a model-specific way. One can convert a path to an iterator by calling gtkTreeModelGetIter. These iterators are the primary way of accessing a model and are similar to the iterators used by GtkTextBuffer. They are generally statically allocated on the stack and only used for a short time. The model interface defines a set of operations using them for navigating the model.

It is expected that models fill in the iterator with private data. For example, the GtkListStore model, which is internally a simple linked list, stores a list node in one of the pointers. The GtkTreeModelSort stores a list and an offset in two of the pointers. Additionally, there is an integer field. This field is generally filled with a unique stamp per model. This stamp is for catching errors resulting from using invalid iterators with a model.

The lifecycle of an iterator can be a little confusing at first. Iterators are expected to always be valid for as long as the model is unchanged (and doesn't emit a signal). The model is considered to own all outstanding iterators and nothing needs to be done to free them from the user's point of view. Additionally, some models guarantee that an iterator is valid for as long as the node it refers to is valid (most notably the GtkTreeStore and GtkListStore). Although generally uninteresting, as one always has to allow for the case where iterators do not persist beyond a signal, some very important performance enhancements were made in the sort model. As a result, the GTK_TREE_MODEL_ITERS_PERSIST flag was added to indicate this behavior.

To help show some common operation of a model, some examples are provided. The first example shows three ways of getting the iter at the location “3:2:5”. While the first method shown is easier, the second is much more common, as you often get paths from callbacks.

Acquiring a GtkTreeIter

## Acquiring a GtkTreeIter

## Three ways of getting the iter pointing to the location

## get the iterator from a string 
model$getIterFromString("3:2:5")$iter

## get the iterator from a path
path <- gtkTreePathNewFromString("3:2:5")
model$getIter(path)$iter

## walk the tree to find the iterator
parent_iter <- model$iterNthChild(NULL, 3)$iter
parent_iter <- model$iterNthChild(parent_iter, 2)$iter
model$iterNthChild(parent_iter, 5)$iter

This second example shows a quick way of iterating through a list and getting a string and an integer from each row. The populateModel function used below is not shown, as it is specific to the GtkListStore. For information on how to write such a function, see the GtkListStore documentation.

Reading data from a GtkTreeModel

## Reading data from a GtkTreeModel

## make a new list_store
list_store <- gtkListStore("character", "integer")

## Fill the list store with data
populate_model(list_store)

## Get the first iter in the list 
result <- list_store$getIterFirst()

row_count <- 1
while(result[[1]]) {
  ## Walk through the list, reading each row
      
  data <- list_store$get(result$iter, 0, 1)
  
  ## Do something with the data
  print(paste("Row ", row_count, ": (", data[[1]], ",", data[[2]], ")", sep=""))
  
  row_count <- row_count + 1
  result <- list_store$iterNext()
}

Structures

GtkTreeModel

undocumented

GtkTreeIter

The GtkTreeIter is the primary structure for accessing a structure. Models are expected to put a unique integer in the stamp member, and put model-specific data in the three user_data members.

GtkTreePath

undocumented

GtkTreeRowReference

undocumented

Enums and Flags

GtkTreeModelFlags

These flags indicate various properties of a GtkTreeModel. They are returned by gtkTreeModelGetFlags, and must be static for the lifetime of the object. A more complete description of GTK_TREE_MODEL_ITERS_PERSIST can be found in the overview of this section.

iters-persist

Iterators survive all signals emitted by the tree.

list-only

The model is a list only, and never has children

User Functions

GtkTreeModelForeachFunc(model, path, iter, data)

undocumented

model

The GtkTreeModel currently being iterated

path

The current GtkTreePath

iter

The current GtkTreeIter

data

The user data passed to gtkTreeModelForeach

Returns: [logical] TRUE to stop iterating, FALSE to continue.

Signals

row-changed(tree.model, path, iter, user.data)

This signal is emitted when a row in the model has changed.

tree.model

the GtkTreeModel on which the signal is emitted

path

a GtkTreePath identifying the changed row

iter

a valid GtkTreeIter pointing to the changed row

user.data

user data set when the signal handler was connected.

row-deleted(tree.model, path, user.data)

This signal is emitted when a row has been deleted.

Note that no iterator is passed to the signal handler, since the row is already deleted.

Implementations of GtkTreeModel must emit row-deleted before removing the node from its internal data structures. This is because models and views which access and monitor this model might have references on the node which need to be released in the row-deleted handler.

tree.model

the GtkTreeModel on which the signal is emitted

path

a GtkTreePath identifying the row

user.data

user data set when the signal handler was connected.

row-has-child-toggled(tree.model, path, iter, user.data)

This signal is emitted when a row has gotten the first child row or lost its last child row.

tree.model

the GtkTreeModel on which the signal is emitted

path

a GtkTreePath identifying the row

iter

a valid GtkTreeIter pointing to the row

user.data

user data set when the signal handler was connected.

row-inserted(tree.model, path, iter, user.data)

This signal is emitted when a new row has been inserted in the model.

Note that the row may still be empty at this point, since it is a common pattern to first insert an empty row, and then fill it with the desired values.

tree.model

the GtkTreeModel on which the signal is emitted

path

a GtkTreePath identifying the new row

iter

a valid GtkTreeIter pointing to the new row

user.data

user data set when the signal handler was connected.

rows-reordered(tree.model, path, iter, new.order, user.data)

This signal is emitted when the children of a node in the GtkTreeModel have been reordered.

Note that this signal is not emitted when rows are reordered by DND, since this is implemented by removing and then reinserting the row.

tree.model

the GtkTreeModel on which the signal is emitted

path

a GtkTreePath identifying the tree node whose children have been reordered

iter

a valid GtkTreeIter pointing to the node whose

new.order

a list of integers mapping the current position of each child to its old position before the re-ordering, i.e. new.order[newpos] = oldpos.

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkTreeModel.html

See Also

GtkTreeView GtkTreeStore GtkListStore GtkTreeSortable

[Package RGtk2 version 2.20.31 Index]