GtkTreeStore {RGtk2}	R Documentation

GtkTreeStore

Description

A tree-like data structure that can be used with the GtkTreeView

Methods and Functions

gtkTreeStoreNew(...)
gtkTreeStoreNewv(types)
gtkTreeStoreSetColumnTypes(object, types)
gtkTreeStoreSetValue(object, iter, column, value)
gtkTreeStoreSet(object, iter, ...)
gtkTreeStoreSetValuesv(object, iter, columns, values)
gtkTreeStoreRemove(object, iter)
gtkTreeStoreInsert(object, parent = NULL, position)
gtkTreeStoreInsertBefore(object, parent, sibling)
gtkTreeStoreInsertAfter(object, parent, sibling)
gtkTreeStoreInsertWithValues(object, parent, position, ...)
gtkTreeStoreInsertWithValuesv(object, parent, position, columns, values)
gtkTreeStorePrepend(object, parent = NULL)
gtkTreeStoreAppend(object, parent = NULL)
gtkTreeStoreIsAncestor(object, iter, descendant)
gtkTreeStoreIterDepth(object, iter)
gtkTreeStoreClear(object)
gtkTreeStoreIterIsValid(object, iter)
gtkTreeStoreReorder(object, parent, new.order)
gtkTreeStoreSwap(object, a, b)
gtkTreeStoreMoveBefore(object, iter, position = NULL)
gtkTreeStoreMoveAfter(object, iter, position = NULL)
gtkTreeStore(..., types)

Hierarchy

GObject
   +----GtkTreeStore

Interfaces

GtkTreeStore implements GtkTreeModel, GtkTreeDragSource, GtkTreeDragDest, GtkTreeSortable and GtkBuildable.

Detailed Description

The GtkTreeStore object is a list model for use with a GtkTreeView widget. It implements the GtkTreeModel interface, and consequentialy, can use all of the methods available there. It also implements the GtkTreeSortable interface so it can be sorted by the view. Finally, it also implements the tree drag and drop interfaces.

GtkTreeStore as GtkBuildable

The GtkTreeStore implementation of the GtkBuildable interface allows to specify the model columns with a <columns> element that may contain multiple <column> elements, each specifying one model column. The "type" attribute specifies the data type for the column.

A UI Definition fragment for a tree store

<object class="GtkTreeStore">
  <columns>
    <column type="gchararray"/>
    <column type="gchararray"/>
    <column type="gint"/>
  </columns>
</object>

Structures

GtkTreeStore

undocumented

Convenient Construction

gtkTreeStore is the result of collapsing the constructors of GtkTreeStore (gtkTreeStoreNew, gtkTreeStoreNewv) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkTreeStore.html

See Also

GtkTreeModel GtkTreeStore

[Package RGtk2 version 2.20.31 Index]