GtkUIManager {RGtk2}	R Documentation

GtkUIManager

Description

Constructing menus and toolbars from an XML description

Methods and Functions

gtkUIManagerNew()
gtkUIManagerSetAddTearoffs(object, add.tearoffs)
gtkUIManagerGetAddTearoffs(object)
gtkUIManagerInsertActionGroup(object, action.group, pos)
gtkUIManagerRemoveActionGroup(object, action.group)
gtkUIManagerGetActionGroups(object)
gtkUIManagerGetAccelGroup(object)
gtkUIManagerGetWidget(object, path)
gtkUIManagerGetToplevels(object, types)
gtkUIManagerGetAction(object, path)
gtkUIManagerAddUiFromString(object, buffer, length = -1, .errwarn = TRUE)
gtkUIManagerNewMergeId(object)
gtkUIManagerAddUi(object, merge.id, path, name, action = NULL, type, top)
gtkUIManagerRemoveUi(object, merge.id)
gtkUIManagerGetUi(object)
gtkUIManagerEnsureUpdate(object)
gtkUIManager()

Hierarchy

GObject
   +----GtkUIManager

Interfaces

GtkUIManager implements GtkBuildable.

Detailed Description

A GtkUIManager constructs a user interface (menus and toolbars) from one or more UI definitions, which reference actions from one or more action groups.

UI Definitions

The UI definitions are specified in an XML format which can be roughly described by the following DTD.

Do not confuse the GtkUIManager UI Definitions described here with the similarly named GtkBuilder UI Definitions.

<!ELEMENT ui          (menubar|toolbar|popup|accelerator)* >
<!ELEMENT menubar     (menuitem|separator|placeholder|menu)* >
<!ELEMENT menu        (menuitem|separator|placeholder|menu)* >
<!ELEMENT popup       (menuitem|separator|placeholder|menu)* >
<!ELEMENT toolbar     (toolitem|separator|placeholder)* >
<!ELEMENT placeholder (menuitem|toolitem|separator|placeholder|menu)* >
<!ELEMENT menuitem     EMPTY >
<!ELEMENT toolitem     (menu?) >
<!ELEMENT separator    EMPTY >
<!ELEMENT accelerator  EMPTY >
<!ATTLIST menubar      name                      #IMPLIED
                       action                    #IMPLIED >
<!ATTLIST toolbar      name                      #IMPLIED
                       action                    #IMPLIED >
<!ATTLIST popup        name                      #IMPLIED
                       action                    #IMPLIED
                       accelerators (true|false) #IMPLIED >
<!ATTLIST placeholder  name                      #IMPLIED
                       action                    #IMPLIED >
<!ATTLIST separator    name                      #IMPLIED
                       action                    #IMPLIED
                       expand       (true|false) #IMPLIED >
<!ATTLIST menu         name                      #IMPLIED
                       action                    #REQUIRED
                       position     (top|bot)    #IMPLIED >
<!ATTLIST menuitem     name                      #IMPLIED
                       action                    #REQUIRED
                       position     (top|bot)    #IMPLIED
                       always-show-image (true|false) #IMPLIED >
<!ATTLIST toolitem     name                      #IMPLIED
                       action                    #REQUIRED
                       position     (top|bot)    #IMPLIED >
<!ATTLIST accelerator  name                      #IMPLIED
                       action                    #REQUIRED >

There are some additional restrictions beyond those specified in the DTD, e.g. every toolitem must have a toolbar in its anchestry and every menuitem must have a menubar or popup in its anchestry. Since a GMarkup parser is used to parse the UI description, it must not only be valid XML, but valid GMarkup.

If a name is not specified, it defaults to the action. If an action is not specified either, the element name is used. The name and action attributes must not contain '/' characters after parsing (since that would mess up path lookup) and must be usable as XML attributes when enclosed in doublequotes, thus they must not '"' characters or references to the &quot; entity.

A UI definition

<ui>
  <menubar>
    <menu name="FileMenu" action="FileMenuAction">
      <menuitem name="New" action="New2Action" />
      <placeholder name="FileMenuAdditions" />
    </menu>
    <menu name="JustifyMenu" action="JustifyMenuAction">
      <menuitem name="Left" action="justify-left"/>
      <menuitem name="Centre" action="justify-center"/>
      <menuitem name="Right" action="justify-right"/>
      <menuitem name="Fill" action="justify-fill"/>
    </menu>
  </menubar>
  <toolbar action="toolbar1">
    <placeholder name="JustifyToolItems">
      <separator/>
      <toolitem name="Left" action="justify-left"/>
      <toolitem name="Centre" action="justify-center"/>
      <toolitem name="Right" action="justify-right"/>
      <toolitem name="Fill" action="justify-fill"/>
      <separator/>
    </placeholder>
  </toolbar>
</ui>

The constructed widget hierarchy is very similar to the element tree of the XML, with the exception that placeholders are merged into their parents. The correspondence of XML elements to widgets should be almost obvious:

menubar

a GtkMenuBar

toolbar

a GtkToolbar

popup

a toplevel GtkMenu

menu

a GtkMenu attached to a menuitem

menuitem

a GtkMenuItem subclass, the exact type depends on the action

toolitem

a GtkToolItem subclass, the exact type depends on the action. Note that toolitem elements may contain a menu element, but only if their associated action specifies a GtkMenuToolButton as proxy.

separator

a GtkSeparatorMenuItem or GtkSeparatorToolItem

accelerator

a keyboard accelerator

The "position" attribute determines where a constructed widget is positioned wrt. to its siblings in the partially constructed tree. If it is "top", the widget is prepended, otherwise it is appended.

UI Merging

The most remarkable feature of GtkUIManager is that it can overlay a set of menuitems and toolitems over another one, and demerge them later.

Merging is done based on the names of the XML elements. Each element is identified by a path which consists of the names of its anchestors, separated by slashes. For example, the menuitem named "Left" in the example above has the path /ui/menubar/JustifyMenu/Left and the toolitem with the same name has path /ui/toolbar1/JustifyToolItems/Left.

Accelerators

Every action has an accelerator path. Accelerators are installed together with menuitem proxies, but they can also be explicitly added with <accelerator> elements in the UI definition. This makes it possible to have accelerators for actions even if they have no visible proxies.

Smart Separators

The separators created by GtkUIManager are "smart", i.e. they do not show up in the UI unless they end up between two visible menu or tool items. Separators which are located at the very beginning or end of the menu or toolbar containing them, or multiple separators next to each other, are hidden. This is a useful feature, since the merging of UI elements from multiple sources can make it hard or impossible to determine in advance whether a separator will end up in such an unfortunate position.

For separators in toolbars, you can set expand="true" to turn them from a small, visible separator to an expanding, invisible one. Toolitems following an expanding separator are effectively right-aligned.

Empty Menus

Submenus pose similar problems to separators inconnection with merging. It is impossible to know in advance whether they will end up empty after merging. GtkUIManager offers two ways to treat empty submenus:

make them disappear by hiding the menu item they're attached to

add an insensitive "Empty" item

The behaviour is chosen based on the "hide_if_empty" property of the action to which the submenu is associated.

GtkUIManager as GtkBuildable

The GtkUIManager implementation of the GtkBuildable interface accepts GtkActionGroup objects as <child> elements in UI definitions.

A GtkUIManager UI definition as described above can be embedded in an GtkUIManager <object> element in a GtkBuilder UI definition.

The widgets that are constructed by a GtkUIManager can be embedded in other parts of the constructed user interface with the help of the "constructor" attribute. See the example below.

An embedded GtkUIManager UI definition

<object class="GtkUIManager" id="uiman">
  <child>
    <object class="GtkActionGroup" id="actiongroup">
      <child>
        <object class="GtkAction" id="file">
          <property name="label">_File</property>
        </object>
      </child>
    </object>
  </child>
  <ui>
    <menubar name="menubar1">
      <menu action="file">
      </menu>
    </menubar>
  </ui>
</object>
<object class="GtkWindow" id="main-window">
  <child>
    <object class="GtkMenuBar" id="menubar1" constructor="uiman"/>
  </child>
</object>

Structures

GtkUIManager

The GtkUIManager struct contains only private members and should not be accessed directly.

Convenient Construction

gtkUIManager is the equivalent of gtkUIManagerNew.

Enums and Flags

GtkUIManagerItemType

These enumeration values are used by gtkUIManagerAddUi to determine what UI element to create.

auto

Pick the type of the UI element according to context.

menubar

Create a menubar.

menu

Create a menu.

toolbar

Create a toolbar.

placeholder

Insert a placeholder.

popup

Create a popup menu.

menuitem

Create a menuitem.

toolitem

Create a toolitem.

separator

Create a separator.

accelerator

Install an accelerator.

Signals

actions-changed(merge, user.data)

The "actions-changed" signal is emitted whenever the set of actions changes. Since 2.4

merge

a GtkUIManager

user.data

user data set when the signal handler was connected.

add-widget(merge, widget, user.data)

The add_widget signal is emitted for each generated menubar and toolbar. It is not emitted for generated popup menus, which can be obtained by gtkUIManagerGetWidget. Since 2.4

merge

a GtkUIManager

widget

the added widget

user.data

user data set when the signal handler was connected.

connect-proxy(uimanager, action, proxy, user.data)

The connect_proxy signal is emitted after connecting a proxy to an action in the group.

This is intended for simple customizations for which a custom action class would be too clumsy, e.g. showing tooltips for menuitems in the statusbar. Since 2.4

uimanager

the ui manager

action

the action

proxy

the proxy

user.data

user data set when the signal handler was connected.

disconnect-proxy(uimanager, action, proxy, user.data)

The disconnect_proxy signal is emitted after disconnecting a proxy from an action in the group. Since 2.4

uimanager

the ui manager

action

the action

proxy

the proxy

user.data

user data set when the signal handler was connected.

post-activate(uimanager, action, user.data)

The post_activate signal is emitted just after the action is activated.

This is intended for applications to get notification just after any action is activated. Since 2.4

uimanager

the ui manager

action

the action

user.data

user data set when the signal handler was connected.

pre-activate(uimanager, action, user.data)

The pre_activate signal is emitted just before the action is activated.

This is intended for applications to get notification just before any action is activated. Since 2.4

uimanager

the ui manager

action

the action

user.data

user data set when the signal handler was connected.

Properties

add-tearoffs [logical : Read / Write]

The "add-tearoffs" property controls whether generated menus have tearoff menu items.

Note that this only affects regular menus. Generated popup menus never have tearoff menu items. Default value: FALSE Since 2.4

ui [character : * : Read]

An XML string describing the merged UI. Default value: "<ui>\n</ui>\n"

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkUIManager.html

See Also

GtkBuilder

[Package RGtk2 version 2.20.31 Index]