GtkVBox {RGtk2}	R Documentation

GtkVBox

Description

A vertical container box

Methods and Functions

gtkVBoxNew(homogeneous = NULL, spacing = NULL, show = TRUE)
gtkVBox(homogeneous = NULL, spacing = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBox
                                 +----GtkVBox
                                       +----GtkColorSelection
                                       +----GtkFileChooserWidget
                                       +----GtkFontSelection
                                       +----GtkGammaCurve
                                       +----GtkRecentChooserWidget

Interfaces

GtkVBox implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

A GtkVBox is a container that organizes child widgets into a single column.

Use the GtkBox packing interface to determine the arrangement, spacing, height, and alignment of GtkVBox children.

All children are allocated the same width.

Structures

GtkVBox

undocumented

Convenient Construction

gtkVBox is the equivalent of gtkVBoxNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkVBox.html

[Package RGtk2 version 2.20.31 Index]