GtkVButtonBox {RGtk2}	R Documentation

GtkVButtonBox

Description

A container for arranging buttons vertically

Methods and Functions

gtkVButtonBoxNew(show = TRUE)
gtkVButtonBoxGetSpacingDefault()
gtkVButtonBoxSetSpacingDefault(spacing)
gtkVButtonBoxGetLayoutDefault()
gtkVButtonBoxSetLayoutDefault(layout)
gtkVButtonBox(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBox
                                 +----GtkButtonBox
                                       +----GtkVButtonBox

Interfaces

GtkVButtonBox implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

A button box should be used to provide a consistent layout of buttons throughout your application. The layout/spacing can be altered by the programmer, or if desired, by the user to alter the 'feel' of a program to a small degree.

A GtkVButtonBox is created with gtkVButtonBoxNew. Buttons are packed into a button box the same way widgets are added to any other container, using gtkContainerAdd. You can also use gtkBoxPackStart or gtkBoxPackEnd, but for button boxes both these functions work just like gtkContainerAdd, ie., they pack the button in a way that depends on the current layout style and on whether the button has had gtkButtonBoxSetChildSecondary called on it.

The spacing between buttons can be set with gtkBoxSetSpacing. The arrangement and layout of the buttons can be changed with gtkButtonBoxSetLayout.

Structures

GtkVButtonBox

GtkVButtonBox does not contain any public fields.

Convenient Construction

gtkVButtonBox is the equivalent of gtkVButtonBoxNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkVButtonBox.html

[Package RGtk2 version 2.20.31 Index]