GtkVPaned {RGtk2}	R Documentation

GtkVPaned

Description

A container with two panes arranged vertically

Methods and Functions

gtkVPanedNew(show = TRUE)
gtkVPaned(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkPaned
                                 +----GtkVPaned

Interfaces

GtkVPaned implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

The VPaned widget is a container widget with two children arranged vertically. The division between the two panes is adjustable by the user by dragging a handle. See GtkPaned for details.

Structures

GtkVPaned

undocumented

Convenient Construction

gtkVPaned is the equivalent of gtkVPanedNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkVPaned.html

[Package RGtk2 version 2.20.31 Index]