GtkVRuler {RGtk2}	R Documentation

GtkVRuler

Description

A vertical ruler

Methods and Functions

gtkVRulerNew(show = TRUE)
gtkVRuler(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkRuler
                           +----GtkVRuler

Interfaces

GtkVRuler implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

PLEASE NOTE: This widget is considered too specialized/little-used for GTK+, and will in the future be moved to some other package. If your application needs this widget, feel free to use it, as the widget does work and is useful in some applications; it's just not of general interest. However, we are not accepting new features for the widget, and it will eventually move out of the GTK+ distribution. The VRuler widget is a widget arranged vertically creating a ruler that is utilized around other widgets such as a text widget. The ruler is used to show the location of the mouse on the window and to show the size of the window in specified units. The available units of measurement are GTK_PIXELS, GTK_INCHES and GTK_CENTIMETERS. GTK_PIXELS is the default. rulers.

Structures

GtkVRuler

The GtkVRuler struct contains private data and should be accessed using the functions below.

Convenient Construction

gtkVRuler is the equivalent of gtkVRulerNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkVRuler.html

[Package RGtk2 version 2.20.31 Index]