GtkVScale {RGtk2}	R Documentation

GtkVScale

Description

A vertical slider widget for selecting a value from a range

Methods and Functions

gtkVScaleNew(adjustment = NULL, show = TRUE)
gtkVScaleNewWithRange(min, max, step, show = TRUE)
gtkVScale(adjustment = NULL, min, max, step, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkRange
                           +----GtkScale
                                 +----GtkVScale

Interfaces

GtkVScale implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

The GtkVScale widget is used to allow the user to select a value using a vertical slider. To create one, use gtkHScaleNewWithRange.

The position to show the current value, and the number of decimal places shown can be set using the parent GtkScale class's functions.

Structures

GtkVScale

The GtkVScale struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkVScale is the result of collapsing the constructors of GtkVScale (gtkVScaleNew, gtkVScaleNewWithRange) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkVScale.html

[Package RGtk2 version 2.20.31 Index]