GtkVScrollbar {RGtk2}	R Documentation

GtkVScrollbar

Description

A vertical scrollbar

Methods and Functions

gtkVScrollbarNew(adjustment = NULL, show = TRUE)
gtkVScrollbar(adjustment = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkRange
                           +----GtkScrollbar
                                 +----GtkVScrollbar

Interfaces

GtkVScrollbar implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

The GtkVScrollbar widget is a widget arranged vertically creating a scrollbar. See GtkScrollbar for details on scrollbars. GtkAdjustment pointers may be added to handle the adjustment of the scrollbar or it may be left NULL in which case one will be created for you. See GtkScrollbar for a description of what the fields in an adjustment represent for a scrollbar.

Structures

GtkVScrollbar

The GtkVScrollbar struct contains private data and should be accessed using the functions below.

Convenient Construction

gtkVScrollbar is the equivalent of gtkVScrollbarNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkVScrollbar.html

[Package RGtk2 version 2.20.31 Index]