GtkVSeparator {RGtk2}	R Documentation

GtkVSeparator

Description

A vertical separator

Methods and Functions

gtkVSeparatorNew(show = TRUE)
gtkVSeparator(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkSeparator
                           +----GtkVSeparator

Interfaces

GtkVSeparator implements AtkImplementorIface, GtkBuildable and GtkOrientable.

Detailed Description

The GtkVSeparator widget is a vertical separator, used to group the widgets within a window. It displays a vertical line with a shadow to make it appear sunken into the interface.

Structures

GtkVSeparator

The GtkVSeparator struct contains private data only, and should be accessed using the functions below.

Convenient Construction

gtkVSeparator is the equivalent of gtkVSeparatorNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkVSeparator.html

[Package RGtk2 version 2.20.31 Index]