GtkViewport {RGtk2}	R Documentation

GtkViewport

Description

An adapter which makes widgets scrollable

Methods and Functions

gtkViewportNew(hadjustment = NULL, vadjustment = NULL, show = TRUE)
gtkViewportGetHadjustment(object)
gtkViewportGetVadjustment(object)
gtkViewportSetHadjustment(object, adjustment = NULL)
gtkViewportSetVadjustment(object, adjustment = NULL)
gtkViewportSetShadowType(object, type)
gtkViewportGetShadowType(object)
gtkViewportGetBinWindow(object)
gtkViewport(hadjustment = NULL, vadjustment = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkViewport

Interfaces

GtkViewport implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkViewport widget acts as an adaptor class, implementing scrollability for child widgets that lack their own scrolling capabilities. Use GtkViewport to scroll child widgets such as GtkTable, GtkBox, and so on.

If a widget has native scrolling abilities, such as GtkTextView, GtkTreeView or GtkIconview, it can be added to a GtkScrolledWindow with gtkContainerAdd. If a widget does not, you must first add the widget to a GtkViewport, then add the viewport to the scrolled window. The convenience function gtkScrolledWindowAddWithViewport does exactly this, so you can ignore the presence of the viewport.

Structures

GtkViewport

undocumented

Convenient Construction

gtkViewport is the equivalent of gtkViewportNew.

Signals

set-scroll-adjustments(horizontal, vertical, user.data)

Set the scroll adjustments for the viewport. Usually scrolled containers like GtkScrolledWindow will emit this signal to connect two instances of GtkScrollbar to the scroll directions of the GtkViewport.

horizontal

the horizontal GtkAdjustment

vertical

the vertical GtkAdjustment

user.data

user data set when the signal handler was connected.

Properties

hadjustment [GtkAdjustment : * : Read / Write / Construct]

The GtkAdjustment that determines the values of the horizontal position for this viewport.

shadow-type [GtkShadowType : Read / Write]

Determines how the shadowed box around the viewport is drawn. Default value: GTK_SHADOW_IN

vadjustment [GtkAdjustment : * : Read / Write / Construct]

The GtkAdjustment that determines the values of the vertical position for this viewport.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkViewport.html

See Also

GtkScrolledWindow GtkAdjustment

[Package RGtk2 version 2.20.31 Index]