GtkVolumeButton {RGtk2}	R Documentation

GtkVolumeButton

Description

A button which pops up a volume control

Methods and Functions

gtkVolumeButtonNew(show = TRUE)
gtkVolumeButton(show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkButton
                                       +----GtkScaleButton
                                             +----GtkVolumeButton

Interfaces

GtkVolumeButton implements AtkImplementorIface, GtkBuildable, GtkActivatable and GtkOrientable.

Detailed Description

GtkVolumeButton is a subclass of GtkScaleButton that has been tailored for use as a volume control widget with suitable icons, tooltips and accessible labels.

Structures

GtkVolumeButton

undocumented

Convenient Construction

gtkVolumeButton is the equivalent of gtkVolumeButtonNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkVolumeButton.html

[Package RGtk2 version 2.20.31 Index]