GtkWindow {RGtk2}	R Documentation

GtkWindow

Description

Toplevel which can contain other widgets

Methods and Functions

gtkWindowNew(type = NULL, show = TRUE)
gtkWindowSetTitle(object, title)
gtkWindowSetWmclass(object, wmclass.name, wmclass.class)
gtkWindowSetPolicy(object, allow.shrink, allow.grow, auto.shrink)
gtkWindowSetResizable(object, resizable)
gtkWindowGetResizable(object)
gtkWindowAddAccelGroup(object, accel.group)
gtkWindowRemoveAccelGroup(object, accel.group)
gtkWindowActivateFocus(object)
gtkWindowActivateDefault(object)
gtkWindowSetModal(object, modal)
gtkWindowSetDefaultSize(object, width, height)
gtkWindowSetGeometryHints(object, geometry.widget, geometry)
gtkWindowSetGravity(object, gravity)
gtkWindowGetGravity(object)
gtkWindowSetPosition(object, position)
gtkWindowSetTransientFor(object, parent = NULL)
gtkWindowSetDestroyWithParent(object, setting)
gtkWindowSetScreen(object, screen)
gtkWindowGetScreen(object)
gtkWindowIsActive(object)
gtkWindowHasToplevelFocus(object)
gtkWindowListToplevels()
gtkWindowAddMnemonic(object, keyval, target)
gtkWindowRemoveMnemonic(object, keyval, target)
gtkWindowMnemonicActivate(object, keyval, modifier)
gtkWindowActivateKey(object, event)
gtkWindowPropagateKeyEvent(object, event)
gtkWindowGetFocus(object)
gtkWindowSetFocus(object, focus = NULL)
gtkWindowGetDefaultWidget(object)
gtkWindowSetDefault(object, default.widget = NULL)
gtkWindowPresent(object)
gtkWindowPresentWithTime(object, timestamp)
gtkWindowIconify(object)
gtkWindowDeiconify(object)
gtkWindowStick(object)
gtkWindowUnstick(object)
gtkWindowMaximize(object)
gtkWindowUnmaximize(object)
gtkWindowFullscreen(object)
gtkWindowUnfullscreen(object)
gtkWindowSetKeepAbove(object, setting)
gtkWindowSetKeepBelow(object, setting)
gtkWindowBeginResizeDrag(object, edge, button, root.x, root.y, timestamp)
gtkWindowBeginMoveDrag(object, button, root.x, root.y, timestamp)
gtkWindowSetDecorated(object, setting)
gtkWindowSetDeletable(object, setting)
gtkWindowSetFrameDimensions(object, left, top, right, bottom)
gtkWindowSetHasFrame(object, setting)
gtkWindowSetMnemonicModifier(object, modifier)
gtkWindowSetTypeHint(object, hint)
gtkWindowSetSkipTaskbarHint(object, setting)
gtkWindowSetSkipPagerHint(object, setting)
gtkWindowSetUrgencyHint(object, setting)
gtkWindowSetAcceptFocus(object, setting)
gtkWindowSetFocusOnMap(object, setting)
gtkWindowSetStartupId(object, startup.id)
gtkWindowSetRole(object, role)
gtkWindowGetDecorated(object)
gtkWindowGetDeletable(object)
gtkWindowGetDefaultIconList()
gtkWindowGetDefaultIconName()
gtkWindowGetDefaultSize(object)
gtkWindowGetDestroyWithParent(object)
gtkWindowGetFrameDimensions(object)
gtkWindowGetHasFrame(object)
gtkWindowGetIcon(object)
gtkWindowGetIconList(object)
gtkWindowGetIconName(object)
gtkWindowGetMnemonicModifier(object)
gtkWindowGetModal(object)
gtkWindowGetPosition(object)
gtkWindowGetRole(object)
gtkWindowGetSize(object)
gtkWindowGetTitle(object)
gtkWindowGetTransientFor(object)
gtkWindowGetTypeHint(object)
gtkWindowGetSkipTaskbarHint(object)
gtkWindowGetSkipPagerHint(object)
gtkWindowGetUrgencyHint(object)
gtkWindowGetAcceptFocus(object)
gtkWindowGetFocusOnMap(object)
gtkWindowGetGroup(object)
gtkWindowGetWindowType(object)
gtkWindowMove(object, x, y)
gtkWindowReshowWithInitialSize(object)
gtkWindowResize(object, width, height)
gtkWindowSetDefaultIconList(list)
gtkWindowSetDefaultIcon(icon)
gtkWindowSetDefaultIconName(name)
gtkWindowSetIcon(object, icon = NULL)
gtkWindowSetIconList(object, list)
gtkWindowSetIconName(object, name = NULL)
gtkWindowSetAutoStartupNotification(setting)
gtkWindowGetOpacity(object)
gtkWindowSetOpacity(object, opacity)
gtkWindowGetMnemonicsVisible(object)
gtkWindowSetMnemonicsVisible(object, setting)
gtkWindow(type = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkWindow
                                       +----GtkDialog
                                       +----GtkAssistant
                                       +----GtkOffscreenWindow
                                       +----GtkPlug

Interfaces

GtkWindow implements AtkImplementorIface and GtkBuildable.

GtkWindow as GtkBuildable

The GtkWindow implementation of the GtkBuildable interface supports a custom <accel-groups> element, which supports any number of <group> elements representing the GtkAccelGroup objects you want to add to your window (synonymous with gtkWindowAddAccelGroup.

A UI definition fragment with accel groups

<object class="GtkWindow">
  <accel-groups>
    <group name="accelgroup1"/>
  </accel-groups>
</object>
...
<object class="GtkAccelGroup" id="accelgroup1"/>

Structures

GtkWindow

undocumented

title

[character]

wmclassName

[character]

wmclassClass

[character]

wmRole

[character]

focusWidget

[GtkWidget]

defaultWidget

[GtkWidget]

transientParent

[GtkWindow]

frame

[GdkWindow]

group

[GtkWindowGroup]

configureRequestCount

[integer]

allowShrink

[numeric]

allowGrow

[numeric]

configureNotifyReceived

[numeric]

needDefaultPosition

[numeric]

needDefaultSize

[numeric]

position

[numeric]

type

[numeric]

hasUserRefCount

[numeric]

hasFocus

[numeric]

modal

[numeric]

destroyWithParent

[numeric]

hasFrame

[numeric]

iconifyInitially

[numeric]

stickInitially

[numeric]

maximizeInitially

[numeric]

decorated

[numeric]

typeHint

[numeric]

gravity

[numeric]

frameLeft

[numeric]

frameTop

[numeric]

frameRight

[numeric]

frameBottom

[numeric]

keysChangedHandler

[numeric]

mnemonicModifier

[GdkModifierType]

Convenient Construction

gtkWindow is the equivalent of gtkWindowNew.

Signals

activate-default(window, user.data)

The ::activate-default signal is a keybinding signal which gets emitted when the user activates the default widget of window.

window

the window which received the signal

user.data

user data set when the signal handler was connected.

activate-focus(window, user.data)

The ::activate-default signal is a keybinding signal which gets emitted when the user activates the currently focused widget of window.

window

the window which received the signal

user.data

user data set when the signal handler was connected.

frame-event(window, user.data)

undocumented

window

the object which received the signal.

user.data

user data set when the signal handler was connected.

keys-changed(window, user.data)

The ::keys-changed signal gets emitted when the set of accelerators or mnemonics that are associated with window changes.

window

the window which received the signal

user.data

user data set when the signal handler was connected.

set-focus(window, user.data)

undocumented

window

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

accept-focus [logical : Read / Write]

Whether the window should receive the input focus. Default value: TRUE Since 2.4

allow-grow [logical : Read / Write]

If TRUE, users can expand the window beyond its minimum size. Default value: TRUE

allow-shrink [logical : Read / Write]

If TRUE, the window has no mimimum size. Setting this to TRUE is 99% of the time a bad idea. Default value: FALSE

decorated [logical : Read / Write]

Whether the window should be decorated by the window manager. Default value: TRUE Since 2.4

default-height [integer : Read / Write]

The default height of the window, used when initially showing the window. Allowed values: >= -1 Default value: -1

default-width [integer : Read / Write]

The default width of the window, used when initially showing the window. Allowed values: >= -1 Default value: -1

deletable [logical : Read / Write]

Whether the window frame should have a close button. Default value: TRUE Since 2.10

destroy-with-parent [logical : Read / Write]

If this window should be destroyed when the parent is destroyed. Default value: FALSE

focus-on-map [logical : Read / Write]

Whether the window should receive the input focus when mapped. Default value: TRUE Since 2.6

gravity [GdkGravity : Read / Write]

The window gravity of the window. See gtkWindowMove and GdkGravity for more details about window gravity. Default value: GDK_GRAVITY_NORTH_WEST Since 2.4

has-toplevel-focus [logical : Read]

Whether the input focus is within this GtkWindow. Default value: FALSE

icon [GdkPixbuf : * : Read / Write]

Icon for this window.

icon-name [character : * : Read / Write]

The :icon-name property specifies the name of the themed icon to use as the window icon. See GtkIconTheme for more details. Default value: NULL Since 2.6

is-active [logical : Read]

Whether the toplevel is the current active window. Default value: FALSE

mnemonics-visible [logical : Read / Write]

Whether mnemonics are currently visible in this window. Default value: TRUE

modal [logical : Read / Write]

If TRUE, the window is modal (other windows are not usable while this one is up). Default value: FALSE

opacity [numeric : Read / Write]

The requested opacity of the window. See gtkWindowSetOpacity for more details about window opacity. Allowed values: [0,1] Default value: 1 Since 2.12

resizable [logical : Read / Write]

If TRUE, users can resize the window. Default value: TRUE

role [character : * : Read / Write]

Unique identifier for the window to be used when restoring a session. Default value: NULL

screen [GdkScreen : * : Read / Write]

The screen where this window will be displayed.

skip-pager-hint [logical : Read / Write]

TRUE if the window should not be in the pager. Default value: FALSE

skip-taskbar-hint [logical : Read / Write]

TRUE if the window should not be in the task bar. Default value: FALSE

startup-id [character : * : Write]

The :startup-id is a write-only property for setting window's startup notification identifier. See gtkWindowSetStartupId for more details. Default value: NULL Since 2.12

title [character : * : Read / Write]

The title of the window. Default value: NULL

transient-for [GtkWindow : * : Read / Write / Construct]

The transient parent of the window. See gtkWindowSetTransientFor for more details about transient windows. Since 2.10

type [GtkWindowType : Read / Write / Construct Only]

The type of the window. Default value: GTK_WINDOW_TOPLEVEL

type-hint [GdkWindowTypeHint : Read / Write]

Hint to help the desktop environment understand what kind of window this is and how to treat it. Default value: GDK_WINDOW_TYPE_HINT_NORMAL

urgency-hint [logical : Read / Write]

TRUE if the window should be brought to the user's attention. Default value: FALSE

window-position [GtkWindowPosition : Read / Write]

The initial position of the window. Default value: GTK_WIN_POS_NONE

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkWindow.html

[Package RGtk2 version 2.20.31 Index]