Pango {RGtk2}	R Documentation

Pango

Description

Pango is a library for internationalized text handling. It centers around the PangoLayout object, representing a paragraph of text. Pango provides the engine for GtkTextView, GtkLabel, GtkEntry, and other widgets that display text.

Details

The RGtk binding to the Pango library consists of the following components:

pango-Bidirectional-Text

Types and functions to help with handling bidirectional text

pango-Coverage-Maps

Unicode character range coverage storage

pango-Fonts

Structures representing abstract fonts

pango-Glyph-Storage

Structures for storing information about glyphs

pango-Layout-Objects

High-level layout driver objects

pango-Text-Processing

Functions to run the rendering pipeline

PangoRenderer

Rendering driver base class

pango-Version-Checking

Tools for checking Pango version at compile- and run-time.

pango-Cairo-Rendering

Rendering with the Cairo backend

pango-Scripts-and-Languages

Identifying writing systems and languages

pango-Tab-Stops

Structures for storing tab stops

pango-Text-Attributes

Font and other attributes for annotating text

pango-Vertical-Text

Laying text out in vertical directions

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//pango

[Package RGtk2 version 2.20.31 Index]