PangoRenderer {RGtk2}	R Documentation

PangoRenderer

Description

Rendering driver base class

Methods and Functions

pangoRendererDrawLayout(object, layout, x, y)
pangoRendererDrawLayoutLine(object, line, x, y)
pangoRendererDrawGlyphs(object, font, glyphs, x, y)
pangoRendererDrawGlyphItem(object, text, glyph.item, x, y)
pangoRendererDrawRectangle(object, part, x, y, width, height)
pangoRendererDrawErrorUnderline(object, x, y, width, height)
pangoRendererDrawTrapezoid(object, part, y1., x11, x21, y2, x12, x22)
pangoRendererDrawGlyph(object, font, glyph, x, y)
pangoRendererActivate(object)
pangoRendererDeactivate(object)
pangoRendererPartChanged(object, part)
pangoRendererSetColor(object, part, color)
pangoRendererGetColor(object, part)
pangoRendererSetMatrix(object, matrix)
pangoRendererGetMatrix(object)
pangoRendererGetLayout(renderer)
pangoRendererGetLayoutLine(renderer)

Hierarchy

GObject
   +----PangoRenderer
         +----PangoXftRenderer

Detailed Description

PangoRenderer is a base class that contains the necessary logic for rendering a PangoLayout or PangoLayoutLine. By subclassing PangoRenderer and overriding operations such as draw.glyphs and draw.rectangle, renderers for particular font backends and destinations can be created.

Structures

PangoRenderer

PangoRenderer is a base class for objects that are used to render Pango objects such as PangoGlyphString and PangoLayout. Since 1.8

matrix

[PangoMatrix] the current transformation matrix for the Renderer; may be NULL, which should be treated the same as the identity matrix.

Enums and Flags

PangoRenderPart

PangoRenderPart defines different items to render for such purposes as setting colors. Since 1.8

foreground

the text itself

background

the area behind the text

underline

underlines

strikethrough

strikethrough lines

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//pango/PangoRenderer.html

[Package RGtk2 version 2.20.31 Index]