RGtk {RGtk2}	R Documentation

The RGtk2 package

Description

RGtk2 provides a set of bindings between R and the GTK+ library and several of its dependent libraries. It allows the user to construct full-featured GUI's completely from within R.

Details

RGtk2 binds to the following libraries:

ATK

ATK is the Accessibility Toolkit. It provides a set of generic interfaces allowing accessibility technologies to interact with a graphical user interface. For example, a screen reader uses ATK to discover the text in an interface and read it to blind users. GTK+ widgets have built-in support for accessibility using the ATK framework.

Pango

Pango is a library for internationalized text handling. It centers around the PangoLayout object, representing a paragraph of text. Pango provides the engine for GtkTextView, GtkLabel, GtkEntry, and other widgets that display text.

GDK

GDK is the abstraction layer that allows GTK+ to support multiple windowing systems. GDK provides drawing and window system facilities on X11, Windows, and the Linux framebuffer device.

GTK

The GTK+ library itself contains widgets, that is, GUI components such as GtkButton or GtkTextView.

GDK-Pixbuf

This is a small library which allows you to create GdkPixbuf ('pixel buffer') objects from image data or image files. Use a GdkPixbuf in combination with GtkImage to display images.

Cairo

Cairo is a 2D graphics library with support for multiple output devices. Currently supported output targets include the X Window System, win32, and image buffers.

RGtk2 also partially binds some lower-level libraries in order to support the bindings to the others. These include GObject and GMainLoop.

R objects passed between the user and RGtk2 are either primitive types (character, logical, etc) or external objects (externalptr). All R objects wrapping external objects extend the RGtkObject class.

Note

As described above, RGtk2 binds many libraries beyond GTK+ itself. Thus, it can serve many purposes besides GUI construction. For example, GDKPixbuf and Cairo allow the R user to produce arbitary high-quality graphics.

Author(s)

Michael Lawrence, with excerpts from library documentation

References

Michael Lawrence, Duncan Temple Lang (2010). RGtk2: A Graphical User Interface Toolkit for R. Journal of Statistical Software, 37(8), 1-52. http://www.jstatsoft.org/v37/i08/.

[Package RGtk2 version 2.20.31 Index]