RGtkDataFrame {RGtk2}	R Documentation

The RGtkDataFrame model

Description

A GtkTreeModel implementation backed by an R data frame

Usage

rGtkDataFrame(frame = data.frame())
rGtkDataFrameNew(frame = data.frame())
rGtkDataFrameAppendColumns(x, ...)
rGtkDataFrameAppendRows(x, ...)
## S3 method for class 'RGtkDataFrame'
as.data.frame(x, ...)
rGtkDataFrameSetFrame(x, frame = data.frame())
## S3 method for class 'RGtkDataFrame'
x[i, j, drop = T]
## S3 replacement method for class 'RGtkDataFrame'
x[i, j] <- value
## S3 method for class 'RGtkDataFrame'
dim(x, ...)
## S3 method for class 'RGtkDataFrame'
dimnames(x, ...)
## S3 replacement method for class 'RGtkDataFrame'
dimnames(x) <- value

Arguments

frame
The frame to use as the backing store of the model

x
An RGtkDataFrame object

i
Row index

j
Column index

value
An R object similar to that accepted by [<-.data.frame or the dimnames for the data frame

drop
Whether to 'drop' the result to the simplest structure

...
Items to append as columns or rows or just additional arguments

Details

The RGtk2 interface carries a lot of overhead, slowing down operations that require large numbers of function calls, such as loading a GtkTreeModel. Under the assumption that R programmers will store large datasets as data frames, a new GtkTreeModel was implemented that draws data directly from an R data frame. This offers not only a dramatic performance gain but also allows efficient addition of columns to a model, which the default GTK implementations do not allow.

The RGtkDataFrame is constructed with a delegate data frame, which can be empty, via either rGtkDataFrameNew or rGtkDataFrame for short. The subset and replacement methods work much the same as for normal data frames, except one should note that removing columns (ie by replacing columns with NULLs) is not supported. Note that even if the initial data frame is empty, one should ensure that the empty vectors representing the column are of the desired types. If one wants to simply replace the backing frame with a new one, then there are two options: create a new RGtkDataFrame and connect it to the views of the old model, or use rGtkDataFrameSetFrame.

The rGtkDataFrameAppendColumns and rGtkDataFrameAppendRows methods allow appending columns and rows, respectively. Note that these are a lot shorter if using the object$appendColumns(...) syntax.

The as.data.frame method retrieves the backing data frame from the model, so that one can perform any data frame operation on the data. Of course, any changes are not propagated back to the model, so it may take some work to efficiently merge any changes, if necessary.

For convenience, one can access the dimensions and dimension names using dim.RGtkDataframe and dimnames.RGtkDataFrame, respectively. It is possible to set the dimension names using the conventional replacement function. Note that rownames mean nothing to GTK.

Value

The constructors return instances of RGtkDataFrame. as.data.frame.RGtkDataFrame returns the data frame backing the model. [.RGtkDataFrame returns the result of the [ method on the backing frame.

Note

It is not yet clear how to encode a tree structure with a data frame, so this is only currently useful for flat tables.

Author(s)

Michael Lawrence

[Package RGtk2 version 2.20.31 Index]