argInfo {RGtk2}	R Documentation

Access information about properties of a GtkObject

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

These allow one to discover what “properties” or variables a Gtk object supports and the details of each property such as its type and whether it can be set or only read.

Usage

## S3 method for class 'GtkObject'
names(x)
gtkObjectGetArgInfo(obj, parents=TRUE, collapse=TRUE)

Arguments

x
the Gtk object the names of whose properties are to be queried

obj
the Gtk object for which the property information is to be retrieved

parents
a logical value indicating whether to get the properties of the parent classes as well as the specific class of the Gtk object obj. If this is true, we iterate over the successive super-classes of obj.

collapse
a logical value indicating whether the information about parent classes should be collapsed into a single list (TRUE) or maintained as separate elements in a top-level list that allows one to determine to which class each property originates. The names of a property contains the class for which it is defined.

Details

A regular user will probably want to call gtkObjectGetArgInfo with parents = TRUE, collapse=TRUE to find all the available properties in the most convenient format. Tool writers using the reflectance may want to get the associated classes and process the properties hierarchically.

Value

names returns a character vector giving the names of the available properties with the class prefix removed. gtkObjectGetArgInfo returns a list giving details of the properties. At the lowest level, each property is described by a list containing

type
an object of class GtkType which identifies the type (!) of the value of the property.

flag
an object which specifies whether the property is read-only or read-write and when it can be specified, etc.

Each property is indexed by its name in the list.

The precise form of the returned value depends on the different arguments. If parents is FALSE, then a simple list giving each property for the class of obj is returned. If parents is TRUE and collapse is FALSE, then a list indexed by class is returned. Each element is a list containing the property information.

Note

THIS STUFF IS VERY OLD AND DEPRECATED (compatibility wrappers for RGtk 1)

Author(s)

Duncan Temple Lang <duncan@research.bell-labs.com>

[Package RGtk2 version 2.20.31 Index]