assertions {RGtk2}	R Documentation

RGtk2 Type Assertion

Description

Assert that an object is of a particular type

Usage

checkPtrType(w, klass = "GtkWidget", nullOk = FALSE, critical = TRUE)
implements(obj, interface)

Arguments

w
An object whose type is to be verified.

klass
The type the object is expected to be.

nullOk
Whether the object is allowed to be NULL.

critical
Whether to stop if the object is not of the specified type. If this is a character vector, then the function will stop on mismatch and report that string as the error message.

obj
A GObject.

interface
The interface that obj is expected to implement.

Details

All RGtk2 functions check that the arguments are of the correct type, if possible. The checkPtrType function is most useful to the user when it is not known if an object is of the required type. A good example is the user data argument of a callback function. To see if a GObject implements a certain interface, use implements.

Author(s)

Michael Lawrence and Duncan Temple Lang

[Package RGtk2 version 2.20.31 Index]