atkActionGetKeybinding {RGtk2}	R Documentation

atkActionGetKeybinding

Description

Returns a keybinding associated with this action, if one exists. The returned string is in the format "<a>;<b>;<c>" (i.e. semicolon-delimited), where <a> is the keybinding which activates the object if it is presently enabled onscreen, <b> corresponds to the keybinding or sequence of keys which invokes the action even if the relevant element is not currently posted on screen (for instance, for a menu item it posts the parent menus before invoking). The last token in the above string, if non-empty, represents a keyboard shortcut which invokes the same action without posting the component or its enclosing menus or dialogs.

Usage

atkActionGetKeybinding(object, i)

Arguments

object
[AtkAction] a GObject instance that implements AtkActionIface

i
[integer] the action index corresponding to the action to be performed

Value

[character] a string representing the available keybindings, or NULL if there is no keybinding for this action.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]