atkActionGetNActions {RGtk2}	R Documentation

atkActionGetNActions

Description

Gets the number of accessible actions available on the object. If there are more than one, the first one is considered the "default" action of the object.

Usage

atkActionGetNActions(object)

Arguments

object
[AtkAction] a GObject instance that implements AtkActionIface

Value

[integer] a the number of actions, or 0 if action does not implement this interface.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]