atkActionGetName {RGtk2}	R Documentation

atkActionGetName

Description

Returns a non-localized string naming the specified action of the object. This name is generally not descriptive of the end result of the action, but instead names the 'interaction type' which the object supports. By convention, the above strings should be used to represent the actions which correspond to the common point-and-click interaction techniques of the same name: i.e. "click", "press", "release", "drag", "drop", "popup", etc. The "popup" action should be used to pop up a context menu for the object, if one exists.

Usage

atkActionGetName(object, i)

Arguments

object
[AtkAction] a GObject instance that implements AtkActionIface

i
[integer] the action index corresponding to the action to be performed

Details

For technical reasons, some toolkits cannot guarantee that the reported action is actually 'bound' to a nontrivial user event; i.e. the result of some actions via atkActionDoAction may be NIL.

Value

[character] a name string, or NULL if action does not implement this interface.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]