atkActionSetDescription {RGtk2}	R Documentation

atkActionSetDescription

Description

Sets a description of the specified action of the object.

Usage

atkActionSetDescription(object, i, desc)

Arguments

object
[AtkAction] a GObject instance that implements AtkActionIface

i
[integer] the action index corresponding to the action to be performed

desc
[character] the description to be assigned to this action

Value

[logical] a gboolean representing if the description was successfully set;

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]