atkAddGlobalEventListener {RGtk2}	R Documentation

atkAddGlobalEventListener

Description

Adds the specified function to the list of functions to be called when an event of type event_type occurs.

Usage

atkAddGlobalEventListener(listener, event.type)

Arguments

listener
[GSignalEmissionHook] the listener to notify

event.type
[character] the type of event for which notification is requested

Value

[numeric] added event listener id, or 0 on failure.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]