atkAddKeyEventListener {RGtk2}	R Documentation

atkAddKeyEventListener

Description

Adds the specified function to the list of functions to be called when a key event occurs. The data element will be passed to the AtkKeySnoopFunc (listener) as the func.data param, on notification.

Usage

atkAddKeyEventListener(listener, data)

Arguments

listener
[AtkKeySnoopFunc] the listener to notify

data
[R object] a R object that points to a block of data that should be sent to the registered listeners, along with the event notification, when it occurs.

Value

[numeric] added event listener id, or 0 on failure.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]