atkComponentAddFocusHandler {RGtk2}	R Documentation

atkComponentAddFocusHandler

Description

Add the specified handler to the set of functions to be called when this object receives focus events (in or out). If the handler is already added it is not added again

Usage

atkComponentAddFocusHandler(object, handler)

Arguments

object
[AtkComponent] The AtkComponent to attach the handler to

handler
[AtkFocusHandler] The AtkFocusHandler to be attached to component

Value

[numeric] a handler id which can be used in atk_component_remove_focus_handler or zero if the handler was already added.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]