atkComponentGetExtents {RGtk2}	R Documentation

atkComponentGetExtents

Description

Gets the rectangle which gives the extent of the component.

Usage

atkComponentGetExtents(object, coord.type)

Arguments

object
[AtkComponent] an AtkComponent

coord.type
[AtkCoordType] specifies whether the coordinates are relative to the screen or to the components top level window

Value

A list containing the following elements:

x
[integer] integer to put x coordinate

y
[integer] integer to put y coordinate

width
[integer] integer to put width

height
[integer] integer to put height

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]