atkComponentGetMdiZorder {RGtk2}	R Documentation

atkComponentGetMdiZorder

Description

Gets the zorder of the component. The value G_MININT will be returned if the layer of the component is not ATK_LAYER_MDI or ATK_LAYER_WINDOW.

Usage

atkComponentGetMdiZorder(object)

Arguments

object
[AtkComponent] an AtkComponent

Value

[integer] a gint which is the zorder of the component, i.e. the depth at which the component is shown in relation to other components in the same container.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]