atkComponentRefAccessibleAtPoint {RGtk2}	R Documentation

atkComponentRefAccessibleAtPoint

Description

Gets a reference to the accessible child, if one exists, at the coordinate point specified by x and y.

Usage

atkComponentRefAccessibleAtPoint(object, x, y, coord.type)

Arguments

object
[AtkComponent] the AtkComponent

x
[integer] x coordinate

y
[integer] y coordinate

coord.type
[AtkCoordType] specifies whether the coordinates are relative to the screen or to the components top level window

Value

[AtkObject] a reference to the accessible child, if one exists

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]