atkComponentSetPosition {RGtk2}	R Documentation

atkComponentSetPosition

Description

Sets the postition of component.

Usage

atkComponentSetPosition(object, x, y, coord.type)

Arguments

object
[AtkComponent] an AtkComponent

x
[integer] x coordinate

y
[integer] y coordinate

coord.type
[AtkCoordType] specifies whether the coordinates are relative to the screen or to the components top level window

Value

[logical] TRUE or FALSE whether or not the position was set or not

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]