atkDocumentGetAttributeValue {RGtk2}	R Documentation

atkDocumentGetAttributeValue

Description

undocumented

Usage

atkDocumentGetAttributeValue(object, attribute.name)

Arguments

object
[AtkDocument] a GObject instance that implements AtkDocumentIface

attribute.name
[character] a character string representing the name of the attribute whose value is being queried.

Details

Since 1.12

Value

[character] a string value associated with the named attribute for this document, or NULL if a value for attribute_name has not been specified for this document.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]