atkDocumentGetLocale {RGtk2}	R Documentation

atkDocumentGetLocale

Description

Gets a UTF-8 string indicating the POSIX-style LC_MESSAGES locale of the content of this document instance. Individual text substrings or images within this document may have a different locale, see atk_text_get_attributes and atk_image_get_image_locale.

Usage

atkDocumentGetLocale(object)

Arguments

object
[AtkDocument] a GObject instance that implements AtkDocumentIface

Value

[character] a UTF-8 string indicating the POSIX-style LC_MESSAGES locale of the document content as a whole, or NULL if the document content does not specify a locale.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]