atkDocumentSetAttributeValue {RGtk2}	R Documentation

atkDocumentSetAttributeValue

Description

undocumented

Usage

atkDocumentSetAttributeValue(object, attribute.name, attribute.value)

Arguments

object
[AtkDocument] a GObject instance that implements AtkDocumentIface

attribute.name
[character] a character string representing the name of the attribute whose value is being set.

attribute.value
[character] a string value to be associated with attribute_name.

Details

Since 1.12

Value

[logical] TRUE if value is successfully associated with attribute_name for this document, FALSE otherwise (e.g. if the document does not allow the attribute to be modified).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]