atkEditableTextInsertText {RGtk2}	R Documentation

atkEditableTextInsertText

Description

Insert text at a given position.

Usage

atkEditableTextInsertText(object, string, position)

Arguments

object
[AtkEditableText] an AtkEditableText

string
[character] the text to insert

position
[integer] The caller initializes this to the position at which to insert the text. After the call it points at the position after the newly inserted text.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]