atkEditableTextSetRunAttributes {RGtk2}	R Documentation

atkEditableTextSetRunAttributes

Description

Sets the attributes for a specified range. See the ATK_ATTRIBUTE functions (such as ATK_ATTRIBUTE_LEFT_MARGIN) for examples of attributes that can be set. Note that other attributes that do not have corresponding ATK_ATTRIBUTE functions may also be set for certain text widgets.

Usage

atkEditableTextSetRunAttributes(object, attrib.set, start.offset, 
    end.offset)

Arguments

object
[AtkEditableText] an AtkEditableText

attrib.set
[AtkAttributeSet] an AtkAttributeSet

start.offset
[integer] start of range in which to set attributes

end.offset
[integer] end of range in which to set attributes

Value

[logical] TRUE if attributes successfully set for the specified range, otherwise FALSE

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]