atkGetDefaultRegistry {RGtk2}	R Documentation

atkGetDefaultRegistry

Description

Gets a default implementation of the AtkObjectFactory/type registry. Note: For most toolkit maintainers, this will be the correct registry for registering new AtkObject factories. Following a call to this function, maintainers may call atkRegistrySetFactoryType to associate an AtkObjectFactory subclass with the GType of objects for whom accessibility information will be provided.

Usage

atkGetDefaultRegistry()

Value

[AtkRegistry] a default implementation of the AtkObjectFactory/type registry

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]