atkHyperlinkIsInline {RGtk2}	R Documentation

atkHyperlinkIsInline

Description

Indicates whether the link currently displays some or all of its content inline. Ordinary HTML links will usually return FALSE, but an inline &lt;src&gt; HTML element will return TRUE. a *

Usage

atkHyperlinkIsInline(object)

Arguments

object
[AtkHyperlink] an AtkHyperlink

Value

[logical] whether or not this link displays its content inline.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]