atkHyperlinkIsSelectedLink {RGtk2}	R Documentation

atkHyperlinkIsSelectedLink

Description

Determines whether this AtkHyperlink is selected WARNING: atk_hyperlink_is_selected_link is deprecated and should not be used in newly-written code. Please use ATK_STATE_SELECTED to indicate when a hyperlink within a Hypertext container is selected.

Usage

atkHyperlinkIsSelectedLink(object)

Arguments

object
[AtkHyperlink] an AtkHyperlink

Details

Since 1.4

Value

[logical] True is the AtkHyperlink is selected, False otherwise

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]