atkHypertextGetLinkIndex {RGtk2}	R Documentation

atkHypertextGetLinkIndex

Description

Gets the index into the list of hyperlinks that is associated with the character specified by char.index.

Usage

atkHypertextGetLinkIndex(object, char.index)

Arguments

object
[AtkHypertext] an AtkHypertext

char.index
[integer] a character index

Value

[integer] an index into the list of hyperlinks in hypertext, or -1 if there is no hyperlink associated with this character.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]