atkImageGetImageLocale {RGtk2}	R Documentation

atkImageGetImageLocale

Description

Since ATK 1.12

Usage

atkImageGetImageLocale(object)

Arguments

object
[AtkImage] An AtkImage

Value

[character] a string corresponding to the POSIX LC_MESSAGES locale used by the image description, or NULL if the image does not specify a locale.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]