atkImageGetImagePosition {RGtk2}	R Documentation

atkImageGetImagePosition

Description

Gets the position of the image in the form of a point specifying the images top-left corner.

Usage

atkImageGetImagePosition(object, coord.type)

Arguments

object
[AtkImage] a GObject instance that implements AtkImageIface

coord.type
[AtkCoordType] specifies whether the coordinates are relative to the screen or to the components top level window

Value

A list containing the following elements:

x
[integer] integer to put x coordinate position; otherwise, -1 if value cannot be obtained.

y
[integer] integer to put y coordinate position; otherwise, -1 if value cannot be obtained.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]