atkImageGetImageSize {RGtk2}	R Documentation

atkImageGetImageSize

Description

Get the width and height in pixels for the specified image. The values of width and height are returned as -1 if the values cannot be obtained (for instance, if the object is not onscreen).

Usage

atkImageGetImageSize(object)

Arguments

object
[AtkImage] a GObject instance that implements AtkImageIface

Value

A list containing the following elements:

width
[integer] filled with the image width, or -1 if the value cannot be obtained.

height
[integer] filled with the image height, or -1 if the value cannot be obtained.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]