atkObjectGetAttributes {RGtk2}	R Documentation

atkObjectGetAttributes

Description

Get a list of properties applied to this object as a whole, as an AtkAttributeSet consisting of name-value pairs. As such these attributes may be considered weakly-typed properties or annotations, as distinct from strongly-typed object data available via other get/set methods. Not all objects have explicit "name-value pair" AtkAttributeSet properties.

Usage

atkObjectGetAttributes(object)

Arguments

object
[AtkObject] An AtkObject.

Details

Since 1.12

Value

[AtkAttributeSet] an AtkAttributeSet consisting of all explicit properties/annotations applied to the object, or an empty set if the object has no name-value pair attributes assigned to it.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]