atkObjectGetIndexInParent {RGtk2}	R Documentation

atkObjectGetIndexInParent

Description

Gets the 0-based index of this accessible in its parent; returns -1 if the accessible does not have an accessible parent.

Usage

atkObjectGetIndexInParent(object)

Arguments

object
[AtkObject] an AtkObject

Value

[integer] an integer which is the index of the accessible in its parent

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]