atkObjectRefAccessibleChild {RGtk2}	R Documentation

atkObjectRefAccessibleChild

Description

Gets a reference to the specified accessible child of the object. The accessible children are 0-based so the first accessible child is at index 0, the second at index 1 and so on.

Usage

atkObjectRefAccessibleChild(object, i)

Arguments

object
[AtkObject] an AtkObject

i
[integer] a gint representing the position of the child, starting from 0

Value

[AtkObject] an AtkObject representing the specified accessible child of the accessible.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]