atkObjectRemoveRelationship {RGtk2}	R Documentation

atkObjectRemoveRelationship

Description

Removes a relationship of the specified type with the specified target.

Usage

atkObjectRemoveRelationship(object, relationship, target)

Arguments

object
[AtkObject] The AtkObject from which an AtkRelation is to be removed.

relationship
[AtkRelationType] The AtkRelationType of the relation

target
[AtkObject] The AtkObject which is the target of the relation to be removed.

Value

[logical] TRUE if the relationship is removed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]