atkRelationSetAdd {RGtk2}	R Documentation

atkRelationSetAdd

Description

Add a new relation to the current relation set if it is not already present. This function ref's the AtkRelation so the caller of this function should unref it to ensure that it will be destroyed when the AtkRelationSet is destroyed.

Usage

atkRelationSetAdd(object, relation)

Arguments

object
[AtkRelationSet] an AtkRelationSet

relation
[AtkRelation] an AtkRelation

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]