atkSelectionGetSelectionCount {RGtk2}	R Documentation

atkSelectionGetSelectionCount

Description

Gets the number of accessible children currently selected. Note: callers should not rely on NULL or on a zero value for indication of whether AtkSelectionIface is implemented, they should use type checking/interface checking functions or the atkGetAccessibleValue() convenience method.

Usage

atkSelectionGetSelectionCount(object)

Arguments

object
[AtkSelection] a GObject instance that implements AtkSelectionIface

Value

[integer] a gint representing the number of items selected, or 0 if selection does not implement this interface.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]