atkSelectionRefSelection {RGtk2}	R Documentation

atkSelectionRefSelection

Description

Gets a reference to the accessible object representing the specified selected child of the object. Note: callers should not rely on NULL or on a zero value for indication of whether AtkSelectionIface is implemented, they should use type checking/interface checking functions or the atkGetAccessibleValue() convenience method.

Usage

atkSelectionRefSelection(object, i)

Arguments

object
[AtkSelection] a GObject instance that implements AtkSelectionIface

i
[integer] a integer specifying the index in the selection set. (e.g. the ith selection as opposed to the ith child).

Value

[AtkObject] an AtkObject representing the selected accessible , or NULL if selection does not implement this interface.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]