atkStateSetContainsStates {RGtk2}	R Documentation

atkStateSetContainsStates

Description

Checks whether the states for all the specified types are in the specified set.

Usage

atkStateSetContainsStates(object, types)

Arguments

object
[AtkStateSet] an AtkStateSet

types
[AtkStateType] a list of AtkStateType

Value

[logical] TRUE if all the states for type are in set.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]