atkStreamableContentGetMimeType {RGtk2}	R Documentation

atkStreamableContentGetMimeType

Description

Gets the character string of the specified mime type. The first mime type is at position 0, the second at position 1, and so on.

Usage

atkStreamableContentGetMimeType(object, i)

Arguments

object
[AtkStreamableContent] a GObject instance that implements AtkStreamableContent

i
[integer] a gint representing the position of the mime type starting from 0

Value

[character] : a gchar* representing the specified mime type; the caller should not free the character string.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]