atkTableGetRowAtIndex {RGtk2}	R Documentation

atkTableGetRowAtIndex

Description

Gets a integer representing the row at the specified index..

Usage

atkTableGetRowAtIndex(object, index)

Arguments

object
[AtkTable] a GObject instance that implements AtkTableInterface

index
[integer] a integer representing an index in table

Value

[integer] a gint representing the row at the specified index, or -1 if the table does not implement this interface

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]