atkTableGetSelectedColumns {RGtk2}	R Documentation

atkTableGetSelectedColumns

Description

Gets the selected columns of the table by initializing **selected with the selected column numbers.

Usage

atkTableGetSelectedColumns(object)

Arguments

object
[AtkTable] a GObject instance that implements AtkTableIface

Value

A list containing the following elements:

retval
[integer] a gint representing the number of selected columns, or 0 if value does not implement this interface.

selected
[integer] a integer** that is to contain the selected columns numbers

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]