atkTableIsColumnSelected {RGtk2}	R Documentation

atkTableIsColumnSelected

Description

Gets a boolean value indicating whether the specified column is selected

Usage

atkTableIsColumnSelected(object, column)

Arguments

object
[AtkTable] a GObject instance that implements AtkTableIface

column
[integer] a integer representing a column in table

Value

[logical] a gboolean representing if the column is selected, or 0 if value does not implement this interface.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]