atkTableSetRowDescription {RGtk2}	R Documentation

atkTableSetRowDescription

Description

Sets the description text for the specified row of table.

Usage

atkTableSetRowDescription(object, row, description)

Arguments

object
[AtkTable] a GObject instance that implements AtkTableIface

row
[integer] a integer representing a row in table

description
[character] a character representing the description text to set for the specified row of table

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]