atkTextAttributeGetValue {RGtk2}	R Documentation

atkTextAttributeGetValue

Description

Gets the value for the index of the AtkTextAttribute

Usage

atkTextAttributeGetValue(attr, index)

Arguments

attr
[AtkTextAttribute] The AtkTextAttribute for which a value is required

index
[integer] The index of the required value

Value

[character] a string containing the value; this string should not be freed; NULL is returned if there are no values maintained for the attr value.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]