atkTextGetBoundedRanges {RGtk2}	R Documentation

atkTextGetBoundedRanges

Description

Get the ranges of text in the specified bounding box.

Usage

atkTextGetBoundedRanges(object, rect, coord.type, x.clip.type, 
    y.clip.type)

Arguments

object
[AtkText] an AtkText

rect
[AtkTextRectangle] An AtkTextRectagle giving the dimensions of the bounding box.

coord.type
[AtkCoordType] Specify whether coordinates are relative to the screen or widget window.

x.clip.type
[AtkTextClipType] Specify the horizontal clip type.

y.clip.type
[AtkTextClipType] Specify the vertical clip type.

Details

Since 1.3

Value

[AtkTextRange] list of AtkTextRange. The last element of the list returned by this function will be NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]