atkTextGetCharacterExtents {RGtk2}	R Documentation

atkTextGetCharacterExtents

Description

Get the bounding box containing the glyph representing the character at a particular text offset.

Usage

atkTextGetCharacterExtents(object, offset, coords)

Arguments

object
[AtkText] an AtkText

offset
[integer] The offset of the text character for which bounding information is required.

coords
[AtkCoordType] specify whether coordinates are relative to the screen or widget window

Value

A list containing the following elements:

x
[integer] Pointer for the x cordinate of the bounding box

y
[integer] Pointer for the y cordinate of the bounding box

width
[integer] Pointer for the width of the bounding box

height
[integer] Pointer for the height of the bounding box

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]