atkTextGetOffsetAtPoint {RGtk2}	R Documentation

atkTextGetOffsetAtPoint

Description

Gets the offset of the character located at coordinates x and y. x and y are interpreted as being relative to the screen or this widget's window depending on coords.

Usage

atkTextGetOffsetAtPoint(object, x, y, coords)

Arguments

object
[AtkText] an AtkText

x
[integer] screen x-position of character

y
[integer] screen y-position of character

coords
[AtkCoordType] specify whether coordinates are relative to the screen or widget window

Value

[integer] the offset to the character which is located at the specified x and y coordinates.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]