atkTextGetRunAttributes {RGtk2}	R Documentation

atkTextGetRunAttributes

Description

Creates an AtkAttributeSet which consists of the attributes explicitly set at the position offset in the text. start.offset and end.offset are set to the start and end of the range around offset where the attributes are invariant. Note that end.offset is the offset of the first character after the range. See the enum AtkTextAttribute for types of text attributes that can be returned. Note that other attributes may also be returned.

Usage

atkTextGetRunAttributes(object, offset)

Arguments

object
[AtkText] an AtkText

offset
[integer] the offset at which to get the attributes, -1 means the offset of the character to be inserted at the caret location.

Value

A list containing the following elements:

retval
[AtkAttributeSet] an AtkAttributeSet which contains the attributes explicitly set at offset.

start.offset
[integer] the address to put the start offset of the range

end.offset
[integer] the address to put the end offset of the range

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]