atkTextGetSelection {RGtk2}	R Documentation

atkTextGetSelection

Description

Gets the text from the specified selection.

Usage

atkTextGetSelection(object, selection.num)

Arguments

object
[AtkText] an AtkText

selection.num
[integer] The selection number. The selected regions are assigned numbers that correspond to how far the region is from the start of the text. The selected region closest to the beginning of the text region is assigned the number 0, etc. Note that adding, moving or deleting a selected region can change the numbering.

Value

A list containing the following elements:

retval
[character] the selected text.

start.offset
[integer] passes back the start position of the selected region

end.offset
[integer] passes back the end position of (e.g. offset immediately past) the selected region

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]