cairo-font-face {RGtk2}	R Documentation

cairo_font_face_t

Description

Base class for font faces

Methods and Functions

cairoFontFaceStatus(font.face)
cairoFontFaceGetType(font.face)
cairoFontFaceSetUserData(font.face, key, user.data)
cairoFontFaceGetUserData(font.face, key)
cairoFontFace(family, slant, weight)

Detailed Description

CairoFontFace represents a particular font at a particular weight, slant, and other characteristic but no size, transformation, or size.

Font faces are created using font-backend-specific constructors, typically of the form cairo_backendfontFaceCreate(), or implicitly using the toy text API by way of cairoSelectFontFace. The resulting face can be accessed using cairoGetFontFace.

Structures

CairoFontFace

A CairoFontFace specifies all aspects of a font other than the size or font matrix (a font matrix is used to distort a font by sheering it or scaling it unequally in the two directions) . A font face can be set on a Cairo by using cairoSetFontFace; the size and font matrix are set with cairoSetFontSize and cairoSetFontMatrix.

There are various types of font faces, depending on the font backend they use. The type of a font face can be queried using cairoFontFaceGetType.

Memory management of CairoFontFace is done with cairoFontFaceReference() and cairoFontFaceDestroy().

Convenient Construction

cairoFontFace is the result of collapsing the constructors of cairo_font_face_t (cairoToyFontFaceCreate, cairoUserFontFaceCreate) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Enums and Flags

CairoFontType

CairoFontType is used to describe the type of a given font face or scaled font. The font types are also known as "font backends" within cairo.

The type of a font face is determined by the function used to create it, which will generally be of the form cairo_typefontFaceCreate(). The font face type can be queried with cairoFontFaceGetType

The various CairoFontFace functions can be used with a font face of any type.

The type of a scaled font is determined by the type of the font face passed to cairoScaledFontCreate. The scaled font type can be queried with cairoScaledFontGetType

The various CairoScaledFont functions can be used with scaled fonts of any type, but some font backends also provide type-specific functions that must only be called with a scaled font of the appropriate type. These functions have names that begin with cairo_typescaledFont() such as cairoFtScaledFontLockFace().

The behavior of calling a type-specific function with a scaled font of the wrong type is undefined.

New entries may be added in future versions. Since 1.2

toy

The font was created using cairo's toy font api

ft

The font is of type FreeType

win32

The font is of type Win32

atsui

The font is of type Quartz (Since: 1.6)

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://www.cairographics.org/manual/cairo-font-face.html

[Package RGtk2 version 2.20.31 Index]