cairo-image-surface {RGtk2}	R Documentation

Image Surfaces

Description

Rendering to memory buffers

Methods and Functions

cairoFormatStrideForWidth(format, width)
cairoImageSurfaceCreate(format, width, height)
cairoImageSurfaceCreateForData(data, format, width, height, stride)
cairoImageSurfaceGetData(surface)
cairoImageSurfaceGetFormat(surface)
cairoImageSurfaceGetWidth(surface)
cairoImageSurfaceGetHeight(surface)
cairoImageSurfaceGetStride(surface)

Detailed Description

Image surfaces provide the ability to render to memory buffers either allocated by cairo or by the calling code. The supported image formats are those defined in CairoFormat.

Enums and Flags

CairoFormat

CairoFormat is used to identify the memory format of image data.

New entries may be added in future versions.

argb32

each pixel is a 32-bit quantity, with alpha in the upper 8 bits, then red, then green, then blue. The 32-bit quantities are stored native-endian. Pre-multiplied alpha is used. (That is, 50% transparent red is 0x80800000, not 0x80ff0000.)

rgb24

each pixel is a 32-bit quantity, with the upper 8 bits unused. Red, Green, and Blue are stored in the remaining 24 bits in that order.

a8

each pixel is a 8-bit quantity holding an alpha value.

a1

each pixel is a 1-bit quantity holding an alpha value. Pixels are packed together into 32-bit quantities. The ordering of the bits matches the endianess of the platform. On a big-endian machine, the first pixel is in the uppermost bit, on a little-endian machine the first pixel is in the least-significant bit.

rgb16-565

undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://www.cairographics.org/manual/cairo-image-surface.html

[Package RGtk2 version 2.20.31 Index]