cairo-ps-surface {RGtk2}	R Documentation

PostScript Surfaces

Description

Rendering PostScript documents

Methods and Functions

cairoPsSurfaceCreate(filename, width.in.points, height.in.points)
cairoPsSurfaceCreateForStream(write.func, closure, width.in.points, height.in.points)
cairoPsSurfaceRestrictToLevel(surface, level)
cairoPsGetLevels()
cairoPsLevelToString(level)
cairoPsSurfaceSetEps(surface, eps)
cairoPsSurfaceGetEps(surface)
cairoPsSurfaceSetSize(surface, width.in.points, height.in.points)
cairoPsSurfaceDscBeginSetup(surface)
cairoPsSurfaceDscBeginPageSetup(surface)
cairoPsSurfaceDscComment(surface, comment)

Detailed Description

The PostScript surface is used to render cairo graphics to Adobe PostScript files and is a multi-page vector surface backend.

Enums and Flags

CairoPsLevel

CairoPsLevel is used to describe the language level of the PostScript Language Reference that a generated PostScript file will conform to.

2

The language level 2 of the PostScript specification.

3

The language level 3 of the PostScript specification.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://www.cairographics.org/manual/cairo-ps-surface.html

[Package RGtk2 version 2.20.31 Index]