cairo-quartz-font {RGtk2}	R Documentation

Quartz (CGFont) Fonts

Description

Font support via CGFont on OS X

Detailed Description

The Quartz font backend is primarily used to render text on Apple MacOS X systems. The CGFont API is used for the internal implementation of the font backend methods.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://www.cairographics.org/manual/cairo-quartz-font.html

[Package RGtk2 version 2.20.31 Index]