cairo-scaled-font {RGtk2}	R Documentation

cairo_scaled_font_t

Description

Font face at particular size and options

Methods and Functions

cairoScaledFontCreate(font.face, font.matrix, ctm, option)
cairoScaledFontStatus(scaled.font)
cairoScaledFontExtents(scaled.font)
cairoScaledFontTextExtents(scaled.font, utf8)
cairoScaledFontGlyphExtents(scaled.font, glyphs, num.glyphs)
cairoScaledFontTextToGlyphs(scaled.font, x, y, utf8, utf8.len = -1)
cairoScaledFontGetFontFace(scaled.font)
cairoScaledFontGetFontOptions(scaled.font)
cairoScaledFontGetFontMatrix(scaled.font)
cairoScaledFontGetCtm(scaled.font)
cairoScaledFontGetScaleMatrix(scaled.font)
cairoScaledFontGetType(scaled.font)
cairoScaledFontSetUserData(scaled.font, key, user.data)
cairoScaledFontGetUserData(scaled.font, key)
cairoScaledFont(font.face, font.matrix, ctm, option)

Detailed Description

CairoScaledFont represents a realization of a font face at a particular size and transformation and a certain set of font options.

Structures

CairoScaledFont

A CairoScaledFont is a font scaled to a particular size and device resolution. A CairoScaledFont is most useful for low-level font usage where a library or application wants to cache a reference to a scaled font to speed up the computation of metrics.

There are various types of scaled fonts, depending on the font backend they use. The type of a scaled font can be queried using cairoScaledFontGetType.

Memory management of CairoScaledFont is done with cairoScaledFontReference() and cairoScaledFontDestroy().

CairoFontExtents

The CairoFontExtents structure stores metric information for a font. Values are given in the current user-space coordinate system.

Because font metrics are in user-space coordinates, they are mostly, but not entirely, independent of the current transformation matrix. If you call cairo_scale(cr, 2.0, 2.0), text will be drawn twice as big, but the reported text extents will not be doubled. They will change slightly due to hinting (so you can't assume that metrics are independent of the transformation matrix), but otherwise will remain unchanged. CairoFontExtents is a transparent-type.

ascent

[numeric] the distance that the font extends above the baseline. Note that this is not always exactly equal to the maximum of the extents of all the glyphs in the font, but rather is picked to express the font designer's intent as to how the font should align with elements above it.

descent

[numeric] the distance that the font extends below the baseline. This value is positive for typical fonts that include portions below the baseline. Note that this is not always exactly equal to the maximum of the extents of all the glyphs in the font, but rather is picked to express the font designer's intent as to how the the font should align with elements below it.

height

[numeric] the recommended vertical distance between baselines when setting consecutive lines of text with the font. This is greater than ascent+descent by a quantity known as the line spacing or external leading. When space is at a premium, most fonts can be set with only a distance of ascent+descent between lines.

maxXAdvance

[numeric] the maximum distance in the X direction that the the origin is advanced for any glyph in the font.

maxYAdvance

[numeric] the maximum distance in the Y direction that the the origin is advanced for any glyph in the font. this will be zero for normal fonts used for horizontal writing. (The scripts of East Asia are sometimes written vertically.)

CairoTextExtents

The CairoTextExtents structure stores the extents of a single glyph or a string of glyphs in user-space coordinates. Because text extents are in user-space coordinates, they are mostly, but not entirely, independent of the current transformation matrix. If you call cairo_scale(cr, 2.0, 2.0), text will be drawn twice as big, but the reported text extents will not be doubled. They will change slightly due to hinting (so you can't assume that metrics are independent of the transformation matrix), but otherwise will remain unchanged.

xBearing

[numeric] the horizontal distance from the origin to the leftmost part of the glyphs as drawn. Positive if the glyphs lie entirely to the right of the origin.

yBearing

[numeric] the vertical distance from the origin to the topmost part of the glyphs as drawn. Positive only if the glyphs lie completely below the origin; will usually be negative.

width

[numeric] width of the glyphs as drawn

height

[numeric] height of the glyphs as drawn

xAdvance

[numeric] distance to advance in the X direction after drawing these glyphs

yAdvance

[numeric] distance to advance in the Y direction after drawing these glyphs. Will typically be zero except for vertical text layout as found in East-Asian languages.

Convenient Construction

cairoScaledFont is the equivalent of cairoScaledFontCreate.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://www.cairographics.org/manual/cairo-scaled-font.html

[Package RGtk2 version 2.20.31 Index]