cairo-svg-surface {RGtk2}	R Documentation

SVG Surfaces

Description

Rendering SVG documents

Methods and Functions

cairoSvgSurfaceCreate(filename, width.in.points, height.in.points)
cairoSvgSurfaceCreateForStream(write.func, closure, width.in.points, height.in.points)
cairoSvgSurfaceRestrictToVersion(surface, version)
cairoSvgGetVersions(versions, num.versions)
cairoSvgVersionToString(version)

Detailed Description

The SVG surface is used to render cairo graphics to SVG files and is a multi-page vector surface backend.

Enums and Flags

CairoSvgVersion

CairoSvgVersion is used to describe the version number of the SVG specification that a generated SVG file will conform to.

1-1

The version 1.1 of the SVG specification.

1-2

The version 1.2 of the SVG specification.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://www.cairographics.org/manual/cairo-svg-surface.html

[Package RGtk2 version 2.20.31 Index]