cairo-transformations {RGtk2}	R Documentation

Transformations

Description

Manipulating the current transformation matrix

Methods and Functions

cairoTranslate(cr, tx, ty)
cairoScale(cr, sx, sy)
cairoRotate(cr, angle)
cairoTransform(cr, matrix)
cairoSetMatrix(cr, matrix)
cairoGetMatrix(cr, matrix)
cairoIdentityMatrix(cr)
cairoUserToDevice(cr, x, y)
cairoUserToDeviceDistance(cr, dx, dy)
cairoDeviceToUser(cr, x, y)
cairoDeviceToUserDistance(cr, dx, dy)

Detailed Description

The current transformation matrix, ctm, is a two-dimensional affine transformation that maps all coordinates and other drawing instruments from the user space into the surface's canonical coordinate system, also known as the device space.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://www.cairographics.org/manual/cairo-transformations.html

[Package RGtk2 version 2.20.31 Index]