cairo-types {RGtk2}	R Documentation

Types

Description

Generic data types

Detailed Description

Generic data types used in the cairo API

Structures

CairoUserDataKey

CairoUserDataKey is used for attaching user data to cairo data structures. The actual contents of the struct is never used, and there is no need to initialize the object; only the unique a cairo_data_key_t object is used. Typically, you would just use the a static cairo_data_key_t object.

unused

[integer] not used; ignore.

User Functions

cairo_destroy_func_t(data)

cairo_destroy_func_t the type of function which is called when a data element is destroyed. It is passed the pointer to the data element and should free any memory and resources allocated for it.

data

[R object] The data element being destroyed.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://www.cairographics.org/manual/cairo-types.html

[Package RGtk2 version 2.20.31 Index]