cairo-user-font {RGtk2}	R Documentation

User Fonts

Description

Font support with font data provided by the user

Methods and Functions

cairoUserFontFaceCreate()
cairoUserFontFaceSetInitFunc(font.face, init.func)
cairoUserFontFaceGetInitFunc(font.face)
cairoUserFontFaceSetRenderGlyphFunc(font.face, render.glyph.func)
cairoUserFontFaceGetRenderGlyphFunc(font.face)
cairoUserFontFaceSetUnicodeToGlyphFunc(font.face, unicode.to.glyph.func)
cairoUserFontFaceGetUnicodeToGlyphFunc(font.face)
cairoUserFontFaceSetTextToGlyphsFunc(font.face, text.to.glyphs.func)
cairoUserFontFaceGetTextToGlyphsFunc(font.face)

Detailed Description

The user-font feature allows the cairo user to provide drawings for glyphs in a font. This is most useful in implementing fonts in non-standard formats, like SVG fonts and Flash fonts, but can also be used by games and other application to draw "funky" fonts.

User Functions

cairo_user_scaled_font_init_func_t(scaled.font, cr, extents)

cairo_user_scaled_font_init_func_t is the type of function which is called when a scaled-font needs to be created for a user font-face.

The cairo context cr is not used by the caller, but is prepared in font space, similar to what the cairo contexts passed to the render_glyph method will look like. The callback can use this context for extents computation for example. After the callback is called, cr is checked for any error status.

The extents argument is where the user font sets the font extents for scaled.font. It is in font space, which means that for most cases its ascent and descent members should add to 1.0. extents is preset to hold a value of 1.0 for ascent, height, and max_x_advance, and 0.0 for descent and max_y_advance members.

The callback is optional. If not set, default font extents as described in the previous paragraph will be used.

Note that scaled.font is not fully initialized at this point and trying to use it for text operations in the callback will result in deadlock. Since 1.8

scaled.font

[CairoScaledFont] the scaled-font being created

cr

[Cairo] a cairo context, in font space

extents

[CairoFontExtents] font extents to fill in, in font space

Returns: [CairoStatus] CAIRO_STATUS_SUCCESS upon success, or CAIRO_STATUS_USER_FONT_ERROR or any other error status on error.

cairo_user_scaled_font_render_glyph_func_t(scaled.font, glyph, cr, extents)

cairo_user_scaled_font_render_glyph_func_t is the type of function which is called when a user scaled-font needs to render a glyph.

The callback is mandatory, and expected to draw the glyph with code glyph to the cairo context cr. cr is prepared such that the glyph drawing is done in font space. That is, the matrix set on cr is the scale matrix of scaled.font, The extents argument is where the user font sets the font extents for scaled.font. However, if user prefers to draw in user space, they can achieve that by changing the matrix on cr. All cairo rendering operations to cr are permitted, however, the result is undefined if any source other than the default source on cr is used. That means, glyph bitmaps should be rendered using cairoMask instead of cairoPaint.

Other non-default settings on cr include a font size of 1.0 (given that it is set up to be in font space), and font options corresponding to scaled.font.

The extents argument is preset to have x_bearing, width, and y_advance of zero, y_bearing set to -font_extents.ascent, height to font_extents.ascent+font_extents.descent, and x_advance to font_extents.max_x_advance. The only field user needs to set in majority of cases is x_advance. If the width field is zero upon the callback returning (which is its preset value), the glyph extents are automatically computed based on the drawings done to cr. This is in most cases exactly what the desired behavior is. However, if for any reason the callback sets the extents, it must be ink extents, and include the extents of all drawing done to cr in the callback. Since 1.8

scaled.font

[CairoScaledFont] user scaled-font

glyph

[unsignedlong] glyph code to render

cr

[Cairo] cairo context to draw to, in font space

extents

[CairoTextExtents] glyph extents to fill in, in font space

Returns: [CairoStatus] CAIRO_STATUS_SUCCESS upon success, or CAIRO_STATUS_USER_FONT_ERROR or any other error status on error.

cairo_user_scaled_font_text_to_glyphs_func_t(scaled.font, utf8, utf8.len, glyphs, num.glyphs, clusters, num.clusters, cluster.flags)

cairo_user_scaled_font_text_to_glyphs_func_t is the type of function which is called to convert input text to a list of glyphs. This is used by the cairoShowText operation.

Using this callback the user-font has full control on glyphs and their positions. That means, it allows for features like ligatures and kerning, as well as complex shaping required for scripts like Arabic and Indic.

The num.glyphs argument is preset to the number of glyph entries available in the glyphs buffer. If the glyphs buffer is NULL, the value of num.glyphs will be zero. If the provided glyph list is too short for the conversion (or for convenience), a new glyph list may be allocated using cairoGlyphAllocate() and placed in glyphs. Upon return, num.glyphs should contain the number of generated glyphs. If the value glyphs points at has changed after the call, the caller will free the allocated glyph list using cairoGlyphFree(). The callback should populate the glyph indices and positions (in font space) assuming that the text is to be shown at the origin.

If clusters is not NULL, num.clusters and cluster.flags are also non-NULL, and cluster mapping should be computed. The semantics of how cluster list allocation works is similar to the glyph list. That is, if clusters initially points to a non-NULL value, that list may be used as a cluster buffer, and num.clusters points to the number of cluster entries available there. If the provided cluster list is too short for the conversion (or for convenience), a new cluster list may be allocated using cairoTextClusterAllocate() and placed in clusters. Upon return, num.clusters should contain the number of generated clusters. If the value clusters points at has changed after the call, the caller will free the allocated cluster list using cairoTextClusterFree().

The callback is optional. If num.glyphs is negative upon the callback returning, the unicode_to_glyph callback is tried. See cairo_user_scaled_font_unicode_to_glyph_func_t.

Note: While cairo does not impose any limitation on glyph indices, some applications may assume that a glyph index fits in a 16-bit unsigned integer. As such, it is advised that user-fonts keep their glyphs in the 0 to 65535 range. Furthermore, some applications may assume that glyph 0 is a special glyph-not-found glyph. User-fonts are advised to use glyph 0 for such purposes and do not use that glyph value for other purposes. Since 1.8

scaled.font

[CairoScaledFont] the scaled-font being created

utf8

[char] a string of text encoded in UTF-8

utf8.len

[integer] length of utf8 in bytes

glyphs

[CairoGlyph] pointer to list of glyphs to fill, in font space

num.glyphs

[integer] pointer to number of glyphs

clusters

[CairoTextCluster] pointer to list of cluster mapping information to fill, or NULL

num.clusters

[integer] pointer to number of clusters

cluster.flags

[CairoTextClusterFlags] pointer to location to store cluster flags corresponding to the output clusters

Returns: [CairoStatus] CAIRO_STATUS_SUCCESS upon success, or CAIRO_STATUS_USER_FONT_ERROR or any other error status on error.

cairo_user_scaled_font_unicode_to_glyph_func_t(scaled.font, unicode, glyph.index)

cairo_user_scaled_font_unicode_to_glyph_func_t is the type of function which is called to convert an input Unicode character to a single glyph. This is used by the cairoShowText operation.

This callback is used to provide the same functionality as the text_to_glyphs callback does (see cairo_user_scaled_font_text_to_glyphs_func_t) but has much less control on the output, in exchange for increased ease of use. The inherent assumption to using this callback is that each character maps to one glyph, and that the mapping is context independent. It also assumes that glyphs are positioned according to their advance width. These mean no ligatures, kerning, or complex scripts can be implemented using this callback.

The callback is optional, and only used if text_to_glyphs callback is not set or fails to return glyphs. If this callback is not set, an identity mapping from Unicode code-points to glyph indices is assumed.

Note: While cairo does not impose any limitation on glyph indices, some applications may assume that a glyph index fits in a 16-bit unsigned integer. As such, it is advised that user-fonts keep their glyphs in the 0 to 65535 range. Furthermore, some applications may assume that glyph 0 is a special glyph-not-found glyph. User-fonts are advised to use glyph 0 for such purposes and do not use that glyph value for other purposes. Since 1.8

scaled.font

[CairoScaledFont] the scaled-font being created

unicode

[unsignedlong] input unicode character code-point

glyph.index

[long] output glyph index

Returns: [CairoStatus] CAIRO_STATUS_SUCCESS upon success, or CAIRO_STATUS_USER_FONT_ERROR or any other error status on error.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://www.cairographics.org/manual/cairo-user-font.html

[Package RGtk2 version 2.20.31 Index]