cairoArcNegative {RGtk2}	R Documentation

cairoArcNegative

Description

Adds a circular arc of the given radius to the current path. The arc is centered at (xc, yc), begins at angle1 and proceeds in the direction of decreasing angles to end at angle2. If angle2 is greater than angle1 it will be progressively decreased by 2*M_PI until it is less than angle1.

Usage

cairoArcNegative(cr, xc, yc, radius, angle1, angle2)

Arguments

cr
[Cairo] a cairo context

xc
[numeric] X position of the center of the arc

yc
[numeric] Y position of the center of the arc

radius
[numeric] the radius of the arc

angle1
[numeric] the start angle, in radians

angle2
[numeric] the end angle, in radians

Details

See cairoArc for more details. This function differs only in the direction of the arc between the two angles.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]