cairoClipPreserve {RGtk2}	R Documentation

cairoClipPreserve

Description

Establishes a new clip region by intersecting the current clip region with the current path as it would be filled by cairoFill and according to the current fill rule (see cairoSetFillRule).

Usage

cairoClipPreserve(cr)

Arguments

cr
[Cairo] a cairo context

Details

Unlike cairoClip, cairoClipPreserve preserves the path within the cairo context.

The current clip region affects all drawing operations by effectively masking out any changes to the surface that are outside the current clip region.

Calling cairoClipPreserve can only make the clip region smaller, never larger. But the current clip is part of the graphics state, so a temporary restriction of the clip region can be achieved by calling cairoClipPreserve within a cairoSave/cairoRestore pair. The only other means of increasing the size of the clip region is cairoResetClip.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]