cairoCopyPage {RGtk2}	R Documentation

cairoCopyPage

Description

Emits the current page for backends that support multiple pages, but doesn't clear it, so, the contents of the current page will be retained for the next page too. Use cairoShowPage if you want to get an empty page after the emission.

Usage

cairoCopyPage(cr)

Arguments

cr
[Cairo] a cairo context

Details

This is a convenience function that simply calls cairoSurfaceCopyPage on cr's target.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]