cairoCopyPath {RGtk2}	R Documentation

cairoCopyPath

Description

Creates a copy of the current path and returns it to the user as a CairoPath. See CairoPathData for hints on how to iterate over the returned data structure.

Usage

cairoCopyPath(cr)

Arguments

cr
[Cairo] a cairo context

Details

This function will always return a valid pointer, but the result will have no data (data== NULL and num_data==0), if either of the following conditions hold:

Value

[CairoPath] the copy of the current path.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]