cairoCreate {RGtk2}	R Documentation

cairoCreate

Description

Creates a new Cairo with all graphics state parameters set to default values and with target as a target surface. The target surface should be constructed with a backend-specific function such as cairoImageSurfaceCreate (or any other cairo_backendsurfaceCreate() variant).

Usage

cairoCreate(target)

Arguments

target
[CairoSurface] target surface for the context

Details

This function references target, so you can immediately call cairoSurfaceDestroy on it if you don't need to maintain a separate reference to it.

Value

[Cairo] a newly allocated Cairo with a reference count of 1. This function never returns NULL. If memory cannot be allocated, a special Cairo object will be returned on which cairoStatus returns CAIRO_STATUS_NO_MEMORY. You can use this object normally, but no drawing will be done.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]