cairoCurveTo {RGtk2}	R Documentation

cairoCurveTo

Description

Adds a cubic B<U+00E9>zier spline to the path from the current point to position (x3, y3) in user-space coordinates, using (x1, y1) and (x2, y2) as the control points. After this call the current point will be (x3, y3).

Usage

cairoCurveTo(cr, x1, y1, x2, y2, x3, y3)

Arguments

cr
[Cairo] a cairo context

x1
[numeric] the X coordinate of the first control point

y1
[numeric] the Y coordinate of the first control point

x2
[numeric] the X coordinate of the second control point

y2
[numeric] the Y coordinate of the second control point

x3
[numeric] the X coordinate of the end of the curve

y3
[numeric] the Y coordinate of the end of the curve

Details

If there is no current point before the call to cairoCurveTo this function will behave as if preceded by a call to cairo_move_to(cr, x1, y1).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]