cairoDeviceToUserDistance {RGtk2}	R Documentation

cairoDeviceToUserDistance

Description

Transform a distance vector from device space to user space. This function is similar to cairoDeviceToUser except that the translation components of the inverse CTM will be ignored when transforming (dx,dy).

Usage

cairoDeviceToUserDistance(cr, dx, dy)

Arguments

cr
[Cairo] a cairo context

dx
[numeric] X component of a distance vector (in/out parameter)

dy
[numeric] Y component of a distance vector (in/out parameter)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]