cairoFontFaceSetUserData {RGtk2}	R Documentation

cairoFontFaceSetUserData

Description

Attach user data to font.face. To remove user data from a font face, call this function with the key that was used to set it and NULL for data.

Usage

cairoFontFaceSetUserData(font.face, key, user.data)

Arguments

font.face
[CairoFontFace] a CairoFontFace

key
[CairoUserDataKey] the a CairoUserDataKey to attach the user data to

user.data
[R object] the user data to attach to the font face

Value

[CairoStatus] CAIRO_STATUS_SUCCESS or CAIRO_STATUS_NO_MEMORY if a slot could not be allocated for the user data.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]