cairoFontOptionsSetSubpixelOrder {RGtk2}	R Documentation

cairoFontOptionsSetSubpixelOrder

Description

Sets the subpixel order for the font options object. The subpixel order specifies the order of color elements within each pixel on the display device when rendering with an antialiasing mode of CAIRO_ANTIALIAS_SUBPIXEL. See the documentation for CairoSubpixelOrder for full details.

Usage

cairoFontOptionsSetSubpixelOrder(options, subpixel.order)

Arguments

options
[CairoFontOptions] a CairoFontOptions

subpixel.order
[CairoSubpixelOrder] the new subpixel order

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]