cairoGetCurrentPoint {RGtk2}	R Documentation

cairoGetCurrentPoint

Description

Gets the current point of the current path, which is conceptually the final point reached by the path so far.

Usage

cairoGetCurrentPoint(cr, x, y)

Arguments

cr
[Cairo] a cairo context

x
[numeric] return value for X coordinate of the current point

y
[numeric] return value for Y coordinate of the current point

Details

The current point is returned in the user-space coordinate system. If there is no defined current point or if cr is in an error status, x and y will both be set to 0.0. It is possible to check this in advance with cairoHasCurrentPoint.

Most path construction functions alter the current point. See the following for details on how they affect the current point: cairoNewPath, cairoNewSubPath, cairoAppendPath, cairoClosePath, cairoMoveTo, cairoLineTo, cairoCurveTo, cairoRelMoveTo, cairoRelLineTo, cairoRelCurveTo, cairoArc, cairoArcNegative, cairoRectangle, cairoTextPath, cairoGlyphPath, cairoStrokeToPath().

Some functions use and alter the current point but do not otherwise change current path: cairoShowText.

Some functions unset the current path and as a result, current point: cairoFill, cairoStroke.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]