cairoGetFontFace {RGtk2}	R Documentation

cairoGetFontFace

Description

Gets the current font face for a Cairo.

Usage

cairoGetFontFace(cr)

Arguments

cr
[Cairo] a Cairo

Value

[CairoFontFace] the current font face. To keep a reference to it, This function never returns NULL. If memory cannot be allocated, a special "nil" CairoFontFace object will be returned on which cairoFontFaceStatus returns CAIRO_STATUS_NO_MEMORY. Using this nil object will cause its error state to propagate to other objects it is passed to, (for example, calling cairoSetFontFace with a nil font will trigger an error that will shutdown the Cairo object).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]