cairoGetLineWidth {RGtk2}	R Documentation

cairoGetLineWidth

Description

This function returns the current line width value exactly as set by cairoSetLineWidth. Note that the value is unchanged even if the CTM has changed between the calls to cairoSetLineWidth and cairoGetLineWidth.

Usage

cairoGetLineWidth(cr)

Arguments

cr
[Cairo] a cairo context

Value

[numeric] the current line width.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]