cairoGetScaledFont {RGtk2}	R Documentation

cairoGetScaledFont

Description

Gets the current scaled font for a Cairo.

Usage

cairoGetScaledFont(cr)

Arguments

cr
[Cairo] a Cairo

Details

Since 1.4

Value

[CairoScaledFont] the current scaled font. To keep a reference to it, This function never returns NULL. If memory cannot be allocated, a special "nil" CairoScaledFont object will be returned on which cairoScaledFontStatus returns CAIRO_STATUS_NO_MEMORY. Using this nil object will cause its error state to propagate to other objects it is passed to, (for example, calling cairoSetScaledFont with a nil font will trigger an error that will shutdown the Cairo object).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]