cairoGlyphPath {RGtk2}	R Documentation

cairoGlyphPath

Description

Adds closed paths for the glyphs to the current path. The generated path if filled, achieves an effect similar to that of cairoShowGlyphs.

Usage

cairoGlyphPath(cr, glyphs)

Arguments

cr
[Cairo] a cairo context

glyphs
[CairoGlyph] list of glyphs to show

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]