cairoImageSurfaceCreate {RGtk2}	R Documentation

cairoImageSurfaceCreate

Description

Creates an image surface of the specified format and dimensions. Initially the surface contents are all 0. (Specifically, within each pixel, each color or alpha channel belonging to format will be 0. The contents of bits within a pixel, but not belonging to the given format are undefined).

Usage

cairoImageSurfaceCreate(format, width, height)

Arguments

format
[CairoFormat] format of pixels in the surface to create

width
[integer] width of the surface, in pixels

height
[integer] height of the surface, in pixels

Value

[CairoSurface] a pointer to the newly created surface. The caller owns the surface and should call cairoSurfaceDestroy when done with it. This function always returns a valid pointer, but it will return a pointer to a "nil" surface if an error such as out of memory occurs. You can use cairoSurfaceStatus to check for this.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]