cairoImageSurfaceCreateFromPngStream {RGtk2}	R Documentation

cairoImageSurfaceCreateFromPngStream

Description

Creates a new image surface from PNG data read incrementally via the read.func function.

Usage

cairoImageSurfaceCreateFromPngStream(con)

Value

[CairoSurface] a new CairoSurface initialized with the contents of the PNG file or a "nil" surface if the data read is not a valid PNG image or memory could not be allocated for the operation. A nil surface can be checked for with cairo_surface_status(surface) which may return one of the following values: CAIRO_STATUS_NO_MEMORY CAIRO_STATUS_READ_ERROR Alternatively, you can allow errors to propagate through the drawing operations and check the status on the context upon completion using cairoStatus.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]