cairoImageSurfaceGetStride {RGtk2}	R Documentation

cairoImageSurfaceGetStride

Description

Get the stride of the image surface in bytes

Usage

cairoImageSurfaceGetStride(surface)

Arguments

surface
[CairoSurface] a cairo_image_surface_t

Details

Since 1.2

Value

[integer] the stride of the image surface in bytes (or 0 if surface is not an image surface). The stride is the distance in bytes from the beginning of one row of the image data to the beginning of the next row.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]