cairoInFill {RGtk2}	R Documentation

cairoInFill

Description

Tests whether the given point is inside the area that would be affected by a cairoFill operation given the current path and filling parameters. Surface dimensions and clipping are not taken into account.

Usage

cairoInFill(cr, x, y)

Arguments

cr
[Cairo] a cairo context

x
[numeric] X coordinate of the point to test

y
[numeric] Y coordinate of the point to test

Details

See cairoFill, cairoSetFillRule and cairoFillPreserve.

Value

[logical] A non-zero value if the point is inside, or zero if outside.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]