cairoLineTo {RGtk2}	R Documentation

cairoLineTo

Description

Adds a line to the path from the current point to position (x, y) in user-space coordinates. After this call the current point will be (x, y).

Usage

cairoLineTo(cr, x, y)

Arguments

cr
[Cairo] a cairo context

x
[numeric] the X coordinate of the end of the new line

y
[numeric] the Y coordinate of the end of the new line

Details

If there is no current point before the call to cairoLineTo this function will behave as cairo_move_to(cr, x, y).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]