cairoMask {RGtk2}	R Documentation

cairoMask

Description

A drawing operator that paints the current source using the alpha channel of pattern as a mask. (Opaque areas of pattern are painted with the source, transparent areas are not painted.)

Usage

cairoMask(cr, pattern)

Arguments

cr
[Cairo] a cairo context

pattern
[CairoPattern] a CairoPattern

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]