cairoMatrixInvert {RGtk2}	R Documentation

cairoMatrixInvert

Description

Changes matrix to be the inverse of it's original value. Not all transformation matrices have inverses; if the matrix collapses points together (it is degenerate), then it has no inverse and this function will fail.

Usage

cairoMatrixInvert(matrix)

Arguments

matrix
[CairoMatrix] a CairoMatrix

Value

[CairoStatus] If matrix has an inverse, modifies matrix to be the inverse matrix and returns CAIRO_STATUS_SUCCESS. Otherwise, returns CAIRO_STATUS_INVALID_MATRIX.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]