cairoMatrixMultiply {RGtk2}	R Documentation

cairoMatrixMultiply

Description

Multiplies the affine transformations in a and b together and stores the result in result. The effect of the resulting transformation is to first apply the transformation in a to the coordinates and then apply the transformation in b to the coordinates.

Usage

cairoMatrixMultiply(result, a, b)

Arguments

result
[CairoMatrix] a CairoMatrix in which to store the result

a
[CairoMatrix] a CairoMatrix

b
[CairoMatrix] a CairoMatrix

Details

It is allowable for result to be identical to either a or b.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]