cairoMatrixRotate {RGtk2}	R Documentation

cairoMatrixRotate

Description

Applies rotation by radians to the transformation in matrix. The effect of the new transformation is to first rotate the coordinates by radians, then apply the original transformation to the coordinates.

Usage

cairoMatrixRotate(matrix, radians)

Arguments

matrix
[CairoMatrix] a CairoMatrix

radians
[numeric] angle of rotation, in radians. The direction of rotation is defined such that positive angles rotate in the direction from the positive X axis toward the positive Y axis. With the default axis orientation of cairo, positive angles rotate in a clockwise direction.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]