cairoMatrixScale {RGtk2}	R Documentation

cairoMatrixScale

Description

Applies scaling by sx, sy to the transformation in matrix. The effect of the new transformation is to first scale the coordinates by sx and sy, then apply the original transformation to the coordinates.

Usage

cairoMatrixScale(matrix, sx, sy)

Arguments

matrix
[CairoMatrix] a CairoMatrix

sx
[numeric] scale factor in the X direction

sy
[numeric] scale factor in the Y direction

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]