cairoMatrixTransformDistance {RGtk2}	R Documentation

cairoMatrixTransformDistance

Description

Transforms the distance vector (dx,dy) by matrix. This is similar to cairoMatrixTransformPoint except that the translation components of the transformation are ignored. The calculation of the returned vector is as follows:

Usage

cairoMatrixTransformDistance(matrix, dx, dy)

Arguments

matrix
[CairoMatrix] a CairoMatrix

dx
[numeric] X component of a distance vector. An in/out parameter

dy
[numeric] Y component of a distance vector. An in/out parameter

Details

dx2 = dx1 * a + dy1 * c;
dy2 = dx1 * b + dy1 * d;

Affine transformations are position invariant, so the same vector always transforms to the same vector. If (x1,y1) transforms to (x2,y2) then (x1+dx1,y1+dy1) will transform to (x1+dx2,y1+dy2) for all values of x1 and x2.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]