cairoPaintWithAlpha {RGtk2}	R Documentation

cairoPaintWithAlpha

Description

A drawing operator that paints the current source everywhere within the current clip region using a mask of constant alpha value alpha. The effect is similar to cairoPaint, but the drawing is faded out using the alpha value.

Usage

cairoPaintWithAlpha(cr, alpha)

Arguments

cr
[Cairo] a cairo context

alpha
[numeric] alpha value, between 0 (transparent) and 1 (opaque)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]