cairoPathExtents {RGtk2}	R Documentation

cairoPathExtents

Description

Computes a bounding box in user-space coordinates covering the points on the current path. If the current path is empty, returns an empty rectangle ((0,0), (0,0)). Stroke parameters, fill rule, surface dimensions and clipping are not taken into account.

Usage

cairoPathExtents(cr)

Arguments

cr
[Cairo] a cairo context

Details

Contrast with cairoFillExtents and cairoStrokeExtents which return the extents of only the area that would be "inked" by the corresponding drawing operations.

The result of cairoPathExtents is defined as equivalent to the limit of cairoStrokeExtents with CAIRO_LINE_CAP_ROUND as the line width approaches 0.0, (but never reaching the empty-rectangle returned by cairoStrokeExtents for a line width of 0.0).

Specifically, this means that zero-area sub-paths such as cairoMoveTo;cairoLineTo segments, (even degenerate cases where the coordinates to both calls are identical), will be considered as contributing to the extents. However, a lone cairoMoveTo will not contribute to the results of cairoPathExtents. Since 1.6

Value

A list containing the following elements:

x1
[numeric] left of the resulting extents

y1
[numeric] top of the resulting extents

x2
[numeric] right of the resulting extents

y2
[numeric] bottom of the resulting extents

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]