cairoPatternAddColorStopRgb {RGtk2}	R Documentation

cairoPatternAddColorStopRgb

Description

Adds an opaque color stop to a gradient pattern. The offset specifies the location along the gradient's control vector. For example, a linear gradient's control vector is from (x0,y0) to (x1,y1) while a radial gradient's control vector is from any point on the start circle to the corresponding point on the end circle.

Usage

cairoPatternAddColorStopRgb(pattern, offset, red, green, blue)

Arguments

pattern
[CairoPattern] a CairoPattern

offset
[numeric] an offset in the range [0.0 .. 1.0]

red
[numeric] red component of color

green
[numeric] green component of color

blue
[numeric] blue component of color

Details

The color is specified in the same way as in cairoSetSourceRgb.

If two (or more) stops are specified with identical offset values, they will be sorted according to the order in which the stops are added, (stops added earlier will compare less than stops added later). This can be useful for reliably making sharp color transitions instead of the typical blend.

Note: If the pattern is not a gradient pattern, (eg. a linear or radial pattern), then the pattern will be put into an error status with a status of CAIRO_STATUS_PATTERN_TYPE_MISMATCH.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]