cairoPatternCreateRadial {RGtk2}	R Documentation

cairoPatternCreateRadial

Description

Creates a new radial gradient CairoPattern between the two circles defined by (cx0, cy0, radius0) and (cx1, cy1, radius1). Before using the gradient pattern, a number of color stops should be defined using cairoPatternAddColorStopRgb or cairoPatternAddColorStopRgba.

Usage

cairoPatternCreateRadial(cx0, cy0, radius0, cx1, cy1, radius1)

Arguments

cx0
[numeric] x coordinate for the center of the start circle

cy0
[numeric] y coordinate for the center of the start circle

radius0
[numeric] radius of the start circle

cx1
[numeric] x coordinate for the center of the end circle

cy1
[numeric] y coordinate for the center of the end circle

radius1
[numeric] radius of the end circle

Details

Note: The coordinates here are in pattern space. For a new pattern, pattern space is identical to user space, but the relationship between the spaces can be changed with cairoPatternSetMatrix.

Value

[CairoPattern] the newly created CairoPattern if successful, or an error pattern in case of no memory. This function will always return a valid pointer, but if an error occurred the pattern status will be set to an error. To inspect the status of a pattern use cairoPatternStatus.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]