cairoPatternCreateRgba {RGtk2}	R Documentation

cairoPatternCreateRgba

Description

Creates a new CairoPattern corresponding to a translucent color. The color components are floating point numbers in the range 0 to 1. If the values passed in are outside that range, they will be clamped.

Usage

cairoPatternCreateRgba(red, green, blue, alpha)

Arguments

red
[numeric] red component of the color

green
[numeric] green component of the color

blue
[numeric] blue component of the color

alpha
[numeric] alpha component of the color

Value

[CairoPattern] the newly created CairoPattern if successful, or an error pattern in case of no memory. This function will always return a valid pointer, but if an error occurred the pattern status will be set to an error. To inspect the status of a pattern use cairoPatternStatus.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]