cairoPatternGetColorStopRgba {RGtk2}	R Documentation

cairoPatternGetColorStopRgba

Description

Gets the color and offset information at the given index for a gradient pattern. Values of index are 0 to 1 less than the number returned by cairoPatternGetColorStopCount.

Usage

cairoPatternGetColorStopRgba(pattern, index)

Arguments

pattern
[CairoPattern] a CairoPattern

index
[integer] index of the stop to return data for

Details

Since 1.4

Value

A list containing the following elements:

retval
[CairoStatus] CAIRO_STATUS_SUCCESS, or CAIRO_STATUS_INVALID_INDEX if index is not valid for the given pattern. If the pattern is not a gradient pattern, CAIRO_STATUS_PATTERN_TYPE_MISMATCH is returned.

offset
[numeric] return value for the offset of the stop, or NULL

red
[numeric] return value for red component of color, or NULL

green
[numeric] return value for green component of color, or NULL

blue
[numeric] return value for blue component of color, or NULL

alpha
[numeric] return value for alpha component of color, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]