cairoPatternGetExtend {RGtk2}	R Documentation

cairoPatternGetExtend

Description

Gets the current extend mode for a pattern. See CairoExtend for details on the semantics of each extend strategy.

Usage

cairoPatternGetExtend(pattern)

Arguments

pattern
[CairoPattern] a CairoPattern

Value

[CairoExtend] the current extend strategy used for drawing the pattern.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]