cairoPatternSetExtend {RGtk2}	R Documentation

cairoPatternSetExtend

Description

Sets the mode to be used for drawing outside the area of a pattern. See CairoExtend for details on the semantics of each extend strategy.

Usage

cairoPatternSetExtend(pattern, extend)

Arguments

pattern
[CairoPattern] a CairoPattern

extend
[CairoExtend] a CairoExtend describing how the area outside of the pattern will be drawn

Details

The default extend mode is CAIRO_EXTEND_NONE for surface patterns and CAIRO_EXTEND_PAD for gradient patterns.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]