cairoPatternSetFilter {RGtk2}	R Documentation

cairoPatternSetFilter

Description

Sets the filter to be used for resizing when using this pattern. See CairoFilter for details on each filter.

Usage

cairoPatternSetFilter(pattern, filter)

Arguments

pattern
[CairoPattern] a CairoPattern

filter
[CairoFilter] a CairoFilter describing the filter to use for resizing the pattern

Details

* Note that you might want to control filtering even when you do not have an explicit CairoPattern object, (for example when using cairoSetSourceSurface). In these cases, it is convenient to use cairoGetSource to get access to the pattern that cairo creates implicitly. For example:

cr$setSourceSurface(image, x, y)
cr$getSource()$setFilter(CairoFilter["nearest"])

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]