cairoPdfSurfaceSetSize {RGtk2}	R Documentation

cairoPdfSurfaceSetSize

Description

Changes the size of a PDF surface for the current (and subsequent) pages.

Usage

cairoPdfSurfaceSetSize(surface, width.in.points, height.in.points)

Arguments

surface
[CairoSurface] a PDF CairoSurface

width.in.points
[numeric] new surface width, in points (1 point == 1/72.0 inch)

height.in.points
[numeric] new surface height, in points (1 point == 1/72.0 inch)

Details

This function should only be called before any drawing operations have been performed on the current page. The simplest way to do this is to call this function immediately after creating the surface or immediately after completing a page with either cairoShowPage or cairoCopyPage. Since 1.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]