cairoPopGroup {RGtk2}	R Documentation

cairoPopGroup

Description

Terminates the redirection begun by a call to cairoPushGroup or cairoPushGroupWithContent and returns a new pattern containing the results of all drawing operations performed to the group.

Usage

cairoPopGroup(cr)

Arguments

cr
[Cairo] a cairo context

Details

The cairoPopGroup function calls cairoRestore, (balancing a call to cairoSave by the push_group function), so that any changes to the graphics state will not be visible outside the group. Since 1.2

Value

[CairoPattern] a newly created (surface) pattern containing the results of all drawing operations performed to the group.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]