cairoPsLevelToString {RGtk2}	R Documentation

cairoPsLevelToString

Description

Get the string representation of the given level id. This function will return NULL if level id isn't valid. See cairoPsGetLevels for a way to get the list of valid level ids.

Usage

cairoPsLevelToString(level)

Arguments

level
[CairoPsLevel] a level id

Details

Since 1.6

Value

[char] the string associated to given level.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]