cairoPsSurfaceCreateForStream {RGtk2}	R Documentation

cairoPsSurfaceCreateForStream

Description

Creates a PostScript surface of the specified size in points to be written incrementally to the stream represented by write.func and closure. See cairoPsSurfaceCreate for a more convenient way to simply direct the PostScript output to a named file.

Usage

cairoPsSurfaceCreateForStream(write.func, closure, width.in.points, 
    height.in.points)

Arguments

write.func
[CairoWriteFunc] a CairoWriteFunc to accept the output data, may be NULL to indicate a no-op write.func. With a no-op write.func, the surface may be queried or used as a source without generating any temporary files.

closure
[R object] the closure argument for write.func

width.in.points
[numeric] width of the surface, in points (1 point == 1/72.0 inch)

height.in.points
[numeric] height of the surface, in points (1 point == 1/72.0 inch)

Details

Note that the size of individual pages of the PostScript output can vary. See cairoPsSurfaceSetSize. Since 1.2

Value

[CairoSurface] a pointer to the newly created surface. The caller owns the surface and should call cairoSurfaceDestroy when done with it. This function always returns a valid pointer, but it will return a pointer to a "nil" surface if an error such as out of memory occurs. You can use cairoSurfaceStatus to check for this.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]