cairoPsSurfaceRestrictToLevel {RGtk2}	R Documentation

cairoPsSurfaceRestrictToLevel

Description

Restricts the generated PostSript file to level. See cairoPsGetLevels for a list of available level values that can be used here.

Usage

cairoPsSurfaceRestrictToLevel(surface, level)

Arguments

surface
[CairoSurface] a PostScript CairoSurface

level
[CairoPsLevel] PostScript level

Details

This function should only be called before any drawing operations have been performed on the given surface. The simplest way to do this is to call this function immediately after creating the surface. Since 1.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]