cairoPsSurfaceSetEps {RGtk2}	R Documentation

cairoPsSurfaceSetEps

Description

If eps is TRUE, the PostScript surface will output Encapsulated PostScript.

Usage

cairoPsSurfaceSetEps(surface, eps)

Arguments

surface
[CairoSurface] a PostScript CairoSurface

eps
[logical] TRUE to output EPS format PostScript

Details

This function should only be called before any drawing operations have been performed on the current page. The simplest way to do this is to call this function immediately after creating the surface. An Encapsulated PostScript file should never contain more than one page. Since 1.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]