cairoPushGroupWithContent {RGtk2}	R Documentation

cairoPushGroupWithContent

Description

Temporarily redirects drawing to an intermediate surface known as a group. The redirection lasts until the group is completed by a call to cairoPopGroup or cairoPopGroupToSource. These calls provide the result of any drawing to the group as a pattern, (either as an explicit object, or set as the source pattern).

Usage

cairoPushGroupWithContent(cr, content)

Arguments

cr
[Cairo] a cairo context

content
[CairoContent] a CairoContent indicating the type of group that will be created

Details

The group will have a content type of content. The ability to control this content type is the only distinction between this function and cairoPushGroup which you should see for a more detailed description of group rendering. Since 1.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]