cairoRectangle {RGtk2}	R Documentation

cairoRectangle

Description

Adds a closed sub-path rectangle of the given size to the current path at position (x, y) in user-space coordinates.

Usage

cairoRectangle(cr, x, y, width, height)

Arguments

cr
[Cairo] a cairo context

x
[numeric] the X coordinate of the top left corner of the rectangle

y
[numeric] the Y coordinate to the top left corner of the rectangle

width
[numeric] the width of the rectangle

height
[numeric] the height of the rectangle

Details

This function is logically equivalent to:

cr$moveTo(x, y)
cr$relLineTo(width, 0)
cr$relLineTo(0, height)
cr$relLineTo(-width, 0)
cr$closePath()

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]