cairoRelCurveTo {RGtk2}	R Documentation

cairoRelCurveTo

Description

Relative-coordinate version of cairoCurveTo. All offsets are relative to the current point. Adds a cubic B<U+00E9>zier spline to the path from the current point to a point offset from the current point by (dx3, dy3), using points offset by (dx1, dy1) and (dx2, dy2) as the control points. After this call the current point will be offset by (dx3, dy3).

Usage

cairoRelCurveTo(cr, dx1, dy1, dx2, dy2, dx3, dy3)

Arguments

cr
[Cairo] a cairo context

dx1
[numeric] the X offset to the first control point

dy1
[numeric] the Y offset to the first control point

dx2
[numeric] the X offset to the second control point

dy2
[numeric] the Y offset to the second control point

dx3
[numeric] the X offset to the end of the curve

dy3
[numeric] the Y offset to the end of the curve

Details

Given a current point of (x, y), cairo_rel_curve_to(cr, dx1, dy1, dx2, dy2, dx3, dy3) is logically equivalent to cairo_curve_to(cr, x+dx1, y+dy1, x+dx2, y+dy2, x+dx3, y+dy3).

It is an error to call this function with no current point. Doing so will cause cr to shutdown with a status of CAIRO_STATUS_NO_CURRENT_POINT.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]