cairoRelMoveTo {RGtk2}	R Documentation

cairoRelMoveTo

Description

Begin a new sub-path. After this call the current point will offset by (x, y).

Usage

cairoRelMoveTo(cr, dx, dy)

Arguments

cr
[Cairo] a cairo context

dx
[numeric] the X offset

dy
[numeric] the Y offset

Details

Given a current point of (x, y), cairo_rel_move_to(cr, dx, dy) is logically equivalent to cairo_move_to(cr, x + dx, y + dy).

It is an error to call this function with no current point. Doing so will cause cr to shutdown with a status of CAIRO_STATUS_NO_CURRENT_POINT.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]