cairoResetClip {RGtk2}	R Documentation

cairoResetClip

Description

Reset the current clip region to its original, unrestricted state. That is, set the clip region to an infinitely large shape containing the target surface. Equivalently, if infinity is too hard to grasp, one can imagine the clip region being reset to the exact bounds of the target surface.

Usage

cairoResetClip(cr)

Arguments

cr
[Cairo] a cairo context

Details

Note that code meant to be reusable should not call cairoResetClip as it will cause results unexpected by higher-level code which calls cairoClip. Consider using cairoSave and cairoRestore around cairoClip as a more robust means of temporarily restricting the clip region.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]