cairoSave {RGtk2}	R Documentation

cairoSave

Description

Makes a copy of the current state of cr and saves it on an internal stack of saved states for cr. When cairoRestore is called, cr will be restored to the saved state. Multiple calls to cairoSave and cairoRestore can be nested; each call to cairoRestore restores the state from the matching paired cairoSave.

Usage

cairoSave(cr)

Arguments

cr
[Cairo] a Cairo

Details

It isn't necessary to clear all saved states before a Cairo is freed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]