cairoScaledFontCreate {RGtk2}	R Documentation

cairoScaledFontCreate

Description

Creates a CairoScaledFont object from a font face and matrices that describe the size of the font and the environment in which it will be used.

Usage

cairoScaledFontCreate(font.face, font.matrix, ctm, option)

Arguments

font.face
[CairoFontFace] a CairoFontFace

font.matrix
[CairoMatrix] font space to user space transformation matrix for the font. In the simplest case of a N point font, this matrix is just a scale by N, but it can also be used to shear the font or stretch it unequally along the two axes. See cairoSetFontMatrix.

ctm
[CairoMatrix] user to device transformation matrix with which the font will be used.

option
[CairoFontOptions] options to use when getting metrics for the font and rendering with it.

Value

[CairoScaledFont] a newly created CairoScaledFont.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]