cairoScaledFontTextExtents {RGtk2}	R Documentation

cairoScaledFontTextExtents

Description

Gets the extents for a string of text. The extents describe a user-space rectangle that encloses the "inked" portion of the text drawn at the origin (0,0) (as it would be drawn by cairoShowText if the cairo graphics state were set to the same font_face, font_matrix, ctm, and font_options as scaled.font). Additionally, the x_advance and y_advance values indicate the amount by which the current point would be advanced by cairoShowText.

Usage

cairoScaledFontTextExtents(scaled.font, utf8)

Arguments

scaled.font
[CairoScaledFont] a CairoScaledFont

utf8
[char] a string of text, encoded in UTF-8

Details

Note that whitespace characters do not directly contribute to the size of the rectangle (extents.width and extents.height). They do contribute indirectly by changing the position of non-whitespace characters. In particular, trailing whitespace characters are likely to not affect the size of the rectangle, though they will affect the x_advance and y_advance values. Since 1.2

Value

A list containing the following elements:

extents
[CairoTextExtents] a CairoTextExtents which to store the retrieved extents.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]