cairoSelectFontFace {RGtk2}	R Documentation

cairoSelectFontFace

Description

Note: The cairoSelectFontFace function call is part of what the cairo designers call the "toy" text API. It is convenient for short demos and simple programs, but it is not expected to be adequate for serious text-using applications.

Usage

cairoSelectFontFace(cr, family, slant, weight)

Arguments

cr
[Cairo] a Cairo

family
[char] a font family name, encoded in UTF-8

slant
[CairoFontSlant] the slant for the font

weight
[CairoFontWeight] the weight for the font

Details

Selects a family and style of font from a simplified description as a family name, slant and weight. Cairo provides no operation to list available family names on the system (this is a "toy", remember), but the standard CSS2 generic family names, ("serif", "sans-serif", "cursive", "fantasy", "monospace"), are likely to work as expected.

For "real" font selection, see the font-backend-specific font_face_create functions for the font backend you are using. (For example, if you are using the freetype-based cairo-ft font backend, see cairoFtFontFaceCreateForFtFace() or cairoFtFontFaceCreateForPattern().) The resulting font face could then be used with cairoScaledFontCreate and cairoSetScaledFont.

Similarly, when using the "real" font support, you can call directly into the underlying font system, (such as fontconfig or freetype), for operations such as listing available fonts, etc.

It is expected that most applications will need to use a more comprehensive font handling and text layout library, (for example, pango), in conjunction with cairo.

If text is drawn without a call to cairoSelectFontFace, (nor cairoSetFontFace nor cairoSetScaledFont), the default family is platform-specific, but is essentially "sans-serif". Default slant is CAIRO_FONT_SLANT_NORMAL, and default weight is CAIRO_FONT_WEIGHT_NORMAL.

This function is equivalent to a call to cairoToyFontFaceCreate followed by cairoSetFontFace.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]