cairoSetFontFace {RGtk2}	R Documentation

cairoSetFontFace

Description

Replaces the current CairoFontFace object in the Cairo with font.face. The replaced font face in the Cairo will be destroyed if there are no other references to it.

Usage

cairoSetFontFace(cr, font.face)

Arguments

cr
[Cairo] a Cairo

font.face
[CairoFontFace] a CairoFontFace, or NULL to restore to the default font

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]