cairoSetFontMatrix {RGtk2}	R Documentation

cairoSetFontMatrix

Description

Sets the current font matrix to matrix. The font matrix gives a transformation from the design space of the font (in this space, the em-square is 1 unit by 1 unit) to user space. Normally, a simple scale is used (see cairoSetFontSize), but a more complex font matrix can be used to shear the font or stretch it unequally along the two axes

Usage

cairoSetFontMatrix(cr, matrix)

Arguments

cr
[Cairo] a Cairo

matrix
[CairoMatrix] a CairoMatrix describing a transform to be applied to the current font.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]