cairoSetFontOptions {RGtk2}	R Documentation

cairoSetFontOptions

Description

Sets a set of custom font rendering options for the Cairo. Rendering options are derived by merging these options with the options derived from underlying surface; if the value in options has a default value (like CAIRO_ANTIALIAS_DEFAULT), then the value from the surface is used.

Usage

cairoSetFontOptions(cr, options)

Arguments

cr
[Cairo] a Cairo

options
[CairoFontOptions] font options to use

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]