cairoSetLineJoin {RGtk2}	R Documentation

cairoSetLineJoin

Description

Sets the current line join style within the cairo context. See CairoLineJoin for details about how the available line join styles are drawn.

Usage

cairoSetLineJoin(cr, line.join)

Arguments

cr
[Cairo] a cairo context

line.join
[CairoLineJoin] a line join style

Details

As with the other stroke parameters, the current line join style is examined by cairoStroke, cairoStrokeExtents, and cairoStrokeToPath(), but does not have any effect during path construction.

The default line join style is CAIRO_LINE_JOIN_MITER.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]