cairoSetLineWidth {RGtk2}	R Documentation

cairoSetLineWidth

Description

Sets the current line width within the cairo context. The line width value specifies the diameter of a pen that is circular in user space, (though device-space pen may be an ellipse in general due to scaling/shear/rotation of the CTM).

Usage

cairoSetLineWidth(cr, width)

Arguments

cr
[Cairo] a Cairo

width
[numeric] a line width

Details

Note: When the description above refers to user space and CTM it refers to the user space and CTM in effect at the time of the stroking operation, not the user space and CTM in effect at the time of the call to cairoSetLineWidth. The simplest usage makes both of these spaces identical. That is, if there is no change to the CTM between a call to cairoSetLineWidth and the stroking operation, then one can just pass user-space values to cairoSetLineWidth and ignore this note.

As with the other stroke parameters, the current line width is examined by cairoStroke, cairoStrokeExtents, and cairoStrokeToPath(), but does not have any effect during path construction.

The default line width value is 2.0.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]