cairoSetSource {RGtk2}	R Documentation

cairoSetSource

Description

Sets the source pattern within cr to source. This pattern will then be used for any subsequent drawing operation until a new source pattern is set.

Usage

cairoSetSource(cr, source)

Arguments

cr
[Cairo] a cairo context

source
[CairoPattern] a CairoPattern to be used as the source for subsequent drawing operations.

Details

Note: The pattern's transformation matrix will be locked to the user space in effect at the time of cairoSetSource. This means that further modifications of the current transformation matrix will not affect the source pattern. See cairoPatternSetMatrix.

The default source pattern is a solid pattern that is opaque black, (that is, it is equivalent to cairo_set_source_rgb(cr, 0.0, 0.0, 0.0)).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]