cairoSetSourceRgb {RGtk2}	R Documentation

cairoSetSourceRgb

Description

Sets the source pattern within cr to an opaque color. This opaque color will then be used for any subsequent drawing operation until a new source pattern is set.

Usage

cairoSetSourceRgb(cr, red, green, blue)

Arguments

cr
[Cairo] a cairo context

red
[numeric] red component of color

green
[numeric] green component of color

blue
[numeric] blue component of color

Details

The color components are floating point numbers in the range 0 to 1. If the values passed in are outside that range, they will be clamped.

The default source pattern is opaque black, (that is, it is equivalent to cairo_set_source_rgb(cr, 0.0, 0.0, 0.0)).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]