cairoSetSourceRgba {RGtk2}	R Documentation

cairoSetSourceRgba

Description

Sets the source pattern within cr to a translucent color. This color will then be used for any subsequent drawing operation until a new source pattern is set.

Usage

cairoSetSourceRgba(cr, red, green, blue, alpha)

Arguments

cr
[Cairo] a cairo context

red
[numeric] red component of color

green
[numeric] green component of color

blue
[numeric] blue component of color

alpha
[numeric] alpha component of color

Details

The color and alpha components are floating point numbers in the range 0 to 1. If the values passed in are outside that range, they will be clamped.

The default source pattern is opaque black, (that is, it is equivalent to cairo_set_source_rgba(cr, 0.0, 0.0, 0.0, 1.0)).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]