cairoSetSourceSurface {RGtk2}	R Documentation

cairoSetSourceSurface

Description

This is a convenience function for creating a pattern from surface and setting it as the source in cr with cairoSetSource.

Usage

cairoSetSourceSurface(cr, surface, x, y)

Arguments

cr
[Cairo] a cairo context

surface
[CairoSurface] a surface to be used to set the source pattern

x
[numeric] User-space X coordinate for surface origin

y
[numeric] User-space Y coordinate for surface origin

Details

The x and y parameters give the user-space coordinate at which the surface origin should appear. (The surface origin is its upper-left corner before any transformation has been applied.) The x and y patterns are negated and then set as translation values in the pattern matrix.

Other than the initial translation pattern matrix, as described above, all other pattern attributes, (such as its extend mode), are set to the default values as in cairoPatternCreateForSurface. The resulting pattern can be queried with cairoGetSource so that these attributes can be modified if desired, (eg. to create a repeating pattern with cairoPatternSetExtend).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]