cairoShowGlyphs {RGtk2}	R Documentation

cairoShowGlyphs

Description

A drawing operator that generates the shape from a list of glyphs, rendered according to the current font face, font size (font matrix), and font options.

Usage

cairoShowGlyphs(cr, glyphs, num.glyphs)

Arguments

cr
[Cairo] a cairo context

glyphs
[CairoGlyph] list of glyphs to show

num.glyphs
[integer] number of glyphs to show

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]