cairoShowPage {RGtk2}	R Documentation

cairoShowPage

Description

Emits and clears the current page for backends that support multiple pages. Use cairoCopyPage if you don't want to clear the page.

Usage

cairoShowPage(cr)

Arguments

cr
[Cairo] a cairo context

Details

This is a convenience function that simply calls cairoSurfaceShowPage on cr's target.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]