cairoShowTextGlyphs {RGtk2}	R Documentation

cairoShowTextGlyphs

Description

This operation has rendering effects similar to cairoShowGlyphs but, if the target surface supports it, uses the provided text and cluster mapping to embed the text for the glyphs shown in the output. If the target does not support the extended attributes, this function acts like the basic cairoShowGlyphs as if it had been passed glyphs and num.glyphs.

Usage

cairoShowTextGlyphs(cr, utf8, glyphs, clusters, cluster.flags)

Arguments

cr
[Cairo] a cairo context

utf8
[char] a string of text encoded in UTF-8

glyphs
[CairoGlyph] list of glyphs to show

clusters
[CairoTextCluster] list of cluster mapping information

cluster.flags
[CairoTextClusterFlags] cluster mapping flags

Details

The mapping between utf8 and glyphs is provided by a list of clusters. Each cluster covers a number of text bytes and glyphs, and neighboring clusters cover neighboring areas of utf8 and glyphs. The clusters should collectively cover utf8 and glyphs in entirety.

The first cluster always covers bytes from the beginning of utf8. If cluster.flags do not have the CAIRO_TEXT_CLUSTER_FLAG_BACKWARD set, the first cluster also covers the beginning of glyphs, otherwise it covers the end of the glyphs list and following clusters move backward.

See CairoTextCluster for constraints on valid clusters. Since 1.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]