cairoStrokeExtents {RGtk2}	R Documentation

cairoStrokeExtents

Description

Computes a bounding box in user coordinates covering the area that would be affected, (the "inked" area), by a cairoStroke operation given the current path and stroke parameters. If the current path is empty, returns an empty rectangle ((0,0), (0,0)). Surface dimensions and clipping are not taken into account.

Usage

cairoStrokeExtents(cr)

Arguments

cr
[Cairo] a cairo context

Details

Note that if the line width is set to exactly zero, then cairoStrokeExtents will return an empty rectangle. Contrast with cairoPathExtents which can be used to compute the non-empty bounds as the line width approaches zero.

Note that cairoStrokeExtents must necessarily do more work to compute the precise inked areas in light of the stroke parameters, so cairoPathExtents may be more desirable for sake of performance if non-inked path extents are desired.

See cairoStroke, cairoSetLineWidth, cairoSetLineJoin, cairoSetLineCap, cairoSetDash, and cairoStrokePreserve.

Value

A list containing the following elements:

x1
[numeric] left of the resulting extents

y1
[numeric] top of the resulting extents

x2
[numeric] right of the resulting extents

y2
[numeric] bottom of the resulting extents

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]