cairoStrokePreserve {RGtk2}	R Documentation

cairoStrokePreserve

Description

A drawing operator that strokes the current path according to the current line width, line join, line cap, and dash settings. Unlike cairoStroke, cairoStrokePreserve preserves the path within the cairo context.

Usage

cairoStrokePreserve(cr)

Arguments

cr
[Cairo] a cairo context

Details

See cairoSetLineWidth, cairoSetLineJoin, cairoSetLineCap, cairoSetDash, and cairoStrokePreserve.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]