cairoSurfaceCopyPage {RGtk2}	R Documentation

cairoSurfaceCopyPage

Description

Emits the current page for backends that support multiple pages, but doesn't clear it, so that the contents of the current page will be retained for the next page. Use cairoSurfaceShowPage if you want to get an empty page after the emission.

Usage

cairoSurfaceCopyPage(surface)

Arguments

surface
[CairoSurface] a CairoSurface

Details

There is a convenience function for this that takes a Cairo, namely cairoCopyPage. Since 1.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]