cairoSurfaceCreateSimilar {RGtk2}	R Documentation

cairoSurfaceCreateSimilar

Description

Create a new surface that is as compatible as possible with an existing surface. For example the new surface will have the same fallback resolution and font options as other. Generally, the new surface will also use the same backend as other, unless that is not possible for some reason. The type of the returned surface may be examined with cairoSurfaceGetType.

Usage

cairoSurfaceCreateSimilar(other, content, width, height)

Arguments

other
[CairoSurface] an existing surface used to select the backend of the new surface

content
[CairoContent] the content for the new surface

width
[integer] width of the new surface, (in device-space units)

height
[integer] height of the new surface (in device-space units)

Details

Initially the surface contents are all 0 (transparent if contents have transparency, black otherwise.)

Value

[CairoSurface] a pointer to the newly allocated surface. The caller owns the surface and should call cairoSurfaceDestroy when done with it. This function always returns a valid pointer, but it will return a pointer to a "nil" surface if other is already in an error state or any other error occurs.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]