cairoSurfaceFlush {RGtk2}	R Documentation

cairoSurfaceFlush

Description

Do any pending drawing for the surface and also restore any temporary modification's cairo has made to the surface's state. This function must be called before switching from drawing on the surface with cairo to drawing on it directly with native APIs. If the surface doesn't support direct access, then this function does nothing.

Usage

cairoSurfaceFlush(surface)

Arguments

surface
[CairoSurface] a CairoSurface

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]