cairoSurfaceGetFontOptions {RGtk2}	R Documentation

cairoSurfaceGetFontOptions

Description

Retrieves the default font rendering options for the surface. This allows display surfaces to report the correct subpixel order for rendering on them, print surfaces to disable hinting of metrics and so forth. The result can then be used with cairoScaledFontCreate.

Usage

cairoSurfaceGetFontOptions(surface)

Arguments

surface
[CairoSurface] a CairoSurface

Value

A list containing the following elements:

options
[CairoFontOptions] a CairoFontOptions object into which to store the retrieved options. All existing values are overwritten

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]