cairoSurfaceGetUserData {RGtk2}	R Documentation

cairoSurfaceGetUserData

Description

Return user data previously attached to surface using the specified key. If no user data has been attached with the given key this function returns NULL.

Usage

cairoSurfaceGetUserData(surface, key)

Arguments

surface
[CairoSurface] a CairoSurface

key
[CairoUserDataKey] the the CairoUserDataKey the user data was attached to

Value

[R object] the user data previously attached or NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]