cairoSurfaceHasShowTextGlyphs {RGtk2}	R Documentation

cairoSurfaceHasShowTextGlyphs

Description

Returns whether the surface supports sophisticated cairoShowTextGlyphs operations. That is, whether it actually uses the provided text and cluster data to a cairoShowTextGlyphs call.

Usage

cairoSurfaceHasShowTextGlyphs(surface)

Arguments

surface
[CairoSurface] a CairoSurface

Details

Note: Even if this function returns FALSE, a cairoShowTextGlyphs operation targeted at surface will still succeed. It just will act like a cairoShowGlyphs operation. Users can use this function to avoid computing UTF-8 text and cluster mapping if the target surface does not use it. Since 1.8

Value

[logical] TRUE if surface supports cairoShowTextGlyphs, FALSE otherwise

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]