cairoSurfaceMarkDirty {RGtk2}	R Documentation

cairoSurfaceMarkDirty

Description

Tells cairo that drawing has been done to surface using means other than cairo, and that cairo should reread any cached areas. Note that you must call cairoSurfaceFlush before doing such drawing.

Usage

cairoSurfaceMarkDirty(surface)

Arguments

surface
[CairoSurface] a CairoSurface

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]