cairoSurfaceMarkDirtyRectangle {RGtk2}	R Documentation

cairoSurfaceMarkDirtyRectangle

Description

Like cairoSurfaceMarkDirty, but drawing has been done only to the specified rectangle, so that cairo can retain cached contents for other parts of the surface.

Usage

cairoSurfaceMarkDirtyRectangle(surface, x, y, width, height)

Arguments

surface
[CairoSurface] a CairoSurface

x
[integer] X coordinate of dirty rectangle

y
[integer] Y coordinate of dirty rectangle

width
[integer] width of dirty rectangle

height
[integer] height of dirty rectangle

Details

Any cached clip set on the surface will be reset by this function, to make sure that future cairo calls have the clip set that they expect.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]