cairoSurfaceSetDeviceOffset {RGtk2}	R Documentation

cairoSurfaceSetDeviceOffset

Description

Sets an offset that is added to the device coordinates determined by the CTM when drawing to surface. One use case for this function is when we want to create a CairoSurface that redirects drawing for a portion of an onscreen surface to an offscreen surface in a way that is completely invisible to the user of the cairo API. Setting a transformation via cairoTranslate isn't sufficient to do this, since functions like cairoDeviceToUser will expose the hidden offset.

Usage

cairoSurfaceSetDeviceOffset(surface, x.offset, y.offset)

Arguments

surface
[CairoSurface] a CairoSurface

x.offset
[numeric] the offset in the X direction, in device units

y.offset
[numeric] the offset in the Y direction, in device units

Details

Note that the offset affects drawing to the surface as well as using the surface in a source pattern.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]