cairoSurfaceShowPage {RGtk2}	R Documentation

cairoSurfaceShowPage

Description

Emits and clears the current page for backends that support multiple pages. Use cairoSurfaceCopyPage if you don't want to clear the page.

Usage

cairoSurfaceShowPage(surface)

Arguments

surface
[CairoSurface] a cairo_Surface_t

Details

There is a convenience function for this that takes a Cairo, namely cairoShowPage. Since 1.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]