cairoSurfaceWriteToPng {RGtk2}	R Documentation

cairoSurfaceWriteToPng

Description

Writes the contents of surface to a new file filename as a PNG image.

Usage

cairoSurfaceWriteToPng(surface, filename)

Arguments

surface
[CairoSurface] a CairoSurface with pixel contents

filename
[char] the name of a file to write to

Value

[CairoStatus] CAIRO_STATUS_SUCCESS if the PNG file was written successfully. Otherwise, CAIRO_STATUS_NO_MEMORY if memory could not be allocated for the operation or CAIRO_STATUS_SURFACE_TYPE_MISMATCH if the surface does not have pixel contents, or CAIRO_STATUS_WRITE_ERROR if an I/O error occurs while attempting to write the file.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]