cairoSurfaceWriteToPngStream {RGtk2}	R Documentation

cairoSurfaceWriteToPngStream

Description

Writes the image surface to the write function.

Usage

cairoSurfaceWriteToPngStream(surface, con)

Arguments

surface
[CairoSurface] a CairoSurface with pixel contents

Value

[CairoStatus] CAIRO_STATUS_SUCCESS if the PNG file was written successfully. Otherwise, CAIRO_STATUS_NO_MEMORY is returned if memory could not be allocated for the operation, CAIRO_STATUS_SURFACE_TYPE_MISMATCH if the surface does not have pixel contents.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]