cairoSvgSurfaceCreate {RGtk2}	R Documentation

cairoSvgSurfaceCreate

Description

Creates a SVG surface of the specified size in points to be written to filename.

Usage

cairoSvgSurfaceCreate(filename, width.in.points, height.in.points)

Arguments

filename
[char] a filename for the SVG output (must be writable), NULL may be used to specify no output. This will generate a SVG surface that may be queried and used as a source, without generating a temporary file.

width.in.points
[numeric] width of the surface, in points (1 point == 1/72.0 inch)

height.in.points
[numeric] height of the surface, in points (1 point == 1/72.0 inch)

Details

Since 1.2

Value

[CairoSurface] a pointer to the newly created surface. The caller owns the surface and should call cairoSurfaceDestroy when done with it. This function always returns a valid pointer, but it will return a pointer to a "nil" surface if an error such as out of memory occurs. You can use cairoSurfaceStatus to check for this.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]