cairoSvgSurfaceRestrictToVersion {RGtk2}	R Documentation

cairoSvgSurfaceRestrictToVersion

Description

Restricts the generated SVG file to version. See cairoSvgGetVersions for a list of available version values that can be used here.

Usage

cairoSvgSurfaceRestrictToVersion(surface, version)

Arguments

surface
[CairoSurface] a SVG CairoSurface

version
[CairoSvgVersion] SVG version

Details

This function should only be called before any drawing operations have been performed on the given surface. The simplest way to do this is to call this function immediately after creating the surface. Since 1.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]