cairoSvgVersionToString {RGtk2}	R Documentation

cairoSvgVersionToString

Description

Get the string representation of the given version id. This function will return NULL if version isn't valid. See cairoSvgGetVersions for a way to get the list of valid version ids.

Usage

cairoSvgVersionToString(version)

Arguments

version
[CairoSvgVersion] a version id

Details

Since 1.2

Value

[char] the string associated to given version.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]