cairoTextExtents {RGtk2}	R Documentation

cairoTextExtents

Description

Gets the extents for a string of text. The extents describe a user-space rectangle that encloses the "inked" portion of the text, (as it would be drawn by cairoShowText). Additionally, the x_advance and y_advance values indicate the amount by which the current point would be advanced by cairoShowText.

Usage

cairoTextExtents(cr, utf8)

Arguments

cr
[Cairo] a Cairo

utf8
[char] a string of text encoded in UTF-8, or NULL

Details

Note that whitespace characters do not directly contribute to the size of the rectangle (extents.width and extents.height). They do contribute indirectly by changing the position of non-whitespace characters. In particular, trailing whitespace characters are likely to not affect the size of the rectangle, though they will affect the x_advance and y_advance values.

Value

A list containing the following elements:

extents
[CairoTextExtents] a CairoTextExtents object into which the results will be stored

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]