cairoTextPath {RGtk2}	R Documentation

cairoTextPath

Description

Adds closed paths for text to the current path. The generated path if filled, achieves an effect similar to that of cairoShowText.

Usage

cairoTextPath(cr, utf8)

Arguments

cr
[Cairo] a cairo context

utf8
[char] a string of text encoded in UTF-8, or NULL

Details

Text conversion and positioning is done similar to cairoShowText.

Like cairoShowText, After this call the current point is moved to the origin of where the next glyph would be placed in this same progression. That is, the current point will be at the origin of the final glyph offset by its advance values. This allows for chaining multiple calls to to cairoTextPath without having to set current point in between.

Note: The cairoTextPath function call is part of what the cairo designers call the "toy" text API. It is convenient for short demos and simple programs, but it is not expected to be adequate for serious text-using applications. See cairoGlyphPath for the "real" text path API in cairo.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]