cairoToyFontFaceCreate {RGtk2}	R Documentation

cairoToyFontFaceCreate

Description

Creates a font face from a triplet of family, slant, and weight. These font faces are used in implementation of the the Cairo "toy" font API.

Usage

cairoToyFontFaceCreate(family, slant, weight)

Arguments

family
[char] a font family name, encoded in UTF-8

slant
[CairoFontSlant] the slant for the font

weight
[CairoFontWeight] the weight for the font

Details

If family is the zero-length string "", the platform-specific default family is assumed. The default family then can be queried using cairoToyFontFaceGetFamily.

The cairoSelectFontFace function uses this to create font faces. See that function for limitations of toy font faces. Since 1.8

Value

[CairoFontFace] a newly created CairoFontFace.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]