cairoUserFontFaceCreate {RGtk2}	R Documentation

cairoUserFontFaceCreate

Description

Creates a new user font-face.

Usage

cairoUserFontFaceCreate()

Details

Use the setter functions to associate callbacks with the returned user font. The only mandatory callback is render_glyph.

After the font-face is created, the user can attach arbitrary data (the actual font data) to it using cairoFontFaceSetUserData and access it from the user-font callbacks by using cairoScaledFontGetFontFace followed by cairoFontFaceGetUserData. Since 1.8

Value

[CairoFontFace] a newly created CairoFontFace.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]