cairoUserFontFaceSetRenderGlyphFunc {RGtk2}	R Documentation

cairoUserFontFaceSetRenderGlyphFunc

Description

Sets the glyph rendering function of a user-font. See cairo_user_scaled_font_render_glyph_func_t for details of how the callback works.

Usage

cairoUserFontFaceSetRenderGlyphFunc(font.face, render.glyph.func)

Arguments

font.face
[CairoFontFace] A user font face

render.glyph.func
[cairo_user_scaled_font_render_glyph_func_t] The render_glyph callback, or NULL

Details

The font-face should not be immutable or a CAIRO_STATUS_USER_FONT_IMMUTABLE error will occur. A user font-face is immutable as soon as a scaled-font is created from it.

The render_glyph callback is the only mandatory callback of a user-font. If the callback is NULL and a glyph is tried to be rendered using font.face, a CAIRO_STATUS_USER_FONT_ERROR will occur. Since 1.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]