cairoVersion {RGtk2}	R Documentation

cairoVersion

Description

Returns the version of the cairo library encoded in a single integer as per CAIRO_VERSION_ENCODE. The encoding ensures that later versions compare greater than earlier versions.

Usage

cairoVersion()

Details

A run-time comparison to check that cairo's version is greater than or equal to version X.Y.Z could be performed as follows:

# This function is not that useful in R.
# Instead, please use cairoVersionString() with compareVersion().

See also cairoVersionString as well as the compile-time equivalents CAIRO_VERSION and CAIRO_VERSION_STRING.

Value

[integer] the encoded version.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]