CairoWriteFunc {RGtk2}	R Documentation

CairoWriteFunc

Description

A callback used by cairo for outputting graphics data in formats like svg and png.

Signature

CairoWriteFunc(data, length, user_data): The type of function which is called when a backend needs to write data to an output stream. It is passed the closure which was specified by the user at the time the write function was registered, the data to write and the length of the data in bytes. The write function should return CAIRO_STATUS_SUCCESS if all the data was successfully written, CAIRO_STATUS_WRITE_ERROR otherwise.

data

[raw] the buffer containing the data to write

length

[numeric] the amount of data to write

user_data

[R object] user-supplied data

Author(s)

Michael Lawrence

References

http://www.cairographics.org/manual/cairo-PNG-Support.html

[Package RGtk2 version 2.20.31 Index]