checkGTK {RGtk2}	R Documentation

Bound versions

Description

These functions are for querying (bound*) and checking (check*) the bound versions of the libraries (GTK, Pango and Cairo). As of RGtk2 2.20.19, the check* functions are deprecated in favor of the more explicit boundVersion() >= version syntax.

Usage

checkGTK(version)
checkPango(version)
checkCairo(version)
boundGTKVersion()
boundPangoVersion()
boundCairoVersion()

Arguments

version
Version description to compare to the bound version, as in: boundGTKVersion() >= version.

Value

The check* functions return TRUE if version is satisfied, otherwise FALSE.

The bound* functions return a numeric_version representation of the bound library version.

Author(s)

Michael Lawrence

Examples

## instead of
# checkGTK("2.12.0")
## do this:
boundGTKVersion() >= "2.12.0"

[Package RGtk2 version 2.20.31 Index]