children {RGtk2}	R Documentation

Get child widgets of a Gtk container

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

This returns a list of the widgets that are currently contained or managed by a particular Gtk container widget.

Usage

gtkChildren(object)
gtkParent(object)

Arguments

w
the Gtk widget object whose children or parent widget(s) are to be retrieved. For gtkChildren, the object should be an object that inherits from the class GtkContainer.

Value

For gtkChildren, a list with each element corresponding to a child in the container widget.

Note

THIS STUFF IS VERY OLD AND DEPRECATED (compatibility wrappers for RGtk 1)

This is a slightly modified function from Duncan's original. It now always behaves as if the former argument FULL is TRUE.

This is an extra-ordinarily early release intended to encourage others to contribute code, etc.

Author(s)

Duncan Temple Lang

References

http://www.gtk.org http://www.omegahat.org/RGtk

See Also

gtkParent

[Package RGtk2 version 2.20.31 Index]