coercion {RGtk2}	R Documentation

Coercion of R lists to C structures

Description

This function coerces an R list to the named C type. This helps form the basis of the transparent type system.

[Package RGtk2 version 2.20.31 Index]