containers {RGtk2}	R Documentation

GtkContainer utilities

Description

This utility gets the nth child from a GtkContainer.

Usage

"[[.GtkContainer(x, child)"

Arguments

x
The GtkContainer object

child
The index of the child to retrieve from the container

Value

The widget at index 'child' in the vector of children in the GtkContainer

Author(s)

Michael Lawrence

[Package RGtk2 version 2.20.31 Index]