findWidgetByType {RGtk2}	R Documentation

Find widget in hierarchy by type

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

This function recursively descends the widget hierarchy identified by the root node win and finds the first widget that matches the specified target type.

Usage

findWidgetByType(win, gtkType = "GtkMenuBar", verbose = FALSE)

Arguments

win
the root node of the widget hierarchy whose children and descendants are to be searced.

gtkType
the name of the Gtk type we are looking for. If this is a function, it is called for each widget in the hierarchy until it returns TRUE. Each call is given the widget as its only argument and expected to return a logical value indicating whether the widget matches or not. One can use a closure and always return FALSE if one wants to iterate over all the widgets and collect the matching ones.

verbose
a logical value indicating whether to print information to the console as we recurse.

Value

The widget that matched or NULL.

Author(s)

Duncan Temple Lang <duncan@research.bell-labs.com>

References

Information on the package is available from http://www.omegahat.org/RGtk.

Information on Gtk is available from http://www.gtk.org.

See Also

gtkContainerGetChildren

[Package RGtk2 version 2.20.31 Index]