gAppInfoCanRemoveSupportsType {RGtk2}	R Documentation

gAppInfoCanRemoveSupportsType

Description

Checks if a supported content type can be removed from an application.

Usage

gAppInfoCanRemoveSupportsType(object)

Arguments

object
a GAppInfo.

Value

[logical] TRUE if it is possible to remove supported content types from a given appinfo, FALSE if not.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]