gAppInfoGetAll {RGtk2}	R Documentation

gAppInfoGetAll

Description

Gets a list of all of the applications currently registered on this system.

Usage

gAppInfoGetAll()

Details

For desktop files, this includes applications that have NoDisplay=true set or are excluded from display by means of OnlyShowIn or NotShowIn. See gAppInfoShouldShow. The returned list does not include applications which have the Hidden key set.

Value

[list] a newly allocated list of references to GAppInfos.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]