gAppInfoGetDefaultForUriScheme {RGtk2}	R Documentation

gAppInfoGetDefaultForUriScheme

Description

Gets the default application for launching applications using this URI scheme. A URI scheme is the initial part of the URI, up to but not including the ':', e.g. "http", "ftp" or "sip".

Usage

gAppInfoGetDefaultForUriScheme(uri.scheme)

Arguments

uri.scheme
a string containing a URI scheme.

Value

[GAppInfo] GAppInfo for given uri.scheme or NULL on error.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]