gAppInfoGetId {RGtk2}	R Documentation

gAppInfoGetId

Description

Gets the ID of an application. An id is a string that identifies the application. The exact format of the id is platform dependent. For instance, on Unix this is the desktop file id from the xdg menu specification.

Usage

gAppInfoGetId(object)

Arguments

object
a GAppInfo.

Details

Note that the returned ID may be NULL, depending on how the appinfo has been constructed.

Value

[char] a string containing the application's ID.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]