gAppInfoLaunch {RGtk2}	R Documentation

gAppInfoLaunch

Description

Launches the application. Passes files to the launched application as arguments, using the optional launch.context to get information about the details of the launcher (like what screen it is on). On error, error will be set accordingly.

Usage

gAppInfoLaunch(object, files, launch.context, .errwarn = TRUE)

Arguments

object
a GAppInfo

files
a list of GFile objects

launch.context
a GAppLaunchContext or NULL

.errwarn
Whether to issue a warning on error or fail silently

Details

To lauch the application without arguments pass a NULL files list.

Note that even if the launch is successful the application launched can fail to start if it runs into problems during startup. There is no way to detect this.

Some URIs can be changed when passed through a GFile (for instance unsupported uris with strange formats like mailto:), so if you have a textual uri you want to pass in as argument, consider using gAppInfoLaunchUris instead.

Value

A list containing the following elements:

retval
[logical] TRUE on successful launch, FALSE otherwise.

error
a GError

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]