gAppInfoSetAsDefaultForExtension {RGtk2}	R Documentation

gAppInfoSetAsDefaultForExtension

Description

Sets the application as the default handler for the given file extension.

Usage

gAppInfoSetAsDefaultForExtension(object, extension, .errwarn = TRUE)

Arguments

object
a GAppInfo.

extension
a string containing the file extension (without the dot).

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[logical] TRUE on success, FALSE on error.

error
a GError.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]