gAppLaunchContextLaunchFailed {RGtk2}	R Documentation

gAppLaunchContextLaunchFailed

Description

Called when an application has failed to launch, so that it can cancel the application startup notification started in gAppLaunchContextGetStartupNotifyId.

Usage

gAppLaunchContextLaunchFailed(object, startup.notify.id)

Arguments

object
a GAppLaunchContext.

startup.notify.id
the startup notification id that was returned by gAppLaunchContextGetStartupNotifyId.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]