gAsyncInitableNewAsync {RGtk2}	R Documentation

gAsyncInitableNewAsync

Description

Helper function for constructing GAsyncInitiable object. This is similar to gObjectNew but also initializes the object asyncronously.

Usage

gAsyncInitableNewAsync(object.type, io.priority, cancellable, 
    callback, user.data, ...)

Arguments

object.type
a GType supporting GAsyncInitable.

io.priority
the I/O priority of the operation.

cancellable
optional GCancellable object, NULL to ignore.

callback
a GAsyncReadyCallback to call when the initialization is finished

user.data
the data to pass to callback function

...
the value if the first property, followed by and other property value pairs, and ended by NULL.

Details

When the initialization is finished, callback will be called. You can then call gAsyncInitableNewFinish to get new object and check for any errors. Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]