gBufferedInputStreamFill {RGtk2}	R Documentation

gBufferedInputStreamFill

Description

Tries to read count bytes from the stream into the buffer. Will block during this read.

Usage

gBufferedInputStreamFill(object, count, cancellable = NULL, .errwarn = TRUE)

Arguments

object
GBufferedInputStream.

count
the number of bytes that will be read from the stream.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If count is zero, returns zero and does nothing. A value of count larger than G_MAXSSIZE will cause a G_IO_ERROR_INVALID_ARGUMENT error.

On success, the number of bytes read into the buffer is returned. It is not an error if this is not the same as the requested size, as it can happen e.g. near the end of a file. Zero is returned on end of file (or if count is zero), but never otherwise.

If count is -1 then the attempted read size is equal to the number of bytes that are required to fill the buffer.

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned. If an operation was partially finished when the operation was cancelled the partial result will be returned, without an error.

On error -1 is returned and error is set accordingly.

For the asynchronous, non-blocking, version of this function, see gBufferedInputStreamFillAsync.

Value

A list containing the following elements:

retval
[integer] the number of bytes read into stream's buffer, up to count, or -1 on error.

error
location to store the error occuring, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]