gBufferedInputStreamFillAsync {RGtk2}	R Documentation

gBufferedInputStreamFillAsync

Description

Reads data into stream's buffer asynchronously, up to count size. io.priority can be used to prioritize reads. For the synchronous version of this function, see gBufferedInputStreamFill.

Usage

gBufferedInputStreamFillAsync(object, count, io.priority = 0, 
    cancellable = NULL, callback, user.data = NULL)

Arguments

object
GBufferedInputStream.

count
the number of bytes that will be read from the stream.

io.priority
the I/O priority of the request.

cancellable
optional GCancellable object

callback
a GAsyncReadyCallback.

user.data
a R object.

Details

If count is -1 then the attempted read size is equal to the number of bytes that are required to fill the buffer.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]