gBufferedInputStreamPeekBuffer {RGtk2}	R Documentation

gBufferedInputStreamPeekBuffer

Description

Returns the buffer with the currently available bytes. The returned buffer must not be modified and will become invalid when reading from the stream or filling the buffer.

Usage

gBufferedInputStreamPeekBuffer(object)

Arguments

object
a GBufferedInputStream.

Value

A list containing the following elements:

retval
[R object] read-only buffer

count
a numeric to get the number of bytes available in the buffer.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]