gBufferedInputStreamSetBufferSize {RGtk2}	R Documentation

gBufferedInputStreamSetBufferSize

Description

Sets the size of the internal buffer of stream to size, or to the size of the contents of the buffer. The buffer can never be resized smaller than its current contents.

Usage

gBufferedInputStreamSetBufferSize(object, size)

Arguments

object
GBufferedInputStream.

size
a numeric.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]