gCancellableDisconnect {RGtk2}	R Documentation

gCancellableDisconnect

Description

Disconnects a handler from an cancellable instance similar to gSignalHandlerDisconnect but ensures that once this function returns the handler will not run anymore in any thread.

Usage

gCancellableDisconnect(object, handler.id)

Arguments

object
A GCancellable or NULL.

handler.id
Handler id of the handler to be disconnected, or 0.

Details

This avoids a race condition where a thread cancels at the same time as the cancellable operation is finished and the signal handler is removed. See "cancelled" for details on how to use this.

If cancellable is NULL or handler.id is 0 this function does nothing. Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]