gCancellableGetFd {RGtk2}	R Documentation

gCancellableGetFd

Description

Gets the file descriptor for a cancellable job. This can be used to implement cancellable operations on Unix systems. The returned fd will turn readable when cancellable is cancelled.

Usage

gCancellableGetFd(object)

Arguments

object
a GCancellable.

Details

You are not supposed to read from the fd yourself, just check for readable status. Reading to unset the readable status is done with gCancellableReset.

After a successful return from this function, you should use gCancellableReleaseFd to free up resources allocated for the returned file descriptor.

See also gCancellableMakePollfd().

Value

[integer] A valid file descriptor. -1 if the file descriptor is not supported, or on errors.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]