gCancellablePushCurrent {RGtk2}	R Documentation

gCancellablePushCurrent

Description

Pushes cancellable onto the cancellable stack. The current cancllable can then be recieved using gCancellableGetCurrent.

Usage

gCancellablePushCurrent(object)

Arguments

object
a GCancellable object

Details

This is useful when implementing cancellable operations in code that does not allow you to pass down the cancellable object.

This is typically called automatically by e.g. GFile operations, so you rarely have to call this yourself.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]