gCancellableReleaseFd {RGtk2}	R Documentation

gCancellableReleaseFd

Description

Releases a resources previously allocated by gCancellableGetFd or gCancellableMakePollfd().

Usage

gCancellableReleaseFd(object)

Arguments

object
a GCancellable

Details

For compatibility reasons with older releases, calling this function is not strictly required, the resources will be automatically freed when the cancellable is finalized. However, the cancellable will block scarce file descriptors until it is finalized if this function is not called. This can cause the application to run out of file descriptors when many GCancellables are used at the same time. Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]