gContentTypeGuess {RGtk2}	R Documentation

gContentTypeGuess

Description

Guesses the content type based on example data. If the function is uncertain, result.uncertain will be set to TRUE. Either filename or data may be NULL, in which case the guess will be based solely on the other argument.

Usage

gContentTypeGuess(filename, data)

Arguments

filename
a string, or NULL

data
a stream of data, or NULL

Value

A list containing the following elements:

retval
[char] a string indicating a guessed content type for the given data.

result.uncertain
a flag indicating the certainty of the result

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]