gContentTypeGuessForTree {RGtk2}	R Documentation

gContentTypeGuessForTree

Description

Tries to guess the type of the tree with root root, by looking at the files it contains. The result is a list of content types, with the best guess coming first.

Usage

gContentTypeGuessForTree(root)

Arguments

root
the root of the tree to guess a type for

Details

The types returned all have the form x-content/foo, e.g. x-content/audio-cdda (for audio CDs) or x-content/image-dcf (for a camera memory card). See the shared-mime-info (http://www.freedesktop.org/wiki/Specifications/shared-mime-info-spec) specification for more on x-content types.

This function is useful in the implementation of gMountGuessContentType. Since 2.18

Value

[char] a list of zero or more content types, or NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]