gDataInputStreamReadInt16 {RGtk2}	R Documentation

gDataInputStreamReadInt16

Description

Reads a 16-bit/2-byte value from stream.

Usage

gDataInputStreamReadInt16(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a given GDataInputStream.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

In order to get the correct byte order for this read operation, see gDataStreamGetByteOrder() and gDataStreamSetByteOrder().

Value

A list containing the following elements:

retval
[integer] a signed 16-bit/2-byte value read from stream or 0 if an error occurred.

error
GError for error reporting.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]