gDataInputStreamReadInt64 {RGtk2}	R Documentation

gDataInputStreamReadInt64

Description

Reads a 64-bit/8-byte value from stream.

Usage

gDataInputStreamReadInt64(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a given GDataInputStream.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

In order to get the correct byte order for this read operation, see gDataStreamGetByteOrder() and gDataStreamSetByteOrder().

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Value

A list containing the following elements:

retval
[numeric] a signed 64-bit/8-byte value read from stream or 0 if an error occurred.

error
GError for error reporting.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]