gDataInputStreamReadLine {RGtk2}	R Documentation

gDataInputStreamReadLine

Description

Reads a line from the data input stream.

Usage

gDataInputStreamReadLine(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a given GDataInputStream.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Value

A list containing the following elements:

retval
[char] a string with the line that was read in (without the newlines). Set length to a numeric to get the length of the read line. On an error, it will return NULL and error will be set. If there's no content to read, it will still return NULL, but error won't be set.

length
a numeric to get the length of the data read in.

error
GError for error reporting.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]