gDataInputStreamReadLineFinish {RGtk2}	R Documentation

gDataInputStreamReadLineFinish

Description

Finish an asynchronous call started by gDataInputStreamReadLineAsync.

Usage

gDataInputStreamReadLineFinish(object, result, .errwarn = TRUE)

Arguments

object
a given GDataInputStream.

result
the GAsyncResult that was provided to the callback.

.errwarn
Whether to issue a warning on error or fail silently

Details

Since 2.20

Value

A list containing the following elements:

retval
[char] a string with the line that was read in (without the newlines). Set length to a numeric to get the length of the read line. On an error, it will return NULL and error will be set. If there's no content to read, it will still return NULL, but error won't be set.

length
a numeric to get the length of the data read in.

error
GError for error reporting.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]