gDataInputStreamReadUntil {RGtk2}	R Documentation

gDataInputStreamReadUntil

Description

Reads a string from the data input stream, up to the first occurrence of any of the stop characters.

Usage

gDataInputStreamReadUntil(object, stop.chars, cancellable = NULL, 
    .errwarn = TRUE)

Arguments

object
a given GDataInputStream.

stop.chars
characters to terminate the read.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[char] a string with the data that was read before encountering any of the stop characters. Set length to a numeric to get the length of the string. This function will return NULL on an error.

length
a numeric to get the length of the data read in.

error
GError for error reporting.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]