gDataInputStreamReadUntilAsync {RGtk2}	R Documentation

gDataInputStreamReadUntilAsync

Description

The asynchronous version of gDataInputStreamReadUntil. It is an error to have two outstanding calls to this function.

Usage

gDataInputStreamReadUntilAsync(object, stop.chars, io.priority, 
    cancellable = NULL, callback, user.data = NULL)

Arguments

object
a given GDataInputStream.

stop.chars
characters to terminate the read.

io.priority
the I/O priority of the request.

cancellable
optional GCancellable object, NULL to ignore.

callback
callback to call when the request is satisfied.

user.data
the data to pass to callback function.

Details

When the operation is finished, callback will be called. You can then call gDataInputStreamReadUntilFinish to get the result of the operation. Since 2.20

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]