gDataInputStreamReadUntilFinish {RGtk2}	R Documentation

gDataInputStreamReadUntilFinish

Description

Finish an asynchronous call started by gDataInputStreamReadUntilAsync.

Usage

gDataInputStreamReadUntilFinish(object, result, length, .errwarn = TRUE)

Arguments

object
a given GDataInputStream.

result
the GAsyncResult that was provided to the callback.

length
a numeric to get the length of the data read in.

.errwarn
Whether to issue a warning on error or fail silently

Details

Since 2.20

Value

A list containing the following elements:

retval
[char] a string with the data that was read before encountering any of the stop characters. Set length to a numeric to get the length of the string. This function will return NULL on an error.

error
GError for error reporting.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]