gDataInputStreamSetByteOrder {RGtk2}	R Documentation

gDataInputStreamSetByteOrder

Description

This function sets the byte order for the given stream. All subsequent reads from the stream will be read in the given order.

Usage

gDataInputStreamSetByteOrder(object, order)

Arguments

object
a given GDataInputStream.

order
a GDataStreamByteOrder to set.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]