gDataInputStreamSetNewlineType {RGtk2}	R Documentation

gDataInputStreamSetNewlineType

Description

Sets the newline type for the stream.

Usage

gDataInputStreamSetNewlineType(object, type)

Arguments

object
a GDataInputStream.

type
the type of new line return as GDataStreamNewlineType.

Details

Note that using G_DATA_STREAM_NEWLINE_TYPE_ANY is slightly unsafe. If a read chunk ends in "CR" we must read an additional byte to know if this is "CR" or "CR LF", and this might block if there is no more data availible.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]