gDataOutputStreamPutString {RGtk2}	R Documentation

gDataOutputStreamPutString

Description

Puts a string into the output stream.

Usage

gDataOutputStreamPutString(object, str, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GDataOutputStream.

str
a string.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[logical] TRUE if string was successfully added to the stream.

error
a GError, NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]