gDataOutputStreamPutUint32 {RGtk2}	R Documentation

gDataOutputStreamPutUint32

Description

Puts an unsigned 32-bit integer into the stream.

Usage

gDataOutputStreamPutUint32(object, data, cancellable = NULL, 
    .errwarn = TRUE)

Arguments

object
a GDataOutputStream.

data
a numeric.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[logical] TRUE if data was successfully added to the stream.

error
a GError, NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]