gDriveGetIdentifier {RGtk2}	R Documentation

gDriveGetIdentifier

Description

Gets the identifier of the given kind for drive.

Usage

gDriveGetIdentifier(object, kind)

Arguments

object
a GDrive

kind
the kind of identifier to return

Value

[char] a newly allocated string containing the requested identfier, or NULL if the GDrive doesn't have this kind of identifier.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]