gDrivePollForMedia {RGtk2}	R Documentation

gDrivePollForMedia

Description

Asynchronously polls drive to see if media has been inserted or removed.

Usage

gDrivePollForMedia(object, cancellable = NULL, callback, user.data = NULL)

Arguments

object
a GDrive.

cancellable
optional GCancellable object, NULL to ignore.

callback
a GAsyncReadyCallback, or NULL.

user.data
user data to pass to callback

Details

When the operation is finished, callback will be called. You can then call gDrivePollForMediaFinish to obtain the result of the operation.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]