gFileAppendTo {RGtk2}	R Documentation

gFileAppendTo

Description

Gets an output stream for appending data to the file. If the file doesn't already exist it is created.

Usage

gFileAppendTo(object, flags = "G_FILE_CREATE_NONE", cancellable = NULL, 
    .errwarn = TRUE)

Arguments

object
input GFile.

flags
a set of GFileCreateFlags.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

By default files created are generally readable by everyone, but if you pass G_FILE_CREATE_PRIVATE in flags the file will be made readable only to the current user, to the level that is supported on the target filesystem.

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Some file systems don't allow all file names, and may return an G_IO_ERROR_INVALID_FILENAME error. If the file is a directory the G_IO_ERROR_IS_DIRECTORY error will be returned. Other errors are possible too, and depend on what kind of filesystem the file is on.

Value

A list containing the following elements:

retval
[GFileOutputStream] a GFileOutputStream, or NULL on error.

error
a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]