gFileAttributeMatcherEnumerateNamespace {RGtk2}	R Documentation

gFileAttributeMatcherEnumerateNamespace

Description

Checks if the matcher will match all of the keys in a given namespace. This will always return TRUE if a wildcard character is in use (e.g. if matcher was created with "standard::*" and ns is "standard", or if matcher was created using "*" and namespace is anything.)

Usage

gFileAttributeMatcherEnumerateNamespace(object, ns)

Arguments

object
a GFileAttributeMatcher.

ns
a string containing a file attribute namespace.

Details

TODO: this is awkwardly worded.

Value

[logical] TRUE if the matcher matches all of the entries in the given ns, FALSE otherwise.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]