gFileAttributeMatcherNew {RGtk2}	R Documentation

gFileAttributeMatcherNew

Description

Creates a new file attribute matcher, which matches attributes against a given string. GFileAttributeMatchers are reference counted structures, and are created with a reference count of 1. If the number of references falls to 0, the GFileAttributeMatcher is automatically destroyed.

Usage

gFileAttributeMatcherNew(attributes)

Arguments

attributes
an attribute string to match.

Details

The attribute string should be formatted with specific keys separated from namespaces with a double colon. Several "namespace::key" strings may be concatenated with a single comma (e.g. "standard::type,standard::is-hidden"). The wildcard "*" may be used to match all keys and namespaces, or "namespace::*" will match all keys in a given namespace.

Examples of strings to use:

"*" 	 matches all attributes.

"standard::is-hidden" 	 matches only the key is-hidden in the standard namespace.

"standard::type,unix::*" 	 matches the type key in the standard namespace and all keys in the unix namespace.

Value

[GFileAttributeMatcher] a GFileAttributeMatcher.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]