gFileCopyAsync {RGtk2}	R Documentation

gFileCopyAsync

Description

Copies the file source to the location specified by destination asynchronously. For details of the behaviour, see gFileCopy.

Usage

gFileCopyAsync(object, destination, flags = "G_FILE_COPY_NONE", 
    io.priority = 0, cancellable = NULL, progress.callback, progress.callback.data, 
    callback, user.data = NULL)

Arguments

object
input GFile.

destination
destination GFile

flags
set of GFileCopyFlags

io.priority
the I/O priority of the request.

cancellable
optional GCancellable object, NULL to ignore.

progress.callback
function to callback with progress information

progress.callback.data
user data to pass to progress.callback

callback
a GAsyncReadyCallback to call when the request is satisfied

user.data
the data to pass to callback function

Details

If progress.callback is not NULL, then that function that will be called just like in gFileCopy, however the callback will run in the main loop, not in the thread that is doing the I/O operation.

When the operation is finished, callback will be called. You can then call gFileCopyFinish to get the result of the operation.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]