gFileCopyAttributes {RGtk2}	R Documentation

gFileCopyAttributes

Description

Copies the file attributes from source to destination.

Usage

gFileCopyAttributes(object, destination, flags = "G_FILE_COPY_NONE", 
    cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GFile with attributes.

destination
a GFile to copy attributes to.

flags
a set of GFileCopyFlags.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

Normally only a subset of the file attributes are copied, those that are copies in a normal file copy operation (which for instance does not include e.g. owner). However if G_FILE_COPY_ALL_METADATA is specified in flags, then all the metadata that is possible to copy is copied. This is useful when implementing move by copy + delete source.

Value

A list containing the following elements:

retval
[logical] TRUE if the attributes were copied successfully, FALSE otherwise.

error
a GError, NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]