gFileCreateAsync {RGtk2}	R Documentation

gFileCreateAsync

Description

Asynchronously creates a new file and returns an output stream for writing to it. The file must not already exist.

Usage

gFileCreateAsync(object, flags = "G_FILE_CREATE_NONE", io.priority = 0, 
    cancellable = NULL, callback, user.data = NULL)

Arguments

object
input GFile.

flags
a set of GFileCreateFlags.

io.priority
the I/O priority of the request.

cancellable
optional GCancellable object, NULL to ignore.

callback
a GAsyncReadyCallback to call when the request is satisfied

user.data
the data to pass to callback function

Details

For more details, see gFileCreate which is the synchronous version of this call.

When the operation is finished, callback will be called. You can then call gFileCreateFinish to get the result of the operation.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]