gFileCreateReadwrite {RGtk2}	R Documentation

gFileCreateReadwrite

Description

Creates a new file and returns a stream for reading and writing to it. The file must not already exist.

Usage

gFileCreateReadwrite(object, flags, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GFile

flags
a set of GFileCreateFlags

cancellable
optional GCancellable object, NULL to ignore

.errwarn
Whether to issue a warning on error or fail silently

Details

By default files created are generally readable by everyone, but if you pass G_FILE_CREATE_PRIVATE in flags the file will be made readable only to the current user, to the level that is supported on the target filesystem.

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

If a file or directory with this name already exists the G_IO_ERROR_EXISTS error will be returned. Some file systems don't allow all file names, and may return an G_IO_ERROR_INVALID_FILENAME error, and if the name is too long, G_IO_ERROR_FILENAME_TOO_LONG will be returned. Other errors are possible too, and depend on what kind of filesystem the file is on.

Note that in many non-local file cases read and write streams are not supported, so make sure you really need to do read and write streaming, rather than just opening for reading or writing. Since 2.22

Value

A list containing the following elements:

retval
[GFileIOStream] a GFileIOStream for the newly created file, or NULL on error.

error
return location for a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]