gFileEjectMountable {RGtk2}	R Documentation

gFileEjectMountable

Description

Starts an asynchronous eject on a mountable. When this operation has completed, callback will be called with user.user data, and the operation can be finalized with gFileEjectMountableFinish. WARNING: g_file_eject_mountable has been deprecated since version 2.22 and should not be used in newly-written code. Use gFileEjectMountableWithOperation instead.

Usage

gFileEjectMountable(object, flags = "G_MOUNT_UNMOUNT_NONE", cancellable = NULL, 
    callback, user.data = NULL)

Arguments

object
input GFile.

flags
flags affecting the operation

cancellable
optional GCancellable object, NULL to ignore.

callback
a GAsyncReadyCallback to call when the request is satisfied, or NULL.

user.data
the data to pass to callback function

Details

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]