gFileEnumerateChildren {RGtk2}	R Documentation

gFileEnumerateChildren

Description

Gets the requested information about the files in a directory. The result is a GFileEnumerator object that will give out GFileInfo objects for all the files in the directory.

Usage

gFileEnumerateChildren(object, attributes, flags = "G_FILE_QUERY_INFO_NONE", 
    cancellable = NULL, .errwarn = TRUE)

Arguments

object
input GFile.

attributes
an attribute query string.

flags
a set of GFileQueryInfoFlags.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

The attribute value is a string that specifies the file attributes that should be gathered. It is not an error if it's not possible to read a particular requested attribute from a file - it just won't be set. attribute should be a comma-separated list of attribute or attribute wildcards. The wildcard "*" means all attributes, and a wildcard like "standard::*" means all attributes in the standard namespace. An example attribute query be "standard::*,owner::user". The standard attributes are available as defines, like G_FILE_ATTRIBUTE_STANDARD_NAME.

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

If the file does not exist, the G_IO_ERROR_NOT_FOUND error will be returned. If the file is not a directory, the G_FILE_ERROR_NOTDIR error will be returned. Other errors are possible too.

Value

A list containing the following elements:

retval
[GFileEnumerator] A GFileEnumerator if successful, NULL on error.

error
GError for error reporting.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]