gFileEnumeratorClose {RGtk2}	R Documentation

gFileEnumeratorClose

Description

Releases all resources used by this enumerator, making the enumerator return G_IO_ERROR_CLOSED on all calls.

Usage

gFileEnumeratorClose(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GFileEnumerator.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

This will be automatically called when the last reference is dropped, but you might want to call this function to make sure resources are released as early as possible.

Value

A list containing the following elements:

retval
[logical] TRUE on success or FALSE on error.

error
location to store the error occuring, or NULL to ignore

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]