gFileEnumeratorCloseFinish {RGtk2}	R Documentation

gFileEnumeratorCloseFinish

Description

Finishes closing a file enumerator, started from gFileEnumeratorCloseAsync.

Usage

gFileEnumeratorCloseFinish(object, result, .errwarn = TRUE)

Arguments

object
a GFileEnumerator.

result
a GAsyncResult.

.errwarn
Whether to issue a warning on error or fail silently

Details

If the file enumerator was already closed when gFileEnumeratorCloseAsync was called, then this function will report G_IO_ERROR_CLOSED in error, and return FALSE. If the file enumerator had pending operation when the close operation was started, then this function will report G_IO_ERROR_PENDING, and return FALSE. If cancellable was not NULL, then the operation may have been cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be set, and FALSE will be returned.

Value

A list containing the following elements:

retval
[logical] TRUE if the close operation has finished successfully.

error
a GError location to store the error occuring, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]