gFileEnumeratorNextFile {RGtk2}	R Documentation

gFileEnumeratorNextFile

Description

Returns information for the next file in the enumerated object. Will block until the information is available. The GFileInfo returned from this function will contain attributes that match the attribute string that was passed when the GFileEnumerator was created.

Usage

gFileEnumeratorNextFile(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GFileEnumerator.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

On error, returns NULL and sets error to the error. If the enumerator is at the end, NULL will be returned and error will be unset.

Value

A list containing the following elements:

retval
[GFileInfo] A GFileInfo or NULL on error or end of enumerator.

error
location to store the error occuring, or NULL to ignore

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]