gFileEnumeratorNextFilesAsync {RGtk2}	R Documentation

gFileEnumeratorNextFilesAsync

Description

Request information for a number of files from the enumerator asynchronously. When all i/o for the operation is finished the callback will be called with the requested information.

Usage

gFileEnumeratorNextFilesAsync(object, num.files, io.priority = 0, 
    cancellable = NULL, callback, user.data = NULL)

Arguments

object
a GFileEnumerator.

num.files
the number of file info objects to request

io.priority
the io priority of the request.

cancellable
optional GCancellable object, NULL to ignore.

callback
a GAsyncReadyCallback to call when the request is satisfied

user.data
the data to pass to callback function

Details

The callback can be called with less than num.files files in case of error or at the end of the enumerator. In case of a partial error the callback will be called with any succeeding items and no error, and on the next request the error will be reported. If a request is cancelled the callback will be called with G_IO_ERROR_CANCELLED.

During an async request no other sync and async calls are allowed, and will result in G_IO_ERROR_PENDING errors.

Any outstanding i/o request with higher priority (lower numerical value) will be executed before an outstanding request with lower priority. Default priority is G_PRIORITY_DEFAULT.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]