gFileEqual {RGtk2}	R Documentation

gFileEqual

Description

Checks equality of two given GFiles. Note that two GFiles that differ can still refer to the same file on the filesystem due to various forms of filename aliasing.

Usage

gFileEqual(object, file2)

Arguments

object
the first GFile.

file2
the second GFile.

Details

This call does no blocking i/o.

Value

[logical] TRUE if file1 and file2 are equal. FALSE if either is not a GFile.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]