gFileFindEnclosingMountAsync {RGtk2}	R Documentation

gFileFindEnclosingMountAsync

Description

Asynchronously gets the mount for the file.

Usage

gFileFindEnclosingMountAsync(object, io.priority = 0, cancellable = NULL, 
    callback, user.data = NULL)

Arguments

object
a GFile

io.priority
the I/O priority of the request.

cancellable
optional GCancellable object, NULL to ignore.

callback
a GAsyncReadyCallback to call when the request is satisfied

user.data
the data to pass to callback function

Details

For more details, see gFileFindEnclosingMount which is the synchronous version of this call.

When the operation is finished, callback will be called. You can then call gFileFindEnclosingMountFinish to get the result of the operation.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]