gFileGetBasename {RGtk2}	R Documentation

gFileGetBasename

Description

Gets the base name (the last component of the path) for a given GFile.

Usage

gFileGetBasename(object)

Arguments

object
input GFile.

Details

If called for the top level of a system (such as the filesystem root or a uri like sftp://host/) it will return a single directory separator (and on Windows, possibly a drive letter).

The base name is a byte string (*not* UTF-8). It has no defined encoding or rules other than it may not contain zero bytes. If you want to use filenames in a user interface you should use the display name that you can get by requesting the G_FILE_ATTRIBUTE_STANDARD_DISPLAY_NAME attribute with gFileQueryInfo.

This call does no blocking i/o.

Value

[char] string containing the GFile's base name, or NULL if given GFile is invalid.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]