gFileGetChild {RGtk2}	R Documentation

gFileGetChild

Description

Gets a child of file with basename equal to name.

Usage

gFileGetChild(object, name)

Arguments

object
input GFile.

name
string containing the child's basename.

Details

Note that the file with that specific name might not exist, but you can still have a GFile that points to it. You can use this for instance to create that file.

This call does no blocking i/o.

Value

[GFile] a GFile to a child specified by name.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]