gFileGetChildForDisplayName {RGtk2}	R Documentation

gFileGetChildForDisplayName

Description

Gets the child of file for a given display.name (i.e. a UTF8 version of the name). If this function fails, it returns NULL and error will be set. This is very useful when constructing a GFile for a new file and the user entered the filename in the user interface, for instance when you select a directory and type a filename in the file selector.

Usage

gFileGetChildForDisplayName(object, display.name, .errwarn = TRUE)

Arguments

object
input GFile.

display.name
string to a possible child.

.errwarn
Whether to issue a warning on error or fail silently

Details

This call does no blocking i/o.

Value

A list containing the following elements:

retval
[GFile] a GFile to the specified child, or NULL if the display name couldn't be converted.

error
GError.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]