gFileGetParent {RGtk2}	R Documentation

gFileGetParent

Description

Gets the parent directory for the file. If the file represents the root directory of the file system, then NULL will be returned.

Usage

gFileGetParent(object)

Arguments

object
input GFile.

Details

This call does no blocking i/o.

Value

[GFile] a GFile structure to the parent of the given GFile or NULL if there is no parent.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]