gFileHash {RGtk2}	R Documentation

gFileHash

Description

Creates a hash value for a GFile.

Usage

gFileHash(file)

Arguments

file
R object to a GFile.

Details

This call does no blocking i/o.

Value

[numeric] 0 if file is not a valid GFile, otherwise an integer that can be used as hash value for the GFile. This function is intended for easily hashing a GFile to add to a GHashTable or similar data structure.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]