gFileIOStreamGetEtag {RGtk2}	R Documentation

gFileIOStreamGetEtag

Description

Gets the entity tag for the file when it has been written. This must be called after the stream has been written and closed, as the etag can change while writing.

Usage

gFileIOStreamGetEtag(object)

Arguments

object
a GFileIOStream.

Details

Since 2.22

Value

[char] the entity tag for the stream.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]