gFileIOStreamQueryInfoAsync {RGtk2}	R Documentation

gFileIOStreamQueryInfoAsync

Description

Asynchronously queries the stream for a GFileInfo. When completed, callback will be called with a GAsyncResult which can be used to finish the operation with gFileIOStreamQueryInfoFinish.

Usage

gFileIOStreamQueryInfoAsync(object, attributes, io.priority, 
    cancellable = NULL, callback, user.data = NULL)

Arguments

object
a GFileIOStream.

attributes
a file attribute query string.

io.priority
the I/O priority of the request.

cancellable
optional GCancellable object, NULL to ignore.

callback
callback to call when the request is satisfied

user.data
the data to pass to callback function

Details

For the synchronous version of this function, see gFileIOStreamQueryInfo. Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]