gFileInfoGetAttributeData {RGtk2}	R Documentation

gFileInfoGetAttributeData

Description

Gets the attribute type, value and status for an attribute key.

Usage

gFileInfoGetAttributeData(object, attribute)

Arguments

object
a GFileInfo

attribute
a file attribute key

Value

A list containing the following elements:

retval
[logical] TRUE if info has an attribute named attribute, FALSE otherwise.

type
return location for the attribute type, or NULL

value.pp
return location for the attribute value, or NULL

status
return location for the attribute status, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]