gFileInfoSetAttributeStatus {RGtk2}	R Documentation

gFileInfoSetAttributeStatus

Description

Sets the attribute status for an attribute key. This is only needed by external code that implement gFileSetAttributesFromInfo or similar functions.

Usage

gFileInfoSetAttributeStatus(object, attribute, status)

Arguments

object
a GFileInfo

attribute
a file attribute key

status
a GFileAttributeStatus

Details

The attribute must exist in info for this to work. Otherwise FALSE is returned and info is unchanged. Since 2.22

Value

[logical] TRUE if the status was changed, FALSE if the key was not set.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]