gFileInfoSetSymlinkTarget {RGtk2}	R Documentation

gFileInfoSetSymlinkTarget

Description

Sets the G_FILE_ATTRIBUTE_STANDARD_SYMLINK_TARGET attribute in the file info to the given symlink target.

Usage

gFileInfoSetSymlinkTarget(object, symlink.target)

Arguments

object
a GFileInfo.

symlink.target
a static string containing a path to a symlink target.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]