gFileInputStreamQueryInfo {RGtk2}	R Documentation

gFileInputStreamQueryInfo

Description

Queries a file input stream the given attributes. This function blocks while querying the stream. For the asynchronous (non-blocking) version of this function, see gFileInputStreamQueryInfoAsync. While the stream is blocked, the stream will set the pending flag internally, and any other operations on the stream will fail with G_IO_ERROR_PENDING.

Usage

gFileInputStreamQueryInfo(object, attributes, cancellable = NULL, 
    .errwarn = TRUE)

Arguments

object
a GFileInputStream.

attributes
a file attribute query string.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[GFileInfo] a GFileInfo, or NULL on error.

error
a GError location to store the error occuring, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]