gFileIsNative {RGtk2}	R Documentation

gFileIsNative

Description

Checks to see if a file is native to the platform.

Usage

gFileIsNative(object)

Arguments

object
input GFile.

Details

A native file s one expressed in the platform-native filename format, e.g. "C:\Windows" or "/usr/bin/". This does not mean the file is local, as it might be on a locally mounted remote filesystem.

On some systems non-native files may be available using the native filesystem via a userspace filesystem (FUSE), in these cases this call will return FALSE, but gFileGetPath will still return a native path.

This call does no blocking i/o.

Value

[logical] TRUE if file is native.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]