gFileLoadContents {RGtk2}	R Documentation

gFileLoadContents

Description

Loads the content of the file into memory. The data is always zero-terminated, but this is not included in the resultant length.

Usage

gFileLoadContents(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
input GFile.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Value

A list containing the following elements:

retval
[logical] TRUE if the file's contents were successfully loaded. FALSE if there were errors.

contents
a location to place the contents of the file.

length
a location to place the length of the contents of the file, or NULL if the length is not needed

etag.out
a location to place the current entity tag for the file, or NULL if the entity tag is not needed

error
a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]