gFileLoadPartialContentsFinish {RGtk2}	R Documentation

gFileLoadPartialContentsFinish

Description

Finishes an asynchronous partial load operation that was started with gFileLoadPartialContentsAsync(). The data is always zero-terminated, but this is not included in the resultant length.

Usage

gFileLoadPartialContentsFinish(object, res, .errwarn = TRUE)

Arguments

object
input GFile.

res
a GAsyncResult.

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[logical] TRUE if the load was successful. If FALSE and error is present, it will be set appropriately.

contents
a location to place the contents of the file.

length
a location to place the length of the contents of the file, or NULL if the length is not needed

etag.out
a location to place the current entity tag for the file, or NULL if the entity tag is not needed

error
a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]