gFileMakeDirectory {RGtk2}	R Documentation

gFileMakeDirectory

Description

Creates a directory. Note that this will only create a child directory of the immediate parent directory of the path or URI given by the GFile. To recursively create directories, see gFileMakeDirectoryWithParents. This function will fail if the parent directory does not exist, setting error to G_IO_ERROR_NOT_FOUND. If the file system doesn't support creating directories, this function will fail, setting error to G_IO_ERROR_NOT_SUPPORTED.

Usage

gFileMakeDirectory(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
input GFile.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

For a local GFile the newly created directory will have the default (current) ownership and permissions of the current process.

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Value

A list containing the following elements:

retval
[logical] TRUE on successful creation, FALSE otherwise.

error
a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]