gFileMakeSymbolicLink {RGtk2}	R Documentation

gFileMakeSymbolicLink

Description

Creates a symbolic link.

Usage

gFileMakeSymbolicLink(object, symlink.value, cancellable = NULL, 
    .errwarn = TRUE)

Arguments

object
input GFile.

symlink.value
a string with the value of the new symlink.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Value

A list containing the following elements:

retval
[logical] TRUE on the creation of a new symlink, FALSE otherwise.

error
a GError.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]