gFileMonitorDirectory {RGtk2}	R Documentation

gFileMonitorDirectory

Description

Obtains a directory monitor for the given file. This may fail if directory monitoring is not supported.

Usage

gFileMonitorDirectory(object, flags = "G_FILE_MONITOR_NONE", 
    cancellable = NULL, .errwarn = TRUE)

Arguments

object
input GFile.

flags
a set of GFileMonitorFlags.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Value

A list containing the following elements:

retval
[GFileMonitor] a GFileMonitor for the given file, or NULL on error.

error
a GError, or NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]