gFileMonitorSetRateLimit {RGtk2}	R Documentation

gFileMonitorSetRateLimit

Description

Sets the rate limit to which the monitor will report consecutive change events to the same file.

Usage

gFileMonitorSetRateLimit(object, limit.msecs)

Arguments

object
a GFileMonitor.

limit.msecs
a integer with the limit in milliseconds to poll for changes.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]