gFileNewForCommandlineArg {RGtk2}	R Documentation

gFileNewForCommandlineArg

Description

Creates a GFile with the given argument from the command line. The value of arg can be either a URI, an absolute path or a relative path resolved relative to the current working directory. This operation never fails, but the returned object might not support any I/O operation if arg points to a malformed path.

Usage

gFileNewForCommandlineArg(arg)

Arguments

arg
a command line string.

Value

[GFile] a new GFile.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]