gFileNewForUri {RGtk2}	R Documentation

gFileNewForUri

Description

Constructs a GFile for a given URI. This operation never fails, but the returned object might not support any I/O operation if uri is malformed or if the uri type is not supported.

Usage

gFileNewForUri(uri)

Arguments

uri
a string containing a URI.

Value

[GFile] a GFile for the given uri.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]