gFileOpenReadwrite {RGtk2}	R Documentation

gFileOpenReadwrite

Description

Opens an existing file for reading and writing. The result is a GFileIOStream that can be used to read and write the contents of the file.

Usage

gFileOpenReadwrite(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
GFile to open

cancellable
a GCancellable

.errwarn
Whether to issue a warning on error or fail silently

Details

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

If the file does not exist, the G_IO_ERROR_NOT_FOUND error will be returned. If the file is a directory, the G_IO_ERROR_IS_DIRECTORY error will be returned. Other errors are possible too, and depend on what kind of filesystem the file is on. Note that in many non-local file cases read and write streams are not supported, so make sure you really need to do read and write streaming, rather than just opening for reading or writing. Since 2.22

Value

A list containing the following elements:

retval
[GFileIOStream] GFileIOStream or NULL on error.

error
a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]