gFileOutputStreamQueryInfo {RGtk2}	R Documentation

gFileOutputStreamQueryInfo

Description

Queries a file output stream for the given attributes. This function blocks while querying the stream. For the asynchronous version of this function, see gFileOutputStreamQueryInfoAsync. While the stream is blocked, the stream will set the pending flag internally, and any other operations on the stream will fail with G_IO_ERROR_PENDING.

Usage

gFileOutputStreamQueryInfo(object, attributes, cancellable = NULL, 
    .errwarn = TRUE)

Arguments

object
a GFileOutputStream.

attributes
a file attribute query string.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

Can fail if the stream was already closed (with error being set to G_IO_ERROR_CLOSED), the stream has pending operations (with error being set to G_IO_ERROR_PENDING), or if querying info is not supported for the stream's interface (with error being set to G_IO_ERROR_NOT_SUPPORTED). In all cases of failure, NULL will be returned.

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be set, and NULL will be returned.

Value

A list containing the following elements:

retval
[GFileInfo] a GFileInfo for the stream, or NULL on error.

error
a GError, NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]