gFilePollMountable {RGtk2}	R Documentation

gFilePollMountable

Description

Polls a file of type G_FILE_TYPE_MOUNTABLE.

Usage

gFilePollMountable(object, cancellable = NULL, callback, user.data = NULL)

Arguments

object
input GFile.

cancellable
optional GCancellable object, NULL to ignore.

callback
a GAsyncReadyCallback to call when the request is satisfied, or NULL.

user.data
the data to pass to callback function

Details

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

When the operation is finished, callback will be called. You can then call gFileMountMountableFinish to get the result of the operation. Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]