gFileQueryDefaultHandler {RGtk2}	R Documentation

gFileQueryDefaultHandler

Description

Returns the GAppInfo that is registered as the default application to handle the file specified by file.

Usage

gFileQueryDefaultHandler(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GFile to open.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Value

A list containing the following elements:

retval
[GAppInfo] a GAppInfo if the handle was found, NULL if there were errors. When you are done with it,

error
a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]