gFileQueryFileType {RGtk2}	R Documentation

gFileQueryFileType

Description

Utility function to inspect the GFileType of a file. This is implemented using gFileQueryInfo and as such does blocking I/O.

Usage

gFileQueryFileType(object, flags, cancellable = NULL)

Arguments

object
input GFile.

flags
a set of GFileQueryInfoFlags passed to gFileQueryInfo.

cancellable
optional GCancellable object, NULL to ignore.

Details

The primary use case of this method is to check if a file is a regular file, directory, or symlink. Since 2.18

Value

[GFileType] The GFileType of the file and G_FILE_TYPE_UNKNOWN if the file does not exist

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]