gFileQueryFilesystemInfo {RGtk2}	R Documentation

gFileQueryFilesystemInfo

Description

Similar to gFileQueryInfo, but obtains information about the filesystem the file is on, rather than the file itself. For instance the amount of space available and the type of the filesystem.

Usage

gFileQueryFilesystemInfo(object, attributes, cancellable = NULL, 
    .errwarn = TRUE)

Arguments

object
input GFile.

attributes
an attribute query string.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

The attribute value is a string that specifies the file attributes that should be gathered. It is not an error if it's not possible to read a particular requested attribute from a file - it just won't be set. attribute should be a comma-separated list of attribute or attribute wildcards. The wildcard "*" means all attributes, and a wildcard like "fs:*" means all attributes in the fs namespace. The standard namespace for filesystem attributes is "fs". Common attributes of interest are G_FILE_ATTRIBUTE_FILESYSTEM_SIZE (the total size of the filesystem in bytes), G_FILE_ATTRIBUTE_FILESYSTEM_FREE (number of bytes available), and G_FILE_ATTRIBUTE_FILESYSTEM_TYPE (type of the filesystem).

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

If the file does not exist, the G_IO_ERROR_NOT_FOUND error will be returned. Other errors are possible too, and depend on what kind of filesystem the file is on.

Value

A list containing the following elements:

retval
[GFileInfo] a GFileInfo or NULL if there was an error.

error
a GError.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]