gFileQueryFilesystemInfoAsync {RGtk2}	R Documentation

gFileQueryFilesystemInfoAsync

Description

Asynchronously gets the requested information about the filesystem that the specified file is on. The result is a GFileInfo object that contains key-value attributes (such as type or size for the file).

Usage

gFileQueryFilesystemInfoAsync(object, attributes, io.priority, 
    cancellable, callback, user.data = NULL)

Arguments

object
input GFile.

attributes
an attribute query string.

io.priority
the I/O priority of the request.

cancellable
optional GCancellable object, NULL to ignore.

callback
a GAsyncReadyCallback to call when the request is satisfied

user.data
the data to pass to callback function

Details

For more details, see gFileQueryFilesystemInfo which is the synchronous version of this call.

When the operation is finished, callback will be called. You can then call gFileQueryInfoFinish to get the result of the operation.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]