gFileQueryWritableNamespaces {RGtk2}	R Documentation

gFileQueryWritableNamespaces

Description

Obtain the list of attribute namespaces where new attributes can be created by a user. An example of this is extended attributes (in the "xattr" namespace).

Usage

gFileQueryWritableNamespaces(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
input GFile.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Value

A list containing the following elements:

retval
[GFileAttributeInfoList] a GFileAttributeInfoList describing the writable namespaces. When you are done with it,

error
a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]