gFileReplaceContentsAsync {RGtk2}	R Documentation

gFileReplaceContentsAsync

Description

Starts an asynchronous replacement of file with the given contents of length bytes. etag will replace the document's current entity tag.

Usage

gFileReplaceContentsAsync(object, contents, length, etag, make.backup, 
    flags = "G_FILE_CREATE_NONE", cancellable = NULL, callback, 
    user.data = NULL)

Arguments

object
input GFile.

contents
string of contents to replace the file with.

length
the length of contents in bytes.

etag
a new entity tag for the file, or NULL

make.backup
TRUE if a backup should be created.

flags
a set of GFileCreateFlags.

cancellable
optional GCancellable object, NULL to ignore.

callback
a GAsyncReadyCallback to call when the request is satisfied

user.data
the data to pass to callback function

Details

When this operation has completed, callback will be called with user.user data, and the operation can be finalized with gFileReplaceContentsFinish.

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

If make.backup is TRUE, this function will attempt to make a backup of file.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]