gFileReplaceReadwrite {RGtk2}	R Documentation

gFileReplaceReadwrite

Description

Returns an output stream for overwriting the file in readwrite mode, possibly creating a backup copy of the file first. If the file doesn't exist, it will be created.

Usage

gFileReplaceReadwrite(object, etag, make.backup, flags, cancellable = NULL, 
    .errwarn = TRUE)

Arguments

object
a GFile

etag
an optional entity tag for the current GFile, or NULL to ignore

make.backup
TRUE if a backup should be created

flags
a set of GFileCreateFlags

cancellable
optional GCancellable object, NULL to ignore

.errwarn
Whether to issue a warning on error or fail silently

Details

For details about the behaviour, see gFileReplace which does the same thing but returns an output stream only.

Note that in many non-local file cases read and write streams are not supported, so make sure you really need to do read and write streaming, rather than just opening for reading or writing. Since 2.22

Value

A list containing the following elements:

retval
[GFileIOStream] a GFileIOStream or NULL on error.

error
return location for a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]