gFileSetAttribute {RGtk2}	R Documentation

gFileSetAttribute

Description

Sets an attribute in the file with attribute name attribute to value.

Usage

gFileSetAttribute(object, attribute, type, value.p, flags = "G_FILE_QUERY_INFO_NONE", 
    cancellable = NULL, .errwarn = TRUE)

Arguments

object
input GFile.

attribute
a string containing the attribute's name.

type
The type of the attribute

value.p
a pointer to the value (or the pointer itself if the type is a pointer type)

flags
a set of GFileQueryInfoFlags.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Value

A list containing the following elements:

retval
[logical] TRUE if the attribute was set, FALSE otherwise.

error
a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]