gFileSetAttributesFromInfo {RGtk2}	R Documentation

gFileSetAttributesFromInfo

Description

Tries to set all attributes in the GFileInfo on the target values, not stopping on the first error.

Usage

gFileSetAttributesFromInfo(object, info, flags = "G_FILE_QUERY_INFO_NONE", 
    cancellable = NULL, .errwarn = TRUE)

Arguments

object
input GFile.

info
a GFileInfo.

flags
GFileQueryInfoFlags

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If there is any error during this operation then error will be set to the first error. Error on particular fields are flagged by setting the "status" field in the attribute value to G_FILE_ATTRIBUTE_STATUS_ERROR_SETTING, which means you can also detect further errors.

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Value

A list containing the following elements:

retval
[logical] TRUE if there was any error, FALSE otherwise.

error
a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]