gIOModulesLoadAllInDirectory {RGtk2}	R Documentation

gIOModulesLoadAllInDirectory

Description

Loads all the modules in the specified directory.

Usage

gIOModulesLoadAllInDirectory(dirname)

Arguments

dirname
pathname for a directory containing modules to load.

Value

[list] a list of GIOModules loaded from the directory, All the modules are loaded into memory, if you want to unload them (enabling on-demand loading) you must call gTypeModuleUnuse() on all the modules.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]