gIOStreamClose {RGtk2}	R Documentation

gIOStreamClose

Description

Closes the stream, releasing resources related to it. This will also closes the individual input and output streams, if they are not already closed.

Usage

gIOStreamClose(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GIOStream

cancellable
optional GCancellable object, NULL to ignore

.errwarn
Whether to issue a warning on error or fail silently

Details

Once the stream is closed, all other operations will return G_IO_ERROR_CLOSED. Closing a stream multiple times will not return an error.

Closing a stream will automatically flush any outstanding buffers in the stream.

Streams will be automatically closed when the last reference is dropped, but you might want to call this function to make sure resources are released as early as possible.

Some streams might keep the backing store of the stream (e.g. a file descriptor) open after the stream is closed. See the documentation for the individual stream for details.

On failure the first error that happened will be reported, but the close operation will finish as much as possible. A stream that failed to close will still return G_IO_ERROR_CLOSED for all operations. Still, it is important to check and report the error to the user, otherwise there might be a loss of data as all data might not be written.

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned. Cancelling a close will still leave the stream closed, but some streams can use a faster close that doesn't block to e.g. check errors.

The default implementation of this method just calls close on the individual input/output streams. Since 2.22

Value

A list containing the following elements:

retval
[logical] TRUE on success, FALSE on failure

error
location to store the error occuring, or NULL to ignore

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]