gIconNewForString {RGtk2}	R Documentation

gIconNewForString

Description

Generate a GIcon instance from str. This function can fail if str is not valid - see gIconToString for discussion.

Usage

gIconNewForString(str, .errwarn = TRUE)

Arguments

str
A string obtained via gIconToString.

.errwarn
Whether to issue a warning on error or fail silently

Details

If your application or library provides one or more GIcon implementations you need to ensure that each GType is registered with the type system prior to calling gIconNewForString. Since 2.20

Value

A list containing the following elements:

retval
[GIcon] An object implementing the GIcon interface or NULL if error is set.

error
Return location for error.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]