gIconToString {RGtk2}	R Documentation

gIconToString

Description

Generates a textual representation of icon that can be used for serialization such as when passing icon to a different process or saving it to persistent storage. Use gIconNewForString to get icon back from the returned string.

Usage

gIconToString(object)

Arguments

object
a GIcon.

Details

The encoding of the returned string is proprietary to GIcon except in the following two cases

If icon is a GFileIcon, the returned string is a native path (such as /path/to/my icon.png) without escaping if the GFile for icon is a native file. If the file is not native, the returned string is the result of gFileGetUri (such as sftp://path/to/my% 20icon .png).

If icon is a GThemedIcon with exactly one name, the encoding is simply the name (such as network-server).

Since 2.20

Value

[character] An allocated UTF8 string or NULL if icon can't be serialized.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]