gInitableInit {RGtk2}	R Documentation

gInitableInit

Description

Initializes the object implementing the interface. This must be done before any real use of the object after initial construction.

Usage

gInitableInit(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GInitable.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

Implementations may also support cancellation. If cancellable is not NULL, then initialization can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned. If cancellable is not NULL and the object doesn't support cancellable initialization the error G_IO_ERROR_NOT_SUPPORTED will be returned.

If this function is not called, or returns with an error then all operations on the object should fail, generally returning the error G_IO_ERROR_NOT_INITIALIZED.

Implementations of this method must be idempotent, i.e. multiple calls to this function with the same argument should return the same results. Only the first call initializes the object, further calls return the result of the first call. This is so that its safe to implement the singleton pattern in the GObject constructor function. Since 2.22

Value

A list containing the following elements:

retval
[logical] TRUE if successful. If an error has occurred, this function will return FALSE and set error appropriately if present.

error
a GError location to store the error occuring, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]