gInitableNew {RGtk2}	R Documentation

gInitableNew

Description

Helper function for constructing GInitiable object. This is similar to gObjectNew but also initializes the object and returns NULL, setting an error on failure.

Usage

gInitableNew(object.type, cancellable, ..., .errwarn = TRUE)

Arguments

object.type
a GType supporting GInitable.

cancellable
optional GCancellable object, NULL to ignore.

...
the value if the first property, followed by and other property value pairs, and ended by NULL.

.errwarn
Whether to issue a warning on error or fail silently

Details

Since 2.22

Value

A list containing the following elements:

retval
[R object] a newly allocated GObject, or NULL on error

error
a GError location to store the error occuring, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]