gInputStreamCloseAsync {RGtk2}	R Documentation

gInputStreamCloseAsync

Description

Requests an asynchronous closes of the stream, releasing resources related to it. When the operation is finished callback will be called. You can then call gInputStreamCloseFinish to get the result of the operation.

Usage

gInputStreamCloseAsync(object, io.priority = 0, cancellable = NULL, 
    callback, user.data = NULL)

Arguments

object
A GInputStream.

io.priority
the I/O priority of the request.

cancellable
optional cancellable object

callback
callback to call when the request is satisfied

user.data
the data to pass to callback function

Details

For behaviour details see gInputStreamClose.

The asyncronous methods have a default fallback that uses threads to implement asynchronicity, so they are optional for inheriting classes. However, if you override one you must override all.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]