gInputStreamReadAll {RGtk2}	R Documentation

gInputStreamReadAll

Description

Tries to read count bytes from the stream into the buffer starting at buffer. Will block during this read.

Usage

gInputStreamReadAll(object, count, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GInputStream.

count
the number of bytes that will be read from the stream

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

This function is similar to gInputStreamRead, except it tries to read as many bytes as requested, only stopping on an error or end of stream.

On a successful read of count bytes, or if we reached the end of the stream, TRUE is returned, and bytes.read is set to the number of bytes read into buffer.

If there is an error during the operation FALSE is returned and error is set to indicate the error status, bytes.read is updated to contain the number of bytes read into buffer before the error occurred.

Value

A list containing the following elements:

retval
[logical] TRUE on success, FALSE if there was an error

buffer
a buffer to read data into (which should be at least count bytes long).

bytes.read
location to store the number of bytes that was read from the stream

error
location to store the error occuring, or NULL to ignore

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]