gInputStreamReadAsync {RGtk2}	R Documentation

gInputStreamReadAsync

Description

Request an asynchronous read of count bytes from the stream into the buffer starting at buffer. When the operation is finished callback will be called. You can then call gInputStreamReadFinish to get the result of the operation.

Usage

gInputStreamReadAsync(object, count, io.priority = 0, cancellable = NULL, 
    callback, user.data = NULL)

Arguments

object
A GInputStream.

count
the number of bytes that will be read from the stream

io.priority
the I/O priority of the request.

cancellable
optional GCancellable object, NULL to ignore.

callback
callback to call when the request is satisfied

user.data
the data to pass to callback function

Details

During an async request no other sync and async calls are allowed on stream, and will result in G_IO_ERROR_PENDING errors.

A value of count larger than G_MAXSSIZE will cause a G_IO_ERROR_INVALID_ARGUMENT error.

On success, the number of bytes read into the buffer will be passed to the callback. It is not an error if this is not the same as the requested size, as it can happen e.g. near the end of a file, but generally we try to read as many bytes as requested. Zero is returned on end of file (or if count is zero), but never otherwise.

Any outstanding i/o request with higher priority (lower numerical value) will be executed before an outstanding request with lower priority. Default priority is G_PRIORITY_DEFAULT.

The asyncronous methods have a default fallback that uses threads to implement asynchronicity, so they are optional for inheriting classes. However, if you override one you must override all.

Value

A list containing the following elements:

buffer
a buffer to read data into (which should be at least count bytes long).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]