gInputStreamSkip {RGtk2}	R Documentation

gInputStreamSkip

Description

Tries to skip count bytes from the stream. Will block during the operation.

Usage

gInputStreamSkip(object, count, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GInputStream.

count
the number of bytes that will be skipped from the stream

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

This is identical to gInputStreamRead, from a behaviour standpoint, but the bytes that are skipped are not returned to the user. Some streams have an implementation that is more efficient than reading the data.

This function is optional for inherited classes, as the default implementation emulates it using read.

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned. If an operation was partially finished when the operation was cancelled the partial result will be returned, without an error.

Value

A list containing the following elements:

retval
[integer] Number of bytes skipped, or -1 on error

error
location to store the error occuring, or NULL to ignore

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]