gIoExtensionPointImplement {RGtk2}	R Documentation

gIoExtensionPointImplement

Description

Registers type as extension for the extension point with name extension.point.name.

Usage

gIoExtensionPointImplement(extension.point.name, type, extension.name, 
    priority)

Arguments

extension.point.name
the name of the extension point

type
the GType to register as extension

extension.name
the name for the extension

priority
the priority for the extension

Details

If type has already been registered as an extension for this extension point, the existing GIOExtension object is returned.

Value

[GIOExtension] a GIOExtension object for GType

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]