gIoSchedulerJobSendToMainloop {RGtk2}	R Documentation

gIoSchedulerJobSendToMainloop

Description

Used from an I/O job to send a callback to be run in the thread that the job was started from, waiting for the result (and thus blocking the I/O job).

Usage

gIoSchedulerJobSendToMainloop(object, func, user.data = NULL)

Arguments

object
a GIOSchedulerJob

func
a GSourceFunc callback that will be called in the original thread

user.data
data to pass to func

Value

[logical] The return value of func

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]