gIoSchedulerJobSendToMainloopAsync {RGtk2}	R Documentation

gIoSchedulerJobSendToMainloopAsync

Description

Used from an I/O job to send a callback to be run asynchronously in the thread that the job was started from. The callback will be run when the main loop is available, but at that time the I/O job might have finished. The return value from the callback is ignored.

Usage

gIoSchedulerJobSendToMainloopAsync(object, func, user.data = NULL)

Arguments

object
a GIOSchedulerJob

func
a GSourceFunc callback that will be called in the original thread

user.data
data to pass to func

Details

either by passing NULL as notify to gIoSchedulerPushJob() or by using refcounting for user.data.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]