gLoadableIconLoad {RGtk2}	R Documentation

gLoadableIconLoad

Description

Loads a loadable icon. For the asynchronous version of this function, see gLoadableIconLoadAsync.

Usage

gLoadableIconLoad(object, size, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GLoadableIcon.

size
an integer.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[GInputStream] a GInputStream to read the icon from.

type
a location to store the type of the loaded icon, NULL to ignore.

error
a GError location to store the error occuring, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]