gMemoryOutputStreamGetSize {RGtk2}	R Documentation

gMemoryOutputStreamGetSize

Description

Gets the size of the currently allocated data area (availible from gMemoryOutputStreamGetData). If the stream isn't growable (no realloc was passed to gMemoryOutputStreamNew) then this is the maximum size of the stream and further writes will return G_IO_ERROR_NO_SPACE.

Usage

gMemoryOutputStreamGetSize(object)

Arguments

object
a GMemoryOutputStream

Details

Note that for growable streams the returned size may become invalid on the next write or truncate operation on the stream.

If you want the number of bytes currently written to the stream, use gMemoryOutputStreamGetDataSize.

Value

[numeric] the number of bytes allocated for the data buffer

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]