gMountGuessContentType {RGtk2}	R Documentation

gMountGuessContentType

Description

Tries to guess the type of content stored on mount. Returns one or more textual identifiers of well-known content types (typically prefixed with "x-content/"), e.g. x-content/image-dcf for camera memory cards. See the shared-mime-info (http://www.freedesktop.org/wiki/Specifications/shared-mime-info-spec) specification for more on x-content types.

Usage

gMountGuessContentType(object, force.rescan, cancellable = NULL, 
    callback, user.data = NULL)

Arguments

object
a GMount

force.rescan
Whether to force a rescan of the content. Otherwise a cached result will be used if available

cancellable
optional GCancellable object, NULL to ignore

callback
a GAsyncReadyCallback

user.data
user data passed to callback

Details

This is an asynchronous operation (see gMountGuessContentTypeSync for the synchronous version), and is finished by calling gMountGuessContentTypeFinish with the mount and GAsyncResult data returned in the callback. Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]