gMountGuessContentTypeFinish {RGtk2}	R Documentation

gMountGuessContentTypeFinish

Description

Finishes guessing content types of mount. If any errors occured during the operation, error will be set to contain the errors and FALSE will be returned. In particular, you may get an G_IO_ERROR_NOT_SUPPORTED if the mount does not support content guessing.

Usage

gMountGuessContentTypeFinish(object, result, .errwarn = TRUE)

Arguments

object
a GMount

result
a GAsyncResult

.errwarn
Whether to issue a warning on error or fail silently

Details

Since 2.18

Value

A list containing the following elements:

retval
[character] a list of content types or NULL on error.

error
a GError location to store the error occuring, or NULL to ignore

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]