gMountIsShadowed {RGtk2}	R Documentation

gMountIsShadowed

Description

Determines if mount is shadowed. Applications or libraries should avoid displaying mount in the user interface if it is shadowed.

Usage

gMountIsShadowed(object)

Arguments

object
A GMount.

Details

A mount is said to be shadowed if there exists one or more user visible objects (currently GMount objects) with a root that is inside the root of mount.

One application of shadow mounts is when exposing a single file system that is used to address several logical volumes. In this situation, a GVolumeMonitor implementation would create two GVolume objects (for example, one for the camera functionality of the device and one for a SD card reader on the device) with activation URIs gphoto2://[usb:001,002]/store1/ and gphoto2://[usb:001,002]/store2/. When the underlying mount (with root gphoto2://[usb:001,002]/) is mounted, said GVolumeMonitor implementation would create two GMount objects (each with their root matching the corresponding volume activation root) that would shadow the original mount.

The proxy monitor in GVfs 2.26 and later, automatically creates and manage shadow mounts (and shadows the underlying mount) if the activation root on a GVolume is set. Since 2.20

Value

[logical] TRUE if mount is shadowed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]