gMountRemount {RGtk2}	R Documentation

gMountRemount

Description

Remounts a mount. This is an asynchronous operation, and is finished by calling gMountRemountFinish with the mount and GAsyncResults data returned in the callback.

Usage

gMountRemount(object, flags, mount.operation, cancellable = NULL, 
    callback, user.data = NULL)

Arguments

object
a GMount.

flags
flags affecting the operation

mount.operation
a GMountOperation or NULL to avoid user interaction.

cancellable
optional GCancellable object, NULL to ignore.

callback
a GAsyncReadyCallback, or NULL.

user.data
user data passed to callback.

Details

Remounting is useful when some setting affecting the operation of the volume has been changed, as these may need a remount to take affect. While this is semantically equivalent with unmounting and then remounting not all backends might need to actually be unmounted.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]