gOutputStreamSpliceAsync {RGtk2}	R Documentation

gOutputStreamSpliceAsync

Description

Splices a stream asynchronously. When the operation is finished callback will be called. You can then call gOutputStreamSpliceFinish to get the result of the operation.

Usage

gOutputStreamSpliceAsync(object, source, flags = "G_OUTPUT_STREAM_SPLICE_NONE", 
    io.priority = 0, cancellable = NULL, callback, user.data = NULL)

Arguments

object
a GOutputStream.

source
a GInputStream.

flags
a set of GOutputStreamSpliceFlags.

io.priority
the io priority of the request.

cancellable
optional GCancellable object, NULL to ignore.

callback
a GAsyncReadyCallback.

user.data
user data passed to callback.

Details

For the synchronous, blocking version of this function, see gOutputStreamSplice.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]