gOutputStreamSpliceFinish {RGtk2}	R Documentation

gOutputStreamSpliceFinish

Description

Finishes an asynchronous stream splice operation.

Usage

gOutputStreamSpliceFinish(object, result, .errwarn = TRUE)

Arguments

object
a GOutputStream.

result
a GAsyncResult.

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[integer] a integer of the number of bytes spliced.

error
a GError location to store the error occuring, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]