gOutputStreamWriteAll {RGtk2}	R Documentation

gOutputStreamWriteAll

Description

Tries to write count bytes from buffer into the stream. Will block during the operation.

Usage

gOutputStreamWriteAll(object, buffer, bytes.written, cancellable = NULL, 
    .errwarn = TRUE)

Arguments

object
a GOutputStream.

buffer
the buffer containing the data to write.

bytes.written
location to store the number of bytes that was written to the stream

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

This function is similar to gOutputStreamWrite, except it tries to write as many bytes as requested, only stopping on an error.

On a successful write of count bytes, TRUE is returned, and bytes.written is set to count.

If there is an error during the operation FALSE is returned and error is set to indicate the error status, bytes.written is updated to contain the number of bytes written into the stream before the error occurred.

Value

A list containing the following elements:

retval
[logical] TRUE on success, FALSE if there was an error

error
location to store the error occuring, or NULL to ignore

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]