gOutputStreamWriteAsync {RGtk2}	R Documentation

gOutputStreamWriteAsync

Description

Request an asynchronous write of count bytes from buffer into the stream. When the operation is finished callback will be called. You can then call gOutputStreamWriteFinish to get the result of the operation.

Usage

gOutputStreamWriteAsync(object, buffer, io.priority = 0, cancellable = NULL, 
    callback, user.data = NULL)

Arguments

object
A GOutputStream.

buffer
the buffer containing the data to write.

io.priority
the io priority of the request.

cancellable
optional GCancellable object, NULL to ignore.

callback
callback to call when the request is satisfied

user.data
the data to pass to callback function

Details

During an async request no other sync and async calls are allowed, and will result in G_IO_ERROR_PENDING errors.

A value of count larger than G_MAXSSIZE will cause a G_IO_ERROR_INVALID_ARGUMENT error.

On success, the number of bytes written will be passed to the callback. It is not an error if this is not the same as the requested size, as it can happen e.g. on a partial I/O error, but generally we try to write as many bytes as requested.

Any outstanding I/O request with higher priority (lower numerical value) will be executed before an outstanding request with lower priority. Default priority is G_PRIORITY_DEFAULT.

The asyncronous methods have a default fallback that uses threads to implement asynchronicity, so they are optional for inheriting classes. However, if you override one you must override all.

For the synchronous, blocking version of this function, see gOutputStreamWrite.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]