gResolverLookupByNameFinish {RGtk2}	R Documentation

gResolverLookupByNameFinish

Description

Retrieves the result of a call to gResolverLookupByNameAsync.

Usage

gResolverLookupByNameFinish(object, result, .errwarn = TRUE)

Arguments

object
a GResolver

result
the result passed to your GAsyncReadyCallback

.errwarn
Whether to issue a warning on error or fail silently

Details

If the DNS resolution failed, error (if non-NULL) will be set to a value from GResolverError. If the operation was cancelled, error will be set to G_IO_ERROR_CANCELLED. Since 2.22

Value

A list containing the following elements:

retval
[list] a list of GInetAddress, or NULL on error. See gResolverLookupByName for more details.

error
return location for a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]