gResolverLookupServiceAsync {RGtk2}	R Documentation

gResolverLookupServiceAsync

Description

Begins asynchronously performing a DNS SRV lookup for the given service and protocol in the given domain, and eventually calls callback, which must call gResolverLookupServiceFinish to get the final result. See gResolverLookupService for more details.

Usage

gResolverLookupServiceAsync(object, service, protocol, domain, 
    cancellable = NULL, callback, user.data = NULL)

Arguments

object
a GResolver

service
the service type to look up (eg, "ldap")

protocol
the networking protocol to use for service (eg, "tcp")

domain
the DNS domain to look up the service in

cancellable
a GCancellable, or NULL

callback
callback to call after resolution completes

user.data
data for callback

Details

Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]