gResolverSetDefault {RGtk2}	R Documentation

gResolverSetDefault

Description

Sets resolver to be the application's default resolver (reffing resolver, and unreffing the previous default resolver, if any). Future calls to gResolverGetDefault will return this resolver.

Usage

gResolverSetDefault(object)

Arguments

object
the new default GResolver

Details

This can be used if an application wants to perform any sort of DNS caching or "pinning"; it can implement its own GResolver that calls the original default resolver for DNS operations, and implements its own cache policies on top of that, and then set itself as the default resolver for all later code to use. Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]