gSeekableSeek {RGtk2}	R Documentation

gSeekableSeek

Description

Seeks in the stream by the given offset, modified by type.

Usage

gSeekableSeek(object, offset, type = "G_SEEK_SET", cancellable = NULL, 
    .errwarn = TRUE)

Arguments

object
a GSeekable.

offset
a numeric.

type
a GSeekType.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned.

Value

A list containing the following elements:

retval
[logical] TRUE if successful. If an error has occurred, this function will return FALSE and set error appropriately if present.

error
a GError location to store the error occuring, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]