gSeekableTruncate {RGtk2}	R Documentation

gSeekableTruncate

Description

Truncates a stream with a given offset.

Usage

gSeekableTruncate(object, offset, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GSeekable.

offset
a numeric.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned. If an operation was partially finished when the operation was cancelled the partial result will be returned, without an error.

Value

A list containing the following elements:

retval
[logical] TRUE if successful. If an error has occurred, this function will return FALSE and set error appropriately if present.

error
a GError location to store the error occuring, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]