gSimpleAsyncReportErrorInIdle {RGtk2}	R Documentation

gSimpleAsyncReportErrorInIdle

Description

Reports an error in an asynchronous function in an idle function by directly setting the contents of the GAsyncResult with the given error information.

Usage

gSimpleAsyncReportErrorInIdle(object, callback, user.data, domain, 
    code, format, ...)

Arguments

object
a GObject.

callback
a GAsyncReadyCallback.

user.data
user data passed to callback.

domain
a GQuark containing the error domain (usually G_IO_ERROR).

code
a specific error code.

format
a formatted error reporting string.

...
a list of variables to fill in format.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]