gSimpleAsyncResultPropagateError {RGtk2}	R Documentation

gSimpleAsyncResultPropagateError

Description

Propagates an error from within the simple asynchronous result to a given destination.

Usage

gSimpleAsyncResultPropagateError(object, .errwarn = TRUE)

Arguments

object
a GSimpleAsyncResult.

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[logical] TRUE if the error was propegated to dest. FALSE otherwise.

dest
a location to propegate the error to.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]