gSocketAddressEnumeratorNext {RGtk2}	R Documentation

gSocketAddressEnumeratorNext

Description

Retrieves the next GSocketAddress from enumerator. Note that this may block for some amount of time. (Eg, a GNetworkAddress may need to do a DNS lookup before it can return an address.) Use gSocketAddressEnumeratorNextAsync if you need to avoid blocking.

Usage

gSocketAddressEnumeratorNext(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GSocketAddressEnumerator

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If enumerator is expected to yield addresses, but for some reason is unable to (eg, because of a DNS error), then the first call to gSocketAddressEnumeratorNext will return an appropriate error in *error. However, if the first call to gSocketAddressEnumeratorNext succeeds, then any further internal errors (other than cancellable being triggered) will be ignored.

Value

A list containing the following elements:

retval
[GSocketAddress] or NULL on error (in which case *error will be set) or if there are no more addresses.

error
a GError.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]