gSocketAddressEnumeratorNextFinish {RGtk2}	R Documentation

gSocketAddressEnumeratorNextFinish

Description

Retrieves the result of a completed call to gSocketAddressEnumeratorNextAsync. See gSocketAddressEnumeratorNext for more information about error handling.

Usage

gSocketAddressEnumeratorNextFinish(object, result, .errwarn = TRUE)

Arguments

object
a GSocketAddressEnumerator

result
a GAsyncResult

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[GSocketAddress] or NULL on error (in which case *error will be set) or if there are no more addresses.

error
a GError

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]