gSocketAddressNewFromNative {RGtk2}	R Documentation

gSocketAddressNewFromNative

Description

Creates a GSocketAddress subclass corresponding to the native structsockaddr native.

Usage

gSocketAddressNewFromNative(native, len)

Arguments

native
a pointer to a structsockaddr

len
the size of the memory location pointed to by native

Details

Since 2.22

Value

[GSocketAddress] a new GSocketAddress if native could successfully be converted, otherwise NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]