gSocketClientConnect {RGtk2}	R Documentation

gSocketClientConnect

Description

Tries to resolve the connectable and make a network connection to it..

Usage

gSocketClientConnect(object, connectable, cancellable = NULL, 
    .errwarn = TRUE)

Arguments

object
a GSocketClient.

connectable
a GSocketConnectable specifying the remote address.

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

Upon a successful connection, a new GSocketConnection is constructed and returned. The caller owns this new object and must drop their reference to it when finished with it.

The type of the GSocketConnection object returned depends on the type of the underlying socket that is used. For instance, for a TCP/IP connection it will be a GTcpConnection.

The socket created will be the same family as the the address that the connectable resolves to, unless family is set with gSocketClientSetFamily or indirectly via gSocketClientSetLocalAddress. The socket type defaults to G_SOCKET_TYPE_STREAM but can be set with gSocketClientSetSocketType.

If a local address is specified with gSocketClientSetLocalAddress the socket will be bound to this address before connecting. Since 2.22

Value

A list containing the following elements:

retval
[GSocketConnection] a GSocketConnection on success, NULL on error.

error
GError for error reporting, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]