gSocketClientConnectToHost {RGtk2}	R Documentation

gSocketClientConnectToHost

Description

This is a helper function for gSocketClientConnect.

Usage

gSocketClientConnectToHost(object, host.and.port, default.port, 
    cancellable = NULL, .errwarn = TRUE)

Arguments

object
a SocketClient

host.and.port
the name and optionally port of the host to connect to

default.port
the default port to connect to

cancellable
a GCancellable, or NULL

.errwarn
Whether to issue a warning on error or fail silently

Details

Attempts to create a TCP connection to the named host. host.and.port may be in any of a number of recognised formats: an IPv6 address, an IPv4 address, or a domain name (in which case a DNS lookup is performed). Quoting with [] is supported for all address types. A port override may be specified in the usual way with a colon. Ports may be given as decimal numbers or symbolic names (in which case an /etc/services lookup is performed).

If no port override is given in host.and.port then default.port will be used as the port number to connect to.

In general, host.and.port is expected to be provided by the user (allowing them to give the hostname, and a port overide if necessary) and default.port is expected to be provided by the application.

In the case that an IP address is given, a single connection attempt is made. In the case that a name is given, multiple connection attempts may be made, in turn and according to the number of address records in DNS, until a connection succeeds.

Upon a successful connection, a new GSocketConnection is constructed and returned. The caller owns this new object and must drop their reference to it when finished with it.

In the event of any failure (DNS error, service not found, no hosts connectable) NULL is returned and error (if non-NULL) is set accordingly. Since 2.22

Value

A list containing the following elements:

retval
[GSocketConnection] a GSocketConnection on success, NULL on error.

error
a pointer to a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]