gSocketClientConnectToService {RGtk2}	R Documentation

gSocketClientConnectToService

Description

Attempts to create a TCP connection to a service.

Usage

gSocketClientConnectToService(object, domain, service, cancellable = NULL, 
    .errwarn = TRUE)

Arguments

object
a GSocketConnection

domain
a domain name

service
the name of the service to connect to

cancellable
a GCancellable, or NULL

.errwarn
Whether to issue a warning on error or fail silently

Details

This call looks up the SRV record for service at domain for the "tcp" protocol. It then attempts to connect, in turn, to each of the hosts providing the service until either a connection succeeds or there are no hosts remaining.

Upon a successful connection, a new GSocketConnection is constructed and returned. The caller owns this new object and must drop their reference to it when finished with it.

In the event of any failure (DNS error, service not found, no hosts connectable) NULL is returned and error (if non-NULL) is set accordingly.

Value

A list containing the following elements:

retval
[GSocketConnection] a GSocketConnection if successful, or NULL on error

error
a pointer to a GError, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]