gSocketClientConnectToServiceFinish {RGtk2}	R Documentation

gSocketClientConnectToServiceFinish

Description

Finishes an async connect operation. See gSocketClientConnectToServiceAsync

Usage

gSocketClientConnectToServiceFinish(object, result, .errwarn = TRUE)

Arguments

object
a GSocketClient.

result
a GAsyncResult.

.errwarn
Whether to issue a warning on error or fail silently

Details

Since 2.22

Value

A list containing the following elements:

retval
[GSocketConnection] a GSocketConnection on success, NULL on error.

error
a GError location to store the error occuring, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]