gSocketClientSetFamily {RGtk2}	R Documentation

gSocketClientSetFamily

Description

Sets the socket family of the socket client. If this is set to something other than G_SOCKET_FAMILY_INVALID then the sockets created by this object will be of the specified family.

Usage

gSocketClientSetFamily(object, family)

Arguments

object
a GSocketClient.

family
a GSocketFamily

Details

This might be useful for instance if you want to force the local connection to be an ipv4 socket, even though the address might be an ipv6 mapped to ipv4 address. Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]