gSocketClientSetLocalAddress {RGtk2}	R Documentation

gSocketClientSetLocalAddress

Description

Sets the local the socket client. The sockets created by this object will bound to the specified address (if not NULL) before connecting.

Usage

gSocketClientSetLocalAddress(object, address)

Arguments

object
a GSocketClient.

address
a GSocketAddress, or NULL

Details

This is useful if you want to ensure the the local side of the connection is on a specific port, or on a specific interface. Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]