gSocketClientSetProtocol {RGtk2}	R Documentation

gSocketClientSetProtocol

Description

Sets the protocol of the socket client. The sockets created by this object will use of the specified protocol.

Usage

gSocketClientSetProtocol(object, protocol)

Arguments

object
a GSocketClient.

protocol
a GSocketProtocol

Details

If protocol is 0 that means to use the default protocol for the socket family and type. Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]