gSocketConditionWait {RGtk2}	R Documentation

gSocketConditionWait

Description

Waits for condition to become true on socket. When the condition is met, TRUE is returned.

Usage

gSocketConditionWait(object, condition, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GSocket

condition
a GIOCondition mask to wait for

cancellable
a GCancellable, or NULL

.errwarn
Whether to issue a warning on error or fail silently

Details

If cancellable is cancelled before the condition is met then FALSE is returned and error, if non-NULL, is set to G_IO_ERROR_CANCELLED. Since 2.22

Value

A list containing the following elements:

retval
[logical] TRUE if the condition was met, FALSE otherwise

error
a GError pointer, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]