gSocketControlMessageDeserialize {RGtk2}	R Documentation

gSocketControlMessageDeserialize

Description

Tries to deserialize a socket control message of a given level and type. This will ask all known (to GType) subclasses of GSocketControlMessage if they can understand this kind of message and if so deserialize it into a GSocketControlMessage.

Usage

gSocketControlMessageDeserialize(level, type, size, data)

Arguments

level
a socket level

type
a socket control message type for the given level

size
the size of the data in bytes

data
pointer to the message data

Details

If there is no implementation for this kind of control message, NULL will be returned. Since 2.22

Value

[GSocketControlMessage] the deserialized message or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]