gSocketCreateSource {RGtk2}	R Documentation

gSocketCreateSource

Description

Creates a GSource that can be attached to a GMainContext to monitor for the availibility of the specified condition on the socket.

Usage

gSocketCreateSource(object, condition, cancellable = NULL)

Arguments

object
a GSocket

condition
a GIOCondition mask to monitor

cancellable
a GCancellable or NULL

Details

The callback on the source is of the GSocketSourceFunc type.

It is meaningless to specify G_IO_ERR or G_IO_HUP in condition; these conditions will always be reported output if they are true. cancellable if not NULL can be used to cancel the source, which will cause the source to trigger, reporting the current condition (which is likely 0 unless cancellation happened at the same time as a condition change). You can check for this in the callback using gCancellableIsCancelled. Since 2.22

Value

[GSource] a newly allocated GSource, free with gSourceUnref().

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]