gSocketListenerAcceptSocket {RGtk2}	R Documentation

gSocketListenerAcceptSocket

Description

Blocks waiting for a client to connect to any of the sockets added to the listener. Returns the GSocket that was accepted.

Usage

gSocketListenerAcceptSocket(object, cancellable = NULL, .errwarn = TRUE)

Arguments

object
a GSocketListener

cancellable
optional GCancellable object, NULL to ignore.

.errwarn
Whether to issue a warning on error or fail silently

Details

If you want to accept the high-level GSocketConnection, not a GSocket, which is often the case, then you should use gSocketListenerAccept instead.

If source.object is not NULL it will be filled out with the source object specified when the corresponding socket or address was added to the listener.

If cancellable is not NULL, then the operation can be cancelled by triggering the cancellable object from another thread. If the operation was cancelled, the error G_IO_ERROR_CANCELLED will be returned. Since 2.22

Value

A list containing the following elements:

retval
[GSocket] a GSocket on success, NULL on error.

source.object
location where GObject pointer will be stored, or NULL

error
GError for error reporting, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]