gSocketListenerAddSocket {RGtk2}	R Documentation

gSocketListenerAddSocket

Description

Adds socket to the set of sockets that we try to accept new clients from. The socket must be bound to a local address and listened to.

Usage

gSocketListenerAddSocket(object, socket, source.object, .errwarn = TRUE)

Arguments

object
a GSocketListener

socket
a listening GSocket

source.object
Optional GObject identifying this source

.errwarn
Whether to issue a warning on error or fail silently

Details

source.object will be passed out in the various calls to accept to identify this particular source, which is useful if you're listening on multiple addresses and do different things depending on what address is connected to. Since 2.22

Value

A list containing the following elements:

retval
[logical] TRUE on success, FALSE on error.

error
GError for error reporting, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]