gSocketNew {RGtk2}	R Documentation

gSocketNew

Description

Creates a new GSocket with the defined family, type and protocol. If protocol is 0 (G_SOCKET_PROTOCOL_DEFAULT) the default protocol type for the family and type is used.

Usage

gSocketNew(family, type, protocol, .errwarn = TRUE)

Arguments

family
the socket family to use, e.g. G_SOCKET_FAMILY_IPV4.

type
the socket type to use.

protocol
the id of the protocol to use, or 0 for default.

.errwarn
Whether to issue a warning on error or fail silently

Details

The protocol is a family and type specific int that specifies what kind of protocol to use. GSocketProtocol lists several common ones. Many families only support one protocol, and use 0 for this, others support several and using 0 means to use the default protocol for the family and type.

The protocol id is passed directly to the operating system, so you can use protocols not listed in GSocketProtocol if you know the protocol number used for it. Since 2.22

Value

A list containing the following elements:

retval
[GSocket] a GSocket or NULL on error.

error
GError for error reporting, or NULL to ignore.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]