gSocketServiceIsActive {RGtk2}	R Documentation

gSocketServiceIsActive

Description

Check whether the service is active or not. An active service will accept new clients that connect, while a non-active service will let connecting clients queue up until the service is started.

Usage

gSocketServiceIsActive(object)

Arguments

object
a GSocketService

Details

Since 2.22

Value

[logical] TRUE if the service is active, FALSE otherwise

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]