gSocketSetBlocking {RGtk2}	R Documentation

gSocketSetBlocking

Description

Sets the blocking mode of the socket. In blocking mode all operations block until they succeed or there is an error. In non-blocking mode all functions return results immediately or with a G_IO_ERROR_WOULD_BLOCK error.

Usage

gSocketSetBlocking(object, blocking)

Arguments

object
a GSocket.

blocking
Whether to use blocking I/O or not.

Details

All sockets are created in blocking mode. However, note that the platform level socket is always non-blocking, and blocking mode is a GSocket level feature. Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]