gSocketSetListenBacklog {RGtk2}	R Documentation

gSocketSetListenBacklog

Description

Sets the maximum number of outstanding connections allowed when listening on this socket. If more clients than this are connecting to the socket and the application is not handling them on time then the new connections will be refused.

Usage

gSocketSetListenBacklog(object, backlog)

Arguments

object
a GSocket.

backlog
the maximum number of pending connections.

Details

Note that this must be called before gSocketListen and has no effect if called after that. Since 2.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]