gSocketSpeaksIpv4 {RGtk2}	R Documentation

gSocketSpeaksIpv4

Description

Checks if a socket is capable of speaking IPv4.

Usage

gSocketSpeaksIpv4(object)

Arguments

object
a GSocket

Details

IPv4 sockets are capable of speaking IPv4. On some operating systems and under some combinations of circumstances IPv6 sockets are also capable of speaking IPv4. See RFC 3493 section 3.7 for more information.

No other types of sockets are currently considered as being capable of speaking IPv4. Since 2.22.

Value

[logical] TRUE if this socket can be used with IPv4.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]