gVolumeEjectFinish {RGtk2}	R Documentation

gVolumeEjectFinish

Description

Finishes ejecting a volume. If any errors occured during the operation, error will be set to contain the errors and FALSE will be returned. WARNING: g_volume_eject_finish has been deprecated since version 2.22 and should not be used in newly-written code. Use gVolumeEjectWithOperationFinish instead.

Usage

gVolumeEjectFinish(object, result, .errwarn = TRUE)

Arguments

object
pointer to a GVolume.

result
a GAsyncResult.

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[logical] TRUE, FALSE if operation failed.

error
a GError location to store an error, or NULL to ignore

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]