gVolumeGetActivationRoot {RGtk2}	R Documentation

gVolumeGetActivationRoot

Description

Gets the activation root for a GVolume if it is known ahead of mount time. Returns NULL otherwise. If not NULL and if volume is mounted, then the result of gMountGetRoot on the GMount object obtained from gVolumeGetMount will always either be equal or a prefix of what this function returns. In other words, in code

Usage

gVolumeGetActivationRoot(object)

Arguments

object
a GVolume

Details

  GMount *mount;
  GFile *mount_root
  GFile *volume_activation_root;

  mount = g_volume_get_mount (volume); /* mounted, so never NULL */
  mount_root = g_mount_get_root (mount);
  volume_activation_root = g_volume_get_activation_root(volume); /* assume not NULL */

then the expression

  (g_file_has_prefix (volume_activation_root, mount_root) ||
      g_file_equal (volume_activation_root, mount_root))

will always be TRUE.

Activation roots are typically used in GVolumeMonitor implementations to find the underlying mount to shadow, see gMountIsShadowed for more details. Since 2.18

Value

[GFile] the activation root of volume or NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]