gVolumeMonitorAdoptOrphanMount {RGtk2}	R Documentation

gVolumeMonitorAdoptOrphanMount

Description

This function should be called by any GVolumeMonitor implementation when a new GMount object is created that is not associated with a GVolume object. It must be called just before emitting the mount.added signal. WARNING: g_volume_monitor_adopt_orphan_mount has been deprecated since version 2.20 and should not be used in newly-written code. Instead of using this function, GVolumeMonitor implementations should instead create shadow mounts with the URI of the mount they intend to adopt. See the proxy volume monitor in gvfs for an example of this. Also see gMountIsShadowed, gMountShadow and gMountUnshadow functions.

Usage

gVolumeMonitorAdoptOrphanMount(mount)

Arguments

mount
a GMount object to find a parent for

Details

If the return value is not NULL, the caller must associate the returned GVolume object with the GMount. This involves returning it in its gMountGetVolume implementation. The caller must also listen for the "removed" signal on the returned object and give up its reference when handling that signal

Similary, if implementing gVolumeMonitorAdoptOrphanMount, the implementor must take a reference to mount and return it in its gVolumeGetMount implemented. Also, the implementor must listen for the "unmounted" signal on mount and give up its reference upon handling that signal.

There are two main use cases for this function.

One is when implementing a user space file system driver that reads blocks of a block device that is already represented by the native volume monitor (for example a CD Audio file system driver). Such a driver will generate its own GMount object that needs to be assoicated with the GVolume object that represents the volume.

The other is for implementing a GVolumeMonitor whose sole purpose is to return GVolume objects representing entries in the users "favorite servers" list or similar.

Value

[GVolume] the GVolume object that is the parent for mount or NULL if no wants to adopt the GMount.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]