gdk-Application-launching {RGtk2}	R Documentation

Application launching

Description

Startup notification for applications

Methods and Functions

gdkAppLaunchContextNew()
gdkAppLaunchContextSetDisplay(object, display)
gdkAppLaunchContextSetScreen(object, screen)
gdkAppLaunchContextSetDesktop(object, desktop)
gdkAppLaunchContextSetTimestamp(object, timestamp)
gdkAppLaunchContextSetIcon(object, icon = NULL)
gdkAppLaunchContextSetIconName(object, icon.name = NULL)
gdkAppLaunchContext()

Detailed Description

GdkAppLaunchContext is an implementation of GAppLaunchContext that handles launching an application in a graphical context. It provides startup notification and allows to launch applications on a specific screen or workspace.

Launching an application

context <- gdkAppLaunchContext()
context$setScreen(my_screen)
context$setTimestamp(event$time)
gAppInfoLaunchDefaultForUri("http://www.gtk.org", context)

Structures

GdkAppLaunchContext

An opaque structure representing an application launch context.

Convenient Construction

gdkAppLaunchContext is the equivalent of gdkAppLaunchContextNew.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/gdk-Application-launching.html

[Package RGtk2 version 2.20.31 Index]