gdk-Bitmaps-and-Pixmaps {RGtk2}	R Documentation

Bitmaps and Pixmaps

Description

Offscreen drawables

Methods and Functions

gdkPixmapNew(drawable = NULL, width, height, depth = -1)
gdkBitmapCreateFromData(drawable = NULL, data, width, height)
gdkPixmapCreateFromData(drawable = NULL, data, height, depth, fg, bg)
gdkPixmapCreateFromXpm(drawable, transparent.color, filename)
gdkPixmapColormapCreateFromXpm(drawable, colormap, transparent.color, filename)
gdkPixmapCreateFromXpmD(drawable, transparent.color, data)
gdkPixmapColormapCreateFromXpmD(drawable, colormap, transparent.color, data)
gdkPixmap(drawable, data, transparent.color, height, width, depth, filename, colormap, fg, bg)

Hierarchy

GObject
   +----GdkDrawable
         +----GdkPixmap

Detailed Description

Pixmaps are offscreen drawables. They can be drawn upon with the standard drawing primitives, then copied to another drawable (such as a GdkWindow) with gdkPixmapDraw(). The depth of a pixmap is the number of bits per pixels. Bitmaps are simply pixmaps with a depth of 1. (That is, they are monochrome bitmaps - each pixel can be either on or off).

Structures

GdkPixmap

An opaque structure representing an offscreen drawable. Pointers to structures of type GdkPixmap, GdkBitmap, and GdkWindow, can often be used interchangeably. The type GdkDrawable refers generically to any of these types.

GdkBitmap

An opaque structure representing an offscreen drawable of depth 1. Pointers to structures of type GdkPixmap, GdkBitmap, and GdkWindow, can often be used interchangeably. The type GdkDrawable refers generically to any of these types.

Convenient Construction

gdkPixmap is the result of collapsing the constructors of GdkPixmap (gdkPixmapNew, gdkBitmapCreateFromData, gdkPixmapCreateFromData, gdkPixmapCreateFromXpm, gdkPixmapColormapCreateFromXpm, gdkPixmapCreateFromXpmD, gdkPixmapColormapCreateFromXpmD, NULL, NULL) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/gdk-Bitmaps-and-Pixmaps.html

[Package RGtk2 version 2.20.31 Index]