gdk-Cairo-Interaction {RGtk2}	R Documentation

Cairo Interaction

Description

Functions to support using Cairo

Methods and Functions

gdkCairoCreate(drawable)
gdkCairoSetSourceColor(cr, color)
gdkCairoSetSourcePixbuf(cr, pixbuf, pixbuf.x, pixbuf.y)
gdkCairoSetSourcePixmap(cr, pixmap, pixmap.x, pixmap.y)
gdkCairoRectangle(cr, rectangle)
gdkCairoRegion(cr, region)
gdkCairoResetClip(cr, drawable)

Detailed Description

Cairo (http://cairographics.org) is a graphics library that supports vector graphics and image compositing that can be used with GDK. Since 2.8, GTK+ does most of its drawing using Cairo.

GDK does not wrap the Cairo API, instead it allows to create Cairo contexts which can be used to draw on GDK drawables. Additional functions allow to convert GDK's rectangles and regions into Cairo paths and to use pixbufs as sources for drawing operations.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/gdk-Cairo-Interaction.html

[Package RGtk2 version 2.20.31 Index]