gdk-Cursors {RGtk2}	R Documentation

Cursors

Description

Standard and pixmap cursors

Methods and Functions

gdkCursorNew(cursor.type)
gdkCursorNewFromPixmap(source, mask, fg, bg, x, y)
gdkCursorNewFromPixbuf(display, source, x, y)
gdkCursorNewFromName(display, name)
gdkCursorNewForDisplay(display, cursor.type)
gdkCursorGetDisplay(object)
gdkCursorGetImage(object)

Detailed Description

These functions are used to create and destroy cursors. There is a number of standard cursors, but it is also possible to construct new cursors from pixmaps and pixbufs. There may be limitations as to what kinds of cursors can be constructed on a given display, see gdkDisplaySupportsCursorAlpha, gdkDisplaySupportsCursorColor, gdkDisplayGetDefaultCursorSize and gdkDisplayGetMaximalCursorSize.

Cursors by themselves are not very interesting, they must be be bound to a window for users to see them. This is done with gdkWindowSetCursor or by setting the cursor member of the GdkWindowAttr struct passed to gdkWindowNew.

Structures

GdkCursor

A GdkCursor structure represents a cursor.

type

[GdkCursorType] the GdkCursorType of the cursor

Enums and Flags

GdkCursorType

The standard cursors available.

x_cursor

undocumented

arrow

undocumented

based_arrow_down

undocumented

based_arrow_up

undocumented

boat

undocumented

bogosity

undocumented

bottom_left_corner

undocumented

bottom_right_corner

undocumented

bottom_side

undocumented

bottom_tee

undocumented

box_spiral

undocumented

center_ptr

undocumented

circle

undocumented

clock

undocumented

coffee_mug

undocumented

cross

undocumented

cross_reverse

undocumented

crosshair

undocumented

diamond_cross

undocumented

dot

undocumented

dotbox

undocumented

double_arrow

undocumented

draft_large

undocumented

draft_small

undocumented

draped_box

undocumented

exchange

undocumented

fleur

undocumented

gobbler

undocumented

gumby

undocumented

hand1

undocumented

hand2

undocumented

heart

undocumented

icon

undocumented

iron_cross

undocumented

left_ptr

undocumented

left_side

undocumented

left_tee

undocumented

leftbutton

undocumented

ll_angle

undocumented

lr_angle

undocumented

man

undocumented

middlebutton

undocumented

mouse

undocumented

pencil

undocumented

pirate

undocumented

plus

undocumented

question_arrow

undocumented

right_ptr

undocumented

right_side

undocumented

right_tee

undocumented

rightbutton

undocumented

rtl_logo

undocumented

sailboat

undocumented

sb_down_arrow

undocumented

sb_h_double_arrow

undocumented

sb_left_arrow

undocumented

sb_right_arrow

undocumented

sb_up_arrow

undocumented

sb_v_double_arrow

undocumented

shuttle

undocumented

sizing

undocumented

spider

undocumented

spraycan

undocumented

star

undocumented

target

undocumented

tcross

undocumented

top_left_arrow

undocumented

top_left_corner

undocumented

top_right_corner

undocumented

top_side

undocumented

top_tee

undocumented

trek

undocumented

ul_angle

undocumented

umbrella

undocumented

ur_angle

undocumented

watch

undocumented

xterm

undocumented

last-cursor

last cursor type

gdk-cursor-is-pixmap

Blank cursor. Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/gdk-Cursors.html

[Package RGtk2 version 2.20.31 Index]