gdk-Event-Structures {RGtk2}	R Documentation

Event Structures

Description

Data structures specific to each type of event

Detailed Description

The event structs contain data specific to each type of event in GDK. PLEASE NOTE: A common mistake is to forget to set the event mask of a widget so that the required events are received. See gtkWidgetSetEvents.

Structures

GdkEventAny

Contains the fields which are common to all event structs. Any event pointer can safely be cast to a pointer to a GdkEventAny to access these fields.

type

[GdkEventType] the type of the event.

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

GdkEventKey

Describes a key press or key release event.

type

[GdkEventType] the type of the event (GDK_KEY_PRESS or GDK_KEY_RELEASE).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

time

[numeric] the time of the event in milliseconds.

state

[numeric] a bit-mask representing the state of the modifier keys (e.g. Control, Shift and Alt) and the pointer buttons. See GdkModifierType.

keyval

[numeric] the key that was pressed or released. See the ‘<gdk/gdkkeysyms.h>’ header file for a complete list of GDK key codes.

length

[integer] the length of string.

string

[character] a string containing the an approximation of the text that would result from this keypress. The only correct way to handle text input of text is using input methods (see GtkIMContext), so this field is deprecated and should never be used. (gdkUnicodeToKeyval provides a non-deprecated way of getting an approximate translation for a key.) The string is encoded in the encoding of the current locale (Note: this for backwards compatibility: strings in GTK+ and GDK are typically in UTF-8.) and NUL-terminated. In some cases, the translation of the key code will be a single NUL byte, in which case looking at length is necessary to distinguish it from the an empty translation.

hardwareKeycode

[integer] the raw code of the key that was pressed or released.

group

[raw] the keyboard group.

GdkEventButton

Used for button press and button release events. The type field will be one of GDK_BUTTON_PRESS, GDK_2BUTTON_PRESS, GDK_3BUTTON_PRESS, and GDK_BUTTON_RELEASE.

Double and triple-clicks result in a sequence of events being received. For double-clicks the order of events will be:

GDK_BUTTON_PRESS

GDK_BUTTON_RELEASE

GDK_BUTTON_PRESS

GDK_2BUTTON_PRESS

GDK_BUTTON_RELEASE

Note that the first click is received just like a normal button press, while the second click results in a GDK_2BUTTON_PRESS being received just after the GDK_BUTTON_PRESS.

Triple-clicks are very similar to double-clicks, except that GDK_3BUTTON_PRESS is inserted after the third click. The order of the events is:

GDK_BUTTON_PRESS

GDK_BUTTON_RELEASE

GDK_BUTTON_PRESS

GDK_2BUTTON_PRESS

GDK_BUTTON_RELEASE

GDK_BUTTON_PRESS

GDK_3BUTTON_PRESS

GDK_BUTTON_RELEASE

For a double click to occur, the second button press must occur within 1/4 of a second of the first. For a triple click to occur, the third button press must also occur within 1/2 second of the first button press.

type

[GdkEventType] the type of the event (GDK_BUTTON_PRESS, GDK_2BUTTON_PRESS, GDK_3BUTTON_PRESS or GDK_BUTTON_RELEASE).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

time

[numeric] the time of the event in milliseconds.

x

[numeric] the x coordinate of the pointer relative to the window.

y

[numeric] the y coordinate of the pointer relative to the window.

axes

[numeric] x, y translated to the axes of device, or NULL if device is the mouse.

state

[numeric] a bit-mask representing the state of the modifier keys (e.g. Control, Shift and Alt) and the pointer buttons. See GdkModifierType.

button

[numeric] the button which was pressed or released, numbered from 1 to 5. Normally button 1 is the left mouse button, 2 is the middle button, and 3 is the right button. On 2-button mice, the middle button can often be simulated by pressing both mouse buttons together.

device

[GdkDevice] the device where the event originated.

xRoot

[numeric] the x coordinate of the pointer relative to the root of the screen.

yRoot

[numeric] the y coordinate of the pointer relative to the root of the screen.

GdkEventScroll

Generated from button presses for the buttons 4 to 7. Wheel mice are usually configured to generate button press events for buttons 4 and 5 when the wheel is turned.

type

[GdkEventType] the type of the event (GDK_SCROLL).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

time

[numeric] the time of the event in milliseconds.

x

[numeric] the x coordinate of the pointer relative to the window.

y

[numeric] the y coordinate of the pointer relative to the window.

state

[numeric] a bit-mask representing the state of the modifier keys (e.g. Control, Shift and Alt) and the pointer buttons. See GdkModifierType.

direction

[GdkScrollDirection] the direction to scroll to (one of GDK_SCROLL_UP, GDK_SCROLL_DOWN, GDK_SCROLL_LEFT and GDK_SCROLL_RIGHT).

device

[GdkDevice] the device where the event originated.

xRoot

[numeric] the x coordinate of the pointer relative to the root of the screen.

yRoot

[numeric] the y coordinate of the pointer relative to the root of the screen.

GdkEventMotion

Generated when the pointer moves.

type

[GdkEventType] the type of the event.

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

time

[numeric] the time of the event in milliseconds.

x

[numeric] the x coordinate of the pointer relative to the window.

y

[numeric] the y coordinate of the pointer relative to the window.

axes

[numeric] x, y translated to the axes of device, or NULL if device is the mouse.

state

[numeric] a bit-mask representing the state of the modifier keys (e.g. Control, Shift and Alt) and the pointer buttons. See GdkModifierType.

isHint

[integer] set to 1 if this event is just a hint, see the GDK_POINTER_MOTION_HINT_MASK value of GdkEventMask.

device

[GdkDevice] the device where the event originated.

xRoot

[numeric] the x coordinate of the pointer relative to the root of the screen.

yRoot

[numeric] the y coordinate of the pointer relative to the root of the screen.

GdkEventExpose

Generated when all or part of a window becomes visible and needs to be redrawn.

type

[GdkEventType] the type of the event (GDK_EXPOSE or GDK_DAMAGE).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

area

[GdkRectangle] bounding box of region.

region

[GdkRegion] the region that needs to be redrawn.

count

[integer] the number of contiguous GDK_EXPOSE events following this one. The only use for this is "exposure compression", i.e. handling all contiguous GDK_EXPOSE events in one go, though GDK performs some exposure compression so this is not normally needed.

GdkEventVisibility

Generated when the window visibility status has changed.

type

[GdkEventType] the type of the event (GDK_VISIBILITY_NOTIFY).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

state

[GdkVisibilityState] the new visibility state (GDK_VISIBILITY_FULLY_OBSCURED, GDK_VISIBILITY_PARTIAL or GDK_VISIBILITY_UNOBSCURED).

GdkEventCrossing

Generated when the pointer enters or leaves a window.

type

[GdkEventType] the type of the event (GDK_ENTER_NOTIFY or GDK_LEAVE_NOTIFY).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

context

[GdkDragContext] the window that was entered or left.

time

[numeric] the time of the event in milliseconds.

xRoot

[integer] the x coordinate of the pointer relative to the window.

yRoot

[integer] the y coordinate of the pointer relative to the window.

GdkEventFocus

Describes a change of keyboard focus.

type

[GdkEventType] the type of the event (GDK_FOCUS_CHANGE).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

in

[integer] TRUE if the window has gained the keyboard focus, FALSE if it has lost the focus.

GdkEventConfigure

Generated when a window size or position has changed.

type

[GdkEventType] the type of the event (GDK_CONFIGURE).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

x

[integer] the new x coordinate of the window, relative to its parent.

y

[integer] the new y coordinate of the window, relative to its parent.

width

[integer] the new width of the window.

height

[integer] the new height of the window.

GdkEventProperty

Describes a property change on a window.

type

[GdkEventType] the type of the event (GDK_PROPERTY_NOTIFY).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

atom

[GdkAtom] the property that was changed.

time

[numeric] the time of the event in milliseconds.

state

[numeric] whether the property was changed (GDK_PROPERTY_NEW_VALUE) or deleted (GDK_PROPERTY_DELETE).

GdkEventSelection

Generated when a selection is requested or ownership of a selection is taken over by another client application.

type

[GdkEventType] the type of the event (GDK_SELECTION_CLEAR, GDK_SELECTION_NOTIFY or GDK_SELECTION_REQUEST).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

selection

[GdkAtom] the selection.

target

[GdkAtom] the target to which the selection should be converted.

property

[GdkAtom] the property in which to place the result of the conversion.

time

[numeric] the time of the event in milliseconds.

requestor

[GdkNativeWindow] the native window on which to place property.

GdkEventDND

Generated during DND operations.

type

[GdkEventType] the type of the event (GDK_DRAG_ENTER, GDK_DRAG_LEAVE, GDK_DRAG_MOTION, GDK_DRAG_STATUS, GDK_DROP_START or GDK_DROP_FINISHED).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

context

[GdkDragContext] the GdkDragContext for the current DND operation.

time

[numeric] the time of the event in milliseconds.

xRoot

[integer] the x coordinate of the pointer relative to the root of the screen, only set for GDK_DRAG_MOTION and GDK_DROP_START.

yRoot

[integer] the y coordinate of the pointer relative to the root of the screen, only set for GDK_DRAG_MOTION and GDK_DROP_START.

GdkEventProximity

Proximity events are generated when using GDK's wrapper for the XInput extension. The XInput extension is an add-on for standard X that allows you to use nonstandard devices such as graphics tablets. A proximity event indicates that the stylus has moved in or out of contact with the tablet, or perhaps that the user's finger has moved in or out of contact with a touch screen.

type

[GdkEventType] the type of the event (GDK_PROXIMITY_IN or GDK_PROXIMITY_OUT).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

time

[numeric] the time of the event in milliseconds.

device

[GdkDevice] the device where the event originated.

GdkEventClient

An event sent by another client application.

type

[GdkEventType] the type of the event (GDK_CLIENT_EVENT).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

messageType

[GdkAtom] the type of the message, which can be defined by the application.

GdkEventNoExpose

Generated when the area of a GdkDrawable being copied, with gdkDrawDrawable or gdkWindowCopyArea(), was completely available.

FIXME: add more here.

type

[GdkEventType] the type of the event (GDK_NO_EXPOSE).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

GdkEventWindowState

Generated when the state of a toplevel window changes.

type

[GdkEventType] the type of the event (GDK_WINDOW_STATE).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

changedMask

[GdkWindowState] mask specifying what flags have changed.

newWindowState

[GdkWindowState] the new window state, a combination of GdkWindowState bits.

GdkEventSetting

Generated when a setting is modified.

type

[GdkEventType] the type of the event (GDK_SETTING).

window

[GdkWindow] the window which received the event.

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

action

[GdkSettingAction] what happened to the setting (GDK_SETTING_ACTION_NEW, GDK_SETTING_ACTION_CHANGED or GDK_SETTING_ACTION_DELETED).

name

[char] the name of the setting.

GdkEventOwnerChange

Generated when the owner of a selection changes. On X11, this information is only available if the X server supports the XFIXES extension. Since 2.6

type

[GdkEventType] the type of the event (GDK_OWNER_CHANGE).

window

[GdkWindow] the window which received the event

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

owner

[GdkNativeWindow] the new owner of the selection

reason

[GdkOwnerChange] the reason for the ownership change as a GdkOwnerChange value

selection

[GdkAtom] the atom identifying the selection

time

[numeric] the timestamp of the event

selectionTime

[numeric] the time at which the selection ownership was taken over

GdkEventGrabBroken

Generated when a pointer or keyboard grab is broken. On X11, this happens when the grab window becomes unviewable (i.e. it or one of its ancestors is unmapped), or if the same application grabs the pointer or keyboard again. Note that implicit grabs (which are initiated by button presses) can also cause GdkEventGrabBroken events. Since 2.8

type

[GdkEventType] the type of the event (GDK_GRAB_BROKEN)

window

[GdkWindow] the window which received the event, i.e. the window that previously owned the grab

sendEvent

[raw] TRUE if the event was sent explicitly (e.g. using xsendevent).

keyboard

[logical] TRUE if a keyboard grab was broken, FALSE if a pointer grab was broken

implicit

[logical] TRUE if the broken grab was implicit

grabWindow

[GdkWindow] If this event is caused by another grab in the same application, grab.window contains the new grab window. Otherwise grab.window is NULL.

Enums and Flags

GdkScrollDirection

Specifies the direction for GdkEventScroll.

up

the window is scrolled up.

down

the window is scrolled down.

left

the window is scrolled to the left.

right

the window is scrolled to the right.

GdkVisibilityState

Specifies the visiblity status of a window for a GdkEventVisibility.

unobscured

the window is completely visible.

partial

the window is partially visible.

fully-obscured

the window is not visible at all.

GdkCrossingMode

Specifies the crossing mode for GdkEventCrossing.

normal

crossing because of pointer motion.

grab

crossing because a grab is activated.

ungrab

crossing because a grab is deactivated.

GdkNotifyType

Specifies the kind of crossing for GdkEventCrossing.

See the X11 protocol specification of LeaveNotify for full details of crossing event generation.

ancestor

the window is entered from an ancestor or left towards an ancestor.

virtual

the pointer moves between an ancestor and an inferior of the window.

inferior

the window is entered from an inferior or left towards an inferior.

nonlinear

the window is entered from or left towards a window which is neither an ancestor nor an inferior.

nonlinear-virtual

the pointer moves between two windows which are not ancestors of each other and the window is part of the ancestor chain between one of these windows and their least common ancestor.

unknown

an unknown type of enter/leave event occurred.

GdkPropertyState

Specifies the type of a property change for a GdkEventProperty.

new-value

the property value was changed.

delete

the property was deleted.

GdkWindowState

Specifies the state of a toplevel window.

withdrawn

the window is not shown.

iconified

the window is minimized.

maximized

the window is maximized.

sticky

the window is sticky.

fullscreen

the window is maximized without decorations.

above

the window is kept above other windows.

below

the window is kept below other windows.

GdkSettingAction

Specifies the kind of modification applied to a setting in a GdkEventSetting.

new

a setting was added.

changed

a setting was changed.

deleted

a setting was deleted.

GdkOwnerChange

Specifies why a selection ownership was changed.

new-owner

some other app claimed the ownership

destroy

the window was destroyed

close

the client was closed

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/gdk-Event-Structures.html

[Package RGtk2 version 2.20.31 Index]