gdk-General {RGtk2}	R Documentation

General

Description

Library initialization and miscellaneous functions

Methods and Functions

gdkNotifyStartupComplete()
gdkNotifyStartupCompleteWithId(id)
gdkGetProgramClass()
gdkSetProgramClass(program.class)
gdkGetDisplay()
gdkFlush()
gdkScreenWidth()
gdkScreenHeight()
gdkScreenWidthMm()
gdkScreenHeightMm()
gdkPointerGrab(window, owner.events = FALSE, event.mask = 0, confine.to = NULL, cursor = NULL, time = "GDK_CURRENT_TIME")
gdkPointerUngrab(time = "GDK_CURRENT_TIME")
gdkPointerIsGrabbed()
gdkSetDoubleClickTime(msec)
gdkKeyboardGrab(window, owner.events = FALSE, time = "GDK_CURRENT_TIME")
gdkKeyboardUngrab(time = "GDK_CURRENT_TIME")
gdkBeep()

Detailed Description

This section describes the GDK initialization functions and miscellaneous utility functions.

Enums and Flags

GdkGrabStatus

Returned by gdkPointerGrab and gdkKeyboardGrab to indicate success or the reason for the failure of the grab attempt.

success

the resource was successfully grabbed.

already-grabbed

the resource is actively grabbed by another client.

invalid-time

the resource was grabbed more recently than the specified time.

not-viewable

the grab window or the confine.to window are not viewable.

frozen

the resource is frozen by an active grab of another client.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/gdk-General.html

[Package RGtk2 version 2.20.31 Index]