gdk-Input-Devices {RGtk2}	R Documentation

Input Devices

Description

Functions for handling extended input devices

Methods and Functions

gdkDevicesList()
gdkDeviceSetSource(object, source)
gdkDeviceSetMode(object, mode)
gdkDeviceSetKey(object, index, keyval, modifiers)
gdkDeviceSetAxisUse(object, index, use)
gdkDeviceGetCorePointer()
gdkDeviceGetState(object, window)
gdkDeviceGetHistory(object, window, start, stop)
gdkDeviceGetAxis(object, axes, use)
gdkInputSetExtensionEvents(object, mask, mode)

Detailed Description

In addition to the normal keyboard and mouse input devices, GTK+ also contains support for extended input devices. In particular, this support is targeted at graphics tablets. Graphics tablets typically return sub-pixel positioning information and possibly information about the pressure and tilt of the stylus. Under X, the support for extended devices is done through the XInput extension.

Because handling extended input devices may involve considerable overhead, they need to be turned on for each GdkWindow individually using gdkInputSetExtensionEvents. (Or, more typically, for GtkWidgets, using gtkWidgetSetExtensionEvents). As an additional complication, depending on the support from the windowing system, its possible that a normal mouse cursor will not be displayed for a particular extension device. If an application does not want to deal with displaying a cursor itself, it can ask only to get extension events from devices that will display a cursor, by passing the GDK_EXTENSION_EVENTS_CURSOR value to gdkInputSetExtensionEvents. Otherwise, the application must retrieve the device information using gdkDevicesList, check the has_cursor field, and, if it is FALSE, draw a cursor itself when it receives motion events.

Each pointing device is assigned a unique integer ID; events from a particular device can be identified by the deviceid field in the event structure. The events generated by pointer devices have also been extended to contain pressure, xtilt and ytilt fields which contain the extended information reported as additional valuators from the device. The pressure field is a a double value ranging from 0.0 to 1.0, while the tilt fields are double values ranging from -1.0 to 1.0. (With -1.0 representing the maximum tilt to the left or up, and 1.0 representing the maximum tilt to the right or down.)

One additional field in each event is the source field, which contains an enumeration value describing the type of device; this currently can be one of GDK_SOURCE_MOUSE, GDK_SOURCE_PEN, GDK_SOURCE_ERASER, or GDK_SOURCE_CURSOR. This field is present to allow simple applications to (for instance) delete when they detect eraser devices without having to keep track of complicated per-device settings.

Various aspects of each device may be configured. The configuration of devices is queried using gdkDevicesList. Each device must be activated using gdkDeviceSetMode, which also controls whether the device's range is mapped to the entire screen or to a single window. The mapping of the valuators of the device onto the predefined valuator types is set using gdkDeviceSetAxisUse. And the source type for each device can be set with gdkDeviceSetSource.

Devices may also have associated keys or function buttons. Such keys can be globally set to map into normal X keyboard events. The mapping is set using gdkDeviceSetKey.

The interfaces in this section will most likely be considerably modified in the future to accomodate devices that may have different sets of additional valuators than the pressure xtilt and ytilt.

Structures

GdkDevice

A GdkDevice structure contains a detailed description of an extended input device. All fields are read-only; but you can use gdkDeviceSetSource, gdkDeviceSetMode, gdkDeviceSetKey and gdkDeviceSetAxisUse to configure various aspects of the device.

name

[character] the parent instance

source

[GdkInputSource] the name of this device.

mode

[GdkInputMode] the type of this device.

hasCursor

[logical] the mode of this device

numAxes

[integer] TRUE if the pointer follows device motion.

axes

[GdkDeviceAxis] the length of the axes list.

numKeys

[integer] a list of GdkDeviceAxis, describing the axes of this device.

keys

[GdkDeviceKey] the length of the keys list.

GdkDeviceKey

The GdkDeviceKey structure contains information about the mapping of one device function button onto a normal X key event. It has the following fields:

keyval

[numeric] the keyval to generate when the function button is pressed. If this is 0, no keypress will be generated.

modifiers

[GdkModifierType] the modifiers set for the generated key event.

GdkDeviceAxis

The GdkDeviceAxis structure contains information about the range and mapping of a device axis.

use

[GdkAxisUse] specifies how the axis is used.

min

[numeric] the minimal value that will be reported by this axis.

max

[numeric] the maximal value that will be reported by this axis.

GdkTimeCoord

The GdkTimeCoord structure stores a single event in a motion history. It contains the following fields: GdkTimeCoord is a transparent-type.

time

The timestamp for this event.

axes

the values of the device's axes.

Enums and Flags

GdkInputSource

An enumeration describing the type of an input device in general terms.

mouse

the device is a mouse. (This will be reported for the core pointer, even if it is something else, such as a trackball.)

pen

the device is a stylus of a graphics tablet or similar device.

eraser

the device is an eraser. Typically, this would be the other end of a stylus on a graphics tablet.

cursor

the device is a graphics tablet "puck" or similar device.

GdkInputMode

An enumeration that describes the mode of an input device.

disabled

the device is disabled and will not report any events.

screen

the device is enabled. The device's coordinate space maps to the entire screen.

window

the device is enabled. The device's coordinate space is mapped to a single window. The manner in which this window is chosen is undefined, but it will typically be the same way in which the focus window for key events is determined.

GdkAxisUse

An enumeration describing the way in which a device axis (valuator) maps onto the predefined valuator types that GTK+ understands.

ignore

the axis is ignored.

x

the axis is used as the x axis.

y

the axis is used as the y axis.

pressure

the axis is used for pressure information.

xtilt

the axis is used for x tilt information.

ytilt

the axis is used for x tilt information.

wheel

the axis is used for wheel information.

last

a constant equal to the numerically highest axis value.

GdkExtensionMode

An enumeration used to specify which extension events are desired for a particular widget.

none

no extension events are desired.

all

all extension events are desired.

cursor

extension events are desired only if a cursor will be displayed for the device.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/gdk-Input-Devices.html

[Package RGtk2 version 2.20.31 Index]