gdk-Keyboard-Handling {RGtk2}	R Documentation

Key Values

Description

Functions for manipulating keyboard codes

Methods and Functions

gdkKeymapGetDefault()
gdkKeymapGetForDisplay(display)
gdkKeymapLookupKey(object, key)
gdkKeymapTranslateKeyboardState(object, hardware.keycode, state, group)
gdkKeymapGetEntriesForKeyval(object, keyval)
gdkKeymapGetEntriesForKeycode(object, hardware.keycode)
gdkKeymapGetDirection(object)
gdkKeymapHaveBidiLayouts(object)
gdkKeymapGetCapsLockState(object)
gdkKeymapAddVirtualModifiers(object)
gdkKeymapMapVirtualModifiers(object)
gdkKeyvalName(keyval)
gdkKeyvalFromName(keyval.name)
gdkKeyvalConvertCase(symbol)
gdkKeyvalToUpper(keyval)
gdkKeyvalToLower(keyval)
gdkKeyvalIsUpper(keyval)
gdkKeyvalIsLower(keyval)
gdkKeyvalToUnicode(keyval)
gdkUnicodeToKeyval(wc)

Hierarchy

GObject
   +----GdkKeymap

Detailed Description

Key values are the codes which are sent whenever a key is pressed or released. They appear in the keyval field of the GdkEventKey structure, which is passed to signal handlers for the "key-press-event" and "key-release-event" signals. The complete list of key values can be found in the ‘<gdk/gdkkeysyms.h>’ header file. ‘<gdk/gdkkeysyms.h>’ is not included in ‘<gdk/gdk.h>’, it must be included independently, because the file is quite large.

Key values are regularly updated from the upstream X.org X11 implementation, so new values are added regularly. They will be prefixed with GDK_ rather than XF86XK_ or XK_ (for older symbols).

Key values can be converted into a string representation using gdkKeyvalName. The reverse function, converting a string to a key value, is provided by gdkKeyvalFromName.

The case of key values can be determined using gdkKeyvalIsUpper and gdkKeyvalIsLower. Key values can be converted to upper or lower case using gdkKeyvalToUpper and gdkKeyvalToLower.

When it makes sense, key values can be converted to and from Unicode characters with gdkKeyvalToUnicode and gdkUnicodeToKeyval.

One GdkKeymap object exists for each user display. gdkKeymapGetDefault returns the GdkKeymap for the default display; to obtain keymaps for other displays, use gdkKeymapGetForDisplay. A keymap is a mapping from GdkKeymapKey to key values. You can think of a GdkKeymapKey as a representation of a symbol printed on a physical keyboard key. That is, it contains three pieces of information. First, it contains the hardware keycode; this is an identifying number for a physical key. Second, it contains the level of the key. The level indicates which symbol on the key will be used, in a vertical direction. So on a standard US keyboard, the key with the number "1" on it also has the exclamation point ("!") character on it. The level indicates whether to use the "1" or the "!" symbol. The letter keys are considered to have a lowercase letter at level 0, and an uppercase letter at level 1, though only the uppercase letter is printed. Third, the GdkKeymapKey contains a group; groups are not used on standard US keyboards, but are used in many other countries. On a keyboard with groups, there can be 3 or 4 symbols printed on a single key. The group indicates movement in a horizontal direction. Usually groups are used for two different languages. In group 0, a key might have two English characters, and in group 1 it might have two Hebrew characters. The Hebrew characters will be printed on the key next to the English characters.

In order to use a keymap to interpret a key event, it's necessary to first convert the keyboard state into an effective group and level. This is done via a set of rules that varies widely according to type of keyboard and user configuration. The function gdkKeymapTranslateKeyboardState accepts a keyboard state – consisting of hardware keycode pressed, active modifiers, and active group – applies the appropriate rules, and returns the group/level to be used to index the keymap, along with the modifiers which did not affect the group and level. i.e. it returns "unconsumed modifiers." The keyboard group may differ from the effective group used for keymap lookups because some keys don't have multiple groups - e.g. the Enter key is always in group 0 regardless of keyboard state.

Note that gdkKeymapTranslateKeyboardState also returns the keyval, i.e. it goes ahead and performs the keymap lookup in addition to telling you which effective group/level values were used for the lookup. GdkEventKey already contains this keyval, however, so you don't normally need to call gdkKeymapTranslateKeyboardState just to get the keyval.

Structures

GdkKeymap

A GdkKeymap defines the translation from keyboard state (including a hardware key, a modifier mask, and active keyboard group) to a keyval. This translation has two phases. The first phase is to determine the effective keyboard group and level for the keyboard state; the second phase is to look up the keycode/group/level triplet in the keymap and see what keyval it corresponds to.

GdkKeymapKey

A GdkKeymapKey is a hardware key that can be mapped to a keyval. GdkKeymapKey is a transparent-type.

keycode

the hardware keycode. This is an identifying number for a physical key.

group

indicates movement in a horizontal direction. Usually groups are used for two different languages. In group 0, a key might have two English characters, and in group 1 it might have two Hebrew characters. The Hebrew characters will be printed on the key next to the English characters.

level

indicates which symbol on the key will be used, in a vertical direction. So on a standard US keyboard, the key with the number "1" on it also has the exclamation point ("!") character on it. The level indicates whether to use the "1" or the "!" symbol. The letter keys are considered to have a lowercase letter at level 0, and an uppercase letter at level 1, though only the uppercase letter is printed.

Signals

direction-changed(keymap, user.data)

The ::direction-changed signal gets emitted when the direction of the keymap changes. Since 2.0

keymap

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

keys-changed(keymap, user.data)

The ::keys-changed signal is emitted when the mapping represented by keymap changes. Since 2.2

keymap

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

state-changed(keymap, user.data)

The ::state-changed signal is emitted when the state of the keyboard changes, e.g when Caps Lock is turned on or off. See gdkKeymapGetCapsLockState. Since 2.16

keymap

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

Note

The keyval constants exist in RGtk2 as .gdkKeyvalName, so .gdkPlus for plus.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/gdk-Keyboard-Handling.html

[Package RGtk2 version 2.20.31 Index]