gdk-Properties-and-Atoms {RGtk2}	R Documentation

Properties and Atoms

Description

Functions to manipulate properties on windows

Methods and Functions

gdkAtomIntern(atom.name, only.if.exists = FALSE)
gdkAtomInternStaticString(atom.name)
gdkAtomName(atom)
gdkPropertyGet(object, property, type, offset, length, pdelete)
gdkPropertyChange(object, property, type, format, mode, data)
gdkPropertyDelete(object, property)

Detailed Description

Each window under X can have any number of associated properties attached to it. Properties are arbitrary chunks of data identified by atoms. (An atom is a numeric index into a string table on the X server. They are used to transfer strings efficiently between clients without having to transfer the entire string.) A property has an associated type, which is also identified using an atom.

A property has an associated format, an integer describing how many bits are in each unit of data inside the property. It must be 8, 16, or 32. When data is transferred between the server and client, if they are of different endianesses it will be byteswapped as necessary according to the format of the property. Note that on the client side, properties of format 32 will be stored with one unit per long, even if a long integer has more than 32 bits on the platform. (This decision was apparently made for Xlib to maintain compatibility with programs that assumed longs were 32 bits, at the expense of programs that knew better.)

The functions in this section are used to add, remove and change properties on windows, to convert atoms to and from strings and to manipulate some types of data commonly stored in X window properties.

Structures

GdkAtom

An opaque type representing a string as an index into a table of strings on the X server. GdkAtom is a transparent-type.

Enums and Flags

GdkPropMode

Describes how existing data is combined with new data when using gdkPropertyChange.

replace

the new data replaces the existing data.

prepend

the new data is prepended to the existing data.

append

the new data is appended to the existing data.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/gdk-Properties-and-Atoms.html

[Package RGtk2 version 2.20.31 Index]