gdk-Visuals {RGtk2}	R Documentation

Visuals

Description

Low-level display hardware information

Methods and Functions

gdkQueryDepths()
gdkQueryVisualTypes()
gdkListVisuals()
gdkVisualGetBestDepth()
gdkVisualGetBestType()
gdkVisualGetSystem()
gdkVisualGetBest()
gdkVisualGetBestWithDepth(depth)
gdkVisualGetBestWithType(visual.type)
gdkVisualGetBestWithBoth(depth, visual.type)
gdkVisualGetScreen(object)

Detailed Description

A GdkVisual describes a particular video hardware display format. It includes information about the number of bits used for each color, the way the bits are translated into an RGB value for display, and the way the bits are stored in memory. For example, a piece of display hardware might support 24-bit color, 16-bit color, or 8-bit color; meaning 24/16/8-bit pixel sizes. For a given pixel size, pixels can be in different formats; for example the "red" element of an RGB pixel may be in the top 8 bits of the pixel, or may be in the lower 4 bits.

Usually you can avoid thinking about visuals in GTK+. Visuals are useful to interpret the contents of a GdkImage, but you should avoid GdkImage precisely because its contents depend on the display hardware; use GdkPixbuf instead, for all but the most low-level purposes. Also, anytime you provide a GdkColormap, the visual is implied as part of the colormap (gdkColormapGetVisual), so you won't have to provide a visual in addition.

There are several standard visuals. The visual returned by gdkVisualGetSystem is the system's default visual. gdkRgbGetVisual return the visual most suited to displaying full-color image data. If you use the calls in GdkRGB, you should create your windows using this visual (and the colormap returned by gdkRgbGetColormap).

A number of functions are provided for determining the "best" available visual. For the purposes of making this determination, higher bit depths are considered better, and for visuals of the same bit depth, GDK_VISUAL_PSEUDO_COLOR is preferred at 8bpp, otherwise, the visual types are ranked in the order of (highest to lowest) GDK_VISUAL_DIRECT_COLOR, GDK_VISUAL_TRUE_COLOR, GDK_VISUAL_PSEUDO_COLOR, GDK_VISUAL_STATIC_COLOR, GDK_VISUAL_GRAYSCALE, then GDK_VISUAL_STATIC_GRAY.

Structures

GdkVisual

The GdkVisual structure contains information about a particular visual.

type

[GdkVisualType] inherited portion from GObject

depth

[integer] The type of this visual.

byteOrder

[GdkByteOrder] The number of bits per pixel.

colormapSize

[integer] The byte-order for this visual.

bitsPerRgb

[integer] The number of entries in the colormap, for visuals of type GDK_VISUAL_PSEUDO_COLOR or GDK_VISUAL_GRAY_SCALE. For other visual types, it is the number of possible levels per color component. If the visual has different numbers of levels for different components, the value of this field is undefined.

redMask

[numeric] The number of significant bits per red, green, or blue when specifying colors for this visual. (For instance, for gdkColormapAllocColor)

redShift

[integer] A mask giving the bits in a pixel value that correspond to the red field. Significant only for GDK_VISUAL_PSEUDOCOLOR and GDK_VISUAL_DIRECTCOLOR.

redPrec

[integer] The red_shift and red_prec give an alternate presentation of the information in red_mask. red_mask is a contiguous sequence of red_prec bits starting at bit number red_shift. For example, shows constructing a pixel value out of three 16 bit color values.

greenMask

[numeric] See above.

greenShift

[integer] A mask giving the bits in a pixel value that correspond to the green field.

greenPrec

[integer] The green_shift and green_prec give an alternate presentation of the information in green_mask.

blueMask

[numeric] See above.

blueShift

[integer] A mask giving the bits in a pixel value that correspond to the blue field.

bluePrec

[integer] The blue_shift and blue_prec give an alternate presentation of the information in blue_mask.

Enums and Flags

GdkVisualType

A set of values that describe the manner in which the pixel values for a visual are converted into RGB values for display.

static-gray

Each pixel value indexes a grayscale value directly.

grayscale

Each pixel is an index into a color map that maps pixel values into grayscale values. The color map can be changed by an application.

static-color

Each pixel value is an index into a predefined, unmodifiable color map that maps pixel values into RGB values.

pseudo-color

Each pixel is an index into a color map that maps pixel values into rgb values. The color map can be changed by an application.

true-color

Each pixel value directly contains red, green, and blue components. The red_mask, green_mask, and blue_mask fields of the GdkVisual structure describe how the components are assembled into a pixel value.

direct-color

Each pixel value contains red, green, and blue components as for GDK_VISUAL_TRUE_COLOR, but the components are mapped via a color table into the final output table instead of being converted directly.

GdkByteOrder

A set of values describing the possible byte-orders for storing pixel values in memory.

lsb-first

The values are stored with the least-significant byte first. For instance, the 32-bit value 0xffeecc would be stored in memory as 0xcc, 0xee, 0xff, 0x00.

msb-first

The values are stored with the most-significant byte first. For instance, the 32-bit value 0xffeecc would be stored in memory as 0x00, 0xcc, 0xee, 0xff.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk/gdk-Visuals.html

See Also

GdkImage GdkColormap

[Package RGtk2 version 2.20.31 Index]