gdk-pixbuf-File-Loading {RGtk2}	R Documentation

File Loading

Description

Loading a pixbuf from a file.

Methods and Functions

gdkPixbufNewFromFile(filename, .errwarn = TRUE)
gdkPixbufNewFromFileAtSize(filename, width, height, .errwarn = TRUE)
gdkPixbufNewFromFileAtScale(filename, width, height, preserve.aspect.ratio, .errwarn = TRUE)
gdkPixbufGetFileInfo(filename)
gdkPixbufNewFromStream(stream, cancellable = NULL, .errwarn = TRUE)
gdkPixbufNewFromStreamAtScale(stream, width = -1, height = -1, preserve.aspect.ratio = 1, cancellable = NULL, .errwarn = TRUE)

Detailed Description

The gdk-pixbuf library provides a simple mechanism for loading an image from a file in synchronous fashion. This means that the library takes control of the application while the file is being loaded; from the user's point of view, the application will block until the image is done loading.

This interface can be used by applications in which blocking is acceptable while an image is being loaded. It can also be used to load small images in general. Applications that need progressive loading can use the GdkPixbufLoader functionality instead.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk-pixbuf/gdk-pixbuf-File-Loading.html

[Package RGtk2 version 2.20.31 Index]