gdk-pixbuf-File-saving {RGtk2}	R Documentation

File saving

Description

Saving a pixbuf to a file.

Methods and Functions

gdkPixbufSavev(object, filename, type, option.keys, option.values, .errwarn = TRUE)
gdkPixbufSave(object, filename, type, ..., .errwarn = TRUE)
gdkPixbufSaveToCallback(object, save.func, user.data, type, ..., .errwarn = TRUE)
gdkPixbufSaveToCallbackv(object, save.func, user.data, type, option.keys, option.values, .errwarn = TRUE)
gdkPixbufSaveToBuffer(object, type, ..., .errwarn = TRUE)
gdkPixbufSaveToBufferv(object, type, option.keys, option.values, .errwarn = TRUE)
gdkPixbufSaveToStream(object, stream, type, cancellable, .errwarn = TRUE)

Detailed Description

These functions allow to save a GdkPixbuf in a number of file formats. The formatted data can be written to a file or to a memory buffer. &gdk-pixbuf; can also call a user-defined callback on the data, which allows to e.g. write the image to a socket or store it in a database.

User Functions

GdkPixbufSaveFunc(buf, count, error, data)

Specifies the type of the function passed to gdkPixbufSaveToCallback. It is called once for each block of bytes that is "written" by gdkPixbufSaveToCallback. If successful it should return TRUE. If an error occurs it should set error and return FALSE, in which case gdkPixbufSaveToCallback will fail with the same error. Since 2.4

buf

bytes to be written.

count

number of bytes in buf.

error

A location to return an error. [ out ]

data

user data passed to gdkPixbufSaveToCallback. [ closure ]

Returns: [logical] TRUE if successful, FALSE (with error set) if failed.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk-pixbuf/gdk-pixbuf-File-saving.html

[Package RGtk2 version 2.20.31 Index]