gdk-pixbuf-animation {RGtk2}	R Documentation

Animations

Description

Animated images.

Methods and Functions

gdkPixbufAnimationGetWidth(object)
gdkPixbufAnimationGetHeight(object)
gdkPixbufAnimationGetIter(object, start.time)
gdkPixbufAnimationIsStaticImage(object)
gdkPixbufAnimationGetStaticImage(object)
gdkPixbufAnimationIterAdvance(object, current.time)
gdkPixbufAnimationIterGetDelayTime(object)
gdkPixbufAnimationIterOnCurrentlyLoadingFrame(object)
gdkPixbufAnimationIterGetPixbuf(object)
gdkPixbufSimpleAnimNew(width, height, rate)
gdkPixbufSimpleAnimAddFrame(object, pixbuf)
gdkPixbufSimpleAnimSetLoop(object, loop)
gdkPixbufSimpleAnimGetLoop(object)
gdkPixbufAnimation(filename, .errwarn = TRUE)
gdkPixbufSimpleAnim(width, height, rate)

Hierarchy

GObject
   +----GdkPixbufAnimation
         +----GdkPixbufSimpleAnim
GObject
   +----GdkPixbufAnimationIter
GObject
   +----GdkPixbufAnimation
         +----GdkPixbufSimpleAnim

Detailed Description

The gdk-pixbuf library provides a simple mechanism to load and represent animations. An animation is conceptually a series of frames to be displayed over time. Each frame is the same size. The animation may not be represented as a series of frames internally; for example, it may be stored as a sprite and instructions for moving the sprite around a background. To display an animation you don't need to understand its representation, however; you just ask gdk-pixbuf what should be displayed at a given point in time.

Structures

GdkPixbufAnimation

An opaque struct representing an animation.

GdkPixbufAnimationIter

An opaque struct representing an iterator which points to a certain position in an animation.

GdkPixbufSimpleAnim

An opaque struct representing a simple animation.

Convenient Construction

gdkPixbufAnimation is the equivalent of gdkPixbufAnimationNewFromFile. gdkPixbufSimpleAnim is the equivalent of gdkPixbufSimpleAnimNew.

Properties

loop [logical : Read / Write]

Whether the animation should loop when it reaches the end. Default value: FALSE Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk-pixbuf/gdk-pixbuf-animation.html

See Also

GdkPixbufLoader

[Package RGtk2 version 2.20.31 Index]