gdk-pixbuf-util {RGtk2}	R Documentation

Utilities

Description

Utility and miscellaneous convenience functions.

Methods and Functions

gdkPixbufAddAlpha(object, substitute.color, r, g, b)
gdkPixbufCopyArea(object, src.x, src.y, width, height, dest.pixbuf, dest.x, dest.y)
gdkPixbufSaturateAndPixelate(object, dest, saturation, pixelate)
gdkPixbufApplyEmbeddedOrientation(object)
gdkPixbufFill(object, pixel)

Detailed Description

These functions provide miscellaneous utilities for manipulating pixbufs. The pixel data in pixbufs may of course be manipulated directly by applications, but several common operations can be performed by these functions instead.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gdk-pixbuf/gdk-pixbuf-util.html

See Also

GdkPixbuf

[Package RGtk2 version 2.20.31 Index]