gdkAppLaunchContextSetDesktop {RGtk2}	R Documentation

gdkAppLaunchContextSetDesktop

Description

Sets the workspace on which applications will be launched when using this context when running under a window manager that supports multiple workspaces, as described in the Extended Window Manager Hints (http://www.freedesktop.org/Standards/wm-spec).

Usage

gdkAppLaunchContextSetDesktop(object, desktop)

Arguments

object
a GdkAppLaunchContext

desktop
the number of a workspace, or -1

Details

When the workspace is not specified or desktop is set to -1, it is up to the window manager to pick one, typically it will be the current workspace. Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]