gdkAppLaunchContextSetIconName {RGtk2}	R Documentation

gdkAppLaunchContextSetIconName

Description

Sets the icon for applications that are launched with this context. The icon.name will be interpreted in the same way as the Icon field in desktop files. See also gdkAppLaunchContextSetIcon.

Usage

gdkAppLaunchContextSetIconName(object, icon.name = NULL)

Arguments

object
a GdkAppLaunchContext

icon.name
an icon name, or NULL. [ allow-none ]

Details

If both icon and icon.name are set, the icon.name takes priority. If neither icon or icon.name is set, the icon is taken from either the file that is passed to launched application or from the GAppInfo for the launched application itself. Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]