gdkAppLaunchContextSetScreen {RGtk2}	R Documentation

gdkAppLaunchContextSetScreen

Description

Sets the screen on which applications will be launched when using this context. See also gdkAppLaunchContextSetDisplay.

Usage

gdkAppLaunchContextSetScreen(object, screen)

Arguments

object
a GdkAppLaunchContext

screen
a GdkScreen

Details

If both screen and display are set, the screen takes priority. If neither screen or display are set, the default screen and display are used. Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]