gdkAppLaunchContextSetTimestamp {RGtk2}	R Documentation

gdkAppLaunchContextSetTimestamp

Description

Sets the timestamp of context. The timestamp should ideally be taken from the event that triggered the launch.

Usage

gdkAppLaunchContextSetTimestamp(object, timestamp)

Arguments

object
a GdkAppLaunchContext

timestamp
a timestamp

Details

Window managers can use this information to avoid moving the focus to the newly launched application when the user is busy typing in another window. This is also known as 'focus stealing prevention'. Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]