gdkAtomIntern {RGtk2}	R Documentation

gdkAtomIntern

Description

Finds or creates an atom corresponding to a given string.

Usage

gdkAtomIntern(atom.name, only.if.exists = FALSE)

Arguments

atom.name
a string.

only.if.exists
if TRUE, GDK is allowed to not create a new atom, but just return GDK_NONE if the requested atom doesn't already exists. Currently, the flag is ignored, since checking the existance of an atom is as expensive as creating it.

Value

[GdkAtom] the atom corresponding to atom.name.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]