gdkAtomInternStaticString {RGtk2}	R Documentation

gdkAtomInternStaticString

Description

Finds or creates an atom corresponding to a given string.

Usage

gdkAtomInternStaticString(atom.name)

Arguments

atom.name
a static string

Details

Note that this function is identical to gdkAtomIntern except that if a new GdkAtom is created the string itself is used rather than a copy. This saves memory, but can only be used if the string will always exist. It can be used with statically allocated strings in the main program, but not with statically allocated memory in dynamically loaded modules, if you expect to ever unload the module again (e.g. do not use this function in GTK+ theme engines). Since 2.10

Value

[GdkAtom] the atom corresponding to atom.name

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]