gdkBitmapCreateFromData {RGtk2}	R Documentation

gdkBitmapCreateFromData

Description

Creates a new bitmap from data in XBM format.

Usage

gdkBitmapCreateFromData(drawable = NULL, data, width, height)

Arguments

drawable
a GdkDrawable, used to determine default values for the new pixmap. Can be NULL, in which case the root window is used.

data
a pointer to the XBM data.

width
the width of the new pixmap in pixels.

height
the height of the new pixmap in pixels.

Value

[GdkBitmap] the GdkBitmap

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]