gdkCairoCreate {RGtk2}	R Documentation

gdkCairoCreate

Description

Creates a Cairo context for drawing to drawable.

Usage

gdkCairoCreate(drawable)

Arguments

drawable
a GdkDrawable

Details

PLEASE NOTE: Note that due to double-buffering, Cairo contexts created in a GTK+ expose event handler cannot be cached and reused between different expose events. Since 2.8

Value

[Cairo] A newly created Cairo context.

Note

Note that due to double-buffering, Cairo contexts created in a GTK+ expose event handler cannot be cached and reused between different expose events.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]