gdkCairoResetClip {RGtk2}	R Documentation

gdkCairoResetClip

Description

Resets the clip region for a Cairo context created by gdkCairoCreate.

Usage

gdkCairoResetClip(cr, drawable)

Arguments

cr
a Cairo

drawable
a GdkDrawable

Details

This resets the clip region to the "empty" state for the given drawable. This is required for non-native windows since a direct call to cairoResetClip would unset the clip region inherited from the drawable (i.e. the window clip region), and thus let you e.g. draw outside your window.

This is rarely needed though, since most code just create a new cairo_t using gdkCairoCreate each time they want to draw something. Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]