gdkCairoSetSourcePixbuf {RGtk2}	R Documentation

gdkCairoSetSourcePixbuf

Description

Sets the given pixbuf as the source pattern for the Cairo context. The pattern has an extend mode of CAIRO_EXTEND_NONE and is aligned so that the origin of pixbuf is pixbuf.x, pixbuf.y

Usage

gdkCairoSetSourcePixbuf(cr, pixbuf, pixbuf.x, pixbuf.y)

Arguments

cr
a Cairo context

pixbuf
a GdkPixbuf

pixbuf.x
X coordinate of location to place upper left corner of pixbuf

pixbuf.y
Y coordinate of location to place upper left corner of pixbuf

Details

Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]