gdkCairoSetSourcePixmap {RGtk2}	R Documentation

gdkCairoSetSourcePixmap

Description

Sets the given pixmap as the source pattern for the Cairo context. The pattern has an extend mode of CAIRO_EXTEND_NONE and is aligned so that the origin of pixmap is pixmap.x, pixmap.y

Usage

gdkCairoSetSourcePixmap(cr, pixmap, pixmap.x, pixmap.y)

Arguments

cr
a Cairo context

pixmap
a GdkPixmap

pixmap.x
X coordinate of location to place upper left corner of pixmap

pixmap.y
Y coordinate of location to place upper left corner of pixmap

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]