gdkCharHeight {RGtk2}	R Documentation

gdkCharHeight

Description

Determines the total height of a given character. This value is not generally useful, because you cannot determine how this total height will be drawn in relation to the baseline. See gdkTextExtents. WARNING: gdk_char_height has been deprecated since version 2.2 and should not be used in newly-written code. Use gdkTextExtents instead.

Usage

gdkCharHeight(object, character)

Arguments

object
a GdkFont

character
the character to measure.

Value

[integer] the height of the character in pixels.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]