gdkColorParse {RGtk2}	R Documentation

gdkColorParse

Description

Parses a textual specification of a color and fill in the red, green, and blue fields of a GdkColor structure. The color is not allocated, you must call gdkColormapAllocColor yourself. The string can either one of a large set of standard names. (Taken from the X11 ‘rgb.txt’ file), or it can be a hex value in the form '#rgb' '#rrggbb' '#rrrgggbbb' or '#rrrrggggbbbb' where 'r', 'g' and 'b' are hex digits of the red, green, and blue components of the color, respectively. (White in the four forms is '#fff' '#ffffff' '#fffffffff' and '#ffffffffffff')

Usage

gdkColorParse(spec)

Arguments

spec
the string specifying the color.

Value

A list containing the following elements:

retval
[logical] TRUE if the parsing succeeded.

color
the GdkColor to fill in. [ out ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]