gdkColorToString {RGtk2}	R Documentation

gdkColorToString

Description

Returns a textual specification of color in the hexadecimal form #rrrrggggbbbb, where r, g and b are hex digits representing the red, green and blue components respectively.

Usage

gdkColorToString(object)

Arguments

object
a GdkColor

Details

Since 2.12

Value

[character] a newly-allocated text string

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]