gdkColormapAllocColor {RGtk2}	R Documentation

gdkColormapAllocColor

Description

Allocates a single color from a colormap.

Usage

gdkColormapAllocColor(object, color, writeable, best.match)

Arguments

object
a GdkColormap.

color
the color to allocate. On return the pixel field will be filled in if allocation succeeds.

writeable
If TRUE, the color is allocated writeable (their values can later be changed using gdkColorChange). Writeable colors cannot be shared between applications.

best.match
If TRUE, GDK will attempt to do matching against existing colors if the color cannot be allocated as requested.

Value

[logical] TRUE if the allocation succeeded.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]