gdkColormapAllocColors {RGtk2}	R Documentation

gdkColormapAllocColors

Description

Allocates colors from a colormap.

Usage

gdkColormapAllocColors(colormap, colors, writeable, best.match)

Arguments

colors
The color values to allocate. On return, the pixel values for allocated colors will be filled in.

writeable
If TRUE, the colors are allocated writeable (their values can later be changed using gdkColorChange). Writeable colors cannot be shared between applications.

best.match
If TRUE, GDK will attempt to do matching against existing colors if the colors cannot be allocated as requested.

Value

A list containing the following elements:

retval
[integer] The number of colors that were not successfully allocated.

success
An list of length ncolors. On return, this indicates whether the corresponding color in colors was successfully allocated or not.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]