gdkColormapGetSystemSize {RGtk2}	R Documentation

gdkColormapGetSystemSize

Description

Returns the size of the system's default colormap. (See the description of struct GdkColormap for an explanation of the size of a colormap.) WARNING: gdk_colormap_get_system_size is deprecated and should not be used in newly-written code.

Usage

gdkColormapGetSystemSize()

Value

[integer] the size of the system's default colormap.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]