gdkColormapNew {RGtk2}	R Documentation

gdkColormapNew

Description

Creates a new colormap for the given visual.

Usage

gdkColormapNew(visual, allocate)

Arguments

visual
a GdkVisual.

allocate
if TRUE, the newly created colormap will be a private colormap, and all colors in it will be allocated for the applications use.

Value

[GdkColormap] the new GdkColormap.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]