gdkColormapQueryColor {RGtk2}	R Documentation

gdkColormapQueryColor

Description

Locates the RGB color in colormap corresponding to the given hardware pixel pixel. pixel must be a valid pixel in the colormap; it's a programmer error to call this function with a pixel which is not in the colormap. Hardware pixels are normally obtained from gdkColormapAllocColors, or from a GdkImage. (A GdkImage contains image data in hardware format, a GdkPixbuf contains image data in a canonical 24-bit RGB format.)

Usage

gdkColormapQueryColor(object, pixel)

Arguments

object
a GdkColormap

pixel
pixel value in hardware display format

Details

This function is rarely useful; it's used for example to implement the eyedropper feature in GtkColorSelection.

Value

A list containing the following elements:

result
GdkColor with red, green, blue fields initialized

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]