gdkCursorGetImage {RGtk2}	R Documentation

gdkCursorGetImage

Description

Returns a GdkPixbuf with the image used to display the cursor.

Usage

gdkCursorGetImage(object)

Arguments

object
a GdkCursor

Details

Note that depending on the capabilities of the windowing system and on the cursor, GDK may not be able to obtain the image data. In this case, NULL is returned. Since 2.8

Value

[GdkPixbuf] a GdkPixbuf representing cursor, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]