gdkCursorNewFromPixbuf {RGtk2}	R Documentation

gdkCursorNewFromPixbuf

Description

Creates a new cursor from a pixbuf.

Usage

gdkCursorNewFromPixbuf(display, source, x, y)

Arguments

display
the GdkDisplay for which the cursor will be created

source
the GdkPixbuf containing the cursor image

x
the horizontal offset of the 'hotspot' of the cursor.

y
the vertical offset of the 'hotspot' of the cursor.

Details

Not all GDK backends support RGBA cursors. If they are not supported, a monochrome approximation will be displayed. The functions gdkDisplaySupportsCursorAlpha and gdkDisplaySupportsCursorColor can be used to determine whether RGBA cursors are supported; gdkDisplayGetDefaultCursorSize and gdkDisplayGetMaximalCursorSize give information about cursor sizes.

On the X backend, support for RGBA cursors requires a sufficently new version of the X Render extension. Since 2.4

Value

[GdkCursor] a new GdkCursor.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]