gdkDeviceGetAxis {RGtk2}	R Documentation

gdkDeviceGetAxis

Description

Interprets a list of double as axis values for a given device, and locates the value in the list for a given axis use.

Usage

gdkDeviceGetAxis(object, axes, use)

Arguments

object
a GdkDevice

axes
pointer to a list of axes

use
the use to look for

Value

A list containing the following elements:

retval
[logical] TRUE if the given axis use was found, otherwise FALSE

value
location to store the found value.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]