gdkDeviceGetHistory {RGtk2}	R Documentation

gdkDeviceGetHistory

Description

Obtains the motion history for a device; given a starting and ending timestamp, return all events in the motion history for the device in the given range of time. Some windowing systems do not support motion history, in which case, FALSE will be returned. (This is not distinguishable from the case where motion history is supported and no events were found.)

Usage

gdkDeviceGetHistory(object, window, start, stop)

Arguments

object
a GdkDevice

window
the window with respect to which which the event coordinates will be reported

start
starting timestamp for range of events to return

stop
ending timestamp for the range of events to return

Value

A list containing the following elements:

retval
[logical] TRUE if the windowing system supports motion history and at least one event was found.

events
location to store a newly-allocated list of GdkTimeCoord, or NULL. [ array length=n_events][ out ][ transfer none ]

n.events
location to store the length of events, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]