gdkDisplayAddClientMessageFilter {RGtk2}	R Documentation

gdkDisplayAddClientMessageFilter

Description

Adds a filter to be called when X ClientMessage events are received. See gdkWindowAddFilter if you are interested in filtering other types of events.

Usage

gdkDisplayAddClientMessageFilter(object, message.type, func, 
    data)

Arguments

object
a GdkDisplay for which this message filter applies

message.type
the type of ClientMessage events to receive. This will be checked against the message.type field of the XClientMessage event struct.

func
the function to call to process the event.

data
user data to pass to func.

Details

Since 2.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]