gdkDisplayGetWindowAtPointer {RGtk2}	R Documentation

gdkDisplayGetWindowAtPointer

Description

Obtains the window underneath the mouse pointer, returning the location of the pointer in that window in win.x, win.y for screen. Returns NULL if the window under the mouse pointer is not known to GDK (for example, belongs to another application).

Usage

gdkDisplayGetWindowAtPointer(object)

Arguments

object
a GdkDisplay

Details

Since 2.2

Value

A list containing the following elements:

retval
[GdkWindow] the window under the mouse pointer, or NULL. [ transfer none ]

win.x
return location for x coordinate of the pointer location relative to the window origin, or NULL. [ out ][ allow-none ]

win.y
return location for y coordinate of the pointer location relative & to the window origin, or NULL. [ out ][ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]