gdkDisplayPeekEvent {RGtk2}	R Documentation

gdkDisplayPeekEvent

Description

Gets a copy of the first GdkEvent in the display's event queue, without removing the event from the queue. (Note that this function will not get more events from the windowing system. It only checks the events that have already been moved to the GDK event queue.)

Usage

gdkDisplayPeekEvent(object)

Arguments

object
a GdkDisplay

Details

Since 2.2

Value

[GdkEvent] a copy of the first GdkEvent on the event queue, or NULL if no events are in the queue.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]