gdkDisplaySetPointerHooks {RGtk2}	R Documentation

gdkDisplaySetPointerHooks

Description

This function allows for hooking into the operation of getting the current location of the pointer on a particular display. This is only useful for such low-level tools as an event recorder. Applications should never have any reason to use this facility.

Usage

gdkDisplaySetPointerHooks(object, new.hooks)

Arguments

object
a GdkDisplay

new.hooks
a table of pointers to functions for getting quantities related to the current pointer position, or NULL to restore the default table.

Details

Since 2.2

Value

[GdkDisplayPointerHooks] the previous pointer hook table

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]