gdkDisplayStoreClipboard {RGtk2}	R Documentation

gdkDisplayStoreClipboard

Description

Issues a request to the clipboard manager to store the clipboard data. On X11, this is a special program that works according to the freedesktop clipboard specification, available at

http://www.freedesktop.org/Standards/clipboard-manager-spec (http://www.freedesktop.org/Standards/clipboard-manager-spec).

Usage

gdkDisplayStoreClipboard(object, clipboard.window, targets)

Arguments

object
a GdkDisplay

clipboard.window
a GdkWindow belonging to the clipboard owner

targets
a list of targets that should be saved, or NULL if all available targets should be saved.

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]