gdkDisplayWarpPointer {RGtk2}	R Documentation

gdkDisplayWarpPointer

Description

Warps the pointer of display to the point x,y on the screen screen, unless the pointer is confined to a window by a grab, in which case it will be moved as far as allowed by the grab. Warping the pointer creates events as if the user had moved the mouse instantaneously to the destination.

Usage

gdkDisplayWarpPointer(object, screen, x, y)

Arguments

object
a GdkDisplay

screen
the screen of display to warp the pointer to

x
the x coordinate of the destination

y
the y coordinate of the destination

Details

Note that the pointer should normally be under the control of the user. This function was added to cover some rare use cases like keyboard navigation support for the color picker in the GtkColorSelectionDialog. Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]