gdkDragFindWindowForScreen {RGtk2}	R Documentation

gdkDragFindWindowForScreen

Description

Finds the destination window and DND protocol to use at the given pointer position.

Usage

gdkDragFindWindowForScreen(object, drag.window, screen, x.root, 
    y.root)

Arguments

object
a GdkDragContext

drag.window
a window which may be at the pointer position, but should be ignored, since it is put up by the drag source as an icon.

screen
the screen where the destination window is sought.

x.root
the x position of the pointer in root coordinates.

y.root
the y position of the pointer in root coordinates.

Details

This function is called by the drag source to obtain the dest.window and protocol parameters for gdkDragMotion. Since 2.2

Value

A list containing the following elements:

dest.window
location to store the destination window in. [ out ]

protocol
location to store the DND protocol in. [ out ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]