gdkDragGetProtocolForDisplay {RGtk2}	R Documentation

gdkDragGetProtocolForDisplay

Description

Finds out the DND protocol supported by a window.

Usage

gdkDragGetProtocolForDisplay(display, xid)

Arguments

display
the GdkDisplay where the destination window resides

xid
the windowing system id of the destination window.

Details

Since 2.2

Value

A list containing the following elements:

retval
[GdkNativeWindow] the windowing system id of the window where the drop should happen. This may be xid or the id of a proxy window, or zero if xid doesn't support Drag and Drop.

protocol
location where the supported DND protocol is returned.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]