gdkDragMotion {RGtk2}	R Documentation

gdkDragMotion

Description

Updates the drag context when the pointer moves or the set of actions changes.

Usage

gdkDragMotion(object, dest.window, protocol, x.root, y.root, 
    suggested.action, possible.actions, time)

Arguments

object
a GdkDragContext.

dest.window
the new destination window, obtained by gdkDragFindWindow.

protocol
the DND protocol in use, obtained by gdkDragFindWindow.

x.root
the x position of the pointer in root coordinates.

y.root
the y position of the pointer in root coordinates.

suggested.action
the suggested action.

possible.actions
the possible actions.

time
the timestamp for this operation.

Details

This function is called by the drag source.

Value

[logical] FIXME

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]