gdkDrawArc {RGtk2}	R Documentation

gdkDrawArc

Description

Draws an arc or a filled 'pie slice'. The arc is defined by the bounding rectangle of the entire ellipse, and the start and end angles of the part of the ellipse to be drawn.

Usage

gdkDrawArc(object, gc, filled, x, y, width, height, angle1, angle2)

Arguments

object
a GdkDrawable (a GdkWindow or a GdkPixmap).

gc
a GdkGC.

filled
TRUE if the arc should be filled, producing a 'pie slice'.

x
the x coordinate of the left edge of the bounding rectangle.

y
the y coordinate of the top edge of the bounding rectangle.

width
the width of the bounding rectangle.

height
the height of the bounding rectangle.

angle1
the start angle of the arc, relative to the 3 o'clock position, counter-clockwise, in 1/64ths of a degree.

angle2
the end angle of the arc, relative to angle1, in 1/64ths of a degree.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]