gdkDrawDrawable {RGtk2}	R Documentation

gdkDrawDrawable

Description

Copies the width x height region of src at coordinates (xsrc, ysrc) to coordinates (xdest, ydest) in drawable. width and/or height may be given as -1, in which case the entire src drawable will be copied.

Usage

gdkDrawDrawable(object, gc, src, xsrc, ysrc, xdest, ydest, width, 
    height)

Arguments

object
a GdkDrawable

gc
a GdkGC sharing the drawable's visual and colormap

src
the source GdkDrawable, which may be the same as drawable

xsrc
X position in src of rectangle to draw

ysrc
Y position in src of rectangle to draw

xdest
X position in drawable where the rectangle should be drawn

ydest
Y position in drawable where the rectangle should be drawn

width
width of rectangle to draw, or -1 for entire src width

height
height of rectangle to draw, or -1 for entire src height

Details

Most fields in gc are not used for this operation, but notably the clip mask or clip region will be honored.

The source and destination drawables must have the same visual and colormap, or errors will result. (On X11, failure to match visual/colormap results in a BadMatch error from the X server.) A common cause of this problem is an attempt to draw a bitmap to a color drawable. The way to draw a bitmap is to set the bitmap as the stipple on the GdkGC, set the fill mode to GDK_STIPPLED, and then draw the rectangle.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]