gdkDrawGlyphs {RGtk2}	R Documentation

gdkDrawGlyphs

Description

This is a low-level function; 99% of text rendering should be done using gdkDrawLayout instead.

Usage

gdkDrawGlyphs(object, gc, font, x, y, glyphs)

Arguments

object
a GdkDrawable

gc
a GdkGC

font
font to be used

x
X coordinate of baseline origin

y
Y coordinate of baseline origin

glyphs
the glyph string to draw

Details

A glyph is a single image in a font. This function draws a sequence of glyphs. To obtain a sequence of glyphs you have to understand a lot about internationalized text handling, which you don't want to understand; thus, use gdkDrawLayout instead of this function, gdkDrawLayout handles the details.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]