gdkDrawGlyphsTransformed {RGtk2}	R Documentation

gdkDrawGlyphsTransformed

Description

Renders a PangoGlyphString onto a drawable, possibly transforming the layed-out coordinates through a transformation matrix. Note that the transformation matrix for font is not changed, so to produce correct rendering results, the font must have been loaded using a PangoContext with an identical transformation matrix to that passed in to this function.

Usage

gdkDrawGlyphsTransformed(drawable, gc, matrix, font, x, y, glyphs)

Arguments

drawable
a GdkDrawable

gc
a GdkGC

matrix
a PangoMatrix, or NULL to use an identity transformation. [ allow-none ]

font
the font in which to draw the string

x
the x position of the start of the string (in Pango units in user space coordinates)

y
the y position of the baseline (in Pango units in user space coordinates)

glyphs
the glyph string to draw

Details

See also gdkDrawGlyphs, gdkDrawLayout. Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]