gdkDrawImage {RGtk2}	R Documentation

gdkDrawImage

Description

Draws a GdkImage onto a drawable. The depth of the GdkImage must match the depth of the GdkDrawable.

Usage

gdkDrawImage(object, gc, image, xsrc, ysrc, xdest, ydest, width, 
    height)

Arguments

object
a GdkDrawable (a GdkWindow or a GdkPixmap).

gc
a GdkGC.

image
the GdkImage to draw.

xsrc
the left edge of the source rectangle within image.

ysrc
the top of the source rectangle within image.

xdest
the x coordinate of the destination within drawable.

ydest
the y coordinate of the destination within drawable.

width
the width of the area to be copied, or -1 to make the area extend to the right edge of image.

height
the height of the area to be copied, or -1 to make the area extend to the bottom edge of image.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]