gdkDrawIndexedImage {RGtk2}	R Documentation

gdkDrawIndexedImage

Description

Draws an indexed image in the drawable, using a GdkRgbCmap to assign actual colors to the color indices.

Usage

gdkDrawIndexedImage(object, gc, x, y, width, height, dith, buf, 
    cmap)

Arguments

object
The GdkDrawable to draw in (usually a GdkWindow).

gc
The graphics context.

x
The x coordinate of the top-left corner in the drawable.

y
The y coordinate of the top-left corner in the drawable.

width
The width of the rectangle to be drawn.

height
The height of the rectangle to be drawn.

dith
A GdkRgbDither value, selecting the desired dither mode.

buf
The pixel data, represented as 8-bit color indices.

cmap
The GdkRgbCmap used to assign colors to the color indices.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]