gdkDrawLayoutLine {RGtk2}	R Documentation

gdkDrawLayoutLine

Description

Render a PangoLayoutLine onto an GDK drawable

Usage

gdkDrawLayoutLine(object, gc, x, y, line)

Arguments

object
the drawable on which to draw the line

gc
base graphics to use

x
the x position of start of string (in pixels)

y
the y position of baseline (in pixels)

line
a PangoLayoutLine

Details

If the layout's PangoContext has a transformation matrix set, then x and y specify the position of the left edge of the baseline (left is in before-tranform user coordinates) in after-transform device coordinates.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]