gdkDrawLayoutLineWithColors {RGtk2}	R Documentation

gdkDrawLayoutLineWithColors

Description

Render a PangoLayoutLine onto a GdkDrawable, overriding the layout's normal colors with foreground and/or background. foreground and background need not be allocated.

Usage

gdkDrawLayoutLineWithColors(drawable, gc, x, y, line, foreground, 
    background)

Arguments

drawable
the drawable on which to draw the line

gc
base graphics to use

x
the x position of start of string (in pixels)

y
the y position of baseline (in pixels)

line
a PangoLayoutLine

foreground
foreground override color, or NULL for none. [ allow-none ]

background
background override color, or NULL for none. [ allow-none ]

Details

If the layout's PangoContext has a transformation matrix set, then x and y specify the position of the left edge of the baseline (left is in before-tranform user coordinates) in after-transform device coordinates.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]