gdkDrawLayoutWithColors {RGtk2}	R Documentation

gdkDrawLayoutWithColors

Description

Render a PangoLayout onto a GdkDrawable, overriding the layout's normal colors with foreground and/or background. foreground and background need not be allocated.

Usage

gdkDrawLayoutWithColors(drawable, gc, x, y, layout, foreground, 
    background)

Arguments

drawable
the drawable on which to draw string

gc
base graphics context to use

x
the X position of the left of the layout (in pixels)

y
the Y position of the top of the layout (in pixels)

layout
a PangoLayout

foreground
foreground override color, or NULL for none. [ allow-none ]

background
background override color, or NULL for none. [ allow-none ]

Details

If the layout's PangoContext has a transformation matrix set, then x and y specify the position of the top left corner of the bounding box (in device space) of the transformed layout.

If you're using GTK+, the ususal way to obtain a PangoLayout is gtkWidgetCreatePangoLayout.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]