gdkDrawLines {RGtk2}	R Documentation

gdkDrawLines

Description

Draws a series of lines connecting the given points. The way in which joins between lines are draw is determined by the GdkCapStyle value in the GdkGC. This can be set with gdkGCSetLineAttributes.

Usage

gdkDrawLines(object, gc, points)

Arguments

object
a GdkDrawable (a GdkWindow or a GdkPixmap).

gc
a GdkGC.

points
a list of GdkPoint structures specifying the endpoints of the

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]