gdkDrawPixbuf {RGtk2}	R Documentation

gdkDrawPixbuf

Description

Renders a rectangular portion of a pixbuf to a drawable. The destination drawable must have a colormap. All windows have a colormap, however, pixmaps only have colormap by default if they were created with a non-NULL window argument. Otherwise a colormap must be set on them with gdkDrawableSetColormap.

Usage

gdkDrawPixbuf(object, gc = NULL, pixbuf, src.x, src.y, dest.x, 
    dest.y, width = -1, height = -1, dither = "GDK_RGB_DITHER_NORMAL", 
    x.dither = 0, y.dither = 0)

Arguments

object
Destination drawable.

gc
a GdkGC, used for clipping, or NULL. [ allow-none ]

pixbuf
a GdkPixbuf

src.x
Source X coordinate within pixbuf.

src.y
Source Y coordinates within pixbuf.

dest.x
Destination X coordinate within drawable.

dest.y
Destination Y coordinate within drawable.

width
Width of region to render, in pixels, or -1 to use pixbuf width.

height
Height of region to render, in pixels, or -1 to use pixbuf height.

dither
Dithering mode for GdkRGB.

x.dither
X offset for dither.

y.dither
Y offset for dither.

Details

On older X servers, rendering pixbufs with an alpha channel involves round trips to the X server, and may be somewhat slow.

If GDK is built with the Sun mediaLib library, the gdk_draw_pixbuf function is accelerated using mediaLib, which provides hardware acceleration on Intel, AMD, and Sparc chipsets. If desired, mediaLib support can be turned off by setting the GDK_DISABLE_MEDIALIB environment variable. Since 2.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]