gdkDrawRectangle {RGtk2}	R Documentation

gdkDrawRectangle

Description

Draws a rectangular outline or filled rectangle, using the foreground color and other attributes of the GdkGC.

Usage

gdkDrawRectangle(object, gc, filled, x, y, width, height)

Arguments

object
a GdkDrawable (a GdkWindow or a GdkPixmap).

gc
a GdkGC.

filled
TRUE if the rectangle should be filled.

x
the x coordinate of the left edge of the rectangle.

y
the y coordinate of the top edge of the rectangle.

width
the width of the rectangle.

height
the height of the rectangle.

Details

A rectangle drawn filled is 1 pixel smaller in both dimensions than a rectangle outlined. Calling gdk_draw_rectangle (window, gc, TRUE, 0, 0, 20, 20) results in a filled rectangle 20 pixels wide and 20 pixels high. Calling gdk_draw_rectangle (window, gc, FALSE, 0, 0, 20, 20) results in an outlined rectangle with corners at (0, 0), (0, 20), (20, 20), and (20, 0), which makes it 21 pixels wide and 21 pixels high.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]