gdkDrawRgb32Image {RGtk2}	R Documentation

gdkDrawRgb32Image

Description

Draws a padded RGB image in the drawable. The image is stored as one pixel per 32-bit word. It is laid out as a red byte, a green byte, a blue byte, and a padding byte.

Usage

gdkDrawRgb32Image(object, gc, x, y, width, height, dith, buf)

Arguments

object
The GdkDrawable to draw in (usually a GdkWindow).

gc
The graphics context.

x
The x coordinate of the top-left corner in the drawable.

y
The y coordinate of the top-left corner in the drawable.

width
The width of the rectangle to be drawn.

height
The height of the rectangle to be drawn.

dith
A GdkRgbDither value, selecting the desired dither mode.

buf
The pixel data, represented as padded 32-bit data.

Details

It's unlikely that this function will give significant performance gains in practice. In my experience, the performance gain from having pixels aligned to 32-bit boundaries is cancelled out by the increased memory bandwidth.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]