gdkDrawRgbImage {RGtk2}	R Documentation

gdkDrawRgbImage

Description

Draws an RGB image in the drawable. This is the core GdkRGB function, and likely the only one you will need to use.

Usage

gdkDrawRgbImage(object, gc, x, y, width, height, dith, rgb.buf, 
    rowstride)

Arguments

object
The GdkDrawable to draw in (usually a GdkWindow).

gc
The graphics context (all GDK drawing operations require one; its contents are ignored).

x
The x coordinate of the top-left corner in the drawable.

y
The y coordinate of the top-left corner in the drawable.

width
The width of the rectangle to be drawn.

height
The height of the rectangle to be drawn.

dith
A GdkRgbDither value, selecting the desired dither mode.

rgb.buf
The pixel data, represented as packed 24-bit data.

rowstride
The number of bytes from the start of one row in rgb.buf to the start of the next.

Details

The rowstride parameter allows for lines to be aligned more flexibly. For example, lines may be allocated to begin on 32-bit boundaries, even if the width of the rectangle is odd. Rowstride is also useful when drawing a subrectangle of a larger image in memory. Finally, to replicate the same line a number of times, the trick of setting rowstride to 0 is allowed.

In general, for 0 <= i < width and 0 <= j < height, the pixel (x + i, y + j) is colored with red value rgb.buf[j * rowstride + i * 3], green value rgb.buf[j * rowstride + i * 3 + 1], and blue value rgb.buf[j * rowstride + i * 3 + 2].

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]