gdkDrawRgbImageDithalign {RGtk2}	R Documentation

gdkDrawRgbImageDithalign

Description

Draws an RGB image in the drawable, with an adjustment for dither alignment.

Usage

gdkDrawRgbImageDithalign(object, gc, x, y, width, height, dith, 
    rgb.buf, xdith, ydith)

Arguments

object
The GdkDrawable to draw in (usually a GdkWindow).

gc
The graphics context.

x
The x coordinate of the top-left corner in the drawable.

y
The y coordinate of the top-left corner in the drawable.

width
The width of the rectangle to be drawn.

height
The height of the rectangle to be drawn.

dith
A GdkRgbDither value, selecting the desired dither mode.

rgb.buf
The pixel data, represented as packed 24-bit data.

xdith
An x offset for dither alignment.

ydith
A y offset for dither alignment.

Details

This function is useful when drawing dithered images into a window that may be scrolled. Pixel (x, y) will be drawn dithered as if its actual location is (x + xdith, y + ydith). Thus, if you draw an image into a window using zero dither alignment, then scroll up one pixel, subsequent draws to the window should have ydith = 1.

Setting the dither alignment correctly allows updating of small parts of the screen while avoiding visible "seams" between the different dither textures.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]