gdkDrawText {RGtk2}	R Documentation

gdkDrawText

Description

Draws a number of characters in the given font or fontset. WARNING: gdk_draw_text has been deprecated since version 2.4 and should not be used in newly-written code. Use gdkDrawLayout instead.

Usage

gdkDrawText(object, font, gc, x, y, text, text.length)

Arguments

object
a GdkDrawable (a GdkWindow or a GdkPixmap).

font
a GdkFont.

gc
a GdkGC.

x
the x coordinate of the left edge of the text.

y
the y coordinate of the baseline of the text.

text
the characters to draw.

text.length
the number of characters of text to draw.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]