gdkDrawTrapezoids {RGtk2}	R Documentation

gdkDrawTrapezoids

Description

Draws a set of anti-aliased trapezoids. The trapezoids are combined using saturation addition, then drawn over the background as a set. This is low level functionality used internally to implement rotated underlines and backgrouds when rendering a PangoLayout and is likely not useful for applications.

Usage

gdkDrawTrapezoids(drawable, gc, trapezoids)

Arguments

drawable
a GdkDrawable

gc
a GdkGC

trapezoids
a list of GdkTrapezoid structures

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]