gdkDrawableCopyToImage {RGtk2}	R Documentation

gdkDrawableCopyToImage

Description

Copies a portion of drawable into the client side image structure image. If image is NULL, creates a new image of size width x height and copies into that. See gdkDrawableGetImage for further details.

Usage

gdkDrawableCopyToImage(object, image = NULL, src.x, src.y, dest.x, 
    dest.y, width, height)

Arguments

object
a GdkDrawable

image
a GdkDrawable, or NULL if a new image should be created. [ allow-none ]

src.x
x coordinate on drawable

src.y
y coordinate on drawable

dest.x
x coordinate within image. Must be 0 if image is NULL

dest.y
y coordinate within image. Must be 0 if image is NULL

width
width of region to get

height
height or region to get

Details

Since 2.4

Value

[GdkImage] image, or a new a GdkImage containing the contents of drawable

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]