gdkDrawableGetSize {RGtk2}	R Documentation

gdkDrawableGetSize

Description

Fills *width and *height with the size of drawable. width or height can be NULL if you only want the other one.

Usage

gdkDrawableGetSize(object)

Arguments

object
a GdkDrawable

Details

On the X11 platform, if drawable is a GdkWindow, the returned size is the size reported in the most-recently-processed configure event, rather than the current size on the X server.

Value

A list containing the following elements:

width
(out): location to store drawable's width, or NULL. [ allow-none ]

height
(out): location to store drawable's height, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]