gdkDrawableSetColormap {RGtk2}	R Documentation

gdkDrawableSetColormap

Description

Sets the colormap associated with drawable. Normally this will happen automatically when the drawable is created; you only need to use this function if the drawable-creating function did not have a way to determine the colormap, and you then use drawable operations that require a colormap. The colormap for all drawables and graphics contexts you intend to use together should match. i.e. when using a GdkGC to draw to a drawable, or copying one drawable to another, the colormaps should match.

Usage

gdkDrawableSetColormap(object, colormap)

Arguments

object
a GdkDrawable

colormap
a GdkColormap

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]