gdkEventGetScreen {RGtk2}	R Documentation

gdkEventGetScreen

Description

Returns the screen for the event. The screen is typically the screen for event->any.window, but for events such as mouse events, it is the screen where the pointer was when the event occurs - that is, the screen which has the root window to which event->motion.x_root and event->motion.y_root are relative.

Usage

gdkEventGetScreen(object)

Arguments

object
a GdkEvent

Details

Since 2.2

Value

[GdkScreen] the screen for the event

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]