gdkEventHandlerSet {RGtk2}	R Documentation

gdkEventHandlerSet

Description

Sets the function to call to handle all events from GDK.

Usage

gdkEventHandlerSet(func, data)

Arguments

func
the function to call to handle events from GDK.

data
user data to pass to the function.

Details

Note that GTK+ uses this to install its own event handler, so it is usually not useful for GTK+ applications. (Although an application can call this function then call gtkMainDoEvent to pass events to GTK+.)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]