gdkEventRequestMotions {RGtk2}	R Documentation

gdkEventRequestMotions

Description

Request more motion notifies if event is a motion notify hint event. This function should be used instead of gdkWindowGetPointer to request further motion notifies, because it also works for extension events where motion notifies are provided for devices other than the core pointer. Coordinate extraction, processing and requesting more motion events from a GDK_MOTION_NOTIFY event usually works like this:

Usage

gdkEventRequestMotions(event)

Arguments

event
a valid GdkEvent

Details

# motion event handler
{
  x <- motion_event$x
  y <- motion_event$y
  # handle (x,y) motion here
  motion_event$request_motions() # handles is_hint events
}

Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]