gdkFlush {RGtk2}	R Documentation

gdkFlush

Description

Flushes the X output buffer and waits until all requests have been processed by the server. This is rarely needed by applications. It's main use is for trapping X errors with gdkErrorTrapPush() and gdkErrorTrapPop().

Usage

gdkFlush()

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]