gdkFontFromDescription {RGtk2}	R Documentation

gdkFontFromDescription

Description

Load a GdkFont based on a Pango font description. This font will only be an approximation of the Pango font, and internationalization will not be handled correctly. This function should only be used for legacy code that cannot be easily converted to use Pango. Using Pango directly will produce better results. WARNING: gdk_font_from_description is deprecated and should not be used in newly-written code.

Usage

gdkFontFromDescription(font.desc)

Arguments

font.desc
a PangoFontDescription.

Value

[GdkFont] the newly loaded font, or NULL if the font cannot be loaded.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]