gdkFontFromDescriptionForDisplay {RGtk2}	R Documentation

gdkFontFromDescriptionForDisplay

Description

Loads a GdkFont based on a Pango font description for use on display. This font will only be an approximation of the Pango font, and internationalization will not be handled correctly. This function should only be used for legacy code that cannot be easily converted to use Pango. Using Pango directly will produce better results. WARNING: gdk_font_from_description_for_display is deprecated and should not be used in newly-written code.

Usage

gdkFontFromDescriptionForDisplay(display, font.desc)

Arguments

display
a GdkDisplay

font.desc
a PangoFontDescription.

Details

Since 2.2

Value

[GdkFont] the newly loaded font, or NULL if the font cannot be loaded.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]