gdkFontLoadForDisplay {RGtk2}	R Documentation

gdkFontLoadForDisplay

Description

Loads a font for use on display. WARNING: gdk_font_load_for_display is deprecated and should not be used in newly-written code.

Usage

gdkFontLoadForDisplay(display, font.name)

Arguments

display
a GdkDisplay

font.name
a XLFD describing the font to load.

Details

The font may be newly loaded or looked up the font in a cache. You should make no assumptions about the initial reference count. Since 2.2

Value

[GdkFont] a GdkFont, or NULL if the font could not be loaded.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]