gdkFontsetLoad {RGtk2}	R Documentation

gdkFontsetLoad

Description

Loads a fontset. WARNING: gdk_fontset_load is deprecated and should not be used in newly-written code.

Usage

gdkFontsetLoad(fontset.name)

Arguments

fontset.name
a comma-separated list of XLFDs describing the component fonts of the fontset to load.

Details

The fontset may be newly loaded or looked up in a cache. You should make no assumptions about the initial reference count.

Value

[GdkFont] a GdkFont, or NULL if the fontset could not be loaded.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]