gdkFontsetLoadForDisplay {RGtk2}	R Documentation

gdkFontsetLoadForDisplay

Description

Loads a fontset for use on display. WARNING: gdk_fontset_load_for_display is deprecated and should not be used in newly-written code.

Usage

gdkFontsetLoadForDisplay(display, fontset.name)

Arguments

display
a GdkDisplay

fontset.name
a comma-separated list of XLFDs describing the component fonts of the fontset to load.

Details

The fontset may be newly loaded or looked up in a cache. You should make no assumptions about the initial reference count. Since 2.2

Value

[GdkFont] a GdkFont, or NULL if the fontset could not be loaded.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]