gdkGCGetValues {RGtk2}	R Documentation

gdkGCGetValues

Description

Retrieves the current values from a graphics context. Note that only the pixel values of the values->foreground and values->background are filled, use gdkColormapQueryColor to obtain the rgb values if you need them.

Usage

gdkGCGetValues(object)

Arguments

object
a GdkGC.

Value

A list containing the following elements:

values
the GdkGCValues structure in which to store the results.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]