gdkGCSetClipRectangle {RGtk2}	R Documentation

gdkGCSetClipRectangle

Description

Sets the clip mask for a graphics context from a rectangle. The clip mask is interpreted relative to the clip origin. (See gdkGCSetClipOrigin).

Usage

gdkGCSetClipRectangle(object, rectangle)

Arguments

object
a GdkGC.

rectangle
the rectangle to clip to.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]