gdkGCSetClipRegion {RGtk2}	R Documentation

gdkGCSetClipRegion

Description

Sets the clip mask for a graphics context from a region structure. The clip mask is interpreted relative to the clip origin. (See gdkGCSetClipOrigin).

Usage

gdkGCSetClipRegion(object, region)

Arguments

object
a GdkGC.

region
the GdkRegion.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]