gdkGCSetDashes {RGtk2}	R Documentation

gdkGCSetDashes

Description

Sets the way dashed-lines are drawn. Lines will be drawn with alternating on and off segments of the lengths specified in dash.list. The manner in which the on and off segments are drawn is determined by the line.style value of the GC. (This can be changed with gdkGCSetLineAttributes.)

Usage

gdkGCSetDashes(object, dash.list)

Arguments

object
a GdkGC.

dash.list
a list of dash lengths.

Details

The dash.offset defines the phase of the pattern, specifying how many pixels into the dash-list the pattern should actually begin.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]