gdkGCSetTsOrigin {RGtk2}	R Documentation

gdkGCSetTsOrigin

Description

Set the origin when using tiles or stipples with the GC. The tile or stipple will be aligned such that the upper left corner of the tile or stipple will coincide with this point.

Usage

gdkGCSetTsOrigin(object, x, y)

Arguments

object
a GdkGC.

x
the x-coordinate of the origin.

y
the y-coordinate of the origin.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]