gdkGCSetValues {RGtk2}	R Documentation

gdkGCSetValues

Description

Sets attributes of a graphics context in bulk. For each flag set in values.mask, the corresponding field will be read from values and set as the new value for gc. If you're only setting a few values on gc, calling individual "setter" functions is likely more convenient.

Usage

gdkGCSetValues(object, values)

Arguments

object
a GdkGC

values
struct containing the new values

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]