gdkGetProgramClass {RGtk2}	R Documentation

gdkGetProgramClass

Description

Gets the program class. Unless the program class has explicitly been set with gdkSetProgramClass or with the --class commandline option, the default value is the program name (determined with gGetPrgname()) with the first character converted to uppercase.

Usage

gdkGetProgramClass()

Value

[char] the program class.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]