gdkImageNew {RGtk2}	R Documentation

gdkImageNew

Description

Creates a new GdkImage.

Usage

gdkImageNew(type, visual, width, height)

Arguments

type
the type of the GdkImage, one of GDK_IMAGE_NORMAL, GDK_IMAGE_SHARED and GDK_IMAGE_FASTEST. GDK_IMAGE_FASTEST is probably the best choice, since it will try creating a GDK_IMAGE_SHARED image first and if that fails it will then use GDK_IMAGE_NORMAL.

visual
the GdkVisual to use for the image.

width
the width of the image in pixels.

height
the height of the image in pixels.

Value

[GdkImage] a new GdkImage, or NULL if the image could not be created.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]