gdkKeyboardGrab {RGtk2}	R Documentation

gdkKeyboardGrab

Description

Grabs the keyboard so that all events are passed to this application until the keyboard is ungrabbed with gdkKeyboardUngrab. This overrides any previous keyboard grab by this client.

Usage

gdkKeyboardGrab(window, owner.events = FALSE, time = "GDK_CURRENT_TIME")

Arguments

window
the GdkWindow which will own the grab (the grab window).

owner.events
if FALSE then all keyboard events are reported with respect to window. If TRUE then keyboard events for this application are reported as normal, but keyboard events outside this application are reported with respect to window. Both key press and key release events are always reported, independant of the event mask set by the application.

time
a timestamp from a GdkEvent, or GDK_CURRENT_TIME if no timestamp is available.

Details

If you set up anything at the time you take the grab that needs to be cleaned up when the grab ends, you should handle the GdkEventGrabBroken events that are emitted when the grab ends unvoluntarily.

Value

[GdkGrabStatus] GDK_GRAB_SUCCESS if the grab was successful.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]