gdkKeymapAddVirtualModifiers {RGtk2}	R Documentation

gdkKeymapAddVirtualModifiers

Description

Adds virtual modifiers (i.e. Super, Hyper and Meta) which correspond to the real modifiers (i.e Mod2, Mod3, ...) in modifiers. are set in state to their non-virtual counterparts (i.e. Mod2, Mod3,...) and set the corresponding bits in state.

Usage

gdkKeymapAddVirtualModifiers(object)

Arguments

object
a GdkKeymap

Details

GDK already does this before delivering key events, but for compatibility reasons, it only sets the first virtual modifier it finds, whereas this function sets all matching virtual modifiers.

This function is useful when matching key events against accelerators. Since 2.20

Value

A list containing the following elements:

state
pointer to the modifier mask to change

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]