gdkKeymapGetEntriesForKeycode {RGtk2}	R Documentation

gdkKeymapGetEntriesForKeycode

Description

Returns the keyvals bound to hardware.keycode. The Nth GdkKeymapKey in keys is bound to the Nth keyval in keyvals. When a keycode is pressed by the user, the keyval from this list of entries is selected by considering the effective keyboard group and level. See gdkKeymapTranslateKeyboardState.

Usage

gdkKeymapGetEntriesForKeycode(object, hardware.keycode)

Arguments

object
a GdkKeymap or NULL to use the default keymap

hardware.keycode
a keycode

Value

A list containing the following elements:

retval
[logical] TRUE if there were any entries

keys
return location for list of GdkKeymapKey, or NULL

keyvals
return location for list of keyvals, or NULL

n.entries
length of keys and keyvals

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]