gdkKeymapGetEntriesForKeyval {RGtk2}	R Documentation

gdkKeymapGetEntriesForKeyval

Description

Obtains a list of keycode/group/level combinations that will generate keyval. Groups and levels are two kinds of keyboard mode; in general, the level determines whether the top or bottom symbol on a key is used, and the group determines whether the left or right symbol is used. On US keyboards, the shift key changes the keyboard level, and there are no groups. A group switch key might convert a keyboard between Hebrew to English modes, for example. GdkEventKey contains a group field that indicates the active keyboard group. The level is computed from the modifier mask.

Usage

gdkKeymapGetEntriesForKeyval(object, keyval)

Arguments

object
a GdkKeymap, or NULL to use the default keymap

keyval
a keyval, such as GDK_a, GDK_Up, GDK_Return, etc.

Value

A list containing the following elements:

retval
[logical] TRUE if keys were found and returned

keys
return location for a list of GdkKeymapKey

n.keys
return location for number of elements in returned list

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]