gdkKeymapLookupKey {RGtk2}	R Documentation

gdkKeymapLookupKey

Description

Looks up the keyval mapped to a keycode/group/level triplet. If no keyval is bound to key, returns 0. For normal user input, you want to use gdkKeymapTranslateKeyboardState instead of this function, since the effective group/level may not be the same as the current keyboard state.

Usage

gdkKeymapLookupKey(object, key)

Arguments

object
a GdkKeymap or NULL to use the default keymap

key
a GdkKeymapKey with keycode, group, and level initialized

Value

[numeric] a keyval, or 0 if none was mapped to the given key

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]