gdkKeymapMapVirtualModifiers {RGtk2}	R Documentation

gdkKeymapMapVirtualModifiers

Description

Maps the virtual modifiers (i.e. Super, Hyper and Meta) which are set in state to their non-virtual counterparts (i.e. Mod2, Mod3,...) and set the corresponding bits in state.

Usage

gdkKeymapMapVirtualModifiers(object)

Arguments

object
a GdkKeymap

Details

This function is useful when matching key events against accelerators. Since 2.20

Value

A list containing the following elements:

retval
[logical] TRUE if no virtual modifiers were mapped to the same non-virtual modifier. Note that FALSE is also returned if a virtual modifier is mapped to a non-virtual modifier that was already set in state.

state
pointer to the modifier state to map

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]