gdkKeyvalName {RGtk2}	R Documentation

gdkKeyvalName

Description

Converts a key value into a symbolic name. The names are the same as those in the ‘<gdk/gdkkeysyms.h>’ header file but without the leading "GDK_".

Usage

gdkKeyvalName(keyval)

Arguments

keyval
a key value.

Value

[character] a string containing the name of the key, or NULL if keyval is not a valid key. The string should not be modified.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]