gdkKeyvalToUpper {RGtk2}	R Documentation

gdkKeyvalToUpper

Description

Converts a key value to upper case, if applicable.

Usage

gdkKeyvalToUpper(keyval)

Arguments

keyval
a key value.

Value

[numeric] the upper case form of keyval, or keyval itself if it is already in upper case or it is not subject to case conversion.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]