gdkListVisuals {RGtk2}	R Documentation

gdkListVisuals

Description

Lists the available visuals for the default screen. (See gdkScreenListVisuals) A visual describes a hardware image data format. For example, a visual might support 24-bit color, or 8-bit color, and might expect pixels to be in a certain format.

Usage

gdkListVisuals()

Value

[list] a list of visuals; the list must be freed, but not its contents

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]