gdkNotifyStartupComplete {RGtk2}	R Documentation

gdkNotifyStartupComplete

Description

Indicates to the GUI environment that the application has finished loading. If the applications opens windows, this function is normally called after opening the application's initial set of windows.

Usage

gdkNotifyStartupComplete()

Details

GTK+ will call this function automatically after opening the first GtkWindow unless gtkWindowSetAutoStartupNotification is called to disable that feature. Since 2.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]