gdkNotifyStartupCompleteWithId {RGtk2}	R Documentation

gdkNotifyStartupCompleteWithId

Description

Indicates to the GUI environment that the application has finished loading, using a given identifier.

Usage

gdkNotifyStartupCompleteWithId(id)

Arguments

id
a startup-notification identifier, for which notification process should be completed

Details

GTK+ will call this function automatically for GtkWindow with custom startup-notification identifier unless gtkWindowSetAutoStartupNotification is called to disable that feature. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]