gdkOffscreenWindowSetEmbedder {RGtk2}	R Documentation

gdkOffscreenWindowSetEmbedder

Description

Sets window to be embedded in embedder.

Usage

gdkOffscreenWindowSetEmbedder(window, embedder)

Arguments

window
a GdkWindow

embedder
the GdkWindow that window gets embedded in

Details

To fully embed an offscreen window, in addition to calling this function, it is also necessary to handle the "pick-embedded-child" signal on the embedder and the "to-embedder" and "from-embedder" signals on window. Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]