gdkPangoContextGetForScreen {RGtk2}	R Documentation

gdkPangoContextGetForScreen

Description

Creates a PangoContext for screen.

Usage

gdkPangoContextGetForScreen(screen)

Arguments

screen
the GdkScreen for which the context is to be created.

Details

When using GTK+, normally you should use gtkWidgetGetPangoContext instead of this function, to get the appropriate context for the widget you intend to render text onto.

The newly created context will have the default font options (see CairoFontOptions) for the screen; if these options change it will not be updated. Using gtkWidgetGetPangoContext is more convenient if you want to keep a context around and track changes to the screen's font rendering settings. Since 2.2

Value

[PangoContext] a new PangoContext for screen

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]