gdkPangoContextSetColormap {RGtk2}	R Documentation

gdkPangoContextSetColormap

Description

This function used to set the colormap to be used for drawing with context. The colormap is now always derived from the graphics context used for drawing, so calling this function is no longer necessary. WARNING: gdk_pango_context_set_colormap is deprecated and should not be used in newly-written code.

Usage

gdkPangoContextSetColormap(context, colormap)

Arguments

context
a PangoContext

colormap
a GdkColormap

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]