gdkPangoLayoutGetClipRegion {RGtk2}	R Documentation

gdkPangoLayoutGetClipRegion

Description

Obtains a clip region which contains the areas where the given ranges of text would be drawn. x.origin and y.origin are the same position you would pass to gdkDrawLayoutLine. index.ranges should contain ranges of bytes in the layout's text.

Usage

gdkPangoLayoutGetClipRegion(layout, x.origin, index.ranges)

Arguments

layout
a PangoLayout

x.origin
X pixel where you intend to draw the layout with this clip

index.ranges
array of byte indexes into the layout, where even members of list are start indexes and odd elements are end indexes

Details

Note that the regions returned correspond to logical extents of the text ranges, not ink extents. So the drawn layout may in fact touch areas out of the clip region. The clip region is mainly useful for highlightling parts of text, such as when text is selected.

Value

[GdkRegion] a clip region containing the given ranges

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]