gdkPangoRendererGetDefault {RGtk2}	R Documentation

gdkPangoRendererGetDefault

Description

Gets the default PangoRenderer for a screen. This default renderer is shared by all users of the display, so properties such as the color or transformation matrix set for the renderer may be overwritten by functions such as gdkDrawLayout.

Usage

gdkPangoRendererGetDefault(screen)

Arguments

screen
a GdkScreen

Details

Before using the renderer, you need to call gdkPangoRendererSetDrawable and gdkPangoRendererSetGc to set the drawable and graphics context to use for drawing. Since 2.6

Value

[PangoRenderer] the default PangoRenderer for screen.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]