gdkPangoRendererSetGc {RGtk2}	R Documentation

gdkPangoRendererSetGc

Description

Sets the GC the renderer draws with. Note that the GC must not be modified until it is unset by calling the function again with NULL for the gc parameter, since GDK may make internal copies of the GC which won't be updated to follow changes to the original GC.

Usage

gdkPangoRendererSetGc(object, gc = NULL)

Arguments

object
a GdkPangoRenderer

gc
the new GC to use for drawing, or NULL. [ allow-none ]

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]