gdkPangoRendererSetOverrideColor {RGtk2}	R Documentation

gdkPangoRendererSetOverrideColor

Description

Sets the color for a particular render part (foreground, background, underline, etc.), overriding any attributes on the layouts renderered with this renderer.

Usage

gdkPangoRendererSetOverrideColor(object, part, color = NULL)

Arguments

object
a GdkPangoRenderer

part
the part to render to set the color of

color
the color to use, or NULL to unset a previously set override color. [ allow-none ]

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]