gdkPixbufAddAlpha {RGtk2}	R Documentation

gdkPixbufAddAlpha

Description

Takes an existing pixbuf and adds an alpha channel to it. If the existing pixbuf already had an alpha channel, the channel values are copied from the original; otherwise, the alpha channel is initialized to 255 (full opacity).

Usage

gdkPixbufAddAlpha(object, substitute.color, r, g, b)

Arguments

object
A GdkPixbuf.

substitute.color
Whether to set a color to zero opacity. If this is FALSE, then the (r, g, b) arguments will be ignored.

r
Red value to substitute.

g
Green value to substitute.

b
Blue value to substitute.

Details

If substitute.color is TRUE, then the color specified by (r, g, b) will be assigned zero opacity. That is, if you pass (255, 255, 255) for the substitute color, all white pixels will become fully transparent.

Value

[GdkPixbuf] A newly-created pixbuf with a reference count of 1.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]