gdkPixbufAnimationGetStaticImage {RGtk2}	R Documentation

gdkPixbufAnimationGetStaticImage

Description

If an animation is really just a plain image (has only one frame), this function returns that image. If the animation is an animation, this function returns a reasonable thing to display as a static unanimated image, which might be the first frame, or something more sophisticated. If an animation hasn't loaded any frames yet, this function will return NULL.

Usage

gdkPixbufAnimationGetStaticImage(object)

Arguments

object
a GdkPixbufAnimation

Value

[GdkPixbuf] unanimated image representing the animation

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]