gdkPixbufAnimationIsStaticImage {RGtk2}	R Documentation

gdkPixbufAnimationIsStaticImage

Description

If you load a file with gdkPixbufAnimationNewFromFile and it turns out to be a plain, unanimated image, then this function will return TRUE. Use gdkPixbufAnimationGetStaticImage to retrieve the image.

Usage

gdkPixbufAnimationIsStaticImage(object)

Arguments

object
a GdkPixbufAnimation

Value

[logical] TRUE if the "animation" was really just an image

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]