gdkPixbufAnimationIterAdvance {RGtk2}	R Documentation

gdkPixbufAnimationIterAdvance

Description

Possibly advances an animation to a new frame. Chooses the frame based on the start time passed to gdkPixbufAnimationGetIter.

Usage

gdkPixbufAnimationIterAdvance(object, current.time)

Arguments

object
a GdkPixbufAnimationIter

current.time
current time

Details

current.time would normally come from gGetCurrentTime(), and must be greater than or equal to the time passed to gdkPixbufAnimationGetIter, and must increase or remain unchanged each time gdkPixbufAnimationIterGetPixbuf is called. That is, you can't go backward in time; animations only play forward.

As a shortcut, pass NULL for the current time and gGetCurrentTime() will be invoked on your behalf. So you only need to explicitly pass current.time if you're doing something odd like playing the animation at double speed.

If this function returns FALSE, there's no need to update the animation display, assuming the display had been rendered prior to advancing; if TRUE, you need to call gdkAnimationIterGetPixbuf() and update the display with the new pixbuf.

Value

[logical] TRUE if the image may need updating

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]