gdkPixbufAnimationIterGetDelayTime {RGtk2}	R Documentation

gdkPixbufAnimationIterGetDelayTime

Description

Gets the number of milliseconds the current pixbuf should be displayed, or -1 if the current pixbuf should be displayed forever. gTimeoutAdd conveniently takes a timeout in milliseconds, so you can use a timeout to schedule the next update.

Usage

gdkPixbufAnimationIterGetDelayTime(object)

Arguments

object
an animation iterator

Value

[integer] delay time in milliseconds (thousandths of a second)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]