gdkPixbufAnimationIterGetPixbuf {RGtk2}	R Documentation

gdkPixbufAnimationIterGetPixbuf

Description

Gets the current pixbuf which should be displayed; the pixbuf will be the same size as the animation itself (gdkPixbufAnimationGetWidth, gdkPixbufAnimationGetHeight). This pixbuf should be displayed for gdkPixbufAnimationIterGetDelayTime milliseconds. The caller of this function does not own a reference to the returned pixbuf; the returned pixbuf will become invalid when the iterator advances to the next frame, which may happen anytime you call gdkPixbufAnimationIterAdvance. Copy the pixbuf to keep it (don't just add a reference), as it may get recycled as you advance the iterator.

Usage

gdkPixbufAnimationIterGetPixbuf(object)

Arguments

object
an animation iterator

Value

[GdkPixbuf] the pixbuf to be displayed

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]