gdkPixbufAnimationNewFromFile {RGtk2}	R Documentation

gdkPixbufAnimationNewFromFile

Description

Creates a new animation by loading it from a file. The file format is detected automatically. If the file's format does not support multi-frame images, then an animation with a single frame will be created. Possible errors are in the GDK_PIXBUF_ERROR and G_FILE_ERROR domains.

Usage

gdkPixbufAnimationNewFromFile(filename, .errwarn = TRUE)

Arguments

filename
Name of file to load, in the GLib file name encoding

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[GdkPixbufAnimation] A newly-created animation with a reference count of 1, or NULL if any of several error conditions ocurred: the file could not be opened, there was no loader for the file's format, there was not enough memory to allocate the image buffer, or the image file contained invalid data.

error
return location for error

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]