gdkPixbufComposite {RGtk2}	R Documentation

gdkPixbufComposite

Description

Creates a transformation of the source image src by scaling by scale.x and scale.y then translating by offset.x and offset.y. This gives an image in the coordinates of the destination pixbuf. The rectangle (dest.x, dest.y, dest.width, dest.height) is then composited onto the corresponding rectangle of the original destination image.

Usage

gdkPixbufComposite(object, dest, dest.x, dest.y, dest.width, 
    dest.height, offset.x, offset.y, scale.x, scale.y, interp.type, 
    overall.alpha)

Arguments

object
a GdkPixbuf

dest
the GdkPixbuf into which to render the results

dest.x
the left coordinate for region to render

dest.y
the top coordinate for region to render

dest.width
the width of the region to render

dest.height
the height of the region to render

offset.x
the offset in the X direction (currently rounded to an integer)

offset.y
the offset in the Y direction (currently rounded to an integer)

scale.x
the scale factor in the X direction

scale.y
the scale factor in the Y direction

interp.type
the interpolation type for the transformation.

overall.alpha
overall alpha for source image (0..255)

Details

When the destination rectangle contains parts not in the source image, the data at the edges of the source image is replicated to infinity.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]