gdkPixbufCompositeColor {RGtk2}	R Documentation

gdkPixbufCompositeColor

Description

Creates a transformation of the source image src by scaling by scale.x and scale.y then translating by offset.x and offset.y, then composites the rectangle (dest.x ,dest.y, dest.width, dest.height) of the resulting image with a checkboard of the colors color1 and color2 and renders it onto the destination image.

Usage

gdkPixbufCompositeColor(object, dest, dest.x, dest.y, dest.width, 
    dest.height, offset.x, offset.y, scale.x, scale.y, interp.type, 
    overall.alpha, check.x, check.y, check.size, color1, color2)

Arguments

object
a GdkPixbuf

dest
the GdkPixbuf into which to render the results

dest.x
the left coordinate for region to render

dest.y
the top coordinate for region to render

dest.width
the width of the region to render

dest.height
the height of the region to render

offset.x
the offset in the X direction (currently rounded to an integer)

offset.y
the offset in the Y direction (currently rounded to an integer)

scale.x
the scale factor in the X direction

scale.y
the scale factor in the Y direction

interp.type
the interpolation type for the transformation.

overall.alpha
overall alpha for source image (0..255)

check.x
the X offset for the checkboard (origin of checkboard is at -check.x, -check.y)

check.y
the Y offset for the checkboard

check.size
the size of checks in the checkboard (must be a power of two)

color1
the color of check at upper left

color2
the color of the other check

Details

See gdkPixbufCompositeColorSimple for a simpler variant of this function suitable for many tasks.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]