gdkPixbufCompositeColorSimple {RGtk2}	R Documentation

gdkPixbufCompositeColorSimple

Description

Creates a new GdkPixbuf by scaling src to dest.width x dest.height and compositing the result with a checkboard of colors color1 and color2.

Usage

gdkPixbufCompositeColorSimple(object, dest.width, dest.height, 
    interp.type, overall.alpha, check.size, color1, color2)

Arguments

object
a GdkPixbuf

dest.width
the width of destination image

dest.height
the height of destination image

interp.type
the interpolation type for the transformation.

overall.alpha
overall alpha for source image (0..255)

check.size
the size of checks in the checkboard (must be a power of two)

color1
the color of check at upper left

color2
the color of the other check

Value

[GdkPixbuf] the new GdkPixbuf, or NULL if not enough memory could be allocated for it.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]