gdkPixbufCopyArea {RGtk2}	R Documentation

gdkPixbufCopyArea

Description

Copies a rectangular area from src.pixbuf to dest.pixbuf. Conversion of pixbuf formats is done automatically.

Usage

gdkPixbufCopyArea(object, src.x, src.y, width, height, dest.pixbuf, 
    dest.x, dest.y)

Arguments

object
Source pixbuf.

src.x
Source X coordinate within src.pixbuf.

src.y
Source Y coordinate within src.pixbuf.

width
Width of the area to copy.

height
Height of the area to copy.

dest.pixbuf
Destination pixbuf.

dest.x
X coordinate within dest.pixbuf.

dest.y
Y coordinate within dest.pixbuf.

Details

If the source rectangle overlaps the destination rectangle on the same pixbuf, it will be overwritten during the copy operation. Therefore, you can not use this function to scroll a pixbuf.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]