gdkPixbufFormatGetLicense {RGtk2}	R Documentation

gdkPixbufFormatGetLicense

Description

Returns information about the license of the image loader for the format. The returned string should be a shorthand for a wellknown license, e.g. "LGPL", "GPL", "QPL", "GPL/QPL", or "other" to indicate some other license.

Usage

gdkPixbufFormatGetLicense(object)

Arguments

object
a GdkPixbufFormat

Details

Since 2.6

Value

[character] a string describing the license of format.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]