gdkPixbufFormatIsScalable {RGtk2}	R Documentation

gdkPixbufFormatIsScalable

Description

Returns whether this image format is scalable. If a file is in a scalable format, it is preferable to load it at the desired size, rather than loading it at the default size and scaling the resulting pixbuf to the desired size.

Usage

gdkPixbufFormatIsScalable(object)

Arguments

object
a GdkPixbufFormat

Details

Since 2.6

Value

[logical] whether this image format is scalable.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]