gdkPixbufFormatSetDisabled {RGtk2}	R Documentation

gdkPixbufFormatSetDisabled

Description

Disables or enables an image format. If a format is disabled, gdk-pixbuf won't use the image loader for this format to load images. Applications can use this to avoid using image loaders with an inappropriate license, see gdkPixbufFormatGetLicense.

Usage

gdkPixbufFormatSetDisabled(object, disabled)

Arguments

object
a GdkPixbufFormat

disabled
TRUE to disable the format format

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]