gdkPixbufGetFileInfo {RGtk2}	R Documentation

gdkPixbufGetFileInfo

Description

Parses an image file far enough to determine its format and size.

Usage

gdkPixbufGetFileInfo(filename)

Arguments

filename
The name of the file to identify.

Details

Since 2.4

Value

A list containing the following elements:

retval
[GdkPixbufFormat] A GdkPixbufFormat describing the image format of the file or NULL if the image format wasn't recognized.

width
Return location for the width of the image, or NULL

height
Return location for the height of the image, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]