gdkPixbufGetFromDrawable {RGtk2}	R Documentation

gdkPixbufGetFromDrawable

Description

Transfers image data from a GdkDrawable and converts it to an RGB(A) representation inside a GdkPixbuf. In other words, copies image data from a server-side drawable to a client-side RGB(A) buffer. This allows you to efficiently read individual pixels on the client side.

Usage

gdkPixbufGetFromDrawable(dest = NULL, src, cmap = NULL, src.x, 
    src.y, dest.x, dest.y, width, height)

Arguments

dest
Destination pixbuf, or NULL if a new pixbuf should be created. [ allow-none ]

src
Source drawable.

cmap
A colormap if src doesn't have one set.

src.x
Source X coordinate within drawable.

src.y
Source Y coordinate within drawable.

dest.x
Destination X coordinate in pixbuf, or 0 if dest is NULL.

dest.y
Destination Y coordinate in pixbuf, or 0 if dest is NULL.

width
Width in pixels of region to get.

height
Height in pixels of region to get.

Details

If the drawable src has no colormap (gdkDrawableGetColormap returns NULL), then a suitable colormap must be specified. Typically a GdkWindow or a pixmap created by passing a GdkWindow to gdkPixmapNew will already have a colormap associated with it. If the drawable has a colormap, the cmap argument will be ignored. If the drawable is a bitmap (1 bit per pixel pixmap), then a colormap is not required; pixels with a value of 1 are assumed to be white, and pixels with a value of 0 are assumed to be black. For taking screenshots, gdkColormapGetSystem returns the correct colormap to use.

If the specified destination pixbuf dest is NULL, then this function will create an RGB pixbuf with 8 bits per channel and no alpha, with the same size specified by the width and height arguments. In this case, the dest.x and dest.y arguments must be specified as 0. If the specified destination pixbuf is not NULL and it contains alpha information, then the filled pixels will be set to full opacity (alpha = 255).

If the specified drawable is a pixmap, then the requested source rectangle must be completely contained within the pixmap, otherwise the function will return NULL. For pixmaps only (not for windows) passing -1 for width or height is allowed to mean the full width or height of the pixmap.

If the specified drawable is a window, and the window is off the screen, then there is no image data in the obscured/offscreen regions to be placed in the pixbuf. The contents of portions of the pixbuf corresponding to the offscreen region are undefined.

If the window you're obtaining data from is partially obscured by other windows, then the contents of the pixbuf areas corresponding to the obscured regions are undefined.

If the target drawable is not mapped (typically because it's iconified/minimized or not on the current workspace), then NULL will be returned.

If memory can't be allocated for the return value, NULL will be returned instead.

(In short, there are several ways this function can fail, and if it fails it returns NULL; so check the return value.)

This function calls gdkDrawableGetImage internally and converts the resulting image to a GdkPixbuf, so the documentation for gdkDrawableGetImage may also be relevant.

Value

[GdkPixbuf] The same pixbuf as dest if it was non-NULL, or a newly-created pixbuf with a reference count of 1 if no destination pixbuf was specified, or NULL on error

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]