gdkPixbufGetOption {RGtk2}	R Documentation

gdkPixbufGetOption

Description

Looks up key in the list of options that may have been attached to the pixbuf when it was loaded, or that may have been attached by another function using gdkPixbufSetOption.

Usage

gdkPixbufGetOption(object, key)

Arguments

object
a GdkPixbuf

key
a string.

Details

For instance, the ANI loader provides "Title" and "Artist" options. The ICO, XBM, and XPM loaders provide "x_hot" and "y_hot" hot-spot options for cursor definitions. The PNG loader provides the tEXt ancillary chunk key/value pairs as options. Since 2.12, the TIFF and JPEG loaders return an "orientation" option string that corresponds to the embedded TIFF/Exif orientation tag (if present).

Value

[character] the value associated with key.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]