gdkPixbufLoaderClose {RGtk2}	R Documentation

gdkPixbufLoaderClose

Description

Informs a pixbuf loader that no further writes with gdkPixbufLoaderWrite will occur, so that it can free its internal loading structures. Also, tries to parse any data that hasn't yet been parsed; if the remaining data is partial or corrupt, an error will be returned. If FALSE is returned, error will be set to an error from the GDK_PIXBUF_ERROR or G_FILE_ERROR domains. If you're just cancelling a load rather than expecting it to be finished, passing NULL for error to ignore it is reasonable.

Usage

gdkPixbufLoaderClose(object, .errwarn = TRUE)

Arguments

object
A pixbuf loader.

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[logical] TRUE if all image data written so far was successfully passed out via the update_area signal

error
return location for a GError, or NULL to ignore errors. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]