gdkPixbufLoaderGetPixbuf {RGtk2}	R Documentation

gdkPixbufLoaderGetPixbuf

Description

Queries the GdkPixbuf that a pixbuf loader is currently creating. In general it only makes sense to call this function after the "area-prepared" signal has been emitted by the loader; this means that enough data has been read to know the size of the image that will be allocated. If the loader has not received enough data via gdkPixbufLoaderWrite, then this function returns NULL. The returned pixbuf will be the same in all future calls to the loader, Additionally, if the loader is an animation, it will return the "static image" of the animation (see gdkPixbufAnimationGetStaticImage).

Usage

gdkPixbufLoaderGetPixbuf(object)

Arguments

object
A pixbuf loader.

Value

[GdkPixbuf] The GdkPixbuf that the loader is creating, or NULL if not enough data has been read to determine how to create the image buffer.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]