gdkPixbufLoaderNewWithType {RGtk2}	R Documentation

gdkPixbufLoaderNewWithType

Description

Creates a new pixbuf loader object that always attempts to parse image data as if it were an image of type image.type, instead of identifying the type automatically. Useful if you want an error if the image isn't the expected type, for loading image formats that can't be reliably identified by looking at the data, or if the user manually forces a specific type.

Usage

gdkPixbufLoaderNewWithType(image.type, .errwarn = TRUE)

Arguments

image.type
name of the image format to be loaded with the image

.errwarn
Whether to issue a warning on error or fail silently

Details

The list of supported image formats depends on what image loaders are installed, but typically "png", "jpeg", "gif", "tiff" and "xpm" are among the supported formats. To obtain the full list of supported image formats, call gdkPixbufFormatGetName on each of the GdkPixbufFormat structs returned by gdkPixbufGetFormats.

Value

A list containing the following elements:

retval
[GdkPixbufLoader] A newly-created pixbuf loader.

error
return location for an allocated GError, or NULL to ignore errors. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]