gdkPixbufLoaderSetSize {RGtk2}	R Documentation

gdkPixbufLoaderSetSize

Description

Causes the image to be scaled while it is loaded. The desired image size can be determined relative to the original size of the image by calling gdkPixbufLoaderSetSize from a signal handler for the ::size-prepared signal.

Usage

gdkPixbufLoaderSetSize(object, width, height)

Arguments

object
A pixbuf loader.

width
The desired width of the image being loaded.

height
The desired height of the image being loaded.

Details

Attempts to set the desired image size are ignored after the emission of the ::size-prepared signal. Since 2.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]