gdkPixbufLoaderWrite {RGtk2}	R Documentation

gdkPixbufLoaderWrite

Description

This will cause a pixbuf loader to parse the next count bytes of an image. It will return TRUE if the data was loaded successfully, and FALSE if an error occurred. In the latter case, the loader will be closed, and will not accept further writes. If FALSE is returned, error will be set to an error from the GDK_PIXBUF_ERROR or G_FILE_ERROR domains.

Usage

gdkPixbufLoaderWrite(object, buf, .errwarn = TRUE)

Arguments

object
A pixbuf loader.

buf
Pointer to image data.

.errwarn
Whether to issue a warning on error or fail silently

Value

A list containing the following elements:

retval
[logical] TRUE if the write was successful, or FALSE if the loader cannot parse the buffer.

error
return location for errors

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]