gdkPixbufNewFromData {RGtk2}	R Documentation

gdkPixbufNewFromData

Description

Creates a new GdkPixbuf out of in-memory image data. Currently only RGB images with 8 bits per sample are supported.

Usage

gdkPixbufNewFromData(data, colorspace, has.alpha, bits.per.sample, 
    width, height, rowstride)

Arguments

data
Image data in 8-bit/sample packed format

colorspace
Colorspace for the image data

has.alpha
Whether the data has an opacity channel

bits.per.sample
Number of bits per sample

width
Width of the image in pixels, must be > 0

height
Height of the image in pixels, must be > 0

rowstride
Distance in bytes between row starts

Value

[GdkPixbuf] A newly-created GdkPixbuf structure with a reference count of 1.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]