gdkPixbufNewFromFileAtScale {RGtk2}	R Documentation

gdkPixbufNewFromFileAtScale

Description

Creates a new pixbuf by loading an image from a file. The file format is detected automatically. If NULL is returned, then error will be set. Possible errors are in the GDK_PIXBUF_ERROR and G_FILE_ERROR domains. The image will be scaled to fit in the requested size, optionally preserving the image's aspect ratio.

Usage

gdkPixbufNewFromFileAtScale(filename, width, height, preserve.aspect.ratio, 
    .errwarn = TRUE)

Arguments

filename
Name of file to load, in the GLib file name encoding

width
The width the image should have or -1 to not constrain the width

height
The height the image should have or -1 to not constrain the height

preserve.aspect.ratio
TRUE to preserve the image's aspect ratio

.errwarn
Whether to issue a warning on error or fail silently

Details

When preserving the aspect ratio, a width of -1 will cause the image to be scaled to the exact given height, and a height of -1 will cause the image to be scaled to the exact given width. When not preserving aspect ratio, a width or height of -1 means to not scale the image at all in that dimension. Negative values for width and height are allowed since 2.8. Since 2.6

Value

A list containing the following elements:

retval
[GdkPixbuf] A newly-created pixbuf with a reference count of 1, or NULL if any of several error conditions occurred: the file could not be opened, there was no loader for the file's format, there was not enough memory to allocate the image buffer, or the image file contained invalid data.

error
Return location for an error

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]