gdkPixbufNewFromStreamAtScale {RGtk2}	R Documentation

gdkPixbufNewFromStreamAtScale

Description

Creates a new pixbuf by loading an image from an input stream.

Usage

gdkPixbufNewFromStreamAtScale(stream, width = -1, height = -1, 
    preserve.aspect.ratio = 1, cancellable = NULL, .errwarn = TRUE)

Arguments

stream
a GInputStream to load the pixbuf from

width
The width the image should have or -1 to not constrain the width

height
The height the image should have or -1 to not constrain the height

preserve.aspect.ratio
TRUE to preserve the image's aspect ratio

cancellable
optional GCancellable object, NULL to ignore

.errwarn
Whether to issue a warning on error or fail silently

Details

The file format is detected automatically. If NULL is returned, then error will be set. The cancellable can be used to abort the operation from another thread. If the operation was cancelled, the error GIO_ERROR_CANCELLED will be returned. Other possible errors are in the GDK_PIXBUF_ERROR and G_IO_ERROR domains.

The image will be scaled to fit in the requested size, optionally preserving the image's aspect ratio. When preserving the aspect ratio, a width of -1 will cause the image to be scaled to the exact given height, and a height of -1 will cause the image to be scaled to the exact given width. When not preserving aspect ratio, a width or height of -1 means to not scale the image at all in that dimension.

The stream is not closed. Since 2.14

Value

A list containing the following elements:

retval
[GdkPixbuf] A newly-created pixbuf, or NULL if any of several error conditions occurred: the file could not be opened, the image format is not supported, there was not enough memory to allocate the image buffer, the stream contained invalid data, or the operation was cancelled.

error
Return location for an error

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]