gdkPixbufNewSubpixbuf {RGtk2}	R Documentation

gdkPixbufNewSubpixbuf

Description

Creates a new pixbuf which represents a sub-region of src.pixbuf. The new pixbuf shares its pixels with the original pixbuf, so writing to one affects both. The new pixbuf holds a reference to src.pixbuf, so src.pixbuf will not be finalized until the new pixbuf is finalized.

Usage

gdkPixbufNewSubpixbuf(object, src.x, src.y, width, height)

Arguments

object
a GdkPixbuf

src.x
X coord in src.pixbuf

src.y
Y coord in src.pixbuf

width
width of region in src.pixbuf

height
height of region in src.pixbuf

Value

[GdkPixbuf] a new pixbuf

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]