gdkPixbufRenderThresholdAlpha {RGtk2}	R Documentation

gdkPixbufRenderThresholdAlpha

Description

Takes the opacity values in a rectangular portion of a pixbuf and thresholds them to produce a bi-level alpha mask that can be used as a clipping mask for a drawable.

Usage

gdkPixbufRenderThresholdAlpha(object, bitmap, src.x, src.y, dest.x, 
    dest.y, width = -1, height = -1, alpha.threshold)

Arguments

object
A pixbuf.

bitmap
Bitmap where the bilevel mask will be painted to.

src.x
Source X coordinate.

src.y
source Y coordinate.

dest.x
Destination X coordinate.

dest.y
Destination Y coordinate.

width
Width of region to threshold, or -1 to use pixbuf width

height
Height of region to threshold, or -1 to use pixbuf height

alpha.threshold
Opacity values below this will be painted as zero; all other values will be painted as one.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]