gdkPixbufRenderToDrawable {RGtk2}	R Documentation

gdkPixbufRenderToDrawable

Description

Renders a rectangular portion of a pixbuf to a drawable while using the specified GC. This is done using GdkRGB, so the specified drawable must have the GdkRGB visual and colormap. Note that this function will ignore the opacity information for images with an alpha channel; the GC must already have the clipping mask set if you want transparent regions to show through. WARNING: gdk_pixbuf_render_to_drawable has been deprecated since version 2.4 and should not be used in newly-written code. This function is obsolete. Use gdkDrawPixbuf instead.

Usage

gdkPixbufRenderToDrawable(object, drawable, gc, src.x = 0, src.y = 0, 
    dest.x, dest.y, width = -1, height = -1, dither = "GDK_RGB_DITHER_NORMAL", 
    x.dither = 0, y.dither = 0)

Arguments

object
A pixbuf.

drawable
Destination drawable.

gc
GC used for rendering.

src.x
Source X coordinate within pixbuf.

src.y
Source Y coordinate within pixbuf.

dest.x
Destination X coordinate within drawable.

dest.y
Destination Y coordinate within drawable.

width
Width of region to render, in pixels, or -1 to use pixbuf width

height
Height of region to render, in pixels, or -1 to use pixbuf height

dither
Dithering mode for GdkRGB.

x.dither
X offset for dither.

y.dither
Y offset for dither.

Details

For an explanation of dither offsets, see the GdkRGB documentation. In brief, the dither offset is important when re-rendering partial regions of an image to a rendered version of the full image, or for when the offsets to a base position change, as in scrolling. The dither matrix has to be shifted for consistent visual results. If you do not have any of these cases, the dither offsets can be both zero.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]