gdkPixbufRenderToDrawableAlpha {RGtk2}	R Documentation

gdkPixbufRenderToDrawableAlpha

Description

Renders a rectangular portion of a pixbuf to a drawable. The destination drawable must have a colormap. All windows have a colormap, however, pixmaps only have colormap by default if they were created with a non-NULL window argument. Otherwise a colormap must be set on them with gdk_drawable_set_colormap. WARNING: gdk_pixbuf_render_to_drawable_alpha has been deprecated since version 2.4 and should not be used in newly-written code. This function is obsolete. Use gdkDrawPixbuf instead.

Usage

gdkPixbufRenderToDrawableAlpha(object, drawable, src.x = 0, src.y = 0, 
    dest.x, dest.y, width = -1, height = -1, alpha.mode = NULL, 
    alpha.threshold = NULL, dither = "GDK_RGB_DITHER_NORMAL", 
    x.dither = 0, y.dither = 0)

Arguments

object
A pixbuf.

drawable
Destination drawable.

src.x
Source X coordinate within pixbuf.

src.y
Source Y coordinates within pixbuf.

dest.x
Destination X coordinate within drawable.

dest.y
Destination Y coordinate within drawable.

width
Width of region to render, in pixels, or -1 to use pixbuf width.

height
Height of region to render, in pixels, or -1 to use pixbuf height.

alpha.mode
Ignored. Present for backwards compatibility.

alpha.threshold
Ignored. Present for backwards compatibility

dither
Dithering mode for GdkRGB.

x.dither
X offset for dither.

y.dither
Y offset for dither.

Details

On older X servers, rendering pixbufs with an alpha channel involves round trips to the X server, and may be somewhat slow.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]