gdkPixbufRotateSimple {RGtk2}	R Documentation

gdkPixbufRotateSimple

Description

Rotates a pixbuf by a multiple of 90 degrees, and returns the result in a new pixbuf.

Usage

gdkPixbufRotateSimple(object, angle)

Arguments

object
a GdkPixbuf

angle
the angle to rotate by

Details

Since 2.6

Value

[GdkPixbuf] the new GdkPixbuf, or NULL if not enough memory could be allocated for it.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]