gdkPixbufSaturateAndPixelate {RGtk2}	R Documentation

gdkPixbufSaturateAndPixelate

Description

Modifies saturation and optionally pixelates src, placing the result in dest. src and dest may be the same pixbuf with no ill effects. If saturation is 1.0 then saturation is not changed. If it's less than 1.0, saturation is reduced (the image turns toward grayscale); if greater than 1.0, saturation is increased (the image gets more vivid colors). If pixelate is TRUE, then pixels are faded in a checkerboard pattern to create a pixelated image. src and dest must have the same image format, size, and rowstride.

Usage

gdkPixbufSaturateAndPixelate(object, dest, saturation, pixelate)

Arguments

object
source image

dest
place to write modified version of src

saturation
saturation factor

pixelate
whether to pixelate

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]