gdkPixbufSaveToBufferv {RGtk2}	R Documentation

gdkPixbufSaveToBufferv

Description

Saves pixbuf to a new buffer in format type, which is currently "jpeg", "tiff", "png", "ico" or "bmp". See gdkPixbufSaveToBuffer for more details.

Usage

gdkPixbufSaveToBufferv(object, type, option.keys, option.values, 
    .errwarn = TRUE)

Arguments

object
a GdkPixbuf.

type
name of file format.

option.keys
name of options to set, NULL-terminated

option.values
values for named options

.errwarn
Whether to issue a warning on error or fail silently

Details

Since 2.4

Value

A list containing the following elements:

retval
[logical] whether an error was set

buffer
location to receive a pointer to the new buffer.

buffer.size
location to receive the size of the new buffer.

error
return location for error, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]