gdkPixbufSaveToCallback {RGtk2}	R Documentation

gdkPixbufSaveToCallback

Description

Saves pixbuf in format type by feeding the produced data to a callback. Can be used when you want to store the image to something other than a file, such as an in-memory buffer or a socket. If error is set, FALSE will be returned. Possible errors include those in the GDK_PIXBUF_ERROR domain and whatever the save function generates.

Usage

gdkPixbufSaveToCallback(object, save.func, user.data, type, ..., 
    .errwarn = TRUE)

Arguments

object
a GdkPixbuf.

save.func
a function that is called to save each block of data that the save routine generates.

user.data
user data to pass to the save function.

type
name of file format.

...
whether an error was set

.errwarn
Whether to issue a warning on error or fail silently

Details

See gdkPixbufSave for more details. Since 2.4

Value

A list containing the following elements:

retval
[logical] whether an error was set

error
return location for error, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]