gdkPixbufSaveToCallbackv {RGtk2}	R Documentation

gdkPixbufSaveToCallbackv

Description

Saves pixbuf to a callback in format type, which is currently "jpeg", "png", "tiff", "ico" or "bmp". If error is set, FALSE will be returned. See gdkPixbufSaveToCallback for more details.

Usage

gdkPixbufSaveToCallbackv(object, save.func, user.data, type, 
    option.keys, option.values, .errwarn = TRUE)

Arguments

object
a GdkPixbuf.

save.func
a function that is called to save each block of data that the save routine generates.

user.data
user data to pass to the save function.

type
name of file format.

option.keys
name of options to set, NULL-terminated

option.values
values for named options

.errwarn
Whether to issue a warning on error or fail silently

Details

Since 2.4

Value

A list containing the following elements:

retval
[logical] whether an error was set

error
return location for error, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]