gdkPixbufSaveToStream {RGtk2}	R Documentation

gdkPixbufSaveToStream

Description

Saves pixbuf to an output stream.

Usage

gdkPixbufSaveToStream(object, stream, type, cancellable, .errwarn = TRUE)

Arguments

object
a GdkPixbuf

stream
a GOutputStream to save the pixbuf to

type
name of file format

cancellable
optional GCancellable object, NULL to ignore

.errwarn
Whether to issue a warning on error or fail silently

Details

Supported file formats are currently "jpeg", "tiff", "png", "ico" or "bmp". See gdkPixbufSaveToBuffer for more details.

The cancellable can be used to abort the operation from another thread. If the operation was cancelled, the error GIO_ERROR_CANCELLED will be returned. Other possible errors are in the GDK_PIXBUF_ERROR and G_IO_ERROR domains.

The stream is not closed. Since 2.14

Value

A list containing the following elements:

retval
[logical] TRUE if the pixbuf was saved successfully, FALSE if an error was set.

error
return location for error, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]