gdkPixbufScale {RGtk2}	R Documentation

gdkPixbufScale

Description

Creates a transformation of the source image src by scaling by scale.x and scale.y then translating by offset.x and offset.y, then renders the rectangle (dest.x, dest.y, dest.width, dest.height) of the resulting image onto the destination image replacing the previous contents.

Usage

gdkPixbufScale(object, dest, dest.x, dest.y, dest.width, dest.height, 
    offset.x, offset.y, scale.x, scale.y, interp.type)

Arguments

object
a GdkPixbuf

dest
the GdkPixbuf into which to render the results

dest.x
the left coordinate for region to render

dest.y
the top coordinate for region to render

dest.width
the width of the region to render

dest.height
the height of the region to render

offset.x
the offset in the X direction (currently rounded to an integer)

offset.y
the offset in the Y direction (currently rounded to an integer)

scale.x
the scale factor in the X direction

scale.y
the scale factor in the Y direction

interp.type
the interpolation type for the transformation.

Details

Try to use gdkPixbufScaleSimple first, this function is the industrial-strength power tool you can fall back to if gdkPixbufScaleSimple isn't powerful enough.

If the source rectangle overlaps the destination rectangle on the same pixbuf, it will be overwritten during the scaling which results in rendering artifacts.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]