gdkPixbufScaleSimple {RGtk2}	R Documentation

gdkPixbufScaleSimple

Description

Create a new GdkPixbuf containing a copy of src scaled to dest.width x dest.height. Leaves src unaffected. interp.type should be GDK_INTERP_NEAREST if you want maximum speed (but when scaling down GDK_INTERP_NEAREST is usually unusably ugly). The default interp.type should be GDK_INTERP_BILINEAR which offers reasonable quality and speed.

Usage

gdkPixbufScaleSimple(object, dest.width, dest.height, interp.type)

Arguments

object
a GdkPixbuf

dest.width
the width of destination image

dest.height
the height of destination image

interp.type
the interpolation type for the transformation.

Details

You can scale a sub-portion of src by creating a sub-pixbuf pointing into src; see gdkPixbufNewSubpixbuf.

For more complicated scaling/compositing see gdkPixbufScale and gdkPixbufComposite.

Value

[GdkPixbuf] the new GdkPixbuf, or NULL if not enough memory could be allocated for it.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]