gdkPixmapColormapCreateFromXpmD {RGtk2}	R Documentation

gdkPixmapColormapCreateFromXpmD

Description

Create a pixmap from data in XPM format using a particular colormap.

Usage

gdkPixmapColormapCreateFromXpmD(drawable, colormap, transparent.color, 
    data)

Arguments

drawable
a GdkDrawable, used to determine default values for the new pixmap. Can be NULL if colormap is given.

colormap
the GdkColormap that the new pixmap will be use. If omitted, the colormap for window will be used.

transparent.color
the color to be used for the pixels that are transparent in the input file. Can be NULL, in which case a default color will be used.

data
Pointer to a string containing the XPM data.

Value

A list containing the following elements:

retval
[GdkPixmap] the GdkPixmap. [ transfer none ]

mask
a pointer to a place to store a bitmap representing the transparency mask of the XPM file. Can be NULL, in which case transparency will be ignored.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]