gdkPixmapCreateFromXpm {RGtk2}	R Documentation

gdkPixmapCreateFromXpm

Description

Create a pixmap from a XPM file.

Usage

gdkPixmapCreateFromXpm(drawable, transparent.color, filename)

Arguments

drawable
a GdkDrawable, used to determine default values for the new pixmap.

transparent.color
the color to be used for the pixels that are transparent in the input file. Can be NULL, in which case a default color will be used.

filename
the filename of a file containing XPM data.

Value

A list containing the following elements:

retval
[GdkPixmap] the GdkPixmap. [ transfer none ]

mask
(out) a pointer to a place to store a bitmap representing the transparency mask of the XPM file. Can be NULL, in which case transparency will be ignored.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]