gdkPixmapNew {RGtk2}	R Documentation

gdkPixmapNew

Description

Create a new pixmap with a given size and depth.

Usage

gdkPixmapNew(drawable = NULL, width, height, depth = -1)

Arguments

drawable
A GdkDrawable, used to determine default values for the new pixmap. Can be NULL if depth is specified,

width
The width of the new pixmap in pixels.

height
The height of the new pixmap in pixels.

depth
The depth (number of bits per pixel) of the new pixmap. If -1, and drawable is not NULL, the depth of the new pixmap will be equal to that of drawable.

Value

[GdkPixmap] the GdkPixmap

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]