gdkPropertyChange {RGtk2}	R Documentation

gdkPropertyChange

Description

Changes the contents of a property on a window.

Usage

gdkPropertyChange(object, property, type, format, mode, data)

Arguments

object
a GdkWindow.

property
the property to change.

type
the new type for the property. If mode is GDK_PROP_MODE_PREPEND or GDK_PROP_MODE_APPEND, then this must match the existing type or an error will occur.

format
the new format for the property. If mode is GDK_PROP_MODE_PREPEND or GDK_PROP_MODE_APPEND, then this must match the existing format or an error will occur.

mode
a value describing how the new data is to be combined with the current data.

data
the data (a guchar * gushort *, or gulong *, depending on format), cast to a guchar *.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]