gdkPropertyGet {RGtk2}	R Documentation

gdkPropertyGet

Description

Retrieves a portion of the contents of a property. If the property does not exist, then the function returns FALSE, and GDK_NONE will be stored in actual.property.type.

Usage

gdkPropertyGet(object, property, type, offset, length, pdelete)

Arguments

object
a GdkWindow.

property
the property to retrieve.

type
the desired property type, or GDK_NONE, if any type of data is acceptable. If this does not match the actual type, then actual.format and actual.length will be filled in, a warning will be printed to stderr and no data will be returned.

offset
the offset into the property at which to begin retrieving data, in 4 byte units.

length
the length of the data to retrieve in bytes. Data is considered to be retrieved in 4 byte chunks, so length will be rounded up to the next highest 4 byte boundary (so be careful not to pass a value that might overflow when rounded up).

pdelete
if TRUE, delete the property after retrieving the data.

Value

A list containing the following elements:

retval
[logical] TRUE if data was successfully received and stored in data, otherwise FALSE.

actual.property.type
location to store the actual type of the property.

actual.format
location to store the actual return format of the data; either 8, 16 or 32 bits.

actual.length
location to store the length of the retrieved data, in bytes. Data returned in the 32 bit format is stored in a long variable, so the actual number of 32 bit elements should be be calculated via actual.length/sizeof(glong) to ensure portability to 64 bit systems.

data
location to store a pointer to the data. The retrieved data should be freed with gFree() when you are finished using it.

Note

The xgetwindowproperty() function that gdkPropertyGet uses has a very confusing and complicated set of semantics. Unfortunately, gdkPropertyGet makes the situation worse instead of better (the semantics should be considered undefined), and also prints warnings to stderr in cases where it should return a useful error to the program. You are advised to use xgetwindowproperty() directly until a replacement function for gdkPropertyGet is provided.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]