gdkQueryDepths {RGtk2}	R Documentation

gdkQueryDepths

Description

This function returns the available bit depths for the default screen. It's equivalent to listing the visuals (gdkListVisuals) and then looking at the depth field in each visual, removing duplicates.

Usage

gdkQueryDepths()

Details

The list returned by this function should not be freed.

Value

A list containing the following elements:

depths
return location for available depths. [ out ][ array ]

count
return location for number of available depths. [ out ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]