gdkRectangleIntersect {RGtk2}	R Documentation

gdkRectangleIntersect

Description

Calculates the intersection of two rectangles. It is allowed for dest to be the same as either src1 or src2. If the rectangles do not intersect, dest's width and height is set to 0 and its x and y values are undefined. If you are only interested in whether the rectangles intersect, but not in the intersecting area itself, pass NULL for dest.

Usage

gdkRectangleIntersect(src1, src2)

Arguments

src1
a GdkRectangle

src2
a GdkRectangle

Value

A list containing the following elements:

retval
[logical] TRUE if the rectangles intersect.

dest
return location for the intersection of src1 and src2, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]