gdkRegionPolygon {RGtk2}	R Documentation

gdkRegionPolygon

Description

Creates a new GdkRegion using the polygon defined by a number of points.

Usage

gdkRegionPolygon(points, fill.rule)

Arguments

points
a list of GdkPoint structs

fill.rule
specifies which pixels are included in the region when the polygon overlaps itself.

Value

[GdkRegion] a new GdkRegion based on the given polygon

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]