gdkRgbDitherable {RGtk2}	R Documentation

gdkRgbDitherable

Description

Determines whether the preferred visual is ditherable. This function may be useful for presenting a user interface choice to the user about which dither mode is desired; if the display is not ditherable, it may make sense to gray out or hide the corresponding UI widget.

Usage

gdkRgbDitherable()

Value

[logical] TRUE if the preferred visual is ditherable.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]