gdkRgbGcSetBackground {RGtk2}	R Documentation

gdkRgbGcSetBackground

Description

Sets the background color in gc to the specified color (or the closest approximation, in the case of limited visuals). WARNING: gdk_rgb_gc_set_background is deprecated and should not be used in newly-written code.

Usage

gdkRgbGcSetBackground(gc, rgb)

Arguments

gc
The GdkGC to modify.

rgb
The color, represented as a 0xRRGGBB integer value.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]