gdkRgbGetColormap {RGtk2}	R Documentation

gdkRgbGetColormap

Description

Get the preferred colormap for rendering image data. Not a very useful function; historically, GDK could only render RGB image data to one colormap and visual, but in the current version it can render to any colormap and visual. So there's no need to call this function.

Usage

gdkRgbGetColormap()

Value

[GdkColormap] the preferred colormap. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]