gdkRgbSetInstall {RGtk2}	R Documentation

gdkRgbSetInstall

Description

If install is TRUE, directs GdkRGB to always install a new "private" colormap rather than trying to find a best fit with the colors already allocated. Ordinarily, GdkRGB will install a colormap only if a sufficient cube cannot be allocated.

Usage

gdkRgbSetInstall(install)

Arguments

install
TRUE to set install mode.

Details

A private colormap has more colors, leading to better quality display, but also leads to the dreaded "colormap flashing" effect.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]