gdkRgbSetMinColors {RGtk2}	R Documentation

gdkRgbSetMinColors

Description

Sets the minimum number of colors for the color cube. Generally, GdkRGB tries to allocate the largest color cube it can. If it can't allocate a color cube at least as large as min.colors, it installs a private colormap.

Usage

gdkRgbSetMinColors(min.colors)

Arguments

min.colors
The minimum number of colors accepted.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]