gdkRgbXpixelFromRgb {RGtk2}	R Documentation

gdkRgbXpixelFromRgb

Description

Finds the X pixel closest in color to the rgb color specified. This value may be used to set the pixel field of a GdkColor struct. WARNING: gdk_rgb_xpixel_from_rgb is deprecated and should not be used in newly-written code.

Usage

gdkRgbXpixelFromRgb(rgb)

Arguments

rgb
The color, represented as a 0xRRGGBB integer value.

Value

[numeric] The X pixel value.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]