gdkScreenBroadcastClientMessage {RGtk2}	R Documentation

gdkScreenBroadcastClientMessage

Description

On X11, sends an X ClientMessage event to all toplevel windows on screen.

Usage

gdkScreenBroadcastClientMessage(object, event)

Arguments

object
the GdkScreen where the event will be broadcasted.

event
the GdkEvent.

Details

Toplevel windows are determined by checking for the WM_STATE property, as described in the Inter-Client Communication Conventions Manual (ICCCM). If no windows are found with the WM_STATE property set, the message is sent to all children of the root window.

On Windows, broadcasts a message registered with the name GDK_WIN32_CLIENT_MESSAGE to all top-level windows. The amount of data is limited to one long, i.e. four bytes. Since 2.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]