gdkScreenGetActiveWindow {RGtk2}	R Documentation

gdkScreenGetActiveWindow

Description

Returns the screen's currently active window.

Usage

gdkScreenGetActiveWindow(object)

Arguments

object
a GdkScreen

Details

On X11, this is done by inspecting the _NET_ACTIVE_WINDOW property on the root window, as described in the Extended Window Manager Hints (http://www.freedesktop.org/Standards/wm-spec). If there is no currently currently active window, or the window manager does not support the _NET_ACTIVE_WINDOW hint, this function returns NULL.

On other platforms, this function may return NULL, depending on whether it is implementable on that platform. Since 2.10

Value

[GdkWindow] the currently active window, or NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]