gdkScreenGetMonitorAtWindow {RGtk2}	R Documentation

gdkScreenGetMonitorAtWindow

Description

Returns the number of the monitor in which the largest area of the bounding rectangle of window resides.

Usage

gdkScreenGetMonitorAtWindow(object, window)

Arguments

object
a GdkScreen.

window
a GdkWindow

Details

Since 2.2

Value

[integer] the monitor number in which most of window is located, or if window does not intersect any monitors, a monitor, close to window.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]