gdkScreenGetMonitorGeometry {RGtk2}	R Documentation

gdkScreenGetMonitorGeometry

Description

Retrieves the GdkRectangle representing the size and position of the individual monitor within the entire screen area.

Usage

gdkScreenGetMonitorGeometry(object, monitor.num)

Arguments

object
a GdkScreen

monitor.num
the monitor number, between 0 and gdk_screen_get_n_monitors (screen)

Details

Note that the size of the entire screen area can be retrieved via gdkScreenGetWidth and gdkScreenGetHeight. Since 2.2

Value

A list containing the following elements:

dest
a GdkRectangle to be filled with the monitor geometry

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]