gdkScreenGetPrimaryMonitor {RGtk2}	R Documentation

gdkScreenGetPrimaryMonitor

Description

Gets the primary monitor for screen. The primary monitor is considered the monitor where the 'main desktop' lives. While normal application windows typically allow the window manager to place the windows, specialized desktop applications such as panels should place themselves on the primary monitor.

Usage

gdkScreenGetPrimaryMonitor(object)

Arguments

object
a GdkScreen.

Details

If no primary monitor is configured by the user, the return value will be 0, defaulting to the first monitor. Since 2.20

Value

[integer] An integer index for the primary monitor, or 0 if none is configured.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]