gdkScreenGetRgbVisual {RGtk2}	R Documentation

gdkScreenGetRgbVisual

Description

Gets a "preferred visual" chosen by GdkRGB for rendering image data on screen. In previous versions of GDK, this was the only visual GdkRGB could use for rendering. In current versions, it's simply the visual GdkRGB would have chosen as the optimal one in those previous versions. GdkRGB can now render to drawables with any visual.

Usage

gdkScreenGetRgbVisual(object)

Arguments

object
a GdkScreen

Details

Since 2.2

Value

[GdkVisual] The GdkVisual chosen by GdkRGB. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]