gdkScreenGetRgbaColormap {RGtk2}	R Documentation

gdkScreenGetRgbaColormap

Description

Gets a colormap to use for creating windows or pixmaps with an alpha channel. The windowing system on which GTK+ is running may not support this capability, in which case NULL will be returned. Even if a non-NULL value is returned, its possible that the window's alpha channel won't be honored when displaying the window on the screen: in particular, for X an appropriate windowing manager and compositing manager must be running to provide appropriate display.

Usage

gdkScreenGetRgbaColormap(object)

Arguments

object
a GdkScreen.

Details

This functionality is not implemented in the Windows backend.

For setting an overall opacity for a top-level window, see gdkWindowSetOpacity. Since 2.8

Value

[GdkColormap] a colormap to use for windows with an alpha channel or NULL if the capability is not available. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]