gdkScreenGetSetting {RGtk2}	R Documentation

gdkScreenGetSetting

Description

Retrieves a desktop-wide setting such as double-click time for the GdkScreen screen.

Usage

gdkScreenGetSetting(object, name)

Arguments

object
the GdkScreen where the setting is located

name
the name of the setting

Details

FIXME needs a list of valid settings here, or a link to more information. Since 2.2

Value

A list containing the following elements:

retval
[logical] TRUE if the setting existed and a value was stored in value, FALSE otherwise.

value
location to store the value of the setting

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]