gdkScreenGetToplevelWindows {RGtk2}	R Documentation

gdkScreenGetToplevelWindows

Description

Obtains a list of all toplevel windows known to GDK on the screen screen. A toplevel window is a child of the root window (see gdkGetDefaultRootWindow).

Usage

gdkScreenGetToplevelWindows(object)

Arguments

object
The GdkScreen where the toplevels are located.

Details

but its elements need not be freed. Since 2.2

Value

[list] list of toplevel windows,

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]