gdkScreenGetWindowStack {RGtk2}	R Documentation

gdkScreenGetWindowStack

Description

Returns a list of GdkWindows representing the current window stack.

Usage

gdkScreenGetWindowStack(object)

Arguments

object
a GdkScreen

Details

On X11, this is done by inspecting the _NET_CLIENT_LIST_STACKING property on the root window, as described in the Extended Window Manager Hints (http://www.freedesktop.org/Standards/wm-spec). If the window manager does not support the _NET_CLIENT_LIST_STACKING hint, this function returns NULL.

On other platforms, this function may return NULL, depending on whether it is implementable on that platform.

The returned list is newly allocated and owns references to the windows it contains, Since 2.10

Value

[list] a list of GdkWindows for the current window stack, or NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]