gdkScreenIsComposited {RGtk2}	R Documentation

gdkScreenIsComposited

Description

Returns whether windows with an RGBA visual can reasonably be expected to have their alpha channel drawn correctly on the screen.

Usage

gdkScreenIsComposited(object)

Arguments

object
a GdkScreen

Details

On X11 this function returns whether a compositing manager is compositing screen. Since 2.10

Value

[logical] Whether windows with RGBA visuals can reasonably be expected to have their alpha channels drawn correctly on the screen.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]