gdkScreenSetFontOptions {RGtk2}	R Documentation

gdkScreenSetFontOptions

Description

Sets the default font options for the screen. These options will be set on any PangoContext's newly created with gdkPangoContextGetForScreen. Changing the default set of font options does not affect contexts that have already been created.

Usage

gdkScreenSetFontOptions(object, options)

Arguments

object
a GdkScreen

options
a CairoFontOptions, or NULL to unset any previously set default font options. [ allow-none ]

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]