gdkScreenSetResolution {RGtk2}	R Documentation

gdkScreenSetResolution

Description

Sets the resolution for font handling on the screen. This is a scale factor between points specified in a PangoFontDescription and cairo units. The default value is 96, meaning that a 10 point font will be 13 units high. (10 * 96. / 72. = 13.3).

Usage

gdkScreenSetResolution(object, dpi)

Arguments

object
a GdkScreen

dpi
the resolution in "dots per inch". (Physical inches aren't actually involved; the terminology is conventional.)

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]