gdkScreenWidthMm {RGtk2}	R Documentation

gdkScreenWidthMm

Description

Returns the width of the default screen in millimeters. Note that on many X servers this value will not be correct.

Usage

gdkScreenWidthMm()

Value

[integer] the width of the default screen in millimeters, though it is not always correct.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]