gdkSpawnCommandLineOnScreen {RGtk2}	R Documentation

gdkSpawnCommandLineOnScreen

Description

Like gSpawnCommandLineAsync(), except the child process is spawned in such an environment that on calling gdkDisplayOpen it would be returned a GdkDisplay with screen as the default screen.

Usage

gdkSpawnCommandLineOnScreen(screen, command.line, .errwarn = TRUE)

Arguments

screen
a GdkScreen

command.line
a command line

.errwarn
Whether to issue a warning on error or fail silently

Details

This is useful for applications which wish to launch an application on a specific screen. Since 2.4

Value

A list containing the following elements:

retval
[logical] TRUE on success, FALSE if error is set.

error
return location for errors

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]