gdkTestRenderSync {RGtk2}	R Documentation

gdkTestRenderSync

Description

This function retrives a pixel from window to force the windowing system to carry out any pending rendering commands. This function is intended to be used to syncronize with rendering pipelines, to benchmark windowing system rendering operations.

Usage

gdkTestRenderSync(window)

Arguments

window
a mapped GdkWindow

Details

Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]