gdkTestSimulateButton {RGtk2}	R Documentation

gdkTestSimulateButton

Description

This function is intended to be used in Gtk+ test programs. It will warp the mouse pointer to the given (x,y) corrdinates within window and simulate a button press or release event. Because the mouse pointer needs to be warped to the target location, use of this function outside of test programs that run in their own virtual windowing system (e.g. Xvfb) is not recommended. Also, gtkTestSimulateButton() is a fairly low level function, for most testing purposes, gtkTestWidgetClick is the right function to call which will generate a button press event followed by its accompanying button release event.

Usage

gdkTestSimulateButton(window, x, y, button, modifiers, button.pressrelease)

Arguments

window
Gdk window to simulate a button event for.

x
x coordinate within window for the button event.

y
y coordinate within window for the button event.

button
Number of the pointer button for the event, usually 1, 2 or 3.

modifiers
Keyboard modifiers the event is setup with.

button.pressrelease
either GDK_BUTTON_PRESS or GDK_BUTTON_RELEASE

Details

Since 2.14

Value

[logical] wether all actions neccessary for a button event simulation were carried out successfully.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]