gdkTextMeasure {RGtk2}	R Documentation

gdkTextMeasure

Description

Determines the distance from the origin to the rightmost portion of a string when drawn. This is not the correct value for determining the origin of the next portion when drawing text in multiple pieces. See gdkTextWidth. WARNING: gdk_text_measure is deprecated and should not be used in newly-written code.

Usage

gdkTextMeasure(object, text, text.length = -1)

Arguments

object
a GdkFont

text
the text to measure.

text.length
the length of the text in bytes.

Value

[integer] the right bearing of the string in pixels.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]