gdkVisualGetBestWithDepth {RGtk2}	R Documentation

gdkVisualGetBestWithDepth

Description

Get the best visual with depth depth for the default GDK screen. Color visuals and visuals with mutable colormaps are preferred over grayscale or fixed-colormap visuals. The return value should not be freed. NULL may be returned if no visual supports depth.

Usage

gdkVisualGetBestWithDepth(depth)

Arguments

depth
a bit depth

Value

[GdkVisual] best visual for the given depth. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]