gdkWindowAddFilter {RGtk2}	R Documentation

gdkWindowAddFilter

Description

Adds an event filter to window, allowing you to intercept events before they reach GDK. This is a low-level operation and makes it easy to break GDK and/or GTK+, so you have to know what you're doing. Pass NULL for window to get all events for all windows, instead of events for a specific window.

Usage

gdkWindowAddFilter(object, fun, data)

Arguments

object
a GdkWindow

data
data to pass to filter callback

Details

See gdkDisplayAddClientMessageFilter if you are interested in X ClientMessage events.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]