gdkWindowAtPointer {RGtk2}	R Documentation

gdkWindowAtPointer

Description

Obtains the window underneath the mouse pointer, returning the location of that window in win.x, win.y. Returns NULL if the window under the mouse pointer is not known to GDK (if the window belongs to another application and a GdkWindow hasn't been created for it with gdkWindowForeignNew())

Usage

gdkWindowAtPointer()

Details

NOTE: For multihead-aware widgets or applications use gdkDisplayGetWindowAtPointer instead.

Value

A list containing the following elements:

retval
[GdkWindow] window under the mouse pointer. [ transfer none ]

win.x
return location for origin of the window under the pointer. [ out ][ allow-none ]

win.y
return location for origin of the window under the pointer. [ out ][ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]