gdkWindowBeginResizeDrag {RGtk2}	R Documentation

gdkWindowBeginResizeDrag

Description

Begins a window resize operation (for a toplevel window). You might use this function to implement a "window resize grip," for example; in fact GtkStatusbar uses it. The function works best with window managers that support the Extended Window Manager Hints (http://www.freedesktop.org/Standards/wm-spec), but has a fallback implementation for other window managers.

Usage

gdkWindowBeginResizeDrag(object, edge, button, root.x, root.y, 
    timestamp)

Arguments

object
a toplevel GdkWindow

edge
the edge or corner from which the drag is started

button
the button being used to drag

root.x
root window X coordinate of mouse click that began the drag

root.y
root window Y coordinate of mouse click that began the drag

timestamp
timestamp of mouse click that began the drag (use gdkEventGetTime)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]