gdkWindowClearAreaE {RGtk2}	R Documentation

gdkWindowClearAreaE

Description

Like gdkWindowClearArea, but also generates an expose event for the cleared area.

Usage

gdkWindowClearAreaE(object, x, y, width, height)

Arguments

object
a GdkWindow

x
x coordinate of rectangle to clear

y
y coordinate of rectangle to clear

width
width of rectangle to clear

height
height of rectangle to clear

Details

This function has a stupid name because it dates back to the mists time, pre-GDK-1.0.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]