gdkWindowDeiconify {RGtk2}	R Documentation

gdkWindowDeiconify

Description

Attempt to deiconify (unminimize) window. On X11 the window manager may choose to ignore the request to deiconify. When using GTK+, use gtkWindowDeiconify instead of the GdkWindow variant. Or better yet, you probably want to use gtkWindowPresent, which raises the window, focuses it, unminimizes it, and puts it on the current desktop.

Usage

gdkWindowDeiconify(object)

Arguments

object
a toplevel GdkWindow

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]