gdkWindowDestroy {RGtk2}	R Documentation

gdkWindowDestroy

Description

Destroys the window system resources associated with window and decrements window's reference count. The window system resources for all children of window are also destroyed, but the children's reference counts are not decremented.

Usage

gdkWindowDestroy(object)

Arguments

object
a GdkWindow

Details

Note that a window will not be destroyed automatically when its reference count reaches zero. You must call this function yourself before that happens.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]