gdkWindowEnableSynchronizedConfigure {RGtk2}	R Documentation

gdkWindowEnableSynchronizedConfigure

Description

Indicates that the application will cooperate with the window system in synchronizing the window repaint with the window manager during resizing operations. After an application calls this function, it must call gdkWindowConfigureFinished every time it has finished all processing associated with a set of Configure events. Toplevel GTK+ windows automatically use this protocol.

Usage

gdkWindowEnableSynchronizedConfigure(object)

Arguments

object
a toplevel GdkWindow

Details

On X, calling this function makes window participate in the _NET_WM_SYNC_REQUEST window manager protocol. Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]