gdkWindowEndPaint {RGtk2}	R Documentation

gdkWindowEndPaint

Description

Indicates that the backing store created by the most recent call to gdkWindowBeginPaintRegion should be copied onscreen and deleted, leaving the next-most-recent backing store or no backing store at all as the active paint region. See gdkWindowBeginPaintRegion for full details. It is an error to call this function without a matching gdkWindowBeginPaintRegion first.

Usage

gdkWindowEndPaint(object)

Arguments

object
a GdkWindow

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]