gdkWindowEnsureNative {RGtk2}	R Documentation

gdkWindowEnsureNative

Description

Tries to ensure that there is a window-system native window for this GdkWindow. This may fail in some situations, returning FALSE.

Usage

gdkWindowEnsureNative(object)

Arguments

object
a GdkWindow

Details

Offscreen window and children of them can never have native windows.

Some backends may not support native child windows. Since 2.18

Value

[logical] TRUE if the window has a native window, FALSE otherwise

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]