gdkWindowGetChildren {RGtk2}	R Documentation

gdkWindowGetChildren

Description

Gets the list of children of window known to GDK. This function only returns children created via GDK, so for example it's useless when used with the root window; it only returns windows an application created itself.

Usage

gdkWindowGetChildren(object)

Arguments

object
a GdkWindow

Details

The returned list must be freed, but the elements in the list need not be.

Value

[list] list of child windows inside window

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]