gdkWindowGetCursor {RGtk2}	R Documentation

gdkWindowGetCursor

Description

Retrieves a GdkCursor pointer for the cursor currently set on the specified GdkWindow, or NULL. If the return value is NULL then there is no custom cursor set on the specified window, and it is using the cursor for its parent window.

Usage

gdkWindowGetCursor(object)

Arguments

object
a GdkWindow

Details

Since 2.18

Value

[GdkCursor] a GdkCursor, or NULL. The returned object is owned by the GdkWindow and should not be unreferenced directly. Use gdkWindowSetCursor to unset the cursor of the window

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]