gdkWindowGetDeskrelativeOrigin {RGtk2}	R Documentation

gdkWindowGetDeskrelativeOrigin

Description

This gets the origin of a GdkWindow relative to an Enlightenment-window-manager desktop. As long as you don't assume that the user's desktop/workspace covers the entire root window (i.e. you don't assume that the desktop begins at root window coordinate 0,0) this function is not necessary. It's deprecated for that reason. WARNING: gdk_window_get_deskrelative_origin is deprecated and should not be used in newly-written code.

Usage

gdkWindowGetDeskrelativeOrigin(object)

Arguments

object
a toplevel GdkWindow

Value

A list containing the following elements:

retval
[logical] not meaningful

x
return location for X coordinate

y
return location for Y coordinate

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]