gdkWindowGetFrameExtents {RGtk2}	R Documentation

gdkWindowGetFrameExtents

Description

Obtains the bounding box of the window, including window manager titlebar/borders if any. The frame position is given in root window coordinates. To get the position of the window itself (rather than the frame) in root window coordinates, use gdkWindowGetOrigin.

Usage

gdkWindowGetFrameExtents(object)

Arguments

object
a GdkWindow

Value

A list containing the following elements:

rect
rectangle to fill with bounding box of the window frame

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]