gdkWindowGetGeometry {RGtk2}	R Documentation

gdkWindowGetGeometry

Description

Any of the return location arguments to this function may be NULL, if you aren't interested in getting the value of that field.

Usage

gdkWindowGetGeometry(object)

Arguments

object
a GdkWindow

Details

The X and Y coordinates returned are relative to the parent window of window, which for toplevels usually means relative to the window decorations (titlebar, etc.) rather than relative to the root window (screen-size background window).

On the X11 platform, the geometry is obtained from the X server, so reflects the latest position of window; this may be out-of-sync with the position of window delivered in the most-recently-processed GdkEventConfigure. gdkWindowGetPosition in contrast gets the position from the most recent configure event. PLEASE NOTE: If window is not a toplevel, it is much better to call gdkWindowGetPosition and gdkDrawableGetSize instead, because it avoids the roundtrip to the X server and because gdkDrawableGetSize supports the full 32-bit coordinate space, whereas gdkWindowGetGeometry is restricted to the 16-bit coordinates of X11.

Value

A list containing the following elements:

x
return location for X coordinate of window (relative to its parent)

y
return location for Y coordinate of window (relative to its parent)

width
return location for width of window

height
return location for height of window

depth
return location for bit depth of window

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]