gdkWindowGetInternalPaintInfo {RGtk2}	R Documentation

gdkWindowGetInternalPaintInfo

Description

If you bypass the GDK layer and use windowing system primitives to draw directly onto a GdkWindow, then you need to deal with two details: there may be an offset between GDK coordinates and windowing system coordinates, and GDK may have redirected drawing to a offscreen pixmap as the result of a gdkWindowBeginPaintRegion calls. This function allows retrieving the information you need to compensate for these effects.

Usage

gdkWindowGetInternalPaintInfo(object)

Arguments

object
a GdkWindow

Details

This function exposes details of the GDK implementation, and is thus likely to change in future releases of GDK.

Value

A list containing the following elements:

real.drawable
location to store the drawable to which drawing should be done. [ out ]

x.offset
location to store the X offset between coordinates in window, and the underlying window system primitive coordinates for *real.drawable. [ out ]

y.offset
location to store the Y offset between coordinates in window, and the underlying window system primitive coordinates for *real.drawable. [ out ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]