gdkWindowGetOrigin {RGtk2}	R Documentation

gdkWindowGetOrigin

Description

Obtains the position of a window in root window coordinates. (Compare with gdkWindowGetPosition and gdkWindowGetGeometry which return the position of a window relative to its parent window.)

Usage

gdkWindowGetOrigin(object)

Arguments

object
a GdkWindow

Value

A list containing the following elements:

retval
[integer] not meaningful, ignore

x
return location for X coordinate

y
return location for Y coordinate

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]