gdkWindowGetParent {RGtk2}	R Documentation

gdkWindowGetParent

Description

Obtains the parent of window, as known to GDK. Does not query the X server; thus this returns the parent as passed to gdkWindowNew, not the actual parent. This should never matter unless you're using Xlib calls mixed with GDK calls on the X11 platform. It may also matter for toplevel windows, because the window manager may choose to reparent them.

Usage

gdkWindowGetParent(object)

Arguments

object
a GdkWindow

Value

[GdkWindow] parent of window

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]