gdkWindowGetPointer {RGtk2}	R Documentation

gdkWindowGetPointer

Description

Obtains the current pointer position and modifier state. The position is given in coordinates relative to the upper left corner of window.

Usage

gdkWindowGetPointer(object)

Arguments

object
a GdkWindow

Value

A list containing the following elements:

retval
[GdkWindow] the window containing the pointer (as with gdkWindowAtPointer), or NULL if the window containing the pointer isn't known to GDK. [ transfer none ]

x
return location for X coordinate of pointer or NULL to not return the X coordinate. [ out ][ allow-none ]

y
return location for Y coordinate of pointer or NULL to not return the Y coordinate. [ out ][ allow-none ]

mask
return location for modifier mask or NULL to not return the modifier mask. [ out ][ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]