gdkWindowGetPosition {RGtk2}	R Documentation

gdkWindowGetPosition

Description

Obtains the position of the window as reported in the most-recently-processed GdkEventConfigure. Contrast with gdkWindowGetGeometry which queries the X server for the current window position, regardless of which events have been received or processed.

Usage

gdkWindowGetPosition(object)

Arguments

object
a GdkWindow

Details

The position coordinates are relative to the window's parent window.

Value

A list containing the following elements:

x
X coordinate of window

y
Y coordinate of window

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]