gdkWindowGetRootCoords {RGtk2}	R Documentation

gdkWindowGetRootCoords

Description

Obtains the position of a window position in root window coordinates. This is similar to gdkWindowGetOrigin but allows you go pass in any position in the window, not just the origin.

Usage

gdkWindowGetRootCoords(object, x, y)

Arguments

object
a GdkWindow

x
X coordinate in window

y
Y coordinate in window

Details

Since 2.18

Value

A list containing the following elements:

root.x
return location for X coordinate

root.y
return location for Y coordinate

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]