gdkWindowGetToplevels {RGtk2}	R Documentation

gdkWindowGetToplevels

Description

Obtains a list of all toplevel windows known to GDK on the default screen (see gdkScreenGetToplevelWindows). A toplevel window is a child of the root window (see gdkGetDefaultRootWindow). WARNING: gdk_window_get_toplevels has been deprecated since version 2.16 and should not be used in newly-written code. Use gdkScreenGetToplevelWindows instead.

Usage

gdkWindowGetToplevels()

Details

but its elements need not be freed.

Value

[list] list of toplevel windows,

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]