gdkWindowGetUpdateArea {RGtk2}	R Documentation

gdkWindowGetUpdateArea

Description

Transfers ownership of the update area from window to the caller of the function. That is, after calling this function, window will no longer have an invalid/dirty region; the update area is removed from window and handed to you. If a window has no update area, gdkWindowGetUpdateArea returns NULL.

Usage

gdkWindowGetUpdateArea(object)

Arguments

object
a GdkWindow

Value

[GdkRegion] the update area for window

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]