gdkWindowInputShapeCombineMask {RGtk2}	R Documentation

gdkWindowInputShapeCombineMask

Description

Like gdkWindowShapeCombineMask, but the shape applies only to event handling. Mouse events which happen while the pointer position corresponds to an unset bit in the mask will be passed on the window below window.

Usage

gdkWindowInputShapeCombineMask(object, mask, x, y)

Arguments

object
a GdkWindow

mask
shape mask, or NULL. [ allow-none ]

x
X position of shape mask with respect to window

y
Y position of shape mask with respect to window

Details

An input shape is typically used with RGBA windows. The alpha channel of the window defines which pixels are invisible and allows for nicely antialiased borders, and the input shape controls where the window is "clickable".

On the X11 platform, this requires version 1.1 of the shape extension.

On the Win32 platform, this functionality is not present and the function does nothing. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]