gdkWindowInvalidateRegion {RGtk2}	R Documentation

gdkWindowInvalidateRegion

Description

Adds region to the update area for window. The update area is the region that needs to be redrawn, or "dirty region." The call gdkWindowProcessUpdates sends one or more expose events to the window, which together cover the entire update area. An application would normally redraw the contents of window in response to those expose events.

Usage

gdkWindowInvalidateRegion(object, region, invalidate.children)

Arguments

object
a GdkWindow

region
a GdkRegion

invalidate.children
TRUE to also invalidate child windows

Details

GDK will call gdkWindowProcessAllUpdates on your behalf whenever your program returns to the main loop and becomes idle, so normally there's no need to do that manually, you just need to invalidate regions that you know should be redrawn.

The invalidate.children parameter controls whether the region of each child window that intersects region will also be invalidated. If FALSE, then the update area for child windows will remain unaffected. See gdk_window_invalidate_maybe_recurse if you need fine grained control over which children are invalidated.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]