gdkWindowMergeChildInputShapes {RGtk2}	R Documentation

gdkWindowMergeChildInputShapes

Description

Merges the input shape masks for any child windows into the input shape mask for window. i.e. the union of all input masks for window and its children will become the new input mask for window. See gdkWindowInputShapeCombineMask.

Usage

gdkWindowMergeChildInputShapes(object)

Arguments

object
a GdkWindow

Details

This function is distinct from gdkWindowSetChildInputShapes because it includes window's input shape mask in the set of shapes to be merged. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]