gdkWindowMove {RGtk2}	R Documentation

gdkWindowMove

Description

Repositions a window relative to its parent window. For toplevel windows, window managers may ignore or modify the move; you should probably use gtkWindowMove on a GtkWindow widget anyway, instead of using GDK functions. For child windows, the move will reliably succeed.

Usage

gdkWindowMove(object, x, y)

Arguments

object
a GdkWindow

x
X coordinate relative to window's parent

y
Y coordinate relative to window's parent

Details

If you're also planning to resize the window, use gdkWindowMoveResize to both move and resize simultaneously, for a nicer visual effect.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]