gdkWindowMoveRegion {RGtk2}	R Documentation

gdkWindowMoveRegion

Description

Move the part of window indicated by region by dy pixels in the Y direction and dx pixels in the X direction. The portions of region that not covered by the new position of region are invalidated.

Usage

gdkWindowMoveRegion(object, region, x, y)

Arguments

object
a GdkWindow

region
The GdkRegion to move

x
Amount to move in the X direction

y
Amount to move in the Y direction

Details

Child windows are not moved. Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]