gdkWindowMoveResize {RGtk2}	R Documentation

gdkWindowMoveResize

Description

Equivalent to calling gdkWindowMove and gdkWindowResize, except that both operations are performed at once, avoiding strange visual effects. (i.e. the user may be able to see the window first move, then resize, if you don't use gdkWindowMoveResize.)

Usage

gdkWindowMoveResize(object, x, y, width, height)

Arguments

object
a GdkWindow

x
new X position relative to window's parent

y
new Y position relative to window's parent

width
new width

height
new height

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]