gdkWindowNew {RGtk2}	R Documentation

gdkWindowNew

Description

Creates a new GdkWindow using the attributes from attributes. See GdkWindowAttr and GdkWindowAttributesType for more details. Note: to use this on displays other than the default display, parent must be specified.

Usage

gdkWindowNew(parent = NULL, attributes)

Arguments

parent
a GdkWindow, or NULL to create the window as a child of the default root window for the default display. [ allow-none ]

attributes
attributes of the new window

Value

[GdkWindow] the new GdkWindow. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]