gdkWindowProcessUpdates {RGtk2}	R Documentation

gdkWindowProcessUpdates

Description

Sends one or more expose events to window. The areas in each expose event will cover the entire update area for the window (see gdkWindowInvalidateRegion for details). Normally GDK calls gdkWindowProcessAllUpdates on your behalf, so there's no need to call this function unless you want to force expose events to be delivered immediately and synchronously (vs. the usual case, where GDK delivers them in an idle handler). Occasionally this is useful to produce nicer scrolling behavior, for example.

Usage

gdkWindowProcessUpdates(object, update.children)

Arguments

object
a GdkWindow

update.children
whether to also process updates for child windows

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]