gdkWindowRedirectToDrawable {RGtk2}	R Documentation

gdkWindowRedirectToDrawable

Description

Redirects drawing into window so that drawing to the window in the rectangle specified by src.x, src.y, width and height is also drawn into drawable at dest.x, dest.y.

Usage

gdkWindowRedirectToDrawable(object, drawable, src.x, src.y, dest.x, 
    dest.y, width, height)

Arguments

object
a GdkWindow

drawable
a GdkDrawable

src.x
x position in window

src.y
y position in window

dest.x
x position in drawable

dest.y
y position in drawable

width
width of redirection, or -1 to use the width of window

height
height of redirection or -1 to use the height of window

Details

Only drawing between gdkWindowBeginPaintRegion or gdkWindowBeginPaintRect and gdkWindowEndPaint is redirected.

Redirection is active until gdkWindowRemoveRedirection is called. Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]