gdkWindowResize {RGtk2}	R Documentation

gdkWindowResize

Description

Resizes window; for toplevel windows, asks the window manager to resize the window. The window manager may not allow the resize. When using GTK+, use gtkWindowResize instead of this low-level GDK function.

Usage

gdkWindowResize(object, width, height)

Arguments

object
a GdkWindow

width
new width of the window

height
new height of the window

Details

Windows may not be resized below 1x1.

If you're also planning to move the window, use gdkWindowMoveResize to both move and resize simultaneously, for a nicer visual effect.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]