gdkWindowRestack {RGtk2}	R Documentation

gdkWindowRestack

Description

Changes the position of window in the Z-order (stacking order), so that it is above sibling (if above is TRUE) or below sibling (if above is FALSE).

Usage

gdkWindowRestack(object, sibling, above)

Arguments

object
a GdkWindow

sibling
a GdkWindow that is a sibling of window, or NULL. [ allow-none ]

above
a boolean

Details

If sibling is NULL, then this either raises (if above is TRUE) or lowers the window.

If window is a toplevel, the window manager may choose to deny the request to move the window in the Z-order, gdkWindowRestack only requests the restack, does not guarantee it. Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]