gdkWindowSetAcceptFocus {RGtk2}	R Documentation

gdkWindowSetAcceptFocus

Description

Setting accept.focus to FALSE hints the desktop environment that the window doesn't want to receive input focus.

Usage

gdkWindowSetAcceptFocus(object, accept.focus)

Arguments

object
a toplevel GdkWindow

accept.focus
TRUE if the window should receive input focus

Details

On X, it is the responsibility of the window manager to interpret this hint. ICCCM-compliant window manager usually respect it. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]