gdkWindowSetBackPixmap {RGtk2}	R Documentation

gdkWindowSetBackPixmap

Description

Sets the background pixmap of window. May also be used to set a background of "None" on window, by setting a background pixmap of NULL.

Usage

gdkWindowSetBackPixmap(object, pixmap = NULL, parent.relative)

Arguments

object
a GdkWindow

pixmap
a GdkPixmap, or NULL. [ allow-none ]

parent.relative
whether the tiling origin is at the origin of window's parent

Details

A background pixmap will be tiled, positioning the first tile at the origin of window, or if parent.relative is TRUE, the tiling will be done based on the origin of the parent window (useful to align tiles in a parent with tiles in a child).

A background pixmap of NULL means that the window will have no background. A window with no background will never have its background filled by the windowing system, instead the window will contain whatever pixels were already in the corresponding area of the display.

The windowing system will normally fill a window with its background when the window is obscured then exposed, and when you call gdkWindowClear.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]