gdkWindowSetChildShapes {RGtk2}	R Documentation

gdkWindowSetChildShapes

Description

Sets the shape mask of window to the union of shape masks for all children of window, ignoring the shape mask of window itself. Contrast with gdkWindowMergeChildShapes which includes the shape mask of window in the masks to be merged.

Usage

gdkWindowSetChildShapes(object)

Arguments

object
a GdkWindow

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]