gdkWindowSetDebugUpdates {RGtk2}	R Documentation

gdkWindowSetDebugUpdates

Description

With update debugging enabled, calls to gdkWindowInvalidateRegion clear the invalidated region of the screen to a noticeable color, and GDK pauses for a short time before sending exposes to windows during gdkWindowProcessUpdates. The net effect is that you can see the invalid region for each window and watch redraws as they occur. This allows you to diagnose inefficiencies in your application.

Usage

gdkWindowSetDebugUpdates(setting)

Arguments

setting
TRUE to turn on update debugging

Details

In essence, because the GDK rendering model prevents all flicker, if you are redrawing the same region 400 times you may never notice, aside from noticing a speed problem. Enabling update debugging causes GTK to flicker slowly and noticeably, so you can see exactly what's being redrawn when, in what order.

The –gtk-debug=updates command line option passed to GTK+ programs enables this debug option at application startup time. That's usually more useful than calling gdkWindowSetDebugUpdates yourself, though you might want to use this function to enable updates sometime after application startup time.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]