gdkWindowSetDecorations {RGtk2}	R Documentation

gdkWindowSetDecorations

Description

"Decorations" are the features the window manager adds to a toplevel GdkWindow. This function sets the traditional Motif window manager hints that tell the window manager which decorations you would like your window to have. Usually you should use gtkWindowSetDecorated on a GtkWindow instead of using the GDK function directly.

Usage

gdkWindowSetDecorations(object, decorations)

Arguments

object
a toplevel GdkWindow

decorations
decoration hint mask

Details

The decorations argument is the logical OR of the fields in the GdkWMDecoration enumeration. If GDK_DECOR_ALL is included in the mask, the other bits indicate which decorations should be turned off. If GDK_DECOR_ALL is not included, then the other bits indicate which decorations should be turned on.

Most window managers honor a decorations hint of 0 to disable all decorations, but very few honor all possible combinations of bits.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]