gdkWindowSetEvents {RGtk2}	R Documentation

gdkWindowSetEvents

Description

The event mask for a window determines which events will be reported for that window. For example, an event mask including GDK_BUTTON_PRESS_MASK means the window should report button press events. The event mask is the bitwise OR of values from the GdkEventMask enumeration.

Usage

gdkWindowSetEvents(object, event.mask)

Arguments

object
a GdkWindow

event.mask
event mask for window

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]