gdkWindowSetFocusOnMap {RGtk2}	R Documentation

gdkWindowSetFocusOnMap

Description

Setting focus.on.map to FALSE hints the desktop environment that the window doesn't want to receive input focus when it is mapped. focus_on_map should be turned off for windows that aren't triggered interactively (such as popups from network activity).

Usage

gdkWindowSetFocusOnMap(object, focus.on.map)

Arguments

object
a toplevel GdkWindow

focus.on.map
TRUE if the window should receive input focus when mapped

Details

On X, it is the responsibility of the window manager to interpret this hint. Window managers following the freedesktop.org window manager extension specification should respect it. Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]