gdkWindowSetFunctions {RGtk2}	R Documentation

gdkWindowSetFunctions

Description

Sets hints about the window management functions to make available via buttons on the window frame.

Usage

gdkWindowSetFunctions(object, functions)

Arguments

object
a toplevel GdkWindow

functions
bitmask of operations to allow on window

Details

On the X backend, this function sets the traditional Motif window manager hint for this purpose. However, few window managers do anything reliable or interesting with this hint. Many ignore it entirely.

The functions argument is the logical OR of values from the GdkWMFunction enumeration. If the bitmask includes GDK_FUNC_ALL, then the other bits indicate which functions to disable; if it doesn't include GDK_FUNC_ALL, it indicates which functions to enable.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]