gdkWindowSetGeometryHints {RGtk2}	R Documentation

gdkWindowSetGeometryHints

Description

Sets the geometry hints for window. Hints flagged in geom.mask are set, hints not flagged in geom.mask are unset. To unset all hints, use a geom.mask of 0 and a geometry of NULL.

Usage

gdkWindowSetGeometryHints(object, geometry)

Arguments

object
a toplevel GdkWindow

geometry
geometry hints

Details

This function provides hints to the windowing system about acceptable sizes for a toplevel window. The purpose of this is to constrain user resizing, but the windowing system will typically (but is not required to) also constrain the current size of the window to the provided values and constrain programatic resizing via gdkWindowResize or gdkWindowMoveResize.

Note that on X11, this effect has no effect on windows of type GDK_WINDOW_TEMP or windows where override redirect has been turned on via gdkWindowSetOverrideRedirect since these windows are not resizable by the user.

Since you can't count on the windowing system doing the constraints for programmatic resizes, you should generally call gdkWindowConstrainSize yourself to determine appropriate sizes.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]