gdkWindowSetGroup {RGtk2}	R Documentation

gdkWindowSetGroup

Description

Sets the group leader window for window. By default, GDK sets the group leader for all toplevel windows to a global window implicitly created by GDK. With this function you can override this default.

Usage

gdkWindowSetGroup(object, leader)

Arguments

object
a toplevel GdkWindow

leader
group leader window, or NULL to restore the default group leader window

Details

The group leader window allows the window manager to distinguish all windows that belong to a single application. It may for example allow users to minimize/unminimize all windows belonging to an application at once. You should only set a non-default group window if your application pretends to be multiple applications.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]