gdkWindowSetHints {RGtk2}	R Documentation

gdkWindowSetHints

Description

This function is broken and useless and you should ignore it. If using GTK+, use functions such as gtkWindowResize, gtkWindowSetSizeRequest(), gtkWindowMove, gtkWindowParseGeometry(), and gtkWindowSetGeometryHints, depending on what you're trying to do. WARNING: gdk_window_set_hints is deprecated and should not be used in newly-written code.

Usage

gdkWindowSetHints(object, x, y, min.width, min.height, max.width, 
    max.height, flags)

Arguments

object
a GdkWindow

x
ignored field, does not matter

y
ignored field, does not matter

min.width
minimum width hint

min.height
minimum height hint

max.width
max width hint

max.height
max height hint

flags
logical OR of GDK_HINT_POS, GDK_HINT_MIN_SIZE, and/or GDK_HINT_MAX_SIZE

Details

If using GDK directly, use gdkWindowSetGeometryHints.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]