gdkWindowSetIcon {RGtk2}	R Documentation

gdkWindowSetIcon

Description

Sets the icon of window as a pixmap or window. If using GTK+, investigate gtkWindowSetDefaultIconList first, and then gtkWindowSetIconList and gtkWindowSetIcon. If those don't meet your needs, look at gdkWindowSetIconList. Only if all those are too high-level do you want to fall back to gdkWindowSetIcon.

Usage

gdkWindowSetIcon(object, icon.window, pixmap, mask)

Arguments

object
a toplevel GdkWindow

icon.window
a GdkWindow to use for the icon, or NULL to unset

pixmap
a GdkPixmap to use as the icon, or NULL to unset

mask
a 1-bit pixmap (GdkBitmap) to use as mask for pixmap, or NULL to have none

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]