gdkWindowSetIconList {RGtk2}	R Documentation

gdkWindowSetIconList

Description

Sets a list of icons for the window. One of these will be used to represent the window when it has been iconified. The icon is usually shown in an icon box or some sort of task bar. Which icon size is shown depends on the window manager. The window manager can scale the icon but setting several size icons can give better image quality since the window manager may only need to scale the icon by a small amount or not at all.

Usage

gdkWindowSetIconList(object, pixbufs)

Arguments

object
The GdkWindow toplevel window to set the icon of.

pixbufs
A list of pixbufs, of different sizes.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]