gdkWindowSetIconName {RGtk2}	R Documentation

gdkWindowSetIconName

Description

Windows may have a name used while minimized, distinct from the name they display in their titlebar. Most of the time this is a bad idea from a user interface standpoint. But you can set such a name with this function, if you like.

Usage

gdkWindowSetIconName(object, name)

Arguments

object
a toplevel GdkWindow

name
name of window while iconified (minimized)

Details

After calling this with a non-NULL name, calls to gdkWindowSetTitle will not update the icon title.

Using NULL for name unsets the icon title; further calls to gdkWindowSetTitle will again update the icon title as well.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]