gdkWindowSetKeepAbove {RGtk2}	R Documentation

gdkWindowSetKeepAbove

Description

Set if window must be kept above other windows. If the window was already above, then this function does nothing.

Usage

gdkWindowSetKeepAbove(object, setting)

Arguments

object
a toplevel GdkWindow

setting
whether to keep window above other windows

Details

On X11, asks the window manager to keep window above, if the window manager supports this operation. Not all window managers support this, and some deliberately ignore it or don't have a concept of "keep above"; so you can't rely on the window being kept above. But it will happen with most standard window managers, and GDK makes a best effort to get it to happen. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]