gdkWindowSetOpacity {RGtk2}	R Documentation

gdkWindowSetOpacity

Description

Request the windowing system to make window partially transparent, with opacity 0 being fully transparent and 1 fully opaque. (Values of the opacity parameter are clamped to the [0,1] range.)

Usage

gdkWindowSetOpacity(object, opacity)

Arguments

object
a top-level GdkWindow

opacity
opacity

Details

On X11, this works only on X screens with a compositing manager running.

For setting up per-pixel alpha, see gdkScreenGetRgbaColormap. For making non-toplevel windows translucent, see gdkWindowSetComposited. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]