gdkWindowSetOverrideRedirect {RGtk2}	R Documentation

gdkWindowSetOverrideRedirect

Description

An override redirect window is not under the control of the window manager. This means it won't have a titlebar, won't be minimizable, etc. - it will be entirely under the control of the application. The window manager can't see the override redirect window at all.

Usage

gdkWindowSetOverrideRedirect(object, override.redirect)

Arguments

object
a toplevel GdkWindow

override.redirect
TRUE if window should be override redirect

Details

Override redirect should only be used for short-lived temporary windows, such as popup menus. GtkMenu uses an override redirect window in its implementation, for example.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]