gdkWindowSetRole {RGtk2}	R Documentation

gdkWindowSetRole

Description

When using GTK+, typically you should use gtkWindowSetRole instead of this low-level function.

Usage

gdkWindowSetRole(object, role)

Arguments

object
a toplevel GdkWindow

role
a string indicating its role

Details

The window manager and session manager use a window's role to distinguish it from other kinds of window in the same application. When an application is restarted after being saved in a previous session, all windows with the same title and role are treated as interchangeable. So if you have two windows with the same title that should be distinguished for session management purposes, you should set the role on those windows. It doesn't matter what string you use for the role, as long as you have a different role for each non-interchangeable kind of window.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]