gdkWindowSetSkipPagerHint {RGtk2}	R Documentation

gdkWindowSetSkipPagerHint

Description

Toggles whether a window should appear in a pager (workspace switcher, or other desktop utility program that displays a small thumbnail representation of the windows on the desktop). If a window's semantic type as specified with gdkWindowSetTypeHint already fully describes the window, this function should not be called in addition, instead you should allow the window to be treated according to standard policy for its semantic type.

Usage

gdkWindowSetSkipPagerHint(object, modal)

Arguments

object
a toplevel GdkWindow

modal
TRUE to skip the pager

Details

Since 2.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]