gdkWindowSetStaticGravities {RGtk2}	R Documentation

gdkWindowSetStaticGravities

Description

Set the bit gravity of the given window to static, and flag it so all children get static subwindow gravity. This is used if you are implementing scary features that involve deep knowledge of the windowing system. Don't worry about it unless you have to.

Usage

gdkWindowSetStaticGravities(object, use.static)

Arguments

object
a GdkWindow

use.static
TRUE to turn on static gravity

Value

[logical] TRUE if the server supports static gravity

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]