gdkWindowSetTransientFor {RGtk2}	R Documentation

gdkWindowSetTransientFor

Description

Indicates to the window manager that window is a transient dialog associated with the application window parent. This allows the window manager to do things like center window on parent and keep window above parent.

Usage

gdkWindowSetTransientFor(object, leader)

Arguments

object
a toplevel GdkWindow

leader
another toplevel GdkWindow

Details

See gtkWindowSetTransientFor if you're using GtkWindow or GtkDialog.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]