gdkWindowSetTypeHint {RGtk2}	R Documentation

gdkWindowSetTypeHint

Description

The application can use this call to provide a hint to the window manager about the functionality of a window. The window manager can use this information when determining the decoration and behaviour of the window.

Usage

gdkWindowSetTypeHint(object, hint)

Arguments

object
A toplevel GdkWindow

hint
A hint of the function this window will have

Details

The hint must be set before the window is mapped.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]