gdkWindowSetUserData {RGtk2}	R Documentation

gdkWindowSetUserData

Description

For most purposes this function is deprecated in favor of gObjectSetData. However, for historical reasons GTK+ stores the GtkWidget that owns a GdkWindow as user data on the GdkWindow. So, custom widget implementations should use this function for that. If GTK+ receives an event for a GdkWindow, and the user data for the window is non-NULL, GTK+ will assume the user data is a GtkWidget, and forward the event to that widget.

Usage

gdkWindowSetUserData(object, user.data = NULL)

Arguments

object
a GdkWindow

user.data
user data

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]