gdkWindowShapeCombineMask {RGtk2}	R Documentation

gdkWindowShapeCombineMask

Description

Applies a shape mask to window. Pixels in window corresponding to set bits in the mask will be visible; pixels in window corresponding to unset bits in the mask will be transparent. This gives a non-rectangular window.

Usage

gdkWindowShapeCombineMask(object, shape.mask = NULL, offset.x, 
    offset.y)

Arguments

object
a GdkWindow

shape.mask
shape mask

offset.x
X position of shape mask with respect to window

offset.y
Y position of shape mask with respect to window

Details

If mask is NULL, the shape mask will be unset, and the x/y parameters are not used.

On the X11 platform, this uses an X server extension which is widely available on most common platforms, but not available on very old X servers, and occasionally the implementation will be buggy. On servers without the shape extension, this function will do nothing.

This function works on both toplevel and child windows.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]