gdkWindowShapeCombineRegion {RGtk2}	R Documentation

gdkWindowShapeCombineRegion

Description

Makes pixels in window outside shape.region be transparent, so that the window may be nonrectangular. See also gdkWindowShapeCombineMask to use a bitmap as the mask.

Usage

gdkWindowShapeCombineRegion(object, shape.region = NULL, offset.x, 
    offset.y)

Arguments

object
a GdkWindow

shape.region
region of window to be non-transparent

offset.x
X position of shape.region in window coordinates

offset.y
Y position of shape.region in window coordinates

Details

If shape.region is NULL, the shape will be unset, so the whole window will be opaque again. offset.x and offset.y are ignored if shape.region is NULL.

On the X11 platform, this uses an X server extension which is widely available on most common platforms, but not available on very old X servers, and occasionally the implementation will be buggy. On servers without the shape extension, this function will do nothing.

This function works on both toplevel and child windows.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]