gdkWindowShow {RGtk2}	R Documentation

gdkWindowShow

Description

Like gdkWindowShowUnraised, but also raises the window to the top of the window stack (moves the window to the front of the Z-order).

Usage

gdkWindowShow(object)

Arguments

object
a GdkWindow

Details

This function maps a window so it's visible onscreen. Its opposite is gdkWindowHide.

When implementing a GtkWidget, you should call this function on the widget's GdkWindow as part of the "map" method.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]