gdkWindowShowUnraised {RGtk2}	R Documentation

gdkWindowShowUnraised

Description

Shows a GdkWindow onscreen, but does not modify its stacking order. In contrast, gdkWindowShow will raise the window to the top of the window stack.

Usage

gdkWindowShowUnraised(object)

Arguments

object
a GdkWindow

Details

On the X11 platform, in Xlib terms, this function calls xmapwindow() (it also updates some internal GDK state, which means that you can't really use xmapwindow() directly on a GDK window).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]