gdkWindowStick {RGtk2}	R Documentation

gdkWindowStick

Description

"Pins" a window such that it's on all workspaces and does not scroll with viewports, for window managers that have scrollable viewports. (When using GtkWindow, gtkWindowStick may be more useful.)

Usage

gdkWindowStick(object)

Arguments

object
a toplevel GdkWindow

Details

On the X11 platform, this function depends on window manager support, so may have no effect with many window managers. However, GDK will do the best it can to convince the window manager to stick the window. For window managers that don't support this operation, there's nothing you can do to force it to happen.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]