gdkWindowUnmaximize {RGtk2}	R Documentation

gdkWindowUnmaximize

Description

Unmaximizes the window. If the window wasn't maximized, then this function does nothing.

Usage

gdkWindowUnmaximize(object)

Arguments

object
a toplevel GdkWindow

Details

On X11, asks the window manager to unmaximize window, if the window manager supports this operation. Not all window managers support this, and some deliberately ignore it or don't have a concept of "maximized"; so you can't rely on the unmaximization actually happening. But it will happen with most standard window managers, and GDK makes a best effort to get it to happen.

On Windows, reliably unmaximizes the window.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]