gtkObjectGetSignals {RGtk2}	R Documentation

Retrieve information about available signals

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

This provides dynamic (run-time) reflectance information about the available signals for a Gtk type/class. It gives the names of the signals, the parameters for the associated callback function and when it is called (i.e. before or after the object calls its own handlers).

Usage

gtkObjectGetSignals(obj)
gtkTypeGetSignals(type)

Arguments

obj
the Gtk object whose signals are to be queried.

type
a string or object of class GtkType whose signals are to be retrieved.

Author(s)

Duncan Temple Lang <duncan@research.bell-labs.com>

References

Information on the package is available from http://www.omegahat.org/RGtk.

Information on Gtk is available from http://www.gtk.org.

See Also

gtkAddCallback

[Package RGtk2 version 2.20.31 Index]