getSignalInfo {RGtk2}	R Documentation

Information for all signals for multiple classes

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

This retrieves the signal information for all of the signals defined for each of the classes specified by classes. This is just a convenient front-end to gtkSignalGetInfo for a large number of signals.

Usage

getSignalInfo(classes=.GtkClasses, load=TRUE)

Arguments

classes
a character vector containing the names of different Gtk classes/types

load
This is now ignored. You don't need to worry about whether types are loaded.

Value

A list with elements for each of the classes. Each element is itself a list with an element for each of the signals defined for that class. Each element within this list is the value returned from gtkSignalGetInfo

Author(s)

Duncan Temple Lang <duncan@research.bell-labs.com>

References

Information on the package is available from http://www.omegahat.org/RGtk.

Information on Gtk is available from http://www.gtk.org.

See Also

gtkSignalGetInfo

Examples

  if (gtkInit())
    getSignalInfo("GtkButton")

[Package RGtk2 version 2.20.31 Index]