gtkGetType {RGtk2}	R Documentation

Get the GtkType object from a Gtk object or name

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

This retrieves the GtkType object for a given GtkObject or directly by name. This can be used to dynamically discover information about a Gtk class such as its signals, properties, inheritance hierarchy, etc.

Usage

gtkGetType(name)
gtkObjectGetType(w, check=TRUE)

Arguments

name
a character vector giving the name of the Gtk class/type to be retrieved.

w
a GtkObject whose type is to be queried and returned.

check
a logical value indicating whether we should first verify that the w object inherits from GtkObject. Typically one passes TRUE for this. It is useful when one gets an incomplete object constructed directly in C code and we want to construct its class information (gtkObjectGetClasses) manually.

Details

If one calls gtkGetType with a string giving the name of the Gtk class, the associated Gtk class must have been initialized first. Specifically an object of that class must have been created earlier or the associated gtk_<class>_get_type called. This can be done using the .C function as in .C("gtk_button_get_type", PACKAGE= "RGtk").

Value

A single real value giving the internal C-level identifier for the Gtk type value. The name of the type is used as the names vector for the numeric vector.

Note

THIS STUFF IS VERY OLD AND DEPRECATED (compatibility wrappers for RGtk 1)

Author(s)

Duncan Temple Lang <duncan@research.bell-labs.com>

See Also

gtkObjectGetClasses gtkTypeGetClasses

[Package RGtk2 version 2.20.31 Index]