gio-GContentType {RGtk2}	R Documentation

GContentType

Description

Platform-specific content typing

Methods and Functions

gContentTypeEquals(type1, type2)
gContentTypeIsA(type, supertype)
gContentTypeIsUnknown(type)
gContentTypeGetDescription(type)
gContentTypeGetMimeType(type)
gContentTypeGetIcon(type)
gContentTypeCanBeExecutable(type)
gContentTypeFromMimeType(mime.type)
gContentTypeGuess(filename, data)
gContentTypeGuessForTree(root)
gContentTypesGetRegistered()

Detailed Description

A content type is a platform specific string that defines the type of a file. On unix it is a mime type, on win32 it is an extension string like ".doc", ".txt" or a percieved string like "audio". Such strings can be looked up in the registry at HKEY_CLASSES_ROOT.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/gio-GContentType.html

[Package RGtk2 version 2.20.31 Index]