gio-GIOError {RGtk2}	R Documentation

GIOError

Description

Error helper functions

Methods and Functions

gIoErrorFromErrno(err.no)

Hierarchy

GEnum
   +----GIOErrorEnum

Detailed Description

Contains helper functions for reporting errors to the user.

Enums and Flags

GIOErrorEnum

Error codes returned by GIO functions.

failed

Generic error condition for when any operation fails.

not-found

File not found error.

exists

File already exists error.

is-directory

File is a directory error.

not-directory

File is not a directory.

not-empty

File is a directory that isn't empty.

not-regular-file

File is not a regular file.

not-symbolic-link

File is not a symbolic link.

not-mountable-file

File cannot be mounted.

filename-too-long

Filename is too many characters.

invalid-filename

Filename is invalid or contains invalid characters.

too-many-links

File contains too many symbolic links.

no-space

No space left on drive.

invalid-argument

Invalid argument.

permission-denied

Permission denied.

not-supported

Operation not supported for the current backend.

not-mounted

File isn't mounted.

already-mounted

File is already mounted.

closed

File was closed.

cancelled

Operation was cancelled. See GCancellable.

pending

Operations are still pending.

read-only

File is read only.

cant-create-backup

Backup couldn't be created.

wrong-etag

File's Entity Tag was incorrect.

timed-out

Operation timed out.

would-recurse

Operation would be recursive.

busy

File is busy.

would-block

Operation would block.

host-not-found

Host couldn't be found (remote operations).

would-merge

Operation would merge files.

failed-handled

Operation failed and a helper program has already interacted with the user. Do not display any error dialog.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/gio-GIOError.html

[Package RGtk2 version 2.20.31 Index]