gio-Unix-Mounts {RGtk2}	R Documentation

Unix Mounts

Description

Unix Mounts

Hierarchy

GObject
   +----GUnixMountMonitor

Detailed Description

Routines for managing mounted UNIX mount points and paths.

Note that ‘<gio/gunixmounts.h>’ belongs to the UNIX-specific GIO interfaces, thus you have to use the ‘gio-unix-2.0.pc’ pkg-config file when using it.

Signals

mountpoints-changed(monitor, user.data)

Emitted when the unix mount points have changed.

monitor

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

mounts-changed(monitor, user.data)

Emitted when the unix mounts have changed.

monitor

the object on which the signal is emitted

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gio/gio-Unix-Mounts.html

[Package RGtk2 version 2.20.31 Index]