gtk-Accelerator-Maps {RGtk2}	R Documentation

Accelerator Maps

Description

Loadable keyboard accelerator specifications

Methods and Functions

gtkAccelMapAddEntry(accel.path, accel.key, accel.mods)
gtkAccelMapLookupEntry(accel.path)
gtkAccelMapChangeEntry(accel.path, accel.key, accel.mods, replace)
gtkAccelMapForeach(data = NULL, foreach.func)
gtkAccelMapLoadFd(fd)
gtkAccelMapSaveFd(fd)
gtkAccelMapLoadScanner(scanner)
gtkAccelMapAddFilter(filter.pattern)
gtkAccelMapForeachUnfiltered(data = NULL, foreach.func)
gtkAccelMapGet()
gtkAccelMapLockPath(accel.path)
gtkAccelMapUnlockPath(accel.path)

Hierarchy

GObject
   +----GtkAccelMap

Structures

GtkAccelMap

undocumented

User Functions

GtkAccelMapForeach()

undocumented

Signals

changed(object, accel.path, accel.key, accel.mods, user.data)

Notifies of a change in the global accelerator map. The path is also used as the detail for the signal, so it is possible to connect to changed::\varaccel_path. Since 2.4

object

the global accel map object

accel.path

the path of the accelerator that changed

accel.key

the key value for the new accelerator

accel.mods

the modifier mask for the new accelerator

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/gtk-Accelerator-Maps.html

[Package RGtk2 version 2.20.31 Index]