gtk-Clipboards {RGtk2}	R Documentation

Clipboards

Description

Storing data on clipboards

Methods and Functions

gtkClipboardGet(selection = "GDK_SELECTION_CLIPBOARD")
gtkClipboardGetForDisplay(display, selection = "GDK_SELECTION_CLIPBOARD")
gtkClipboardGetDisplay(object)
gtkClipboardSetWithData(object, targets, get.func, user.data = NULL)
gtkClipboardSetWithOwner(object, targets, get.func, owner = NULL)
gtkClipboardGetOwner(object)
gtkClipboardClear(object)
gtkClipboardSetText(object, text, len = -1)
gtkClipboardSetImage(object, pixbuf)
gtkClipboardRequestContents(object, target, callback, user.data = NULL)
gtkClipboardRequestText(object, callback, user.data = NULL)
gtkClipboardRequestImage(object, callback, user.data = NULL)
gtkClipboardRequestTargets(object, callback, user.data = NULL)
gtkClipboardRequestRichText(object, buffer, callback, user.data)
gtkClipboardRequestUris(object, callback, user.data)
gtkClipboardWaitForContents(object, target)
gtkClipboardWaitForText(object)
gtkClipboardWaitForImage(object)
gtkClipboardWaitForRichText(object, buffer)
gtkClipboardWaitForUris(object)
gtkClipboardWaitIsTextAvailable(object)
gtkClipboardWaitIsImageAvailable(object)
gtkClipboardWaitIsRichTextAvailable(object, buffer)
gtkClipboardWaitIsUrisAvailable(object)
gtkClipboardWaitForTargets(object)
gtkClipboardWaitIsTargetAvailable(object, target)
gtkClipboardSetCanStore(object, targets)
gtkClipboardStore(object)
gtkClipboard(display, selection = "GDK_SELECTION_CLIPBOARD")

Hierarchy

GObject
   +----GtkClipboard

Detailed Description

The GtkClipboard object represents a clipboard of data shared between different processes or between different widgets in the same process. Each clipboard is identified by a name encoded as a GdkAtom. (Conversion to and from strings can be done with gdkAtomIntern and gdkAtomName.) The default clipboard corresponds to the "CLIPBOARD" atom; another commonly used clipboard is the "PRIMARY" clipboard, which, in X, traditionally contains the currently selected text.

To support having a number of different formats on the clipboard at the same time, the clipboard mechanism allows providing callbacks instead of the actual data. When you set the contents of the clipboard, you can either supply the data directly (via functions like gtkClipboardSetText), or you can supply a callback to be called at a later time when the data is needed (via gtkClipboardSetWithData or gtkClipboardSetWithOwner.) Providing a callback also avoids having to make copies of the data when it is not needed. gtkClipboardSetWithData and gtkClipboardSetWithOwner are quite similar; the choice between the two depends mostly on which is more convenient in a particular situation. The former is most useful when you want to have a blob of data with callbacks to convert it into the various data types that you advertise. When the clear.func you provided is called, you simply free the data blob. The latter is more useful when the contents of clipboard reflect the internal state of a GObject (As an example, for the PRIMARY clipboard, when an entry widget provides the clipboard's contents the contents are simply the text within the selected region.) If the contents change, the entry widget can call gtkClipboardSetWithOwner to update the timestamp for clipboard ownership, without having to worry about clear.func being called.

Requesting the data from the clipboard is essentially asynchronous. If the contents of the clipboard are provided within the same process, then a direct function call will be made to retrieve the data, but if they are provided by another process, then the data needs to be retrieved from the other process, which may take some time. To avoid blocking the user interface, the call to request the selection, gtkClipboardRequestContents takes a callback that will be called when the contents are received (or when the request fails.) If you don't want to deal with providing a separate callback, you can also use gtkClipboardWaitForContents. What this does is run the GLib main loop recursively waiting for the contents. This can simplify the code flow, but you still have to be aware that other callbacks in your program can be called while this recursive mainloop is running.

Along with the functions to get the clipboard contents as an arbitrary data chunk, there are also functions to retrieve it as text, gtkClipboardRequestText and gtkClipboardWaitForText. These functions take care of determining which formats are advertised by the clipboard provider, asking for the clipboard in the best available format and converting the results into the UTF-8 encoding. (The standard form for representing strings in GTK+.)

Structures

GtkClipboard

undocumented

Convenient Construction

gtkClipboard is the equivalent of gtkClipboardGetForDisplay.

User Functions

GtkClipboardReceivedFunc(clipboard, selection.data, data)

A function to be called when the results of gtkClipboardRequestContents are received, or when the request fails.

clipboard

the GtkClipboard

selection.data

a GtkSelectionData containing the data was received. If retrieving the data failed, then then length field of selection.data will be negative.

data

the user.data supplied to gtkClipboardRequestContents.

GtkClipboardTextReceivedFunc(clipboard, text, data)

A function to be called when the results of gtkClipboardRequestText are received, or when the request fails.

clipboard

the GtkClipboard

text

the text received, as a UTF-8 encoded string, or NULL if retrieving the data failed.

data

the user.data supplied to gtkClipboardRequestText.

GtkClipboardImageReceivedFunc(clipboard, pixbuf, data)

A function to be called when the results of gtkClipboardRequestImage are received, or when the request fails. Since 2.6

clipboard

the GtkClipboard

pixbuf

the received image

data

the user.data supplied to gtkClipboardRequestImage.

GtkClipboardTargetsReceivedFunc(clipboard, atoms, n.atoms, data)

A function to be called when the results of gtkClipboardRequestTargets are received, or when the request fails. Since 2.4

clipboard

the GtkClipboard

atoms

the supported targets, as list of GdkAtom, or NULL if retrieving the data failed.

n.atoms

the length of the atoms list.

data

the user.data supplied to gtkClipboardRequestTargets.

GtkClipboardRichTextReceivedFunc()

undocumented

GtkClipboardURIReceivedFunc()

undocumented

GtkClipboardGetFunc(clipboard, selection.data, info, user.data.or.owner)

A function that will be called to provide the contents of the selection. If multiple types of data were advertised, the requested type can be determined from the info parameter or by checking the target field of selection.data. If the data could successfully be converted into then it should be stored into the selection.data object by calling gtkSelectionDataSet (or related functions such as gtkSelectionDataSetText). If no data is set, the requestor will be informed that the attempt to get the data failed.

clipboard

the GtkClipboard

selection.data

a GtkSelectionData argument in which the requested data should be stored.

info

the info field corresponding to the requested target from the GtkTargetEntry list passed to gtkClipboardSetWithData or gtkClipboardSetWithOwner.

user.data.or.owner

the user.data argument passed to gtkClipboardSetWithData, or the owner argument passed to gtkClipboardSetWithOwner

GtkClipboardClearFunc(clipboard, user.data.or.owner)

A function that will be called when the contents of the clipboard are changed or cleared. Once this has called, the user.data.or.owner argument will not be used again.

clipboard

the GtkClipboard

user.data.or.owner

the user.data argument passed to gtkClipboardSetWithData, or the owner argument passed to gtkClipboardSetWithOwner

Signals

owner-change(clipboard, event, user.data)

The ::owner-change signal is emitted when GTK+ receives an event that indicates that the ownership of the selection associated with clipboard has changed. Since 2.6

clipboard

the GtkClipboard on which the signal is emitted

event

the GdkEventOwnerChange event

user.data

user data set when the signal handler was connected.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/gtk-Clipboards.html

[Package RGtk2 version 2.20.31 Index]