gtk-Filesystem-utilities {RGtk2}	R Documentation

Filesystem utilities

Description

Functions for working with GIO

Methods and Functions

gtkMountOperationNew(parent = NULL)
gtkMountOperationIsShowing(object)
gtkMountOperationSetParent(object, parent)
gtkMountOperationGetParent(object)
gtkMountOperationSetScreen(object, screen)
gtkMountOperationGetScreen(object)
gtkShowUri(screen = NULL, uri, timestamp, .errwarn = TRUE)
gtkMountOperation(parent = NULL)

Hierarchy

GObject
   +----GMountOperation
         +----GtkMountOperation

Detailed Description

The functions and objects described here make working with GTK+ and GIO more convenient. GtkMountOperation is needed when mounting volumes: It is an implementation of GMountOperation that can be used with GIO functions for mounting volumes such as gFileMountEnclosingVolume, gFileMountMountable, gVolumeMount, gMountUnmountWithOperation and others.

When necessary, GtkMountOperation shows dialogs to ask for passwords, questions or show processes blocking unmount. gtkShowUri is a convenient way to launch applications for URIs.

Another object that is worth mentioning in this context is GdkAppLaunchContext, which provides visual feedback when lauching applications.

Structures

GtkMountOperation

This should not be accessed directly. Use the accessor functions below.

Convenient Construction

gtkMountOperation is the equivalent of gtkMountOperationNew.

Properties

is-showing [logical : Read]

Are we showing a dialog. Default value: FALSE

parent [GtkWindow : * : Read / Write]

The parent window.

screen [GdkScreen : * : Read / Write]

The screen where this window will be displayed.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/gtk-Filesystem-utilities.html

[Package RGtk2 version 2.20.31 Index]