gtk-Graphics-Contexts {RGtk2}	R Documentation

Graphics Contexts

Description

A shared pool of GdkGC objects

Methods and Functions

gtkGcGet(depth, colormap, values)
gtkGcRelease(gc)

Detailed Description

These functions provide access to a shared pool of GdkGC objects. When a new GdkGC is needed, gtkGcGet is called with the required depth, colormap and GdkGCValues. If a GdkGC with the required properties already exists then that is returned. If not, a new GdkGC is created. When the GdkGC is no longer needed, gtkGcRelease should be called.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/gtk-Graphics-Contexts.html

[Package RGtk2 version 2.20.31 Index]