gtk-Orientable {RGtk2}	R Documentation

GtkOrientable

Description

An interface for flippable widgets

Methods and Functions

gtkOrientableGetOrientation(object)
gtkOrientableSetOrientation(object, orientation)

Hierarchy

GInterface
   +----GtkOrientable

Implementations

GtkOrientable is implemented by GtkBox, GtkButtonBox, GtkColorSelection, GtkCombo, GtkFileChooserButton, GtkFileChooserWidget, GtkFontSelection, GtkGammaCurve, GtkHBox, GtkHButtonBox, GtkHPaned, GtkHRuler, GtkHScale, GtkHScrollbar, GtkHSeparator, GtkInfoBar, GtkPaned, GtkRange, GtkRecentChooserWidget, GtkRuler, GtkScale, GtkScaleButton, GtkScrollbar, GtkSeparator, GtkStatusbar, GtkToolPalette, GtkToolbar, GtkVBox, GtkVButtonBox, GtkVPaned, GtkVRuler, GtkVScale, GtkVScrollbar, GtkVSeparator and GtkVolumeButton.

Detailed Description

The GtkOrientable interface is implemented by all widgets that can be oriented horizontally or vertically. Historically, such widgets have been realized as subclasses of a common base class (e.g GtkBox/GtkHBox/GtkVBox and GtkScale/GtkHScale/GtkVScale). GtkOrientable is more flexible in that it allows the orientation to be changed at runtime, allowing the widgets to 'flip'.

GtkOrientable was introduced in GTK+ 2.16.

Structures

GtkOrientable

undocumented

Properties

orientation [GtkOrientation : Read / Write]

The orientation of the orientable. Default value: GTK_ORIENTATION_HORIZONTAL Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/gtk-Orientable.html

[Package RGtk2 version 2.20.31 Index]