gtk-Selections {RGtk2}	R Documentation

Selections

Description

Functions for handling inter-process communication via selections

Methods and Functions

gtkTargetListNew(targets)
gtkTargetListAdd(object, target, flags, info)
gtkTargetListAddTable(object, targets)
gtkTargetListAddTextTargets(list, info)
gtkTargetListAddImageTargets(list, info, writable)
gtkTargetListAddUriTargets(list, info)
gtkTargetListAddRichTextTargets(list, info, deserializable, buffer)
gtkTargetListRemove(object, target)
gtkTargetListFind(object, target)
gtkTargetTableNewFromList(list)
gtkSelectionOwnerSet(object, selection, time = "GDK_CURRENT_TIME")
gtkSelectionOwnerSetForDisplay(display, widget = NULL, selection, time = "GDK_CURRENT_TIME")
gtkSelectionAddTarget(object, selection, target, info)
gtkSelectionAddTargets(object, selection, targets)
gtkSelectionClearTargets(object, selection)
gtkSelectionConvert(object, selection, target, time = "GDK_CURRENT_TIME")
gtkSelectionDataSet(object, type = object[["target"]], format = 8, data)
gtkSelectionDataSetText(object, str, len = -1)
gtkSelectionDataGetText(object)
gtkSelectionDataSetPixbuf(object, pixbuf)
gtkSelectionDataGetPixbuf(object)
gtkSelectionDataSetUris(object, uris)
gtkSelectionDataGetUris(object)
gtkSelectionDataGetTargets(object)
gtkSelectionDataTargetsIncludeImage(object, writable)
gtkSelectionDataTargetsIncludeText(object)
gtkSelectionDataTargetsIncludeUri(object)
gtkSelectionDataTargetsIncludeRichText(object, buffer)
gtkSelectionDataGetSelection(object)
gtkSelectionDataGetData(object)
gtkSelectionDataGetLength(object)
gtkSelectionDataGetDataType(object)
gtkSelectionDataGetDisplay(object)
gtkSelectionDataGetFormat(object)
gtkSelectionDataGetTarget(object)
gtkTargetsIncludeImage(targets, writable)
gtkTargetsIncludeText(targets)
gtkTargetsIncludeUri(targets)
gtkTargetsIncludeRichText(targets, buffer)
gtkSelectionRemoveAll(object)
gtkSelectionClear(object, event)
gtkSelectionClear(object, event)
gtkSelectionDataCopy(object)

Hierarchy

GBoxed
   +----GtkTargetList

Detailed Description

The selection mechanism provides the basis for different types of communication between processes. In particular, drag and drop and GtkClipboard work via selections. You will very seldom or never need to use most of the functions in this section directly; GtkClipboard provides a nicer interface to the same functionality.

Some of the datatypes defined this section are used in the GtkClipboard and drag-and-drop API's as well. The GtkTargetEntry structure and GtkTargetList objects represent lists of data types that are supported when sending or receiving data. The GtkSelectionData object is used to store a chunk of data along with the data type and other associated information.

Structures

GtkTargetEntry

A GtkTargetEntry structure represents a single type of data than can be supplied for by a widget for a selection or for supplied or received during drag-and-drop. It contains a string representing the drag type, a flags field (used only for drag and drop - see GtkTargetFlags), and an application assigned integer ID. The integer ID will later be passed as a signal parameter for signals like "selection_get". It allows the application to identify the target type without extensive string compares. GtkTargetEntry is a transparent-type.

target

[character]

flags

[numeric]

info

[integer]

GtkTargetList

A GtkTargetList structure is a reference counted list of GtkTargetPair. It is used to represent the same information as a table of GtkTargetEntry, but in an efficient form. This structure should be treated as opaque.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/gtk-Selections.html

[Package RGtk2 version 2.20.31 Index]