gtk-Standard-Enumerations {RGtk2}	R Documentation

Standard Enumerations

Description

Public enumerated types used throughout GTK+

Enums and Flags

GtkAccelFlags

undocumented

visible

undocumented

locked

undocumented

mask

undocumented

GtkAnchorType

undocumented

center

undocumented

north

undocumented

north-west

undocumented

north-east

undocumented

south

undocumented

south-west

undocumented

south-east

undocumented

west

undocumented

east

undocumented

n

undocumented

nw

undocumented

ne

undocumented

s

undocumented

sw

undocumented

se

undocumented

w

undocumented

e

undocumented

GtkArrowType

Used to indicate the direction in which a GtkArrow should point.

up

Represents an upward pointing arrow.

down

Represents a downward pointing arrow.

left

Represents a left pointing arrow.

right

Represents a right pointing arrow.

GtkAttachOptions

Denotes the expansion properties that a widget will have when it (or its parent) is resized.

expand

the widget should expand to take up any extra space in its container that has been allocated.

shrink

the widget should shrink as and when possible.

fill

the widget should fill the space allocated to it.

GtkButtonBoxStyle

Used to dictate the style that a GtkButtonBox uses to layout the buttons it contains. (See also: GtkVButtonBox and GtkHButtonBox).

default-style

Default packing.

spread

Buttons are evenly spread across the box.

edge

Buttons are placed at the edges of the box.

start

Buttons are grouped towards the start of the box, (on the left for a HBox, or the top for a VBox).

end

Buttons are grouped towards the end of the box, (on the right for a HBox, or the bottom for a VBox).

GtkCornerType

Specifies which corner a child widget should be placed in when packed into a GtkScrolledWindow. This is effectively the opposite of where the scroll bars are placed.

top-left

Place the scrollbars on the right and bottom of the widget (default behaviour).

bottom-left

Place the scrollbars on the top and right of the widget.

top-right

Place the scrollbars on the left and bottom of the widget.

bottom-right

Place the scrollbars on the top and left of the widget.

GtkCurveType

WARNING: GtkCurveType is deprecated and should not be used in newly-written code. undocumented

linear

undocumented

spline

undocumented

free

undocumented

GtkDeleteType

undocumented

chars

undocumented

word-ends

undocumented

words

undocumented

display-lines

undocumented

display-line-ends

undocumented

paragraph-ends

undocumented

paragraphs

undocumented

whitespace

undocumented

GtkDirectionType

undocumented

tab-forward

undocumented

tab-backward

undocumented

up

undocumented

down

undocumented

left

undocumented

right

undocumented

GtkExpanderStyle

Used to specify the style of the expanders drawn by a GtkTreeView.

collapsed

The style used for a collapsed subtree.

semi-collapsed

Intermediate style used during animation.

semi-expanded

Intermediate style used during animation.

expanded

The style used for an expanded subtree.

GtkIMPreeditStyle

undocumented

nothing

undocumented

callback

undocumented

none

undocumented

GtkIMStatusStyle

undocumented

nothing

undocumented

callback

undocumented

GtkJustification

Used for justifying the text inside a GtkLabel widget. (See also GtkAlignment).

left

The text is placed at the left edge of the label.

right

The text is placed at the right edge of the label.

center

The text is placed in the center of the label.

fill

The text is placed is distributed across the label.

GtkMatchType

WARNING: GtkMatchType is deprecated and should not be used in newly-written code. undocumented

all

undocumented

all-tail

undocumented

head

undocumented

tail

undocumented

exact

undocumented

last

undocumented

GtkMetricType

Used to indicate which metric is used by a GtkRuler.

pixels

Pixels.

inches

Inches.

centimeters

Centimeters.

GtkMovementStep

undocumented

logical-positions

undocumented

visual-positions

undocumented

words

undocumented

display-lines

undocumented

display-line-ends

undocumented

paragraphs

undocumented

paragraph-ends

undocumented

pages

undocumented

buffer-ends

undocumented

horizontal-pages

undocumented

GtkOrientation

Represents the orientation of widgets which can be switched between horizontal and vertical orientation on the fly, like GtkToolbar.

horizontal

The widget is in horizontal orientation.

vertical

The widget is in vertical orientation.

GtkPackType

Represents the packing location GtkBox children. (See: GtkVBox, GtkHBox, and GtkButtonBox).

start

The child is packed into the start of the box

end

The child is packed into the end of the box

GtkPathPriorityType

undocumented

lowest

undocumented

gtk

undocumented

application

undocumented

theme

undocumented

rc

undocumented

highest

undocumented

GtkPathType

undocumented

widget

undocumented

widget-class

undocumented

class

undocumented

GtkPolicyType

Determines when a scroll bar will be visible.

always

The scrollbar is always visible.

automatic

The scrollbar will appear and disappear as necessary. For example, when all of a GtkCList can not be seen.

never

The scrollbar will never appear.

GtkPositionType

Describes which edge of a widget a certain feature is positioned at, e.g. the tabs of a GtkNotebook, the handle of a GtkHandleBox or the label of a GtkScale.

left

The feature is at the left edge.

right

The feature is at the right edge.

top

The feature is at the top edge.

bottom

The feature is at the bottom edge.

GtkPreviewType

WARNING: GtkPreviewType is deprecated and should not be used in newly-written code. An enumeration which describes whether a preview contains grayscale or red-green-blue data.

color

the preview contains red-green-blue data.

grayscale

The preview contains grayscale data.

GtkReliefStyle

Indicated the relief to be drawn around a GtkButton.

normal

Draw a normal relief.

half

A half relief.

none

No relief.

GtkResizeMode

undocumented

parent

undocumented

queue

undocumented

immediate

Deprecated.

GtkScrollStep

undocumented

steps

undocumented

pages

undocumented

ends

undocumented

horizontal-steps

undocumented

horizontal-pages

undocumented

horizontal-ends

undocumented

GtkScrollType

undocumented

none

undocumented

jump

undocumented

step-backward

undocumented

step-forward

undocumented

page-backward

undocumented

page-forward

undocumented

step-up

undocumented

step-down

undocumented

page-up

undocumented

page-down

undocumented

step-left

undocumented

step-right

undocumented

page-left

undocumented

page-right

undocumented

start

undocumented

end

undocumented

GtkSelectionMode

Used to control what selections users are allowed to make.

none

No selection is possible.

single

Zero or one element may be selected.

browse

Exactly one element is selected. In some circumstances, such as initially or during a search operation, it's possible for no element to be selected with GTK_SELECTION_BROWSE. What is really enforced is that the user can't deselect a currently selected element except by selecting another element.

multiple

Any number of elements may be selected. Clicks toggle the state of an item. Any number of elements may be selected. The Ctrl key may be used to enlarge the selection, and Shift key to select between the focus and the child pointed to. Some widgets may also allow Click-drag to select a range of elements.

extended

Deprecated, behaves identical to GTK_SELECTION_MULTIPLE.

GtkShadowType

Used to change the appearance of an outline typically provided by a GtkFrame.

none

No outline.

in

The outline is bevelled inwards.

out

The outline is bevelled outwards like a button.

etched-in

The outline has a sunken 3d appearance.

etched-out

The outline has a raised 3d appearance

GtkSideType

WARNING: GtkSideType is deprecated and should not be used in newly-written code. undocumented

top

undocumented

bottom

undocumented

left

undocumented

right

undocumented

GtkStateType

This type indicates the current state of a widget; the state determines how the widget is drawn. The GtkStateType enumeration is also used to identify different colors in a GtkStyle for drawing, so states can be used for subparts of a widget as well as entire widgets.

normal

State during normal operation.

active

State of a currently active widget, such as a depressed button.

prelight

State indicating that the mouse pointer is over the widget and the widget will respond to mouse clicks.

selected

State of a selected item, such the selected row in a list.

insensitive

State indicating that the widget is unresponsive to user actions.

GtkSubmenuDirection

WARNING: GtkSubmenuDirection is deprecated and should not be used in newly-written code. Indicates the direction a sub-menu will appear.

left

A sub-menu will appear to the left of the current menu.

right

A sub-menu will appear to the right of the current menu.

GtkSubmenuPlacement

WARNING: GtkSubmenuPlacement is deprecated and should not be used in newly-written code. undocumented

top-bottom

undocumented

left-right

undocumented

GtkToolbarStyle

Used to customize the appearance of a GtkToolbar. Note that setting the toolbar style overrides the user's preferences for the default toolbar style. Note that if the button has only a label set and GTK_TOOLBAR_ICONS is used, the label will be visible, and vice versa.

icons

Buttons display only icons in the toolbar.

text

Buttons display only text labels in the toolbar.

both

Buttons display text and icons in the toolbar.

both-horiz

Buttons display icons and text alongside each other, rather than vertically stacked

GtkUpdateType

Used by GtkRange to control the policy for notifying value changes.

continuous

Notify updates whenever the value changed

discontinuous

Notify updates when the mouse button has been released

delayed

Space out updates with a small timeout

GtkVisibility

Used by GtkCList and GtkCTree to indicate whether a row is visible.

none

The row is not visible.

partial

The row is partially visible.

full

The row is fully visible.

GtkWindowPosition

Window placement can be influenced using this enumeration. Note that using GTK_WIN_POS_CENTER_ALWAYS is almost always a bad idea. It won't necessarily work well with all window managers or on all windowing systems.

none

No influence is made on placement.

center

Windows should be placed in the center of the screen.

mouse

Windows should be placed at the current mouse position.

center-always

Keep window centered as it changes size, etc.

center-on-parent

Center the window on its transient parent (see gtkWindowSetTransientFor).

GtkWindowType

A GtkWindow can be one of these types. Most things you'd consider a "window" should have type GTK_WINDOW_TOPLEVEL; windows with this type are managed by the window manager and have a frame by default (call gtkWindowSetDecorated to toggle the frame). Windows with type GTK_WINDOW_POPUP are ignored by the window manager; window manager keybindings won't work on them, the window manager won't decorate the window with a frame, many GTK+ features that rely on the window manager will not work (e.g. resize grips and maximization/minimization). GTK_WINDOW_POPUP is used to implement widgets such as GtkMenu or tooltips that you normally don't think of as windows per se. Nearly all windows should be GTK_WINDOW_TOPLEVEL. In particular, do not use GTK_WINDOW_POPUP just to turn off the window borders; use gtkWindowSetDecorated for that.

toplevel

A regular window, such as a dialog.

popup

A special window such as a tooltip.

GtkSortType

Determines the direction of a sort.

ascending

Sorting is in ascending order.

descending

Sorting is in descending order.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/gtk-Standard-Enumerations.html

[Package RGtk2 version 2.20.31 Index]