gtk-Stock-Items {RGtk2}	R Documentation

Stock Items

Description

Prebuilt common menu/toolbar items and corresponding icons

Methods and Functions

gtkStockAdd(items)
gtkStockAddStatic(items)
gtkStockItemCopy(object)
gtkStockListIds()
gtkStockLookup(stock.id)
gtkStockSetTranslateFunc(domain, func, data)

Detailed Description

Stock items represent commonly-used menu or toolbar items such as "Open" or "Exit". Each stock item is identified by a stock ID; stock IDs are just strings, but functions such as GTK_STOCK_OPEN are provided to avoid typing mistakes in the strings. Applications can register their own stock items in addition to those built-in to GTK+.

Each stock ID can be associated with a GtkStockItem, which contains the user-visible label, keyboard accelerator, and translation domain of the menu or toolbar item; and/or with an icon stored in a GtkIconFactory. See GtkIconFactory for more information on stock icons. The connection between a GtkStockItem and stock icons is purely conventional (by virtue of using the same stock ID); it's possible to register a stock item but no icon, and vice versa. Stock icons may have a RTL variant which gets used for right-to-left locales.

Structures

GtkStockItem

undocumented GtkStockItem is a transparent-type.

stockId

[character]

label

[character]

modifier

[GdkModifierType]

keyval

[numeric]

translationDomain

[character]

Note

Please see the reference for a detailed list of the stock items

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/gtk-Stock-Items.html

[Package RGtk2 version 2.20.31 Index]