gtk-gtkbuildable {RGtk2}	R Documentation

GtkBuildable

Description

Interface for objects that can be built by GtkBuilder

Methods and Functions

gtkBuildableSetName(object, name)
gtkBuildableGetName(object)
gtkBuildableAddChild(object, builder, child, type)
gtkBuildableSetBuildableProperty(object, builder, name, value)
gtkBuildableConstructChild(object, builder, name)
gtkBuildableCustomTagStart(object, builder, child, tagname, parser, data)
gtkBuildableCustomTagEnd(object, builder, child, tagname, data)
gtkBuildableCustomFinished(object, builder, child, tagname, data)
gtkBuildableParserFinished(object, builder)
gtkBuildableGetInternalChild(object, builder, childname)

Hierarchy

GInterface
   +----GtkBuildable

Implementations

GtkBuildable is implemented by GtkAboutDialog, GtkAccelLabel, GtkAction, GtkActionGroup, GtkAlignment, GtkArrow, GtkAspectFrame, GtkAssistant, GtkBin, GtkBox, GtkButton, GtkButtonBox, GtkCList, GtkCTree, GtkCalendar, GtkCellView, GtkCheckButton, GtkCheckMenuItem, GtkColorButton, GtkColorSelection, GtkColorSelectionDialog, GtkCombo, GtkComboBox, GtkComboBoxEntry, GtkContainer, GtkCurve, GtkDialog, GtkDrawingArea, GtkEntry, GtkEntryCompletion, GtkEventBox, GtkExpander, GtkFileChooserButton, GtkFileChooserDialog, GtkFileChooserWidget, GtkFileSelection, GtkFixed, GtkFontButton, GtkFontSelection, GtkFontSelectionDialog, GtkFrame, GtkGammaCurve, GtkHBox, GtkHButtonBox, GtkHPaned, GtkHRuler, GtkHSV, GtkHScale, GtkHScrollbar, GtkHSeparator, GtkHandleBox, GtkIconFactory, GtkIconView, GtkImage, GtkImageMenuItem, GtkInfoBar, GtkInputDialog, GtkInvisible, GtkItem, GtkLabel, GtkLayout, GtkLinkButton, GtkList, GtkListItem, GtkListStore, GtkMenu, GtkMenuBar, GtkMenuItem, GtkMenuShell, GtkMenuToolButton, GtkMessageDialog, GtkMisc, GtkNotebook, GtkOffscreenWindow, GtkOldEditable, GtkOptionMenu, GtkPageSetupUnixDialog, GtkPaned, GtkPixmap, GtkPlug, GtkPreview, GtkPrintUnixDialog, GtkProgress, GtkProgressBar, GtkRadioAction, GtkRadioButton, GtkRadioMenuItem, GtkRadioToolButton, GtkRange, GtkRecentAction, GtkRecentChooserDialog, GtkRecentChooserMenu, GtkRecentChooserWidget, GtkRuler, GtkScale, GtkScaleButton, GtkScrollbar, GtkScrolledWindow, GtkSeparator, GtkSeparatorMenuItem, GtkSeparatorToolItem, GtkSizeGroup, GtkSocket, GtkSpinButton, GtkSpinner, GtkStatusbar, GtkTable, GtkTearoffMenuItem, GtkText, GtkTextView, GtkTipsQuery, GtkToggleAction, GtkToggleButton, GtkToggleToolButton, GtkToolButton, GtkToolItem, GtkToolItemGroup, GtkToolPalette, GtkToolbar, GtkTree, GtkTreeItem, GtkTreeStore, GtkTreeView, GtkTreeViewColumn, GtkUIManager, GtkVBox, GtkVButtonBox, GtkVPaned, GtkVRuler, GtkVScale, GtkVScrollbar, GtkVSeparator, GtkViewport, GtkVolumeButton, GtkWidget and GtkWindow.

Detailed Description

In order to allow construction from a GtkBuilder UI description, an object class must implement the GtkBuildable interface. The interface includes methods for setting names and properties of objects, parsing custom tags, constructing child objects.

The GtkBuildable interface is implemented by all widgets and many of the non-widget objects that are provided by GTK+. The main user of this interface is GtkBuilder, there should be very little need for applications to call any gtkBuildable... functions.

Structures

GtkBuildable

undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/gtk-gtkbuildable.html

[Package RGtk2 version 2.20.31 Index]