gtk-gtkcheckmenuitem {RGtk2}	R Documentation

GtkCheckMenuItem

Description

A menu item with a check box

Methods and Functions

gtkCheckMenuItemNew(show = TRUE)
gtkCheckMenuItemNewWithLabel(label, show = TRUE)
gtkCheckMenuItemNewWithMnemonic(label, show = TRUE)
gtkCheckMenuItemGetActive(object)
gtkCheckMenuItemSetActive(object, is.active)
gtkCheckMenuItemSetShowToggle(object, always)
gtkCheckMenuItemToggled(object)
gtkCheckMenuItemGetInconsistent(object)
gtkCheckMenuItemSetInconsistent(object, setting)
gtkCheckMenuItemSetDrawAsRadio(object, draw.as.radio)
gtkCheckMenuItemGetDrawAsRadio(object)
gtkCheckMenuItem(label, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkContainer
                           +----GtkBin
                                 +----GtkItem
                                       +----GtkMenuItem
                                             +----GtkCheckMenuItem
                                                   +----GtkRadioMenuItem

Interfaces

GtkCheckMenuItem implements AtkImplementorIface, GtkBuildable and GtkActivatable.

Detailed Description

A GtkCheckMenuItem is a menu item that maintains the state of a boolean value in addition to a GtkMenuItem's usual role in activating application code.

A check box indicating the state of the boolean value is displayed at the left side of the GtkMenuItem. Activating the GtkMenuItem toggles the value.

Structures

GtkCheckMenuItem

The GtkCheckMenuItem struct contains the following fields. (These fields should be considered read-only. They should never be set by an application.)

numeric active; 	 TRUE if the check box is active.

Convenient Construction

gtkCheckMenuItem is the result of collapsing the constructors of GtkCheckMenuItem (gtkCheckMenuItemNew, gtkCheckMenuItemNewWithLabel, gtkCheckMenuItemNewWithMnemonic) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Signals

toggled(checkmenuitem, user.data)

This signal is emitted when the state of the check box is changed.

A signal handler can examine the active field of the GtkCheckMenuItem struct to discover the new state.

checkmenuitem

the object which received the signal.

user.data

user data set when the signal handler was connected.

Properties

active [logical : Read / Write]

Whether the menu item is checked. Default value: FALSE

draw-as-radio [logical : Read / Write]

Whether the menu item looks like a radio menu item. Default value: FALSE

inconsistent [logical : Read / Write]

Whether to display an "inconsistent" state. Default value: FALSE

Style Properties

indicator-size [integer : Read]

Size of check or radio indicator. Allowed values: >= 0 Default value: 13

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/gtk-gtkcheckmenuitem.html

[Package RGtk2 version 2.20.31 Index]