gtkAboutDialogGetLogoIconName {RGtk2}	R Documentation

gtkAboutDialogGetLogoIconName

Description

Returns the icon name displayed as logo in the about dialog.

Usage

gtkAboutDialogGetLogoIconName(object)

Arguments

object
a GtkAboutDialog

Details

Since 2.6

Value

[character] the icon name displayed as logo. If you want to keep a reference to it, you have to call gStrdup() on it.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]