gtkAboutDialogSetComments {RGtk2}	R Documentation

gtkAboutDialogSetComments

Description

Sets the comments string to display in the about dialog. This should be a short string of one or two lines.

Usage

gtkAboutDialogSetComments(object, comments = NULL)

Arguments

object
a GtkAboutDialog

comments
a comments string. [ allow-none ]

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]