gtkAboutDialogSetTranslatorCredits {RGtk2}	R Documentation

gtkAboutDialogSetTranslatorCredits

Description

Sets the translator credits string which is displayed in the translators tab of the secondary credits dialog.

Usage

gtkAboutDialogSetTranslatorCredits(object, translator.credits = NULL)

Arguments

object
a GtkAboutDialog

translator.credits
the translator credits. [ allow-none ]

Details

The intended use for this string is to display the translator of the language which is currently used in the user interface. Using gettext(), a simple way to achieve that is to mark the string for translation:

about$setTranslatorCredits(gettext("translator-credits"))

It is a good idea to use the customary msgid "translator-credits" for this purpose, since translators will already know the purpose of that msgid, and since GtkAboutDialog will detect if "translator-credits" is untranslated and hide the tab. Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]