gtkAccelGroupConnect {RGtk2}	R Documentation

gtkAccelGroupConnect

Description

Installs an accelerator in this group. When accel.group is being activated in response to a call to gtkAccelGroupsActivate, closure will be invoked if the accel.key and accel.mods from gtkAccelGroupsActivate match those of this connection.

Usage

gtkAccelGroupConnect(object, accel.key, accel.mods, accel.flags, 
    closure)

Arguments

object
the accelerator group to install an accelerator in

accel.key
key value of the accelerator

accel.mods
modifier combination of the accelerator

accel.flags
a flag mask to configure this accelerator

closure
closure to be executed upon accelerator activation

Details

The signature used for the closure is that of GtkAccelGroupActivate.

Note that, due to implementation details, a single closure can only be connected to one accelerator group.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]