gtkAccelGroupConnectByPath {RGtk2}	R Documentation

gtkAccelGroupConnectByPath

Description

Installs an accelerator in this group, using an accelerator path to look up the appropriate key and modifiers (see gtkAccelMapAddEntry). When accel.group is being activated in response to a call to gtkAccelGroupsActivate, closure will be invoked if the accel.key and accel.mods from gtkAccelGroupsActivate match the key and modifiers for the path.

Usage

gtkAccelGroupConnectByPath(object, accel.path, closure)

Arguments

object
the accelerator group to install an accelerator in

accel.path
path used for determining key and modifiers.

closure
closure to be executed upon accelerator activation

Details

The signature used for the closure is that of GtkAccelGroupActivate.

Note that accel.path string will be stored in a GQuark. Therefore, if you pass a static string, you can save some memory by interning it first with gInternStaticString().

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]