gtkAccelGroupDisconnect {RGtk2}	R Documentation

gtkAccelGroupDisconnect

Description

Removes an accelerator previously installed through gtkAccelGroupConnect.

Usage

gtkAccelGroupDisconnect(object, closure)

Arguments

object
the accelerator group to remove an accelerator from

closure
the closure to remove from this accelerator group, or NULL to remove all closures. [ allow-none ]

Details

Since 2.20 closure can be NULL.

Value

[logical] TRUE if the closure was found and got disconnected

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]