gtkAccelMapAddEntry {RGtk2}	R Documentation

gtkAccelMapAddEntry

Description

Registers a new accelerator with the global accelerator map. This function should only be called once per accel.path with the canonical accel.key and accel.mods for this path. To change the accelerator during runtime programatically, use gtkAccelMapChangeEntry. The accelerator path must consist of "<WINDOWTYPE>/Category1/Category2/.../Action", where <WINDOWTYPE> should be a unique application-specific identifier, that corresponds to the kind of window the accelerator is being used in, e.g. "Gimp-Image", "Abiword-Document" or "Gnumeric-Settings". The Category1/.../Action portion is most appropriately chosen by the action the accelerator triggers, i.e. for accelerators on menu items, choose the item's menu path, e.g. "File/Save As", "Image/View/Zoom" or "Edit/Select All". So a full valid accelerator path may look like: "<Gimp-Toolbox>/File/Dialogs/Tool Options...".

Usage

gtkAccelMapAddEntry(accel.path, accel.key, accel.mods)

Arguments

accel.path
valid accelerator path

accel.key
the accelerator key

accel.mods
the accelerator modifiers

Details

Note that accel.path string will be stored in a GQuark. Therefore, if you pass a static string, you can save some memory by interning it first with gInternStaticString().

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]