gtkAccelMapAddFilter {RGtk2}	R Documentation

gtkAccelMapAddFilter

Description

Adds a filter to the global list of accel path filters.

Usage

gtkAccelMapAddFilter(filter.pattern)

Arguments

filter.pattern
a pattern (see GPatternSpec)

Details

Accel map entries whose accel path matches one of the filters are skipped by gtkAccelMapForeach.

This function is intended for GTK+ modules that create their own menus, but don't want them to be saved into the applications accelerator map dump.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]