gtkAccelMapChangeEntry {RGtk2}	R Documentation

gtkAccelMapChangeEntry

Description

Changes the accel.key and accel.mods currently associated with accel.path. Due to conflicts with other accelerators, a change may not always be possible, replace indicates whether other accelerators may be deleted to resolve such conflicts. A change will only occur if all conflicts could be resolved (which might not be the case if conflicting accelerators are locked). Successful changes are indicated by a TRUE return value.

Usage

gtkAccelMapChangeEntry(accel.path, accel.key, accel.mods, replace)

Arguments

accel.path
a valid accelerator path

accel.key
the new accelerator key

accel.mods
the new accelerator modifiers

replace
TRUE if other accelerators may be deleted upon conflicts

Details

Note that accel.path string will be stored in a GQuark. Therefore, if you pass a static string, you can save some memory by interning it first with gInternStaticString().

Value

[logical] TRUE if the accelerator could be changed, FALSE otherwise

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]