gtkAccelMapForeachUnfiltered {RGtk2}	R Documentation

gtkAccelMapForeachUnfiltered

Description

Loops over all entries in the accelerator map, and execute foreach.func on each. The signature of foreach.func is that of GtkAccelMapForeach, the changed parameter indicates whether this accelerator was changed during runtime (thus, would need saving during an accelerator map dump).

Usage

gtkAccelMapForeachUnfiltered(data = NULL, foreach.func)

Arguments

data
data to be passed into foreach.func

foreach.func
function to be executed for each accel map entry

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]