gtkAccelMapLockPath {RGtk2}	R Documentation

gtkAccelMapLockPath

Description

Locks the given accelerator path. If the accelerator map doesn't yet contain an entry for accel.path, a new one is created.

Usage

gtkAccelMapLockPath(accel.path)

Arguments

accel.path
a valid accelerator path

Details

Locking an accelerator path prevents its accelerator from being changed during runtime. A locked accelerator path can be unlocked by gtkAccelMapUnlockPath. Refer to gtkAccelMapChangeEntry for information about runtime accelerator changes.

If called more than once, accel.path remains locked until gtkAccelMapUnlockPath has been called an equivalent number of times.

Note that locking of individual accelerator paths is independent from locking the GtkAccelGroup containing them. For runtime accelerator changes to be possible both the accelerator path and its GtkAccelGroup have to be unlocked. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]