gtkAccelMapSave {RGtk2}	R Documentation

gtkAccelMapSave

Description

Saves current accelerator specifications (accelerator path, key and modifiers) to file.name. The file is written in a format suitable to be read back in by gtkAccelMapLoad.

Usage

gtkAccelMapSave(file.name)

Arguments

file.name
the name of the file to contain accelerator specifications, in the GLib file name encoding

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]