gtkAcceleratorGetLabel {RGtk2}	R Documentation

gtkAcceleratorGetLabel

Description

Converts an accelerator keyval and modifier mask into a string which can be used to represent the accelerator to the user.

Usage

gtkAcceleratorGetLabel(accelerator.key, accelerator.mods)

Arguments

accelerator.key
accelerator keyval

accelerator.mods
accelerator modifier mask

Details

Since 2.6

Value

[character] a newly-allocated string representing the accelerator.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]