gtkAcceleratorName {RGtk2}	R Documentation

gtkAcceleratorName

Description

Converts an accelerator keyval and modifier mask into a string parseable by gtkAcceleratorParse. For example, if you pass in GDK_q and GDK_CONTROL_MASK, this function returns "<Control>q".

Usage

gtkAcceleratorName(accelerator.key, accelerator.mods)

Arguments

accelerator.key
accelerator keyval

accelerator.mods
accelerator modifier mask

Details

If you need to display accelerators in the user interface, see gtkAcceleratorGetLabel.

Value

[character] a newly-allocated accelerator name

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]