gtkAcceleratorParse {RGtk2}	R Documentation

gtkAcceleratorParse

Description

Parses a string representing an accelerator. The format looks like "<Control>a" or "<Shift><Alt>F1" or "<Release>z" (the last one is for key release). The parser is fairly liberal and allows lower or upper case, and also abbreviations such as "<Ctl>" and "<Ctrl>".

Usage

gtkAcceleratorParse(accelerator)

Arguments

accelerator
string representing an accelerator

Details

If the parse fails, accelerator.key and accelerator.mods will be set to 0 (zero).

Value

A list containing the following elements:

accelerator.key
return location for accelerator keyval

accelerator.mods
return location for accelerator modifier mask

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]