gtkAcceleratorSetDefaultModMask {RGtk2}	R Documentation

gtkAcceleratorSetDefaultModMask

Description

Sets the modifiers that will be considered significant for keyboard accelerators. The default mod mask is GDK_CONTROL_MASK | GDK_SHIFT_MASK | GDK_MOD1_MASK | GDK_SUPER_MASK | GDK_HYPER_MASK | GDK_META_MASK, that is, Control, Shift, Alt, Super, Hyper and Meta. Other modifiers will by default be ignored by GtkAccelGroup. You must include at least the three modifiers Control, Shift and Alt in any value you pass to this function.

Usage

gtkAcceleratorSetDefaultModMask(default.mod.mask)

Arguments

default.mod.mask
accelerator modifier mask

Details

The default mod mask should be changed on application startup, before using any accelerator groups.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]