gtkActionBlockActivateFrom {RGtk2}	R Documentation

gtkActionBlockActivateFrom

Description

Disables calls to the gtkActionActivate function by signals on the given proxy widget. This is used to break notification loops for things like check or radio actions. WARNING: gtk_action_block_activate_from has been deprecated since version 2.16 and should not be used in newly-written code. activatables are now responsible for activating the action directly so this doesnt apply anymore.

Usage

gtkActionBlockActivateFrom(object, proxy)

Arguments

object
the action object

proxy
a proxy widget

Details

This function is intended for use by action implementations. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]