gtkActionConnectAccelerator {RGtk2}	R Documentation

gtkActionConnectAccelerator

Description

Installs the accelerator for action if action has an accel path and group. See gtkActionSetAccelPath and gtkActionSetAccelGroup

Usage

gtkActionConnectAccelerator(object)

Arguments

object
a GtkAction

Details

Since multiple proxies may independently trigger the installation of the accelerator, the action counts the number of times this function has been called and doesn't remove the accelerator until gtkActionDisconnectAccelerator has been called as many times. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]