gtkActionGroupAddAction {RGtk2}	R Documentation

gtkActionGroupAddAction

Description

Adds an action object to the action group. Note that this function does not set up the accel path of the action, which can lead to problems if a user tries to modify the accelerator of a menuitem associated with the action. Therefore you must either set the accel path yourself with gtkActionSetAccelPath, or use gtk_action_group_add_action_with_accel (..., NULL).

Usage

gtkActionGroupAddAction(object, action)

Arguments

object
the action group

action
an action

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]