gtkActionGroupAddActions {RGtk2}	R Documentation

gtkActionGroupAddActions

Description

This is a convenience function to create a number of actions and add them to the action group.

Usage

gtkActionGroupAddActions(object, entries, user.data = NULL)

Arguments

object
the action group

entries
a list of action descriptions

user.data
data to pass to the action callbacks

Details

The "activate" signals of the actions are connected to the callbacks and their accel paths are set to <Actions>/ \var{group-name} / \var{action-name}. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]