gtkActionGroupAddRadioActions {RGtk2}	R Documentation

gtkActionGroupAddRadioActions

Description

This is a convenience routine to create a group of radio actions and add them to the action group.

Usage

gtkActionGroupAddRadioActions(object, entries, value, on.change = NULL, 
    user.data = NULL)

Arguments

object
the action group

entries
a list of radio action descriptions

value
the value of the action to activate initially, or -1 if no action should be activated

on.change
the callback to connect to the changed signal

user.data
data to pass to the action callbacks

Details

The "changed" signal of the first radio action is connected to the on.change callback and the accel paths of the actions are set to <Actions>/ \var{group-name} / \var{action-name}. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]