gtkActionGroupSetTranslateFunc {RGtk2}	R Documentation

gtkActionGroupSetTranslateFunc

Description

Sets a function to be used for translating the label and tooltip of GtkActionGroupEntrys added by gtkActionGroupAddActions.

Usage

gtkActionGroupSetTranslateFunc(object, func, data = NULL)

Arguments

object
a GtkActionGroup

func
a GtkTranslateFunc

data
data to be passed to func and notify

Details

If you're using gettext(), it is enough to set the translation domain with gtkActionGroupSetTranslationDomain. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]