gtkActivatableDoSetRelatedAction {RGtk2}	R Documentation

gtkActivatableDoSetRelatedAction

Description

This is a utility function for GtkActivatable implementors.

Usage

gtkActivatableDoSetRelatedAction(object, action)

Arguments

object
a GtkActivatable

action
the GtkAction to set

Details

When implementing GtkActivatable you must call this when handling changes of the "related-action", and you must also use this to break references in GObject->dispose().

This function adds a reference to the currently set related action for you, it also makes sure the GtkActivatable->update() method is called when the related GtkAction properties change and registers to the action's proxy list. PLEASE NOTE: Be careful to call this before setting the local copy of the GtkAction property, since this function uses gtkActivatableGetAction() to retrieve the previous action Since 2.16

Note

Be careful to call this before setting the local copy of the GtkAction property, since this function uses gtkActivatableGetAction() to retrieve the previous action

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]