gtkActivatableGetUseActionAppearance {RGtk2}	R Documentation

gtkActivatableGetUseActionAppearance

Description

Gets whether this activatable should reset its layout and appearance when setting the related action or when the action changes appearance.

Usage

gtkActivatableGetUseActionAppearance(object)

Arguments

object
a GtkActivatable

Details

Since 2.16

Value

[logical] whether activatable uses its actions appearance.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]