gtkActivatableSetUseActionAppearance {RGtk2}	R Documentation

gtkActivatableSetUseActionAppearance

Description

Sets whether this activatable should reset its layout and appearance when setting the related action or when the action changes appearance

Usage

gtkActivatableSetUseActionAppearance(object, use.appearance)

Arguments

object
a GtkActivatable

use.appearance
whether to use the actions appearance

Details

PLEASE NOTE: GtkActivatable implementors need to handle the "use-action-appearance" property and call gtkActivatableSyncActionProperties to update activatable if needed. Since 2.16

Note

GtkActivatable implementors need to handle the "use-action-appearance" property and call gtkActivatableSyncActionProperties to update activatable if needed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]