gtkActivatableSyncActionProperties {RGtk2}	R Documentation

gtkActivatableSyncActionProperties

Description

This is called to update the activatable completely, this is called internally when the "related-action" property is set or unset and by the implementing class when "use-action-appearance" changes.

Usage

gtkActivatableSyncActionProperties(object, action = NULL)

Arguments

object
a GtkActivatable

action
the related GtkAction or NULL. [ allow-none ]

Details

Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]