gtkAdd {RGtk2}	R Documentation

Add Gtk wdget to parent container

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

This adds the specified Gtk widget to the hierarchy defining a graphical interface by giving it a parent and a place in the parent's container. This is then followed by a call to gtkShow to make the widget visible.

Usage

gtkAdd(parent, ...)

Arguments

parent
the parent container widget (e.g. a box, etc. ) into which the widget ‘w’ is to be added.

...
one or more widgets that are to be added to the parent container.

Value

Currently, a logical value indicating whether the, typically a logical value, TRUE, indicating success. This will be changed to be a widget. Perhaps we will add an option to allow a scrolled container or a pane to be added in which case the result will be that newly created container widget.

Author(s)

Duncan Temple Lang

References

http://www.gtk.org http://www.omegahat.org/RGtk

See Also

gtkContainerAdd gtkWindow gtkShow

[Package RGtk2 version 2.20.31 Index]