gtkAdjustmentClampPage {RGtk2}	R Documentation

gtkAdjustmentClampPage

Description

Updates the GtkAdjustment value to ensure that the range between lower and upper is in the current page (i.e. between value and value + page.size). If the range is larger than the page size, then only the start of it will be in the current page. A "changed" signal will be emitted if the value is changed.

Usage

gtkAdjustmentClampPage(object, lower, upper)

Arguments

object
a GtkAdjustment.

lower
the lower value.

upper
the upper value.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]