gtkAdjustmentConfigure {RGtk2}	R Documentation

gtkAdjustmentConfigure

Description

Sets all properties of the adjustment at once.

Usage

gtkAdjustmentConfigure(object, value, lower, upper, step.increment, 
    page.increment, page.size)

Arguments

object
a GtkAdjustment

value
the new value

lower
the new minimum value

upper
the new maximum value

step.increment
the new step increment

page.increment
the new page increment

page.size
the new page size

Details

Use this function to avoid multiple emissions of the "changed" signal. See gtkAdjustmentSetLower for an alternative way of compressing multiple emissions of "changed" into one. Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]