gtkAdjustmentSetLower {RGtk2}	R Documentation

gtkAdjustmentSetLower

Description

Sets the minimum value of the adjustment.

Usage

gtkAdjustmentSetLower(object, lower)

Arguments

object
a GtkAdjustment

lower
the new minimum value

Details

When setting multiple adjustment properties via their individual setters, multiple "changed" signals will be emitted. However, since the emission of the "changed" signal is tied to the emission of the "GObject::notify" signals of the changed properties, it's possible to compress the "changed" signals into one by calling gObjectFreezeNotify() and gObjectThawNotify() around the calls to the individual setters.

Alternatively, using a single gObjectSet for all the properties to change, or using gtkAdjustmentConfigure has the same effect of compressing "changed" emissions. Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]