gtkAdjustmentSetUpper {RGtk2}	R Documentation

gtkAdjustmentSetUpper

Description

Sets the maximum value of the adjustment.

Usage

gtkAdjustmentSetUpper(object, upper)

Arguments

object
a GtkAdjustment

upper
the new maximum value

Details

Note that values will be restricted by upper - page-size if the page-size property is nonzero.

See gtkAdjustmentSetLower about how to compress multiple emissions of the "changed" signal when setting multiple adjustment properties. Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]