gtkAdjustmentSetValue {RGtk2}	R Documentation

gtkAdjustmentSetValue

Description

Sets the GtkAdjustment value. The value is clamped to lie between adjustment->lower and adjustment->upper.

Usage

gtkAdjustmentSetValue(object, value)

Arguments

object
a GtkAdjustment.

value
the new value.

Details

Note that for adjustments which are used in a GtkScrollbar, the effective range of allowed values goes from adjustment->lower to adjustment->upper - adjustment->page_size.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]