gtkAlignmentGetPadding {RGtk2}	R Documentation

gtkAlignmentGetPadding

Description

Gets the padding on the different sides of the widget. See gtkAlignmentSetPadding.

Usage

gtkAlignmentGetPadding(object)

Arguments

object
a GtkAlignment

Details

Since 2.4

Value

A list containing the following elements:

padding.top
location to store the padding for the top of the widget, or NULL. [ allow-none ]

padding.bottom
location to store the padding for the bottom of the widget, or NULL. [ allow-none ]

padding.left
location to store the padding for the left of the widget, or NULL. [ allow-none ]

padding.right
location to store the padding for the right of the widget, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]