gtkAlignmentNew {RGtk2}	R Documentation

gtkAlignmentNew

Description

Creates a new GtkAlignment.

Usage

gtkAlignmentNew(xalign = NULL, yalign = NULL, xscale = NULL, 
    yscale = NULL, show = TRUE)

Arguments

xalign
the horizontal alignment of the child widget, from 0 (left) to 1 (right).

yalign
the vertical alignment of the child widget, from 0 (top) to 1 (bottom).

xscale
the amount that the child widget expands horizontally to fill up unused space, from 0 to 1. A value of 0 indicates that the child widget should never expand. A value of 1 indicates that the child widget will expand to fill all of the space allocated for the GtkAlignment.

yscale
the amount that the child widget expands vertically to fill up unused space, from 0 to 1. The values are similar to xscale.

Value

[GtkWidget] the new GtkAlignment.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]