gtkAlignmentSetPadding {RGtk2}	R Documentation

gtkAlignmentSetPadding

Description

Sets the padding on the different sides of the widget. The padding adds blank space to the sides of the widget. For instance, this can be used to indent the child widget towards the right by adding padding on the left.

Usage

gtkAlignmentSetPadding(object, padding.top, padding.bottom, padding.left, 
    padding.right)

Arguments

object
a GtkAlignment

padding.top
the padding at the top of the widget

padding.bottom
the padding at the bottom of the widget

padding.left
the padding at the left of the widget

padding.right
the padding at the right of the widget.

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]