gtkAlternativeDialogButtonOrder {RGtk2}	R Documentation

gtkAlternativeDialogButtonOrder

Description

Returns TRUE if dialogs are expected to use an alternative button order on the screen screen. See gtkDialogSetAlternativeButtonOrder for more details about alternative button order.

Usage

gtkAlternativeDialogButtonOrder(object)

Arguments

object
a GdkScreen, or NULL to use the default screen. [ allow-none ]

Details

If you need to use this function, you should probably connect to the ::notify:gtk-alternative-button-order signal on the GtkSettings object associated to screen, in order to be notified if the button order setting changes. Since 2.6

Value

[logical] Whether the alternative button order should be used

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]