gtkAspectFrameNew {RGtk2}	R Documentation

gtkAspectFrameNew

Description

Create a new GtkAspectFrame.

Usage

gtkAspectFrameNew(label = NULL, xalign = NULL, yalign = NULL, 
    ratio = NULL, obey.child = NULL, show = TRUE)

Arguments

label
Label text.

xalign
Horizontal alignment of the child within the allocation of the GtkAspectFrame. This ranges from 0.0 (left aligned) to 1.0 (right aligned)

yalign
Vertical alignment of the child within the allocation of the GtkAspectFrame. This ranges from 0.0 (left aligned) to 1.0 (right aligned)

ratio
The desired aspect ratio.

obey.child
If TRUE, ratio is ignored, and the aspect ratio is taken from the requistion of the child.

Value

[GtkWidget] the new GtkAspectFrame.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]