gtkAssistantSetCurrentPage {RGtk2}	R Documentation

gtkAssistantSetCurrentPage

Description

Switches the page to page.num. Note that this will only be necessary in custom buttons, as the assistant flow can be set with gtkAssistantSetForwardPageFunc.

Usage

gtkAssistantSetCurrentPage(object, page.num)

Arguments

object
a GtkAssistant

page.num
index of the page to switch to, starting from 0. If negative, the last page will be used. If greater than the number of pages in the assistant, nothing will be done.

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]