gtkAssistantSetForwardPageFunc {RGtk2}	R Documentation

gtkAssistantSetForwardPageFunc

Description

Sets the page forwarding function to be page.func, this function will be used to determine what will be the next page when the user presses the forward button. Setting page.func to NULL will make the assistant to use the default forward function, which just goes to the next visible page.

Usage

gtkAssistantSetForwardPageFunc(object, page.func, data)

Arguments

object
a GtkAssistant

page.func
the GtkAssistantPageFunc, or NULL to use the default one. [ allow-none ]

data
user data for page.func

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]