gtkAssistantSetPageHeaderImage {RGtk2}	R Documentation

gtkAssistantSetPageHeaderImage

Description

Sets a header image for page. This image is displayed in the header area of the assistant when page is the current page.

Usage

gtkAssistantSetPageHeaderImage(object, page, pixbuf = NULL)

Arguments

object
a GtkAssistant

page
a page of assistant

pixbuf
the new header image page. [ allow-none ]

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]