gtkAssistantUpdateButtonsState {RGtk2}	R Documentation

gtkAssistantUpdateButtonsState

Description

Forces assistant to recompute the buttons state.

Usage

gtkAssistantUpdateButtonsState(object)

Arguments

object
a GtkAssistant

Details

GTK+ automatically takes care of this in most situations, e.g. when the user goes to a different page, or when the visibility or completeness of a page changes.

One situation where it can be necessary to call this function is when changing a value on the current page affects the future page flow of the assistant. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]