gtkBoxPackEndDefaults {RGtk2}	R Documentation

gtkBoxPackEndDefaults

Description

Adds widget to box, packed with reference to the end of box. The child is packed after any other child packed with reference to the start of box. WARNING: gtk_box_pack_end_defaults has been deprecated since version 2.14 and should not be used in newly-written code. Use gtkBoxPackEnd

Usage

gtkBoxPackEndDefaults(object, widget)

Arguments

object
a GtkBox

widget
the GtkWidget to be added to box

Details

Parameters for how to pack the child widget, "expand", "fill" and "padding", are given their default values, TRUE, TRUE, and 0, respectively.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]