gtkBoxPackStart {RGtk2}	R Documentation

gtkBoxPackStart

Description

Adds child to box, packed with reference to the start of box. The child is packed after any other child packed with reference to the start of box.

Usage

gtkBoxPackStart(object, child, expand = TRUE, fill = TRUE, padding = 0)

Arguments

object
a GtkBox

child
the GtkWidget to be added to box

expand
TRUE if the new child is to be given extra space allocated to box. The extra space will be divided evenly between all children of box that use this option

fill
TRUE if space given to child by the expand option is actually allocated to child, rather than just padding it. This parameter has no effect if expand is set to FALSE. A child is always allocated the full height of a GtkHBox and the full width of a GtkVBox. This option affects the other dimension

padding
extra space in pixels to put between this child and its neighbors, over and above the global amount specified by "spacing" property. If child is a widget at one of the reference ends of box, then padding pixels are also put between child and the reference edge of box

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]