gtkBoxReorderChild {RGtk2}	R Documentation

gtkBoxReorderChild

Description

Moves child to a new position in the list of box children. The list is the children field of GtkBox, and contains both widgets packed GTK_PACK_START as well as widgets packed GTK_PACK_END, in the order that these widgets were added to box.

Usage

gtkBoxReorderChild(object, child, position)

Arguments

object
a GtkBox

child
the GtkWidget to move

position
the new position for child in the list of children of box, starting from 0. If negative, indicates the end of the list

Details

A widget's position in the box children list determines where the widget is packed into box. A child widget at some position in the list will be packed just after all other widgets of the same packing type that appear earlier in the list.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]