gtkBuildableCustomTagEnd {RGtk2}	R Documentation

gtkBuildableCustomTagEnd

Description

This is called at the end of each custom element handled by the buildable.

Usage

gtkBuildableCustomTagEnd(object, builder, child, tagname, data)

Arguments

object
A GtkBuildable

builder
GtkBuilder used to construct this object

child
child object or NULL for non-child tags. [ allow-none ]

tagname
name of tag

data
user data that will be passed in to parser functions

Details

Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]