gtkBuildableCustomTagStart {RGtk2}	R Documentation

gtkBuildableCustomTagStart

Description

This is called for each unknown element under <child>.

Usage

gtkBuildableCustomTagStart(object, builder, child, tagname, parser, 
    data)

Arguments

object
a GtkBuildable

builder
a GtkBuilder used to construct this object

child
child object or NULL for non-child tags. [ allow-none ]

tagname
name of tag

parser
a GMarkupParser structure to fill in

data
return location for user data that will be passed in to parser functions

Details

Since 2.12

Value

[logical] TRUE if a object has a custom implementation, FALSE if it doesn't.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]