gtkBuildableParserFinished {RGtk2}	R Documentation

gtkBuildableParserFinished

Description

Called when the builder finishes the parsing of a GtkBuilder UI definition. Note that this will be called once for each time gtkBuilderAddFromFile or gtkBuilderAddFromString is called on a builder.

Usage

gtkBuildableParserFinished(object, builder)

Arguments

object
a GtkBuildable

builder
a GtkBuilder

Details

Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]