gtkBuilderAddFromFile {RGtk2}	R Documentation

gtkBuilderAddFromFile

Description

Parses a file containing a GtkBuilder UI definition and merges it with the current contents of builder.

Usage

gtkBuilderAddFromFile(object, filename, .errwarn = TRUE)

Arguments

object
a GtkBuilder

filename
the name of the file to parse

.errwarn
Whether to issue a warning on error or fail silently

Details

Upon errors 0 will be returned and error will be assigned a GError from the GTK_BUILDER_ERROR, G_MARKUP_ERROR or G_FILE_ERROR domain. Since 2.12

Value

A list containing the following elements:

retval
[numeric] A positive value on success, 0 if an error occurred

error
return location for an error, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]