gtkBuilderAddFromString {RGtk2}	R Documentation

gtkBuilderAddFromString

Description

Parses a string containing a GtkBuilder UI definition and merges it with the current contents of builder.

Usage

gtkBuilderAddFromString(object, buffer, length, .errwarn = TRUE)

Arguments

object
a GtkBuilder

buffer
the string to parse

length
the length of buffer (may be -1 if buffer is nul-terminated)

.errwarn
Whether to issue a warning on error or fail silently

Details

Upon errors 0 will be returned and error will be assigned a GError from the GTK_BUILDER_ERROR or G_MARKUP_ERROR domain. Since 2.12

Value

A list containing the following elements:

retval
[numeric] A positive value on success, 0 if an error occurred

error
return location for an error, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]