gtkBuilderAddObjectsFromFile {RGtk2}	R Documentation

gtkBuilderAddObjectsFromFile

Description

Parses a file containing a GtkBuilder UI definition building only the requested objects and merges them with the current contents of builder.

Usage

gtkBuilderAddObjectsFromFile(object, filename, object.ids, .errwarn = TRUE)

Arguments

object
a GtkBuilder

filename
the name of the file to parse

object.ids
array of objects to build

.errwarn
Whether to issue a warning on error or fail silently

Details

Upon errors 0 will be returned and error will be assigned a GError from the GTK_BUILDER_ERROR, G_MARKUP_ERROR or G_FILE_ERROR domain. PLEASE NOTE: If you are adding an object that depends on an object that is not its child (for instance a GtkTreeView that depends on its GtkTreeModel), you have to explicitely list all of them in object.ids. Since 2.14

Value

A list containing the following elements:

retval
[numeric] A positive value on success, 0 if an error occurred

error
return location for an error, or NULL. [ allow-none ]

Note

If you are adding an object that depends on an object that is not its child (for instance a GtkTreeView that depends on its GtkTreeModel), you have to explicitely list all of them in object.ids.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]