gtkBuilderConnectSignals {RGtk2}	R Documentation

gtkBuilderConnectSignals

Description

This method is a simpler variation of gtkBuilderConnectSignalsFull. It uses GModule's introspective features (by opening the module NULL) to look at the application's symbol table. From here it tries to match the signal handler names given in the interface description with symbols in the application and connects the signals.

Usage

gtkBuilderConnectSignals(object, user.data = NULL)

Arguments

object
a GtkBuilder

user.data
a pointer to a structure sent in as user data to all signals

Details

Note that this function will not work correctly if GModule is not supported on the platform.

When compiling applications for Windows, you must declare signal callbacks with G_MODULE_EXPORT, or they will not be put in the symbol table. On Linux and Unices, this is not necessary; applications should instead be compiled with the -Wl,–export-dynamic CFLAGS, and linked against gmodule-export-2.0. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]