gtkBuilderGetObjects {RGtk2}	R Documentation

gtkBuilderGetObjects

Description

Gets all objects that have been constructed by builder. Note that this function does not increment the reference counts of the returned objects.

Usage

gtkBuilderGetObjects(object)

Arguments

object
a GtkBuilder

Details

Since 2.12

Value

[list] a newly-allocated list containing all the objects constructed by the GtkBuilder instance. [ element-type GObject][ transfer container ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]