gtkBuilderValueFromStringType {RGtk2}	R Documentation

gtkBuilderValueFromStringType

Description

Like gtkBuilderValueFromString, this function demarshals a value from a string, but takes a GType instead of GParamSpec. This function calls gValueInit() on the value argument, so it need not be initialised beforehand.

Usage

gtkBuilderValueFromStringType(object, type, string, .errwarn = TRUE)

Arguments

object
a GtkBuilder

type
the GType of the value

string
the string representation of the value

.errwarn
Whether to issue a warning on error or fail silently

Details

Upon errors FALSE will be returned and error will be assigned a GError from the GTK_BUILDER_ERROR domain. Since 2.12

Value

A list containing the following elements:

retval
[logical] TRUE on success

value
the R object to store the result in

error
return location for an error, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]