gtkButtonBoxGetChildIpadding {RGtk2}	R Documentation

gtkButtonBoxGetChildIpadding

Description

Gets the default number of pixels that pad the buttons in a given button box. WARNING: gtk_button_box_get_child_ipadding is deprecated and should not be used in newly-written code. Use the style properties "child-internal-pad-x" and "child-internal-pad-y" instead.

Usage

gtkButtonBoxGetChildIpadding(object)

Arguments

object
a GtkButtonBox.

Value

A list containing the following elements:

ipad.x
the horizontal padding used by buttons in widget.

ipad.y
the vertical padding used by buttons in widget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]