gtkButtonBoxSetChildSecondary {RGtk2}	R Documentation

gtkButtonBoxSetChildSecondary

Description

Sets whether child should appear in a secondary group of children. A typical use of a secondary child is the help button in a dialog.

Usage

gtkButtonBoxSetChildSecondary(object, child, is.secondary)

Arguments

object
a GtkButtonBox

child
a child of widget

is.secondary
if TRUE, the child appears in a secondary group of the button box.

Details

This group appears after the other children if the style is GTK_BUTTONBOX_START, GTK_BUTTONBOX_SPREAD or GTK_BUTTONBOX_EDGE, and before the other children if the style is GTK_BUTTONBOX_END. For horizontal button boxes, the definition of before/after depends on direction of the widget (see gtkWidgetSetDirection). If the style is GTK_BUTTONBOX_START or GTK_BUTTONBOX_END, then the secondary children are aligned at the other end of the button box from the main children. For the other styles, they appear immediately next to the main children.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]