gtkButtonBoxSetChildSize {RGtk2}	R Documentation

gtkButtonBoxSetChildSize

Description

Sets a new default size for the children of a given button box. WARNING: gtk_button_box_set_child_size is deprecated and should not be used in newly-written code. Use the style properties "child-min-width" and "child-min-height" instead.

Usage

gtkButtonBoxSetChildSize(object, min.width, min.height)

Arguments

object
a GtkButtonBox

min.width
a default width for buttons in widget

min.height
a default height for buttons in widget

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]