gtkButtonNewWithMnemonic {RGtk2}	R Documentation

gtkButtonNewWithMnemonic

Description

Creates a new GtkButton containing a label. If characters in label are preceded by an underscore, they are underlined. If you need a literal underscore character in a label, use '__' (two underscores). The first underlined character represents a keyboard accelerator called a mnemonic. Pressing Alt and that key activates the button.

Usage

gtkButtonNewWithMnemonic(label, show = TRUE)

Arguments

label
The text of the button, with an underscore in front of the mnemonic character

Value

[GtkWidget] a new GtkButton

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]