gtkButtonSetImage {RGtk2}	R Documentation

gtkButtonSetImage

Description

Set the image of button to the given widget. Note that it depends on the "gtk-button-images" setting whether the image will be displayed or not, you don't have to call gtkWidgetShow on image yourself.

Usage

gtkButtonSetImage(object, image)

Arguments

object
a GtkButton

image
a widget to set as the image for the button

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]