gtkButtonSetLabel {RGtk2}	R Documentation

gtkButtonSetLabel

Description

Sets the text of the label of the button to str. This text is also used to select the stock item if gtkButtonSetUseStock is used.

Usage

gtkButtonSetLabel(object, label)

Arguments

object
a GtkButton

label
a string

Details

This will also clear any previously set labels.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]