gtkButtonSetUseStock {RGtk2}	R Documentation

gtkButtonSetUseStock

Description

If TRUE, the label set on the button is used as a stock id to select the stock item for the button.

Usage

gtkButtonSetUseStock(object, use.stock)

Arguments

object
a GtkButton

use.stock
TRUE if the button should use a stock item

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]