gtkCListFreeze {RGtk2}	R Documentation

gtkCListFreeze

Description

Causes the GtkCList to stop updating its visuals until a matching call to gtkCListThaw is made. This function is useful if a lot of changes will be made to the widget that may cause a lot of visual updating to occur. Note that calls to gtkCListFreeze can be nested. WARNING: gtk_clist_freeze is deprecated and should not be used in newly-written code.

Usage

gtkCListFreeze(object)

Arguments

object
The GtkCList to freeze.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]