gtkCListGetCellType {RGtk2}	R Documentation

gtkCListGetCellType

Description

Checks the type of cell at the location specified. WARNING: gtk_clist_get_cell_type is deprecated and should not be used in newly-written code.

Usage

gtkCListGetCellType(object, row, column)

Arguments

object
The GtkCList to affect.

row
The row of the cell.

column
The column of the cell.

Value

[GtkCellType] A GtkCellType value describing the cell.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]