gtkCListGetColumnWidget {RGtk2}	R Documentation

gtkCListGetColumnWidget

Description

Gets the widget in the column header for the specified column. WARNING: gtk_clist_get_column_widget is deprecated and should not be used in newly-written code.

Usage

gtkCListGetColumnWidget(object, column)

Arguments

object
The GtkCList to affect.

column
The column to query.

Value

[GtkWidget] Pointer to a GtkWidget for the column header.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]