gtkCListGetRowData {RGtk2}	R Documentation

gtkCListGetRowData

Description

Gets the currently set data for the specified row. WARNING: gtk_clist_get_row_data is deprecated and should not be used in newly-written code.

Usage

gtkCListGetRowData(object, row)

Arguments

object
The GtkCList to affect.

row
The row to query.

Value

[R object] The data set for the row.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]