gtkCListMoveto {RGtk2}	R Documentation

gtkCListMoveto

Description

Tells the CList widget to visually move to the specified row and column. WARNING: gtk_clist_moveto is deprecated and should not be used in newly-written code.

Usage

gtkCListMoveto(object, row, column, row.align, col.align)

Arguments

object
The GtkCList to affect.

row
The row to which to move.

column
The column to which to move.

row.align
A value between 0 and 1 that describes the positioning of the row in relation to the viewable area of the CList's contents.

col.align
A value between 0 and 1 that describes the positioning of the column in relation to the viewable area of the CList's contents.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]