gtkCListNewWithTitles {RGtk2}	R Documentation

gtkCListNewWithTitles

Description

Creates a new GtkCList widget with column titles for use. WARNING: gtk_clist_new_with_titles is deprecated and should not be used in newly-written code.

Usage

gtkCListNewWithTitles(columns = 1, titles, show = TRUE)

Arguments

columns
The number of columns the GtkCList should have.

titles
A string list of titles for the widget. There should be enough strings in the list for the specified number of columns.

Value

[GtkWidget] A pointer to a new GtkCList object.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]