gtkCListOptimalColumnWidth {RGtk2}	R Documentation

gtkCListOptimalColumnWidth

Description

Gets the required width in pixels that is needed to show everything in the specified column. WARNING: gtk_clist_optimal_column_width is deprecated and should not be used in newly-written code.

Usage

gtkCListOptimalColumnWidth(object, column)

Arguments

object
The GtkCList to check.

column
The column to check.

Value

[integer] The required width in pixels for the column.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]