gtkCListRowIsVisible {RGtk2}	R Documentation

gtkCListRowIsVisible

Description

Checks how the specified row is visible. WARNING: gtk_clist_row_is_visible is deprecated and should not be used in newly-written code.

Usage

gtkCListRowIsVisible(object, row)

Arguments

object
The GtkCList to affect.

row
The row to query.

Value

[GtkVisibility] A GtkVisibility value that tells you how the row is visible.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]