gtkCListRowMove {RGtk2}	R Documentation

gtkCListRowMove

Description

Allows you to move a row from one position to another in the list. WARNING: gtk_clist_row_move is deprecated and should not be used in newly-written code.

Usage

gtkCListRowMove(object, source.row, dest.row)

Arguments

object
The GtkCList to affect.

source.row
The original position of the row to move.

dest.row
The position to which the row should be moved.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]