gtkCListSetColumnAutoResize {RGtk2}	R Documentation

gtkCListSetColumnAutoResize

Description

Lets you specify whether a column should be automatically resized by the widget when data is added or removed. Enabling auto-resize on a column explicity disallows user-resizing of the column. WARNING: gtk_clist_set_column_auto_resize is deprecated and should not be used in newly-written code.

Usage

gtkCListSetColumnAutoResize(object, column, auto.resize)

Arguments

object
The GtkCList to affect.

column
The column on which to set auto-resizing.

auto.resize
TRUE or FALSE.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]