gtkCListSetColumnResizeable {RGtk2}	R Documentation

gtkCListSetColumnResizeable

Description

Lets you specify whether a specified column should be resizeable by the user. Note that turning on resizeability for the column will automatically shut off auto-resizing, but turning off resizeability will NOT turn on auto-resizing. This must be done manually via a call to gtkCListSetColumnAutoResize. WARNING: gtk_clist_set_column_resizeable is deprecated and should not be used in newly-written code.

Usage

gtkCListSetColumnResizeable(object, column, resizeable)

Arguments

object
The GtkCList to affect.

column
The column on which to set resizeability.

resizeable
TRUE or FALSE.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]