gtkCListSetColumnVisibility {RGtk2}	R Documentation

gtkCListSetColumnVisibility

Description

Allows you to set whether a specified column in the GtkCList should be hidden or shown. Note that at least one column must always be showing, so attempting to hide the last visible column will be ignored. WARNING: gtk_clist_set_column_visibility is deprecated and should not be used in newly-written code.

Usage

gtkCListSetColumnVisibility(object, column, visible)

Arguments

object
The GtkCList to affect.

column
The column to set visibility.

visible
TRUE or FALSE.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]