gtkCListSetColumnWidget {RGtk2}	R Documentation

gtkCListSetColumnWidget

Description

Sets a widget to be used as the specified column's title. This can be used to place a pixmap or something else as the column title, instead of the standard text. WARNING: gtk_clist_set_column_widget is deprecated and should not be used in newly-written code.

Usage

gtkCListSetColumnWidget(object, column, widget)

Arguments

object
The GtkCList to affect.

column
The column whose title should be a widget.

widget
A pointer to a previously create widget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]