gtkCListSetCompareFunc {RGtk2}	R Documentation

gtkCListSetCompareFunc

Description

Sets the compare function of the GtkClist to cmp.func. If cmp.func is NULL, then the default compare function is used. The default compare function sorts ascending or with the type set by gtkCListSetSortType by the column set by gtkCListSetSortColumn. WARNING: gtk_clist_set_compare_func is deprecated and should not be used in newly-written code.

Usage

gtkCListSetCompareFunc(object, cmp.func)

Arguments

object
The GtkCList to affect.

cmp.func
The GtkCompareFunction to use.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]