gtkCListSetRowStyle {RGtk2}	R Documentation

gtkCListSetRowStyle

Description

Sets the style for all cells in the specified row. WARNING: gtk_clist_set_row_style is deprecated and should not be used in newly-written code.

Usage

gtkCListSetRowStyle(object, row, style)

Arguments

object
The GtkCList to affect.

row
The row to affect.

style
A pointer to a GtkStyle to set.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]