gtkCListSetSelectionMode {RGtk2}	R Documentation

gtkCListSetSelectionMode

Description

Sets the selection mode for the specified CList. This allows you to set whether only one or more than one item can be selected at a time in the widget. Note that setting the widget's selection mode to one of GTK_SELECTION_BROWSE or GTK_SELECTION_SINGLE will cause all the items in the GtkCList to become deselected. WARNING: gtk_clist_set_selection_mode is deprecated and should not be used in newly-written code.

Usage

gtkCListSetSelectionMode(object, mode)

Arguments

object
The GtkCList to affect.

mode
The GtkSelectionMode type to set for this CList.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]