gtkCListSetShift {RGtk2}	R Documentation

gtkCListSetShift

Description

Sets the vertical and horizontal shift of the specified cell. WARNING: gtk_clist_set_shift is deprecated and should not be used in newly-written code.

Usage

gtkCListSetShift(object, row, column, vertical, horizontal)

Arguments

object
The GtkCList to affect.

row
The row of the cell.

column
The column of the cell.

vertical
The value to set for the vertical shift.

horizontal
The value to set for the vertical shift.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]