gtkCTreeInsertNode {RGtk2}	R Documentation

gtkCTreeInsertNode

Description

Insert a new node to the tree. The position is specified through the parent-sibling notation, as explained in the introduction above. WARNING: gtk_ctree_insert_node is deprecated and should not be used in newly-written code.

Usage

gtkCTreeInsertNode(object, parent, sibling, text, spacing = 5, 
    pixmap.closed = NULL, mask.closed = NULL, pixmap.opened = NULL, 
    mask.opened = NULL, is.leaf = 1, expanded = 0)

Arguments

object
The GtkCTree widget.

parent
The parent node to be.

sibling
The sibling node to be.

text
The texts to be shown in each column.

spacing
The extra space between the pixmap and the text.

pixmap.closed
. [ allow-none ]

mask.closed
. [ allow-none ]

pixmap.opened
. [ allow-none ]

mask.opened
. [ allow-none ]

is.leaf
Whether this node is going to be a leaf.

expanded
Whether this node should start out expanded or not.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]