gtkCTreeIsViewable {RGtk2}	R Documentation

gtkCTreeIsViewable

Description

This function checks whether the given node is viewable i.e. so that all of its parent nodes are expanded. This is different from being actually visible: the node can be viewable but outside the scrolling area of the window. WARNING: gtk_ctree_is_viewable is deprecated and should not be used in newly-written code.

Usage

gtkCTreeIsViewable(object, node)

Arguments

object
Whether the node is viewable.

node
undocumented

Value

[logical] Whether the node is viewable.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]