gtkCTreeNewWithTitles {RGtk2}	R Documentation

gtkCTreeNewWithTitles

Description

Create a new GtkCTree widget with the given titles for the columns. WARNING: gtk_ctree_new_with_titles is deprecated and should not be used in newly-written code.

Usage

gtkCTreeNewWithTitles(columns = 1, tree.column = 0, titles, show = TRUE)

Arguments

columns
Number of columns.

tree.column
Which column has the tree graphic; 0 = leftmost.

titles
The titles for the columns.

Value

[GtkWidget] The GtkCTree widget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]