gtkCTreeNodeGetPixmap {RGtk2}	R Documentation

gtkCTreeNodeGetPixmap

Description

WARNING: gtk_ctree_node_get_pixmap is deprecated and should not be used in newly-written code.

Usage

gtkCTreeNodeGetPixmap(object, node, column)

Arguments

object
If nonnull, the pointer to the pixmap is returned through this.

node
If nonnull, the pointer to the mask is returned through this.

column
True if the given cell contains a pixmap.

Value

A list containing the following elements:

retval
[logical] True if the given cell contains a pixmap.

pixmap
undocumented

mask
undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]