gtkCTreeNodeGetText {RGtk2}	R Documentation

gtkCTreeNodeGetText

Description

WARNING: gtk_ctree_node_get_text is deprecated and should not be used in newly-written code.

Usage

gtkCTreeNodeGetText(object, node, column)

Arguments

object
If nonnull, the pointer to the text string is assigned to *text.

node
True if the given cell has text in it.

column
undocumented

Value

A list containing the following elements:

retval
[logical] True if the given cell has text in it.

text
undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]