gtkCTreeNodeIsVisible {RGtk2}	R Documentation

gtkCTreeNodeIsVisible

Description

WARNING: gtk_ctree_node_is_visible is deprecated and should not be used in newly-written code.

Usage

gtkCTreeNodeIsVisible(object, node)

Arguments

object
True if the node is currently inside the bounds of the window. Note that this function can return true even if the node is not viewable, if the node's ancestor is visible.

node
undocumented

Value

[GtkVisibility] True if the node is currently inside the bounds of the window. Note that this function can return true even if the node is not viewable, if the node's ancestor is visible.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]