gtkCTreeNodeMoveto {RGtk2}	R Documentation

gtkCTreeNodeMoveto

Description

This function makes the given column of the given node visible by scrolling. WARNING: gtk_ctree_node_moveto is deprecated and should not be used in newly-written code.

Usage

gtkCTreeNodeMoveto(object, node, column, row.align, col.align)

Arguments

object
The node to be made visible.

node
The column to be made visible.

column
Where in the window the row should appear.

row.align
Where in the window the column should appear.

col.align
undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]