gtkCTreeNodeSetRowDataFull {RGtk2}	R Documentation

gtkCTreeNodeSetRowDataFull

Description

This is the full interface to setting row data, so that a destructor can be given for the data. WARNING: gtk_ctree_node_set_row_data_full is deprecated and should not be used in newly-written code.

Usage

gtkCTreeNodeSetRowDataFull(object, node, data)

Arguments

object
The routine to be called when data is no longer needed.

node
undocumented

data
undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]