gtkCTreeNodeSetShift {RGtk2}	R Documentation

gtkCTreeNodeSetShift

Description

Shift the given cell the given amounts in pixels. WARNING: gtk_ctree_node_set_shift is deprecated and should not be used in newly-written code.

Usage

gtkCTreeNodeSetShift(object, node, column, vertical, horizontal)

Arguments

object
undocumented

node
undocumented

column
undocumented

vertical
undocumented

horizontal
undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]