gtkCTreePreRecursive {RGtk2}	R Documentation

gtkCTreePreRecursive

Description

Recursively apply a function to all nodes of the tree at or below a certain node. The function is called for each node after it has been called for its parent. WARNING: gtk_ctree_pre_recursive is deprecated and should not be used in newly-written code.

Usage

gtkCTreePreRecursive(object, node, func, data = NULL)

Arguments

object
The node where to start. NULL means to start at the root.

node
The function to apply to each node.

func
A closure argument given to each invocation of the function.

data
undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]