gtkCellLayoutPackStart {RGtk2}	R Documentation

gtkCellLayoutPackStart

Description

Packs the cell into the beginning of cell.layout. If expand is FALSE, then the cell is allocated no more space than it needs. Any unused space is divided evenly between cells for which expand is TRUE.

Usage

gtkCellLayoutPackStart(object, cell, expand = TRUE)

Arguments

object
A GtkCellLayout.

cell
A GtkCellRenderer.

expand
TRUE if cell is to be given extra space allocated to cell.layout.

Details

Note that reusing the same cell renderer is not supported. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]