gtkCellLayoutSetAttributes {RGtk2}	R Documentation

gtkCellLayoutSetAttributes

Description

Sets the attributes in list as the attributes of cell.layout. The attributes should be in attribute/column order, as in gtkCellLayoutAddAttribute. All existing attributes are removed, and replaced with the new attributes.

Usage

gtkCellLayoutSetAttributes(object, cell, ...)

Arguments

object
A GtkCellLayout.

cell
A GtkCellRenderer.

...
undocumented

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]