gtkCellRendererComboNew {RGtk2}	R Documentation

gtkCellRendererComboNew

Description

Creates a new GtkCellRendererCombo. Adjust how text is drawn using object properties. Object properties can be set globally (with gObjectSet). Also, with GtkTreeViewColumn, you can bind a property to a value in a GtkTreeModel. For example, you can bind the "text" property on the cell renderer to a string value in the model, thus rendering a different string in each row of the GtkTreeView.

Usage

gtkCellRendererComboNew()

Details

Since 2.6

Value

[GtkCellRenderer] the new cell renderer

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]