gtkCellRendererEditingCanceled {RGtk2}	R Documentation

gtkCellRendererEditingCanceled

Description

Causes the cell renderer to emit the gtkCellRendererEditingCanceled signal. WARNING: gtk_cell_renderer_editing_canceled has been deprecated since version 2.6 and should not be used in newly-written code. Use gtkCellRendererStopEditing instead

Usage

gtkCellRendererEditingCanceled(object)

Arguments

object
A GtkCellRenderer

Details

This function is for use only by implementations of cell renderers that need to notify the client program that an editing process was canceled and the changes were not committed. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]