gtkCellRendererGetPadding {RGtk2}	R Documentation

gtkCellRendererGetPadding

Description

Fills in xpad and ypad with the appropriate values of cell.

Usage

gtkCellRendererGetPadding(object, xpad, ypad)

Arguments

object
A GtkCellRenderer

xpad
location to fill in with the x padding of the cell, or NULL. [ allow-none ]

ypad
location to fill in with the y padding of the cell, or NULL. [ allow-none ]

Details

Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]