gtkCellRendererGetSize {RGtk2}	R Documentation

gtkCellRendererGetSize

Description

Obtains the width and height needed to render the cell. Used by view widgets to determine the appropriate size for the cell_area passed to gtkCellRendererRender. If cell.area is not NULL, fills in the x and y offsets (if set) of the cell relative to this location.

Usage

gtkCellRendererGetSize(object, widget, cell.area = NULL)

Arguments

object
a GtkCellRenderer

widget
the widget the renderer is rendering to

cell.area
The area a cell will be allocated, or NULL. [ allow-none ]

Details

Please note that the values set in width and height, as well as those in x.offset and y.offset are inclusive of the xpad and ypad properties.

Value

A list containing the following elements:

x.offset
location to return x offset of cell relative to cell.area, or NULL. [ allow-none ]

y.offset
location to return y offset of cell relative to cell.area, or NULL. [ allow-none ]

width
location to return width needed to render a cell, or NULL. [ allow-none ]

height
location to return height needed to render a cell, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]