gtkCellRendererRender {RGtk2}	R Documentation

gtkCellRendererRender

Description

Invokes the virtual render function of the GtkCellRenderer. The three passed-in rectangles are areas of window. Most renderers will draw within cell.area; the xalign, yalign, xpad, and ypad fields of the GtkCellRenderer should be honored with respect to cell.area. background.area includes the blank space around the cell, and also the area containing the tree expander; so the background.area rectangles for all cells tile to cover the entire window. expose.area is a clip rectangle.

Usage

gtkCellRendererRender(object, window, widget, background.area, 
    cell.area, expose.area, flags)

Arguments

object
a GtkCellRenderer

window
a GdkDrawable to draw to

widget
the widget owning window

background.area
entire cell area (including tree expanders and maybe padding on the sides)

cell.area
area normally rendered by a cell renderer

expose.area
area that actually needs updating

flags
flags that affect rendering

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]