gtkCellRendererToggleNew {RGtk2}	R Documentation

gtkCellRendererToggleNew

Description

Creates a new GtkCellRendererToggle. Adjust rendering parameters using object properties. Object properties can be set globally (with gObjectSet). Also, with GtkTreeViewColumn, you can bind a property to a value in a GtkTreeModel. For example, you can bind the "active" property on the cell renderer to a boolean value in the model, thus causing the check button to reflect the state of the model.

Usage

gtkCellRendererToggleNew()

Value

[GtkCellRenderer] the new cell renderer

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]