gtkCellViewSetDisplayedRow {RGtk2}	R Documentation

gtkCellViewSetDisplayedRow

Description

Sets the row of the model that is currently displayed by the GtkCellView. If the path is unset, then the contents of the cellview "stick" at their last value; this is not normally a desired result, but may be a needed intermediate state if say, the model for the GtkCellView becomes temporarily empty.

Usage

gtkCellViewSetDisplayedRow(object, path = NULL)

Arguments

object
a GtkCellView

path
a GtkTreePath or NULL to unset. [ allow-none ]

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]