gtkCheckMenuItemSetInconsistent {RGtk2}	R Documentation

gtkCheckMenuItemSetInconsistent

Description

If the user has selected a range of elements (such as some text or spreadsheet cells) that are affected by a boolean setting, and the current values in that range are inconsistent, you may want to display the check in an "in between" state. This function turns on "in between" display. Normally you would turn off the inconsistent state again if the user explicitly selects a setting. This has to be done manually, gtkCheckMenuItemSetInconsistent only affects visual appearance, it doesn't affect the semantics of the widget.

Usage

gtkCheckMenuItemSetInconsistent(object, setting)

Arguments

object
a GtkCheckMenuItem

setting
TRUE to display an "inconsistent" third state check

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]