gtkCheckMenuItemSetShowToggle {RGtk2}	R Documentation

gtkCheckMenuItemSetShowToggle

Description

Controls whether the check box is shown at all times. Normally the check box is shown only when it is active or while the menu item is selected. WARNING: gtk_check_menu_item_set_show_toggle is deprecated and should not be used in newly-written code.

Usage

gtkCheckMenuItemSetShowToggle(object, always)

Arguments

object
a GtkCheckMenuItem.

always
boolean value indicating whether to always show the check box.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]