gtkClipboardGet {RGtk2}	R Documentation

gtkClipboardGet

Description

Returns the clipboard object for the given selection. See gtkClipboardGetForDisplay for complete details.

Usage

gtkClipboardGet(selection = "GDK_SELECTION_CLIPBOARD")

Arguments

selection
a GdkAtom which identifies the clipboard to use

Value

[GtkClipboard] the appropriate clipboard object. If no clipboard already exists, a new one will be created. Once a clipboard object has been created, it is persistent and, [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]