gtkClipboardRequestContents {RGtk2}	R Documentation

gtkClipboardRequestContents

Description

Requests the contents of clipboard as the given target. When the results of the result are later received the supplied callback will be called.

Usage

gtkClipboardRequestContents(object, target, callback, user.data = NULL)

Arguments

object
a GtkClipboard

target
an atom representing the form into which the clipboard owner should convert the selection.

callback
A function to call when the results are received (or the retrieval fails). If the retrieval fails the length field of selection.data will be negative.

user.data
user data to pass to callback

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]