gtkClipboardRequestImage {RGtk2}	R Documentation

gtkClipboardRequestImage

Description

Requests the contents of the clipboard as image. When the image is later received, it will be converted to a GdkPixbuf, and callback will be called.

Usage

gtkClipboardRequestImage(object, callback, user.data = NULL)

Arguments

object
a GtkClipboard

callback
a function to call when the image is received, or the retrieval fails. (It will always be called one way or the other.)

user.data
user data to pass to callback.

Details

The pixbuf parameter to callback will contain the resulting GdkPixbuf if the request succeeded, or NULL if it failed. This could happen for various reasons, in particular if the clipboard was empty or if the contents of the clipboard could not be converted into an image. Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]