gtkClipboardRequestRichText {RGtk2}	R Documentation

gtkClipboardRequestRichText

Description

Requests the contents of the clipboard as rich text. When the rich text is later received, callback will be called.

Usage

gtkClipboardRequestRichText(object, buffer, callback, user.data)

Arguments

object
a GtkClipboard

buffer
a GtkTextBuffer

callback
a function to call when the text is received, or the retrieval fails. (It will always be called one way or the other.)

user.data
user data to pass to callback.

Details

The text parameter to callback will contain the resulting rich text if the request succeeded, or NULL if it failed. The length parameter will contain text's length. This function can fail for various reasons, in particular if the clipboard was empty or if the contents of the clipboard could not be converted into rich text form. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]