gtkClipboardRequestTargets {RGtk2}	R Documentation

gtkClipboardRequestTargets

Description

Requests the contents of the clipboard as list of supported targets. When the list is later received, callback will be called.

Usage

gtkClipboardRequestTargets(object, callback, user.data = NULL)

Arguments

object
a GtkClipboard

callback
a function to call when the targets are received, or the retrieval fails. (It will always be called one way or the other.)

user.data
user data to pass to callback.

Details

The targets parameter to callback will contain the resulting targets if the request succeeded, or NULL if it failed. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]