gtkClipboardSetCanStore {RGtk2}	R Documentation

gtkClipboardSetCanStore

Description

Hints that the clipboard data should be stored somewhere when the application exits or when gtkClipboardStore is called.

Usage

gtkClipboardSetCanStore(object, targets)

Arguments

object
a GtkClipboard

targets
array containing information about which forms should be stored or NULL to indicate that all forms should be stored.

Details

This value is reset when the clipboard owner changes. Where the clipboard data is stored is platform dependent, see gdkDisplayStoreClipboard for more information. Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]