gtkClipboardSetImage {RGtk2}	R Documentation

gtkClipboardSetImage

Description

Sets the contents of the clipboard to the given GdkPixbuf. GTK+ will take responsibility for responding for requests for the image, and for converting the image into the requested format.

Usage

gtkClipboardSetImage(object, pixbuf)

Arguments

object
a GtkClipboard object

pixbuf
a GdkPixbuf

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]