gtkClipboardWaitForRichText {RGtk2}	R Documentation

gtkClipboardWaitForRichText

Description

Requests the contents of the clipboard as rich text. This function waits for the data to be received using the main loop, so events, timeouts, etc, may be dispatched during the wait.

Usage

gtkClipboardWaitForRichText(object, buffer)

Arguments

object
a GtkClipboard

buffer
a GtkTextBuffer

Details

Since 2.10

Value

A list containing the following elements:

retval
[raw] or NULL if retrieving the selection data failed. (This could happen for various reasons, in particular if the clipboard was empty or if the contents of the clipboard could not be converted into text form.)

format
return location for the format of the returned data

length
return location for the length of the returned data

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]