gtkClipboardWaitForTargets {RGtk2}	R Documentation

gtkClipboardWaitForTargets

Description

Returns a list of targets that are present on the clipboard, or NULL if there aren't any targets available. This function waits for the data to be received using the main loop, so events, timeouts, etc, may be dispatched during the wait.

Usage

gtkClipboardWaitForTargets(object)

Arguments

object
a GtkClipboard

Details

Since 2.4

Value

A list containing the following elements:

retval
[logical] TRUE if any targets are present on the clipboard, otherwise FALSE.

targets
location to store a list of targets. The result stored here must be freed with gFree().

n.targets
location to store number of items in targets.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]