gtkClipboardWaitForUris {RGtk2}	R Documentation

gtkClipboardWaitForUris

Description

Requests the contents of the clipboard as URIs. This function waits for the data to be received using the main loop, so events, timeouts, etc, may be dispatched during the wait.

Usage

gtkClipboardWaitForUris(object)

Arguments

object
a GtkClipboard

Details

Since 2.14

Value

[character] or NULL if retrieving the selection data failed. (This could happen for various reasons, in particular if the clipboard was empty or if the contents of the clipboard could not be converted into URI form.). [ array zero-terminated=1][ element-type utf8][ transfer full ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]