gtkClipboardWaitIsImageAvailable {RGtk2}	R Documentation

gtkClipboardWaitIsImageAvailable

Description

Test to see if there is an image available to be pasted This is done by requesting the TARGETS atom and checking if it contains any of the supported image targets. This function waits for the data to be received using the main loop, so events, timeouts, etc, may be dispatched during the wait.

Usage

gtkClipboardWaitIsImageAvailable(object)

Arguments

object
a GtkClipboard

Details

This function is a little faster than calling gtkClipboardWaitForImage since it doesn't need to retrieve the actual image data. Since 2.6

Value

[logical] TRUE is there is an image available, FALSE otherwise.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]