gtkClipboardWaitIsTargetAvailable {RGtk2}	R Documentation

gtkClipboardWaitIsTargetAvailable

Description

Checks if a clipboard supports pasting data of a given type. This function can be used to determine if a "Paste" menu item should be insensitive or not.

Usage

gtkClipboardWaitIsTargetAvailable(object, target)

Arguments

object
a GtkClipboard

target
A GdkAtom indicating which target to look for.

Details

If you want to see if there's text available on the clipboard, use gtkClipboardWaitIsTextAvailable instead. Since 2.6

Value

[logical] TRUE if the target is available, FALSE otherwise.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]