gtkClipboardWaitIsUrisAvailable {RGtk2}	R Documentation

gtkClipboardWaitIsUrisAvailable

Description

Test to see if there is a list of URIs available to be pasted This is done by requesting the TARGETS atom and checking if it contains the URI targets. This function waits for the data to be received using the main loop, so events, timeouts, etc, may be dispatched during the wait.

Usage

gtkClipboardWaitIsUrisAvailable(object)

Arguments

object
a GtkClipboard

Details

This function is a little faster than calling gtkClipboardWaitForUris since it doesn't need to retrieve the actual URI data. Since 2.14

Value

[logical] TRUE is there is an URI list available, FALSE otherwise.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]