gtkColorSelectionSetChangePaletteHook {RGtk2}	R Documentation

gtkColorSelectionSetChangePaletteHook

Description

Installs a global function to be called whenever the user tries to modify the palette in a color selection. This function should save the new palette contents, and update the GtkSettings property "gtk-color-palette" so all GtkColorSelection widgets will be modified. WARNING: gtk_color_selection_set_change_palette_hook has been deprecated since version 2.4 and should not be used in newly-written code. This function does not work in multihead environments. Use gtkColorSelectionSetChangePaletteWithScreenHook instead.

Usage

gtkColorSelectionSetChangePaletteHook(func)

Arguments

func
a function to call when the custom palette needs saving.

Value

[GtkColorSelectionChangePaletteFunc] the previous change palette hook (that was replaced).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]