gtkColorSelectionSetChangePaletteWithScreenHook {RGtk2}	R Documentation

gtkColorSelectionSetChangePaletteWithScreenHook

Description

Installs a global function to be called whenever the user tries to modify the palette in a color selection. This function should save the new palette contents, and update the GtkSettings property "gtk-color-palette" so all GtkColorSelection widgets will be modified.

Usage

gtkColorSelectionSetChangePaletteWithScreenHook(func)

Arguments

func
a function to call when the custom palette needs saving.

Details

Since 2.2

Value

[GtkColorSelectionChangePaletteWithScreenFunc] the previous change palette hook (that was replaced).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]