gtkColorSelectionSetColor {RGtk2}	R Documentation

gtkColorSelectionSetColor

Description

Sets the current color to be color. The first time this is called, it will also set the original color to be color too. WARNING: gtk_color_selection_set_color has been deprecated since version 2.0 and should not be used in newly-written code. Use gtkColorSelectionSetCurrentColor instead.

Usage

gtkColorSelectionSetColor(object, color)

Arguments

object
a GtkColorSelection.

color
a list of 4 doubles specifying the red, green, blue and opacity to set the current color to.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]