gtkColorSelectionSetPreviousAlpha {RGtk2}	R Documentation

gtkColorSelectionSetPreviousAlpha

Description

Sets the 'previous' alpha to be alpha. This function should be called with some hesitations, as it might seem confusing to have that alpha change.

Usage

gtkColorSelectionSetPreviousAlpha(object, alpha)

Arguments

object
a GtkColorSelection.

alpha
an integer between 0 and 65535.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]