gtkColorSelectionSetPreviousColor {RGtk2}	R Documentation

gtkColorSelectionSetPreviousColor

Description

Sets the 'previous' color to be color. This function should be called with some hesitations, as it might seem confusing to have that color change. Calling gtkColorSelectionSetCurrentColor will also set this color the first time it is called.

Usage

gtkColorSelectionSetPreviousColor(object, color)

Arguments

object
a GtkColorSelection.

color
a GdkColor to set the previous color with.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]