gtkColorSelectionSetUpdatePolicy {RGtk2}	R Documentation

gtkColorSelectionSetUpdatePolicy

Description

Sets the policy controlling when the color_changed signals are emitted. The available policies are:

GTK_UPDATE_CONTINUOUS - signals are sent continuously as the color selection changes.

GTK_UPDATE_DISCONTINUOUS - signals are sent only when the mouse button is released.

GTK_UPDATE_DELAYED - signals are sent when the mouse button is released or when the mouse has been motionless for a period of time.

WARNING: gtk_color_selection_set_update_policy is deprecated and should not be used in newly-written code.

Usage

gtkColorSelectionSetUpdatePolicy(object, policy)

Arguments

object
a GtkColorSelection.

policy
a GtkUpdateType value indicating the desired policy.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]