gtkComboBoxAppendText {RGtk2}	R Documentation

gtkComboBoxAppendText

Description

Appends string to the list of strings stored in combo.box. Note that you can only use this function with combo boxes constructed with gtkComboBoxNewText.

Usage

gtkComboBoxAppendText(object, text)

Arguments

object
A GtkComboBox constructed using gtkComboBoxNewText

text
A string

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]