gtkComboBoxEntryNewText {RGtk2}	R Documentation

gtkComboBoxEntryNewText

Description

Convenience function which constructs a new editable text combo box, which is a GtkComboBoxEntry just displaying strings. If you use this function to create a text combo box, you should only manipulate its data source with the following convenience functions: gtkComboBoxAppendText, gtkComboBoxInsertText, gtkComboBoxPrependText and gtkComboBoxRemoveText.

Usage

gtkComboBoxEntryNewText()

Details

Since 2.4

Value

[GtkWidget] A new text GtkComboBoxEntry.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]