gtkComboBoxEntryNewWithModel {RGtk2}	R Documentation

gtkComboBoxEntryNewWithModel

Description

Creates a new GtkComboBoxEntry which has a GtkEntry as child and a list of strings as popup. You can get the GtkEntry from a GtkComboBoxEntry using GTK_ENTRY (GTK_BIN (combo_box_entry)->child). To add and remove strings from the list, just modify model using its data manipulation API.

Usage

gtkComboBoxEntryNewWithModel(model, text.column, show = TRUE)

Arguments

model
A GtkTreeModel.

text.column
A column in model to get the strings from.

Details

Since 2.4

Value

[GtkWidget] A new GtkComboBoxEntry.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]