gtkComboBoxGetActive {RGtk2}	R Documentation

gtkComboBoxGetActive

Description

Returns the index of the currently active item, or -1 if there's no active item. If the model is a non-flat treemodel, and the active item is not an immediate child of the root of the tree, this function returns gtk_tree_path_get_indices (path)[0], where path is the GtkTreePath of the active item.

Usage

gtkComboBoxGetActive(object)

Arguments

object
A GtkComboBox

Details

Since 2.4

Value

[integer] An integer which is the index of the currently active item, or -1 if there's no active item.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]