gtkComboBoxGetPopupAccessible {RGtk2}	R Documentation

gtkComboBoxGetPopupAccessible

Description

Gets the accessible object corresponding to the combo box's popup.

Usage

gtkComboBoxGetPopupAccessible(object)

Arguments

object
a GtkComboBox

Details

This function is mostly intended for use by accessibility technologies; applications should have little use for it. Since 2.6

Value

[AtkObject] the accessible object corresponding to the combo box's popup.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]