gtkComboBoxSetColumnSpanColumn {RGtk2}	R Documentation

gtkComboBoxSetColumnSpanColumn

Description

Sets the column with column span information for combo.box to be column.span. The column span column contains integers which indicate how many columns an item should span.

Usage

gtkComboBoxSetColumnSpanColumn(object, column.span)

Arguments

object
A GtkComboBox

column.span
A column in the model passed during construction

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]