gtkComboBoxSetModel {RGtk2}	R Documentation

gtkComboBoxSetModel

Description

Sets the model used by combo.box to be model. Will unset a previously set model (if applicable). If model is NULL, then it will unset the model.

Usage

gtkComboBoxSetModel(object, model = NULL)

Arguments

object
A GtkComboBox

model
A GtkTreeModel. [ allow-none ]

Details

Note that this function does not clear the cell renderers, you have to call gtkCellLayoutClear yourself if you need to set up different cell renderers for the new model. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]