gtkComboBoxSetWrapWidth {RGtk2}	R Documentation

gtkComboBoxSetWrapWidth

Description

Sets the wrap width of combo.box to be width. The wrap width is basically the preferred number of columns when you want the popup to be layed out in a table.

Usage

gtkComboBoxSetWrapWidth(object, width)

Arguments

object
A GtkComboBox

width
Preferred number of columns

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]