gtkComboDisableActivate {RGtk2}	R Documentation

gtkComboDisableActivate

Description

Stops the GtkCombo widget from showing the popup list when the GtkEntry emits the "activate" signal, i.e. when the Return key is pressed. This may be useful if, for example, you want the Return key to close a dialog instead. WARNING: gtk_combo_disable_activate has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkComboBox instead.

Usage

gtkComboDisableActivate(object)

Arguments

object
a GtkCombo.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]