gtkComboSetCaseSensitive {RGtk2}	R Documentation

gtkComboSetCaseSensitive

Description

Specifies whether the text entered into the GtkEntry field and the text in the list items is case sensitive. WARNING: gtk_combo_set_case_sensitive has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkComboBox instead.

Usage

gtkComboSetCaseSensitive(object, val)

Arguments

object
a GtkCombo.

val
TRUE if the text in the list items is case sensitive.

Details

This may be useful, for example, when you have called gtkComboSetValueInList to limit the values entered, but you are not worried about differences in case.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]