gtkComboSetUseArrows {RGtk2}	R Documentation

gtkComboSetUseArrows

Description

Specifies if the arrow (cursor) keys can be used to step through the items in the list. This is on by default. WARNING: gtk_combo_set_use_arrows has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkComboBox instead.

Usage

gtkComboSetUseArrows(object, val)

Arguments

object
a GtkCombo.

val
TRUE if the arrow keys can be used to step through the items in the list.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]