gtkComboSetValueInList {RGtk2}	R Documentation

gtkComboSetValueInList

Description

Specifies whether the value entered in the text entry field must match one of the values in the list. If this is set then the user will not be able to perform any other action until a valid value has been entered. WARNING: gtk_combo_set_value_in_list has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkComboBox instead.

Usage

gtkComboSetValueInList(object, val, ok.if.empty)

Arguments

object
a GtkCombo.

val
TRUE if the value entered must match one of the values in the list.

ok.if.empty
TRUE if an empty value is considered valid.

Details

If an empty field is acceptable, the ok.if.empty parameter should be TRUE.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]