gtkContainerAdd {RGtk2}	R Documentation

gtkContainerAdd

Description

Adds widget to container. Typically used for simple containers such as GtkWindow, GtkFrame, or GtkButton; for more complicated layout containers such as GtkBox or GtkTable, this function will pick default packing parameters that may not be correct. So consider functions such as gtkBoxPackStart and gtkTableAttach as an alternative to gtkContainerAdd in those cases. A widget may be added to only one container at a time; you can't place the same widget inside two different containers.

Usage

gtkContainerAdd(object, widget)

Arguments

object
a GtkContainer

widget
a widget to be placed inside container

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]