gtkContainerChildType {RGtk2}	R Documentation

gtkContainerChildType

Description

Returns the type of the children supported by the container.

Usage

gtkContainerChildType(object)

Arguments

object
a GtkContainer

Details

Note that this may return G_TYPE_NONE to indicate that no more children can be added, e.g. for a GtkPaned which already has two children.

Value

[GType] a GType.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]