gtkContainerClassFindChildProperty {RGtk2}	R Documentation

gtkContainerClassFindChildProperty

Description

Finds a child property of a container class by name.

Usage

gtkContainerClassFindChildProperty(cclass, property.name)

Arguments

cclass
a GtkContainerClass

property.name
the name of the child property to find

Value

[GParamSpec] the GParamSpec of the child property or NULL if class has no child property with that name. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]