gtkContainerForall {RGtk2}	R Documentation

gtkContainerForall

Description

Invokes callback on each child of container, including children that are considered "internal" (implementation details of the container). "Internal" children generally weren't added by the user of the container, but were added by the container implementation itself. Most applications should use gtkContainerForeach, rather than gtkContainerForall.

Usage

gtkContainerForall(object, callback, callback.data = NULL)

Arguments

object
a GtkContainer

callback
a callback

callback.data
callback user data

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]