gtkContainerRemove {RGtk2}	R Documentation

gtkContainerRemove

Description

Removes widget from container. widget must be inside container. Note that container will own a reference to widget, and that this may be the last reference held; so removing a widget from its container can destroy that widget. If you want to use widget again, you need to add a reference to it while it's not inside a container, If you don't want to use widget again it's usually more efficient to simply destroy it directly using gtkWidgetDestroy since this will remove it from the container and help break any circular reference count cycles.

Usage

gtkContainerRemove(object, widget)

Arguments

object
a GtkContainer

widget
a current child of container

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]