gtkContainerSetBorderWidth {RGtk2}	R Documentation

gtkContainerSetBorderWidth

Description

Sets the border width of the container.

Usage

gtkContainerSetBorderWidth(object, border.width)

Arguments

object
a GtkContainer

border.width
amount of blank space to leave outside the container. Valid values are in the range 0-65535 pixels.

Details

The border width of a container is the amount of space to leave around the outside of the container. The only exception to this is GtkWindow; because toplevel windows can't leave space outside, they leave the space inside. The border is added on all sides of the container. To add space to only one side, one approach is to create a GtkAlignment widget, call gtkWidgetSetSizeRequest to give it a size, and place it on the side of the container as a spacer.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]