gtkContainerSetFocusHadjustment {RGtk2}	R Documentation

gtkContainerSetFocusHadjustment

Description

Hooks up an adjustment to focus handling in a container, so when a child of the container is focused, the adjustment is scrolled to show that widget. This function sets the horizontal alignment. See gtkScrolledWindowGetHadjustment for a typical way of obtaining the adjustment and gtkContainerSetFocusVadjustment for setting the vertical adjustment.

Usage

gtkContainerSetFocusHadjustment(object, adjustment)

Arguments

object
a GtkContainer

adjustment
an adjustment which should be adjusted when the focus is moved among the descendents of container

Details

The adjustments have to be in pixel units and in the same coordinate system as the allocation for immediate children of the container.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]