gtkContainerSetReallocateRedraws {RGtk2}	R Documentation

gtkContainerSetReallocateRedraws

Description

Sets the reallocate.redraws flag of the container to the given value.

Usage

gtkContainerSetReallocateRedraws(object, needs.redraws)

Arguments

object
a GtkContainer

needs.redraws
the new value for the container's reallocate.redraws flag

Details

Containers requesting reallocation redraws get automatically redrawn if any of their children changed allocation.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]