gtkCurveSetCurveType {RGtk2}	R Documentation

gtkCurveSetCurveType

Description

Sets the type of the curve. The curve will remain unchanged except when changing from a free curve to a linear or spline curve, in which case the curve will be changed as little as possible. WARNING: gtk_curve_set_curve_type has been deprecated since version 2.20 and should not be used in newly-written code. Don't use this widget anymore.

Usage

gtkCurveSetCurveType(object, type)

Arguments

object
a GtkCurve.

type
the type of the curve.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]