gtkDialogAddActionWidget {RGtk2}	R Documentation

gtkDialogAddActionWidget

Description

Adds an activatable widget to the action area of a GtkDialog, connecting a signal handler that will emit the gtkDialogResponse signal on the dialog when the widget is activated. The widget is appended to the end of the dialog's action area. If you want to add a non-activatable widget, simply pack it into the action.area field of the GtkDialog struct.

Usage

gtkDialogAddActionWidget(object, child, response.id)

Arguments

object
a GtkDialog

child
an activatable widget

response.id
response ID for child

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]