gtkDialogAddButton {RGtk2}	R Documentation

gtkDialogAddButton

Description

Adds a button with the given text (or a stock button, if button.text is a stock ID) and sets things up so that clicking the button will emit the gtkDialogResponse signal with the given response.id. The button is appended to the end of the dialog's action area. The button widget is returned, but usually you don't need it.

Usage

gtkDialogAddButton(object, button.text, response.id)

Arguments

object
a GtkDialog

button.text
text of button, or stock ID

response.id
response ID for the button

Value

[GtkWidget] the button widget that was added

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]