gtkDialogSetAlternativeButtonOrder {RGtk2}	R Documentation

gtkDialogSetAlternativeButtonOrder

Description

Sets an alternative button order. If the "gtk-alternative-button-order" setting is set to TRUE, the dialog buttons are reordered according to the order of the response ids passed to this function.

Usage

gtkDialogSetAlternativeButtonOrder(object, ...)

Arguments

object
a GtkDialog

...
undocumented

Details

By default, GTK+ dialogs use the button order advocated by the Gnome Human Interface Guidelines (http://developer.gnome.org/projects/gup/hig/2.0/) with the affirmative button at the far right, and the cancel button left of it. But the builtin GTK+ dialogs and GtkMessageDialogs do provide an alternative button order, which is more suitable on some platforms, e.g. Windows.

Use this function after adding all the buttons to your dialog, as the following example shows:

cancel_button <- dialog$addButton("gtk-cancel", "cancel")
 
ok_button <- dialog$addButton("gtk-ok", "ok")
ok_button$grabDefault()
  
help_button <- dialog$addButton("gtk-help", "help")

dialog$setAlternativeButtonOrder("ok", "cancel", "help")

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]