gtkDragBegin {RGtk2}	R Documentation

gtkDragBegin

Description

Initiates a drag on the source side. The function only needs to be used when the application is starting drags itself, and is not needed when gtkDragSourceSet is used.

Usage

gtkDragBegin(object, targets, actions, button, event)

Arguments

object
the source widget.

targets
The targets (data formats) in which the source can provide the data.

actions
A bitmask of the allowed drag actions for this drag.

button
The button the user clicked to start the drag.

event
The event that triggered the start of the drag.

Value

[GdkDragContext] the context for this drag.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]