gtkDragDestFindTarget {RGtk2}	R Documentation

gtkDragDestFindTarget

Description

Looks for a match between context->targets and the dest.target.list, returning the first matching target, otherwise returning GDK_NONE. dest.target.list should usually be the return value from gtkDragDestGetTargetList, but some widgets may have different valid targets for different parts of the widget; in that case, they will have to implement a drag_motion handler that passes the correct target list to this function.

Usage

gtkDragDestFindTarget(object, context, target.list)

Arguments

object
drag destination widget

context
drag context

target.list
list of droppable targets, or NULL to use gtk_drag_dest_get_target_list (widget). [ allow-none ]

Value

[GdkAtom] first target that the source offers and the dest can accept, or GDK_NONE

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]