gtkDragDestSetProxy {RGtk2}	R Documentation

gtkDragDestSetProxy

Description

Sets this widget as a proxy for drops to another window.

Usage

gtkDragDestSetProxy(object, proxy.window, protocol, use.coordinates)

Arguments

object
a GtkWidget

proxy.window
the window to which to forward drag events

protocol
the drag protocol which the proxy.window accepts (You can use gdkDragGetProtocol to determine this)

use.coordinates
If TRUE, send the same coordinates to the destination, because it is an embedded subwindow.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]