gtkDragFinish {RGtk2}	R Documentation

gtkDragFinish

Description

Informs the drag source that the drop is finished, and that the data of the drag will no longer be required.

Usage

gtkDragFinish(object, success, del, time = "GDK_CURRENT_TIME")

Arguments

object
the drag context.

success
a flag indicating whether the drop was successful

del
a flag indicating whether the source should delete the original data. (This should be TRUE for a move)

time
the timestamp from the "drag_data_drop" signal.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]