gtkDragGetData {RGtk2}	R Documentation

gtkDragGetData

Description

Gets the data associated with a drag. When the data is received or the retrieval fails, GTK+ will emit a "drag_data_received" signal. Failure of the retrieval is indicated by the length field of the selection.data signal parameter being negative. However, when gtkDragGetData is called implicitely because the GTK_DEST_DEFAULT_DROP was set, then the widget will not receive notification of failed drops.

Usage

gtkDragGetData(object, context, target, time = "GDK_CURRENT_TIME")

Arguments

object
the widget that will receive the "drag_data_received" signal.

context
the drag context

target
the target (form of the data) to retrieve.

time
a timestamp for retrieving the data. This will generally be the time received in a "drag_data_motion" or "drag_data_drop" signal.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]