gtkDragSetIconName {RGtk2}	R Documentation

gtkDragSetIconName

Description

Sets the icon for a given drag from a named themed icon. See the docs for GtkIconTheme for more details. Note that the size of the icon depends on the icon theme (the icon is loaded at the symbolic size GTK_ICON_SIZE_DND), thus hot.x and hot.y have to be used with care.

Usage

gtkDragSetIconName(object, icon.name, hot.x, hot.y)

Arguments

object
the context for a drag. (This must be called with a context for the source side of a drag)

icon.name
name of icon to use

hot.x
the X offset of the hotspot within the icon

hot.y
the Y offset of the hotspot within the icon

Details

Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]