gtkDragSetIconPixbuf {RGtk2}	R Documentation

gtkDragSetIconPixbuf

Description

Sets pixbuf as the icon for a given drag.

Usage

gtkDragSetIconPixbuf(object, pixbuf, hot.x, hot.y)

Arguments

object
the context for a drag. (This must be called with a context for the source side of a drag)

pixbuf
the GdkPixbuf to use as the drag icon.

hot.x
the X offset within widget of the hotspot.

hot.y
the Y offset within widget of the hotspot.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]