gtkDragSetIconPixmap {RGtk2}	R Documentation

gtkDragSetIconPixmap

Description

Sets pixmap as the icon for a given drag. GTK+ retains references for the arguments, and will release them when they are no longer needed. In general, gtkDragSetIconPixbuf will be more convenient to use.

Usage

gtkDragSetIconPixmap(object, colormap, pixmap, mask, hot.x, hot.y)

Arguments

object
the context for a drag. (This must be called with a context for the source side of a drag)

colormap
the colormap of the icon

pixmap
the image data for the icon

mask
the transparency mask for the icon or NULL for none. [ allow-none ]

hot.x
the X offset within pixmap of the hotspot.

hot.y
the Y offset within pixmap of the hotspot.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]