gtkDragSetIconWidget {RGtk2}	R Documentation

gtkDragSetIconWidget

Description

Changes the icon for a widget to a given widget. GTK+ will not destroy the icon, so if you don't want it to persist, you should connect to the "drag-end" signal and destroy it yourself.

Usage

gtkDragSetIconWidget(object, widget, hot.x, hot.y)

Arguments

object
the context for a drag. (This must be called with a context for the source side of a drag)

widget
a toplevel window to use as an icon.

hot.x
the X offset within widget of the hotspot.

hot.y
the Y offset within widget of the hotspot.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]