gtkDragSourceSet {RGtk2}	R Documentation

gtkDragSourceSet

Description

Sets up a widget so that GTK+ will start a drag operation when the user clicks and drags on the widget. The widget must have a window.

Usage

gtkDragSourceSet(object, start.button.mask, targets, actions)

Arguments

object
a GtkWidget

start.button.mask
the bitmask of buttons that can start the drag

targets
the table of targets that the drag will support, may be NULL. [ allow-none ][ array length=n_targets]

actions
the bitmask of possible actions for a drag from this widget

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]