gtkDragSourceSetIcon {RGtk2}	R Documentation

gtkDragSourceSetIcon

Description

Sets the icon that will be used for drags from a particular widget from a pixmap/mask. GTK+ retains references for the arguments, and will release them when they are no longer needed. Use gtkDragSourceSetIconPixbuf instead.

Usage

gtkDragSourceSetIcon(object, colormap, pixmap, mask = NULL)

Arguments

object
a GtkWidget

colormap
the colormap of the icon

pixmap
the image data for the icon

mask
the transparency mask for an image. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]