gtkDragSourceSetIconPixbuf {RGtk2}	R Documentation

gtkDragSourceSetIconPixbuf

Description

Sets the icon that will be used for drags from a particular widget from a GdkPixbuf. GTK+ retains a reference for pixbuf and will release it when it is no longer needed.

Usage

gtkDragSourceSetIconPixbuf(object, pixbuf)

Arguments

object
a GtkWidget

pixbuf
the GdkPixbuf for the drag icon

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]