gtkDragSourceSetTargetList {RGtk2}	R Documentation

gtkDragSourceSetTargetList

Description

Changes the target types that this widget offers for drag-and-drop. The widget must first be made into a drag source with gtkDragSourceSet.

Usage

gtkDragSourceSetTargetList(object, target.list)

Arguments

object
a GtkWidget that's a drag source

target.list
list of draggable targets, or NULL for none. [ allow-none ]

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]