gtkDrawInsertionCursor {RGtk2}	R Documentation

gtkDrawInsertionCursor

Description

Draws a text caret on drawable at location. This is not a style function but merely a convenience function for drawing the standard cursor shape.

Usage

gtkDrawInsertionCursor(widget, drawable, area = NULL, location, 
    is.primary, direction, draw.arrow)

Arguments

widget
a GtkWidget

drawable
a GdkDrawable

area
rectangle to which the output is clipped, or NULL if the output should not be clipped. [ allow-none ]

location
location where to draw the cursor (location->width is ignored)

is.primary
if the cursor should be the primary cursor color.

direction
whether the cursor is left-to-right or right-to-left. Should never be GTK_TEXT_DIR_NONE

draw.arrow
TRUE to draw a directional arrow on the cursor. Should be FALSE unless the cursor is split.

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]