gtkDrawLayout {RGtk2}	R Documentation

gtkDrawLayout

Description

Draws a layout on window using the given parameters. WARNING: gtk_draw_layout is deprecated and should not be used in newly-written code.

Usage

gtkDrawLayout(object, window, state.type, use.text, x, y, layout)

Arguments

object
a GtkStyle

window
a GdkWindow

state.type
a state

use.text
whether to use the text or foreground graphics context of style

x
x origin

y
y origin

layout
the layout to draw

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]