gtkDrawResizeGrip {RGtk2}	R Documentation

gtkDrawResizeGrip

Description

Draws a resize grip in the given rectangle on window using the given parameters. WARNING: gtk_draw_resize_grip has been deprecated since version 2.0 and should not be used in newly-written code. Use gtkPaintResizeGrip instead.

Usage

gtkDrawResizeGrip(object, window, state.type, edge, x, y, width, 
    height)

Arguments

object
a GtkStyle

window
a GdkWindow

state.type
a state

edge
the edge in which to draw the resize grip

x
the x origin of the rectangle in which to draw the resize grip

y
the y origin of the rectangle in which to draw the resize grip

width
the width of the rectangle in which to draw the resize grip

height
the height of the rectangle in which to draw the resize grip

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]