gtkDrawShadow {RGtk2}	R Documentation

gtkDrawShadow

Description

Draws a shadow around the given rectangle in window using the given style and state and shadow type. WARNING: gtk_draw_shadow has been deprecated since version 2.0 and should not be used in newly-written code. Use gtkPaintShadow instead.

Usage

gtkDrawShadow(object, window, state.type, shadow.type, x, y, 
    width, height)

Arguments

object
a GtkStyle

window
a GdkWindow

state.type
a state

shadow.type
type of shadow to draw

x
x origin of the rectangle

y
y origin of the rectangle

width
width of the rectangle

height
width of the rectangle

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]