gtkDrawSlider {RGtk2}	R Documentation

gtkDrawSlider

Description

Draws a slider in the given rectangle on window using the given style and orientation. WARNING: gtk_draw_slider is deprecated and should not be used in newly-written code.

Usage

gtkDrawSlider(object, window, state.type, shadow.type, x, y, 
    width, height, orientation)

Arguments

object
a GtkStyle

window
a GdkWindow

state.type
a state

shadow.type
a shadow

x
the x origin of the rectangle in which to draw a slider

y
the y origin of the rectangle in which to draw a slider

width
the width of the rectangle in which to draw a slider

height
the height of the rectangle in which to draw a slider

orientation
the orientation to be used

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]