gtkDrawingAreaSize {RGtk2}	R Documentation

gtkDrawingAreaSize

Description

Sets the size that the drawing area will request in response to a "size_request" signal. The drawing area may actually be allocated a size larger than this depending on how it is packed within the enclosing containers. WARNING: gtk_drawing_area_size is deprecated and should not be used in newly-written code. Use gtkWidgetSetSizeRequest instead.

Usage

gtkDrawingAreaSize(object, width, height)

Arguments

object
a GtkDrawingArea

width
the width to request

height
the height to request

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]