gtkEditableDeleteText {RGtk2}	R Documentation

gtkEditableDeleteText

Description

Deletes a sequence of characters. The characters that are deleted are those characters at positions from start.pos up to, but not including end.pos. If end.pos is negative, then the the characters deleted are those from start.pos to the end of the text.

Usage

gtkEditableDeleteText(object, start.pos, end.pos)

Arguments

object
a GtkEditable

start.pos
start position

end.pos
end position

Details

Note that the positions are specified in characters, not bytes.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]