gtkEditableSelectRegion {RGtk2}	R Documentation

gtkEditableSelectRegion

Description

Selects a region of text. The characters that are selected are those characters at positions from start.pos up to, but not including end.pos. If end.pos is negative, then the the characters selected are those characters from start.pos to the end of the text.

Usage

gtkEditableSelectRegion(object, start, end)

Arguments

object
a GtkEditable

start
start of region

end
end of region

Details

Note that positions are specified in characters, not bytes.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]