gtkEditableSetEditable {RGtk2}	R Documentation

gtkEditableSetEditable

Description

Determines if the user can edit the text in the editable widget or not.

Usage

gtkEditableSetEditable(object, is.editable)

Arguments

object
a GtkEditable

is.editable
TRUE if the user is allowed to edit the text in the widget

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]