gtkEditableSetPosition {RGtk2}	R Documentation

gtkEditableSetPosition

Description

Sets the cursor position in the editable to the given value.

Usage

gtkEditableSetPosition(object, position)

Arguments

object
a GtkEditable

position
the position of the cursor

Details

The cursor is displayed before the character with the given (base 0) index in the contents of the editable. The value must be less than or equal to the number of characters in the editable. A value of -1 indicates that the position should be set after the last character of the editable. Note that position is in characters, not in bytes.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]