gtkEntryBufferDeleteText {RGtk2}	R Documentation

gtkEntryBufferDeleteText

Description

Deletes a sequence of characters from the buffer. n.chars characters are deleted starting at position. If n.chars is negative, then all characters until the end of the text are deleted.

Usage

gtkEntryBufferDeleteText(object, position, n.chars)

Arguments

object
a GtkEntryBuffer

position
position at which to delete text

n.chars
number of characters to delete

Details

If position or n.chars are out of bounds, then they are coerced to sane values.

Note that the positions are specified in characters, not bytes. Since 2.18

Value

[numeric] The number of characters deleted.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]