gtkEntryBufferInsertText {RGtk2}	R Documentation

gtkEntryBufferInsertText

Description

Inserts n.chars characters of chars into the contents of the buffer, at position position.

Usage

gtkEntryBufferInsertText(object, position, chars, n.chars)

Arguments

object
a GtkEntryBuffer

position
the position at which to insert text.

chars
the text to insert into the buffer.

n.chars
the length of the text in characters, or -1

Details

If n.chars is negative, then characters from chars will be inserted until a null-terminator is found. If position or n.chars are out of bounds, or the maximum buffer text length is exceeded, then they are coerced to sane values.

Note that the position and length are in characters, not in bytes. Since 2.18

Value

[numeric] The number of characters actually inserted.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]