gtkEntryBufferSetMaxLength {RGtk2}	R Documentation

gtkEntryBufferSetMaxLength

Description

Sets the maximum allowed length of the contents of the buffer. If the current contents are longer than the given length, then they will be truncated to fit.

Usage

gtkEntryBufferSetMaxLength(object, max.length)

Arguments

object
a GtkEntryBuffer

max.length
the maximum length of the entry buffer, or 0 for no maximum. (other than the maximum length of entries.) The value passed in will be clamped to the range 0-65536.

Details

Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]