gtkEntryCompletionSetMatchFunc {RGtk2}	R Documentation

gtkEntryCompletionSetMatchFunc

Description

Sets the match function for completion to be func. The match function is used to determine if a row should or should not be in the completion list.

Usage

gtkEntryCompletionSetMatchFunc(object, func, func.data = NULL)

Arguments

object
A GtkEntryCompletion.

func
The GtkEntryCompletionMatchFunc to use.

func.data
The user data for func.

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]