gtkEntryCompletionSetMinimumKeyLength {RGtk2}	R Documentation

gtkEntryCompletionSetMinimumKeyLength

Description

Requires the length of the search key for completion to be at least length. This is useful for long lists, where completing using a small key takes a lot of time and will come up with meaningless results anyway (ie, a too large dataset).

Usage

gtkEntryCompletionSetMinimumKeyLength(object, length)

Arguments

object
A GtkEntryCompletion.

length
The minimum length of the key in order to start completing.

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]