gtkEntryCompletionSetTextColumn {RGtk2}	R Documentation

gtkEntryCompletionSetTextColumn

Description

Convenience function for setting up the most used case of this code: a completion list with just strings. This function will set up completion to have a list displaying all (and just) strings in the completion list, and to get those strings from column in the model of completion.

Usage

gtkEntryCompletionSetTextColumn(object, column)

Arguments

object
A GtkEntryCompletion.

column
The column in the model of completion to get strings from.

Details

This functions creates and adds a GtkCellRendererText for the selected column. If you need to set the text column, but don't want the cell renderer, use gObjectSet to set the ::text_column property directly. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]