gtkEntryGetCurrentIconDragSource {RGtk2}	R Documentation

gtkEntryGetCurrentIconDragSource

Description

Returns the index of the icon which is the source of the current DND operation, or -1.

Usage

gtkEntryGetCurrentIconDragSource(object)

Arguments

object
a GtkIconEntry

Details

This function is meant to be used in a "drag-data-get" callback. Since 2.16

Value

[integer] index of the icon which is the source of the current DND operation, or -1.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]