gtkEntryGetIconAtPos {RGtk2}	R Documentation

gtkEntryGetIconAtPos

Description

Finds the icon at the given position and return its index. If x, y doesn't lie inside an icon, -1 is returned. This function is intended for use in a "query-tooltip" signal handler.

Usage

gtkEntryGetIconAtPos(object, x, y)

Arguments

object
a GtkEntry

x
the x coordinate of the position to find

y
the y coordinate of the position to find

Details

Since 2.16

Value

[integer] the index of the icon at the given position, or -1

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]