gtkEntryGetIconGicon {RGtk2}	R Documentation

gtkEntryGetIconGicon

Description

Retrieves the GIcon used for the icon, or NULL if there is no icon or if the icon was set by some other method (e.g., by stock, pixbuf, or icon name).

Usage

gtkEntryGetIconGicon(object, icon.pos)

Arguments

object
A GtkEntry

icon.pos
Icon position

Details

Since 2.16

Value

[GIcon] A GIcon, or NULL if no icon is set or if the icon is not a GIcon

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]