gtkEntryGetIconName {RGtk2}	R Documentation

gtkEntryGetIconName

Description

Retrieves the icon name used for the icon, or NULL if there is no icon or if the icon was set by some other method (e.g., by pixbuf, stock or gicon).

Usage

gtkEntryGetIconName(object, icon.pos)

Arguments

object
A GtkEntry

icon.pos
Icon position

Details

Since 2.16

Value

[character] An icon name, or NULL if no icon is set or if the icon wasn't set from an icon name

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]