gtkEntryGetIconPixbuf {RGtk2}	R Documentation

gtkEntryGetIconPixbuf

Description

Retrieves the image used for the icon.

Usage

gtkEntryGetIconPixbuf(object, icon.pos)

Arguments

object
A GtkEntry

icon.pos
Icon position

Details

Unlike the other methods of setting and getting icon data, this method will work regardless of whether the icon was set using a GdkPixbuf, a GIcon, a stock item, or an icon name. Since 2.16

Value

[GdkPixbuf] A GdkPixbuf, or NULL if no icon is set for this position.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]