gtkEntryGetIconWindow {RGtk2}	R Documentation

gtkEntryGetIconWindow

Description

Returns the GdkWindow which contains the entry's icon at icon.pos. This function is useful when drawing something to the entry in an expose-event callback because it enables the callback to distinguish between the text window and entry's icon windows.

Usage

gtkEntryGetIconWindow(object, icon.pos)

Arguments

object
A GtkEntry

icon.pos
Icon position

Details

See also gtkEntryGetTextWindow. Since 2.20

Value

[GdkWindow] the entry's icon window at icon.pos.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]