gtkEntryGetLayout {RGtk2}	R Documentation

gtkEntryGetLayout

Description

Gets the PangoLayout used to display the entry. The layout is useful to e.g. convert text positions to pixel positions, in combination with gtkEntryGetLayoutOffsets.

Usage

gtkEntryGetLayout(object)

Arguments

object
a GtkEntry

Details

Keep in mind that the layout text may contain a preedit string, so gtkEntryLayoutIndexToTextIndex and gtkEntryTextIndexToLayoutIndex are needed to convert byte indices in the layout to byte indices in the entry contents.

Value

[PangoLayout] the PangoLayout for this entry. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]