gtkEntryGetLayoutOffsets {RGtk2}	R Documentation

gtkEntryGetLayoutOffsets

Description

Obtains the position of the PangoLayout used to render text in the entry, in widget coordinates. Useful if you want to line up the text in an entry with some other text, e.g. when using the entry to implement editable cells in a sheet widget.

Usage

gtkEntryGetLayoutOffsets(object)

Arguments

object
a GtkEntry

Details

Also useful to convert mouse events into coordinates inside the PangoLayout, e.g. to take some action if some part of the entry text is clicked.

Note that as the user scrolls around in the entry the offsets will change; you'll need to connect to the "notify::scroll-offset" signal to track this. Remember when using the PangoLayout functions you need to convert to and from pixels using pangoPixels() or PANGO_SCALE.

Keep in mind that the layout text may contain a preedit string, so gtkEntryLayoutIndexToTextIndex and gtkEntryTextIndexToLayoutIndex are needed to convert byte indices in the layout to byte indices in the entry contents.

Value

A list containing the following elements:

x
location to store X offset of layout, or NULL. [ allow-none ]

y
location to store Y offset of layout, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]