gtkEntryGetTextWindow {RGtk2}	R Documentation

gtkEntryGetTextWindow

Description

Returns the GdkWindow which contains the text. This function is useful when drawing something to the entry in an expose-event callback because it enables the callback to distinguish between the text window and entry's icon windows.

Usage

gtkEntryGetTextWindow(object)

Arguments

object
a GtkEntry

Details

See also gtkEntryGetIconWindow. Since 2.20

Value

[GdkWindow] the entry's text window.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]