gtkEntrySelectRegion {RGtk2}	R Documentation

gtkEntrySelectRegion

Description

Selects a region of text. The characters that are selected are those characters at positions from start.pos up to, but not including end.pos. If end.pos is negative, then the the characters selected will be those characters from start.pos to the end of the text. WARNING: gtk_entry_select_region has been deprecated since version 2.0 and should not be used in newly-written code. Use gtkEditableSelectRegion instead.

Usage

gtkEntrySelectRegion(object, start, end)

Arguments

object
a GtkEntry

start
the starting position

end
the end position

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]