gtkEntrySetActivatesDefault {RGtk2}	R Documentation

gtkEntrySetActivatesDefault

Description

If setting is TRUE, pressing Enter in the entry will activate the default widget for the window containing the entry. This usually means that the dialog box containing the entry will be closed, since the default widget is usually one of the dialog buttons.

Usage

gtkEntrySetActivatesDefault(object, setting)

Arguments

object
a GtkEntry

setting
TRUE to activate window's default widget on Enter keypress

Details

(For experts: if setting is TRUE, the entry calls gtkWindowActivateDefault on the window containing the entry, in the default handler for the gtkWidgetActivate signal.)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]