gtkEntrySetAlignment {RGtk2}	R Documentation

gtkEntrySetAlignment

Description

Sets the alignment for the contents of the entry. This controls the horizontal positioning of the contents when the displayed text is shorter than the width of the entry.

Usage

gtkEntrySetAlignment(object, xalign)

Arguments

object
a GtkEntry

xalign
The horizontal alignment, from 0 (left) to 1 (right). Reversed for RTL layouts

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]