gtkEntrySetCursorHadjustment {RGtk2}	R Documentation

gtkEntrySetCursorHadjustment

Description

Hooks up an adjustment to the cursor position in an entry, so that when the cursor is moved, the adjustment is scrolled to show that position. See gtkScrolledWindowGetHadjustment for a typical way of obtaining the adjustment.

Usage

gtkEntrySetCursorHadjustment(object, adjustment)

Arguments

object
a GtkEntry

adjustment
an adjustment which should be adjusted when the cursor is moved, or NULL

Details

The adjustment has to be in pixel units and in the same coordinate system as the entry. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]