gtkEntrySetEditable {RGtk2}	R Documentation

gtkEntrySetEditable

Description

Determines if the user can edit the text in the editable widget or not. WARNING: gtk_entry_set_editable has been deprecated since version 2.0 and should not be used in newly-written code. Use gtkEditableSetEditable instead.

Usage

gtkEntrySetEditable(object, editable)

Arguments

object
a GtkEntry

editable
TRUE if the user is allowed to edit the text in the widget

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]