gtkEntrySetIconDragSource {RGtk2}	R Documentation

gtkEntrySetIconDragSource

Description

Sets up the icon at the given position so that GTK+ will start a drag operation when the user clicks and drags the icon.

Usage

gtkEntrySetIconDragSource(object, icon.pos, target.list, actions)

Arguments

object
a GtkIconEntry

icon.pos
icon position

target.list
the targets (data formats) in which the data can be provided

actions
a bitmask of the allowed drag actions

Details

To handle the drag operation, you need to connect to the usual "drag-data-get" (or possibly "drag-data-delete") signal, and use gtkEntryGetCurrentIconDragSource in your signal handler to find out if the drag was started from an icon.

By default, GTK+ uses the icon as the drag icon. You can use the "drag-begin" signal to set a different icon. Note that you have to use gSignalConnectAfter() to ensure that your signal handler gets executed after the default handler. Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]