gtkEntrySetIconFromIconName {RGtk2}	R Documentation

gtkEntrySetIconFromIconName

Description

Sets the icon shown in the entry at the specified position from the current icon theme.

Usage

gtkEntrySetIconFromIconName(object, icon.pos, icon.name = NULL)

Arguments

object
A GtkEntry

icon.pos
The position at which to set the icon

icon.name
An icon name, or NULL. [ allow-none ]

Details

If the icon name isn't known, a "broken image" icon will be displayed instead.

If icon.name is NULL, no icon will be shown in the specified position. Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]