gtkEntrySetIconTooltipText {RGtk2}	R Documentation

gtkEntrySetIconTooltipText

Description

Sets tooltip as the contents of the tooltip for the icon at the specified position.

Usage

gtkEntrySetIconTooltipText(object, icon.pos, tooltip = NULL)

Arguments

object
a GtkEntry

icon.pos
the icon position

tooltip
the contents of the tooltip for the icon, or NULL. [ allow-none ]

Details

Use NULL for tooltip to remove an existing tooltip.

See also gtkWidgetSetTooltipText and gtkEntrySetIconTooltipMarkup. Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]