gtkEntrySetInnerBorder {RGtk2}	R Documentation

gtkEntrySetInnerBorder

Description

Sets entry's inner-border property to border, or clears it if NULL is passed. The inner-border is the area around the entry's text, but inside its frame.

Usage

gtkEntrySetInnerBorder(object, border = NULL)

Arguments

object
a GtkEntry

border
a GtkBorder, or NULL. [ allow-none ]

Details

If set, this property overrides the inner-border style property. Overriding the style-provided border is useful when you want to do in-place editing of some text in a canvas or list widget, where pixel-exact positioning of the entry is important. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]