gtkEntrySetInvisibleChar {RGtk2}	R Documentation

gtkEntrySetInvisibleChar

Description

Sets the character to use in place of the actual text when gtkEntrySetVisibility has been called to set text visibility to FALSE. i.e. this is the character used in "password mode" to show the user how many characters have been typed. By default, GTK+ picks the best invisible char available in the current font. If you set the invisible char to 0, then the user will get no feedback at all; there will be no text on the screen as they type.

Usage

gtkEntrySetInvisibleChar(object, ch)

Arguments

object
a GtkEntry

ch
a Unicode character

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]