gtkEntrySetMaxLength {RGtk2}	R Documentation

gtkEntrySetMaxLength

Description

Sets the maximum allowed length of the contents of the widget. If the current contents are longer than the given length, then they will be truncated to fit.

Usage

gtkEntrySetMaxLength(object, max)

Arguments

object
a GtkEntry

max
the maximum length of the entry, or 0 for no maximum. (other than the maximum length of entries.) The value passed in will be clamped to the range 0-65536.

Details

This is equivalent to:

gtk_entry_buffer_set_max_length (gtk_entry_get_buffer (entry), max);

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]