gtkEntrySetPosition {RGtk2}	R Documentation

gtkEntrySetPosition

Description

Sets the cursor position in an entry to the given value. WARNING: gtk_entry_set_position has been deprecated since version 2.0 and should not be used in newly-written code. Use gtkEditableSetPosition instead.

Usage

gtkEntrySetPosition(object, position)

Arguments

object
a GtkEntry

position
the position of the cursor. The cursor is displayed before the character with the given (base 0) index in the widget. The value must be less than or equal to the number of characters in the widget. A value of -1 indicates that the position should be set after the last character in the entry. Note that this position is in characters, not in bytes.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]