gtkEntrySetVisibility {RGtk2}	R Documentation

gtkEntrySetVisibility

Description

Sets whether the contents of the entry are visible or not. When visibility is set to FALSE, characters are displayed as the invisible char, and will also appear that way when the text in the entry widget is copied elsewhere.

Usage

gtkEntrySetVisibility(object, visible)

Arguments

object
a GtkEntry

visible
TRUE if the contents of the entry are displayed as plaintext

Details

By default, GTK+ picks the best invisible character available in the current font, but it can be changed with gtkEntrySetInvisibleChar.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]