gtkEntrySetWidthChars {RGtk2}	R Documentation

gtkEntrySetWidthChars

Description

Changes the size request of the entry to be about the right size for n.chars characters. Note that it changes the size request, the size can still be affected by how you pack the widget into containers. If n.chars is -1, the size reverts to the default entry size.

Usage

gtkEntrySetWidthChars(object, n.chars)

Arguments

object
a GtkEntry

n.chars
width in chars

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]