gtkEventBoxSetVisibleWindow {RGtk2}	R Documentation

gtkEventBoxSetVisibleWindow

Description

Set whether the event box uses a visible or invisible child window. The default is to use visible windows.

Usage

gtkEventBoxSetVisibleWindow(object, visible.window)

Arguments

object
a GtkEventBox

visible.window
boolean value

Details

In an invisible window event box, the window that the event box creates is a GDK_INPUT_ONLY window, which means that it is invisible and only serves to receive events.

A visible window event box creates a visible (GDK_INPUT_OUTPUT) window that acts as the parent window for all the widgets contained in the event box.

You should generally make your event box invisible if you just want to trap events. Creating a visible window may cause artifacts that are visible to the user, especially if the user is using a theme with gradients or pixmaps.

The main reason to create a non input-only event box is if you want to set the background to a different color or draw on it. PLEASE NOTE: There is one unexpected issue for an invisible event box that has its window below the child. (See gtkEventBoxSetAboveChild.) Since the input-only window is not an ancestor window of any windows that descendent widgets of the event box create, events on these windows aren't propagated up by the windowing system, but only by GTK+. The practical effect of this is if an event isn't in the event mask for the descendant window (see gtkWidgetAddEvents), it won't be received by the event box.

This problem doesn't occur for visible event boxes, because in that case, the event box window is actually the ancestor of the descendant windows, not just at the same place on the screen. Since 2.4

Note

There is one unexpected issue for an invisible event box that has its window below the child. (See gtkEventBoxSetAboveChild.) Since the input-only window is not an ancestor window of any windows that descendent widgets of the event box create, events on these windows aren't propagated up by the windowing system, but only by GTK+. The practical effect of this is if an event isn't in the event mask for the descendant window (see gtkWidgetAddEvents), it won't be received by the event box.

This problem doesn't occur for visible event boxes, because in that case, the event box window is actually the ancestor of the descendant windows, not just at the same place on the screen.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]