gtkExit {RGtk2}	R Documentation

gtkExit

Description

Terminates the program and returns the given exit code to the caller. This function will shut down the GUI and free all resources allocated for GTK+. WARNING: gtk_exit is deprecated and should not be used in newly-written code. Use the standard exit() function instead.

Usage

gtkExit(error.code)

Arguments

error.code
Return value to pass to the caller. This is dependent on the target system but at least on Unix systems 0 means success.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]