gtkExpanderGetLabel {RGtk2}	R Documentation

gtkExpanderGetLabel

Description

Fetches the text from a label widget including any embedded underlines indicating mnemonics and Pango markup, as set by gtkExpanderSetLabel. If the label text has not been set the return value will be NULL. This will be the case if you create an empty button with gtkButtonNew to use as a container.

Usage

gtkExpanderGetLabel(object)

Arguments

object
a GtkExpander

Details

Note that this function behaved differently in versions prior to 2.14 and used to return the label text stripped of embedded underlines indicating mnemonics and Pango markup. This problem can be avoided by fetching the label text directly from the label widget. Since 2.4

Value

[character] The text of the label widget. This string is owned by the widget and must not be modified or freed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]