gtkExpanderSetExpanded {RGtk2}	R Documentation

gtkExpanderSetExpanded

Description

Sets the state of the expander. Set to TRUE, if you want the child widget to be revealed, and FALSE if you want the child widget to be hidden.

Usage

gtkExpanderSetExpanded(object, expanded)

Arguments

object
a GtkExpander

expanded
whether the child widget is revealed

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]