gtkFileChooserButtonNewWithDialog {RGtk2}	R Documentation

gtkFileChooserButtonNewWithDialog

Description

Creates a GtkFileChooserButton widget which uses dialog as its file-picking window.

Usage

gtkFileChooserButtonNewWithDialog(dialog)

Arguments

dialog
the widget to use as dialog

Details

Note that dialog must be a GtkDialog (or subclass) which implements the GtkFileChooser interface and must not have GTK_DIALOG_DESTROY_WITH_PARENT set.

Also note that the dialog needs to have its confirmative button added with response GTK_RESPONSE_ACCEPT or GTK_RESPONSE_OK in order for the button to take over the file selected in the dialog. Since 2.6

Value

[GtkWidget] a new button widget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]