gtkFileChooserGetCurrentFolder {RGtk2}	R Documentation

gtkFileChooserGetCurrentFolder

Description

Gets the current folder of chooser as a local filename. See gtkFileChooserSetCurrentFolder.

Usage

gtkFileChooserGetCurrentFolder(object)

Arguments

object
a GtkFileChooser

Details

Note that this is the folder that the file chooser is currently displaying (e.g. "/home/username/Documents"), which is not the same as the currently-selected folder if the chooser is in GTK_FILE_CHOOSER_SELECT_FOLDER mode (e.g. "/home/username/Documents/selected-folder/". To get the currently-selected folder in that mode, use gtkFileChooserGetUri as the usual way to get the selection. Since 2.4

Value

[character] the full path of the current folder, or NULL if the current path cannot be represented as a local filename. This function will also return NULL if the file chooser was unable to load the last folder that was requested from it; for example, as would be for calling gtkFileChooserSetCurrentFolder on a nonexistent folder.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]