gtkFileChooserGetFilename {RGtk2}	R Documentation

gtkFileChooserGetFilename

Description

Gets the filename for the currently selected file in the file selector. If multiple files are selected, one of the filenames will be returned at random.

Usage

gtkFileChooserGetFilename(object)

Arguments

object
a GtkFileChooser

Details

If the file chooser is in folder mode, this function returns the selected folder. Since 2.4

Value

[character] The currently selected filename, or NULL if no file is selected, or the selected file can't be represented with a local filename.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]