gtkFileChooserGetFilenames {RGtk2}	R Documentation

gtkFileChooserGetFilenames

Description

Lists all the selected files and subfolders in the current folder of chooser. The returned names are full absolute paths. If files in the current folder cannot be represented as local filenames they will be ignored. (See gtkFileChooserGetUris)

Usage

gtkFileChooserGetFilenames(object)

Arguments

object
a GtkFileChooser

Details

Since 2.4

Value

[list] a list containing the filenames of all selected files and subfolders in the current folder. [ element-type utf8][ transfer full utf8]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]