gtkFileChooserGetFiles {RGtk2}	R Documentation

gtkFileChooserGetFiles

Description

Lists all the selected files and subfolders in the current folder of chooser as GFile. An internal function, see gtkFileChooserGetUris.

Usage

gtkFileChooserGetFiles(object)

Arguments

object
a GtkFileChooser

Details

Since 2.14

Value

[list] a list containing a GFile for each selected file and subfolder in the current folder. [ element-type utf8][ transfer full ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]