gtkFileChooserGetUri {RGtk2}	R Documentation

gtkFileChooserGetUri

Description

Gets the URI for the currently selected file in the file selector. If multiple files are selected, one of the filenames will be returned at random.

Usage

gtkFileChooserGetUri(object)

Arguments

object
a GtkFileChooser

Details

If the file chooser is in folder mode, this function returns the selected folder. Since 2.4

Value

[character] The currently selected URI, or NULL if no file is selected.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]