gtkFileChooserGetUris {RGtk2}	R Documentation

gtkFileChooserGetUris

Description

Lists all the selected files and subfolders in the current folder of chooser. The returned names are full absolute URIs.

Usage

gtkFileChooserGetUris(object)

Arguments

object
a GtkFileChooser

Details

Since 2.4

Value

[list] a list containing the URIs of all selected files and subfolders in the current folder. [ element-type utf8][ transfer full ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]