gtkFileChooserRemoveShortcutFolderUri {RGtk2}	R Documentation

gtkFileChooserRemoveShortcutFolderUri

Description

Removes a folder URI from a file chooser's list of shortcut folders.

Usage

gtkFileChooserRemoveShortcutFolderUri(object, uri, .errwarn = TRUE)

Arguments

object
a GtkFileChooser

uri
URI of the folder to remove

.errwarn
Whether to issue a warning on error or fail silently

Details

Since 2.4

Value

A list containing the following elements:

retval
[logical] TRUE if the operation succeeds, FALSE otherwise. In the latter case, the error will be set as appropriate. See also: gtkFileChooserAddShortcutFolderUri

error
location to store error, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]