gtkFileChooserSelectFilename {RGtk2}	R Documentation

gtkFileChooserSelectFilename

Description

Selects a filename. If the file name isn't in the current folder of chooser, then the current folder of chooser will be changed to the folder containing filename.

Usage

gtkFileChooserSelectFilename(object, filename)

Arguments

object
a GtkFileChooser

filename
the filename to select

Details

Since 2.4

Value

[logical] TRUE if both the folder could be changed and the file was selected successfully, FALSE otherwise.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]