gtkFileChooserSelectUri {RGtk2}	R Documentation

gtkFileChooserSelectUri

Description

Selects the file to by uri. If the URI doesn't refer to a file in the current folder of chooser, then the current folder of chooser will be changed to the folder containing filename.

Usage

gtkFileChooserSelectUri(object, uri)

Arguments

object
a GtkFileChooser

uri
the URI to select

Details

Since 2.4

Value

[logical] TRUE if both the folder could be changed and the URI was selected successfully, FALSE otherwise.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]