gtkFileChooserSetAction {RGtk2}	R Documentation

gtkFileChooserSetAction

Description

Sets the type of operation that the chooser is performing; the user interface is adapted to suit the selected action. For example, an option to create a new folder might be shown if the action is GTK_FILE_CHOOSER_ACTION_SAVE but not if the action is GTK_FILE_CHOOSER_ACTION_OPEN.

Usage

gtkFileChooserSetAction(object, action)

Arguments

object
a GtkFileChooser

action
the action that the file selector is performing

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]