gtkFileChooserSetCurrentName {RGtk2}	R Documentation

gtkFileChooserSetCurrentName

Description

Sets the current name in the file selector, as if entered by the user. Note that the name passed in here is a UTF-8 string rather than a filename. This function is meant for such uses as a suggested name in a "Save As..." dialog.

Usage

gtkFileChooserSetCurrentName(object, name)

Arguments

object
a GtkFileChooser

name
the filename to use, as a UTF-8 string

Details

If you want to preselect a particular existing file, you should use gtkFileChooserSetFilename or gtkFileChooserSetUri instead. Please see the documentation for those functions for an example of using gtkFileChooserSetCurrentName as well. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]