gtkFileChooserSetDoOverwriteConfirmation {RGtk2}	R Documentation

gtkFileChooserSetDoOverwriteConfirmation

Description

Sets whether a file chooser in GTK_FILE_CHOOSER_ACTION_SAVE mode will present a confirmation dialog if the user types a file name that already exists. This is FALSE by default.

Usage

gtkFileChooserSetDoOverwriteConfirmation(object, do.overwrite.confirmation)

Arguments

object
a GtkFileChooser

do.overwrite.confirmation
whether to confirm overwriting in save mode

Details

Regardless of this setting, the chooser will emit the "confirm-overwrite" signal when appropriate.

If all you need is the stock confirmation dialog, set this property to TRUE. You can override the way confirmation is done by actually handling the "confirm-overwrite" signal; please refer to its documentation for the details. Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]