gtkFileChooserSetLocalOnly {RGtk2}	R Documentation

gtkFileChooserSetLocalOnly

Description

Sets whether only local files can be selected in the file selector. If local.only is TRUE (the default), then the selected file are files are guaranteed to be accessible through the operating systems native file file system and therefore the application only needs to worry about the filename functions in GtkFileChooser, like gtkFileChooserGetFilename, rather than the URI functions like gtkFileChooserGetUri,

Usage

gtkFileChooserSetLocalOnly(object, local.only)

Arguments

object
a GtkFileChooser

local.only
TRUE if only local files can be selected

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]