gtkFileChooserSetPreviewWidget {RGtk2}	R Documentation

gtkFileChooserSetPreviewWidget

Description

Sets an application-supplied widget to use to display a custom preview of the currently selected file. To implement a preview, after setting the preview widget, you connect to the "update-preview" signal, and call gtkFileChooserGetPreviewFilename or gtkFileChooserGetPreviewUri on each change. If you can display a preview of the new file, update your widget and set the preview active using gtkFileChooserSetPreviewWidgetActive. Otherwise, set the preview inactive.

Usage

gtkFileChooserSetPreviewWidget(object, preview.widget)

Arguments

object
a GtkFileChooser

preview.widget
widget for displaying preview.

Details

When there is no application-supplied preview widget, or the application-supplied preview widget is not active, the file chooser may display an internally generated preview of the current file or it may display no preview at all. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]