gtkFileChooserSetUsePreviewLabel {RGtk2}	R Documentation

gtkFileChooserSetUsePreviewLabel

Description

Sets whether the file chooser should display a stock label with the name of the file that is being previewed; the default is TRUE. Applications that want to draw the whole preview area themselves should set this to FALSE and display the name themselves in their preview widget.

Usage

gtkFileChooserSetUsePreviewLabel(object, use.label)

Arguments

object
a GtkFileChooser

use.label
whether to display a stock label with the name of the previewed file

Details

See also: gtkFileChooserSetPreviewWidget Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]