gtkFileChooserUnselectFile {RGtk2}	R Documentation

gtkFileChooserUnselectFile

Description

Unselects the file referred to by file. If the file is not in the current directory, does not exist, or is otherwise not currently selected, does nothing.

Usage

gtkFileChooserUnselectFile(object, file)

Arguments

object
a GtkFileChooser

file
a GFile

Details

Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]