gtkFileChooserWidgetNewWithBackend {RGtk2}	R Documentation

gtkFileChooserWidgetNewWithBackend

Description

Creates a new GtkFileChooserWidget with a specified backend. This is especially useful if you use gtkFileChooserSetLocalOnly to allow non-local files. This is a file chooser widget that can be embedded in custom windows and it is the same widget that is used by GtkFileChooserDialog. WARNING: gtk_file_chooser_widget_new_with_backend has been deprecated since version 2.14 and should not be used in newly-written code. Use gtkFileChooserWidgetNew instead.

Usage

gtkFileChooserWidgetNewWithBackend(action, backend, show = TRUE)

Arguments

action
Open or save mode for the widget

backend
The name of the specific filesystem backend to use.

Details

Since 2.4

Value

[GtkWidget] a new GtkFileChooserWidget

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]