gtkFileFilterFilter {RGtk2}	R Documentation

gtkFileFilterFilter

Description

Tests whether a file should be displayed according to filter. The GtkFileFilterInfo structure filter.info should include the fields returned from gtkFileFilterGetNeeded.

Usage

gtkFileFilterFilter(object, filter.info)

Arguments

object
a GtkFileFilter

filter.info
a GtkFileFilterInfo structure containing information about a file.

Details

This function will not typically be used by applications; it is intended principally for use in the implementation of GtkFileChooser. Since 2.4

Value

[logical] TRUE if the file should be displayed

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]