gtkFileFilterNew {RGtk2}	R Documentation

gtkFileFilterNew

Description

Creates a new GtkFileFilter with no rules added to it. Such a filter doesn't accept any files, so is not particularly useful until you add rules with gtkFileFilterAddMimeType, gtkFileFilterAddPattern, or gtkFileFilterAddCustom. To create a filter that accepts any file, use:

filter <- gtkFileFilter()
filter$addPattern("*")

Usage

gtkFileFilterNew()

Details

Since 2.4

Value

[GtkFileFilter] a new GtkFileFilter

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]