gtkFileSelectionComplete {RGtk2}	R Documentation

gtkFileSelectionComplete

Description

Will attempt to match pattern to a valid filenames or subdirectories in the current directory. If a match can be made, the matched filename will appear in the text entry field of the file selection dialog. If a partial match can be made, the "Files" list will contain those file names which have been partially matched, and the "Folders" list those directories which have been partially matched. WARNING: gtk_file_selection_complete is deprecated and should not be used in newly-written code.

Usage

gtkFileSelectionComplete(object, pattern)

Arguments

object
a GtkFileSelection.

pattern
a string of characters which may or may not match any filenames in the current directory.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]