gtkFileSelectionGetFilename {RGtk2}	R Documentation

gtkFileSelectionGetFilename

Description

This function returns the selected filename in the GLib file name encoding. To convert to UTF-8, call gFilenameToUtf8(). The returned string points to a statically allocated buffer and should be copied if you plan to keep it around. WARNING: gtk_file_selection_get_filename is deprecated and should not be used in newly-written code.

Usage

gtkFileSelectionGetFilename(object)

Arguments

object
a GtkFileSelection

Details

If no file is selected then the selected directory path is returned.

Value

[character] currently-selected filename in the on-disk encoding.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]