gtkFileSelectionGetSelectMultiple {RGtk2}	R Documentation

gtkFileSelectionGetSelectMultiple

Description

Determines whether or not the user is allowed to select multiple files in the file list. See gtkFileSelectionSetSelectMultiple. WARNING: gtk_file_selection_get_select_multiple is deprecated and should not be used in newly-written code.

Usage

gtkFileSelectionGetSelectMultiple(object)

Arguments

object
a GtkFileSelection

Value

[logical] TRUE if the user is allowed to select multiple files in the file list

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]