gtkFileSelectionHideFileopButtons {RGtk2}	R Documentation

gtkFileSelectionHideFileopButtons

Description

Hides the file operation buttons that normally appear at the top of the dialog. Useful if you wish to create a custom file selector, based on GtkFileSelection. WARNING: gtk_file_selection_hide_fileop_buttons is deprecated and should not be used in newly-written code.

Usage

gtkFileSelectionHideFileopButtons(object)

Arguments

object
a GtkFileSelection.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]