gtkFileSelectionSetFilename {RGtk2}	R Documentation

gtkFileSelectionSetFilename

Description

Sets a default path for the file requestor. If filename includes a directory path, then the requestor will open with that path as its current working directory. WARNING: gtk_file_selection_set_filename is deprecated and should not be used in newly-written code.

Usage

gtkFileSelectionSetFilename(object, filename)

Arguments

object
a GtkFileSelection.

filename
a string to set as the default file name.

Details

This has the consequence that in order to open the requestor with a working directory and an empty filename, filename must have a trailing directory separator.

The encoding of filename is preferred GLib file name encoding, which may not be UTF-8. See gFilenameFromUtf8().

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]