gtkFileSelectionShowFileopButtons {RGtk2}	R Documentation

gtkFileSelectionShowFileopButtons

Description

Shows the file operation buttons, if they have previously been hidden. The rest of the widgets in the dialog will be resized accordingly. WARNING: gtk_file_selection_show_fileop_buttons is deprecated and should not be used in newly-written code.

Usage

gtkFileSelectionShowFileopButtons(object)

Arguments

object
a GtkFileSelection.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]