gtkFixedSetHasWindow {RGtk2}	R Documentation

gtkFixedSetHasWindow

Description

Sets whether a GtkFixed widget is created with a separate GdkWindow for widget->window or not. (By default, it will be created with no separate GdkWindow). This function must be called while the GtkFixed is not realized, for instance, immediately after the window is created. WARNING: gtk_fixed_set_has_window has been deprecated since version 2.20 and should not be used in newly-written code. Use gtkWidgetSetHasWindow instead.

Usage

gtkFixedSetHasWindow(object, has.window)

Arguments

object
a GtkFixed

has.window
TRUE if a separate window should be created

Details

This function was added to provide an easy migration path for older applications which may expect GtkFixed to have a separate window.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]