gtkFontButtonGetFontName {RGtk2}	R Documentation

gtkFontButtonGetFontName

Description

Retrieves the name of the currently selected font. This name includes style and size information as well. If you want to render something with the font, use this string with pangoFontDescriptionFromString . If you're interested in peeking certain values (family name, style, size, weight) just query these properties from the PangoFontDescription object.

Usage

gtkFontButtonGetFontName(object)

Arguments

object
a GtkFontButton

Details

Since 2.4

Value

[character] an internal copy of the font name which must not be freed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]