gtkFontButtonSetFontName {RGtk2}	R Documentation

gtkFontButtonSetFontName

Description

Sets or updates the currently-displayed font in font picker dialog.

Usage

gtkFontButtonSetFontName(object, fontname)

Arguments

object
a GtkFontButton

fontname
Name of font to display in font selection dialog

Details

Since 2.4

Value

[logical] Return value of gtkFontSelectionDialogSetFontName if the font selection dialog exists, otherwise FALSE.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]