gtkFontButtonSetShowSize {RGtk2}	R Documentation

gtkFontButtonSetShowSize

Description

If show.size is TRUE, the font size will be displayed along with the name of the selected font.

Usage

gtkFontButtonSetShowSize(object, show.size)

Arguments

object
a GtkFontButton

show.size
TRUE if font size should be displayed in dialog.

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]