gtkFontSelectionGetFontName {RGtk2}	R Documentation

gtkFontSelectionGetFontName

Description

Gets the currently-selected font name.

Usage

gtkFontSelectionGetFontName(object)

Arguments

object
a GtkFontSelection

Details

Note that this can be a different string than what you set with gtkFontSelectionSetFontName, as the font selection widget may normalize font names and thus return a string with a different structure. For example, "Helvetica Italic Bold 12" could be normalized to "Helvetica Bold Italic 12". Use pangoFontDescriptionEqual if you want to compare two font descriptions.

Value

[character] A string with the name of the current font, or NULL if no font is selected.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]