gtkFontSelectionSetFontName {RGtk2}	R Documentation

gtkFontSelectionSetFontName

Description

Sets the currently-selected font.

Usage

gtkFontSelectionSetFontName(object, fontname)

Arguments

object
a GtkFontSelection

fontname
a font name like "Helvetica 12" or "Times Bold 18"

Details

Note that the fontsel needs to know the screen in which it will appear for this to work; this can be guaranteed by simply making sure that the fontsel is inserted in a toplevel window before you call this function.

Value

[logical] TRUE if the font could be set successfully; FALSE if no such font exists or if the fontsel doesn't belong to a particular screen yet.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]