gtkFrameSetLabelAlign {RGtk2}	R Documentation

gtkFrameSetLabelAlign

Description

Sets the alignment of the frame widget's label. The default values for a newly created frame are 0.0 and 0.5.

Usage

gtkFrameSetLabelAlign(object, xalign, yalign)

Arguments

object
a GtkFrame

xalign
The position of the label along the top edge of the widget. A value of 0.0 represents left alignment; 1.0 represents right alignment.

yalign
The y alignment of the label. A value of 0.0 aligns under the frame; 1.0 aligns above the frame. If the values are exactly 0.0 or 1.0 the gap in the frame won't be painted because the label will be completely above or below the frame.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]