gtkFrameSetLabelWidget {RGtk2}	R Documentation

gtkFrameSetLabelWidget

Description

Sets the label widget for the frame. This is the widget that will appear embedded in the top edge of the frame as a title.

Usage

gtkFrameSetLabelWidget(object, label.widget)

Arguments

object
a GtkFrame

label.widget
the new label widget

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]