gtkGetCurrentEvent {RGtk2}	R Documentation

gtkGetCurrentEvent

Description

Obtains a copy of the event currently being processed by GTK+. For example, if you get a "clicked" signal from GtkButton, the current event will be the GdkEventButton that triggered the "clicked" signal. If there is no current event, the function returns NULL.

Usage

gtkGetCurrentEvent()

Value

[GdkEvent] a copy of the current event, or NULL if no current event.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]