gtkGetCurrentEventState {RGtk2}	R Documentation

gtkGetCurrentEventState

Description

If there is a current event and it has a state field, place that state field in state and return TRUE, otherwise return FALSE.

Usage

gtkGetCurrentEventState()

Value

A list containing the following elements:

retval
[logical] TRUE if there was a current event and it had a state field

state
a location to store the state of the current event

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]