gtkGetDefaultLanguage {RGtk2}	R Documentation

gtkGetDefaultLanguage

Description

Returns the PangoLanguage for the default language currently in effect. (Note that this can change over the life of an application.) The default language is derived from the current locale. It determines, for example, whether GTK+ uses the right-to-left or left-to-right text direction.

Usage

gtkGetDefaultLanguage()

Details

This function is equivalent to pangoLanguageGetDefault. See that function for details.

Value

[PangoLanguage] the default language as a PangoLanguage, must not be freed

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]