gtkGetText {RGtk2}	R Documentation

Access contents of Gtk text widget

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

These functions allow one to retrieve or set the contents of a Gtk text widget.

Usage

gtkTextSetText(w, contents="", append = FALSE)
gtkTextGetText(w)
gtkTextClearText(w, start=0, end=-1)

Arguments

w
the GtkTextWidget whose contents are to be accessed

contents
a character vector (of length 1) giving the contents to be either appended or inserted into the text widget.

append
a logical value indicating whether the new text is to be appended to the existing contents or used to replace it. If this is FALSE, the text widget is currently cleared first.

start
locations in the text widget where the to start clearing the text. These are essentially character offsets into the string at present.

end
locations in the text widget where the to complete clearing the text. These are essentially character offsets into the string at present.

Note

THIS STUFF IS VERY OLD AND DEPRECATED (compatibility wrappers for RGtk 1)

This is an extra-ordinarily early release intended to encourage others to contribute code, etc.

Author(s)

Duncan Temple Lang

References

http://www.gtk.org http://www.omegahat.org/RGtk

[Package RGtk2 version 2.20.31 Index]