gtkGrabAdd {RGtk2}	R Documentation

gtkGrabAdd

Description

Makes widget the current grabbed widget. This means that interaction with other widgets in the same application is blocked and mouse as well as keyboard events are delivered to this widget.

Usage

gtkGrabAdd(object)

Arguments

object
The widget that grabs keyboard and pointer events.

Details

If widget is not sensitive, it is not set as the current grabbed widget and this function does nothing.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]