gtkHBoxNew {RGtk2}	R Documentation

gtkHBoxNew

Description

Creates a new GtkHBox.

Usage

gtkHBoxNew(homogeneous = NULL, spacing = NULL, show = TRUE)

Arguments

homogeneous
TRUE if all children are to be given equal space allotments.

spacing
the number of pixels to place by default between children.

Value

[GtkWidget] a new GtkHBox.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]