gtkHButtonBoxGetSpacingDefault {RGtk2}	R Documentation

gtkHButtonBoxGetSpacingDefault

Description

Retrieves the current default spacing for horizontal button boxes. This is the number of pixels to be placed between the buttons when they are arranged. WARNING: gtk_hbutton_box_get_spacing_default is deprecated and should not be used in newly-written code.

Usage

gtkHButtonBoxGetSpacingDefault()

Value

[integer] the default number of pixels between buttons.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]