gtkHSVIsAdjusting {RGtk2}	R Documentation

gtkHSVIsAdjusting

Description

An HSV color selector can be said to be adjusting if multiple rapid changes are being made to its value, for example, when the user is adjusting the value with the mouse. This function queries whether the HSV color selector is being adjusted or not.

Usage

gtkHSVIsAdjusting(object)

Arguments

object
A GtkHSV

Details

Since 2.14

Value

[logical] TRUE if clients can ignore changes to the color value, since they may be transitory, or FALSE if they should consider the color value status to be final.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]