gtkHandleBoxSetHandlePosition {RGtk2}	R Documentation

gtkHandleBoxSetHandlePosition

Description

Sets the side of the handlebox where the handle is drawn.

Usage

gtkHandleBoxSetHandlePosition(object, position)

Arguments

object
a GtkHandleBox

position
the side of the handlebox where the handle should be drawn.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]