gtkIMContextDeleteSurrounding {RGtk2}	R Documentation

gtkIMContextDeleteSurrounding

Description

Asks the widget that the input context is attached to to delete characters around the cursor position by emitting the GtkIMContext::delete_surrounding signal. Note that offset and n.chars are in characters not in bytes which differs from the usage other places in GtkIMContext.

Usage

gtkIMContextDeleteSurrounding(object, offset, n.chars)

Arguments

object
a GtkIMContext

offset
offset from cursor position in chars; a negative value means start before the cursor.

n.chars
number of characters to delete.

Details

In order to use this function, you should first call gtkIMContextGetSurrounding to get the current context, and call this function immediately afterwards to make sure that you know what you are deleting. You should also account for the fact that even if the signal was handled, the input context might not have deleted all the characters that were requested to be deleted.

This function is used by an input method that wants to make subsitutions in the existing text in response to new input. It is not useful for applications.

Value

[logical] TRUE if the signal was handled.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]