gtkIMContextFilterKeypress {RGtk2}	R Documentation

gtkIMContextFilterKeypress

Description

Allow an input method to internally handle key press and release events. If this function returns TRUE, then no further processing should be done for this key event.

Usage

gtkIMContextFilterKeypress(object, event)

Arguments

object
a GtkIMContext

event
the key event

Value

[logical] TRUE if the input method handled the key event.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]