gtkIMContextGetPreeditString {RGtk2}	R Documentation

gtkIMContextGetPreeditString

Description

Retrieve the current preedit string for the input context, and a list of attributes to apply to the string. This string should be displayed inserted at the insertion point.

Usage

gtkIMContextGetPreeditString(object)

Arguments

object
a GtkIMContext

Value

A list containing the following elements:

str
location to store the retrieved string. The string retrieved must be freed with gFree().

attrs
location to store the retrieved attribute list. When you are done with this list, you must unreference it with pangoAttrListUnref().

cursor.pos
location to store position of cursor (in characters) within the preedit string.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]