gtkIMContextReset {RGtk2}	R Documentation

gtkIMContextReset

Description

Notify the input method that a change such as a change in cursor position has been made. This will typically cause the input method to clear the preedit state.

Usage

gtkIMContextReset(object)

Arguments

object
a GtkIMContext

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]