gtkIMContextSetClientWindow {RGtk2}	R Documentation

gtkIMContextSetClientWindow

Description

Set the client window for the input context; this is the GdkWindow in which the input appears. This window is used in order to correctly position status windows, and may also be used for purposes internal to the input method.

Usage

gtkIMContextSetClientWindow(object, window)

Arguments

object
a GtkIMContext

window
the client window. This may be NULL to indicate that the previous client window no longer exists. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]