gtkIMContextSetSurrounding {RGtk2}	R Documentation

gtkIMContextSetSurrounding

Description

Sets surrounding context around the insertion point and preedit string. This function is expected to be called in response to the GtkIMContext::retrieve_surrounding signal, and will likely have no effect if called at other times.

Usage

gtkIMContextSetSurrounding(object, text, cursor.index)

Arguments

object
a GtkIMContext

text
text surrounding the insertion point, as UTF-8. the preedit string should not be included within text.

cursor.index
the byte index of the insertion cursor within text.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]