gtkIMContextSimpleAddTable {RGtk2}	R Documentation

gtkIMContextSimpleAddTable

Description

Adds an additional table to search to the input context. Each row of the table consists of max.seq.len key symbols followed by two integer interpreted as the high and low words of a gunicode value. Tables are searched starting from the last added.

Usage

gtkIMContextSimpleAddTable(object, data, max.seq.len, n.seqs)

Arguments

object
A GtkIMContextSimple

data
the table

max.seq.len
Maximum length of a sequence in the table (cannot be greater than GTK_MAX_COMPOSE_LEN)

n.seqs
number of sequences in the table

Details

The table must be sorted in dictionary order on the numeric value of the key symbol fields. (Values beyond the length of the sequence should be zero.)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]