gtkIMMulticontextSetContextId {RGtk2}	R Documentation

gtkIMMulticontextSetContextId

Description

Sets the context id for context.

Usage

gtkIMMulticontextSetContextId(object, context.id)

Arguments

object
a GtkIMMulticontext

context.id
the id to use

Details

This causes the currently active slave of context to be replaced by the slave corresponding to the new context id. Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]