gtkIconFactoryAdd {RGtk2}	R Documentation

gtkIconFactoryAdd

Description

Adds the given icon.set to the icon factory, under the name stock.id. stock.id should be namespaced for your application, e.g. "myapp-whatever-icon". Normally applications create a GtkIconFactory, then add it to the list of default factories with gtkIconFactoryAddDefault. Then they pass the stock.id to widgets such as GtkImage to display the icon. Themes can provide an icon with the same name (such as "myapp-whatever-icon") to override your application's default icons. If an icon already existed in factory for stock.id, it is unreferenced and replaced with the new icon.set.

Usage

gtkIconFactoryAdd(object, stock.id, icon.set)

Arguments

object
a GtkIconFactory

stock.id
icon name

icon.set
icon set

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]