gtkIconFactoryAddDefault {RGtk2}	R Documentation

gtkIconFactoryAddDefault

Description

Adds an icon factory to the list of icon factories searched by gtkStyleLookupIconSet. This means that, for example, gtkImageNewFromStock will be able to find icons in factory. There will normally be an icon factory added for each library or application that comes with icons. The default icon factories can be overridden by themes.

Usage

gtkIconFactoryAddDefault(object)

Arguments

object
a GtkIconFactory

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]