gtkIconFactoryRemoveDefault {RGtk2}	R Documentation

gtkIconFactoryRemoveDefault

Description

Removes an icon factory from the list of default icon factories. Not normally used; you might use it for a library that can be unloaded or shut down.

Usage

gtkIconFactoryRemoveDefault(object)

Arguments

object
a GtkIconFactory previously added with gtkIconFactoryAddDefault

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]