gtkIconInfoGetAttachPoints {RGtk2}	R Documentation

gtkIconInfoGetAttachPoints

Description

Fetches the set of attach points for an icon. An attach point is a location in the icon that can be used as anchor points for attaching emblems or overlays to the icon.

Usage

gtkIconInfoGetAttachPoints(object)

Arguments

object
a GtkIconInfo

Details

Since 2.4

Value

A list containing the following elements:

retval
[logical] TRUE if there are any attach points for the icon.

points
(array length=n_points) (out): location to store pointer to a list of points, or NULL free the list of points with gFree(). [ allow-none ]

n.points
location to store the number of points in points, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]