gtkIconInfoGetBaseSize {RGtk2}	R Documentation

gtkIconInfoGetBaseSize

Description

Gets the base size for the icon. The base size is a size for the icon that was specified by the icon theme creator. This may be different than the actual size of image; an example of this is small emblem icons that can be attached to a larger icon. These icons will be given the same base size as the larger icons to which they are attached.

Usage

gtkIconInfoGetBaseSize(object)

Arguments

object
a GtkIconInfo

Details

Since 2.4

Value

[integer] the base size, or 0, if no base size is known for the icon.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]