gtkIconInfoGetBuiltinPixbuf {RGtk2}	R Documentation

gtkIconInfoGetBuiltinPixbuf

Description

Gets the built-in image for this icon, if any. To allow GTK+ to use built in icon images, you must pass the GTK_ICON_LOOKUP_USE_BUILTIN to gtkIconThemeLookupIcon.

Usage

gtkIconInfoGetBuiltinPixbuf(object)

Arguments

object
a GtkIconInfo structure

Details

Since 2.4

Value

[GdkPixbuf] the built-in image pixbuf, or NULL. No extra reference is added to the returned pixbuf, so if you want to keep it around, The returned image must not be modified. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]