gtkIconInfoGetDisplayName {RGtk2}	R Documentation

gtkIconInfoGetDisplayName

Description

Gets the display name for an icon. A display name is a string to be used in place of the icon name in a user visible context like a list of icons.

Usage

gtkIconInfoGetDisplayName(object)

Arguments

object
a GtkIconInfo

Details

Since 2.4

Value

[character] the display name for the icon or NULL, if the icon doesn't have a specified display name. This value is owned icon.info and must not be modified or free.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]