gtkIconInfoGetEmbeddedRect {RGtk2}	R Documentation

gtkIconInfoGetEmbeddedRect

Description

Gets the coordinates of a rectangle within the icon that can be used for display of information such as a preview of the contents of a text file. See gtkIconInfoSetRawCoordinates for further information about the coordinate system.

Usage

gtkIconInfoGetEmbeddedRect(object)

Arguments

object
a GtkIconInfo

Details

Since 2.4

Value

A list containing the following elements:

retval
[logical] TRUE if the icon has an embedded rectangle

rectangle
GdkRectangle in which to store embedded rectangle coordinates; coordinates are only stored when this function returns TRUE.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]