gtkIconInfoLoadIcon {RGtk2}	R Documentation

gtkIconInfoLoadIcon

Description

Renders an icon previously looked up in an icon theme using gtkIconThemeLookupIcon; the size will be based on the size passed to gtkIconThemeLookupIcon. Note that the resulting pixbuf may not be exactly this size; an icon theme may have icons that differ slightly from their nominal sizes, and in addition GTK+ will avoid scaling icons that it considers sufficiently close to the requested size or for which the source image would have to be scaled up too far. (This maintains sharpness.). This behaviour can be changed by passing the GTK_ICON_LOOKUP_FORCE_SIZE flag when obtaining the GtkIconInfo. If this flag has been specified, the pixbuf returned by this function will be scaled to the exact size.

Usage

gtkIconInfoLoadIcon(object, .errwarn = TRUE)

Arguments

object
a GtkIconInfo structure from gtkIconThemeLookupIcon

.errwarn
Whether to issue a warning on error or fail silently

Details

Since 2.4

Value

A list containing the following elements:

retval
[GdkPixbuf] the rendered icon; this may be a newly created icon or a new reference to an internal icon, so you must not modify the icon.

error
location to store error information on failure, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]