gtkIconInfoSetRawCoordinates {RGtk2}	R Documentation

gtkIconInfoSetRawCoordinates

Description

Sets whether the coordinates returned by gtkIconInfoGetEmbeddedRect and gtkIconInfoGetAttachPoints should be returned in their original form as specified in the icon theme, instead of scaled appropriately for the pixbuf returned by gtkIconInfoLoadIcon.

Usage

gtkIconInfoSetRawCoordinates(object, raw.coordinates)

Arguments

object
a GtkIconInfo

raw.coordinates
whether the coordinates of embedded rectangles and attached points should be returned in their original (unscaled) form.

Details

Raw coordinates are somewhat strange; they are specified to be with respect to the unscaled pixmap for PNG and XPM icons, but for SVG icons, they are in a 1000x1000 coordinate space that is scaled to the final size of the icon. You can determine if the icon is an SVG icon by using gtkIconInfoGetFilename, and seeing if it is non-NULL and ends in '.svg'.

This function is provided primarily to allow compatibility wrappers for older API's, and is not expected to be useful for applications. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]