gtkIconSetNew {RGtk2}	R Documentation

gtkIconSetNew

Description

Creates a new GtkIconSet. A GtkIconSet represents a single icon in various sizes and widget states. It can provide a GdkPixbuf for a given size and state on request, and automatically caches some of the rendered GdkPixbuf objects.

Usage

gtkIconSetNew()

Details

Normally you would use gtkWidgetRenderIcon instead of using GtkIconSet directly. The one case where you'd use GtkIconSet is to create application-specific icon sets to place in a GtkIconFactory.

Value

[GtkIconSet] a new GtkIconSet

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]