gtkIconSetNewFromPixbuf {RGtk2}	R Documentation

gtkIconSetNewFromPixbuf

Description

Creates a new GtkIconSet with pixbuf as the default/fallback source image. If you don't add any additional GtkIconSource to the icon set, all variants of the icon will be created from pixbuf, using scaling, pixelation, etc. as required to adjust the icon size or make the icon look insensitive/prelighted.

Usage

gtkIconSetNewFromPixbuf(pixbuf)

Arguments

pixbuf
a GdkPixbuf

Value

[GtkIconSet] a new GtkIconSet

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]