gtkIconSetRenderIcon {RGtk2}	R Documentation

gtkIconSetRenderIcon

Description

Renders an icon using gtkStyleRenderIcon. In most cases, gtkWidgetRenderIcon is better, since it automatically provides most of the arguments from the current widget settings. This function never returns NULL; if the icon can't be rendered (perhaps because an image file fails to load), a default "missing image" icon will be returned instead.

Usage

gtkIconSetRenderIcon(object, style, direction, state, size, widget = NULL, 
    detail = NULL)

Arguments

object
a GtkIconSet

style
a GtkStyle associated with widget, or NULL. [ allow-none ]

direction
text direction

state
widget state

size
icon size. A size of (GtkIconSize)-1 means render at the size of the source and don't scale. [ type int]

widget
widget that will display the icon, or NULL. The only use that is typically made of this is to determine the appropriate GdkScreen. [ allow-none ]

detail
detail to pass to the theme engine, or NULL. Note that passing a detail of anything but NULL will disable caching. [ allow-none ]

Value

[GdkPixbuf] a GdkPixbuf to be displayed

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]