gtkIconSizeGetName {RGtk2}	R Documentation

gtkIconSizeGetName

Description

Gets the canonical name of the given icon size. The returned string is statically allocated and should not be freed.

Usage

gtkIconSizeGetName(size)

Arguments

size
a GtkIconSize. [ type int]

Value

[character] the name of the given icon size.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]