gtkIconSizeLookup {RGtk2}	R Documentation

gtkIconSizeLookup

Description

Obtains the pixel size of a semantic icon size, possibly modified by user preferences for the default GtkSettings. (See gtkIconSizeLookupForSettings.) Normally size would be GTK_ICON_SIZE_MENU, GTK_ICON_SIZE_BUTTON, etc. This function isn't normally needed, gtkWidgetRenderIcon is the usual way to get an icon for rendering, then just look at the size of the rendered pixbuf. The rendered pixbuf may not even correspond to the width/height returned by gtkIconSizeLookup, because themes are free to render the pixbuf however they like, including changing the usual size.

Usage

gtkIconSizeLookup(size)

Arguments

size
an icon size. [ type int]

Value

A list containing the following elements:

retval
[logical] TRUE if size was a valid size

width
location to store icon width

height
location to store icon height

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]