gtkIconSourceGetFilename {RGtk2}	R Documentation

gtkIconSourceGetFilename

Description

Retrieves the source filename, or NULL if none is set. The filename is not a copy, and should not be modified or expected to persist beyond the lifetime of the icon source.

Usage

gtkIconSourceGetFilename(object)

Arguments

object
a GtkIconSource

Value

[character] image filename. This string must not be modified or freed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]