gtkIconSourceGetPixbuf {RGtk2}	R Documentation

gtkIconSourceGetPixbuf

Description

Retrieves the source pixbuf, or NULL if none is set. In addition, if a filename source is in use, this function in some cases will return the pixbuf from loaded from the filename. This is, for example, true for the GtkIconSource passed to the GtkStyle::renderIcon() virtual function. The reference count on the pixbuf is not incremented.

Usage

gtkIconSourceGetPixbuf(object)

Arguments

object
a GtkIconSource

Value

[GdkPixbuf] source pixbuf

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]