gtkIconSourceSetDirection {RGtk2}	R Documentation

gtkIconSourceSetDirection

Description

Sets the text direction this icon source is intended to be used with.

Usage

gtkIconSourceSetDirection(object, direction)

Arguments

object
a GtkIconSource

direction
text direction this source applies to

Details

Setting the text direction on an icon source makes no difference if the text direction is wildcarded. Therefore, you should usually call gtkIconSourceSetDirectionWildcarded to un-wildcard it in addition to calling this function.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]