gtkIconSourceSetDirectionWildcarded {RGtk2}	R Documentation

gtkIconSourceSetDirectionWildcarded

Description

If the text direction is wildcarded, this source can be used as the base image for an icon in any GtkTextDirection. If the text direction is not wildcarded, then the text direction the icon source applies to should be set with gtkIconSourceSetDirection, and the icon source will only be used with that text direction.

Usage

gtkIconSourceSetDirectionWildcarded(object, setting)

Arguments

object
a GtkIconSource

setting
TRUE to wildcard the text direction

Details

GtkIconSet prefers non-wildcarded sources (exact matches) over wildcarded sources, and will use an exact match when possible.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]