gtkIconSourceSetSize {RGtk2}	R Documentation

gtkIconSourceSetSize

Description

Sets the icon size this icon source is intended to be used with.

Usage

gtkIconSourceSetSize(object, size)

Arguments

object
a GtkIconSource

size
icon size this source applies to. [ type int]

Details

Setting the icon size on an icon source makes no difference if the size is wildcarded. Therefore, you should usually call gtkIconSourceSetSizeWildcarded to un-wildcard it in addition to calling this function.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]