gtkIconSourceSetState {RGtk2}	R Documentation

gtkIconSourceSetState

Description

Sets the widget state this icon source is intended to be used with.

Usage

gtkIconSourceSetState(object, state)

Arguments

object
a GtkIconSource

state
widget state this source applies to

Details

Setting the widget state on an icon source makes no difference if the state is wildcarded. Therefore, you should usually call gtkIconSourceSetStateWildcarded to un-wildcard it in addition to calling this function.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]