gtkIconSourceSetStateWildcarded {RGtk2}	R Documentation

gtkIconSourceSetStateWildcarded

Description

If the widget state is wildcarded, this source can be used as the base image for an icon in any GtkStateType. If the widget state is not wildcarded, then the state the source applies to should be set with gtkIconSourceSetState and the icon source will only be used with that specific state.

Usage

gtkIconSourceSetStateWildcarded(object, setting)

Arguments

object
a GtkIconSource

setting
TRUE to wildcard the widget state

Details

GtkIconSet prefers non-wildcarded sources (exact matches) over wildcarded sources, and will use an exact match when possible. GtkIconSet will normally transform wildcarded source images to produce an appropriate icon for a given state, for example lightening an image on prelight, but will not modify source images that match exactly.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]