gtkIconThemeAddBuiltinIcon {RGtk2}	R Documentation

gtkIconThemeAddBuiltinIcon

Description

Registers a built-in icon for icon theme lookups. The idea of built-in icons is to allow an application or library that uses themed icons to function requiring files to be present in the file system. For instance, the default images for all of GTK+'s stock icons are registered as built-icons.

Usage

gtkIconThemeAddBuiltinIcon(icon.name, size, pixbuf)

Arguments

icon.name
the name of the icon to register

size
the size at which to register the icon (different images can be registered for the same icon name at different sizes.)

pixbuf
GdkPixbuf that contains the image to use for icon.name.

Details

In general, if you use gtkIconThemeAddBuiltinIcon you should also install the icon in the icon theme, so that the icon is generally available.

This function will generally be used with pixbufs loaded via gdkPixbufNewFromInline(). Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]