gtkIconThemeChooseIcon {RGtk2}	R Documentation

gtkIconThemeChooseIcon

Description

Looks up a named icon and returns a structure containing information such as the filename of the icon. The icon can then be rendered into a pixbuf using gtkIconInfoLoadIcon. (gtkIconThemeLoadIcon combines these two steps if all you need is the pixbuf.)

Usage

gtkIconThemeChooseIcon(object, icon.names, size, flags)

Arguments

object
a GtkIconTheme

icon.names
array of icon names to lookup

size
desired icon size

flags
flags modifying the behavior of the icon lookup

Details

If icon.names contains more than one name, this function tries them all in the given order before falling back to inherited icon themes. Since 2.12

Value

[GtkIconInfo] a GtkIconInfo structure containing information about the icon, or NULL if the icon wasn't found.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]