gtkIconThemeGetForScreen {RGtk2}	R Documentation

gtkIconThemeGetForScreen

Description

Gets the icon theme object associated with screen; if this function has not previously been called for the given screen, a new icon theme object will be created and associated with the screen. Icon theme objects are fairly expensive to create, so using this function is usually a better choice than calling than gtkIconThemeNew and setting the screen yourself; by using this function a single icon theme object will be shared between users.

Usage

gtkIconThemeGetForScreen(screen)

Arguments

screen
a GdkScreen

Details

Since 2.4

Value

[GtkIconTheme] A unique GtkIconTheme associated with the given screen. This icon theme is associated with the screen and can be used as long as the screen is open. Do not ref or unref it. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]