gtkIconThemeGetSearchPath {RGtk2}	R Documentation

gtkIconThemeGetSearchPath

Description

Gets the current search path. See gtkIconThemeSetSearchPath.

Usage

gtkIconThemeGetSearchPath(object)

Arguments

object
a GtkIconTheme

Details

Since 2.4

Value

A list containing the following elements:

path
location to store a list of icon theme path directories or NULL The stored value should be freed with gStrfreev(). [ allow-none ][ out ]

n.elements
location to store number of elements in path, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]