gtkIconThemeListContexts {RGtk2}	R Documentation

gtkIconThemeListContexts

Description

Gets the list of contexts available within the current hierarchy of icon themes

Usage

gtkIconThemeListContexts(object)

Arguments

object
a GtkIconTheme

Details

Since 2.12

Value

[list] a list list holding the names of all the contexts in the theme. [ element-type utf8][ transfer full ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]