gtkIconThemeListIcons {RGtk2}	R Documentation

gtkIconThemeListIcons

Description

Lists the icons in the current icon theme. Only a subset of the icons can be listed by providing a context string. The set of values for the context string is system dependent, but will typically include such values as "Applications" and "MimeTypes".

Usage

gtkIconThemeListIcons(object, context = NULL)

Arguments

object
a GtkIconTheme

context
a string identifying a particular type of icon, or NULL to list all icons.

Details

Since 2.4

Value

[list] a list list holding the names of all the icons in the theme. [ element-type utf8][ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]