gtkIconThemeLoadIcon {RGtk2}	R Documentation

gtkIconThemeLoadIcon

Description

Looks up an icon in an icon theme, scales it to the given size and renders it into a pixbuf. This is a convenience function; if more details about the icon are needed, use gtkIconThemeLookupIcon followed by gtkIconInfoLoadIcon.

Usage

gtkIconThemeLoadIcon(object, icon.name, size, flags, .errwarn = TRUE)

Arguments

object
a GtkIconTheme

icon.name
the name of the icon to lookup

size
the desired icon size. The resulting icon may not be exactly this size; see gtkIconInfoLoadIcon.

flags
flags modifying the behavior of the icon lookup

.errwarn
Whether to issue a warning on error or fail silently

Details

Note that you probably want to listen for icon theme changes and update the icon. This is usually done by connecting to the GtkWidget::style-set signal. If for some reason you do not want to update the icon when the icon theme changes, you should consider using gdkPixbufCopy to make a private copy of the pixbuf returned by this function. Otherwise GTK+ may need to keep the old icon theme loaded, which would be a waste of memory. Since 2.4

Value

A list containing the following elements:

retval
[GdkPixbuf] the rendered icon; this may be a newly created icon or a new reference to an internal icon, so you must not modify the icon. NULL if the icon isn't found.

error
Location to store error information on failure, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]