gtkIconThemeNew {RGtk2}	R Documentation

gtkIconThemeNew

Description

Creates a new icon theme object. Icon theme objects are used to lookup up an icon by name in a particular icon theme. Usually, you'll want to use gtkIconThemeGetDefault or gtkIconThemeGetForScreen rather than creating a new icon theme object for scratch.

Usage

gtkIconThemeNew()

Details

Since 2.4

Value

[GtkIconTheme] the newly created GtkIconTheme object.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]