gtkIconThemeRescanIfNeeded {RGtk2}	R Documentation

gtkIconThemeRescanIfNeeded

Description

Checks to see if the icon theme has changed; if it has, any currently cached information is discarded and will be reloaded next time icon.theme is accessed.

Usage

gtkIconThemeRescanIfNeeded(object)

Arguments

object
a GtkIconTheme

Details

Since 2.4

Value

[logical] TRUE if the icon theme has changed and needed to be reloaded.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]