gtkIconViewEnableModelDragDest {RGtk2}	R Documentation

gtkIconViewEnableModelDragDest

Description

Turns icon.view into a drop destination for automatic DND. Calling this method sets "reorderable" to FALSE.

Usage

gtkIconViewEnableModelDragDest(object, targets, actions)

Arguments

object
a GtkIconView

targets
the table of targets that the drag will support

actions
the bitmask of possible actions for a drag to this widget

Details

Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]