gtkIconViewGetDestItemAtPos {RGtk2}	R Documentation

gtkIconViewGetDestItemAtPos

Description

Determines the destination item for a given position.

Usage

gtkIconViewGetDestItemAtPos(object, drag.x, drag.y)

Arguments

object
a GtkIconView

drag.x
the position to determine the destination item for

drag.y
the position to determine the destination item for

Details

Since 2.8

Value

A list containing the following elements:

retval
[logical] whether there is an item at the given position.

path
Return location for the path of the item, or NULL. [ allow-none ]

pos
Return location for the drop position, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]