gtkIconViewGetItemAtPos {RGtk2}	R Documentation

gtkIconViewGetItemAtPos

Description

Finds the path at the point (x, y), relative to bin_window coordinates. In contrast to gtkIconViewGetPathAtPos, this function also obtains the cell at the specified position. See gtkIconViewConvertWidgetToBinWindowCoords for converting widget coordinates to bin_window coordinates.

Usage

gtkIconViewGetItemAtPos(object, x, y)

Arguments

object
A GtkIconView.

x
The x position to be identified

y
The y position to be identified

Details

Since 2.8

Value

A list containing the following elements:

retval
[logical] TRUE if an item exists at the specified position

path
Return location for the path, or NULL. [ allow-none ]

cell
Return location for the renderer responsible for the cell at (x, y), or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]