gtkIconViewGetTooltipContext {RGtk2}	R Documentation

gtkIconViewGetTooltipContext

Description

This function is supposed to be used in a "query-tooltip" signal handler for GtkIconView. The x, y and keyboard.tip values which are received in the signal handler, should be passed to this function without modification.

Usage

gtkIconViewGetTooltipContext(object, x, y, keyboard.tip)

Arguments

object
an GtkIconView

x
the x coordinate (relative to widget coordinates)

y
the y coordinate (relative to widget coordinates)

keyboard.tip
whether this is a keyboard tooltip or not

Details

The return value indicates whether there is an icon view item at the given coordinates (TRUE) or not (FALSE) for mouse tooltips. For keyboard tooltips the item returned will be the cursor item. When TRUE, then any of model, path and iter which have been provided will be set to point to that row and the corresponding model. x and y will always be converted to be relative to icon.view's bin_window if keyboard.tooltip is FALSE. Since 2.12

Value

A list containing the following elements:

retval
[logical] whether or not the given tooltip context points to a item

model
a pointer to receive a GtkTreeModel or NULL. [ allow-none ]

path
a pointer to receive a GtkTreePath or NULL. [ allow-none ]

iter
a pointer to receive a GtkTreeIter or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]