gtkIconViewGetVisibleRange {RGtk2}	R Documentation

gtkIconViewGetVisibleRange

Description

Sets start.path and end.path to be the first and last visible path. Note that there may be invisible paths in between.

Usage

gtkIconViewGetVisibleRange(object)

Arguments

object
A GtkIconView

Details

Since 2.8

Value

A list containing the following elements:

retval
[logical] TRUE, if valid paths were placed in start.path and end.path

start.path
Return location for start of region, or NULL. [ allow-none ]

end.path
Return location for end of region, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]