gtkIconViewScrollToPath {RGtk2}	R Documentation

gtkIconViewScrollToPath

Description

Moves the alignments of icon.view to the position specified by path. row.align determines where the row is placed, and col.align determines where column is placed. Both are expected to be between 0.0 and 1.0. 0.0 means left/top alignment, 1.0 means right/bottom alignment, 0.5 means center.

Usage

gtkIconViewScrollToPath(object, path, use.align, row.align, col.align)

Arguments

object
A GtkIconView.

path
The path of the item to move to.

use.align
whether to use alignment arguments, or FALSE.

row.align
The vertical alignment of the item specified by path.

col.align
The horizontal alignment of the item specified by path.

Details

If use.align is FALSE, then the alignment arguments are ignored, and the tree does the minimum amount of work to scroll the item onto the screen. This means that the item will be scrolled to the edge closest to its current position. If the item is currently visible on the screen, nothing is done.

This function only works if the model is set, and path is a valid row on the model. If the model changes before the icon.view is realized, the centered path will be modified to reflect this change. Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]