gtkIconViewSetCursor {RGtk2}	R Documentation

gtkIconViewSetCursor

Description

Sets the current keyboard focus to be at path, and selects it. This is useful when you want to focus the user's attention on a particular item. If cell is not NULL, then focus is given to the cell specified by it. Additionally, if start.editing is TRUE, then editing should be started in the specified cell.

Usage

gtkIconViewSetCursor(object, path, cell, start.editing)

Arguments

object
A GtkIconView

path
A GtkTreePath

cell
One of the cell renderers of icon.view, or NULL. [ allow-none ]

start.editing
TRUE if the specified cell should start being edited.

Details

This function is often followed by gtk_widget_grab_focus (icon_view) in order to give keyboard focus to the widget. Please note that editing can only happen when the widget is realized. Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]