gtkIconViewSetItemWidth {RGtk2}	R Documentation

gtkIconViewSetItemWidth

Description

Sets the ::item-width property which specifies the width to use for each item. If it is set to -1, the icon view will automatically determine a suitable item size.

Usage

gtkIconViewSetItemWidth(object, item.width)

Arguments

object
a GtkIconView

item.width
the width for each item

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]