gtkIconViewSetMarkupColumn {RGtk2}	R Documentation

gtkIconViewSetMarkupColumn

Description

Sets the column with markup information for icon.view to be column. The markup column must be of type G_TYPE_STRING. If the markup column is set to something, it overrides the text column set by gtkIconViewSetTextColumn.

Usage

gtkIconViewSetMarkupColumn(object, column)

Arguments

object
A GtkIconView.

column
A column in the currently used model, or -1 to display no text

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]