gtkIconViewSetTooltipCell {RGtk2}	R Documentation

gtkIconViewSetTooltipCell

Description

Sets the tip area of tooltip to the area which cell occupies in the item pointed to by path. See also gtkTooltipSetTipArea.

Usage

gtkIconViewSetTooltipCell(object, tooltip, path, cell)

Arguments

object
a GtkIconView

tooltip
a GtkTooltip

path
a GtkTreePath

cell
a GtkCellRenderer or NULL. [ allow-none ]

Details

See also gtkIconViewSetTooltipColumn for a simpler alternative. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]