gtkIconViewSetTooltipColumn {RGtk2}	R Documentation

gtkIconViewSetTooltipColumn

Description

If you only plan to have simple (text-only) tooltips on full items, you can use this function to have GtkIconView handle these automatically for you. column should be set to the column in icon.view's model containing the tooltip texts, or -1 to disable this feature.

Usage

gtkIconViewSetTooltipColumn(object, column)

Arguments

object
a GtkIconView

column
an integer, which is a valid column number for icon.view's model

Details

When enabled, "has-tooltip" will be set to TRUE and icon.view will connect a "query-tooltip" signal handler. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]