gtkIdleAdd {RGtk2}	R Documentation

gtkIdleAdd

Description

Causes the mainloop to call the given function whenever no events with higher priority are to be processed. The default priority is GTK_PRIORITY_DEFAULT, which is rather low. WARNING: gtk_idle_add has been deprecated since version 2.4 and should not be used in newly-written code. Use gIdleAdd instead.

Usage

gtkIdleAdd(fun, data = NULL)

Arguments

data
The information to pass to the function.

Value

[numeric] a unique handle for this registration.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]