gtkIdleAddPriority {RGtk2}	R Documentation

gtkIdleAddPriority

Description

Like gtkIdleAdd this function allows you to have a function called when the event loop is idle. The difference is that you can give a priority different from GTK_PRIORITY_DEFAULT to the idle function. WARNING: gtk_idle_add_priority has been deprecated since version 2.4 and should not be used in newly-written code. Use gIdleAddFull() instead.

Usage

gtkIdleAddPriority(priority, fun, data = NULL)

Arguments

priority
The priority which should not be above G_PRIORITY_HIGH_IDLE. Note that you will interfere with GTK+ if you use a priority above GTK_PRIORITY_RESIZE.

data
Data to pass to that function.

Value

[numeric] A unique id for the event source.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]