gtkImageGetAnimation {RGtk2}	R Documentation

gtkImageGetAnimation

Description

Gets the GdkPixbufAnimation being displayed by the GtkImage. The storage type of the image must be GTK_IMAGE_EMPTY or GTK_IMAGE_ANIMATION (see gtkImageGetStorageType). The caller of this function does not own a reference to the returned animation.

Usage

gtkImageGetAnimation(object)

Arguments

object
a GtkImage

Value

[GdkPixbufAnimation] the displayed animation, or NULL if the image is empty. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]