gtkImageGetIconSet {RGtk2}	R Documentation

gtkImageGetIconSet

Description

Gets the icon set and size being displayed by the GtkImage. The storage type of the image must be GTK_IMAGE_EMPTY or GTK_IMAGE_ICON_SET (see gtkImageGetStorageType).

Usage

gtkImageGetIconSet(object)

Arguments

object
a GtkImage

Value

A list containing the following elements:

icon.set
location to store a GtkIconSet, or NULL. [ out ][ transfer none ][ allow-none ]

size
location to store a stock icon size, or NULL. [ out ][ allow-none ][ type int]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]