gtkImageGetImage {RGtk2}	R Documentation

gtkImageGetImage

Description

Gets the GdkImage and mask being displayed by the GtkImage. The storage type of the image must be GTK_IMAGE_EMPTY or GTK_IMAGE_IMAGE (see gtkImageGetStorageType). The caller of this function does not own a reference to the returned image and mask.

Usage

gtkImageGetImage(object)

Arguments

object
a GtkImage

Value

A list containing the following elements:

gdk.image
return location for a GtkImage, or NULL. [ out ][ transfer none ][ allow-none ]

mask
return location for a GdkBitmap, or NULL. [ out ][ transfer none ][ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]