gtkImageMenuItemNewFromStock {RGtk2}	R Documentation

gtkImageMenuItemNewFromStock

Description

Creates a new GtkImageMenuItem containing the image and text from a stock item. Some stock ids have preprocessor functions like GTK_STOCK_OK and GTK_STOCK_APPLY.

Usage

gtkImageMenuItemNewFromStock(stock.id, accel.group, show = TRUE)

Arguments

stock.id
the name of the stock item.

accel.group
the GtkAccelGroup to add the menu items accelerator to, or NULL.

Details

If you want this menu item to have changeable accelerators, then pass in NULL for accel_group. Next call gtkMenuItemSetAccelPath with an appropriate path for the menu item, use gtkStockLookup to look up the standard accelerator for the stock item, and if one is found, call gtkAccelMapAddEntry to register it.

Value

[GtkWidget] a new GtkImageMenuItem.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]