gtkImageMenuItemSetAccelGroup {RGtk2}	R Documentation

gtkImageMenuItemSetAccelGroup

Description

Specifies an accel.group to add the menu items accelerator to (this only applies to stock items so a stock item must already be set, make sure to call gtkImageMenuItemSetUseStock and gtkMenuItemSetLabel with a valid stock item first).

Usage

gtkImageMenuItemSetAccelGroup(object, accel.group)

Arguments

object
a GtkImageMenuItem

accel.group
the GtkAccelGroup

Details

If you want this menu item to have changeable accelerators then you shouldnt need this (see gtkImageMenuItemNewFromStock). Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]