gtkImageMenuItemSetAlwaysShowImage {RGtk2}	R Documentation

gtkImageMenuItemSetAlwaysShowImage

Description

If TRUE, the menu item will ignore the "gtk-menu-images" setting and always show the image, if available.

Usage

gtkImageMenuItemSetAlwaysShowImage(object, always.show)

Arguments

object
a GtkImageMenuItem

always.show
TRUE if the menuitem should always show the image

Details

Use this property if the menuitem would be useless or hard to use without the image. Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]