gtkImageMenuItemSetImage {RGtk2}	R Documentation

gtkImageMenuItemSetImage

Description

Sets the image of image.menu.item to the given widget. Note that it depends on the show-menu-images setting whether the image will be displayed or not.

Usage

gtkImageMenuItemSetImage(object, image = NULL)

Arguments

object
a GtkImageMenuItem.

image
a widget to set as the image for the menu item. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]