gtkImageNewFromAnimation {RGtk2}	R Documentation

gtkImageNewFromAnimation

Description

Creates a GtkImage displaying the given animation. The GtkImage does not assume a reference to the animation; you still need to unref it if you own references. GtkImage will add its own reference rather than adopting yours.

Usage

gtkImageNewFromAnimation(animation, show = TRUE)

Arguments

animation
an animation

Details

Note that the animation frames are shown using a timeout with G_PRIORITY_DEFAULT. When using animations to indicate busyness, keep in mind that the animation will only be shown if the main loop is not busy with something that has a higher priority.

Value

[GtkWidget] a new GtkImage widget

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]