gtkImageNewFromFile {RGtk2}	R Documentation

gtkImageNewFromFile

Description

Creates a new GtkImage displaying the file filename. If the file isn't found or can't be loaded, the resulting GtkImage will display a "broken image" icon. This function never returns NULL, it always returns a valid GtkImage widget.

Usage

gtkImageNewFromFile(filename, show = TRUE)

Arguments

filename
a filename

Details

If the file contains an animation, the image will contain an animation.

If you need to detect failures to load the file, use gdkPixbufNewFromFile to load the file yourself, then create the GtkImage from the pixbuf. (Or for animations, use gdkPixbufAnimationNewFromFile).

The storage type (gtkImageGetStorageType) of the returned image is not defined, it will be whatever is appropriate for displaying the file.

Value

[GtkWidget] a new GtkImage

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]