gtkImageNewFromGicon {RGtk2}	R Documentation

gtkImageNewFromGicon

Description

Creates a GtkImage displaying an icon from the current icon theme. If the icon name isn't known, a "broken image" icon will be displayed instead. If the current icon theme is changed, the icon will be updated appropriately.

Usage

gtkImageNewFromGicon(icon, size, show = TRUE)

Arguments

icon
an icon

size
a stock icon size. [ type int]

Details

Since 2.14

Value

[GtkWidget] a new GtkImage displaying the themed icon

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]