gtkImageNewFromIconSet {RGtk2}	R Documentation

gtkImageNewFromIconSet

Description

Creates a GtkImage displaying an icon set. Sample stock sizes are GTK_ICON_SIZE_MENU, GTK_ICON_SIZE_SMALL_TOOLBAR. Instead of using this function, usually it's better to create a GtkIconFactory, put your icon sets in the icon factory, add the icon factory to the list of default factories with gtkIconFactoryAddDefault, and then use gtkImageNewFromStock. This will allow themes to override the icon you ship with your application.

Usage

gtkImageNewFromIconSet(icon.set, size, show = TRUE)

Arguments

icon.set
a GtkIconSet

size
a stock icon size. [ type int]

Details

The GtkImage does not assume a reference to the icon set; you still need to unref it if you own references. GtkImage will add its own reference rather than adopting yours.

Value

[GtkWidget] a new GtkImage

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]