gtkImageNewFromImage {RGtk2}	R Documentation

gtkImageNewFromImage

Description

Creates a GtkImage widget displaying a image with a mask. A GdkImage is a client-side image buffer in the pixel format of the current display. The GtkImage does not assume a reference to the image or mask; you still need to unref them if you own references. GtkImage will add its own reference rather than adopting yours.

Usage

gtkImageNewFromImage(image = NULL, mask = NULL, show = TRUE)

Arguments

image
a GdkImage, or NULL. [ allow-none ]

mask
a GdkBitmap, or NULL. [ allow-none ]

Value

[GtkWidget] a new GtkImage

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]