gtkImageNewFromPixmap {RGtk2}	R Documentation

gtkImageNewFromPixmap

Description

Creates a GtkImage widget displaying pixmap with a mask. A GdkPixmap is a server-side image buffer in the pixel format of the current display. The GtkImage does not assume a reference to the pixmap or mask; you still need to unref them if you own references. GtkImage will add its own reference rather than adopting yours.

Usage

gtkImageNewFromPixmap(pixmap = NULL, mask = NULL, show = TRUE)

Arguments

pixmap
a GdkPixmap, or NULL. [ allow-none ]

mask
a GdkBitmap, or NULL. [ allow-none ]

Value

[GtkWidget] a new GtkImage

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]