gtkInfoBarAddActionWidget {RGtk2}	R Documentation

gtkInfoBarAddActionWidget

Description

Add an activatable widget to the action area of a GtkInfoBar, connecting a signal handler that will emit the gtkInfoBarResponse signal on the message area when the widget is activated. The widget is appended to the end of the message areas action area.

Usage

gtkInfoBarAddActionWidget(object, child, response.id)

Arguments

object
a GtkInfoBar

child
an activatable widget

response.id
response ID for child

Details

Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]