gtkInfoBarAddButton {RGtk2}	R Documentation

gtkInfoBarAddButton

Description

Adds a button with the given text (or a stock button, if button_text is a stock ID) and sets things up so that clicking the button will emit the "response" signal with the given response_id. The button is appended to the end of the info bars's action area. The button widget is returned, but usually you don't need it.

Usage

gtkInfoBarAddButton(object, button.text, response.id)

Arguments

object
a GtkInfoBar

button.text
text of button, or stock ID

response.id
response ID for the button

Details

Since 2.18

Value

[GtkWidget] the button widget that was added

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]