gtkInfoBarNewWithButtons {RGtk2}	R Documentation

gtkInfoBarNewWithButtons

Description

Creates a new GtkInfoBar with buttons. Button text/response ID pairs should be listed, with a NULL pointer ending the list. Button text can be either a stock ID such as GTK_STOCK_OK, or some arbitrary text. A response ID can be any positive number, or one of the values in the GtkResponseType enumeration. If the user clicks one of these dialog buttons, GtkInfoBar will emit the "response" signal with the corresponding response ID.

Usage

gtkInfoBarNewWithButtons(first.button.text, ..., show = TRUE)

Arguments

first.button.text
stock ID or text to go in first button, or NULL. [ allow-none ]

...
response ID for first button, then additional buttons, ending with NULL

Value

[GtkWidget] a new GtkInfoBar

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]