gtkInit {RGtk2}	R Documentation

gtkInit

Description

Call this function before using any other GTK+ functions in your GUI applications. It will initialize everything needed to operate the toolkit and parses some standard command line options. argc and argv are adjusted accordingly so your own code will never see those standard arguments.

Usage

gtkInit(args = "R")

Details

Note that there are some alternative ways to initialize GTK+: if you are calling gtkParseArgs(), gtkInitCheck(), gtkInitWithArgs() or gOptionContextParse() with the option group returned by gtkGetOptionGroup(), you don't have to call gtkInit. PLEASE NOTE: This function will terminate your program if it was unable to initialize the GUI for some reason. If you want your program to fall back to a textual interface you want to call gtkInitCheck() instead. PLEASE NOTE: Since 2.18, GTK+ calls signal (SIGPIPE, SIG_IGN) during initialization, to ignore SIGPIPE signals, since these are almost never wanted in graphical applications. If you do need to handle SIGPIPE for some reason, reset the handler after gtkInit, but notice that other libraries (e.g. libdbus or gvfs) might do similar things.

Note

This function will terminate your program if it was unable to initialize the GUI for some reason. If you want your program to fall back to a textual interface you want to call gtkInitCheck() instead. Since 2.18, GTK+ calls signal (SIGPIPE, SIG_IGN) during initialization, to ignore SIGPIPE signals, since these are almost never wanted in graphical applications. If you do need to handle SIGPIPE for some reason, reset the handler after gtkInit, but notice that other libraries (e.g. libdbus or gvfs) might do similar things.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]