gtkItemFactoryAddForeign {RGtk2}	R Documentation

gtkItemFactoryAddForeign

Description

Installs an accelerator for accel.widget in accel.group, that causes the ::activate signal to be emitted if the accelerator is activated. WARNING: gtk_item_factory_add_foreign has been deprecated since version 2.4 and should not be used in newly-written code. The recommended API for this purpose are the functions gtkMenuItemSetAccelPath and gtkWidgetSetAccelPath; don't use gtkItemFactoryAddForeign in new code, since it is likely to be removed in the future.

Usage

gtkItemFactoryAddForeign(accel.widget, full.path, accel.group, 
    keyval, modifiers)

Arguments

accel.widget
widget to install an accelerator on

full.path
the full path for the accel.widget

accel.group
the accelerator group to install the accelerator in

keyval
key value of the accelerator

modifiers
modifier combination of the accelerator

Details

This function can be used to make widgets participate in the accel saving/restoring functionality provided by gtkAccelMapSave and gtkAccelMapLoad, even if they haven't been created by an item factory.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]