gtkItemFactoryConstruct {RGtk2}	R Documentation

gtkItemFactoryConstruct

Description

Initializes an item factory. WARNING: gtk_item_factory_construct has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkUIManager instead.

Usage

gtkItemFactoryConstruct(object, container.type, path, accel.group)

Arguments

object
a GtkItemFactory

container.type
the kind of menu to create; can be GTK_TYPE_MENU_BAR, GTK_TYPE_MENU or GTK_TYPE_OPTION_MENU

path
the factory path of ifactory, a string of the form "<name>"

accel.group
a GtkAccelGroup to which the accelerators for the menu items will be added, or NULL to create a new one

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]