gtkItemFactoryCreateItems {RGtk2}	R Documentation

gtkItemFactoryCreateItems

Description

Creates the menu items from the entries. WARNING: gtk_item_factory_create_items has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkUIManager instead.

Usage

gtkItemFactoryCreateItems(object, entries, callback.data = NULL)

Arguments

object
a GtkItemFactory

entries
a list of GtkItemFactoryEntrys whose callback members must by of type GtkItemFactoryCallback1

callback.data
data passed to the callback functions of all entries

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]