gtkItemFactoryFromPath {RGtk2}	R Documentation

gtkItemFactoryFromPath

Description

Finds an item factory which has been constructed using the "<name>" prefix of path as the path argument for gtkItemFactoryNew. WARNING: gtk_item_factory_from_path has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkUIManager instead.

Usage

gtkItemFactoryFromPath(path)

Arguments

path
a string starting with a factory path of the form "<name>"

Value

[GtkItemFactory] the GtkItemFactory created for the given factory path, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]