gtkItemFactoryGetItem {RGtk2}	R Documentation

gtkItemFactoryGetItem

Description

Obtains the menu item which corresponds to path. WARNING: gtk_item_factory_get_item has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkUIManager instead.

Usage

gtkItemFactoryGetItem(object, path)

Arguments

object
a GtkItemFactory

path
the path to the menu item

Details

If the widget corresponding to path is a menu item which opens a submenu, then the item is returned. If you are interested in the submenu, use gtkItemFactoryGetWidget instead.

Value

[GtkWidget] the menu item for the given path, or NULL if path doesn't lead to a menu item. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]