gtkItemFactoryGetItemByAction {RGtk2}	R Documentation

gtkItemFactoryGetItemByAction

Description

Obtains the menu item which was constructed from the first GtkItemFactoryEntry with the given action. WARNING: gtk_item_factory_get_item_by_action has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkUIManager instead.

Usage

gtkItemFactoryGetItemByAction(object, action)

Arguments

object
a GtkItemFactory

action
an action as specified in the callback.action field of GtkItemFactoryEntry

Value

[GtkWidget] the menu item which corresponds to the given action, or NULL if no menu item was found. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]