gtkItemFactoryNew {RGtk2}	R Documentation

gtkItemFactoryNew

Description

Creates a new GtkItemFactory. WARNING: gtk_item_factory_new has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkUIManager instead.

Usage

gtkItemFactoryNew(container.type, path, accel.group = NULL)

Arguments

container.type
the kind of menu to create; can be GTK_TYPE_MENU_BAR, GTK_TYPE_MENU or GTK_TYPE_OPTION_MENU

path
the factory path of the new item factory, a string of the form "<name>"

accel.group
a GtkAccelGroup to which the accelerators for the menu items will be added, or NULL to create a new one. [ allow-none ]

Details

Beware that the returned object does not have a floating reference.

Value

[GtkItemFactory] a new GtkItemFactory

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]