gtkItemFactoryPathFromWidget {RGtk2}	R Documentation

gtkItemFactoryPathFromWidget

Description

If widget has been created by an item factory, returns the full path to it. (The full path of a widget is the concatenation of the factory path specified in gtkItemFactoryNew with the path specified in the GtkItemFactoryEntry from which the widget was created.) WARNING: gtk_item_factory_path_from_widget has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkUIManager instead.

Usage

gtkItemFactoryPathFromWidget(widget)

Arguments

widget
a widget

Value

[character] the full path to widget if it has been created by an item factory, NULL otherwise.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]