gtkItemFactoryPopupDataFromWidget {RGtk2}	R Documentation

gtkItemFactoryPopupDataFromWidget

Description

Obtains the popup.data which was passed to gtkItemFactoryPopupWithData. This data is available until the menu is popped down again. WARNING: gtk_item_factory_popup_data_from_widget has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkUIManager instead.

Usage

gtkItemFactoryPopupDataFromWidget(widget)

Arguments

widget
a widget

Value

[R object] popup.data associated with the item factory from which widget was created, or NULL if widget wasn't created by an item factory

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]