gtkItemFactoryPopupWithData {RGtk2}	R Documentation

gtkItemFactoryPopupWithData

Description

Pops up the menu constructed from the item factory at (x, y). Callbacks can access the popup.data while the menu is posted via gtkItemFactoryPopupData and gtkItemFactoryPopupDataFromWidget. WARNING: gtk_item_factory_popup_with_data has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkUIManager instead.

Usage

gtkItemFactoryPopupWithData(object, popup.data, x, y, mouse.button, 
    time = "GDK_CURRENT_TIME")

Arguments

object
a GtkItemFactory of type GTK_TYPE_MENU (see gtkItemFactoryNew)

popup.data
data available for callbacks while the menu is posted

x
the x position

y
the y position

mouse.button
the mouse button which was pressed to initiate the popup

time
the time at which the activation event occurred

Details

The mouse.button parameter should be the mouse button pressed to initiate the menu popup. If the menu popup was initiated by something other than a mouse button press, such as a mouse button release or a keypress, mouse.button should be 0.

The time. parameter should be the time stamp of the event that initiated the popup. If such an event is not available, use gtkGetCurrentEventTime instead.

The operation of the mouse.button and the time. parameters is the same as the button and activation.time parameters for gtkMenuPopup.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]