gtkItemFactorySetTranslateFunc {RGtk2}	R Documentation

gtkItemFactorySetTranslateFunc

Description

Sets a function to be used for translating the path elements before they are displayed. WARNING: gtk_item_factory_set_translate_func has been deprecated since version 2.4 and should not be used in newly-written code. Use GtkUIManager instead.

Usage

gtkItemFactorySetTranslateFunc(object, func, data = NULL)

Arguments

object
a GtkItemFactory

func
the GtkTranslateFunc function to be used to translate path elements

data
data to pass to func and notify

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]