gtkKeySnooperInstall {RGtk2}	R Documentation

gtkKeySnooperInstall

Description

Installs a key snooper function, which will get called on all key events before delivering them normally.

Usage

gtkKeySnooperInstall(snooper, func.data = NULL)

Arguments

snooper
a GtkKeySnoopFunc.

func.data
data to pass to snooper.

Value

[numeric] a unique id for this key snooper for use with gtkKeySnooperRemove.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]