gtkLabelGet {RGtk2}	R Documentation

gtkLabelGet

Description

Gets the current string of text within the GtkLabel and writes it to the given str argument. It does not make a copy of this string so you must not write to it. WARNING: gtk_label_get is deprecated and should not be used in newly-written code. Use gtkLabelGetText instead.

Usage

gtkLabelGet(object)

Arguments

object
The GtkLabel widget you want to get the text from.

Value

A list containing the following elements:

str
The reference to the pointer you want to point to the text.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]