gtkLabelGetAttributes {RGtk2}	R Documentation

gtkLabelGetAttributes

Description

Gets the attribute list that was set on the label using gtkLabelSetAttributes, if any. This function does not reflect attributes that come from the labels markup (see gtkLabelSetMarkup). If you want to get the effective attributes for the label, use pango_layout_get_attribute (gtk_label_get_layout (label)).

Usage

gtkLabelGetAttributes(object)

Arguments

object
a GtkLabel

Value

[PangoAttrList] the attribute list, or NULL if none was set.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]