gtkLabelGetCurrentUri {RGtk2}	R Documentation

gtkLabelGetCurrentUri

Description

Returns the URI for the currently active link in the label. The active link is the one under the mouse pointer or, in a selectable label, the link in which the text cursor is currently positioned.

Usage

gtkLabelGetCurrentUri(object)

Arguments

object
a GtkLabel

Details

This function is intended for use in a "activate-link" handler or for use in a "query-tooltip" handler. Since 2.18

Value

[character] the currently active URI.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]