gtkLabelGetLayoutOffsets {RGtk2}	R Documentation

gtkLabelGetLayoutOffsets

Description

Obtains the coordinates where the label will draw the PangoLayout representing the text in the label; useful to convert mouse events into coordinates inside the PangoLayout, e.g. to take some action if some part of the label is clicked. Of course you will need to create a GtkEventBox to receive the events, and pack the label inside it, since labels are a GTK_NO_WINDOW widget. Remember when using the PangoLayout functions you need to convert to and from pixels using pangoPixels() or PANGO_SCALE.

Usage

gtkLabelGetLayoutOffsets(object)

Arguments

object
a GtkLabel

Value

A list containing the following elements:

x
location to store X offset of layout, or NULL. [ allow-none ]

y
location to store Y offset of layout, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]