gtkLabelGetMnemonicKeyval {RGtk2}	R Documentation

gtkLabelGetMnemonicKeyval

Description

If the label has been set so that it has an mnemonic key this function returns the keyval used for the mnemonic accelerator. If there is no mnemonic set up it returns GDK_VoidSymbol.

Usage

gtkLabelGetMnemonicKeyval(object)

Arguments

object
a GtkLabel

Value

[numeric] GDK keyval usable for accelerators, or GDK_VoidSymbol

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]