gtkLabelGetText {RGtk2}	R Documentation

gtkLabelGetText

Description

Fetches the text from a label widget, as displayed on the screen. This does not include any embedded underlines indicating mnemonics or Pango markup. (See gtkLabelGetLabel)

Usage

gtkLabelGetText(object)

Arguments

object
a GtkLabel

Value

[character] the text in the label widget. This is the internal string used by the label, and must not be modified.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]