gtkLabelNewWithMnemonic {RGtk2}	R Documentation

gtkLabelNewWithMnemonic

Description

Creates a new GtkLabel, containing the text in str.

Usage

gtkLabelNewWithMnemonic(str = NULL, show = TRUE)

Arguments

str
The text of the label, with an underscore in front of the mnemonic character

Details

If characters in str are preceded by an underscore, they are underlined. If you need a literal underscore character in a label, use '__' (two underscores). The first underlined character represents a keyboard accelerator called a mnemonic. The mnemonic key can be used to activate another widget, chosen automatically, or explicitly using gtkLabelSetMnemonicWidget.

If gtkLabelSetMnemonicWidget is not called, then the first activatable ancestor of the GtkLabel will be chosen as the mnemonic widget. For instance, if the label is inside a button or menu item, the button or menu item will automatically become the mnemonic widget and be activated by the mnemonic.

Value

[GtkWidget] the new GtkLabel

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]