gtkLabelParseUline {RGtk2}	R Documentation

gtkLabelParseUline

Description

Parses the given string for underscores and converts the next character to an underlined character. The last character that was underlined will have its lower-cased accelerator keyval returned (i.e. "_File" would return the keyval for "f". This is probably only used within the GTK+ library itself for menu items and such. WARNING: gtk_label_parse_uline is deprecated and should not be used in newly-written code. Use gtkLabelSetUseUnderline instead.

Usage

gtkLabelParseUline(object, string)

Arguments

object
The GtkLabel you want to affect.

string
The string you want to parse for underlines.

Value

[numeric] The lowercase keyval of the last character underlined.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]