gtkLabelSetAngle {RGtk2}	R Documentation

gtkLabelSetAngle

Description

Sets the angle of rotation for the label. An angle of 90 reads from from bottom to top, an angle of 270, from top to bottom. The angle setting for the label is ignored if the label is selectable, wrapped, or ellipsized.

Usage

gtkLabelSetAngle(object, angle)

Arguments

object
a GtkLabel

angle
the angle that the baseline of the label makes with the horizontal, in degrees, measured counterclockwise

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]