gtkLabelSetAttributes {RGtk2}	R Documentation

gtkLabelSetAttributes

Description

Sets a PangoAttrList; the attributes in the list are applied to the label text.

Usage

gtkLabelSetAttributes(object, attrs)

Arguments

object
a GtkLabel

attrs
a PangoAttrList

Details

PLEASE NOTE: The attributes set with this function will be applied and merged with any other attributes previously effected by way of the "use-underline" or "use-markup" properties. While it is not recommended to mix markup strings with manually set attributes, if you must; know that the attributes will be applied to the label after the markup string is parsed.

Note

The attributes set with this function will be applied and merged with any other attributes previously effected by way of the "use-underline" or "use-markup" properties. While it is not recommended to mix markup strings with manually set attributes, if you must; know that the attributes will be applied to the label after the markup string is parsed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]