gtkLabelSetJustify {RGtk2}	R Documentation

gtkLabelSetJustify

Description

Sets the alignment of the lines in the text of the label relative to each other. GTK_JUSTIFY_LEFT is the default value when the widget is first created with gtkLabelNew. If you instead want to set the alignment of the label as a whole, use gtkMiscSetAlignment instead. gtkLabelSetJustify has no effect on labels containing only a single line.

Usage

gtkLabelSetJustify(object, jtype)

Arguments

object
a GtkLabel

jtype
a GtkJustification

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]