gtkLabelSetLineWrap {RGtk2}	R Documentation

gtkLabelSetLineWrap

Description

Toggles line wrapping within the GtkLabel widget. TRUE makes it break lines if text exceeds the widget's size. FALSE lets the text get cut off by the edge of the widget if it exceeds the widget size.

Usage

gtkLabelSetLineWrap(object, wrap)

Arguments

object
a GtkLabel

wrap
the setting

Details

Note that setting line wrapping to TRUE does not make the label wrap at its parent container's width, because GTK+ widgets conceptually can't make their requisition depend on the parent container's size. For a label that wraps at a specific position, set the label's width using gtkWidgetSetSizeRequest.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]