gtkLabelSetMarkup {RGtk2}	R Documentation

gtkLabelSetMarkup

Description

Parses str which is marked up with the Pango text markup language, setting the label's text and attribute list based on the parse results. If the str is external data, you may need to escape it with gMarkupEscapeText() or gMarkupPrintfEscaped():

## RGtk2 has no support for GMarkup

Usage

gtkLabelSetMarkup(object, str)

Arguments

object
a GtkLabel

str
a markup string (see Pango markup format)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]