gtkLabelSetMarkupWithMnemonic {RGtk2}	R Documentation

gtkLabelSetMarkupWithMnemonic

Description

Parses str which is marked up with the Pango text markup language, setting the label's text and attribute list based on the parse results. If characters in str are preceded by an underscore, they are underlined indicating that they represent a keyboard accelerator called a mnemonic.

Usage

gtkLabelSetMarkupWithMnemonic(object, str)

Arguments

object
a GtkLabel

str
a markup string (see Pango markup format)

Details

The mnemonic key can be used to activate another widget, chosen automatically, or explicitly using gtkLabelSetMnemonicWidget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]