gtkLabelSetMnemonicWidget {RGtk2}	R Documentation

gtkLabelSetMnemonicWidget

Description

If the label has been set so that it has an mnemonic key (using i.e. gtkLabelSetMarkupWithMnemonic, gtkLabelSetTextWithMnemonic, gtkLabelNewWithMnemonic or the "use_underline" property) the label can be associated with a widget that is the target of the mnemonic. When the label is inside a widget (like a GtkButton or a GtkNotebook tab) it is automatically associated with the correct widget, but sometimes (i.e. when the target is a GtkEntry next to the label) you need to set it explicitly using this function.

Usage

gtkLabelSetMnemonicWidget(object, widget)

Arguments

object
a GtkLabel

widget
the target GtkWidget. [ allow-none ]

Details

The target widget will be accelerated by emitting the GtkWidget::mnemonic-activate signal on it. The default handler for this signal will activate the widget if there are no mnemonic collisions and toggle focus between the colliding widgets otherwise.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]