gtkLabelSetTextWithMnemonic {RGtk2}	R Documentation

gtkLabelSetTextWithMnemonic

Description

Sets the label's text from the string str. If characters in str are preceded by an underscore, they are underlined indicating that they represent a keyboard accelerator called a mnemonic. The mnemonic key can be used to activate another widget, chosen automatically, or explicitly using gtkLabelSetMnemonicWidget.

Usage

gtkLabelSetTextWithMnemonic(object, str)

Arguments

object
a GtkLabel

str
a string

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]