gtkLabelSetUseUnderline {RGtk2}	R Documentation

gtkLabelSetUseUnderline

Description

If true, an underline in the text indicates the next character should be used for the mnemonic accelerator key.

Usage

gtkLabelSetUseUnderline(object, setting)

Arguments

object
a GtkLabel

setting
TRUE if underlines in the text indicate mnemonics

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]