gtkLayoutGetSize {RGtk2}	R Documentation

gtkLayoutGetSize

Description

Gets the size that has been set on the layout, and that determines the total extents of the layout's scrollbar area. See gtkLayoutSetSize.

Usage

gtkLayoutGetSize(object)

Arguments

object
a GtkLayout

Value

A list containing the following elements:

width
location to store the width set on layout, or NULL. [ allow-none ]

height
location to store the height set on layout, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]