gtkLayoutGetVadjustment {RGtk2}	R Documentation

gtkLayoutGetVadjustment

Description

This function should only be called after the layout has been placed in a GtkScrolledWindow or otherwise configured for scrolling. It returns the GtkAdjustment used for communication between the vertical scrollbar and layout.

Usage

gtkLayoutGetVadjustment(object)

Arguments

object
a GtkLayout

Details

See GtkScrolledWindow, GtkScrollbar, GtkAdjustment for details.

Value

[GtkAdjustment] vertical scroll adjustment

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]