gtkLinkButtonSetUriHook {RGtk2}	R Documentation

gtkLinkButtonSetUriHook

Description

Sets func as the function that should be invoked every time a user clicks a GtkLinkButton. This function is called before every callback registered for the "clicked" signal.

Usage

gtkLinkButtonSetUriHook(func, data)

Arguments

func
a function called each time a GtkLinkButton is clicked, or NULL. [ allow-none ]

data
user data to be passed to func, or NULL. [ allow-none ]

Details

If no uri hook has been set, GTK+ defaults to calling gtkShowUri. Since 2.10

Value

[GtkLinkButtonUriFunc] the previously set hook function.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]