gtkListClearItems {RGtk2}	R Documentation

gtkListClearItems

Description

Removes the items between index start (included) and end (excluded) from the list. If end is negative, or greater than the number of children of list, it's assumed to be exactly the number of elements. If start is greater than or equal to end, nothing is done. WARNING: gtk_list_clear_items is deprecated and should not be used in newly-written code.

Usage

gtkListClearItems(object, start, end)

Arguments

object
the list widget.

start
the index of the first item to remove.

end
the index of the lest item to remove plus one.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]