gtkListExtendSelection {RGtk2}	R Documentation

gtkListExtendSelection

Description

Extends the selection by moving the anchor according to scroll.type. Only in GTK_SELECTION_EXTENDED. WARNING: gtk_list_extend_selection is deprecated and should not be used in newly-written code.

Usage

gtkListExtendSelection(object, scroll.type, position, auto.start.selection)

Arguments

object
the list widget.

scroll.type
the direction and length.

position
the position if scroll.type is GTK_SCROLL_JUMP.

auto.start.selection
if TRUE, gtkListStartSelection is automatically carried out before extending the selection.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]