gtkListItemNewWithLabel {RGtk2}	R Documentation

gtkListItemNewWithLabel

Description

Creates a new GtkListItem with a child label containing the given string. WARNING: gtk_list_item_new_with_label is deprecated and should not be used in newly-written code.

Usage

gtkListItemNewWithLabel(label, show = TRUE)

Arguments

label
the string to use for the child label.

Value

[GtkWidget] a new GtkListItem with a child GtkLabel with the text set to label.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]