gtkListScrollVertical {RGtk2}	R Documentation

gtkListScrollVertical

Description

Scrolls list vertically. This supposes that the list is packed into a scrolled window or something similar, and adjustments are well set. Step and page increment are those from the vertical adjustment of list. Backward means up, and forward down. Out of bounds values are truncated. scroll.type may be any valid GtkScrollType. If scroll.type is GTK_SCROLL_NONE, nothing is done. If it's GTK_SCROLL_JUMP, the list scrolls to the ratio position: 0 is top, 1 is bottom. WARNING: gtk_list_scroll_vertical is deprecated and should not be used in newly-written code.

Usage

gtkListScrollVertical(object, scroll.type, position)

Arguments

object
the list widget.

scroll.type
the scrolling type.

position
the position if scroll.type is GTK_SCROLL_JUMP

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]