gtkListSelectAll {RGtk2}	R Documentation

gtkListSelectAll

Description

Selects all children of list. A signal will be emitted for each newly selected child. WARNING: gtk_list_select_all is deprecated and should not be used in newly-written code.

Usage

gtkListSelectAll(object)

Arguments

object
the list widget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]