gtkListSetSelectionMode {RGtk2}	R Documentation

gtkListSetSelectionMode

Description

Set the list selection mode. The selection mode can be any value in GtkSelectionMode:

GTK_SELECTION_SINGLE

Zero or one element may be selected.

GTK_SELECTION_BROWSE

Exactly one element is always selected (this can be false after you have changed the selection mode).

GTK_SELECTION_MULTIPLE

Any number of elements may be selected. Clicks toggle the state of an item.

GTK_SELECTION_EXTENDED

Any number of elements may be selected. Click-drag selects a range of elements; the Ctrl key may be used to enlarge the selection, and Shift key to select between the focus and the child pointed to.

WARNING: gtk_list_set_selection_mode is deprecated and should not be used in newly-written code.

Usage

gtkListSetSelectionMode(object, mode)

Arguments

object
the list widget.

mode
the new selection mode.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]