gtkListStoreInsertWithValues {RGtk2}	R Documentation

gtkListStoreInsertWithValues

Description

Creates a new row at position. iter will be changed to point to this new row. If position is larger than the number of rows on the list, then the new row will be appended to the list. The row will be filled with the values given to this function.

Usage

gtkListStoreInsertWithValues(object, position, ...)

Arguments

object
A GtkListStore

position
position to insert the new row

...
undocumented

Details

Calling gtk_list_store_insert_with_values(list_store, iter, position...) has the same effect as calling

list_store$insert(iter, position)
list_store$set(iter, ...)

with the difference that the former will only emit a row_inserted signal, while the latter will emit row_inserted, row_changed and, if the list store is sorted, rows_reordered. Since emitting the rows_reordered signal repeatedly can affect the performance of the program, gtkListStoreInsertWithValues should generally be preferred when inserting rows in a sorted list store. Since 2.6

Value

A list containing the following elements:

iter
An unset GtkTreeIter to set to the new row, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]