gtkListStoreRemove {RGtk2}	R Documentation

gtkListStoreRemove

Description

Removes the given row from the list store. After being removed, iter is set to be the next valid row, or invalidated if it pointed to the last row in list.store.

Usage

gtkListStoreRemove(object, iter)

Arguments

object
A GtkListStore

iter
A valid GtkTreeIter

Value

[logical] TRUE if iter is valid, FALSE if not.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]