gtkListStoreSet {RGtk2}	R Documentation

gtkListStoreSet

Description

Sets the value of one or more cells in the row referenced by iter. The variable argument list should contain integer column numbers, each column number followed by the value to be set. The list is terminated by a -1. For example, to set column 0 with type G_TYPE_STRING to "Foo", you would write gtk_list_store_set (store, iter, 0, "Foo", -1). The value will be copied or referenced by the store if appropriate.

Usage

gtkListStoreSet(object, iter, ...)

Arguments

object
a GtkListStore

iter
row iterator

...
undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]