gtkListToggleFocusRow {RGtk2}	R Documentation

gtkListToggleFocusRow

Description

Toggles the focus row. If the focus row is selected, it's unselected. If the focus row is unselected, it's selected. If the selection mode of list is GTK_SELECTION_BROWSE, this has no effect, as the selection is always at the focus row. WARNING: gtk_list_toggle_focus_row is deprecated and should not be used in newly-written code.

Usage

gtkListToggleFocusRow(object)

Arguments

object
the list widget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]