gtkMainDoEvent {RGtk2}	R Documentation

gtkMainDoEvent

Description

Processes a single GDK event. This is public only to allow filtering of events between GDK and GTK+. You will not usually need to call this function directly.

Usage

gtkMainDoEvent(event)

Arguments

event
An event to process (normally) passed by GDK.

Details

While you should not call this function directly, you might want to know how exactly events are handled. So here is what this function does with the event:

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]