gtkMainIteration {RGtk2}	R Documentation

gtkMainIteration

Description

Runs a single iteration of the mainloop. If no events are waiting to be processed GTK+ will block until the next event is noticed. If you don't want to block look at gtkMainIterationDo or check if any events are pending with gtkEventsPending first.

Usage

gtkMainIteration()

Value

[logical] TRUE if gtkMainQuit has been called for the innermost mainloop.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]