gtkMainLevel {RGtk2}	R Documentation

gtkMainLevel

Description

Asks for the current nesting level of the main loop. This can be useful when calling gtkQuitAdd.

Usage

gtkMainLevel()

Value

[numeric] the nesting level of the current invocation of the main loop.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]