gtkMenuAttach {RGtk2}	R Documentation

gtkMenuAttach

Description

Adds a new GtkMenuItem to a (table) menu. The number of 'cells' that an item will occupy is specified by left.attach, right.attach, top.attach and bottom.attach. These each represent the leftmost, rightmost, uppermost and lower column and row numbers of the table. (Columns and rows are indexed from zero).

Usage

gtkMenuAttach(object, child, left.attach, right.attach, top.attach, 
    bottom.attach)

Arguments

object
a GtkMenu.

child
a GtkMenuItem.

left.attach
The column number to attach the left side of the item to.

right.attach
The column number to attach the right side of the item to.

top.attach
The row number to attach the top of the item to.

bottom.attach
The row number to attach the bottom of the item to.

Details

Note that this function is not related to gtkMenuDetach. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]