gtkMenuDetach {RGtk2}	R Documentation

gtkMenuDetach

Description

Detaches the menu from the widget to which it had been attached. This function will call the callback function, detacher, provided when the gtkMenuAttachToWidget function was called.

Usage

gtkMenuDetach(object)

Arguments

object
a GtkMenu.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]