gtkMenuItemNewWithMnemonic {RGtk2}	R Documentation

gtkMenuItemNewWithMnemonic

Description

Creates a new GtkMenuItem containing a label. The label will be created using gtkLabelNewWithMnemonic, so underscores in label indicate the mnemonic for the menu item.

Usage

gtkMenuItemNewWithMnemonic(label, show = TRUE)

Arguments

label
The text of the button, with an underscore in front of the mnemonic character

Value

[GtkWidget] a new GtkMenuItem

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]