gtkMenuItemSetRightJustified {RGtk2}	R Documentation

gtkMenuItemSetRightJustified

Description

Sets whether the menu item appears justified at the right side of a menu bar. This was traditionally done for "Help" menu items, but is now considered a bad idea. (If the widget layout is reversed for a right-to-left language like Hebrew or Arabic, right-justified-menu-items appear at the left.)

Usage

gtkMenuItemSetRightJustified(object, right.justified)

Arguments

object
a GtkMenuItem.

right.justified
if TRUE the menu item will appear at the far right if added to a menu bar.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]