gtkMenuPopup {RGtk2}	R Documentation

gtkMenuPopup

Description

Displays a menu and makes it available for selection. Applications can use this function to display context-sensitive menus, and will typically supply NULL for the parent.menu.shell, parent.menu.item, func and data parameters. The default menu positioning function will position the menu at the current mouse cursor position.

Usage

gtkMenuPopup(object, parent.menu.shell = NULL, parent.menu.item = NULL, 
    func = NULL, data = NULL, button, activate.time)

Arguments

object
a GtkMenu.

parent.menu.shell
the menu shell containing the triggering menu item, or NULL. [ allow-none ]

parent.menu.item
the menu item whose activation triggered the popup, or NULL. [ allow-none ]

func
a user supplied function used to position the menu, or NULL. [ allow-none ]

data
user supplied data to be passed to func. [ allow-none ]

button
the mouse button which was pressed to initiate the event.

activate.time
the time at which the activation event occurred.

Details

The button parameter should be the mouse button pressed to initiate the menu popup. If the menu popup was initiated by something other than a mouse button press, such as a mouse button release or a keypress, button should be 0.

The activate.time parameter is used to conflict-resolve initiation of concurrent requests for mouse/keyboard grab requests. To function properly, this needs to be the time stamp of the user event (such as a mouse click or key press) that caused the initiation of the popup. Only if no such event is available, gtkGetCurrentEventTime can be used instead.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]