gtkMenuSetAccelGroup {RGtk2}	R Documentation

gtkMenuSetAccelGroup

Description

Set the GtkAccelGroup which holds global accelerators for the menu. This accelerator group needs to also be added to all windows that this menu is being used in with gtkWindowAddAccelGroup, in order for those windows to support all the accelerators contained in this group.

Usage

gtkMenuSetAccelGroup(object, accel.group)

Arguments

object
a GtkMenu.

accel.group
. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]