gtkMenuSetAccelPath {RGtk2}	R Documentation

gtkMenuSetAccelPath

Description

Sets an accelerator path for this menu from which accelerator paths for its immediate children, its menu items, can be constructed. The main purpose of this function is to spare the programmer the inconvenience of having to call gtkMenuItemSetAccelPath on each menu item that should support runtime user changable accelerators. Instead, by just calling gtkMenuSetAccelPath on their parent, each menu item of this menu, that contains a label describing its purpose, automatically gets an accel path assigned. For example, a menu containing menu items "New" and "Exit", will, after gtk_menu_set_accel_path (menu, "<Gnumeric-Sheet>/File"); has been called, assign its items the accel paths: "<Gnumeric-Sheet>/File/New" and "<Gnumeric-Sheet>/File/Exit". Assigning accel paths to menu items then enables the user to change their accelerators at runtime. More details about accelerator paths and their default setups can be found at gtkAccelMapAddEntry.

Usage

gtkMenuSetAccelPath(object, accel.path)

Arguments

object
a valid GtkMenu

accel.path
a valid accelerator path. [ allow-none ]

Details

Note that accel.path string will be stored in a GQuark. Therefore, if you pass a static string, you can save some memory by interning it first with gInternStaticString().

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]