gtkMenuSetMonitor {RGtk2}	R Documentation

gtkMenuSetMonitor

Description

Informs GTK+ on which monitor a menu should be popped up. See gdkScreenGetMonitorGeometry.

Usage

gtkMenuSetMonitor(object, monitor.num)

Arguments

object
a GtkMenu

monitor.num
the number of the monitor on which the menu should be popped up

Details

This function should be called from a GtkMenuPositionFunc if the menu should not appear on the same monitor as the pointer. This information can't be reliably inferred from the coordinates returned by a GtkMenuPositionFunc, since, for very long menus, these coordinates may extend beyond the monitor boundaries or even the screen boundaries. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]