gtkMenuSetTearoffState {RGtk2}	R Documentation

gtkMenuSetTearoffState

Description

Changes the tearoff state of the menu. A menu is normally displayed as drop down menu which persists as long as the menu is active. It can also be displayed as a tearoff menu which persists until it is closed or reattached.

Usage

gtkMenuSetTearoffState(object, torn.off)

Arguments

object
a GtkMenu.

torn.off
If TRUE, menu is displayed as a tearoff menu.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]