gtkMenuSetTitle {RGtk2}	R Documentation

gtkMenuSetTitle

Description

Sets the title string for the menu. The title is displayed when the menu is shown as a tearoff menu. If title is NULL, the menu will see if it is attached to a parent menu item, and if so it will try to use the same text as that menu item's label.

Usage

gtkMenuSetTitle(object, title)

Arguments

object
a GtkMenu

title
a string containing the title for the menu.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]