gtkMenuShellInsert {RGtk2}	R Documentation

gtkMenuShellInsert

Description

Adds a new GtkMenuItem to the menu shell's item list at the position indicated by position.

Usage

gtkMenuShellInsert(object, child, position)

Arguments

object
a GtkMenuShell.

child
The GtkMenuItem to add.

position
The position in the item list where child is added. Positions are numbered from 0 to n-1.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]