gtkMenuShellSelectFirst {RGtk2}	R Documentation

gtkMenuShellSelectFirst

Description

Select the first visible or selectable child of the menu shell; don't select tearoff items unless the only item is a tearoff item.

Usage

gtkMenuShellSelectFirst(object, search.sensitive)

Arguments

object
a GtkMenuShell

search.sensitive
if TRUE, search for the first selectable menu item, otherwise select nothing if the first item isn't sensitive. This should be FALSE if the menu is being popped up initially.

Details

Since 2.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]