gtkMessageDialogFormatSecondaryMarkup {RGtk2}	R Documentation

gtkMessageDialogFormatSecondaryMarkup

Description

Sets the secondary text of the message dialog to be message.format (with printf()-style), which is marked up with the Pango text markup language.

Usage

gtkMessageDialogFormatSecondaryMarkup(object, ...)

Arguments

object
a GtkMessageDialog

...
undocumented

Details

Note that setting a secondary text makes the primary text become bold, unless you have provided explicit markup.

Due to an oversight, this function does not escape special XML characters like gtkMessageDialogNewWithMarkup does. Thus, if the arguments may contain special XML characters, you should use gMarkupPrintfEscaped() to escape it.

# GMarkup is not yet supported in RGtk2

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]