gtkMessageDialogFormatSecondaryText {RGtk2}	R Documentation

gtkMessageDialogFormatSecondaryText

Description

Sets the secondary text of the message dialog to be message.format (with printf()-style).

Usage

gtkMessageDialogFormatSecondaryText(object, ...)

Arguments

object
a GtkMessageDialog

...
undocumented

Details

Note that setting a secondary text makes the primary text become bold, unless you have provided explicit markup. Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]