gtkMessageDialogNew {RGtk2}	R Documentation

gtkMessageDialogNew

Description

Creates a new message dialog, which is a simple dialog with an icon indicating the dialog type (error, warning, etc.) and some text the user may want to see. When the user clicks a button a "response" signal is emitted with response IDs from GtkResponseType. See GtkDialog for more details.

Usage

gtkMessageDialogNew(parent = NULL, flags, type, buttons, ..., 
    show = TRUE)

Arguments

parent
transient parent, or NULL for none. [ allow-none ]

flags
flags

type
type of message

buttons
set of buttons to use

...
a new GtkMessageDialog. [ transfer none ]

Value

[GtkWidget] a new GtkMessageDialog. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]