gtkMessageDialogNewWithMarkup {RGtk2}	R Documentation

gtkMessageDialogNewWithMarkup

Description

Creates a new message dialog, which is a simple dialog with an icon indicating the dialog type (error, warning, etc.) and some text which is marked up with the Pango text markup language. When the user clicks a button a "response" signal is emitted with response IDs from GtkResponseType. See GtkDialog for more details.

Usage

gtkMessageDialogNewWithMarkup(parent, flags, type, buttons, ..., 
    show = TRUE)

Arguments

parent
transient parent, or NULL for none. [ allow-none ]

flags
flags

type
type of message

buttons
set of buttons to use

...
a new GtkMessageDialog

Details

Special XML characters in the printf() arguments passed to this function will automatically be escaped as necessary. (See gMarkupPrintfEscaped() for how this is implemented.) Usually this is what you want, but if you have an existing Pango markup string that you want to use literally as the label, then you need to use gtkMessageDialogSetMarkup instead, since you can't pass the markup string either as the format (it might contain '%' characters) or as a string argument.

dialog <- gtkMessageDialog(main_application_window,  "destroy-with-parent",
                           "error", "close")
dialog$setMarkup(message)

Since 2.4

Value

[GtkWidget] a new GtkMessageDialog

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]