gtkMessageDialogSetMarkup {RGtk2}	R Documentation

gtkMessageDialogSetMarkup

Description

Sets the text of the message dialog to be str, which is marked up with the Pango text markup language.

Usage

gtkMessageDialogSetMarkup(object, str)

Arguments

object
a GtkMessageDialog

str
markup string (see Pango markup format)

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]