gtkMiscGetAlignment {RGtk2}	R Documentation

gtkMiscGetAlignment

Description

Gets the X and Y alignment of the widget within its allocation. See gtkMiscSetAlignment.

Usage

gtkMiscGetAlignment(object)

Arguments

object
a GtkMisc

Value

A list containing the following elements:

xalign
location to store X alignment of misc, or NULL. [ allow-none ]

yalign
location to store Y alignment of misc, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]