gtkMiscSetPadding {RGtk2}	R Documentation

gtkMiscSetPadding

Description

Sets the amount of space to add around the widget.

Usage

gtkMiscSetPadding(object, xpad, ypad)

Arguments

object
a GtkMisc.

xpad
the amount of space to add on the left and right of the widget, in pixels.

ypad
the amount of space to add on the top and bottom of the widget, in pixels.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]