gtkNotebookGetMenuLabel {RGtk2}	R Documentation

gtkNotebookGetMenuLabel

Description

Retrieves the menu label widget of the page containing child.

Usage

gtkNotebookGetMenuLabel(object, child)

Arguments

object
a GtkNotebook

child
a widget contained in a page of notebook

Value

[GtkWidget] the menu label, or NULL if the notebook page does not have a menu label other than the default (the tab label).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]