gtkNotebookGetNthPage {RGtk2}	R Documentation

gtkNotebookGetNthPage

Description

Returns the child widget contained in page number page.num.

Usage

gtkNotebookGetNthPage(object, page.num)

Arguments

object
a GtkNotebook

page.num
the index of a page in the notebook, or -1 to get the last page.

Value

[GtkWidget] the child widget, or NULL if page.num is out of bounds. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]