gtkNotebookGetTabLabel {RGtk2}	R Documentation

gtkNotebookGetTabLabel

Description

Returns the tab label widget for the page child. NULL is returned if child is not in notebook or if no tab label has specifically been set for child.

Usage

gtkNotebookGetTabLabel(object, child)

Arguments

object
a GtkNotebook

child
the page

Value

[GtkWidget] the tab label. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]