gtkNotebookInsertPage {RGtk2}	R Documentation

gtkNotebookInsertPage

Description

Insert a page into notebook at the given position.

Usage

gtkNotebookInsertPage(object, child, tab.label = NULL, position = -1)

Arguments

object
a GtkNotebook

child
the GtkWidget to use as the contents of the page.

tab.label
the GtkWidget to be used as the label for the page, or NULL to use the default label, 'page N'. [ allow-none ]

position
the index (starting at 0) at which to insert the page, or -1 to append the page after all other pages.

Value

[integer] the index (starting from 0) of the inserted page in the notebook, or -1 if function fails

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]