gtkNotebookPageNum {RGtk2}	R Documentation

gtkNotebookPageNum

Description

Finds the index of the page which contains the given child widget.

Usage

gtkNotebookPageNum(object, child)

Arguments

object
a GtkNotebook

child
a GtkWidget

Value

[integer] the index of the page containing child, or -1 if child is not in the notebook.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]