gtkNotebookPrependPage {RGtk2}	R Documentation

gtkNotebookPrependPage

Description

Prepends a page to notebook.

Usage

gtkNotebookPrependPage(object, child, tab.label = NULL)

Arguments

object
a GtkNotebook

child
the GtkWidget to use as the contents of the page.

tab.label
the GtkWidget to be used as the label for the page, or NULL to use the default label, 'page N'. [ allow-none ]

Value

[integer] the index (starting from 0) of the prepended page in the notebook, or -1 if function fails

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]