gtkNotebookRemovePage {RGtk2}	R Documentation

gtkNotebookRemovePage

Description

Removes a page from the notebook given its index in the notebook.

Usage

gtkNotebookRemovePage(object, page.num)

Arguments

object
a GtkNotebook.

page.num
the index of a notebook page, starting from 0. If -1, the last page will be removed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]