gtkNotebookReorderChild {RGtk2}	R Documentation

gtkNotebookReorderChild

Description

Reorders the page containing child, so that it appears in position position. If position is greater than or equal to the number of children in the list or negative, child will be moved to the end of the list.

Usage

gtkNotebookReorderChild(object, child, position)

Arguments

object
a GtkNotebook

child
the child to move

position
the new position, or -1 to move to the end

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]