gtkNotebookSetCurrentPage {RGtk2}	R Documentation

gtkNotebookSetCurrentPage

Description

Switches to the page number page.num.

Usage

gtkNotebookSetCurrentPage(object, page.num)

Arguments

object
a GtkNotebook

page.num
index of the page to switch to, starting from 0. If negative, the last page will be used. If greater than the number of pages in the notebook, nothing will be done.

Details

Note that due to historical reasons, GtkNotebook refuses to switch to a page unless the child widget is visible. Therefore, it is recommended to show child widgets before adding them to a notebook.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]