gtkNotebookSetGroupId {RGtk2}	R Documentation

gtkNotebookSetGroupId

Description

Sets an group identificator for notebook, notebooks sharing the same group identificator will be able to exchange tabs via drag and drop. A notebook with group identificator -1 will not be able to exchange tabs with any other notebook. WARNING: gtk_notebook_set_group_id has been deprecated since version 2.12 and should not be used in newly-written code. use gtkNotebookSetGroup instead.

Usage

gtkNotebookSetGroupId(object, group.id)

Arguments

object
a GtkNotebook

group.id
a group identificator, or -1 to unset it

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]