gtkNotebookSetShowBorder {RGtk2}	R Documentation

gtkNotebookSetShowBorder

Description

Sets whether a bevel will be drawn around the notebook pages. This only has a visual effect when the tabs are not shown. See gtkNotebookSetShowTabs.

Usage

gtkNotebookSetShowBorder(object, show.border)

Arguments

object
a GtkNotebook

show.border
TRUE if a bevel should be drawn around the notebook.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]