gtkNotebookSetTabBorder {RGtk2}	R Documentation

gtkNotebookSetTabBorder

Description

Sets the width the border around the tab labels in a notebook. This is equivalent to calling gtk_notebook_set_tab_hborder (notebook, border.width) followed by gtk_notebook_set_tab_vborder (notebook, border.width). WARNING: gtk_notebook_set_tab_border is deprecated and should not be used in newly-written code.

Usage

gtkNotebookSetTabBorder(object, border.width)

Arguments

object
a GtkNotebook

border.width
width of the border around the tab labels.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]