gtkNotebookSetTabLabelPacking {RGtk2}	R Documentation

gtkNotebookSetTabLabelPacking

Description

Sets the packing parameters for the tab label of the page containing child. See gtkBoxPackStart for the exact meaning of the parameters. WARNING: gtk_notebook_set_tab_label_packing has been deprecated since version 2.20 and should not be used in newly-written code. Modify the "tab-expand" and "tab-fill" child properties instead. Modifying the packing of the tab label is a deprecated feature and shouldn't be done anymore.

Usage

gtkNotebookSetTabLabelPacking(object, child, expand, fill, pack.type)

Arguments

object
a GtkNotebook

child
the child widget

expand
whether to expand the tab label or not

fill
whether the tab label should fill the allocated area or not

pack.type
the position of the tab label

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]