gtkNotebookSetTabVborder {RGtk2}	R Documentation

gtkNotebookSetTabVborder

Description

Sets the width of the vertical border of tab labels. WARNING: gtk_notebook_set_tab_vborder is deprecated and should not be used in newly-written code.

Usage

gtkNotebookSetTabVborder(object, tab.vborder)

Arguments

object
a GtkNotebook

tab.vborder
width of the vertical border of tab labels.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]