gtkObjectSignalEmit {RGtk2}	R Documentation

Emit a Gtk signal from a Gtk object.

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

This allows us to emit or raise a Gtk signal from a Gtk object directly from R. This should not be used widely. The conversion of the arguments to their C equivalents is complex.

Usage

gtkObjectSignalEmit(obj, signal, ...)

Arguments

obj
the GtkObject from which the event is to be emitted.

signal
the identifier of the signal/event, typically a string giving the name of the event.

...
the arguments that are passed in the signal emission to the callbacks/signal handlers.

Details

One should not rely on this, and fortunately one rarely needs to when developing regular GUIs in R. Typically, we only need this when we are developing Gtk classes.

Author(s)

Duncan Temple Lang <duncan@research.bell-labs.com>

References

Information on the package is available from http://www.omegahat.org/RGtk.

Information on Gtk is available from http://www.gtk.org.

See Also

gtkObjectAddCallback

[Package RGtk2 version 2.20.31 Index]