gtkOffscreenWindowGetPixmap {RGtk2}	R Documentation

gtkOffscreenWindowGetPixmap

Description

Retrieves a snapshot of the contained widget in the form of a GdkPixmap. If you need to keep this around over window resizes then you should add a reference to it.

Usage

gtkOffscreenWindowGetPixmap(object)

Arguments

object
the GtkOffscreenWindow contained widget.

Details

Since 2.20

Value

[GdkPixmap] A GdkPixmap pointer to the offscreen pixmap, or NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]