gtkOffscreenWindowNew {RGtk2}	R Documentation

gtkOffscreenWindowNew

Description

Creates a toplevel container widget that is used to retrieve snapshots of widgets without showing them on the screen. For widgets that are on the screen and part of a normal widget hierarchy, gtkWidgetGetSnapshot can be used instead.

Usage

gtkOffscreenWindowNew(show = TRUE)

Details

Since 2.20

Value

[GtkWidget] A pointer to a GtkWidget

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]