gtkPaintExpander {RGtk2}	R Documentation

gtkPaintExpander

Description

Draws an expander as used in GtkTreeView. x and y specify the center the expander. The size of the expander is determined by the "expander-size" style property of widget. (If widget is not specified or doesn't have an "expander-size" property, an unspecified default size will be used, since the caller doesn't have sufficient information to position the expander, this is likely not useful.) The expander is expander_size pixels tall in the collapsed position and expander_size pixels wide in the expanded position.

Usage

gtkPaintExpander(object, window, state.type, area = NULL, widget = NULL, 
    detail = NULL, x, y, expander.style)

Arguments

object
a GtkStyle

window
a GdkWindow

state.type
a state

area
clip rectangle, or NULL if the output should not be clipped. [ allow-none ]

widget
the widget. [ allow-none ]

detail
a style detail. [ allow-none ]

x
the x position to draw the expander at

y
the y position to draw the expander at

expander.style
the style to draw the expander in; determines whether the expander is collapsed, expanded, or in an intermediate state.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]