gtkPaintSpinner {RGtk2}	R Documentation

gtkPaintSpinner

Description

Draws a spinner on window using the given parameters.

Usage

gtkPaintSpinner(object, window, state.type, area, widget, detail, 
    step, x, y, width, height)

Arguments

object
a GtkStyle

window
a GdkWindow

state.type
a state

area
clip rectangle, or NULL if the output should not be clipped. [ allow-none ]

widget
the widget (may be NULL). [ allow-none ]

detail
a style detail (may be NULL). [ allow-none ]

step
the nth step, a value between 0 and "num-steps"

x
the x origin of the rectangle in which to draw the spinner

y
the y origin of the rectangle in which to draw the spinner

width
the width of the rectangle in which to draw the spinner

height
the height of the rectangle in which to draw the spinner

Details

Since 2.20

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]