gtkPanedPack1 {RGtk2}	R Documentation

gtkPanedPack1

Description

Adds a child to the top or left pane.

Usage

gtkPanedPack1(object, child, resize = FALSE, shrink = TRUE)

Arguments

object
a paned widget

child
the child to add

resize
should this child expand when the paned widget is resized.

shrink
can this child be made smaller than its requisition.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]