gtkPixmapSet {RGtk2}	R Documentation

gtkPixmapSet

Description

Sets the GdkPixmap and GdkBitmap mask. WARNING: gtk_pixmap_set is deprecated and should not be used in newly-written code.

Usage

gtkPixmapSet(object, val, mask = NULL)

Arguments

object
a GtkPixmap.

val
a GdkPixmap.

mask
a GdkBitmap, which indicates which parts of the pixmap should be transparent. This can be NULL, in which case none of the pixmap is transparent.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]