gtkPreviewDrawRow {RGtk2}	R Documentation

gtkPreviewDrawRow

Description

Sets the data for a portion of a row. WARNING: gtk_preview_draw_row is deprecated and should not be used in newly-written code.

Usage

gtkPreviewDrawRow(object, data, y, w)

Arguments

object
a GtkPreview.

data
the new data for the portion. It should contain w bytes of data if the preview is of type GTK_TYPE_GRAYSCALE, and 3*w bytes of data if the preview is of type GTK_TYPE_COLOR.

y
the row to change.

w
the number of pixels in the row to change.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]