gtkPreviewGetVisual {RGtk2}	R Documentation

gtkPreviewGetVisual

Description

Returns the visual used by preview widgets. This function is deprecated, and you should use gdkRgbGetVisual instead. WARNING: gtk_preview_get_visual is deprecated and should not be used in newly-written code.

Usage

gtkPreviewGetVisual()

Value

[GdkVisual] the visual for previews.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]