gtkPreviewPut {RGtk2}	R Documentation

gtkPreviewPut

Description

Takes a portion of the contents of a preview widget and draws it onto the given drawable, window. WARNING: gtk_preview_put is deprecated and should not be used in newly-written code.

Usage

gtkPreviewPut(object, window, gc, srcx, srcy, destx, desty, width, 
    height)

Arguments

object
a GtkPreview.

window
a window or pixmap.

gc
The graphics context for the operation. Only the clip mask for this GC matters.

srcx
the x coordinate of the upper left corner in the source image.

srcy
the y coordinate of the upper left corner in the source image.

destx
the x coordinate of the upper left corner in the destination image.

desty
the y coordinate of the upper left corner in the destination image.

width
the width of the rectangular portion to draw.

height
the height of the rectangular portion to draw.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]