gtkPreviewSetColorCube {RGtk2}	R Documentation

gtkPreviewSetColorCube

Description

This function is deprecated and does nothing. GdkRGB automatically picks an optimium color cube for the display. WARNING: gtk_preview_set_color_cube is deprecated and should not be used in newly-written code.

Usage

gtkPreviewSetColorCube(nred.shades, ngreen.shades, nblue.shades, 
    ngray.shades)

Arguments

nred.shades
ignored

ngreen.shades
ignored

nblue.shades
ignored

ngray.shades
ignored

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]