gtkPreviewSetExpand {RGtk2}	R Documentation

gtkPreviewSetExpand

Description

Determines the way that the the preview widget behaves when the size it is allocated is larger than the requested size. If expand is FALSE, then the preview's window and buffer will be no larger than the size set with gtkPreviewSize, and the data set will be centered in the allocation if it is larger. If expand is TRUE then the window and buffer will expand with the allocation; the application is responsible for catching the "size_allocate" signal and providing the data appropriate for this size. WARNING: gtk_preview_set_expand is deprecated and should not be used in newly-written code.

Usage

gtkPreviewSetExpand(object, expand)

Arguments

object
a GtkPreview.

expand
whether the preview's window should expand or not.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]