gtkPreviewSetGamma {RGtk2}	R Documentation

gtkPreviewSetGamma

Description

Set the gamma-correction value for all preview widgets. (This function will eventually be replaced with a function that sets a per-preview-widget gamma value). The resulting intensity is given by: destination_value * pow (source_value/255, 1/gamma). The gamma value is applied when the data is set with gtkPreviewDrawRow so changing this value will not affect existing data in preview widgets. WARNING: gtk_preview_set_gamma is deprecated and should not be used in newly-written code.

Usage

gtkPreviewSetGamma(gamma)

Arguments

gamma
the new gamma value.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]