gtkPreviewSize {RGtk2}	R Documentation

gtkPreviewSize

Description

Set the size that the preview widget will request in response to a "size_request" signal. The drawing area may actually be allocated a size larger than this depending on how it is packed within the enclosing containers. The effect of this is determined by whether the preview is set to expand or not (see gtkPreviewExpand()) WARNING: gtk_preview_size is deprecated and should not be used in newly-written code.

Usage

gtkPreviewSize(object, width, height)

Arguments

object
a GtkPreview.

width
the new width.

height
the new height.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]