gtkPrintContextSetCairoContext {RGtk2}	R Documentation

gtkPrintContextSetCairoContext

Description

Sets a new cairo context on a print context.

Usage

gtkPrintContextSetCairoContext(object, cr, dpi.x, dpi.y)

Arguments

object
a GtkPrintContext

cr
the cairo context

dpi.x
the horizontal resolution to use with cr

dpi.y
the vertical resolution to use with cr

Details

This function is intended to be used when implementing an internal print preview, it is not needed for printing, since GTK+ itself creates a suitable cairo context in that case. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]