gtkPrintOperationDrawPageFinish {RGtk2}	R Documentation

gtkPrintOperationDrawPageFinish

Description

Signalize that drawing of particular page is complete.

Usage

gtkPrintOperationDrawPageFinish(object)

Arguments

object
a GtkPrintOperation

Details

It is called after completion of page drawing (e.g. drawing in another thread). If gtkPrintOperationSetDeferDrawing was called before, then this function has to be called by application. In another case it is called by the library itself. Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]