gtkPrintOperationGetError {RGtk2}	R Documentation

gtkPrintOperationGetError

Description

Call this when the result of a print operation is GTK_PRINT_OPERATION_RESULT_ERROR, either as returned by gtkPrintOperationRun, or in the "done" signal handler. The returned GError will contain more details on what went wrong.

Usage

gtkPrintOperationGetError(object, .errwarn = TRUE)

Arguments

object
a GtkPrintOperation

.errwarn
Whether to issue a warning on error or fail silently

Details

Since 2.10

Value

A list containing the following elements:

error
return location for the error

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]