gtkPrintOperationGetNPagesToPrint {RGtk2}	R Documentation

gtkPrintOperationGetNPagesToPrint

Description

Returns the number of pages that will be printed.

Usage

gtkPrintOperationGetNPagesToPrint(object)

Arguments

object
a GtkPrintOperation

Details

Note that this value is set during print preparation phase (GTK_PRINT_STATUS_PREPARING), so this function should never be called before the data generation phase (GTK_PRINT_STATUS_GENERATING_DATA). You can connect to the "status-changed" signal and call gtkPrintOperationGetNPagesToPrint when print status is GTK_PRINT_STATUS_GENERATING_DATA. This is typically used to track the progress of print operation. Since 2.18

Value

[integer] the number of pages that will be printed

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]