gtkPrintOperationIsFinished {RGtk2}	R Documentation

gtkPrintOperationIsFinished

Description

A convenience function to find out if the print operation is finished, either successfully (GTK_PRINT_STATUS_FINISHED) or unsuccessfully (GTK_PRINT_STATUS_FINISHED_ABORTED).

Usage

gtkPrintOperationIsFinished(object)

Arguments

object
a GtkPrintOperation

Details

Note: when you enable print status tracking the print operation can be in a non-finished state even after done has been called, as the operation status then tracks the print job status on the printer. Since 2.10

Value

[logical] TRUE, if the print operation is finished.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]