gtkPrintOperationPreviewRenderPage {RGtk2}	R Documentation

gtkPrintOperationPreviewRenderPage

Description

Renders a page to the preview, using the print context that was passed to the "preview" handler together with preview.

Usage

gtkPrintOperationPreviewRenderPage(object, page.nr)

Arguments

object
a GtkPrintOperationPreview

page.nr
the page to render

Details

A custom iprint preview should use this function in its ::expose handler to render the currently selected page.

Note that this function requires a suitable cairo context to be associated with the print context. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]