gtkPrintOperationSetDeferDrawing {RGtk2}	R Documentation

gtkPrintOperationSetDeferDrawing

Description

Sets up the GtkPrintOperation to wait for calling of gtkPrintOperationDrawPageFinish from application. It can be used for drawing page in another thread.

Usage

gtkPrintOperationSetDeferDrawing(object)

Arguments

object
a GtkPrintOperation

Details

This function must be called in the callback of "draw-page" signal. Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]