gtkPrintOperationSetEmbedPageSetup {RGtk2}	R Documentation

gtkPrintOperationSetEmbedPageSetup

Description

Embed page size combo box and orientation combo box into page setup page. Selected page setup is stored as default page setup in GtkPrintOperation.

Usage

gtkPrintOperationSetEmbedPageSetup(object, embed)

Arguments

object
a GtkPrintOperation

embed
TRUE to embed page setup selection in the GtkPrintDialog

Details

Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]