gtkPrintOperationSetExportFilename {RGtk2}	R Documentation

gtkPrintOperationSetExportFilename

Description

Sets up the GtkPrintOperation to generate a file instead of showing the print dialog. The indended use of this function is for implementing "Export to PDF" actions. Currently, PDF is the only supported format.

Usage

gtkPrintOperationSetExportFilename(object, filename)

Arguments

object
a GtkPrintOperation

filename
the filename for the exported file

Details

"Print to PDF" support is independent of this and is done by letting the user pick the "Print to PDF" item from the list of printers in the print dialog. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]