gtkPrintOperationSetJobName {RGtk2}	R Documentation

gtkPrintOperationSetJobName

Description

Sets the name of the print job. The name is used to identify the job (e.g. in monitoring applications like eggcups).

Usage

gtkPrintOperationSetJobName(object, job.name)

Arguments

object
a GtkPrintOperation

job.name
a string that identifies the print job

Details

If you don't set a job name, GTK+ picks a default one by numbering successive print jobs. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]