gtkPrintOperationSetNPages {RGtk2}	R Documentation

gtkPrintOperationSetNPages

Description

Sets the number of pages in the document.

Usage

gtkPrintOperationSetNPages(object, n.pages)

Arguments

object
a GtkPrintOperation

n.pages
the number of pages

Details

This must be set to a positive number before the rendering starts. It may be set in a "begin-print" signal hander.

Note that the page numbers passed to the "request-page-setup" and "draw-page" signals are 0-based, i.e. if the user chooses to print all pages, the last ::draw-page signal will be for page n.pages - 1. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]