gtkPrintOperationSetUseFullPage {RGtk2}	R Documentation

gtkPrintOperationSetUseFullPage

Description

If full.page is TRUE, the transformation for the cairo context obtained from GtkPrintContext puts the origin at the top left corner of the page (which may not be the top left corner of the sheet, depending on page orientation and the number of pages per sheet). Otherwise, the origin is at the top left corner of the imageable area (i.e. inside the margins).

Usage

gtkPrintOperationSetUseFullPage(object, full.page)

Arguments

object
a GtkPrintOperation

full.page
TRUE to set up the GtkPrintContext for the full page

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]