gtkPrintRunPageSetupDialog {RGtk2}	R Documentation

gtkPrintRunPageSetupDialog

Description

Runs a page setup dialog, letting the user modify the values from page.setup. If the user cancels the dialog, the returned GtkPageSetup is identical to the passed in page.setup, otherwise it contains the modifications done in the dialog.

Usage

gtkPrintRunPageSetupDialog(parent, page.setup = NULL, settings)

Arguments

parent
transient parent. [ allow-none ]

page.setup
an existing GtkPageSetup. [ allow-none ]

settings
a GtkPrintSettings

Details

Note that this function may use a recursive mainloop to show the page setup dialog. See gtkPrintRunPageSetupDialogAsync if this is a problem. Since 2.10

Value

[GtkPageSetup] a new GtkPageSetup

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]