gtkPrintRunPageSetupDialogAsync {RGtk2}	R Documentation

gtkPrintRunPageSetupDialogAsync

Description

Runs a page setup dialog, letting the user modify the values from page.setup.

Usage

gtkPrintRunPageSetupDialogAsync(parent, page.setup, settings, 
    done.cb, data)

Arguments

parent
transient parent, or NULL. [ allow-none ]

page.setup
an existing GtkPageSetup, or NULL. [ allow-none ]

settings
a GtkPrintSettings

done.cb
a function to call when the user saves the modified page setup

data
user data to pass to done.cb

Details

In contrast to gtkPrintRunPageSetupDialog, this function returns after showing the page setup dialog on platforms that support this, and calls done.cb from a signal handler for the ::response signal of the dialog. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]