gtkPrintSettingsLoadKeyFile {RGtk2}	R Documentation

gtkPrintSettingsLoadKeyFile

Description

Reads the print settings from the group group.name in key.file. If the file could not be loaded then error is set to either a GFileError or GKeyFileError.

Usage

gtkPrintSettingsLoadKeyFile(object, key.file, group.name, .errwarn = TRUE)

Arguments

object
a GtkPrintSettings

key.file
the GKeyFile to retrieve the settings from

group.name
the name of the group to use, or NULL to use the default "Print Settings". [ allow-none ]

.errwarn
Whether to issue a warning on error or fail silently

Details

Since 2.14

Value

A list containing the following elements:

retval
[logical] TRUE on success

error
return location for errors, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]