gtkPrintSettingsToKeyFile {RGtk2}	R Documentation

gtkPrintSettingsToKeyFile

Description

This function adds the print settings from settings to key.file.

Usage

gtkPrintSettingsToKeyFile(object, key.file, group.name)

Arguments

object
a GtkPrintSettings

key.file
the GKeyFile to save the print settings to

group.name
the group to add the settings to in key.file, or NULL to use the default "Print Settings"

Details

Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]