gtkProgressBarPulse {RGtk2}	R Documentation

gtkProgressBarPulse

Description

Indicates that some progress is made, but you don't know how much. Causes the progress bar to enter "activity mode," where a block bounces back and forth. Each call to gtkProgressBarPulse causes the block to move by a little bit (the amount of movement per pulse is determined by gtkProgressBarSetPulseStep).

Usage

gtkProgressBarPulse(object)

Arguments

object
a GtkProgressBar

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]