gtkProgressBarSetActivityStep {RGtk2}	R Documentation

gtkProgressBarSetActivityStep

Description

Sets the step value used when the progress bar is in activity mode. The step is the amount by which the progress is incremented each iteration. WARNING: gtk_progress_bar_set_activity_step is deprecated and should not be used in newly-written code.

Usage

gtkProgressBarSetActivityStep(object, step)

Arguments

object
a GtkProgressBar.

step
the amount which the progress is incremented in activity mode.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]