gtkProgressBarSetDiscreteBlocks {RGtk2}	R Documentation

gtkProgressBarSetDiscreteBlocks

Description

Sets the number of blocks that the progress bar is divided into when the style is GTK_PROGRESS_DISCRETE. WARNING: gtk_progress_bar_set_discrete_blocks is deprecated and should not be used in newly-written code.

Usage

gtkProgressBarSetDiscreteBlocks(object, blocks)

Arguments

object
a GtkProgressBar.

blocks
number of individual blocks making up the bar.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]