gtkProgressConfigure {RGtk2}	R Documentation

gtkProgressConfigure

Description

Allows the configuration of the minimum, maximum, and current values for the GtkProgress. WARNING: gtk_progress_configure is deprecated and should not be used in newly-written code.

Usage

gtkProgressConfigure(object, value, min, max)

Arguments

object
a GtkProgress.

value
the current progress value.

min
the minimum progress value.

max
the maximum progress value.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]