gtkProgressGetTextFromValue {RGtk2}	R Documentation

gtkProgressGetTextFromValue

Description

Returns the text indicating the progress based on the supplied value. The current value for the GtkProgress remains unchanged. WARNING: gtk_progress_get_text_from_value is deprecated and should not be used in newly-written code.

Usage

gtkProgressGetTextFromValue(object, value)

Arguments

object
a GtkProgress.

value
an absolute progress value to use when formatting the progress text.

Value

[character] a string indicating the progress.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]