gtkProgressSetActivityMode {RGtk2}	R Documentation

gtkProgressSetActivityMode

Description

A GtkProgress can be in one of two different modes: percentage mode (the default) and activity mode. In activity mode, the progress is simply indicated as activity rather than as a percentage complete. WARNING: gtk_progress_set_activity_mode is deprecated and should not be used in newly-written code.

Usage

gtkProgressSetActivityMode(object, activity.mode)

Arguments

object
a GtkProgress.

activity.mode
a boolean, TRUE for activity mode.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]