gtkProgressSetFormatString {RGtk2}	R Documentation

gtkProgressSetFormatString

Description

Sets a format string used to display text indicating the current progress. The string can contain the following substitution characters:

%v - the current progress value.

%l - the lower bound for the progress value.

%u - the upper bound for the progress value.

%p - the current progress percentage.

WARNING: gtk_progress_set_format_string is deprecated and should not be used in newly-written code.

Usage

gtkProgressSetFormatString(object, format)

Arguments

object
a GtkProgress.

format
a string used to display progress text, or NULL to restore to the default format.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]