gtkPropagateEvent {RGtk2}	R Documentation

gtkPropagateEvent

Description

Sends an event to a widget, propagating the event to parent widgets if the event remains unhandled. Events received by GTK+ from GDK normally begin in gtkMainDoEvent. Depending on the type of event, existence of modal dialogs, grabs, etc., the event may be propagated; if so, this function is used. gtkPropagateEvent calls gtkWidgetEvent on each widget it decides to send the event to. So gtkWidgetEvent is the lowest-level function; it simply emits the "event" and possibly an event-specific signal on a widget. gtkPropagateEvent is a bit higher-level, and gtkMainDoEvent is the highest level.

Usage

gtkPropagateEvent(object, event)

Arguments

object
a GtkWidget

event
an event

Details

All that said, you most likely don't want to use any of these functions; synthesizing events is rarely needed. Consider asking on the mailing list for better ways to achieve your goals. For example, use gdkWindowInvalidateRect or gtkWidgetQueueDraw instead of making up expose events.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]