gtkQuitAddDestroy {RGtk2}	R Documentation

gtkQuitAddDestroy

Description

Trigger destruction of object in case the mainloop at level main.level is quit.

Usage

gtkQuitAddDestroy(main.level, object)

Arguments

main.level
Level of the mainloop which shall trigger the destruction.

object
Object to be destroyed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]