gtkQuitAddFull {RGtk2}	R Documentation

gtkQuitAddFull

Description

Registers a function to be called when an instance of the mainloop is left. In comparison to gtkQuitAdd this function adds the possibility to pass a marshaller and a function to be called when the quit handler is freed.

Usage

gtkQuitAddFull(main.level, fun, data = NULL)

Arguments

main.level
Level at which termination the function shall be called. You can pass 0 here to have the function run at the termination of the current mainloop.

data
Pointer to pass when calling function.

Details

The former can be used to run interpreted code instead of a compiled function while the latter can be used to free the information stored in data (while you can do this in function as well)... So this function will mostly be used by GTK+ wrappers for languages other than C.

Value

[numeric] A handle for this quit handler (you need this for gtkQuitRemove) or 0 if you passed a NULL pointer in function.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]