gtkRadioActionNew {RGtk2}	R Documentation

gtkRadioActionNew

Description

Creates a new GtkRadioAction object. To add the action to a GtkActionGroup and set the accelerator for the action, call gtkActionGroupAddActionWithAccel.

Usage

gtkRadioActionNew(name = NULL, label = NULL, tooltip = NULL, 
    stock.id = NULL, value = NULL)

Arguments

name
A unique name for the action

label
The label displayed in menu items and on buttons, or NULL. [ allow-none ]

tooltip
A tooltip for this action, or NULL. [ allow-none ]

stock.id
The stock icon to display in widgets representing this action, or NULL

value
The value which gtkRadioActionGetCurrentValue should return if this action is selected.

Details

Since 2.4

Value

[GtkRadioAction] a new GtkRadioAction

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]