gtkRadioButtonSetGroup {RGtk2}	R Documentation

gtkRadioButtonSetGroup

Description

Sets a GtkRadioButton's group. It should be noted that this does not change the layout of your interface in any way, so if you are changing the group, it is likely you will need to re-arrange the user interface to reflect these changes.

Usage

gtkRadioButtonSetGroup(object, group)

Arguments

object
a GtkRadioButton.

group
an existing radio button group, such as one returned from gtkRadioButtonGetGroup.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]