gtkRadioMenuItemNewWithMnemonicFromWidget {RGtk2}	R Documentation

gtkRadioMenuItemNewWithMnemonicFromWidget

Description

Creates a new GtkRadioMenuItem containing a label. The label will be created using gtkLabelNewWithMnemonic, so underscores in label indicate the mnemonic for the menu item.

Usage

gtkRadioMenuItemNewWithMnemonicFromWidget(group = NULL, label, 
    show = TRUE)

Arguments

group
An existing GtkRadioMenuItem

label
the text of the button, with an underscore in front of the mnemonic character

Details

The new GtkRadioMenuItem is added to the same group as group. Since 2.4

Value

[GtkWidget] The new GtkRadioMenuItem

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]