gtkRangeSetAdjustment {RGtk2}	R Documentation

gtkRangeSetAdjustment

Description

Sets the adjustment to be used as the "model" object for this range widget. The adjustment indicates the current range value, the minimum and maximum range values, the step/page increments used for keybindings and scrolling, and the page size. The page size is normally 0 for GtkScale and nonzero for GtkScrollbar, and indicates the size of the visible area of the widget being scrolled. The page size affects the size of the scrollbar slider.

Usage

gtkRangeSetAdjustment(object, adjustment)

Arguments

object
a GtkRange

adjustment
a GtkAdjustment

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]