gtkRangeSetFillLevel {RGtk2}	R Documentation

gtkRangeSetFillLevel

Description

Set the new position of the fill level indicator.

Usage

gtkRangeSetFillLevel(object, fill.level)

Arguments

object
a GtkRange

fill.level
the new position of the fill level indicator

Details

The "fill level" is probably best described by its most prominent use case, which is an indicator for the amount of pre-buffering in a streaming media player. In that use case, the value of the range would indicate the current play position, and the fill level would be the position up to which the file/stream has been downloaded.

This amount of prebuffering can be displayed on the range's trough and is themeable separately from the trough. To enable fill level display, use gtkRangeSetShowFillLevel. The range defaults to not showing the fill level.

Additionally, it's possible to restrict the range's slider position to values which are smaller than the fill level. This is controller by gtkRangeSetRestrictToFillLevel and is by default enabled. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]