gtkRangeSetFlippable {RGtk2}	R Documentation

gtkRangeSetFlippable

Description

If a range is flippable, it will switch its direction if it is horizontal and its direction is GTK_TEXT_DIR_RTL.

Usage

gtkRangeSetFlippable(object, flippable)

Arguments

object
a GtkRange

flippable
TRUE to make the range flippable

Details

See gtkWidgetGetDirection. Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]