gtkRangeSetIncrements {RGtk2}	R Documentation

gtkRangeSetIncrements

Description

Sets the step and page sizes for the range. The step size is used when the user clicks the GtkScrollbar arrows or moves GtkScale via arrow keys. The page size is used for example when moving via Page Up or Page Down keys.

Usage

gtkRangeSetIncrements(object, step, page)

Arguments

object
a GtkRange

step
step size

page
page size

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]