gtkRangeSetInverted {RGtk2}	R Documentation

gtkRangeSetInverted

Description

Ranges normally move from lower to higher values as the slider moves from top to bottom or left to right. Inverted ranges have higher values at the top or on the right rather than on the bottom or left.

Usage

gtkRangeSetInverted(object, setting)

Arguments

object
a GtkRange

setting
TRUE to invert the range

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]