gtkRangeSetRange {RGtk2}	R Documentation

gtkRangeSetRange

Description

Sets the allowable values in the GtkRange, and clamps the range value to be between min and max. (If the range has a non-zero page size, it is clamped between min and max - page-size.)

Usage

gtkRangeSetRange(object, min, max)

Arguments

object
a GtkRange

min
minimum range value

max
maximum range value

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]