gtkRangeSetUpdatePolicy {RGtk2}	R Documentation

gtkRangeSetUpdatePolicy

Description

Sets the update policy for the range. GTK_UPDATE_CONTINUOUS means that anytime the range slider is moved, the range value will change and the value_changed signal will be emitted. GTK_UPDATE_DELAYED means that the value will be updated after a brief timeout where no slider motion occurs, so updates are spaced by a short time rather than continuous. GTK_UPDATE_DISCONTINUOUS means that the value will only be updated when the user releases the button and ends the slider drag operation.

Usage

gtkRangeSetUpdatePolicy(object, policy)

Arguments

object
a GtkRange

policy
update policy

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]