gtkRangeSetValue {RGtk2}	R Documentation

gtkRangeSetValue

Description

Sets the current value of the range; if the value is outside the minimum or maximum range values, it will be clamped to fit inside them. The range emits the "value-changed" signal if the value changes.

Usage

gtkRangeSetValue(object, value)

Arguments

object
a GtkRange

value
new value of the range

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]