gtkRcAddClassStyle {RGtk2}	R Documentation

gtkRcAddClassStyle

Description

Adds a GtkRcStyle that will be looked up by a matching against the class hierarchy of the widget. This is equivalent to a: class PATTERN style STYLE statement in a RC file. WARNING: gtk_rc_add_class_style is deprecated and should not be used in newly-written code. Use gtkRcParseString with a suitable string instead.

Usage

gtkRcAddClassStyle(object, pattern)

Arguments

object
the GtkRcStyle to use for widgets deriving from pattern

pattern
the pattern

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]