gtkRcAddDefaultFile {RGtk2}	R Documentation

gtkRcAddDefaultFile

Description

Adds a file to the list of files to be parsed at the end of gtkInit.

Usage

gtkRcAddDefaultFile(filename)

Arguments

filename
the pathname to the file. If filename is not absolute, it is searched in the current directory.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]