gtkRcAddWidgetClassStyle {RGtk2}	R Documentation

gtkRcAddWidgetClassStyle

Description

Adds a GtkRcStyle that will be looked up by a match against the widget's class pathname. This is equivalent to a: widget_class PATTERN style STYLE statement in a RC file. WARNING: gtk_rc_add_widget_class_style is deprecated and should not be used in newly-written code. Use gtkRcParseString with a suitable string instead.

Usage

gtkRcAddWidgetClassStyle(object, pattern)

Arguments

object
the GtkRcStyle to use for widgets matching pattern

pattern
the pattern

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]