gtkRcGetStyle {RGtk2}	R Documentation

gtkRcGetStyle

Description

Finds all matching RC styles for a given widget, composites them together, and then creates a GtkStyle representing the composite appearance. (GTK+ actually keeps a cache of previously created styles, so a new style may not be created.)

Usage

gtkRcGetStyle(object)

Arguments

object
a GtkWidget

Value

[GtkStyle] the resulting style. No refcount is added to the returned style, so if you want to save this style around, you should add a reference yourself.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]