gtkRcGetStyleByPaths {RGtk2}	R Documentation

gtkRcGetStyleByPaths

Description

Creates up a GtkStyle from styles defined in a RC file by providing the raw components used in matching. This function may be useful when creating pseudo-widgets that should be themed like widgets but don't actually have corresponding GTK+ widgets. An example of this would be items inside a GNOME canvas widget.

Usage

gtkRcGetStyleByPaths(settings, widget.path, class.path, type)

Arguments

settings
a GtkSettings object

widget.path
the widget path to use when looking up the style, or NULL if no matching against the widget path should be done. [ allow-none ]

class.path
the class path to use when looking up the style, or NULL if no matching against the class path should be done. [ allow-none ]

type
a type that will be used along with parent types of this type when matching against class styles, or G_TYPE_NONE

Details

The action of gtkRcGetStyle is similar to:

path <- widget$path()$path
class_path <- widget$classPath()$path
gtkRcGetStyleByPaths(widget$getSettings(), path, class_path, class(widget)[1])

Value

[GtkStyle] A style created by matching with the supplied paths, or NULL if nothing matching was specified and the default style should be used.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]