gtkRcParseColor {RGtk2}	R Documentation

gtkRcParseColor

Description

Parses a color in the format expected in a RC file.

Usage

gtkRcParseColor(scanner, color)

Arguments

scanner
a GScanner

color
a pointer to a GdkColor structure in which to store the result

Details

Note that theme engines should use gtkRcParseColorFull in order to support symbolic colors.

Value

[numeric] G_TOKEN_NONE if parsing succeeded, otherwise the token that was expected but not found

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]