gtkRcParseColorFull {RGtk2}	R Documentation

gtkRcParseColorFull

Description

Parses a color in the format expected in a RC file. If style is not NULL, it will be consulted to resolve references to symbolic colors.

Usage

gtkRcParseColorFull(scanner, style)

Arguments

scanner
a GScanner

style
a GtkRcStyle, or NULL. [ allow-none ]

Details

Since 2.12

Value

A list containing the following elements:

retval
[numeric] G_TOKEN_NONE if parsing succeeded, otherwise the token that was expected but not found

color
a pointer to a GdkColor structure in which to store the result

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]