gtkRcParsePriority {RGtk2}	R Documentation

gtkRcParsePriority

Description

Parses a GtkPathPriorityType variable from the format expected in a RC file.

Usage

gtkRcParsePriority(scanner)

Arguments

scanner
a GtkScanner (must be initialized for parsing an RC file)

Value

A list containing the following elements:

retval
[numeric] G_TOKEN_NONE if parsing succeeded, otherwise the token that was expected but not found.

priority
A pointer to GtkPathPriorityType variable in which to store the result.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]