gtkRcReparseAll {RGtk2}	R Documentation

gtkRcReparseAll

Description

If the modification time on any previously read file for the default GtkSettings has changed, discard all style information and then reread all previously read RC files.

Usage

gtkRcReparseAll()

Value

[logical] TRUE if the files were reread.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]