gtkRcReparseAllForSettings {RGtk2}	R Documentation

gtkRcReparseAllForSettings

Description

If the modification time on any previously read file for the given GtkSettings has changed, discard all style information and then reread all previously read RC files.

Usage

gtkRcReparseAllForSettings(settings, force.load)

Arguments

settings
a GtkSettings

force.load
load whether or not anything changed

Value

[logical] TRUE if the files were reread.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]