gtkRecentActionSetShowNumbers {RGtk2}	R Documentation

gtkRecentActionSetShowNumbers

Description

Sets whether a number should be added to the items shown by the widgets representing action. The numbers are shown to provide a unique character for a mnemonic to be used inside the menu item's label. Only the first ten items get a number to avoid clashes.

Usage

gtkRecentActionSetShowNumbers(object, show.numbers)

Arguments

object
a GtkRecentAction

show.numbers
TRUE if the shown items should be numbered

Details

Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]