gtkRecentChooserGetItems {RGtk2}	R Documentation

gtkRecentChooserGetItems

Description

Gets the list of recently used resources in form of GtkRecentInfo objects.

Usage

gtkRecentChooserGetItems(object)

Arguments

object
a GtkRecentChooser

Details

The return value of this function is affected by the "sort-type" and "limit" properties of chooser. Since 2.10

Value

[list] A newly allocated list of GtkRecentInfo objects. You should use gtkRecentInfoUnref on every item of the list, [ element-type GtkRecentInfo][ transfer full ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]