gtkRecentChooserMenuNew {RGtk2}	R Documentation

gtkRecentChooserMenuNew

Description

Creates a new GtkRecentChooserMenu widget.

Usage

gtkRecentChooserMenuNew()

Details

This kind of widget shows the list of recently used resources as a menu, each item as a menu item. Each item inside the menu might have an icon, representing its MIME type, and a number, for mnemonic access.

This widget implements the GtkRecentChooser interface.

This widget creates its own GtkRecentManager object. See the gtkRecentChooserMenuNewForManager function to know how to create a GtkRecentChooserMenu widget bound to another GtkRecentManager object. Since 2.10

Value

[GtkWidget] a new GtkRecentChooserMenu

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]