gtkRecentChooserMenuNewForManager {RGtk2}	R Documentation

gtkRecentChooserMenuNewForManager

Description

Creates a new GtkRecentChooserMenu widget using manager as the underlying recently used resources manager.

Usage

gtkRecentChooserMenuNewForManager(manager = NULL, show = TRUE)

Arguments

manager
a GtkRecentManager

Details

This is useful if you have implemented your own recent manager, or if you have a customized instance of a GtkRecentManager object or if you wish to share a common GtkRecentManager object among multiple GtkRecentChooser widgets. Since 2.10

Value

[GtkWidget] a new GtkRecentChooserMenu, bound to manager.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]