gtkRecentChooserMenuSetShowNumbers {RGtk2}	R Documentation

gtkRecentChooserMenuSetShowNumbers

Description

Sets whether a number should be added to the items of menu. The numbers are shown to provide a unique character for a mnemonic to be used inside ten menu item's label. Only the first the items get a number to avoid clashes.

Usage

gtkRecentChooserMenuSetShowNumbers(object, show.numbers)

Arguments

object
a GtkRecentChooserMenu

show.numbers
whether to show numbers

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]