gtkRecentChooserSetSortFunc {RGtk2}	R Documentation

gtkRecentChooserSetSortFunc

Description

Sets the comparison function used when sorting to be sort.func. If the chooser has the sort type set to GTK_RECENT_SORT_CUSTOM then the chooser will sort using this function.

Usage

gtkRecentChooserSetSortFunc(object, sort.func, sort.data)

Arguments

object
a GtkRecentChooser

sort.func
the comparison function

sort.data
user data to pass to sort.func, or NULL. [ allow-none ]

Details

To the comparison function will be passed two GtkRecentInfo structs and sort.data; sort.func should return a positive integer if the first item comes before the second, zero if the two items are equal and a negative integer if the first item comes after the second. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]