gtkRecentFilterGetNeeded {RGtk2}	R Documentation

gtkRecentFilterGetNeeded

Description

Gets the fields that need to be filled in for the structure passed to gtkRecentFilterFilter

Usage

gtkRecentFilterGetNeeded(object)

Arguments

object
a GtkRecentFilter

Details

This function will not typically be used by applications; it is intended principally for use in the implementation of GtkRecentChooser. Since 2.10

Value

[GtkRecentFilterFlags] bitfield of flags indicating needed fields when calling gtkRecentFilterFilter

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]