gtkRecentFilterNew {RGtk2}	R Documentation

gtkRecentFilterNew

Description

Creates a new GtkRecentFilter with no rules added to it. Such filter does not accept any recently used resources, so is not particularly useful until you add rules with gtkRecentFilterAddPattern, gtkRecentFilterAddMimeType, gtkRecentFilterAddApplication, gtkRecentFilterAddAge. To create a filter that accepts any recently used resource, use:

filter <- gtkRecentFilter()
filter$addPattern("*")

Usage

gtkRecentFilterNew()

Details

Since 2.10

Value

[GtkRecentFilter] a new GtkRecentFilter

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]