gtkRecentFilterSetName {RGtk2}	R Documentation

gtkRecentFilterSetName

Description

Sets the human-readable name of the filter; this is the string that will be displayed in the recently used resources selector user interface if there is a selectable list of filters.

Usage

gtkRecentFilterSetName(object, name)

Arguments

object
a GtkRecentFilter

name
then human readable name of filter

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]