gtkRecentInfoGetApplicationInfo {RGtk2}	R Documentation

gtkRecentInfoGetApplicationInfo

Description

Gets the data regarding the application that has registered the resource pointed by info.

Usage

gtkRecentInfoGetApplicationInfo(object, app.name)

Arguments

object
a GtkRecentInfo

app.name
the name of the application that has registered this item

Details

If the command line contains any escape characters defined inside the storage specification, they will be expanded. Since 2.10

Value

A list containing the following elements:

retval
[logical] TRUE if an application with app.name has registered this resource inside the recently used list, or FALSE otherwise.

app.exec
return location for the string containing the command line. [ transfer none ][ out ]

count
return location for the number of times this item was registered. [ out ]

time.
return location for the timestamp this item was last registered for this application. [ out ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]