gtkRecentInfoGetApplications {RGtk2}	R Documentation

gtkRecentInfoGetApplications

Description

Retrieves the list of applications that have registered this resource.

Usage

gtkRecentInfoGetApplications(object, length)

Arguments

object
a GtkRecentInfo

length
return location for the length of the returned list. [ out ][ allow-none ]

Details

Since 2.10

Value

[character] a newly list of strings. [ array length=length zero-terminated=1]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]