gtkRecentInfoGetGroups {RGtk2}	R Documentation

gtkRecentInfoGetGroups

Description

Returns all groups registered for the recently used item info. The array of returned group names will be NULL terminated, so length might optionally be NULL.

Usage

gtkRecentInfoGetGroups(object)

Arguments

object
a GtkRecentInfo

Details

Since 2.10

Value

A list containing the following elements:

retval
[character] a newly allocated NULL terminated list of strings. [ array length=length zero-terminated=1]

length
return location for the number of groups returned. [ out ][ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]