gtkRecentManagerAddFull {RGtk2}	R Documentation

gtkRecentManagerAddFull

Description

Adds a new resource, pointed by uri, into the recently used resources list, using the metadata specified inside the GtkRecentData structure passed in recent.data.

Usage

gtkRecentManagerAddFull(object, uri, recent.data)

Arguments

object
a GtkRecentManager

uri
a valid URI

recent.data
metadata of the resource

Details

The passed URI will be used to identify this resource inside the list.

In order to register the new recently used resource, metadata about the resource must be passed as well as the URI; the metadata is stored in a GtkRecentData structure, which must contain the MIME type of the resource pointed by the URI; the name of the application that is registering the item, and a command line to be used when launching the item.

Optionally, a GtkRecentData structure might contain a UTF-8 string to be used when viewing the item instead of the last component of the URI; a short description of the item; whether the item should be considered private - that is, should be displayed only by the applications that have registered it. Since 2.10

Value

[logical] TRUE if the new item was successfully added to the recently used resources list, FALSE otherwise.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]