gtkRecentManagerAddItem {RGtk2}	R Documentation

gtkRecentManagerAddItem

Description

Adds a new resource, pointed by uri, into the recently used resources list.

Usage

gtkRecentManagerAddItem(object, uri)

Arguments

object
a GtkRecentManager

uri
a valid URI

Details

This function automatically retrieves some of the needed metadata and setting other metadata to common default values; it then feeds the data to gtkRecentManagerAddFull.

See gtkRecentManagerAddFull if you want to explicitly define the metadata for the resource pointed by uri. Since 2.10

Value

[logical] TRUE if the new item was successfully added to the recently used resources list

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]