gtkRecentManagerGetForScreen {RGtk2}	R Documentation

gtkRecentManagerGetForScreen

Description

Gets the recent manager object associated with screen; if this function has not previously been called for the given screen, a new recent manager object will be created and associated with the screen. Recent manager objects are fairly expensive to create, so using this function is usually a better choice than calling gtkRecentManagerNew and setting the screen yourself; by using this function a single recent manager object will be shared between users. WARNING: gtk_recent_manager_get_for_screen has been deprecated since version 2.12 and should not be used in newly-written code. This function has been deprecated and should not be used in newly written code. Calling this function is equivalent to calling gtkRecentManagerGetDefault.

Usage

gtkRecentManagerGetForScreen(screen)

Arguments

screen
a GdkScreen

Details

Since 2.10

Value

[GtkRecentManager] A unique GtkRecentManager associated with the given screen. This recent manager is associated to the with the screen and can be used as long as the screen is open. Do not ref or unref it.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]