gtkRecentManagerNew {RGtk2}	R Documentation

gtkRecentManagerNew

Description

Creates a new recent manager object. Recent manager objects are used to handle the list of recently used resources. A GtkRecentManager object monitors the recently used resources list, and emits the "changed" signal each time something inside the list changes.

Usage

gtkRecentManagerNew()

Details

GtkRecentManager objects are expensive: be sure to create them only when needed. You should use gtkRecentManagerGetDefault instead. Since 2.10

Value

[GtkRecentManager] A newly created GtkRecentManager object.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]