gtkRecentManagerRemoveItem {RGtk2}	R Documentation

gtkRecentManagerRemoveItem

Description

Removes a resource pointed by uri from the recently used resources list handled by a recent manager.

Usage

gtkRecentManagerRemoveItem(object, uri, .errwarn = TRUE)

Arguments

object
a GtkRecentManager

uri
the URI of the item you wish to remove

.errwarn
Whether to issue a warning on error or fail silently

Details

Since 2.10

Value

A list containing the following elements:

retval
[logical] TRUE if the item pointed by uri has been successfully removed by the recently used resources list, and FALSE otherwise.

error
return location for a GError, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]