gtkRulerGetRange {RGtk2}	R Documentation

gtkRulerGetRange

Description

Retrieves values indicating the range and current position of a GtkRuler. See gtkRulerSetRange.

Usage

gtkRulerGetRange(object)

Arguments

object
a GtkRuler

Value

A list containing the following elements:

lower
location to store lower limit of the ruler, or NULL. [ allow-none ]

upper
location to store upper limit of the ruler, or NULL. [ allow-none ]

position
location to store the current position of the mark on the ruler, or NULL. [ allow-none ]

max.size
location to store the maximum size of the ruler used when calculating the space to leave for the text, or NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]