gtkScaleAddMark {RGtk2}	R Documentation

gtkScaleAddMark

Description

Adds a mark at value.

Usage

gtkScaleAddMark(object, value, position, markup = NULL)

Arguments

object
a GtkScale

value
the value at which the mark is placed, must be between the lower and upper limits of the scales' adjustment

position
where to draw the mark. For a horizontal scale, GTK_POS_TOP is drawn above the scale, anything else below. For a vertical scale, GTK_POS_LEFT is drawn to the left of the scale, anything else to the right.

markup
Text to be shown at the mark, using Pango markup, or NULL. [ allow-none ]

Details

A mark is indicated visually by drawing a tick mark next to the scale, and GTK+ makes it easy for the user to position the scale exactly at the marks value.

If markup is not NULL, text is shown next to the tick mark.

To remove marks from a scale, use gtkScaleClearMarks. Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]