gtkScaleButtonNew {RGtk2}	R Documentation

gtkScaleButtonNew

Description

Creates a GtkScaleButton, with a range between min and max, with a stepping of step.

Usage

gtkScaleButtonNew(size, min, max, step, icons, show = TRUE)

Arguments

size
a stock icon size. [ in ]

min
the minimum value of the scale (usually 0)

max
the maximum value of the scale (usually 100)

step
the stepping of value when a scroll-wheel event, or up/down arrow event occurs (usually 2)

icons
a list of icon names, or NULL if you want to set the list later with gtkScaleButtonSetIcons. [ allow-none ]

Details

Since 2.12

Value

[GtkWidget] a new GtkScaleButton

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]