gtkScaleButtonSetValue {RGtk2}	R Documentation

gtkScaleButtonSetValue

Description

Sets the current value of the scale; if the value is outside the minimum or maximum range values, it will be clamped to fit inside them. The scale button emits the "value-changed" signal if the value changes.

Usage

gtkScaleButtonSetValue(object, value)

Arguments

object
a GtkScaleButton

value
new value of the scale button

Details

Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]