gtkScaleSetDigits {RGtk2}	R Documentation

gtkScaleSetDigits

Description

Sets the number of decimal places that are displayed in the value. Also causes the value of the adjustment to be rounded off to this number of digits, so the retrieved value matches the value the user saw.

Usage

gtkScaleSetDigits(object, digits)

Arguments

object
a GtkScale

digits
the number of decimal places to display, e.g. use 1 to display 1.0, 2 to display 1.00, etc

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]