gtkScrolledWindowGetPolicy {RGtk2}	R Documentation

gtkScrolledWindowGetPolicy

Description

Retrieves the current policy values for the horizontal and vertical scrollbars. See gtkScrolledWindowSetPolicy.

Usage

gtkScrolledWindowGetPolicy(object)

Arguments

object
a GtkScrolledWindow

Value

A list containing the following elements:

hscrollbar.policy
location to store the policy for the horizontal scrollbar, or NULL.

vscrollbar.policy
location to store the policy for the vertical scrollbar, or NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]