gtkScrolledWindowSetPlacement {RGtk2}	R Documentation

gtkScrolledWindowSetPlacement

Description

Sets the placement of the contents with respect to the scrollbars for the scrolled window.

Usage

gtkScrolledWindowSetPlacement(object, window.placement)

Arguments

object
a GtkScrolledWindow

window.placement
position of the child window

Details

The default is GTK_CORNER_TOP_LEFT, meaning the child is in the top left, with the scrollbars underneath and to the right. Other values in GtkCornerType are GTK_CORNER_TOP_RIGHT, GTK_CORNER_BOTTOM_LEFT, and GTK_CORNER_BOTTOM_RIGHT.

See also gtkScrolledWindowGetPlacement and gtkScrolledWindowUnsetPlacement.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]