gtkScrolledWindowSetPolicy {RGtk2}	R Documentation

gtkScrolledWindowSetPolicy

Description

Sets the scrollbar policy for the horizontal and vertical scrollbars.

Usage

gtkScrolledWindowSetPolicy(object, hscrollbar.policy, vscrollbar.policy)

Arguments

object
a GtkScrolledWindow

hscrollbar.policy
policy for horizontal bar

vscrollbar.policy
policy for vertical bar

Details

The policy determines when the scrollbar should appear; it is a value from the GtkPolicyType enumeration. If GTK_POLICY_ALWAYS, the scrollbar is always present; if GTK_POLICY_NEVER, the scrollbar is never present; if GTK_POLICY_AUTOMATIC, the scrollbar is present only if needed (that is, if the slider part of the bar would be smaller than the trough - the display is larger than the page size).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]