gtkSelectionClear {RGtk2}	R Documentation

gtkSelectionClear

Description

The default handler for the "selection-clear-event" signal. WARNING: gtk_selection_clear has been deprecated since version 2.4 and should not be used in newly-written code. Instead of calling this function, chain up from your selection-clear-event handler. Calling this function from any other context is illegal.

Usage

gtkSelectionClear(object, event)

Arguments

object
a GtkWidget

event
the event

Details

Since 2.2

Value

[logical] TRUE if the event was handled, otherwise false

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]