gtkSelectionDataGetPixbuf {RGtk2}	R Documentation

gtkSelectionDataGetPixbuf

Description

Gets the contents of the selection data as a GdkPixbuf.

Usage

gtkSelectionDataGetPixbuf(object)

Arguments

object
a GtkSelectionData

Details

Since 2.6

Value

[GdkPixbuf] if the selection data contained a recognized image type and it could be converted to a GdkPixbuf, a newly allocated pixbuf is returned, otherwise NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]