gtkSelectionDataGetTargets {RGtk2}	R Documentation

gtkSelectionDataGetTargets

Description

Gets the contents of selection.data as a list of targets. This can be used to interpret the results of getting the standard TARGETS target that is always supplied for any selection.

Usage

gtkSelectionDataGetTargets(object)

Arguments

object
a GtkSelectionData object

Value

A list containing the following elements:

retval
[logical] TRUE if selection.data contains a valid array of targets, otherwise FALSE.

targets
location to store a list of targets. The result stored here must be freed with gFree().

n.atoms
location to store number of items in targets.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]