gtkSelectionDataGetText {RGtk2}	R Documentation

gtkSelectionDataGetText

Description

Gets the contents of the selection data as a UTF-8 string.

Usage

gtkSelectionDataGetText(object)

Arguments

object
a GtkSelectionData

Value

[raw] if the selection data contained a recognized text type and it could be converted to UTF-8, a newly allocated string containing the converted text, otherwise NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]