gtkSelectionDataGetUris {RGtk2}	R Documentation

gtkSelectionDataGetUris

Description

Gets the contents of the selection data as list of URIs.

Usage

gtkSelectionDataGetUris(object)

Arguments

object
a GtkSelectionData

Details

Since 2.6

Value

[character] if the selection data contains a list of URIs, a newly allocated string list containing the URIs, otherwise NULL. [ array zero-terminated=1][ element-type utf8][ transfer full ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]