gtkShow {RGtk2}	R Documentation

Show/realize one or more Widgets

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

This causes each widget given in the call to be realized or actually displayed

Usage

gtkShow(..., all = T)

Arguments

...
any collection of widgets that are to be realized. These are processed in the order they are given.

all
a logical value indicating whether to show the particular widget (FALSE), or both it and all of the sub-widgets which it contains (TRUE). This is the difference between a call to gtk_widget_show and gtk_widget_show_all in the lower-level Gtk interface.

Details

This calls gtk_widget_show or gtk_widget_show_all for each widget.

Value

An integer giving the number of widgets that were realized in the call.

Author(s)

Duncan Temple Lang

References

http://www.gtk.org http://www.omegahat.org/RGtk

See Also

gtkWidgetShow

Examples

  if (gtkInit()) {
    w <- gtkWindow(show = FALSE) 
    b <- gtkButton("Hit me") # automatically shown in the creation.
    w$add(b)
    gtkShow(w)
  }

[Package RGtk2 version 2.20.31 Index]