gtkShowAboutDialog {RGtk2}	R Documentation

gtkShowAboutDialog

Description

This is a convenience function for showing an application's about box. The constructed dialog is associated with the parent window and reused for future invocations of this function.

Usage

gtkShowAboutDialog(parent, ...)

Arguments

parent
transient parent, or NULL for none. [ allow-none ]

...
undocumented

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]