gtkShowUri {RGtk2}	R Documentation

gtkShowUri

Description

This is a convenience function for launching the default application to show the uri. The uri must be of a form understood by GIO. Typical examples are

‘file:///home/gnome/pict.jpg’

‘http://www.gnome.org’

‘mailto:me’

Ideally the timestamp is taken from the event triggering the gtkShowUri call. If timestamp is not known you can take GDK_CURRENT_TIME.

Usage

gtkShowUri(screen = NULL, uri, timestamp, .errwarn = TRUE)

Arguments

screen
screen to show the uri on or NULL for the default screen. [ allow-none ]

uri
the uri to show

timestamp
a timestamp to prevent focus stealing.

.errwarn
Whether to issue a warning on error or fail silently

Details

This function can be used as a replacement for gnomeVfsUrlShow() and gnomeUrlShow(). Since 2.14

Value

A list containing the following elements:

retval
[logical] TRUE on success, FALSE on error.

error
a GError that is returned in case of errors

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]