gtkSizeGroupAddWidget {RGtk2}	R Documentation

gtkSizeGroupAddWidget

Description

Adds a widget to a GtkSizeGroup. In the future, the requisition of the widget will be determined as the maximum of its requisition and the requisition of the other widgets in the size group. Whether this applies horizontally, vertically, or in both directions depends on the mode of the size group. See gtkSizeGroupSetMode.

Usage

gtkSizeGroupAddWidget(object, widget)

Arguments

object
a GtkSizeGroup

widget
the GtkWidget to add

Details

When the widget is destroyed or no longer referenced elsewhere, it will be removed from the size group.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]