gtkSizeGroupSetMode {RGtk2}	R Documentation

gtkSizeGroupSetMode

Description

Sets the GtkSizeGroupMode of the size group. The mode of the size group determines whether the widgets in the size group should all have the same horizontal requisition (GTK_SIZE_GROUP_MODE_HORIZONTAL) all have the same vertical requisition (GTK_SIZE_GROUP_MODE_VERTICAL), or should all have the same requisition in both directions (GTK_SIZE_GROUP_MODE_BOTH).

Usage

gtkSizeGroupSetMode(object, mode)

Arguments

object
a GtkSizeGroup

mode
the mode to set for the size group.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]