gtkSocketAddId {RGtk2}	R Documentation

gtkSocketAddId

Description

Adds an XEMBED client, such as a GtkPlug, to the GtkSocket. The client may be in the same process or in a different process.

Usage

gtkSocketAddId(object, window.id)

Arguments

object
a GtkSocket

window.id
the window ID of a client participating in the XEMBED protocol.

Details

To embed a GtkPlug in a GtkSocket, you can either create the GtkPlug with gtk_plug_new (0), call gtkPlugGetId to get the window ID of the plug, and then pass that to the gtkSocketAddId, or you can call gtkSocketGetId to get the window ID for the socket, and call gtkPlugNew passing in that ID.

The GtkSocket must have already be added into a toplevel window before you can make this call.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]