gtkSocketGetId {RGtk2}	R Documentation

gtkSocketGetId

Description

Gets the window ID of a GtkSocket widget, which can then be used to create a client embedded inside the socket, for instance with gtkPlugNew.

Usage

gtkSocketGetId(object)

Arguments

object
a GtkSocket.

Details

The GtkSocket must have already be added into a toplevel window before you can make this call.

Value

[GdkNativeWindow] the window ID for the socket

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]