gtkSocketSteal {RGtk2}	R Documentation

gtkSocketSteal

Description

Reparents a pre-existing toplevel window into a GtkSocket. This is meant to embed clients that do not know about embedding into a GtkSocket, however doing so is inherently unreliable, and using this function is not recommended. WARNING: gtk_socket_steal is deprecated and should not be used in newly-written code.

Usage

gtkSocketSteal(object, wid)

Arguments

object
a GtkSocket

wid
the window ID of an existing toplevel window.

Details

The GtkSocket must have already be added into a toplevel window before you can make this call.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]