gtkSpinButtonConfigure {RGtk2}	R Documentation

gtkSpinButtonConfigure

Description

Changes the properties of an existing spin button. The adjustment, climb rate, and number of decimal places are all changed accordingly, after this function call.

Usage

gtkSpinButtonConfigure(object, adjustment = NULL, climb.rate, 
    digits)

Arguments

object
a GtkSpinButton

adjustment
a GtkAdjustment. [ allow-none ]

climb.rate
the new climb rate.

digits
the number of decimal places to display in the spin button.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]