gtkSpinButtonGetWrap {RGtk2}	R Documentation

gtkSpinButtonGetWrap

Description

Returns whether the spin button's value wraps around to the opposite limit when the upper or lower limit of the range is exceeded. See gtkSpinButtonSetWrap.

Usage

gtkSpinButtonGetWrap(object)

Arguments

object
a GtkSpinButton

Value

[logical] TRUE if the spin button wraps around

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]