gtkSpinButtonNewWithRange {RGtk2}	R Documentation

gtkSpinButtonNewWithRange

Description

This is a convenience constructor that allows creation of a numeric GtkSpinButton without manually creating an adjustment. The value is initially set to the minimum value and a page increment of 10 * step is the default. The precision of the spin button is equivalent to the precision of step.

Usage

gtkSpinButtonNewWithRange(min, max, step, show = TRUE)

Arguments

min
Minimum allowable value

max
Maximum allowable value

step
Increment added or subtracted by spinning the widget

Details

Note that the way in which the precision is derived works best if step is a power of ten. If the resulting precision is not suitable for your needs, use gtkSpinButtonSetDigits to correct it.

Value

[GtkWidget] The new spin button as a GtkWidget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]