gtkSpinButtonSetSnapToTicks {RGtk2}	R Documentation

gtkSpinButtonSetSnapToTicks

Description

Sets the policy as to whether values are corrected to the nearest step increment when a spin button is activated after providing an invalid value.

Usage

gtkSpinButtonSetSnapToTicks(object, snap.to.ticks)

Arguments

object
a GtkSpinButton

snap.to.ticks
a flag indicating if invalid values should be corrected.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]