gtkSpinButtonSetUpdatePolicy {RGtk2}	R Documentation

gtkSpinButtonSetUpdatePolicy

Description

Sets the update behavior of a spin button. This determines whether the spin button is always updated or only when a valid value is set.

Usage

gtkSpinButtonSetUpdatePolicy(object, policy)

Arguments

object
a GtkSpinButton

policy
a GtkSpinButtonUpdatePolicy value

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]