gtkSpinButtonSetWrap {RGtk2}	R Documentation

gtkSpinButtonSetWrap

Description

Sets the flag that determines if a spin button value wraps around to the opposite limit when the upper or lower limit of the range is exceeded.

Usage

gtkSpinButtonSetWrap(object, wrap)

Arguments

object
a GtkSpinButton

wrap
a flag indicating if wrapping behavior is performed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]