gtkStatusIconGetGeometry {RGtk2}	R Documentation

gtkStatusIconGetGeometry

Description

Obtains information about the location of the status icon on screen. This information can be used to e.g. position popups like notification bubbles.

Usage

gtkStatusIconGetGeometry(object)

Arguments

object
a GtkStatusIcon

Details

See gtkStatusIconPositionMenu for a more convenient alternative for positioning menus.

Note that some platforms do not allow GTK+ to provide this information, and even on platforms that do allow it, the information is not reliable unless the status icon is embedded in a notification area, see gtkStatusIconIsEmbedded. Since 2.10

Value

A list containing the following elements:

retval
[logical] TRUE if the location information has been filled in

screen
return location for the screen, or NULL if the information is not needed. [ out ][ transfer none ][ allow-none ]

area
return location for the area occupied by the status icon, or NULL. [ out ][ allow-none ]

orientation
return location for the orientation of the panel in which the status icon is embedded, or NULL. A panel at the top or bottom of the screen is horizontal, a panel at the left or right is vertical. [ out ][ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]