gtkStatusIconGetPixbuf {RGtk2}	R Documentation

gtkStatusIconGetPixbuf

Description

Gets the GdkPixbuf being displayed by the GtkStatusIcon. The storage type of the status icon must be GTK_IMAGE_EMPTY or GTK_IMAGE_PIXBUF (see gtkStatusIconGetStorageType). The caller of this function does not own a reference to the returned pixbuf.

Usage

gtkStatusIconGetPixbuf(object)

Arguments

object
a GtkStatusIcon

Details

Since 2.10

Value

[GdkPixbuf] the displayed pixbuf, or NULL if the image is empty.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]