gtkStatusIconGetSize {RGtk2}	R Documentation

gtkStatusIconGetSize

Description

Gets the size in pixels that is available for the image. Stock icons and named icons adapt their size automatically if the size of the notification area changes. For other storage types, the size-changed signal can be used to react to size changes.

Usage

gtkStatusIconGetSize(object)

Arguments

object
a GtkStatusIcon

Details

Note that the returned size is only meaningful while the status icon is embedded (see gtkStatusIconIsEmbedded). Since 2.10

Value

[integer] the size that is available for the image

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]