gtkStatusIconGetX11WindowId {RGtk2}	R Documentation

gtkStatusIconGetX11WindowId

Description

This function is only useful on the X11/freedesktop.org platform. It returns a window ID for the widget in the underlying status icon implementation. This is useful for the Galago notification service, which can send a window ID in the protocol in order for the server to position notification windows pointing to a status icon reliably.

Usage

gtkStatusIconGetX11WindowId(object)

Arguments

object
a GtkStatusIcon

Details

This function is not intended for other use cases which are more likely to be met by one of the non-X11 specific methods, such as gtkStatusIconPositionMenu. Since 2.14

Value

[numeric] An 32 bit unsigned integer identifier for the underlying X11 Window

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]