gtkStatusIconNewFromFile {RGtk2}	R Documentation

gtkStatusIconNewFromFile

Description

Creates a status icon displaying the file filename.

Usage

gtkStatusIconNewFromFile(filename)

Arguments

filename
a filename

Details

The image will be scaled down to fit in the available space in the notification area, if necessary. Since 2.10

Value

[GtkStatusIcon] a new GtkStatusIcon

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]