gtkStatusIconPositionMenu {RGtk2}	R Documentation

gtkStatusIconPositionMenu

Description

Menu positioning function to use with gtkMenuPopup to position menu aligned to the status icon user.data.

Usage

gtkStatusIconPositionMenu(menu, user.data)

Arguments

menu
the GtkMenu

user.data
the status icon to position the menu on

Details

Since 2.10

Value

A list containing the following elements:

x
return location for the x position

y
return location for the y position

push.in
whether the first menu item should be offset (pushed in) to be aligned with the menu popup position (only useful for GtkOptionMenu).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]