gtkStatusIconSetBlinking {RGtk2}	R Documentation

gtkStatusIconSetBlinking

Description

Makes the status icon start or stop blinking. Note that blinking user interface elements may be problematic for some users, and thus may be turned off, in which case this setting has no effect.

Usage

gtkStatusIconSetBlinking(object, blinking)

Arguments

object
a GtkStatusIcon

blinking
TRUE to turn blinking on, FALSE to turn it off

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]