gtkStatusIconSetName {RGtk2}	R Documentation

gtkStatusIconSetName

Description

Sets the name of this tray icon. This should be a string identifying this icon. It is may be used for sorting the icons in the tray and will not be shown to the user.

Usage

gtkStatusIconSetName(object, name)

Arguments

object
a GtkStatusIcon

name
the name

Details

Since 2.20

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]