gtkStatusIconSetTooltipMarkup {RGtk2}	R Documentation

gtkStatusIconSetTooltipMarkup

Description

Sets markup as the contents of the tooltip, which is marked up with the Pango text markup language.

Usage

gtkStatusIconSetTooltipMarkup(object, markup = NULL)

Arguments

object
a GtkStatusIcon

markup
the contents of the tooltip for status.icon, or NULL. [ allow-none ]

Details

This function will take care of setting "has-tooltip" to TRUE and of the default handler for the "query-tooltip" signal.

See also the "tooltip-markup" property and gtkTooltipSetMarkup. Since 2.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]