gtkStatusbarPop {RGtk2}	R Documentation

gtkStatusbarPop

Description

Removes the first message in the GtkStatusBar's stack with the given context id.

Usage

gtkStatusbarPop(object, context.id)

Arguments

object
a GtkStatusBar

context.id
a context identifier

Details

Note that this may not change the displayed message, if the message at the top of the stack has a different context id.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]