gtkStockAdd {RGtk2}	R Documentation

gtkStockAdd

Description

Registers each of the stock items in items. If an item already exists with the same stock ID as one of the items, the old item gets replaced. The stock items are copied, so GTK+ does not hold any pointer into items and items can be freed. Use gtkStockAddStatic if items is persistent and GTK+ need not copy the list.

Usage

gtkStockAdd(items)

Arguments

items
a GtkStockItem or list of items

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]