gtkStyleAttach {RGtk2}	R Documentation

gtkStyleAttach

Description

Attaches a style to a window; this process allocates the colors and creates the GC's for the style - it specializes it to a particular visual and colormap. The process may involve the creation of a new style if the style has already been attached to a window with a different style and colormap.

Usage

gtkStyleAttach(object, window)

Arguments

object
a GtkStyle.

window
a GdkWindow.

Details

Since this function may return a new object, you have to use it in the following way: style = gtk_style_attach (style, window)

Value

[GtkStyle] Either style, or a newly-created GtkStyle. If the style is newly created, the style parameter will be unref'ed, and the new style will have a reference count belonging to the caller.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]