gtkStyleGetStyleProperty {RGtk2}	R Documentation

gtkStyleGetStyleProperty

Description

Queries the value of a style property corresponding to a widget class is in the given style.

Usage

gtkStyleGetStyleProperty(object, widget.type, property.name)

Arguments

object
a GtkStyle

widget.type
the GType of a descendant of GtkWidget

property.name
the name of the style property to get

Details

Since 2.16

Value

A list containing the following elements:

value
a R object where the value of the property being queried will be stored

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]