gtkStyleLookupColor {RGtk2}	R Documentation

gtkStyleLookupColor

Description

Looks up color.name in the style's logical color mappings, filling in color and returning TRUE if found, otherwise returning FALSE. Do not cache the found mapping, because it depends on the GtkStyle and might change when a theme switch occurs.

Usage

gtkStyleLookupColor(object, color.name)

Arguments

object
a GtkStyle

color.name
the name of the logical color to look up

Details

Since 2.10

Value

A list containing the following elements:

retval
[logical] TRUE if the mapping was found.

color
the GdkColor to fill in

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]