gtkStyleRenderIcon {RGtk2}	R Documentation

gtkStyleRenderIcon

Description

Renders the icon specified by source at the given size according to the given parameters and returns the result in a pixbuf.

Usage

gtkStyleRenderIcon(object, source, direction, state, size, widget = NULL, 
    detail = NULL)

Arguments

object
a GtkStyle

source
the GtkIconSource specifying the icon to render

direction
a text direction

state
a state

size
(type int) the size to render the icon at. A size of (GtkIconSize)-1 means render at the size of the source and don't scale.

widget
the widget. [ allow-none ]

detail
a style detail. [ allow-none ]

Value

[GdkPixbuf] a newly-created GdkPixbuf containing the rendered icon

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]