gtkStyleSetFont {RGtk2}	R Documentation

gtkStyleSetFont

Description

Sets the GdkFont to use for a given style. This is meant only as a replacement for direct access to style->font and should not be used in new code. New code should use style->font_desc instead. WARNING: gtk_style_set_font is deprecated and should not be used in newly-written code.

Usage

gtkStyleSetFont(object, font)

Arguments

object
a GtkStyle.

font
a GdkFont, or NULL to use the GdkFont corresponding to style->font_desc. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]