gtkTableAttach {RGtk2}	R Documentation

gtkTableAttach

Description

Adds a widget to a table. The number of 'cells' that a widget will occupy is specified by left.attach, right.attach, top.attach and bottom.attach. These each represent the leftmost, rightmost, uppermost and lowest column and row numbers of the table. (Columns and rows are indexed from zero).

Usage

gtkTableAttach(object, child, left.attach, right.attach, top.attach, 
    bottom.attach, xoptions = 5, yoptions = 5, xpadding = 0, 
    ypadding = 0)

Arguments

object
The GtkTable to add a new widget to.

child
The widget to add.

left.attach
the column number to attach the left side of a child widget to.

right.attach
the column number to attach the right side of a child widget to.

top.attach
the row number to attach the top of a child widget to.

bottom.attach
the row number to attach the bottom of a child widget to.

xoptions
Used to specify the properties of the child widget when the table is resized.

yoptions
The same as xoptions, except this field determines behaviour of vertical resizing.

xpadding
An integer value specifying the padding on the left and right of the widget being added to the table.

ypadding
The amount of padding above and below the child widget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]