gtkTableAttachDefaults {RGtk2}	R Documentation

gtkTableAttachDefaults

Description

As there are many options associated with gtkTableAttach, this convenience function provides the programmer with a means to add children to a table with identical padding and expansion options. The values used for the GtkAttachOptions are GTK_EXPAND | GTK_FILL, and the padding is set to 0.

Usage

gtkTableAttachDefaults(object, widget, left.attach, right.attach, 
    top.attach, bottom.attach)

Arguments

object
The table to add a new child widget to.

widget
The child widget to add.

left.attach
The column number to attach the left side of the child widget to.

right.attach
The column number to attach the right side of the child widget to.

top.attach
The row number to attach the top of the child widget to.

bottom.attach
The row number to attach the bottom of the child widget to.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]