gtkTableNew {RGtk2}	R Documentation

gtkTableNew

Description

Used to create a new table widget. An initial size must be given by specifying how many rows and columns the table should have, although this can be changed later with gtkTableResize. rows and columns must both be in the range 0 .. 65535.

Usage

gtkTableNew(rows = NULL, columns = NULL, homogeneous = NULL, 
    show = TRUE)

Arguments

rows
The number of rows the new table should have.

columns
The number of columns the new table should have.

homogeneous
If set to TRUE, all table cells are resized to the size of the cell containing the largest widget.

Value

[GtkWidget] A pointer to the the newly created table widget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]