gtkTableSetColSpacing {RGtk2}	R Documentation

gtkTableSetColSpacing

Description

Alters the amount of space between a given table column and the following column.

Usage

gtkTableSetColSpacing(object, column, spacing)

Arguments

object
a GtkTable.

column
the column whose spacing should be changed.

spacing
number of pixels that the spacing should take up.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]