gtkTableSetRowSpacing {RGtk2}	R Documentation

gtkTableSetRowSpacing

Description

Changes the space between a given table row and the subsequent row.

Usage

gtkTableSetRowSpacing(object, row, spacing)

Arguments

object
a GtkTable containing the row whose properties you wish to change.

row
row number whose spacing will be changed.

spacing
number of pixels that the spacing should take up.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]