gtkTargetEntry {RGtk2}	R Documentation

Create a new target for drag-n-drop(?)

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

This creates a new GtkTarget object. This is used in transferring data in the form of selections between processes or widgets.

Usage

gtkTargetEntry(target, flags, info)
gtkTargetEntryNew(target, flags, info)

Arguments

target
a string

flags
an integer value

info
an integer value

Details

This allocates a new GtkTargetEntry object in C and fills in the targer, flags and info fields.

Value

A reference to the newly creatd GtkTargetEntry.

Author(s)

Duncan Temple Lang <duncan@research.bell-labs.com>

References

Information on the package is available from http://www.omegahat.org/RGtk.

Information on Gtk is available from http://www.gtk.org.

[Package RGtk2 version 2.20.31 Index]