gtkTargetListAddRichTextTargets {RGtk2}	R Documentation

gtkTargetListAddRichTextTargets

Description

Appends the rich text targets registered with gtkTextBufferRegisterSerializeFormat or gtkTextBufferRegisterDeserializeFormat to the target list. All targets are added with the same info.

Usage

gtkTargetListAddRichTextTargets(list, info, deserializable, buffer)

Arguments

list
a GtkTargetList

info
an ID that will be passed back to the application

deserializable
if TRUE, then deserializable rich text formats will be added, serializable formats otherwise.

buffer
a GtkTextBuffer.

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]