gtkTargetsIncludeImage {RGtk2}	R Documentation

gtkTargetsIncludeImage

Description

Determines if any of the targets in targets can be used to provide a GdkPixbuf.

Usage

gtkTargetsIncludeImage(targets, writable)

Arguments

targets
a list of GdkAtoms

writable
whether to accept only targets for which GTK+ knows how to convert a pixbuf into the format

Details

Since 2.10

Value

[logical] TRUE if targets include a suitable target for images, otherwise FALSE.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]