gtkTestFindLabel {RGtk2}	R Documentation

gtkTestFindLabel

Description

This function will search widget and all its descendants for a GtkLabel widget with a text string matching label.pattern. The label.pattern may contain asterisks '*' and question marks '?' as placeholders, gPatternMatch() is used for the matching. Note that locales other than "C" tend to alter (translate" label strings, so this function is genrally only useful in test programs with predetermined locales, see gtkTestInit() for more details.

Usage

gtkTestFindLabel(widget, label.pattern)

Arguments

widget
Valid label or container widget.

label.pattern
Shell-glob pattern to match a label string.

Details

Since 2.14

Value

[GtkWidget] a GtkLabel widget if any is found.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]