gtkTestFindWidget {RGtk2}	R Documentation

gtkTestFindWidget

Description

This function will search the descendants of widget for a widget of type widget.type that has a label matching label.pattern next to it. This is most useful for automated GUI testing, e.g. to find the "OK" button in a dialog and synthesize clicks on it. However see gtkTestFindLabel, gtkTestFindSibling and gtkTestWidgetClick for possible caveats involving the search of such widgets and synthesizing widget events.

Usage

gtkTestFindWidget(widget, label.pattern, widget.type)

Arguments

widget
Container widget, usually a GtkWindow.

label.pattern
Shell-glob pattern to match a label string.

widget.type
Type of a aearched for label sibling widget.

Details

Since 2.14

Value

[GtkWidget] a valid widget if any is found or NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]