gtkTestSliderGetValue {RGtk2}	R Documentation

gtkTestSliderGetValue

Description

Retrive the literal adjustment value for GtkRange based widgets and spin buttons. Note that the value returned by this function is anything between the lower and upper bounds of the adjustment belonging to widget, and is not a percentage as passed in to gtkTestSliderSetPerc.

Usage

gtkTestSliderGetValue(widget)

Arguments

widget
valid widget pointer.

Details

Since 2.14

Value

[numeric] adjustment->value for an adjustment belonging to widget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]