gtkTestSliderSetPerc {RGtk2}	R Documentation

gtkTestSliderSetPerc

Description

This function will adjust the slider position of all GtkRange based widgets, such as scrollbars or scales, it'll also adjust spin buttons. The adjustment value of these widgets is set to a value between the lower and upper limits, according to the percentage argument.

Usage

gtkTestSliderSetPerc(widget, percentage)

Arguments

widget
valid widget pointer.

percentage
value between 0 and 100.

Details

Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]