gtkTestSpinButtonClick {RGtk2}	R Documentation

gtkTestSpinButtonClick

Description

This function will generate a button click in the upwards or downwards spin button arrow areas, usually leading to an increase or decrease of spin button's value.

Usage

gtkTestSpinButtonClick(spinner, button, upwards)

Arguments

spinner
valid GtkSpinButton widget.

button
Number of the pointer button for the event, usually 1, 2 or 3.

upwards
TRUE for upwards arrow click, FALSE for downwards arrow click.

Details

Since 2.14

Value

[logical] wether all actions neccessary for the button click simulation were carried out successfully.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]