gtkTestWidgetClick {RGtk2}	R Documentation

gtkTestWidgetClick

Description

This function will generate a button click (button press and button release event) in the middle of the first GdkWindow found that belongs to widget. For GTK_NO_WINDOW widgets like GtkButton, this will often be an input-only event window. For other widgets, this is usually widget->window. Certain caveats should be considered when using this function, in particular because the mouse pointer is warped to the button click location, see gdkTestSimulateButton for details.

Usage

gtkTestWidgetClick(widget, button, modifiers)

Arguments

widget
Widget to generate a button click on.

button
Number of the pointer button for the event, usually 1, 2 or 3.

modifiers
Keyboard modifiers the event is setup with.

Details

Since 2.14

Value

[logical] wether all actions neccessary for the button click simulation were carried out successfully.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]