gtkTestWidgetSendKey {RGtk2}	R Documentation

gtkTestWidgetSendKey

Description

This function will generate keyboard press and release events in the middle of the first GdkWindow found that belongs to widget. For GTK_NO_WINDOW widgets like GtkButton, this will often be an input-only event window. For other widgets, this is usually widget->window. Certain caveats should be considered when using this function, in particular because the mouse pointer is warped to the key press location, see gdkTestSimulateKey for details.

Usage

gtkTestWidgetSendKey(widget, keyval, modifiers)

Arguments

widget
Widget to generate a key press and release on.

keyval
A Gdk keyboard value.

modifiers
Keyboard modifiers the event is setup with.

Details

Since 2.14

Value

[logical] wether all actions neccessary for the key event simulation were carried out successfully.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]