gtkTextBufferAddMark {RGtk2}	R Documentation

gtkTextBufferAddMark

Description

Adds the mark at position where. The mark must not be added to another buffer, and if its name is not NULL then there must not be another mark in the buffer with the same name.

Usage

gtkTextBufferAddMark(object, mark, where)

Arguments

object
a GtkTextBuffer

mark
the mark to add

where
location to place mark

Details

Emits the "mark-set" signal as notification of the mark's initial placement. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]