gtkTextBufferApplyTag {RGtk2}	R Documentation

gtkTextBufferApplyTag

Description

Emits the "apply-tag" signal on buffer. The default handler for the signal applies tag to the given range. start and end do not have to be in order.

Usage

gtkTextBufferApplyTag(object, tag, start, end)

Arguments

object
a GtkTextBuffer

tag
a GtkTextTag

start
one bound of range to be tagged

end
other bound of range to be tagged

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]