gtkTextBufferBackspace {RGtk2}	R Documentation

gtkTextBufferBackspace

Description

Performs the appropriate action as if the user hit the delete key with the cursor at the position specified by iter. In the normal case a single character will be deleted, but when combining accents are involved, more than one character can be deleted, and when precomposed character and accent combinations are involved, less than one character will be deleted.

Usage

gtkTextBufferBackspace(object, iter, interactive, default.editable)

Arguments

object
a GtkTextBuffer

iter
a position in buffer

interactive
whether the deletion is caused by user interaction

default.editable
whether the buffer is editable by default

Details

Because the buffer is modified, all outstanding iterators become invalid after calling this function; however, the iter will be re-initialized to point to the location where text was deleted. Since 2.6

Value

[logical] TRUE if the buffer was modified

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]