gtkTextBufferBeginUserAction {RGtk2}	R Documentation

gtkTextBufferBeginUserAction

Description

Called to indicate that the buffer operations between here and a call to gtkTextBufferEndUserAction are part of a single user-visible operation. The operations between gtkTextBufferBeginUserAction and gtkTextBufferEndUserAction can then be grouped when creating an undo stack. GtkTextBuffer maintains a count of calls to gtkTextBufferBeginUserAction that have not been closed with a call to gtkTextBufferEndUserAction, and emits the "begin-user-action" and "end-user-action" signals only for the outermost pair of calls. This allows you to build user actions from other user actions.

Usage

gtkTextBufferBeginUserAction(object)

Arguments

object
a GtkTextBuffer

Details

The "interactive" buffer mutation functions, such as gtkTextBufferInsertInteractive, automatically call begin/end user action around the buffer operations they perform, so there's no need to add extra calls if you user action consists solely of a single call to one of those functions.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]