gtkTextBufferCreateMark {RGtk2}	R Documentation

gtkTextBufferCreateMark

Description

Creates a mark at position where. If mark.name is NULL, the mark is anonymous; otherwise, the mark can be retrieved by name using gtkTextBufferGetMark. If a mark has left gravity, and text is inserted at the mark's current location, the mark will be moved to the left of the newly-inserted text. If the mark has right gravity (left.gravity = FALSE), the mark will end up on the right of newly-inserted text. The standard left-to-right cursor is a mark with right gravity (when you type, the cursor stays on the right side of the text you're typing).

Usage

gtkTextBufferCreateMark(object, mark.name = NULL, where, left.gravity = FALSE)

Arguments

object
a GtkTextBuffer

mark.name
name for mark, or NULL. [ allow-none ]

where
location to place mark

left.gravity
whether the mark has left gravity

Details

The caller of this function does not own a reference to the returned GtkTextMark, so you can ignore the return value if you like.

Emits the "mark-set" signal as notification of the mark's initial placement.

Value

[GtkTextMark] the new GtkTextMark object

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]