gtkTextBufferCreateTag {RGtk2}	R Documentation

gtkTextBufferCreateTag

Description

Creates a tag and adds it to the tag table for buffer. Equivalent to calling gtkTextTagNew and then adding the tag to the buffer's tag table. so the ref count will be equal to one.

Usage

gtkTextBufferCreateTag(object, tag.name, ...)

Arguments

object
a GtkTextBuffer

tag.name
name of the new tag, or NULL. [ allow-none ]

...
a new tag

Details

If tag.name is NULL, the tag is anonymous.

If tag.name is non-NULL, a tag called tag.name must not already exist in the tag table for this buffer.

The first.property.name argument and subsequent arguments are a list of properties to set on the tag, as with gObjectSet.

Value

[GtkTextTag] a new tag

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]