gtkTextBufferDelete {RGtk2}	R Documentation

gtkTextBufferDelete

Description

Deletes text between start and end. The order of start and end is not actually relevant; gtkTextBufferDelete will reorder them. This function actually emits the "delete-range" signal, and the default handler of that signal deletes the text. Because the buffer is modified, all outstanding iterators become invalid after calling this function; however, the start and end will be re-initialized to point to the location where text was deleted.

Usage

gtkTextBufferDelete(object, start, end)

Arguments

object
a GtkTextBuffer

start
a position in buffer

end
another position in buffer

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]