gtkTextBufferDeleteInteractive {RGtk2}	R Documentation

gtkTextBufferDeleteInteractive

Description

Deletes all editable text in the given range. Calls gtkTextBufferDelete for each editable sub-range of [start,end). start and end are revalidated to point to the location of the last deleted range, or left untouched if no text was deleted.

Usage

gtkTextBufferDeleteInteractive(object, start.iter, end.iter, 
    default.editable)

Arguments

object
a GtkTextBuffer

start.iter
start of range to delete

end.iter
end of range

default.editable
whether the buffer is editable by default

Value

[logical] whether some text was actually deleted

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]