gtkTextBufferDeleteMark {RGtk2}	R Documentation

gtkTextBufferDeleteMark

Description

Deletes mark, so that it's no longer located anywhere in the buffer. Removes the reference the buffer holds to the mark, it will be freed. Even if the mark isn't freed, most operations on mark become invalid, until it gets added to a buffer again with gtkTextBufferAddMark. Use gtkTextMarkGetDeleted to find out if a mark has been removed from its buffer. The "mark-deleted" signal will be emitted as notification after the mark is deleted.

Usage

gtkTextBufferDeleteMark(object, mark)

Arguments

object
a GtkTextBuffer

mark
a GtkTextMark in buffer

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]