gtkTextBufferDeserialize {RGtk2}	R Documentation

gtkTextBufferDeserialize

Description

This function deserializes rich text in format format and inserts it at iter.

Usage

gtkTextBufferDeserialize(object, content.buffer, format, iter, 
    data, .errwarn = TRUE)

Arguments

object
the GtkTextBuffer format is registered with

content.buffer
the GtkTextBuffer to deserialize into

format
the rich text format to use for deserializing

iter
insertion point for the deserialized text

data
data to deserialize

.errwarn
Whether to issue a warning on error or fail silently

Details

formats to be used must be registered using gtkTextBufferRegisterDeserializeFormat or gtkTextBufferRegisterDeserializeTagset beforehand. Since 2.10

Value

A list containing the following elements:

retval
[logical] TRUE on success, FALSE otherwise.

error
return location for a GError

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]