gtkTextBufferGetEndIter {RGtk2}	R Documentation

gtkTextBufferGetEndIter

Description

Initializes iter with the "end iterator," one past the last valid character in the text buffer. If dereferenced with gtkTextIterGetChar, the end iterator has a character value of 0. The entire buffer lies in the range from the first position in the buffer (call gtkTextBufferGetStartIter to get character position 0) to the end iterator.

Usage

gtkTextBufferGetEndIter(object)

Arguments

object
a GtkTextBuffer

Value

A list containing the following elements:

iter
iterator to initialize

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]