gtkTextBufferGetSelectionBound {RGtk2}	R Documentation

gtkTextBufferGetSelectionBound

Description

Returns the mark that represents the selection bound. Equivalent to calling gtkTextBufferGetMark to get the mark named "selection_bound", but very slightly more efficient, and involves less typing.

Usage

gtkTextBufferGetSelectionBound(object)

Arguments

object
a GtkTextBuffer

Details

The currently-selected text in buffer is the region between the "selection_bound" and "insert" marks. If "selection_bound" and "insert" are in the same place, then there is no current selection. gtkTextBufferGetSelectionBounds is another convenient function for handling the selection, if you just want to know whether there's a selection and what its bounds are.

Value

[GtkTextMark] selection bound mark. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]