gtkTextBufferGetSelectionBounds {RGtk2}	R Documentation

gtkTextBufferGetSelectionBounds

Description

Returns TRUE if some text is selected; places the bounds of the selection in start and end (if the selection has length 0, then start and end are filled in with the same value). start and end will be in ascending order. If start and end are NULL, then they are not filled in, but the return value still indicates whether text is selected.

Usage

gtkTextBufferGetSelectionBounds(object)

Arguments

object
a GtkTextBuffer a GtkTextBuffer

Value

A list containing the following elements:

retval
[logical] whether the selection has nonzero length

start
iterator to initialize with selection start

end
iterator to initialize with selection end

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]