gtkTextBufferInsert {RGtk2}	R Documentation

gtkTextBufferInsert

Description

Inserts len bytes of text at position iter. If len is -1, text must be and will be inserted in its entirety. Emits the "insert-text" signal; insertion actually occurs in the default handler for the signal. iter is invalidated when insertion occurs (because the buffer contents change), but the default signal handler revalidates it to point to the end of the inserted text.

Usage

gtkTextBufferInsert(object, iter, text, len = -1)

Arguments

object
a GtkTextBuffer

iter
a position in the buffer

text
UTF-8 format text to insert

len
length of text in bytes, or -1

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]