gtkTextBufferInsertChildAnchor {RGtk2}	R Documentation

gtkTextBufferInsertChildAnchor

Description

Inserts a child widget anchor into the text buffer at iter. The anchor will be counted as one character in character counts, and when obtaining the buffer contents as a string, will be represented by the Unicode "object replacement character" 0xFFFC. Note that the "slice" variants for obtaining portions of the buffer as a string include this character for child anchors, but the "text" variants do not. E.g. see gtkTextBufferGetSlice and gtkTextBufferGetText. Consider gtkTextBufferCreateChildAnchor as a more convenient alternative to this function. The buffer will add a reference to the anchor, so you can unref it after insertion.

Usage

gtkTextBufferInsertChildAnchor(object, iter, anchor)

Arguments

object
a GtkTextBuffer

iter
location to insert the anchor

anchor
a GtkTextChildAnchor

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]