gtkTextBufferInsertInteractive {RGtk2}	R Documentation

gtkTextBufferInsertInteractive

Description

Like gtkTextBufferInsert, but the insertion will not occur if iter is at a non-editable location in the buffer. Usually you want to prevent insertions at ineditable locations if the insertion results from a user action (is interactive).

Usage

gtkTextBufferInsertInteractive(object, iter, text, default.editable)

Arguments

object
a GtkTextBuffer

iter
a position in buffer

text
some UTF-8 text

default.editable
default editability of buffer

Details

default.editable indicates the editability of text that doesn't have a tag affecting editability applied to it. Typically the result of gtkTextViewGetEditable is appropriate here.

Value

[logical] whether text was actually inserted

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]