gtkTextBufferInsertPixbuf {RGtk2}	R Documentation

gtkTextBufferInsertPixbuf

Description

Inserts an image into the text buffer at iter. The image will be counted as one character in character counts, and when obtaining the buffer contents as a string, will be represented by the Unicode "object replacement character" 0xFFFC. Note that the "slice" variants for obtaining portions of the buffer as a string include this character for pixbufs, but the "text" variants do not. e.g. see gtkTextBufferGetSlice and gtkTextBufferGetText.

Usage

gtkTextBufferInsertPixbuf(object, iter, pixbuf)

Arguments

object
a GtkTextBuffer

iter
location to insert the pixbuf

pixbuf
a GdkPixbuf

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]