gtkTextBufferInsertRange {RGtk2}	R Documentation

gtkTextBufferInsertRange

Description

Copies text, tags, and pixbufs between start and end (the order of start and end doesn't matter) and inserts the copy at iter. Used instead of simply getting/inserting text because it preserves images and tags. If start and end are in a different buffer from buffer, the two buffers must share the same tag table.

Usage

gtkTextBufferInsertRange(object, iter, start, end)

Arguments

object
a GtkTextBuffer

iter
a position in buffer

start
a position in a GtkTextBuffer

end
another position in the same buffer as start

Details

Implemented via emissions of the insert_text and apply_tag signals, so expect those.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]