gtkTextBufferInsertRangeInteractive {RGtk2}	R Documentation

gtkTextBufferInsertRangeInteractive

Description

Same as gtkTextBufferInsertRange, but does nothing if the insertion point isn't editable. The default.editable parameter indicates whether the text is editable at iter if no tags enclosing iter affect editability. Typically the result of gtkTextViewGetEditable is appropriate here.

Usage

gtkTextBufferInsertRangeInteractive(object, iter, start, end, 
    default.editable)

Arguments

object
a GtkTextBuffer

iter
a position in buffer

start
a position in a GtkTextBuffer

end
another position in the same buffer as start

default.editable
default editability of the buffer

Value

[logical] whether an insertion was possible at iter

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]