gtkTextBufferRegisterSerializeTagset {RGtk2}	R Documentation

gtkTextBufferRegisterSerializeTagset

Description

This function registers GTK+'s internal rich text serialization format with the passed buffer. The internal format does not comply to any standard rich text format and only works between GtkTextBuffer instances. It is capable of serializing all of a text buffer's tags and embedded pixbufs.

Usage

gtkTextBufferRegisterSerializeTagset(object, tagset.name = NULL)

Arguments

object
a GtkTextBuffer

tagset.name
an optional tagset name, on NULL. [ allow-none ]

Details

This function is just a wrapper around gtkTextBufferRegisterSerializeFormat. The mime type used for registering is "application/x-gtk-text-buffer-rich-text", or "application/x-gtk-text-buffer-rich-text;format=tagset.name" if a tagset.name was passed.

The tagset.name can be used to restrict the transfer of rich text to buffers with compatible sets of tags, in order to avoid unknown tags from being pasted. It is probably the common case to pass an identifier != NULL here, since the NULL tagset requires the receiving buffer to deal with with pasting of arbitrary tags. Since 2.10

Value

[GdkAtom] the GdkAtom that corresponds to the newly registered format's mime-type.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]