gtkTextBufferRemoveAllTags {RGtk2}	R Documentation

gtkTextBufferRemoveAllTags

Description

Removes all tags in the range between start and end. Be careful with this function; it could remove tags added in code unrelated to the code you're currently writing. That is, using this function is probably a bad idea if you have two or more unrelated code sections that add tags.

Usage

gtkTextBufferRemoveAllTags(object, start, end)

Arguments

object
a GtkTextBuffer

start
one bound of range to be untagged

end
other bound of range to be untagged

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]