gtkTextBufferSerialize {RGtk2}	R Documentation

gtkTextBufferSerialize

Description

This function serializes the portion of text between start and end in the rich text format represented by format.

Usage

gtkTextBufferSerialize(object, content.buffer, format, start, 
    end)

Arguments

object
the GtkTextBuffer format is registered with

content.buffer
the GtkTextBuffer to serialize

format
the rich text format to use for serializing

start
start of block of text to serialize

end
end of block of test to serialize

Details

formats to be used must be registered using gtkTextBufferRegisterSerializeFormat or gtkTextBufferRegisterSerializeTagset beforehand. Since 2.10

Value

A list containing the following elements:

retval
[raw] the serialized data, encoded as format

length
return location for the length of the serialized data

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]