gtkTextBufferSetModified {RGtk2}	R Documentation

gtkTextBufferSetModified

Description

Used to keep track of whether the buffer has been modified since the last time it was saved. Whenever the buffer is saved to disk, call gtk_text_buffer_set_modified (buffer, FALSE). When the buffer is modified, it will automatically toggled on the modified bit again. When the modified bit flips, the buffer emits a "modified-changed" signal.

Usage

gtkTextBufferSetModified(object, setting)

Arguments

object
a GtkTextBuffer

setting
modification flag setting

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]