gtkTextChildAnchorNew {RGtk2}	R Documentation

gtkTextChildAnchorNew

Description

Creates a new GtkTextChildAnchor. Usually you would then insert it into a GtkTextBuffer with gtkTextBufferInsertChildAnchor. To perform the creation and insertion in one step, use the convenience function gtkTextBufferCreateChildAnchor.

Usage

gtkTextChildAnchorNew()

Value

[GtkTextChildAnchor] a new GtkTextChildAnchor

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]