gtkTextIterBackwardToTagToggle {RGtk2}	R Documentation

gtkTextIterBackwardToTagToggle

Description

Moves backward to the next toggle (on or off) of the GtkTextTag tag, or to the next toggle of any tag if tag is NULL. If no matching tag toggles are found, returns FALSE, otherwise TRUE. Does not return toggles located at iter, only toggles before iter. Sets iter to the location of the toggle, or the start of the buffer if no toggle is found.

Usage

gtkTextIterBackwardToTagToggle(object, tag = NULL)

Arguments

object
a GtkTextIter

tag
a GtkTextTag, or NULL. [ allow-none ]

Value

[logical] whether we found a tag toggle before iter

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]