gtkTextIterBackwardVisibleLine {RGtk2}	R Documentation

gtkTextIterBackwardVisibleLine

Description

Moves iter to the start of the previous visible line. Returns TRUE if iter could be moved; i.e. if iter was at character offset 0, this function returns FALSE. Therefore if iter was already on line 0, but not at the start of the line, iter is snapped to the start of the line and the function returns TRUE. (Note that this implies that in a loop calling this function, the line number may not change on every iteration, if your first iteration is on line 0.)

Usage

gtkTextIterBackwardVisibleLine(object)

Arguments

object
an iterator

Details

Since 2.8

Value

[logical] whether iter moved

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]