gtkTextIterBeginsTag {RGtk2}	R Documentation

gtkTextIterBeginsTag

Description

Returns TRUE if tag is toggled on at exactly this point. If tag is NULL, returns TRUE if any tag is toggled on at this point. Note that the gtkTextIterBeginsTag returns TRUE if iter is the start of the tagged range; gtkTextIterHasTag tells you whether an iterator is within a tagged range.

Usage

gtkTextIterBeginsTag(object, tag = NULL)

Arguments

object
an iterator

tag
a GtkTextTag, or NULL. [ allow-none ]

Value

[logical] whether iter is the start of a range tagged with tag

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]