gtkTextIterCanInsert {RGtk2}	R Documentation

gtkTextIterCanInsert

Description

Considering the default editability of the buffer, and tags that affect editability, determines whether text inserted at iter would be editable. If text inserted at iter would be editable then the user should be allowed to insert text at iter. gtkTextBufferInsertInteractive uses this function to decide whether insertions are allowed at a given position.

Usage

gtkTextIterCanInsert(object, default.editability)

Arguments

object
an iterator

default.editability
TRUE if text is editable by default

Value

[logical] whether text inserted at iter would be editable

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]