gtkTextIterCompare {RGtk2}	R Documentation

gtkTextIterCompare

Description

A qsort()-style function that returns negative if lhs is less than rhs, positive if lhs is greater than rhs, and 0 if they're equal. Ordering is in character offset order, i.e. the first character in the buffer is less than the second character in the buffer.

Usage

gtkTextIterCompare(object, rhs)

Arguments

object
a GtkTextIter

rhs
another GtkTextIter

Value

[integer] -1 if lhs is less than rhs, 1 if lhs is greater, 0 if they are equal

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]