gtkTextIterEditable {RGtk2}	R Documentation

gtkTextIterEditable

Description

Returns whether the character at iter is within an editable region of text. Non-editable text is "locked" and can't be changed by the user via GtkTextView. This function is simply a convenience wrapper around gtkTextIterGetAttributes. If no tags applied to this text affect editability, default.setting will be returned.

Usage

gtkTextIterEditable(object, default.setting)

Arguments

object
an iterator

default.setting
TRUE if text is editable by default

Details

You don't want to use this function to decide whether text can be inserted at iter, because for insertion you don't want to know whether the char at iter is inside an editable range, you want to know whether a new character inserted at iter would be inside an editable range. Use gtkTextIterCanInsert to handle this case.

Value

[logical] whether iter is inside an editable range

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]