gtkTextIterEndsLine {RGtk2}	R Documentation

gtkTextIterEndsLine

Description

Returns TRUE if iter points to the start of the paragraph delimiter characters for a line (delimiters will be either a newline, a carriage return, a carriage return followed by a newline, or a Unicode paragraph separator character). Note that an iterator pointing to the \n of a \r\n pair will not be counted as the end of a line, the line ends before the \r. The end iterator is considered to be at the end of a line, even though there are no paragraph delimiter chars there.

Usage

gtkTextIterEndsLine(object)

Arguments

object
an iterator

Value

[logical] whether iter is at the end of a line

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]