gtkTextIterEndsWord {RGtk2}	R Documentation

gtkTextIterEndsWord

Description

Determines whether iter ends a natural-language word. Word breaks are determined by Pango and should be correct for nearly any language (if not, the correct fix would be to the Pango word break algorithms).

Usage

gtkTextIterEndsWord(object)

Arguments

object
a GtkTextIter

Value

[logical] TRUE if iter is at the end of a word

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]