gtkTextIterEqual {RGtk2}	R Documentation

gtkTextIterEqual

Description

Tests whether two iterators are equal, using the fastest possible mechanism. This function is very fast; you can expect it to perform better than e.g. getting the character offset for each iterator and comparing the offsets yourself. Also, it's a bit faster than gtkTextIterCompare.

Usage

gtkTextIterEqual(object, rhs)

Arguments

object
a GtkTextIter

rhs
another GtkTextIter

Value

[logical] TRUE if the iterators point to the same place in the buffer

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]