gtkTextIterForwardChar {RGtk2}	R Documentation

gtkTextIterForwardChar

Description

Moves iter forward by one character offset. Note that images embedded in the buffer occupy 1 character slot, so gtkTextIterForwardChar may actually move onto an image instead of a character, if you have images in your buffer. If iter is the end iterator or one character before it, iter will now point at the end iterator, and gtkTextIterForwardChar returns FALSE for convenience when writing loops.

Usage

gtkTextIterForwardChar(object)

Arguments

object
an iterator

Value

[logical] whether iter moved and is dereferenceable

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]