gtkTextIterForwardChars {RGtk2}	R Documentation

gtkTextIterForwardChars

Description

Moves count characters if possible (if count would move past the start or end of the buffer, moves to the start or end of the buffer). The return value indicates whether the new position of iter is different from its original position, and dereferenceable (the last iterator in the buffer is not dereferenceable). If count is 0, the function does nothing and returns FALSE.

Usage

gtkTextIterForwardChars(object, count)

Arguments

object
an iterator

count
number of characters to move, may be negative

Value

[logical] whether iter moved and is dereferenceable

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]