gtkTextIterForwardLine {RGtk2}	R Documentation

gtkTextIterForwardLine

Description

Moves iter to the start of the next line. If the iter is already on the last line of the buffer, moves the iter to the end of the current line. If after the operation, the iter is at the end of the buffer and not dereferencable, returns FALSE. Otherwise, returns TRUE.

Usage

gtkTextIterForwardLine(object)

Arguments

object
an iterator

Value

[logical] whether iter can be dereferenced

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]