gtkTextIterForwardSearch {RGtk2}	R Documentation

gtkTextIterForwardSearch

Description

Searches forward for str. Any match is returned by setting match.start to the first character of the match and match.end to the first character after the match. The search will not continue past limit. Note that a search is a linear or O(n) operation, so you may wish to use limit to avoid locking up your UI on large buffers.

Usage

gtkTextIterForwardSearch(object, str, flags, limit = NULL)

Arguments

object
start of search

str
a search string

flags
flags affecting how the search is done

limit
bound for the search, or NULL for the end of the buffer. [ allow-none ]

Details

If the GTK_TEXT_SEARCH_VISIBLE_ONLY flag is present, the match may have invisible text interspersed in str. i.e. str will be a possibly-noncontiguous subsequence of the matched range. similarly, if you specify GTK_TEXT_SEARCH_TEXT_ONLY, the match may have pixbufs or child widgets mixed inside the matched range. If these flags are not given, the match must be exact; the special 0xFFFC character in str will match embedded pixbufs or child widgets.

Value

A list containing the following elements:

retval
[logical] whether a match was found

match.start
return location for start of match, or NULL. [ allow-none ]

match.end
return location for end of match, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]