gtkTextIterForwardToLineEnd {RGtk2}	R Documentation

gtkTextIterForwardToLineEnd

Description

Moves the iterator to point to the paragraph delimiter characters, which will be either a newline, a carriage return, a carriage return/newline in sequence, or the Unicode paragraph separator character. If the iterator is already at the paragraph delimiter characters, moves to the paragraph delimiter characters for the next line. If iter is on the last line in the buffer, which does not end in paragraph delimiters, moves to the end iterator (end of the last line), and returns FALSE.

Usage

gtkTextIterForwardToLineEnd(object)

Arguments

object
a GtkTextIter

Value

[logical] TRUE if we moved and the new location is not the end iterator

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]