gtkTextIterForwardVisibleLine {RGtk2}	R Documentation

gtkTextIterForwardVisibleLine

Description

Moves iter to the start of the next visible line. Returns TRUE if there was a next line to move to, and FALSE if iter was simply moved to the end of the buffer and is now not dereferenceable, or if iter was already at the end of the buffer.

Usage

gtkTextIterForwardVisibleLine(object)

Arguments

object
an iterator

Details

Since 2.8

Value

[logical] whether iter can be dereferenced

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]