gtkTextIterForwardVisibleLines {RGtk2}	R Documentation

gtkTextIterForwardVisibleLines

Description

Moves count visible lines forward, if possible (if count would move past the start or end of the buffer, moves to the start or end of the buffer). The return value indicates whether the iterator moved onto a dereferenceable position; if the iterator didn't move, or moved onto the end iterator, then FALSE is returned. If count is 0, the function does nothing and returns FALSE. If count is negative, moves backward by 0 - count lines.

Usage

gtkTextIterForwardVisibleLines(object, count)

Arguments

object
a GtkTextIter

count
number of lines to move forward

Details

Since 2.8

Value

[logical] whether iter moved and is dereferenceable

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]