gtkTextIterGetAttributes {RGtk2}	R Documentation

gtkTextIterGetAttributes

Description

Computes the effect of any tags applied to this spot in the text. The values parameter should be initialized to the default settings you wish to use if no tags are in effect. You'd typically obtain the defaults from gtkTextViewGetDefaultAttributes.

Usage

gtkTextIterGetAttributes(object, values)

Arguments

object
an iterator

values
a GtkTextAttributes to be filled in

Details

gtkTextIterGetAttributes will modify values, applying the effects of any tags present at iter. If any tags affected values, the function returns TRUE.

Value

[logical] TRUE if values was modified

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]