gtkTextIterGetLineIndex {RGtk2}	R Documentation

gtkTextIterGetLineIndex

Description

Returns the byte index of the iterator, counting from the start of a line. Remember that GtkTextBuffer encodes text in UTF-8, and that characters can require a variable number of bytes to represent.

Usage

gtkTextIterGetLineIndex(object)

Arguments

object
an iterator

Value

[integer] distance from start of line, in bytes

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]