gtkTextIterGetLineOffset {RGtk2}	R Documentation

gtkTextIterGetLineOffset

Description

Returns the character offset of the iterator, counting from the start of a line. The first character on the line has offset 0.

Usage

gtkTextIterGetLineOffset(object)

Arguments

object
an iterator

Value

[integer] offset from start of line

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]