gtkTextIterGetMarks {RGtk2}	R Documentation

gtkTextIterGetMarks

Description

Returns a list of all GtkTextMark at this location. Because marks are not iterable (they don't take up any "space" in the buffer, they are just marks in between iterable locations), multiple marks can exist in the same place. The returned list is not in any meaningful order.

Usage

gtkTextIterGetMarks(object)

Arguments

object
an iterator

Value

[list] list of GtkTextMark. [ element-type GtkTextMark][ transfer container ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]