gtkTextIterGetTags {RGtk2}	R Documentation

gtkTextIterGetTags

Description

Returns a list of tags that apply to iter, in ascending order of priority (highest-priority tags are last). The GtkTextTag in the list don't have a reference added, but you have to free the list itself.

Usage

gtkTextIterGetTags(object)

Arguments

object
a GtkTextIter

Value

[list] list of GtkTextTag. [ element-type GtkTextTag][ transfer container ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]