gtkTextIterGetText {RGtk2}	R Documentation

gtkTextIterGetText

Description

Returns text in the given range. If the range contains non-text elements such as images, the character and byte offsets in the returned string will not correspond to character and byte offsets in the buffer. If you want offsets to correspond, see gtkTextIterGetSlice.

Usage

gtkTextIterGetText(object, end)

Arguments

object
iterator at start of a range

end
iterator at end of a range

Value

[character] array of characters from the buffer

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]