gtkTextIterGetToggledTags {RGtk2}	R Documentation

gtkTextIterGetToggledTags

Description

Returns a list of GtkTextTag that are toggled on or off at this point. (If toggled.on is TRUE, the list contains tags that are toggled on.) If a tag is toggled on at iter, then some non-empty range of characters following iter has that tag applied to it. If a tag is toggled off, then some non-empty range following iter does not have the tag applied to it.

Usage

gtkTextIterGetToggledTags(object, toggled.on)

Arguments

object
an iterator

toggled.on
TRUE to get toggled-on tags

Value

[list] tags toggled at this point. [ element-type GtkTextTag][ transfer container ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]