gtkTextIterInsideSentence {RGtk2}	R Documentation

gtkTextIterInsideSentence

Description

Determines whether iter is inside a sentence (as opposed to in between two sentences, e.g. after a period and before the first letter of the next sentence). Sentence boundaries are determined by Pango and should be correct for nearly any language (if not, the correct fix would be to the Pango text boundary algorithms).

Usage

gtkTextIterInsideSentence(object)

Arguments

object
a GtkTextIter

Value

[logical] TRUE if iter is inside a sentence.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]