gtkTextIterSetLineOffset {RGtk2}	R Documentation

gtkTextIterSetLineOffset

Description

Moves iter within a line, to a new character (not byte) offset. The given character offset must be less than or equal to the number of characters in the line; if equal, iter moves to the start of the next line. See gtkTextIterSetLineIndex if you have a byte index rather than a character offset.

Usage

gtkTextIterSetLineOffset(object, char.on.line)

Arguments

object
a GtkTextIter

char.on.line
a character offset relative to the start of iter's current line

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]