gtkTextIterSetVisibleLineIndex {RGtk2}	R Documentation

gtkTextIterSetVisibleLineIndex

Description

Like gtkTextIterSetLineIndex, but the index is in visible bytes, i.e. text with a tag making it invisible is not counted in the index.

Usage

gtkTextIterSetVisibleLineIndex(object, byte.on.line)

Arguments

object
a GtkTextIter

byte.on.line
a byte index

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]