gtkTextMarkGetDeleted {RGtk2}	R Documentation

gtkTextMarkGetDeleted

Description

Returns TRUE if the mark has been removed from its buffer with gtkTextBufferDeleteMark. See gtkTextBufferAddMark for a way to add it to a buffer again.

Usage

gtkTextMarkGetDeleted(object)

Arguments

object
a GtkTextMark

Value

[logical] whether the mark is deleted

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]