gtkTextMarkNew {RGtk2}	R Documentation

gtkTextMarkNew

Description

Creates a text mark. Add it to a buffer using gtkTextBufferAddMark. If name is NULL, the mark is anonymous; otherwise, the mark can be retrieved by name using gtkTextBufferGetMark. If a mark has left gravity, and text is inserted at the mark's current location, the mark will be moved to the left of the newly-inserted text. If the mark has right gravity (left.gravity = FALSE), the mark will end up on the right of newly-inserted text. The standard left-to-right cursor is a mark with right gravity (when you type, the cursor stays on the right side of the text you're typing).

Usage

gtkTextMarkNew(name, left.gravity)

Arguments

name
mark name or NULL. [ allow-none ]

left.gravity
whether the mark should have left gravity

Details

Since 2.12

Value

[GtkTextMark] new GtkTextMark

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]