gtkTextMarkSetVisible {RGtk2}	R Documentation

gtkTextMarkSetVisible

Description

Sets the visibility of mark; the insertion point is normally visible, i.e. you can see it as a vertical bar. Also, the text widget uses a visible mark to indicate where a drop will occur when dragging-and-dropping text. Most other marks are not visible. Marks are not visible by default.

Usage

gtkTextMarkSetVisible(object, setting)

Arguments

object
a GtkTextMark

setting
visibility of mark

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]