gtkTextTagSetPriority {RGtk2}	R Documentation

gtkTextTagSetPriority

Description

Sets the priority of a GtkTextTag. Valid priorities are start at 0 and go to one less than gtkTextTagTableGetSize. Each tag in a table has a unique priority; setting the priority of one tag shifts the priorities of all the other tags in the table to maintain a unique priority for each tag. Higher priority tags "win" if two tags both set the same text attribute. When adding a tag to a tag table, it will be assigned the highest priority in the table by default; so normally the precedence of a set of tags is the order in which they were added to the table, or created with gtkTextBufferCreateTag, which adds the tag to the buffer's table automatically.

Usage

gtkTextTagSetPriority(object, priority)

Arguments

object
a GtkTextTag

priority
the new priority

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]