gtkTextTagTableAdd {RGtk2}	R Documentation

gtkTextTagTableAdd

Description

Add a tag to the table. The tag is assigned the highest priority in the table.

Usage

gtkTextTagTableAdd(object, tag)

Arguments

object
a GtkTextTagTable

tag
a GtkTextTag

Details

tag must not be in a tag table already, and may not have the same name as an already-added tag.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]