gtkTextTagTableForeach {RGtk2}	R Documentation

gtkTextTagTableForeach

Description

Calls func on each tag in table, with user data data. Note that the table may not be modified while iterating over it (you can't add/remove tags).

Usage

gtkTextTagTableForeach(object, func, data = NULL)

Arguments

object
a GtkTextTagTable

func
a function to call on each tag

data
user data

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]