gtkTextViewAddChildInWindow {RGtk2}	R Documentation

gtkTextViewAddChildInWindow

Description

Adds a child at fixed coordinates in one of the text widget's windows. The window must have nonzero size (see gtkTextViewSetBorderWindowSize). Note that the child coordinates are given relative to the GdkWindow in question, and that these coordinates have no sane relationship to scrolling. When placing a child in GTK_TEXT_WINDOW_WIDGET, scrolling is irrelevant, the child floats above all scrollable areas. But when placing a child in one of the scrollable windows (border windows or text window), you'll need to compute the child's correct position in buffer coordinates any time scrolling occurs or buffer changes occur, and then call gtkTextViewMoveChild to update the child's position. Unfortunately there's no good way to detect that scrolling has occurred, using the current API; a possible hack would be to update all child positions when the scroll adjustments change or the text buffer changes. See bug 64518 on bugzilla.gnome.org for status of fixing this issue.

Usage

gtkTextViewAddChildInWindow(object, child, which.window, xpos, 
    ypos)

Arguments

object
a GtkTextView

child
a GtkWidget

which.window
which window the child should appear in

xpos
X position of child in window coordinates

ypos
Y position of child in window coordinates

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]