gtkTextViewBackwardDisplayLine {RGtk2}	R Documentation

gtkTextViewBackwardDisplayLine

Description

Moves the given iter backward by one display (wrapped) line. A display line is different from a paragraph. Paragraphs are separated by newlines or other paragraph separator characters. Display lines are created by line-wrapping a paragraph. If wrapping is turned off, display lines and paragraphs will be the same. Display lines are divided differently for each view, since they depend on the view's width; paragraphs are the same in all views, since they depend on the contents of the GtkTextBuffer.

Usage

gtkTextViewBackwardDisplayLine(object, iter)

Arguments

object
a GtkTextView

iter
a GtkTextIter

Value

[logical] TRUE if iter was moved and is not on the end iterator

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]