gtkTextViewGetAcceptsTab {RGtk2}	R Documentation

gtkTextViewGetAcceptsTab

Description

Returns whether pressing the Tab key inserts a tab characters. gtkTextViewSetAcceptsTab.

Usage

gtkTextViewGetAcceptsTab(object)

Arguments

object
A GtkTextView

Details

Since 2.4

Value

[logical] TRUE if pressing the Tab key inserts a tab character, FALSE if pressing the Tab key moves the keyboard focus.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]