gtkTextViewGetIterAtLocation {RGtk2}	R Documentation

gtkTextViewGetIterAtLocation

Description

Retrieves the iterator at buffer coordinates x and y. Buffer coordinates are coordinates for the entire buffer, not just the currently-displayed portion. If you have coordinates from an event, you have to convert those to buffer coordinates with gtkTextViewWindowToBufferCoords.

Usage

gtkTextViewGetIterAtLocation(object, x, y)

Arguments

object
a GtkTextView

x
x position, in buffer coordinates

y
y position, in buffer coordinates

Value

A list containing the following elements:

iter
a GtkTextIter. [ out ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]