gtkTextViewGetIterAtPosition {RGtk2}	R Documentation

gtkTextViewGetIterAtPosition

Description

Retrieves the iterator pointing to the character at buffer coordinates x and y. Buffer coordinates are coordinates for the entire buffer, not just the currently-displayed portion. If you have coordinates from an event, you have to convert those to buffer coordinates with gtkTextViewWindowToBufferCoords.

Usage

gtkTextViewGetIterAtPosition(object, x, y)

Arguments

object
a GtkTextView

x
x position, in buffer coordinates

y
y position, in buffer coordinates

Details

Note that this is different from gtkTextViewGetIterAtLocation, which returns cursor locations, i.e. positions between characters. Since 2.6

Value

A list containing the following elements:

iter
a GtkTextIter. [ out ]

trailing
if non-NULL, location to store an integer indicating where in the grapheme the user clicked. It will either be zero, or the number of characters in the grapheme. 0 represents the trailing edge of the grapheme. [ out ][ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]