gtkTextViewGetIterLocation {RGtk2}	R Documentation

gtkTextViewGetIterLocation

Description

Gets a rectangle which roughly contains the character at iter. The rectangle position is in buffer coordinates; use gtkTextViewBufferToWindowCoords to convert these coordinates to coordinates for one of the windows in the text view.

Usage

gtkTextViewGetIterLocation(object, iter)

Arguments

object
a GtkTextView

iter
a GtkTextIter

Value

A list containing the following elements:

location
bounds of the character at iter. [ out ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]