gtkTextViewGetLineAtY {RGtk2}	R Documentation

gtkTextViewGetLineAtY

Description

Gets the GtkTextIter at the start of the line containing the coordinate y. y is in buffer coordinates, convert from window coordinates with gtkTextViewWindowToBufferCoords. If non-NULL, line.top will be filled with the coordinate of the top edge of the line.

Usage

gtkTextViewGetLineAtY(object, y)

Arguments

object
a GtkTextView

y
a y coordinate

Value

A list containing the following elements:

target.iter
a GtkTextIter. [ out ]

line.top
return location for top coordinate of the line. [ out ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]