gtkTextViewGetWindow {RGtk2}	R Documentation

gtkTextViewGetWindow

Description

Retrieves the GdkWindow corresponding to an area of the text view; possible windows include the overall widget window, child windows on the left, right, top, bottom, and the window that displays the text buffer. Windows are NULL and nonexistent if their width or height is 0, and are nonexistent before the widget has been realized.

Usage

gtkTextViewGetWindow(object, win)

Arguments

object
a GtkTextView

win
window to get

Value

[GdkWindow] a GdkWindow, or NULL. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]