gtkTextViewGetWindowType {RGtk2}	R Documentation

gtkTextViewGetWindowType

Description

Usually used to find out which window an event corresponds to. If you connect to an event signal on text.view, this function should be called on event->window to see which window it was.

Usage

gtkTextViewGetWindowType(object, window)

Arguments

object
a GtkTextView

window
a window type

Value

[GtkTextWindowType] the window type.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]