gtkTextViewMoveVisually {RGtk2}	R Documentation

gtkTextViewMoveVisually

Description

Move the iterator a given number of characters visually, treating it as the strong cursor position. If count is positive, then the new strong cursor position will be count positions to the right of the old cursor position. If count is negative then the new strong cursor position will be count positions to the left of the old cursor position.

Usage

gtkTextViewMoveVisually(object, iter, count)

Arguments

object
a GtkTextView

iter
a GtkTextIter

count
number of characters to move (negative moves left, positive moves right)

Details

In the presence of bi-directional text, the correspondence between logical and visual order will depend on the direction of the current run, and there may be jumps when the cursor is moved off of the end of a run.

Value

[logical] TRUE if iter moved and is not on the end iterator

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]