gtkTextViewNew {RGtk2}	R Documentation

gtkTextViewNew

Description

Creates a new GtkTextView. If you don't call gtkTextViewSetBuffer before using the text view, an empty default buffer will be created for you. Get the buffer with gtkTextViewGetBuffer. If you want to specify your own buffer, consider gtkTextViewNewWithBuffer.

Usage

gtkTextViewNew(show = TRUE)

Value

[GtkWidget] a new GtkTextView

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]