gtkTextViewNewWithBuffer {RGtk2}	R Documentation

gtkTextViewNewWithBuffer

Description

Creates a new GtkTextView widget displaying the buffer buffer. One buffer can be shared among many widgets. buffer may be NULL to create a default buffer, in which case this function is equivalent to gtkTextViewNew. The text view adds its own reference count to the buffer; it does not take over an existing reference.

Usage

gtkTextViewNewWithBuffer(buffer = NULL, show = TRUE)

Arguments

buffer
a GtkTextBuffer

Value

[GtkWidget] a new GtkTextView.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]