gtkTextViewScrollToMark {RGtk2}	R Documentation

gtkTextViewScrollToMark

Description

Scrolls text.view so that mark is on the screen in the position indicated by xalign and yalign. An alignment of 0.0 indicates left or top, 1.0 indicates right or bottom, 0.5 means center. If use.align is FALSE, the text scrolls the minimal distance to get the mark onscreen, possibly not scrolling at all. The effective screen for purposes of this function is reduced by a margin of size within.margin.

Usage

gtkTextViewScrollToMark(object, mark, within.margin, use.align = FALSE, 
    xalign = 0.5, yalign = 0.5)

Arguments

object
a GtkTextView

mark
a GtkTextMark

within.margin
margin as a [0.0,0.5) fraction of screen size

use.align
whether to use alignment arguments (if FALSE, just get the mark onscreen)

xalign
horizontal alignment of mark within visible area

yalign
vertical alignment of mark within visible area

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]