gtkTextViewSetAcceptsTab {RGtk2}	R Documentation

gtkTextViewSetAcceptsTab

Description

Sets the behavior of the text widget when the Tab key is pressed. If accepts.tab is TRUE, a tab character is inserted. If accepts.tab is FALSE the keyboard focus is moved to the next widget in the focus chain.

Usage

gtkTextViewSetAcceptsTab(object, accepts.tab)

Arguments

object
A GtkTextView

accepts.tab
TRUE if pressing the Tab key should insert a tab character, FALSE, if pressing the Tab key should move the keyboard focus.

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]