gtkTextViewSetBuffer {RGtk2}	R Documentation

gtkTextViewSetBuffer

Description

Sets buffer as the buffer being displayed by text.view. The previous buffer displayed by the text view is unreferenced, and a reference is added to buffer. If you owned a reference to buffer before passing it to this function, you must remove that reference yourself; GtkTextView will not "adopt" it.

Usage

gtkTextViewSetBuffer(object, buffer)

Arguments

object
a GtkTextView

buffer
a GtkTextBuffer. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]