gtkTimeoutAddFull {RGtk2}	R Documentation

gtkTimeoutAddFull

Description

Registers a function to be called periodically. The function will be called repeatedly after interval milliseconds until it returns FALSE at which point the timeout is destroyed and will not be called again. WARNING: gtk_timeout_add_full has been deprecated since version 2.4 and should not be used in newly-written code. Use gTimeoutAddFull() instead.

Usage

gtkTimeoutAddFull(interval, fun, data = NULL)

Arguments

interval
The time between calls to the function, in milliseconds (1/1000ths of a second.)

data
The data to pass to the function.

Value

[numeric] A unique id for the event source.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]