gtkTipsQuerySetLabels {RGtk2}	R Documentation

gtkTipsQuerySetLabels

Description

Sets the text to display when the query is not in effect, and the text to display when the query is in effect but the widget beneath the pointer has no tooltip. WARNING: gtk_tips_query_set_labels is deprecated and should not be used in newly-written code.

Usage

gtkTipsQuerySetLabels(object, label.inactive, label.no.tip)

Arguments

object
a GtkTipsQuery.

label.inactive
the text to display when the query is not running.

label.no.tip
the text to display when the query is running but the widget beneath the pointer has no tooltip.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]