gtkToggleButtonNewWithMnemonic {RGtk2}	R Documentation

gtkToggleButtonNewWithMnemonic

Description

Creates a new GtkToggleButton containing a label. The label will be created using gtkLabelNewWithMnemonic, so underscores in label indicate the mnemonic for the button.

Usage

gtkToggleButtonNewWithMnemonic(label, show = TRUE)

Arguments

label
the text of the button, with an underscore in front of the mnemonic character

Value

[GtkWidget] a new GtkToggleButton

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]