gtkToggleButtonSetInconsistent {RGtk2}	R Documentation

gtkToggleButtonSetInconsistent

Description

If the user has selected a range of elements (such as some text or spreadsheet cells) that are affected by a toggle button, and the current values in that range are inconsistent, you may want to display the toggle in an "in between" state. This function turns on "in between" display. Normally you would turn off the inconsistent state again if the user toggles the toggle button. This has to be done manually, gtkToggleButtonSetInconsistent only affects visual appearance, it doesn't affect the semantics of the button.

Usage

gtkToggleButtonSetInconsistent(object, setting)

Arguments

object
a GtkToggleButton

setting
TRUE if state is inconsistent

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]