gtkToggleButtonSetMode {RGtk2}	R Documentation

gtkToggleButtonSetMode

Description

Sets whether the button is displayed as a separate indicator and label. You can call this function on a checkbutton or a radiobutton with draw.indicator = FALSE to make the button look like a normal button

Usage

gtkToggleButtonSetMode(object, draw.indicator)

Arguments

object
a GtkToggleButton

draw.indicator
if TRUE, draw the button as a separate indicator and label; if FALSE, draw the button like a normal button

Details

This function only affects instances of classes like GtkCheckButton and GtkRadioButton that derive from GtkToggleButton, not instances of GtkToggleButton itself.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]