gtkToolButtonGetIconName {RGtk2}	R Documentation

gtkToolButtonGetIconName

Description

Returns the name of the themed icon for the tool button, see gtkToolButtonSetIconName.

Usage

gtkToolButtonGetIconName(object)

Arguments

object
a GtkToolButton

Details

Since 2.8

Value

[character] the icon name or NULL if the tool button has no themed icon

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]