gtkToolButtonGetUseUnderline {RGtk2}	R Documentation

gtkToolButtonGetUseUnderline

Description

Returns whether underscores in the label property are used as mnemonics on menu items on the overflow menu. See gtkToolButtonSetUseUnderline.

Usage

gtkToolButtonGetUseUnderline(object)

Arguments

object
a GtkToolButton

Details

Since 2.4

Value

[logical] TRUE if underscores in the label property are used as mnemonics on menu items on the overflow menu.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]