gtkToolButtonNewFromStock {RGtk2}	R Documentation

gtkToolButtonNewFromStock

Description

Creates a new GtkToolButton containing the image and text from a stock item. Some stock ids have preprocessor functions like GTK_STOCK_OK and GTK_STOCK_APPLY.

Usage

gtkToolButtonNewFromStock(stock.id, show = TRUE)

Arguments

stock.id
the name of the stock item

Details

It is an error if stock.id is not a name of a stock item. Since 2.4

Value

[GtkToolItem] A new GtkToolButton

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]