gtkToolButtonSetIconName {RGtk2}	R Documentation

gtkToolButtonSetIconName

Description

Sets the icon for the tool button from a named themed icon. See the docs for GtkIconTheme for more details. The "icon_name" property only has an effect if not overridden by non-NULL "label", "icon_widget" and "stock_id" properties.

Usage

gtkToolButtonSetIconName(object, icon.name)

Arguments

object
a GtkToolButton

icon.name
the name of the themed icon. [ allow-none ]

Details

Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]