gtkToolButtonSetLabelWidget {RGtk2}	R Documentation

gtkToolButtonSetLabelWidget

Description

Sets label.widget as the widget that will be used as the label for button. If label.widget is NULL the "label" property is used as label. If "label" is also NULL, the label in the stock item determined by the "stock_id" property is used as label. If "stock_id" is also NULL, button does not have a label.

Usage

gtkToolButtonSetLabelWidget(object, label.widget = NULL)

Arguments

object
a GtkToolButton

label.widget
the widget used as label, or NULL. [ allow-none ]

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]