gtkToolButtonSetUseUnderline {RGtk2}	R Documentation

gtkToolButtonSetUseUnderline

Description

If set, an underline in the label property indicates that the next character should be used for the mnemonic accelerator key in the overflow menu. For example, if the label property is "_Open" and use.underline is TRUE, the label on the tool button will be "Open" and the item on the overflow menu will have an underlined 'O'.

Usage

gtkToolButtonSetUseUnderline(object, use.underline)

Arguments

object
a GtkToolButton

use.underline
whether the button label has the form "_Open"

Details

Labels shown on tool buttons never have mnemonics on them; this property only affects the menu item on the overflow menu. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]