gtkToolItemGetProxyMenuItem {RGtk2}	R Documentation

gtkToolItemGetProxyMenuItem

Description

If menu.item.id matches the string passed to gtkToolItemSetProxyMenuItem return the corresponding GtkMenuItem.

Usage

gtkToolItemGetProxyMenuItem(object, menu.item.id)

Arguments

object
a GtkToolItem

menu.item.id
a string used to identify the menu item

Details

Custom subclasses of GtkToolItem should use this function to update their menu item when the GtkToolItem changes. That the menu.item.ids must match ensures that a GtkToolItem will not inadvertently change a menu item that they did not create. Since 2.4

Value

[GtkWidget] The GtkMenuItem passed to gtkToolItemSetProxyMenuItem, if the menu.item.ids match.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]