gtkToolItemGetTextSizeGroup {RGtk2}	R Documentation

gtkToolItemGetTextSizeGroup

Description

Returns the size group used for labels in tool.item. Custom subclasses of GtkToolItem should call this function and use the size group for labels.

Usage

gtkToolItemGetTextSizeGroup(object)

Arguments

object
a GtkToolItem

Details

Since 2.20

Value

[GtkSizeGroup] a GtkSizeGroup

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]