gtkToolItemGetVisibleVertical {RGtk2}	R Documentation

gtkToolItemGetVisibleVertical

Description

Returns whether tool.item is visible when the toolbar is docked vertically. See gtkToolItemSetVisibleVertical.

Usage

gtkToolItemGetVisibleVertical(object)

Arguments

object
a GtkToolItem

Details

Since 2.4

Value

[logical] Whether tool.item is visible when the toolbar is docked vertically

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]