gtkToolItemGroupInsert {RGtk2}	R Documentation

gtkToolItemGroupInsert

Description

Inserts item at position in the list of children of group.

Usage

gtkToolItemGroupInsert(object, item, position)

Arguments

object
a GtkToolItemGroup

item
the GtkToolItem to insert into group

position
the position of item in group, starting with 0. The position -1 means end of list.

Details

Since 2.20

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]