gtkToolItemRetrieveProxyMenuItem {RGtk2}	R Documentation

gtkToolItemRetrieveProxyMenuItem

Description

Returns the GtkMenuItem that was last set by gtkToolItemSetProxyMenuItem, ie. the GtkMenuItem that is going to appear in the overflow menu.

Usage

gtkToolItemRetrieveProxyMenuItem(object)

Arguments

object
a GtkToolItem

Details

Since 2.4

Value

[GtkWidget] The GtkMenuItem that is going to appear in the overflow menu for tool.item. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]