gtkToolItemSetHomogeneous {RGtk2}	R Documentation

gtkToolItemSetHomogeneous

Description

Sets whether tool.item is to be allocated the same size as other homogeneous items. The effect is that all homogeneous items will have the same width as the widest of the items.

Usage

gtkToolItemSetHomogeneous(object, homogeneous)

Arguments

object
a GtkToolItem

homogeneous
whether tool.item is the same size as other homogeneous items

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]