gtkToolItemSetIsImportant {RGtk2}	R Documentation

gtkToolItemSetIsImportant

Description

Sets whether tool.item should be considered important. The GtkToolButton class uses this property to determine whether to show or hide its label when the toolbar style is GTK_TOOLBAR_BOTH_HORIZ. The result is that only tool buttons with the "is_important" property set have labels, an effect known as "priority text"

Usage

gtkToolItemSetIsImportant(object, is.important)

Arguments

object
a GtkToolItem

is.important
whether the tool item should be considered important

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]