gtkToolItemSetProxyMenuItem {RGtk2}	R Documentation

gtkToolItemSetProxyMenuItem

Description

Sets the GtkMenuItem used in the toolbar overflow menu. The menu.item.id is used to identify the caller of this function and should also be used with gtkToolItemGetProxyMenuItem.

Usage

gtkToolItemSetProxyMenuItem(object, menu.item.id, menu.item = NULL)

Arguments

object
a GtkToolItem

menu.item.id
a string used to identify menu.item

menu.item
a GtkMenuItem to be used in the overflow menu

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]