gtkToolItemSetTooltip {RGtk2}	R Documentation

gtkToolItemSetTooltip

Description

Sets the GtkTooltips object to be used for tool.item, the text to be displayed as tooltip on the item and the private text to be used. See gtkTooltipsSetTip. WARNING: gtk_tool_item_set_tooltip has been deprecated since version 2.12 and should not be used in newly-written code. Use gtkToolItemSetTooltipText instead.

Usage

gtkToolItemSetTooltip(object, tooltips, tip.text = NULL, tip.private = NULL)

Arguments

object
a GtkToolItem

tooltips
The GtkTooltips object to be used

tip.text
text to be used as tooltip text for tool.item. [ allow-none ]

tip.private
text to be used as private tooltip text. [ allow-none ]

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]