gtkToolItemSetTooltipMarkup {RGtk2}	R Documentation

gtkToolItemSetTooltipMarkup

Description

Sets the markup text to be displayed as tooltip on the item. See gtkWidgetSetTooltipMarkup.

Usage

gtkToolItemSetTooltipMarkup(object, markup)

Arguments

object
a GtkToolItem

markup
markup text to be used as tooltip for tool.item

Details

Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]