gtkToolShellRebuildMenu {RGtk2}	R Documentation

gtkToolShellRebuildMenu

Description

Calling this function signals the tool shell that the overflow menu item for tool items have changed. If there is an overflow menu and if it is visible when this function it called, the menu will be rebuilt.

Usage

gtkToolShellRebuildMenu(object)

Arguments

object
a GtkToolShell

Details

Tool items must not call this function directly, but rely on gtkToolItemRebuildMenu instead. Since 2.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]