gtkToolbarAppendElement {RGtk2}	R Documentation

gtkToolbarAppendElement

Description

Adds a new element to the end of a toolbar. WARNING: gtk_toolbar_append_element has been deprecated since version 2.4 and should not be used in newly-written code. Use gtkToolbarInsert instead.

Usage

gtkToolbarAppendElement(object, type, widget, text, tooltip.text, 
    tooltip.private.text, icon, callback, user.data = NULL)

Arguments

object
a GtkToolbar.

type
a value of type GtkToolbarChildType that determines what widget will be.

widget
a GtkWidget, or NULL. [ allow-none ]

text
the element's label.

tooltip.text
the element's tooltip.

tooltip.private.text
used for context-sensitive help about this toolbar element.

icon
a GtkWidget that provides pictorial representation of the element's function.

callback
the function to be executed when the button is pressed.

user.data
any data you wish to pass to the callback.

Details

If type == GTK_TOOLBAR_CHILD_WIDGET, widget is used as the new element. If type == GTK_TOOLBAR_CHILD_RADIOBUTTON, widget is used to determine the radio group for the new element. In all other cases, widget must be NULL. callback must be a pointer to a function taking a GtkWidget and a gpointer as arguments. Use gCallback() to cast the function to GCallback.

Value

[GtkWidget] the new toolbar element as a GtkWidget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]