gtkToolbarGetItemIndex {RGtk2}	R Documentation

gtkToolbarGetItemIndex

Description

Returns the position of item on the toolbar, starting from 0. It is an error if item is not a child of the toolbar.

Usage

gtkToolbarGetItemIndex(object, item)

Arguments

object
a GtkToolbar

item
a GtkToolItem that is a child of toolbar

Details

Since 2.4

Value

[integer] the position of item on the toolbar.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]