gtkToolbarInsert {RGtk2}	R Documentation

gtkToolbarInsert

Description

Insert a GtkToolItem into the toolbar at position pos. If pos is 0 the item is prepended to the start of the toolbar. If pos is negative, the item is appended to the end of the toolbar.

Usage

gtkToolbarInsert(object, item, pos)

Arguments

object
a GtkToolbar

item
a GtkToolItem

pos
the position of the new item

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]