gtkToolbarInsertStock {RGtk2}	R Documentation

gtkToolbarInsertStock

Description

Inserts a stock item at the specified position of the toolbar. If stock.id is not a known stock item ID, it's inserted verbatim, except that underscores used to mark mnemonics are removed. WARNING: gtk_toolbar_insert_stock has been deprecated since version 2.4 and should not be used in newly-written code. Use gtkToolbarInsert instead.

Usage

gtkToolbarInsertStock(object, stock.id, tooltip.text, tooltip.private.text, 
    callback, user.data, position)

Arguments

object
A GtkToolbar

stock.id
The id of the stock item you want to insert

tooltip.text
The text in the tooltip of the toolbar button

tooltip.private.text
The private text of the tooltip

callback
The callback called when the toolbar button is clicked.

user.data
user data passed to callback

position
The position the button shall be inserted at. -1 means at the end.

Details

callback must be a pointer to a function taking a GtkWidget and a gpointer as arguments. Use gCallback() to cast the function to GCallback.

Value

[GtkWidget] the inserted widget

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]