gtkToolbarPrependItem {RGtk2}	R Documentation

gtkToolbarPrependItem

Description

Adds a new button to the beginning (top or left edges) of the given toolbar. WARNING: gtk_toolbar_prepend_item has been deprecated since version 2.4 and should not be used in newly-written code. Use gtkToolbarInsert instead.

Usage

gtkToolbarPrependItem(object, text, tooltip.text, tooltip.private.text, 
    icon, callback, user.data)

Arguments

object
a GtkToolbar.

text
give your toolbar button a label.

tooltip.text
a string that appears when the user holds the mouse over this item.

tooltip.private.text
use with GtkTipsQuery.

icon
a GtkWidget that should be used as the button's icon.

callback
the function to be executed when the button is pressed.

user.data
a pointer to any data you wish to be passed to the callback.

Details

callback must be a pointer to a function taking a GtkWidget and a gpointer as arguments. Use gCallback() to cast the function to GCallback.

Value

[GtkWidget] the new toolbar item as a GtkWidget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]