gtkToolbarSetDropHighlightItem {RGtk2}	R Documentation

gtkToolbarSetDropHighlightItem

Description

Highlights toolbar to give an idea of what it would look like if item was added to toolbar at the position indicated by index.. If item is NULL, highlighting is turned off. In that case index. is ignored.

Usage

gtkToolbarSetDropHighlightItem(object, tool.item = NULL, index)

Arguments

object
a GtkToolbar

tool.item
a GtkToolItem, or NULL to turn of highlighting. [ allow-none ]

index
a position on toolbar

Details

The tool.item passed to this function must not be part of any widget hierarchy. When an item is set as drop highlight item it can not added to any widget hierarchy or used as highlight item for another toolbar. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]