gtkToolbarSetIconSize {RGtk2}	R Documentation

gtkToolbarSetIconSize

Description

This function sets the size of stock icons in the toolbar. You can call it both before you add the icons and after they've been added. The size you set will override user preferences for the default icon size.

Usage

gtkToolbarSetIconSize(object, icon.size)

Arguments

object
A GtkToolbar

icon.size
The GtkIconSize that stock icons in the toolbar shall have. [ type int]

Details

This should only be used for special-purpose toolbars, normal application toolbars should respect the user preferences for the size of icons.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]