gtkTooltipSetCustom {RGtk2}	R Documentation

gtkTooltipSetCustom

Description

Replaces the widget packed into the tooltip with custom.widget. custom.widget does not get destroyed when the tooltip goes away. By default a box with a GtkImage and GtkLabel is embedded in the tooltip, which can be configured using gtkTooltipSetMarkup and gtkTooltipSetIcon.

Usage

gtkTooltipSetCustom(object, custom.widget)

Arguments

object
a GtkTooltip

custom.widget
a GtkWidget, or NULL to unset the old custom widget. [ allow-none ]

Details

Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]