gtkTooltipsForceWindow {RGtk2}	R Documentation

gtkTooltipsForceWindow

Description

Ensures that the window used for displaying the given tooltips is created. WARNING: gtk_tooltips_force_window has been deprecated since version 2.12 and should not be used in newly-written code.

Usage

gtkTooltipsForceWindow(object)

Arguments

object
a GtkToolTips

Details

Applications should never have to call this function, since GTK+ takes care of this.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]