gtkTooltipsGetInfoFromTipWindow {RGtk2}	R Documentation

gtkTooltipsGetInfoFromTipWindow

Description

Determines the tooltips and the widget they belong to from the window in which they are displayed. WARNING: gtk_tooltips_get_info_from_tip_window has been deprecated since version 2.12 and should not be used in newly-written code.

Usage

gtkTooltipsGetInfoFromTipWindow(object)

Arguments

object
a GtkWindow

Details

This function is mostly intended for use by accessibility technologies; applications should have little use for it. Since 2.4

Value

A list containing the following elements:

retval
[logical] TRUE if tip.window is displaying tooltips, otherwise FALSE.

tooltips
the return location for the tooltips which are displayed in tip.window, or NULL

current.widget
the return location for the widget whose tooltips are displayed, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]