gtkTopWindow {RGtk2}	R Documentation

Create a Gtk Window

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

This creates a top-level Gtk window (as opposed to a dialog or popup window). This is a convenience function for the more general gtkWindow function which is just as easy to use, given its defaults.

Usage

gtkTopWindow(title="My Window", show=TRUE)

Arguments

title
a string to use in the title bar for the window.

show
a logical value indicating whether the window should be displayed (TRUE) or left unshown so that other widgets can be added to it and then shown.

Value

An object of class GtkWindow that has the same inheritance as the low-level inheritance in Gtk.

Note

THIS STUFF IS VERY OLD AND DEPRECATED (compatibility wrappers for RGtk 1)

Author(s)

Duncan Temple Lang <duncan@research.bell-labs.com>

See Also

gtkWindow gtkWindowNew

[Package RGtk2 version 2.20.31 Index]