gtkTreeDragSourceDragDataGet {RGtk2}	R Documentation

gtkTreeDragSourceDragDataGet

Description

Asks the GtkTreeDragSource to fill in selection.data with a representation of the row at path. selection.data->target gives the required type of the data. Should robustly handle a path no longer found in the model!

Usage

gtkTreeDragSourceDragDataGet(object, path)

Arguments

object
a GtkTreeDragSource

path
row that was dragged

Value

A list containing the following elements:

retval
[logical] TRUE if data of the required type was provided

selection.data
a GtkSelectionData to fill with data from the dragged row

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]