gtkTreeGetRowDragData {RGtk2}	R Documentation

gtkTreeGetRowDragData

Description

Obtains a tree.model and path from selection data of target type GTK_TREE_MODEL_ROW. Normally called from a drag_data_received handler. This function can only be used if selection.data originates from the same process that's calling this function, because a pointer to the tree model is being passed around. If you aren't in the same process, then you'll get memory corruption. In the GtkTreeDragDest drag_data_received handler, you can assume that selection data of type GTK_TREE_MODEL_ROW is in from the current process.

Usage

gtkTreeGetRowDragData(object)

Arguments

object
a GtkSelectionData

Value

A list containing the following elements:

retval
[logical] TRUE if selection.data had target type GTK_TREE_MODEL_ROW and is otherwise valid

tree.model
a GtkTreeModel

path
row in tree.model

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]