gtkTreeModelFilterConvertChildIterToIter {RGtk2}	R Documentation

gtkTreeModelFilterConvertChildIterToIter

Description

Sets filter.iter to point to the row in filter that corresponds to the row pointed at by child.iter. If filter.iter was not set, FALSE is returned.

Usage

gtkTreeModelFilterConvertChildIterToIter(object, child.iter)

Arguments

object
A GtkTreeModelFilter.

child.iter
A valid GtkTreeIter pointing to a row on the child model.

Details

Since 2.4

Value

A list containing the following elements:

retval
[logical] TRUE, if filter.iter was set, i.e. if child.iter is a valid iterator pointing to a visible row in child model.

filter.iter
An uninitialized GtkTreeIter.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]