gtkTreeModelFilterConvertChildPathToPath {RGtk2}	R Documentation

gtkTreeModelFilterConvertChildPathToPath

Description

Converts child.path to a path relative to filter. That is, child.path points to a path in the child model. The rerturned path will point to the same row in the filtered model. If child.path isn't a valid path on the child model or points to a row which is not visible in filter, then NULL is returned.

Usage

gtkTreeModelFilterConvertChildPathToPath(object, child.path)

Arguments

object
A GtkTreeModelFilter.

child.path
A GtkTreePath to convert.

Details

Since 2.4

Value

[GtkTreePath] A newly allocated GtkTreePath, or NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]