gtkTreeModelFilterSetModifyFunc {RGtk2}	R Documentation

gtkTreeModelFilterSetModifyFunc

Description

With the n.columns and types parameters, you give a list of column types for this model (which will be exposed to the parent model/view). The func, data and destroy parameters are for specifying the modify function. The modify function will get called for each data access, the goal of the modify function is to return the data which should be displayed at the location specified using the parameters of the modify function.

Usage

gtkTreeModelFilterSetModifyFunc(object, types, func, data = NULL)

Arguments

object
A GtkTreeModelFilter.

types
The GTypes of the columns.

func
A GtkTreeModelFilterModifyFunc

data
User data to pass to the modify function, or NULL. [ allow-none ]

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]