gtkTreeModelGet {RGtk2}	R Documentation

gtkTreeModelGet

Description

Gets the value of one or more cells in the row referenced by iter. The variable argument list should contain integer column numbers, each column number followed by a place to store the value being retrieved. The list is terminated by a -1. For example, to get a value from column 0 with type G_TYPE_STRING, you would write: gtk_tree_model_get (model, iter, 0, &place_string_here, -1), where place_string_here is a character to be filled with the string. If appropriate,

Usage

gtkTreeModelGet(object, iter, ...)

Arguments

object
a GtkTreeModel

iter
a row in tree.model

...
undocumented

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]