gtkTreeModelIterChildren {RGtk2}	R Documentation

gtkTreeModelIterChildren

Description

Sets iter to point to the first child of parent. If parent has no children, FALSE is returned and iter is set to be invalid. parent will remain a valid node after this function has been called.

Usage

gtkTreeModelIterChildren(object, parent = NULL)

Arguments

object
A GtkTreeModel.

parent
The GtkTreeIter, or NULL. [ allow-none ]

Details

If parent is NULL returns the first node, equivalent to gtk_tree_model_get_iter_first (tree_model, iter);

Value

A list containing the following elements:

retval
[logical] TRUE, if child has been set to the first child.

iter
The new GtkTreeIter to be set to the child.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]