gtkTreeModelIterNthChild {RGtk2}	R Documentation

gtkTreeModelIterNthChild

Description

Sets iter to be the child of parent, using the given index. The first index is 0. If n is too big, or parent has no children, iter is set to an invalid iterator and FALSE is returned. parent will remain a valid node after this function has been called. As a special case, if parent is NULL, then the nth root node is set.

Usage

gtkTreeModelIterNthChild(object, parent = NULL, n)

Arguments

object
A GtkTreeModel.

parent
The GtkTreeIter to get the child from, or NULL. [ allow-none ]

n
Then index of the desired child.

Value

A list containing the following elements:

retval
[logical] TRUE, if parent has an nth child.

iter
The GtkTreeIter to set to the nth child.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]