gtkTreeModelIterParent {RGtk2}	R Documentation

gtkTreeModelIterParent

Description

Sets iter to be the parent of child. If child is at the toplevel, and doesn't have a parent, then iter is set to an invalid iterator and FALSE is returned. child will remain a valid node after this function has been called.

Usage

gtkTreeModelIterParent(object, child)

Arguments

object
A GtkTreeModel

child
The GtkTreeIter.

Value

A list containing the following elements:

retval
[logical] TRUE, if iter is set to the parent of child.

iter
The new GtkTreeIter to set to the parent.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]