gtkTreeModelRefNode {RGtk2}	R Documentation

gtkTreeModelRefNode

Description

Lets the tree ref the node. This is an optional method for models to implement. To be more specific, models may ignore this call as it exists primarily for performance reasons.

Usage

gtkTreeModelRefNode(object, iter)

Arguments

object
A GtkTreeModel.

iter
The GtkTreeIter.

Details

This function is primarily meant as a way for views to let caching model know when nodes are being displayed (and hence, whether or not to cache that node.) For example, a file-system based model would not want to keep the entire file-hierarchy in memory, just the sections that are currently being displayed by every current view.

A model should be expected to be able to get an iter independent of its reffed state.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]