gtkTreeModelRowDeleted {RGtk2}	R Documentation

gtkTreeModelRowDeleted

Description

Emits the "row-deleted" signal on tree.model. This should be called by models after a row has been removed. The location pointed to by path should be the location that the row previously was at. It may not be a valid location anymore.

Usage

gtkTreeModelRowDeleted(object, path)

Arguments

object
A GtkTreeModel

path
A GtkTreePath pointing to the previous location of the deleted row.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]