gtkTreeModelRowsReordered {RGtk2}	R Documentation

gtkTreeModelRowsReordered

Description

Emits the "rows-reordered" signal on tree.model. This should be called by models when their rows have been reordered.

Usage

gtkTreeModelRowsReordered(object, path, iter, new.order)

Arguments

object
A GtkTreeModel

path
A GtkTreePath pointing to the tree node whose children have been reordered

iter
A valid GtkTreeIter pointing to the node whose children have been reordered, or NULL if the depth of path is 0.

new.order
a list of integers mapping the current position of each child to its old position before the re-ordering, i.e. new.order[newpos] = oldpos.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]