gtkTreeModelSortClearCache {RGtk2}	R Documentation

gtkTreeModelSortClearCache

Description

This function should almost never be called. It clears the tree.model.sort of any cached iterators that haven't been reffed with gtkTreeModelRefNode. This might be useful if the child model being sorted is static (and doesn't change often) and there has been a lot of unreffed access to nodes. As a side effect of this function, all unreffed iters will be invalid.

Usage

gtkTreeModelSortClearCache(object)

Arguments

object
A GtkTreeModelSort

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]