gtkTreeModelSortConvertChildIterToIter {RGtk2}	R Documentation

gtkTreeModelSortConvertChildIterToIter

Description

Sets sort.iter to point to the row in tree.model.sort that corresponds to the row pointed at by child.iter. If sort.iter was not set, FALSE is returned. Note: a boolean is only returned since 2.14.

Usage

gtkTreeModelSortConvertChildIterToIter(object, child.iter)

Arguments

object
A GtkTreeModelSort

child.iter
A valid GtkTreeIter pointing to a row on the child model

Value

A list containing the following elements:

retval
[logical] TRUE, if sort.iter was set, i.e. if sort.iter is a valid iterator pointer to a visible row in the child model.

sort.iter
An uninitialized GtkTreeIter.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]