gtkTreePathNewFromString {RGtk2}	R Documentation

gtkTreePathNewFromString

Description

Creates a new GtkTreePath initialized to path. path is expected to be a colon separated list of numbers. For example, the string "10:4:0" would create a path of depth 3 pointing to the 11th child of the root node, the 5th child of that 11th child, and the 1st child of that 5th child. If an invalid path string is passed in, NULL is returned.

Usage

gtkTreePathNewFromString(path)

Arguments

path
The string representation of a path.

Value

[GtkTreePath] A newly-created GtkTreePath, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]