gtkTreePathPrev {RGtk2}	R Documentation

gtkTreePathPrev

Description

Moves the path to point to the previous node at the current depth, if it exists.

Usage

gtkTreePathPrev(object)

Arguments

object
A GtkTreePath.

Value

[logical] TRUE if path has a previous node, and the move was made.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]