gtkTreeRowReferenceNew {RGtk2}	R Documentation

gtkTreeRowReferenceNew

Description

Creates a row reference based on path. This reference will keep pointing to the node pointed to by path, so long as it exists. It listens to all signals emitted by model, and updates its path appropriately. If path isn't a valid path in model, then NULL is returned.

Usage

gtkTreeRowReferenceNew(model, path)

Arguments

model
A GtkTreeModel

path
A valid GtkTreePath to monitor

Value

[GtkTreeRowReference] A newly allocated GtkTreeRowReference, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]