gtkTreeRowReferenceReordered {RGtk2}	R Documentation

gtkTreeRowReferenceReordered

Description

Lets a set of row reference created by gtkTreeRowReferenceNewProxy know that the model emitted the "rows_reordered" signal.

Usage

gtkTreeRowReferenceReordered(proxy, path, iter, new.order)

Arguments

proxy
A GObject

path
The parent path of the reordered signal

iter
The iter pointing to the parent of the reordered

new.order
The new order of rows

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]