gtkTreeSelectionGetSelectedRows {RGtk2}	R Documentation

gtkTreeSelectionGetSelectedRows

Description

Creates a list of path of all selected rows. Additionally, if you are planning on modifying the model after calling this function, you may want to convert the returned list into a list of GtkTreeRowReferences. To do this, you can use gtkTreeRowReferenceNew.

Usage

gtkTreeSelectionGetSelectedRows(object)

Arguments

object
A GtkTreeSelection.

Details

To free the return value, use:

# You don't have to free anything in RGtk, silly

Since 2.2

Value

A list containing the following elements:

retval
[list] A list containing a GtkTreePath for each selected row. [ element-type GtkTreePath][ transfer full ]

model
A pointer to set to the GtkTreeModel, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]