gtkTreeSelectionSetSelectFunction {RGtk2}	R Documentation

gtkTreeSelectionSetSelectFunction

Description

Sets the selection function. If set, this function is called before any node is selected or unselected, giving some control over which nodes are selected. The select function should return TRUE if the state of the node may be toggled, and FALSE if the state of the node should be left unchanged.

Usage

gtkTreeSelectionSetSelectFunction(object, func, data = NULL)

Arguments

object
A GtkTreeSelection.

func
The selection function.

data
The selection function's data.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]