gtkTreeSortableGetSortColumnId {RGtk2}	R Documentation

gtkTreeSortableGetSortColumnId

Description

Fills in sort.column.id and order with the current sort column and the order. It returns TRUE unless the sort.column.id is GTK_TREE_SORTABLE_DEFAULT_SORT_COLUMN_ID or GTK_TREE_SORTABLE_UNSORTED_SORT_COLUMN_ID.

Usage

gtkTreeSortableGetSortColumnId(object)

Arguments

object
A GtkTreeSortable

Value

A list containing the following elements:

retval
[logical] TRUE if the sort column is not one of the special sort column ids.

sort.column.id
The sort column id to be filled in

order
The GtkSortType to be filled in

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]