gtkTreeSortableSetDefaultSortFunc {RGtk2}	R Documentation

gtkTreeSortableSetDefaultSortFunc

Description

Sets the default comparison function used when sorting to be sort.func. If the current sort column id of sortable is GTK_TREE_SORTABLE_DEFAULT_SORT_COLUMN_ID, then the model will sort using this function.

Usage

gtkTreeSortableSetDefaultSortFunc(object, sort.func, user.data = NULL)

Arguments

object
A GtkTreeSortable

sort.func
The comparison function

user.data
User data to pass to sort.func, or NULL. [ allow-none ]

Details

If sort.func is NULL, then there will be no default comparison function. This means that once the model has been sorted, it can't go back to the default state. In this case, when the current sort column id of sortable is GTK_TREE_SORTABLE_DEFAULT_SORT_COLUMN_ID, the model will be unsorted.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]