gtkTreeSortableSetSortColumnId {RGtk2}	R Documentation

gtkTreeSortableSetSortColumnId

Description

Sets the current sort column to be sort.column.id. The sortable will resort itself to reflect this change, after emitting a gtkTreeSortableSortColumnChanged signal. sortable may either be a regular column id, or one of the following special values:

GTK_TREE_SORTABLE_DEFAULT_SORT_COLUMN_ID

undocumented

GTK_TREE_SORTABLE_UNSORTED_SORT_COLUMN_ID

undocumented

Usage

gtkTreeSortableSetSortColumnId(object, sort.column.id, order)

Arguments

object
A GtkTreeSortable

sort.column.id
the sort column id to set

order
The sort order of the column

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]