gtkTreeStoreInsertBefore {RGtk2}	R Documentation

gtkTreeStoreInsertBefore

Description

Inserts a new row before sibling. If sibling is NULL, then the row will be appended to parent 's children. If parent and sibling are NULL, then the row will be appended to the toplevel. If both sibling and parent are set, then parent must be the parent of sibling. When sibling is set, parent is optional.

Usage

gtkTreeStoreInsertBefore(object, parent, sibling)

Arguments

object
A GtkTreeStore

parent
A valid GtkTreeIter, or NULL. [ allow-none ]

sibling
A valid GtkTreeIter, or NULL. [ allow-none ]

Details

iter will be changed to point to this new row. The row will be empty after this function is called. To fill in values, you need to call gtkTreeStoreSet or gtkTreeStoreSetValue.

Value

A list containing the following elements:

iter
An unset GtkTreeIter to set to the new row

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]