gtkTreeStoreMoveBefore {RGtk2}	R Documentation

gtkTreeStoreMoveBefore

Description

Moves iter in tree.store to the position before position. iter and position should be in the same level. Note that this function only works with unsorted stores. If position is NULL, iter will be moved to the end of the level.

Usage

gtkTreeStoreMoveBefore(object, iter, position = NULL)

Arguments

object
A GtkTreeStore.

iter
A GtkTreeIter.

position
A GtkTreeIter or NULL. [ allow-none ]

Details

Since 2.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]