gtkTreeStoreSetColumnTypes {RGtk2}	R Documentation

gtkTreeStoreSetColumnTypes

Description

This function is meant primarily for GObjects that inherit from GtkTreeStore, and should only be used when constructing a new GtkTreeStore. It will not function after a row has been added, or a method on the GtkTreeModel interface is called.

Usage

gtkTreeStoreSetColumnTypes(object, types)

Arguments

object
A GtkTreeStore

types
An list of GType types, one for each column. [ array length=n_columns]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]