gtkTreeStoreSwap {RGtk2}	R Documentation

gtkTreeStoreSwap

Description

Swaps a and b in the same level of tree.store. Note that this function only works with unsorted stores.

Usage

gtkTreeStoreSwap(object, a, b)

Arguments

object
A GtkTreeStore.

a
A GtkTreeIter.

b
Another GtkTreeIter.

Details

Since 2.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]