gtkTreeViewColumnAddAttribute {RGtk2}	R Documentation

gtkTreeViewColumnAddAttribute

Description

Adds an attribute mapping to the list in tree.column. The column is the column of the model to get a value from, and the attribute is the parameter on cell.renderer to be set from the value. So for example if column 2 of the model contains strings, you could have the "text" attribute of a GtkCellRendererText get its values from column 2.

Usage

gtkTreeViewColumnAddAttribute(object, cell.renderer, attribute, 
    column)

Arguments

object
A GtkTreeViewColumn.

cell.renderer
the GtkCellRenderer to set attributes on

attribute
An attribute on the renderer

column
The column position on the model to get the attribute from.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]