gtkTreeViewColumnCellGetPosition {RGtk2}	R Documentation

gtkTreeViewColumnCellGetPosition

Description

Obtains the horizontal position and size of a cell in a column. If the cell is not found in the column, start.pos and width are not changed and FALSE is returned.

Usage

gtkTreeViewColumnCellGetPosition(object, cell.renderer)

Arguments

object
a GtkTreeViewColumn

cell.renderer
a GtkCellRenderer

Value

A list containing the following elements:

retval
[logical] TRUE if cell belongs to tree.column.

start.pos
return location for the horizontal position of cell within tree.column, may be NULL

width
return location for the width of cell, may be NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]