gtkTreeViewColumnCellGetSize {RGtk2}	R Documentation

gtkTreeViewColumnCellGetSize

Description

Obtains the width and height needed to render the column. This is used primarily by the GtkTreeView.

Usage

gtkTreeViewColumnCellGetSize(object)

Arguments

object
A GtkTreeViewColumn.

Value

A list containing the following elements:

cell.area
The area a cell in the column will be allocated, or NULL. [ allow-none ]

x.offset
location to return x offset of a cell relative to cell.area, or NULL. [ allow-none ]

y.offset
location to return y offset of a cell relative to cell.area, or NULL. [ allow-none ]

width
location to return width needed to render a cell, or NULL. [ allow-none ]

height
location to return height needed to render a cell, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]