gtkTreeViewColumnGetFixedWidth {RGtk2}	R Documentation

gtkTreeViewColumnGetFixedWidth

Description

Gets the fixed width of the column. This value is only meaning may not be the actual width of the column on the screen, just what is requested.

Usage

gtkTreeViewColumnGetFixedWidth(object)

Arguments

object
a GtkTreeViewColumn

Value

[integer] the fixed width of the column

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]