gtkTreeViewColumnPackStart {RGtk2}	R Documentation

gtkTreeViewColumnPackStart

Description

Packs the cell into the beginning of the column. If expand is FALSE, then the cell is allocated no more space than it needs. Any unused space is divided evenly between cells for which expand is TRUE.

Usage

gtkTreeViewColumnPackStart(object, cell, expand = TRUE)

Arguments

object
A GtkTreeViewColumn.

cell
The GtkCellRenderer.

expand
TRUE if cell is to be given extra space allocated to tree.column.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]