gtkTreeViewColumnSetAlignment {RGtk2}	R Documentation

gtkTreeViewColumnSetAlignment

Description

Sets the alignment of the title or custom widget inside the column header. The alignment determines its location inside the button – 0.0 for left, 0.5 for center, 1.0 for right.

Usage

gtkTreeViewColumnSetAlignment(object, xalign)

Arguments

object
A GtkTreeViewColumn.

xalign
The alignment, which is between [0.0 and 1.0] inclusive.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]