gtkTreeViewColumnSetCellDataFunc {RGtk2}	R Documentation

gtkTreeViewColumnSetCellDataFunc

Description

Sets the GtkTreeViewColumnFunc to use for the column. This function is used instead of the standard attributes mapping for setting the column value, and should set the value of tree.column's cell renderer as appropriate. func may be NULL to remove an older one.

Usage

gtkTreeViewColumnSetCellDataFunc(object, cell.renderer, func, 
    func.data = NULL)

Arguments

object
A GtkTreeViewColumn

cell.renderer
A GtkCellRenderer

func
The GtkTreeViewColumnFunc to use.

func.data
The user data for func.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]