gtkTreeViewColumnSetExpand {RGtk2}	R Documentation

gtkTreeViewColumnSetExpand

Description

Sets the column to take available extra space. This space is shared equally amongst all columns that have the expand set to TRUE. If no column has this option set, then the last column gets all extra space. By default, every column is created with this FALSE.

Usage

gtkTreeViewColumnSetExpand(object, expand)

Arguments

object
A GtkTreeViewColumn

expand
TRUE if the column should take available extra space, FALSE if not

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]