gtkTreeViewColumnSetFixedWidth {RGtk2}	R Documentation

gtkTreeViewColumnSetFixedWidth

Description

Sets the size of the column in pixels. This is meaningful only if the sizing type is GTK_TREE_VIEW_COLUMN_FIXED. The size of the column is clamped to the min/max width for the column. Please note that the min/max width of the column doesn't actually affect the "fixed_width" property of the widget, just the actual size when displayed.

Usage

gtkTreeViewColumnSetFixedWidth(object, fixed.width)

Arguments

object
A GtkTreeViewColumn.

fixed.width
The size to set tree.column to. Must be greater than 0.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]