gtkTreeViewColumnSetMaxWidth {RGtk2}	R Documentation

gtkTreeViewColumnSetMaxWidth

Description

Sets the maximum width of the tree.column. If max.width is -1, then the maximum width is unset. Note, the column can actually be wider than max width if it's the last column in a view. In this case, the column expands to fill any extra space.

Usage

gtkTreeViewColumnSetMaxWidth(object, max.width)

Arguments

object
A GtkTreeViewColumn.

max.width
The maximum width of the column in pixels, or -1.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]