gtkTreeViewColumnSetResizable {RGtk2}	R Documentation

gtkTreeViewColumnSetResizable

Description

If resizable is TRUE, then the user can explicitly resize the column by grabbing the outer edge of the column button. If resizable is TRUE and sizing mode of the column is GTK_TREE_VIEW_COLUMN_AUTOSIZE, then the sizing mode is changed to GTK_TREE_VIEW_COLUMN_GROW_ONLY.

Usage

gtkTreeViewColumnSetResizable(object, resizable)

Arguments

object
A GtkTreeViewColumn

resizable
TRUE, if the column can be resized

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]