gtkTreeViewColumnSetSortIndicator {RGtk2}	R Documentation

gtkTreeViewColumnSetSortIndicator

Description

Call this function with a setting of TRUE to display an arrow in the header button indicating the column is sorted. Call gtkTreeViewColumnSetSortOrder to change the direction of the arrow.

Usage

gtkTreeViewColumnSetSortIndicator(object, setting)

Arguments

object
a GtkTreeViewColumn

setting
TRUE to display an indicator that the column is sorted

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]