gtkTreeViewColumnSetSortOrder {RGtk2}	R Documentation

gtkTreeViewColumnSetSortOrder

Description

Changes the appearance of the sort indicator.

Usage

gtkTreeViewColumnSetSortOrder(object, order)

Arguments

object
a GtkTreeViewColumn

order
sort order that the sort indicator should indicate

Details

This does not actually sort the model. Use gtkTreeViewColumnSetSortColumnId if you want automatic sorting support. This function is primarily for custom sorting behavior, and should be used in conjunction with gtkTreeSortableSetSortColumn() to do that. For custom models, the mechanism will vary.

The sort indicator changes direction to indicate normal sort or reverse sort. Note that you must have the sort indicator enabled to see anything when calling this function; see gtkTreeViewColumnSetSortIndicator.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]