gtkTreeViewColumnSetWidget {RGtk2}	R Documentation

gtkTreeViewColumnSetWidget

Description

Sets the widget in the header to be widget. If widget is NULL, then the header button is set with a GtkLabel set to the title of tree.column.

Usage

gtkTreeViewColumnSetWidget(object, widget = NULL)

Arguments

object
A GtkTreeViewColumn.

widget
A child GtkWidget, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]