gtkTreeViewGetCellArea {RGtk2}	R Documentation

gtkTreeViewGetCellArea

Description

Fills the bounding rectangle in bin_window coordinates for the cell at the row specified by path and the column specified by column. If path is NULL, or points to a path not currently displayed, the y and height fields of the rectangle will be filled with 0. If column is NULL, the x and width fields will be filled with 0. The sum of all cell rects does not cover the entire tree; there are extra pixels in between rows, for example. The returned rectangle is equivalent to the cell.area passed to gtkCellRendererRender. This function is only valid if tree.view is realized.

Usage

gtkTreeViewGetCellArea(object, path, column)

Arguments

object
a GtkTreeView

path
a GtkTreePath for the row, or NULL to get only horizontal coordinates. [ allow-none ]

column
a GtkTreeViewColumn for the column, or NULL to get only vertical coordinates. [ allow-none ]

Value

A list containing the following elements:

rect
rectangle to fill with cell rect

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]