gtkTreeViewGetDestRowAtPos {RGtk2}	R Documentation

gtkTreeViewGetDestRowAtPos

Description

Determines the destination row for a given position. drag.x and drag.y are expected to be in widget coordinates. This function is only meaningful if tree.view is realized. Therefore this function will always return FALSE if tree.view is not realized or does not have a model.

Usage

gtkTreeViewGetDestRowAtPos(object, drag.x, drag.y)

Arguments

object
a GtkTreeView

drag.x
the position to determine the destination row for

drag.y
the position to determine the destination row for

Value

A list containing the following elements:

retval
[logical] whether there is a row at the given position, TRUE if this is indeed the case.

path
Return location for the path of the highlighted row, or NULL. [ allow-none ]

pos
Return location for the drop position, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]