gtkTreeViewInsertColumnWithDataFunc {RGtk2}	R Documentation

gtkTreeViewInsertColumnWithDataFunc

Description

Convenience function that inserts a new column into the GtkTreeView with the given cell renderer and a GtkCellDataFunc to set cell renderer attributes (normally using data from the model). See also gtkTreeViewColumnSetCellDataFunc, gtkTreeViewColumnPackStart. If tree.view has "fixed_height" mode enabled, then the new column will have its "sizing" property set to be GTK_TREE_VIEW_COLUMN_FIXED.

Usage

gtkTreeViewInsertColumnWithDataFunc(object, position, title, 
    cell, func, data = NULL)

Arguments

object
a GtkTreeView

position
Position to insert, -1 for append

title
column title

cell
cell renderer for column

func
function to set attributes of cell renderer

data
data for func

Value

[integer] number of columns in the tree view post-insert

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]