gtkTreeViewScrollToCell {RGtk2}	R Documentation

gtkTreeViewScrollToCell

Description

Moves the alignments of tree.view to the position specified by column and path. If column is NULL, then no horizontal scrolling occurs. Likewise, if path is NULL no vertical scrolling occurs. At a minimum, one of column or path need to be non-NULL. row.align determines where the row is placed, and col.align determines where column is placed. Both are expected to be between 0.0 and 1.0. 0.0 means left/top alignment, 1.0 means right/bottom alignment, 0.5 means center.

Usage

gtkTreeViewScrollToCell(object, path, column = NULL, use.align = FALSE, 
    row.align = 0, col.align = 0)

Arguments

object
A GtkTreeView.

path
The path of the row to move to, or NULL. [ allow-none ]

column
The GtkTreeViewColumn to move horizontally to, or NULL. [ allow-none ]

use.align
whether to use alignment arguments, or FALSE.

row.align
The vertical alignment of the row specified by path.

col.align
The horizontal alignment of the column specified by column.

Details

If use.align is FALSE, then the alignment arguments are ignored, and the tree does the minimum amount of work to scroll the cell onto the screen. This means that the cell will be scrolled to the edge closest to its current position. If the cell is currently visible on the screen, nothing is done.

This function only works if the model is set, and path is a valid row on the model. If the model changes before the tree.view is realized, the centered path will be modified to reflect this change.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]