gtkTreeViewSetColumnDragFunction {RGtk2}	R Documentation

gtkTreeViewSetColumnDragFunction

Description

Sets a user function for determining where a column may be dropped when dragged. This function is called on every column pair in turn at the beginning of a column drag to determine where a drop can take place. The arguments passed to func are: the tree.view, the GtkTreeViewColumn being dragged, the two GtkTreeViewColumn s determining the drop spot, and user.data. If either of the GtkTreeViewColumn arguments for the drop spot are NULL, then they indicate an edge. If func is set to be NULL, then tree.view reverts to the default behavior of allowing all columns to be dropped everywhere.

Usage

gtkTreeViewSetColumnDragFunction(object, func, user.data = NULL)

Arguments

object
A GtkTreeView.

func
A function to determine which columns are reorderable, or NULL. [ allow-none ]

user.data
User data to be passed to func, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]