gtkTreeViewSetCursor {RGtk2}	R Documentation

gtkTreeViewSetCursor

Description

Sets the current keyboard focus to be at path, and selects it. This is useful when you want to focus the user's attention on a particular row. If focus.column is not NULL, then focus is given to the column specified by it. Additionally, if focus.column is specified, and start.editing is TRUE, then editing should be started in the specified cell. This function is often followed by gtk.widget.grab.focus (tree.view) in order to give keyboard focus to the widget. Please note that editing can only happen when the widget is realized.

Usage

gtkTreeViewSetCursor(object, path, focus.column = NULL, start.editing = FALSE)

Arguments

object
A GtkTreeView

path
A GtkTreePath

focus.column
A GtkTreeViewColumn, or NULL. [ allow-none ]

start.editing
TRUE if the specified cell should start being edited.

Details

If path is invalid for model, the current cursor (if any) will be unset and the function will return without failing.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]