gtkTreeViewSetDestroyCountFunc {RGtk2}	R Documentation

gtkTreeViewSetDestroyCountFunc

Description

This function should almost never be used. It is meant for private use by ATK for determining the number of visible children that are removed when the user collapses a row, or a row is deleted.

Usage

gtkTreeViewSetDestroyCountFunc(object, func, data = NULL)

Arguments

object
A GtkTreeView

func
Function to be called when a view row is destroyed, or NULL. [ allow-none ]

data
User data to be passed to func, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]