gtkTreeViewSetEnableSearch {RGtk2}	R Documentation

gtkTreeViewSetEnableSearch

Description

If enable.search is set, then the user can type in text to search through the tree interactively (this is sometimes called "typeahead find").

Usage

gtkTreeViewSetEnableSearch(object, enable.search)

Arguments

object
A GtkTreeView

enable.search
TRUE, if the user can search interactively

Details

Note that even if this is FALSE, the user can still initiate a search using the "start-interactive-search" key binding.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]