gtkTreeViewSetExpanderColumn {RGtk2}	R Documentation

gtkTreeViewSetExpanderColumn

Description

Sets the column to draw the expander arrow at. It must be in tree.view. If column is NULL, then the expander arrow is always at the first visible column.

Usage

gtkTreeViewSetExpanderColumn(object, column)

Arguments

object
A GtkTreeView

column
NULL, or the column to draw the expander arrow at.

Details

If you do not want expander arrow to appear in your tree, set the expander column to a hidden column.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]