gtkTreeViewSetHoverSelection {RGtk2}	R Documentation

gtkTreeViewSetHoverSelection

Description

Enables of disables the hover selection mode of tree.view. Hover selection makes the selected row follow the pointer. Currently, this works only for the selection modes GTK_SELECTION_SINGLE and GTK_SELECTION_BROWSE.

Usage

gtkTreeViewSetHoverSelection(object, hover)

Arguments

object
a GtkTreeView

hover
TRUE to enable hover selection mode

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]