gtkTreeViewSetLevelIndentation {RGtk2}	R Documentation

gtkTreeViewSetLevelIndentation

Description

Sets the amount of extra indentation for child levels to use in tree.view in addition to the default indentation. The value should be specified in pixels, a value of 0 disables this feature and in this case only the default indentation will be used. This does not have any visible effects for lists.

Usage

gtkTreeViewSetLevelIndentation(object, indentation)

Arguments

object
a GtkTreeView

indentation
the amount, in pixels, of extra indentation in tree.view.

Details

Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]