gtkTreeViewSetSearchColumn {RGtk2}	R Documentation

gtkTreeViewSetSearchColumn

Description

Sets column as the column where the interactive search code should search in for the current model.

Usage

gtkTreeViewSetSearchColumn(object, column)

Arguments

object
A GtkTreeView

column
the column of the model to search in, or -1 to disable searching

Details

If the search column is set, users can use the "start-interactive-search" key binding to bring up search popup. The enable-search property controls whether simply typing text will also start an interactive search.

Note that column refers to a column of the current model. The search column is reset to -1 when the model is changed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]