gtkTreeViewSetSearchEntry {RGtk2}	R Documentation

gtkTreeViewSetSearchEntry

Description

Sets the entry which the interactive search code will use for this tree.view. This is useful when you want to provide a search entry in our interface at all time at a fixed position. Passing NULL for entry will make the interactive search code use the built-in popup entry again.

Usage

gtkTreeViewSetSearchEntry(object, entry = NULL)

Arguments

object
A GtkTreeView

entry
the entry the interactive search code of tree.view should use or NULL. [ allow-none ]

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]