gtkTreeViewSetShowExpanders {RGtk2}	R Documentation

gtkTreeViewSetShowExpanders

Description

Sets whether to draw and enable expanders and indent child rows in tree.view. When disabled there will be no expanders visible in trees and there will be no way to expand and collapse rows by default. Also note that hiding the expanders will disable the default indentation. You can set a custom indentation in this case using gtkTreeViewSetLevelIndentation. This does not have any visible effects for lists.

Usage

gtkTreeViewSetShowExpanders(object, enabled)

Arguments

object
a GtkTreeView

enabled
TRUE to enable expander drawing, FALSE otherwise.

Details

Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]