gtkTreeViewSetTooltipCell {RGtk2}	R Documentation

gtkTreeViewSetTooltipCell

Description

Sets the tip area of tooltip to the area path, column and cell have in common. For example if path is NULL and column is set, the tip area will be set to the full area covered by column. See also gtkTooltipSetTipArea.

Usage

gtkTreeViewSetTooltipCell(object, tooltip, path, column, cell)

Arguments

object
a GtkTreeView

tooltip
a GtkTooltip

path
a GtkTreePath or NULL. [ allow-none ]

column
a GtkTreeViewColumn or NULL. [ allow-none ]

cell
a GtkCellRenderer or NULL. [ allow-none ]

Details

Note that if path is not specified and cell is set and part of a column containing the expander, the tooltip might not show and hide at the correct position. In such cases path must be set to the current node under the mouse cursor for this function to operate correctly.

See also gtkTreeViewSetTooltipColumn for a simpler alternative. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]