gtkTreeViewSetTooltipColumn {RGtk2}	R Documentation

gtkTreeViewSetTooltipColumn

Description

If you only plan to have simple (text-only) tooltips on full rows, you can use this function to have GtkTreeView handle these automatically for you. column should be set to the column in tree.view's model containing the tooltip texts, or -1 to disable this feature.

Usage

gtkTreeViewSetTooltipColumn(object, column)

Arguments

object
a GtkTreeView

column
an integer, which is a valid column number for tree.view's model

Details

When enabled, "has-tooltip" will be set to TRUE and tree.view will connect a "query-tooltip" signal handler.

Note that the signal handler sets the text with gtkTooltipSetMarkup, so &, <, etc have to be escaped in the text. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]