gtkTreeViewWidgetToTreeCoords {RGtk2}	R Documentation

gtkTreeViewWidgetToTreeCoords

Description

Converts bin_window coordinates to coordinates for the tree (the full scrollable area of the tree). WARNING: gtk_tree_view_widget_to_tree_coords has been deprecated since version 2.12 and should not be used in newly-written code. Due to historial reasons the name of this function is incorrect. For converting coordinates relative to the widget to bin_window coordinates, please see gtkTreeViewConvertWidgetToBinWindowCoords.

Usage

gtkTreeViewWidgetToTreeCoords(object, wx, wy)

Arguments

object
a GtkTreeView

wx
X coordinate relative to bin_window

wy
Y coordinate relative to bin_window

Value

A list containing the following elements:

tx
return location for tree X coordinate

ty
return location for tree Y coordinate

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]