gtkUIManagerAddUi {RGtk2}	R Documentation

gtkUIManagerAddUi

Description

Adds a UI element to the current contents of self.

Usage

gtkUIManagerAddUi(object, merge.id, path, name, action = NULL, 
    type, top)

Arguments

object
a GtkUIManager

merge.id
the merge id for the merged UI, see gtkUIManagerNewMergeId

path
a path

name
the name for the added UI element

action
the name of the action to be proxied, or NULL to add a separator. [ allow-none ]

type
the type of UI element to add.

top
if TRUE, the UI element is added before its siblings, otherwise it is added after its siblings.

Details

If type is GTK_UI_MANAGER_AUTO, GTK+ inserts a menuitem, toolitem or separator if such an element can be inserted at the place determined by path. Otherwise type must indicate an element that can be inserted at the place determined by path.

If path points to a menuitem or toolitem, the new element will be inserted before or after this item, depending on top. Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]