gtkUIManagerGetWidget {RGtk2}	R Documentation

gtkUIManagerGetWidget

Description

Looks up a widget by following a path. The path consists of the names specified in the XML description of the UI. separated by '/'. Elements which don't have a name or action attribute in the XML (e.g. <popup>) can be addressed by their XML element name (e.g. "popup"). The root element ("/ui") can be omitted in the path.

Usage

gtkUIManagerGetWidget(object, path)

Arguments

object
a GtkUIManager

path
a path

Details

Note that the widget found by following a path that ends in a <menu> element is the menuitem to which the menu is attached, not the menu itself.

Also note that the widgets constructed by a ui manager are not tied to the lifecycle of the ui manager. If you add the widgets returned by this function to some container or explicitly ref them, they will survive the destruction of the ui manager. Since 2.4

Value

[GtkWidget] the widget found by following the path, or NULL if no widget was found. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]