gtkVButtonBoxGetLayoutDefault {RGtk2}	R Documentation

gtkVButtonBoxGetLayoutDefault

Description

Retrieves the current layout used to arrange buttons in button box widgets. WARNING: gtk_vbutton_box_get_layout_default is deprecated and should not be used in newly-written code.

Usage

gtkVButtonBoxGetLayoutDefault()

Value

[GtkButtonBoxStyle] the current GtkButtonBoxStyle.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]