gtkWidgetAddAccelerator {RGtk2}	R Documentation

gtkWidgetAddAccelerator

Description

Installs an accelerator for this widget in accel.group that causes accel.signal to be emitted if the accelerator is activated. The accel.group needs to be added to the widget's toplevel via gtkWindowAddAccelGroup, and the signal must be of type G_RUN_ACTION. Accelerators added through this function are not user changeable during runtime. If you want to support accelerators that can be changed by the user, use gtkAccelMapAddEntry and gtkWidgetSetAccelPath or gtkMenuItemSetAccelPath instead.

Usage

gtkWidgetAddAccelerator(object, accel.signal, accel.group, accel.key, 
    accel.mods, accel.flags)

Arguments

object
widget to install an accelerator on

accel.signal
widget signal to emit on accelerator activation

accel.group
accel group for this widget, added to its toplevel

accel.key
GDK keyval of the accelerator

accel.mods
modifier key combination of the accelerator

accel.flags
flag accelerators, e.g. GTK_ACCEL_VISIBLE

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]