gtkWidgetAddMnemonicLabel {RGtk2}	R Documentation

gtkWidgetAddMnemonicLabel

Description

Adds a widget to the list of mnemonic labels for this widget. (See gtkWidgetListMnemonicLabels). Note the list of mnemonic labels for the widget is cleared when the widget is destroyed, so the caller must make sure to update its internal state at this point as well, by using a connection to the gtkWidgetDestroy signal or a weak notifier.

Usage

gtkWidgetAddMnemonicLabel(object, label)

Arguments

object
a GtkWidget

label
a GtkWidget that acts as a mnemonic label for widget

Details

Since 2.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]