gtkWidgetCanActivateAccel {RGtk2}	R Documentation

gtkWidgetCanActivateAccel

Description

Determines whether an accelerator that activates the signal identified by signal.id can currently be activated. This is done by emitting the gtkWidgetCanActivateAccel signal on widget; if the signal isn't overridden by a handler or in a derived widget, then the default check is that the widget must be sensitive, and the widget and all its ancestors mapped.

Usage

gtkWidgetCanActivateAccel(object, signal.id)

Arguments

object
a GtkWidget

signal.id
the ID of a signal installed on widget

Details

Since 2.4

Value

[logical] TRUE if the accelerator can be activated.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]