gtkWidgetClassFindStyleProperty {RGtk2}	R Documentation

gtkWidgetClassFindStyleProperty

Description

Finds a style property of a widget class by name.

Usage

gtkWidgetClassFindStyleProperty(klass, property.name)

Arguments

klass
a GtkWidgetClass

property.name
the name of the style property to find

Details

Since 2.2

Value

[GParamSpec] the GParamSpec of the style property or NULL if class has no style property with that name. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]