gtkWidgetDraw {RGtk2}	R Documentation

gtkWidgetDraw

Description

In GTK+ 1.2, this function would immediately render the region area of a widget, by invoking the virtual draw method of a widget. In GTK+ 2.0, the draw method is gone, and instead gtkWidgetDraw simply invalidates the specified region of the widget, then updates the invalid region of the widget immediately. Usually you don't want to update the region immediately for performance reasons, so in general gtkWidgetQueueDrawArea is a better choice if you want to draw a region of a widget. WARNING: gtk_widget_draw is deprecated and should not be used in newly-written code.

Usage

gtkWidgetDraw(object, area)

Arguments

object
a GtkWidget

area
area to draw

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]