gtkWidgetEnsureStyle {RGtk2}	R Documentation

gtkWidgetEnsureStyle

Description

Ensures that widget has a style (widget->style). Not a very useful function; most of the time, if you want the style, the widget is realized, and realized widgets are guaranteed to have a style already.

Usage

gtkWidgetEnsureStyle(object)

Arguments

object
a GtkWidget

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]