gtkWidgetErrorBell {RGtk2}	R Documentation

gtkWidgetErrorBell

Description

Notifies the user about an input-related error on this widget. If the "gtk-error-bell" setting is TRUE, it calls gdkWindowBeep, otherwise it does nothing.

Usage

gtkWidgetErrorBell(object)

Arguments

object
a GtkWidget

Details

Note that the effect of gdkWindowBeep can be configured in many ways, depending on the windowing backend and the desktop environment or window manager that is used. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]