gtkWidgetEvent {RGtk2}	R Documentation

gtkWidgetEvent

Description

Rarely-used function. This function is used to emit the event signals on a widget (those signals should never be emitted without using this function to do so). If you want to synthesize an event though, don't use this function; instead, use gtkMainDoEvent so the event will behave as if it were in the event queue. Don't synthesize expose events; instead, use gdkWindowInvalidateRect to invalidate a region of the window.

Usage

gtkWidgetEvent(object, event)

Arguments

object
a GtkWidget

event
a GdkEvent

Value

[logical] return from the event signal emission (TRUE if the event was handled)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]