gtkWidgetGetAccessible {RGtk2}	R Documentation

gtkWidgetGetAccessible

Description

Returns the accessible object that describes the widget to an assistive technology.

Usage

gtkWidgetGetAccessible(object)

Arguments

object
a GtkWidget

Details

If no accessibility library is loaded (i.e. no ATK implementation library is loaded via GTK_MODULES or via another application library, such as libgnome), then this AtkObject instance may be a no-op. Likewise, if no class-specific AtkObject implementation is available for the widget instance in question, it will inherit an AtkObject implementation from the first ancestor class for which such an implementation is defined.

The documentation of the ATK (http://developer.gnome.org/doc/API/2.0/atk/index.html) library contains more information about accessible objects and their uses.

Value

[AtkObject] the AtkObject associated with widget. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]