gtkWidgetGetAncestor {RGtk2}	R Documentation

gtkWidgetGetAncestor

Description

Gets the first ancestor of widget with type widget.type. For example, gtk_widget_get_ancestor (widget, GTK_TYPE_BOX) gets the first GtkBox that's an ancestor of widget. No reference will be added to the returned widget; it should not be unreferenced. See note about checking for a toplevel GtkWindow in the docs for gtkWidgetGetToplevel.

Usage

gtkWidgetGetAncestor(object, widget.type)

Arguments

object
a GtkWidget

widget.type
ancestor type

Details

Note that unlike gtkWidgetIsAncestor, gtkWidgetGetAncestor considers widget to be an ancestor of itself.

Value

[GtkWidget] the ancestor widget, or NULL if not found. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]