gtkWidgetGetChildRequisition {RGtk2}	R Documentation

gtkWidgetGetChildRequisition

Description

This function is only for use in widget implementations. Obtains widget->requisition, unless someone has forced a particular geometry on the widget (e.g. with gtkWidgetSetSizeRequest), in which case it returns that geometry instead of the widget's requisition.

Usage

gtkWidgetGetChildRequisition(object)

Arguments

object
a GtkWidget

Details

This function differs from gtkWidgetSizeRequest in that it retrieves the last size request value from widget->requisition, while gtkWidgetSizeRequest actually calls the "size_request" method on widget to compute the size request and fill in widget->requisition, and only then returns widget->requisition.

Because this function does not call the "size_request" method, it can only be used when you know that widget->requisition is up-to-date, that is, gtkWidgetSizeRequest has been called since the last time a resize was queued. In general, only container implementations have this information; applications should use gtkWidgetSizeRequest.

Value

A list containing the following elements:

requisition
a GtkRequisition to be filled in. [ out ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]