gtkWidgetGetClipboard {RGtk2}	R Documentation

gtkWidgetGetClipboard

Description

Returns the clipboard object for the given selection to be used with widget. widget must have a GdkDisplay associated with it, so must be attached to a toplevel window.

Usage

gtkWidgetGetClipboard(object, selection)

Arguments

object
a GtkWidget

selection
a GdkAtom which identifies the clipboard to use. GDK_SELECTION_CLIPBOARD gives the default clipboard. Another common value is GDK_SELECTION_PRIMARY, which gives the primary X selection.

Details

Since 2.2

Value

[GtkClipboard] the appropriate clipboard object. If no clipboard already exists, a new one will be created. Once a clipboard object has been created, it is persistent for all time. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]