gtkWidgetGetColormap {RGtk2}	R Documentation

gtkWidgetGetColormap

Description

Gets the colormap that will be used to render widget. No reference will be added to the returned colormap; it should not be unreferenced.

Usage

gtkWidgetGetColormap(object)

Arguments

object
a GtkWidget

Value

[GdkColormap] the colormap used by widget. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]