gtkWidgetGetDisplay {RGtk2}	R Documentation

gtkWidgetGetDisplay

Description

Get the GdkDisplay for the toplevel window associated with this widget. This function can only be called after the widget has been added to a widget hierarchy with a GtkWindow at the top.

Usage

gtkWidgetGetDisplay(object)

Arguments

object
a GtkWidget

Details

In general, you should only create display specific resources when a widget has been realized, and you should free those resources when the widget is unrealized. Since 2.2

Value

[GdkDisplay] the GdkDisplay for the toplevel for this widget. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]