gtkWidgetGetModifierStyle {RGtk2}	R Documentation

gtkWidgetGetModifierStyle

Description

Returns the current modifier style for the widget. (As set by gtkWidgetModifyStyle.) If no style has previously set, a new GtkRcStyle will be created with all values unset, and set as the modifier style for the widget. If you make changes to this rc style, you must call gtkWidgetModifyStyle, passing in the returned rc style, to make sure that your changes take effect.

Usage

gtkWidgetGetModifierStyle(object)

Arguments

object
a GtkWidget

Details

Caution: passing the style back to gtkWidgetModifyStyle will normally end up destroying it, because gtkWidgetModifyStyle copies the passed-in style and sets the copy as the new modifier style, thus dropping any reference to the old modifier style. Add a reference to the modifier style if you want to keep it alive.

Value

[GtkRcStyle] the modifier style for the widget. If you want to keep a pointer to value this around, [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]