gtkWidgetGetPangoContext {RGtk2}	R Documentation

gtkWidgetGetPangoContext

Description

Gets a PangoContext with the appropriate font map, font description, and base direction for this widget. Unlike the context returned by gtkWidgetCreatePangoContext, and will be updated to match any changes to the widget's attributes.

Usage

gtkWidgetGetPangoContext(object)

Arguments

object
a GtkWidget

Details

If you create and keep a PangoLayout using this context, you must deal with changes to the context by calling pangoLayoutContextChanged on the layout in response to the "style-set" and "direction-changed" signals for the widget.

Value

[PangoContext] the PangoContext for the widget. [ transfer none ]

Note

Unlike the context returned by gtkWidgetCreatePangoContext, this context is owned by the widget (it can be used until the screen for the widget changes or the widget is removed from its toplevel), and will be updated to match any changes to the widget's attributes.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]