gtkWidgetGetSettings {RGtk2}	R Documentation

gtkWidgetGetSettings

Description

Gets the settings object holding the settings (global property settings, RC file information, etc) used for this widget.

Usage

gtkWidgetGetSettings(object)

Arguments

object
a GtkWidget

Details

Note that this function can only be called when the GtkWidget is attached to a toplevel, since the settings object is specific to a particular GdkScreen.

Value

[GtkSettings] the relevant GtkSettings object. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]