gtkWidgetGetSizeRequest {RGtk2}	R Documentation

gtkWidgetGetSizeRequest

Description

Gets the size request that was explicitly set for the widget using gtkWidgetSetSizeRequest. A value of -1 stored in width or height indicates that that dimension has not been set explicitly and the natural requisition of the widget will be used intead. See gtkWidgetSetSizeRequest. To get the size a widget will actually use, call gtkWidgetSizeRequest instead of this function.

Usage

gtkWidgetGetSizeRequest(object)

Arguments

object
a GtkWidget

Value

A list containing the following elements:

width
(out): return location for width, or NULL. [ allow-none ]

height
(out): return location for height, or NULL. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]