gtkWidgetGrabDefault {RGtk2}	R Documentation

gtkWidgetGrabDefault

Description

Causes widget to become the default widget. widget must have the GTK_CAN_DEFAULT flag set; typically you have to set this flag yourself by calling gtk_widget_set_can_default ( widget , TRUE ). The default widget is activated when the user presses Enter in a window. Default widgets must be activatable, that is, gtkWidgetActivate should affect them.

Usage

gtkWidgetGrabDefault(object)

Arguments

object
a GtkWidget

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]