gtkWidgetGrabFocus {RGtk2}	R Documentation

gtkWidgetGrabFocus

Description

Causes widget to have the keyboard focus for the GtkWindow it's inside. widget must be a focusable widget, such as a GtkEntry; something like GtkFrame won't work.

Usage

gtkWidgetGrabFocus(object)

Arguments

object
a GtkWidget

Details

More precisely, it must have the GTK_CAN_FOCUS flag set. Use gtkWidgetSetCanFocus to modify that flag.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]