gtkWidgetHasScreen {RGtk2}	R Documentation

gtkWidgetHasScreen

Description

Checks whether there is a GdkScreen is associated with this widget. All toplevel widgets have an associated screen, and all widgets added into a hierarchy with a toplevel window at the top.

Usage

gtkWidgetHasScreen(object)

Arguments

object
a GtkWidget

Details

Since 2.2

Value

[logical] TRUE if there is a GdkScreen associcated with the widget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]