gtkWidgetInputShapeCombineMask {RGtk2}	R Documentation

gtkWidgetInputShapeCombineMask

Description

Sets an input shape for this widget's GDK window. This allows for windows which react to mouse click in a nonrectangular region, see gdkWindowInputShapeCombineMask for more information.

Usage

gtkWidgetInputShapeCombineMask(object, shape.mask = NULL, offset.x, 
    offset.y)

Arguments

object
a GtkWidget

shape.mask
shape to be added, or NULL to remove an existing shape. [ allow-none ]

offset.x
X position of shape mask with respect to window

offset.y
Y position of shape mask with respect to window

Details

Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]