gtkWidgetIntersect {RGtk2}	R Documentation

gtkWidgetIntersect

Description

Computes the intersection of a widget's area and area, storing the intersection in intersection, and returns TRUE if there was an intersection. intersection may be NULL if you're only interested in whether there was an intersection.

Usage

gtkWidgetIntersect(object, area, intersection)

Arguments

object
a GtkWidget

area
a rectangle

intersection
rectangle to store intersection of widget and area

Value

[logical] TRUE if there was an intersection

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]