gtkWidgetIsComposited {RGtk2}	R Documentation

gtkWidgetIsComposited

Description

Whether widget can rely on having its alpha channel drawn correctly. On X11 this function returns whether a compositing manager is running for widget's screen.

Usage

gtkWidgetIsComposited(object)

Arguments

object
a GtkWidget

Details

Please note that the semantics of this call will change in the future if used on a widget that has a composited window in its hierarchy (as set by gdkWindowSetComposited). Since 2.10

Value

[logical] TRUE if the widget can rely on its alpha channel being drawn correctly.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]