gtkWidgetListAccelClosures {RGtk2}	R Documentation

gtkWidgetListAccelClosures

Description

Lists the closures used by widget for accelerator group connections with gtkAccelGroupConnectByPath or gtkAccelGroupConnect. The closures can be used to monitor accelerator changes on widget, by connecting to the GtkAccelGroup::accel-changed signal of the GtkAccelGroup of a closure which can be found out with gtkAccelGroupFromAccelClosure.

Usage

gtkWidgetListAccelClosures(object)

Arguments

object
widget to list accelerator closures for

Value

[list] a newly allocated list of closures

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]