gtkWidgetMap {RGtk2}	R Documentation

gtkWidgetMap

Description

This function is only for use in widget implementations. Causes a widget to be mapped if it isn't already.

Usage

gtkWidgetMap(object)

Arguments

object
a GtkWidget

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]