gtkWidgetModifyBase {RGtk2}	R Documentation

gtkWidgetModifyBase

Description

Sets the base color for a widget in a particular state. All other style values are left untouched. The base color is the background color used along with the text color (see gtkWidgetModifyText) for widgets such as GtkEntry and GtkTextView. See also gtkWidgetModifyStyle.

Usage

gtkWidgetModifyBase(object, state, color = NULL)

Arguments

object
a GtkWidget

state
the state for which to set the base color

color
the color to assign (does not need to be allocated), or NULL to undo the effect of previous calls to of gtkWidgetModifyBase. [ allow-none ]

Details

Note that "no window" widgets (which have the GTK_NO_WINDOW flag set) draw on their parent container's window and thus may not draw any background themselves. This is the case for e.g. GtkLabel. To modify the background of such widgets, you have to set the base color on their parent; if you want to set the background of a rectangular area around a label, try placing the label in a GtkEventBox widget and setting the base color on that.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]