gtkWidgetModifyCursor {RGtk2}	R Documentation

gtkWidgetModifyCursor

Description

Sets the cursor color to use in a widget, overriding the "cursor-color" and "secondary-cursor-color" style properties. All other style values are left untouched. See also gtkWidgetModifyStyle.

Usage

gtkWidgetModifyCursor(object, primary, secondary)

Arguments

object
a GtkWidget

primary
the color to use for primary cursor (does not need to be allocated), or NULL to undo the effect of previous calls to of gtkWidgetModifyCursor.

secondary
the color to use for secondary cursor (does not need to be allocated), or NULL to undo the effect of previous calls to of gtkWidgetModifyCursor.

Details

Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]