gtkWidgetModifyStyle {RGtk2}	R Documentation

gtkWidgetModifyStyle

Description

Modifies style values on the widget. Modifications made using this technique take precedence over style values set via an RC file, however, they will be overriden if a style is explicitely set on the widget using gtkWidgetSetStyle. The GtkRcStyle structure is designed so each field can either be set or unset, so it is possible, using this function, to modify some style values and leave the others unchanged.

Usage

gtkWidgetModifyStyle(object, style)

Arguments

object
a GtkWidget

style
the GtkRcStyle holding the style modifications

Details

Note that modifications made with this function are not cumulative with previous calls to gtkWidgetModifyStyle or with such functions as gtkWidgetModifyFg. If you wish to retain previous values, you must first call gtkWidgetGetModifierStyle, make your modifications to the returned style, then call gtkWidgetModifyStyle with that style. On the other hand, if you first call gtkWidgetModifyStyle, subsequent calls to such functions gtkWidgetModifyFg will have a cumulative effect with the initial modifications.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]