gtkWidgetModifyText {RGtk2}	R Documentation

gtkWidgetModifyText

Description

Sets the text color for a widget in a particular state. All other style values are left untouched. The text color is the foreground color used along with the base color (see gtkWidgetModifyBase) for widgets such as GtkEntry and GtkTextView. See also gtkWidgetModifyStyle.

Usage

gtkWidgetModifyText(object, state, color = NULL)

Arguments

object
a GtkWidget

state
the state for which to set the text color

color
the color to assign (does not need to be allocated), or NULL to undo the effect of previous calls to of gtkWidgetModifyText. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]