gtkWidgetPushCompositeChild {RGtk2}	R Documentation

gtkWidgetPushCompositeChild

Description

Makes all newly-created widgets as composite children until the corresponding gtkWidgetPopCompositeChild call.

Usage

gtkWidgetPushCompositeChild()

Details

A composite child is a child that's an implementation detail of the container it's inside and should not be visible to people using the container. Composite children aren't treated differently by GTK (but see gtkContainerForeach vs. gtkContainerForall), but e.g. GUI builders might want to treat them in a different way.

Here is a simple example:

gtkWidgetPushCompositeChild()
hscrollbar <- gtkHScrollbarNew(hadjustment)
hscrollbar$setCompositeName("hscrollbar")
gtkWidgetPopCompositeChild()
hscrollbar$setParent(scrolled_window)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]