gtkWidgetQueueClearArea {RGtk2}	R Documentation

gtkWidgetQueueClearArea

Description

This function is no longer different from gtkWidgetQueueDrawArea, though it once was. Now it just calls gtkWidgetQueueDrawArea. Originally gtkWidgetQueueClearArea would force a redraw of the background for GTK_NO_WINDOW widgets, and gtkWidgetQueueDrawArea would not. Now both functions ensure the background will be redrawn. WARNING: gtk_widget_queue_clear_area has been deprecated since version 2.2 and should not be used in newly-written code. Use gtkWidgetQueueDrawArea instead.

Usage

gtkWidgetQueueClearArea(object, x, y, width, height)

Arguments

object
a GtkWidget

x
x coordinate of upper-left corner of rectangle to redraw

y
y coordinate of upper-left corner of rectangle to redraw

width
width of region to draw

height
height of region to draw

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]