gtkWidgetQueueDrawArea {RGtk2}	R Documentation

gtkWidgetQueueDrawArea

Description

Invalidates the rectangular area of widget defined by x, y, width and height by calling gdkWindowInvalidateRect on the widget's window and all its child windows. Once the main loop becomes idle (after the current batch of events has been processed, roughly), the window will receive expose events for the union of all regions that have been invalidated.

Usage

gtkWidgetQueueDrawArea(object, x, y, width, height)

Arguments

object
a GtkWidget

x
x coordinate of upper-left corner of rectangle to redraw

y
y coordinate of upper-left corner of rectangle to redraw

width
width of region to draw

height
height of region to draw

Details

Normally you would only use this function in widget implementations. You might also use it, or gdkWindowInvalidateRect directly, to schedule a redraw of a GtkDrawingArea or some portion thereof.

Frequently you can just call gdkWindowInvalidateRect or gdkWindowInvalidateRegion instead of this function. Those functions will invalidate only a single window, instead of the widget and all its children.

The advantage of adding to the invalidated region compared to simply drawing immediately is efficiency; using an invalid region ensures that you only have to redraw one time.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]