gtkWidgetRemoveAccelerator {RGtk2}	R Documentation

gtkWidgetRemoveAccelerator

Description

Removes an accelerator from widget, previously installed with gtkWidgetAddAccelerator.

Usage

gtkWidgetRemoveAccelerator(object, accel.group, accel.key, accel.mods)

Arguments

object
widget to install an accelerator on

accel.group
accel group for this widget

accel.key
GDK keyval of the accelerator

accel.mods
modifier key combination of the accelerator

Value

[logical] whether an accelerator was installed and could be removed

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]