gtkWidgetRenderIcon {RGtk2}	R Documentation

gtkWidgetRenderIcon

Description

A convenience function that uses the theme engine and RC file settings for widget to look up stock.id and render it to a pixbuf. stock.id should be a stock icon ID such as GTK_STOCK_OPEN or GTK_STOCK_OK. size should be a size such as GTK_ICON_SIZE_MENU. detail should be a string that identifies the widget or code doing the rendering, so that theme engines can special-case rendering for that widget or code.

Usage

gtkWidgetRenderIcon(object, stock.id, size, detail = NULL)

Arguments

object
a GtkWidget

stock.id
a stock ID

size
(type int) a stock size. A size of (GtkIconSize)-1 means render at the size of the source and don't scale (if there are multiple source sizes, GTK+ picks one of the available sizes).

detail
render detail to pass to theme engine. [ allow-none ]

Details

The pixels in the returned GdkPixbuf are shared with the rest of the application and should not be modified.

Value

[GdkPixbuf] a new pixbuf, or NULL if the stock ID wasn't known

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]