gtkWidgetSendExpose {RGtk2}	R Documentation

gtkWidgetSendExpose

Description

Very rarely-used function. This function is used to emit an expose event signals on a widget. This function is not normally used directly. The only time it is used is when propagating an expose event to a child NO_WINDOW widget, and that is normally done using gtkContainerPropagateExpose.

Usage

gtkWidgetSendExpose(object, event)

Arguments

object
a GtkWidget

event
a expose GdkEvent

Details

If you want to force an area of a window to be redrawn, use gdkWindowInvalidateRect or gdkWindowInvalidateRegion. To cause the redraw to be done immediately, follow that call with a call to gdkWindowProcessUpdates.

Value

[integer] return from the event signal emission (TRUE if the event was handled)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]