gtkWidgetSetAccelPath {RGtk2}	R Documentation

gtkWidgetSetAccelPath

Description

Given an accelerator group, accel.group, and an accelerator path, accel.path, sets up an accelerator in accel.group so whenever the key binding that is defined for accel.path is pressed, widget will be activated. This removes any accelerators (for any accelerator group) installed by previous calls to gtkWidgetSetAccelPath. Associating accelerators with paths allows them to be modified by the user and the modifications to be saved for future use. (See gtkAccelMapSave.)

Usage

gtkWidgetSetAccelPath(object, accel.path, accel.group)

Arguments

object
a GtkWidget

accel.path
path used to look up the accelerator. [ allow-none ]

accel.group
a GtkAccelGroup. [ allow-none ]

Details

This function is a low level function that would most likely be used by a menu creation system like GtkUIManager. If you use GtkUIManager, setting up accelerator paths will be done automatically.

Even when you you aren't using GtkUIManager, if you only want to set up accelerators on menu items gtkMenuItemSetAccelPath provides a somewhat more convenient interface.

Note that accel.path string will be stored in a GQuark. Therefore, if you pass a static string, you can save some memory by interning it first with gInternStaticString().

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]