gtkWidgetSetAppPaintable {RGtk2}	R Documentation

gtkWidgetSetAppPaintable

Description

Sets whether the application intends to draw on the widget in an "expose-event" handler.

Usage

gtkWidgetSetAppPaintable(object, app.paintable)

Arguments

object
a GtkWidget

app.paintable
TRUE if the application will paint on the widget

Details

This is a hint to the widget and does not affect the behavior of the GTK+ core; many widgets ignore this flag entirely. For widgets that do pay attention to the flag, such as GtkEventBox and GtkWindow, the effect is to suppress default themed drawing of the widget's background. (Children of the widget will still be drawn.) The application is then entirely responsible for drawing the widget background.

Note that the background is still drawn when the widget is mapped. If this is not suitable (e.g. because you want to make a transparent window using an RGBA visual), you can work around this by doing:

window$realize()
window$window$setBackPixmap(NULL, FALSE)
window$show()

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]