gtkWidgetSetChildVisible {RGtk2}	R Documentation

gtkWidgetSetChildVisible

Description

Sets whether widget should be mapped along with its when its parent is mapped and widget has been shown with gtkWidgetShow.

Usage

gtkWidgetSetChildVisible(object, is.visible)

Arguments

object
a GtkWidget

is.visible
if TRUE, widget should be mapped along with its parent.

Details

The child visibility can be set for widget before it is added to a container with gtkWidgetSetParent, to avoid mapping children unnecessary before immediately unmapping them. However it will be reset to its default state of TRUE when the widget is removed from a container.

Note that changing the child visibility of a widget does not queue a resize on the widget. Most of the time, the size of a widget is computed from all visible children, whether or not they are mapped. If this is not the case, the container can queue a resize itself.

This function is only useful for container implementations and never should be called by an application.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]