gtkWidgetSetDefaultDirection {RGtk2}	R Documentation

gtkWidgetSetDefaultDirection

Description

Sets the default reading direction for widgets where the direction has not been explicitly set by gtkWidgetSetDirection.

Usage

gtkWidgetSetDefaultDirection(dir)

Arguments

dir
the new default direction. This cannot be GTK_TEXT_DIR_NONE.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]