gtkWidgetSetDirection {RGtk2}	R Documentation

gtkWidgetSetDirection

Description

Sets the reading direction on a particular widget. This direction controls the primary direction for widgets containing text, and also the direction in which the children of a container are packed. The ability to set the direction is present in order so that correct localization into languages with right-to-left reading directions can be done. Generally, applications will let the default reading direction present, except for containers where the containers are arranged in an order that is explicitely visual rather than logical (such as buttons for text justification).

Usage

gtkWidgetSetDirection(object, dir)

Arguments

object
a GtkWidget

dir
the new direction

Details

If the direction is set to GTK_TEXT_DIR_NONE, then the value set by gtkWidgetSetDefaultDirection will be used.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]