gtkWidgetSetDoubleBuffered {RGtk2}	R Documentation

gtkWidgetSetDoubleBuffered

Description

Widgets are double buffered by default; you can use this function to turn off the buffering. "Double buffered" simply means that gdkWindowBeginPaintRegion and gdkWindowEndPaint are called automatically around expose events sent to the widget. gdkWindowBeginPaint() diverts all drawing to a widget's window to an offscreen buffer, and gdkWindowEndPaint draws the buffer to the screen. The result is that users see the window update in one smooth step, and don't see individual graphics primitives being rendered.

Usage

gtkWidgetSetDoubleBuffered(object, double.buffered)

Arguments

object
a GtkWidget

double.buffered
TRUE to double-buffer a widget

Details

In very simple terms, double buffered widgets don't flicker, so you would only use this function to turn off double buffering if you had special needs and really knew what you were doing.

Note: if you turn off double-buffering, you have to handle expose events, since even the clearing to the background color or pixmap will not happen automatically (as it is done in gdkWindowBeginPaint()).

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]