gtkWidgetSetParent {RGtk2}	R Documentation

gtkWidgetSetParent

Description

This function is useful only when implementing subclasses of GtkContainer. Sets the container as the parent of widget, and takes care of some details such as updating the state and style of the child to reflect its new location. The opposite function is gtkWidgetUnparent.

Usage

gtkWidgetSetParent(object, parent)

Arguments

object
a GtkWidget

parent
parent container

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]