gtkWidgetSetReceivesDefault {RGtk2}	R Documentation

gtkWidgetSetReceivesDefault

Description

Specifies whether widget will be treated as the default widget within its toplevel when it has the focus, even if another widget is the default.

Usage

gtkWidgetSetReceivesDefault(object, receives.default)

Arguments

object
a GtkWidget

receives.default
whether or not widget can be a default widget.

Details

See gtkWidgetGrabDefault for details about the meaning of "default". Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]