gtkWidgetSetScrollAdjustments {RGtk2}	R Documentation

gtkWidgetSetScrollAdjustments

Description

For widgets that support scrolling, sets the scroll adjustments and returns TRUE. For widgets that don't support scrolling, does nothing and returns FALSE. Widgets that don't support scrolling can be scrolled by placing them in a GtkViewport, which does support scrolling.

Usage

gtkWidgetSetScrollAdjustments(object, hadjustment = NULL, vadjustment = NULL)

Arguments

object
a GtkWidget

hadjustment
an adjustment for horizontal scrolling, or NULL. [ allow-none ]

vadjustment
an adjustment for vertical scrolling, or NULL. [ allow-none ]

Value

[logical] TRUE if the widget supports scrolling

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]