gtkWidgetSetStyle {RGtk2}	R Documentation

gtkWidgetSetStyle

Description

Sets the GtkStyle for a widget (widget->style). You probably don't want to use this function; it interacts badly with themes, because themes work by replacing the GtkStyle. Instead, use gtkWidgetModifyStyle.

Usage

gtkWidgetSetStyle(object, style = NULL)

Arguments

object
a GtkWidget

style
a GtkStyle, or NULL to remove the effect of a previous gtkWidgetSetStyle and go back to the default style. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]