gtkWidgetSetUposition {RGtk2}	R Documentation

gtkWidgetSetUposition

Description

Sets the position of a widget. The funny "u" in the name comes from the "user position" hint specified by the X Window System, and exists for legacy reasons. This function doesn't work if a widget is inside a container; it's only really useful on GtkWindow. WARNING: gtk_widget_set_uposition is deprecated and should not be used in newly-written code.

Usage

gtkWidgetSetUposition(object, x, y)

Arguments

object
a GtkWidget

x
x position; -1 to unset x; -2 to leave x unchanged

y
y position; -1 to unset y; -2 to leave y unchanged

Details

Don't use this function to center dialogs over the main application window; most window managers will do the centering on your behalf if you call gtkWindowSetTransientFor, and it's really not possible to get the centering to work correctly in all cases from application code. But if you insist, use gtkWindowSetPosition to set GTK_WIN_POS_CENTER_ON_PARENT, don't do the centering manually.

Note that although x and y can be individually unset, the position is not honoured unless both x and y are set.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]