gtkWidgetSetUsize {RGtk2}	R Documentation

gtkWidgetSetUsize

Description

Sets the minimum size of a widget; that is, the widget's size request will be width by height. You can use this function to force a widget to be either larger or smaller than it is. The strange "usize" name dates from the early days of GTK+, and derives from X Window System terminology. In many cases, gtkWindowSetDefaultSize is a better choice for toplevel windows than this function; setting the default size will still allow users to shrink the window. Setting the usize will force them to leave the window at least as large as the usize. When dealing with window sizes, gtkWindowSetGeometryHints can be a useful function as well. WARNING: gtk_widget_set_usize has been deprecated since version 2.2 and should not be used in newly-written code. Use gtkWidgetSetSizeRequest instead.

Usage

gtkWidgetSetUsize(object, width, height)

Arguments

object
a GtkWidget

width
minimum width, or -1 to unset

height
minimum height, or -1 to unset

Details

Note the inherent danger of setting any fixed size - themes, translations into other languages, different fonts, and user action can all change the appropriate size for a given widget. So, it's basically impossible to hardcode a size that will always be correct.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]