gtkWidgetSetVisible {RGtk2}	R Documentation

gtkWidgetSetVisible

Description

Sets the visibility state of widget. Note that setting this to TRUE doesn't mean the widget is actually viewable, see gtkWidgetGetVisible.

Usage

gtkWidgetSetVisible(object, visible)

Arguments

object
a GtkWidget

visible
whether the widget should be shown or not

Details

This function simply calls gtkWidgetShow or gtkWidgetHide but is nicer to use when the visibility of the widget depends on some condition. Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]