gtkWidgetSetWindow {RGtk2}	R Documentation

gtkWidgetSetWindow

Description

Sets a widget's window. This function should only be used in a widget's GtkWidget::realize() implementation. The window passed is usually either new window created with gdkWindowNew, or the window of its parent widget as returned by gtkWidgetGetParentWindow.

Usage

gtkWidgetSetWindow(object, window)

Arguments

object
a GtkWidget

window
a GdkWindow

Details

Widgets must indicate whether they will create their own GdkWindow by calling gtkWidgetSetHasWindow. This is usually done in the widget's init() function. Since 2.18

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]