gtkWidgetShapeCombineMask {RGtk2}	R Documentation

gtkWidgetShapeCombineMask

Description

Sets a shape for this widget's GDK window. This allows for transparent windows etc., see gdkWindowShapeCombineMask for more information.

Usage

gtkWidgetShapeCombineMask(object, shape.mask, offset.x, offset.y)

Arguments

object
a GtkWidget

shape.mask
shape to be added, or NULL to remove an existing shape. [ allow-none ]

offset.x
X position of shape mask with respect to window

offset.y
Y position of shape mask with respect to window

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]