gtkWidgetSizeRequest {RGtk2}	R Documentation

gtkWidgetSizeRequest

Description

This function is typically used when implementing a GtkContainer subclass. Obtains the preferred size of a widget. The container uses this information to arrange its child widgets and decide what size allocations to give them with gtkWidgetSizeAllocate.

Usage

gtkWidgetSizeRequest(object)

Arguments

object
a GtkWidget

Details

You can also call this function from an application, with some caveats. Most notably, getting a size request requires the widget to be associated with a screen, because font information may be needed. Multihead-aware applications should keep this in mind.

Also remember that the size request is not necessarily the size a widget will actually be allocated.

See also gtkWidgetGetChildRequisition.

Value

A list containing the following elements:

requisition
a GtkRequisition to be filled in. [ out ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]