gtkWidgetStyleAttach {RGtk2}	R Documentation

gtkWidgetStyleAttach

Description

This function attaches the widget's GtkStyle to the widget's GdkWindow. It is a replacement for

Usage

gtkWidgetStyleAttach(object)

Arguments

object
undocumented

Details

widget->style = gtk_style_attach (widget->style, widget->window);

and should only ever be called in a derived widget's "realize" implementation which does not chain up to its parent class' "realize" implementation, because one of the parent classes (finally GtkWidget) would attach the style itself. Since 2.20

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]