gtkWidgetTranslateCoordinates {RGtk2}	R Documentation

gtkWidgetTranslateCoordinates

Description

Translate coordinates relative to src.widget's allocation to coordinates relative to dest.widget's allocations. In order to perform this operation, both widgets must be realized, and must share a common toplevel.

Usage

gtkWidgetTranslateCoordinates(object, dest.widget, src.x, src.y)

Arguments

object
a GtkWidget

dest.widget
a GtkWidget

src.x
X position relative to src.widget

src.y
Y position relative to src.widget

Value

A list containing the following elements:

retval
[logical] FALSE if either widget was not realized, or there was no common ancestor. In this case, nothing is stored in *dest.x and *dest.y. Otherwise TRUE.

dest.x
location to store X position relative to dest.widget. [ out ]

dest.y
location to store Y position relative to dest.widget. [ out ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]