gtkWindowBeginMoveDrag {RGtk2}	R Documentation

gtkWindowBeginMoveDrag

Description

Starts moving a window. This function is used if an application has window movement grips. When GDK can support it, the window movement will be done using the standard mechanism for the window manager or windowing system. Otherwise, GDK will try to emulate window movement, potentially not all that well, depending on the windowing system.

Usage

gtkWindowBeginMoveDrag(object, button, root.x, root.y, timestamp)

Arguments

object
a GtkWindow

button
mouse button that initiated the drag

root.x
X position where the user clicked to initiate the drag, in root window coordinates

root.y
Y position where the user clicked to initiate the drag

timestamp
timestamp from the click event that initiated the drag

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]