gtkWindowBeginResizeDrag {RGtk2}	R Documentation

gtkWindowBeginResizeDrag

Description

Starts resizing a window. This function is used if an application has window resizing controls. When GDK can support it, the resize will be done using the standard mechanism for the window manager or windowing system. Otherwise, GDK will try to emulate window resizing, potentially not all that well, depending on the windowing system.

Usage

gtkWindowBeginResizeDrag(object, edge, button, root.x, root.y, 
    timestamp)

Arguments

object
a GtkWindow

edge
position of the resize control

button
mouse button that initiated the drag

root.x
X position where the user clicked to initiate the drag, in root window coordinates

root.y
Y position where the user clicked to initiate the drag

timestamp
timestamp from the click event that initiated the drag

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]