gtkWindowDeiconify {RGtk2}	R Documentation

gtkWindowDeiconify

Description

Asks to deiconify (i.e. unminimize) the specified window. Note that you shouldn't assume the window is definitely deiconified afterward, because other entities (e.g. the user or window manager) could iconify it again before your code which assumes deiconification gets to run.

Usage

gtkWindowDeiconify(object)

Arguments

object
a GtkWindow

Details

You can track iconification via the "window-state-event" signal on GtkWidget.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]