gtkWindowGetDefaultIconName {RGtk2}	R Documentation

gtkWindowGetDefaultIconName

Description

Returns the fallback icon name for windows that has been set with gtkWindowSetDefaultIconName. It is only valid until the next call to gtkWindowSetDefaultIconName.

Usage

gtkWindowGetDefaultIconName()

Details

Since 2.16

Value

[character] the fallback icon name for windows

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]