gtkWindowGetFocus {RGtk2}	R Documentation

gtkWindowGetFocus

Description

Retrieves the current focused widget within the window. Note that this is the widget that would have the focus if the toplevel window focused; if the toplevel window is not focused then gtk_widget_has_focus (widget) will not be TRUE for the widget.

Usage

gtkWindowGetFocus(object)

Arguments

object
a GtkWindow

Value

[GtkWidget] the currently focused widget, or NULL if there is none. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]