gtkWindowGetFrameDimensions {RGtk2}	R Documentation

gtkWindowGetFrameDimensions

Description

(Note: this is a special-purpose function intended for the framebuffer port; see gtkWindowSetHasFrame. It will not return the size of the window border drawn by the window manager, which is the normal case when using a windowing system. See gdkWindowGetFrameExtents to get the standard window border extents.)

Usage

gtkWindowGetFrameDimensions(object)

Arguments

object
a GtkWindow

Details

Retrieves the dimensions of the frame window for this toplevel. See gtkWindowSetHasFrame, gtkWindowSetFrameDimensions.

Value

A list containing the following elements:

left
location to store the width of the frame at the left, or NULL. [ allow-none ][ out ]

top
location to store the height of the frame at the top, or NULL. [ allow-none ][ out ]

right
location to store the width of the frame at the returns, or NULL. [ allow-none ][ out ]

bottom
location to store the height of the frame at the bottom, or NULL. [ allow-none ][ out ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]