gtkWindowGetGroup {RGtk2}	R Documentation

gtkWindowGetGroup

Description

Returns the group for window or the default group, if window is NULL or if window does not have an explicit window group.

Usage

gtkWindowGetGroup(object)

Arguments

object
a GtkWindow, or NULL. [ allow-none ]

Details

Since 2.10

Value

[GtkWindowGroup] the GtkWindowGroup for a window or the default group. [ transfer none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]