gtkWindowIsActive {RGtk2}	R Documentation

gtkWindowIsActive

Description

Returns whether the window is part of the current active toplevel. (That is, the toplevel window receiving keystrokes.) The return value is TRUE if the window is active toplevel itself, but also if it is, say, a GtkPlug embedded in the active toplevel. You might use this function if you wanted to draw a widget differently in an active window from a widget in an inactive window. See gtkWindowHasToplevelFocus

Usage

gtkWindowIsActive(object)

Arguments

object
a GtkWindow

Details

Since 2.4

Value

[logical] TRUE if the window part of the current active window.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]