gtkWindowListToplevels {RGtk2}	R Documentation

gtkWindowListToplevels

Description

Returns a list of all existing toplevel windows. The widgets in the list are not individually referenced. If you want to iterate through the list and perform actions involving callbacks that might destroy the widgets, you must call g_list_foreach (result, (GFunc)g_object_ref, NULL) first, and then unref all the widgets afterwards.

Usage

gtkWindowListToplevels()

Value

[list] list of toplevel widgets. [ element-type GtkWidget][ transfer container ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]