gtkWindowNew {RGtk2}	R Documentation

gtkWindowNew

Description

Creates a new GtkWindow, which is a toplevel window that can contain other widgets. Nearly always, the type of the window should be GTK_WINDOW_TOPLEVEL. If you're implementing something like a popup menu from scratch (which is a bad idea, just use GtkMenu), you might use GTK_WINDOW_POPUP. GTK_WINDOW_POPUP is not for dialogs, though in some other toolkits dialogs are called "popups". In GTK+, GTK_WINDOW_POPUP means a pop-up menu or pop-up tooltip. On X11, popup windows are not controlled by the window manager.

Usage

gtkWindowNew(type = NULL, show = TRUE)

Arguments

type
type of window

Details

If you simply want an undecorated window (no window borders), use gtkWindowSetDecorated, don't use GTK_WINDOW_POPUP.

Value

[GtkWidget] a new GtkWindow.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]