gtkWindowPresent {RGtk2}	R Documentation

gtkWindowPresent

Description

Presents a window to the user. This may mean raising the window in the stacking order, deiconifying it, moving it to the current desktop, and/or giving it the keyboard focus, possibly dependent on the user's platform, window manager, and preferences.

Usage

gtkWindowPresent(object)

Arguments

object
a GtkWindow

Details

If window is hidden, this function calls gtkWidgetShow as well.

This function should be used when the user tries to open a window that's already open. Say for example the preferences dialog is currently open, and the user chooses Preferences from the menu a second time; use gtkWindowPresent to move the already-open dialog where the user can see it.

If you are calling this function in response to a user interaction, it is preferable to use gtkWindowPresentWithTime.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]