gtkWindowPresentWithTime {RGtk2}	R Documentation

gtkWindowPresentWithTime

Description

Presents a window to the user in response to a user interaction. If you need to present a window without a timestamp, use gtkWindowPresent. See gtkWindowPresent for details.

Usage

gtkWindowPresentWithTime(object, timestamp)

Arguments

object
a GtkWindow

timestamp
the timestamp of the user interaction (typically a button or key press event) which triggered this call

Details

Since 2.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]