gtkWindowPropagateKeyEvent {RGtk2}	R Documentation

gtkWindowPropagateKeyEvent

Description

Propagate a key press or release event to the focus widget and up the focus container chain until a widget handles event. This is normally called by the default ::key_press_event and ::key_release_event handlers for toplevel windows, however in some cases it may be useful to call this directly when overriding the standard key handling for a toplevel window.

Usage

gtkWindowPropagateKeyEvent(object, event)

Arguments

object
a GtkWindow

event
a GdkEventKey

Details

Since 2.4

Value

[logical] TRUE if a widget in the focus chain handled the event.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]