gtkWindowResize {RGtk2}	R Documentation

gtkWindowResize

Description

Resizes the window as if the user had done so, obeying geometry constraints. The default geometry constraint is that windows may not be smaller than their size request; to override this constraint, call gtkWidgetSetSizeRequest to set the window's request to a smaller value.

Usage

gtkWindowResize(object, width, height)

Arguments

object
a GtkWindow

width
width in pixels to resize the window to

height
height in pixels to resize the window to

Details

If gtkWindowResize is called before showing a window for the first time, it overrides any default size set with gtkWindowSetDefaultSize.

Windows may not be resized smaller than 1 by 1 pixels.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]