gtkWindowSetAutoStartupNotification {RGtk2}	R Documentation

gtkWindowSetAutoStartupNotification

Description

By default, after showing the first GtkWindow, GTK+ calls gdkNotifyStartupComplete. Call this function to disable the automatic startup notification. You might do this if your first window is a splash screen, and you want to delay notification until after your real main window has been shown, for example.

Usage

gtkWindowSetAutoStartupNotification(setting)

Arguments

setting
TRUE to automatically do startup notification

Details

In that example, you would disable startup notification temporarily, show your splash screen, then re-enable it so that showing the main window would automatically result in notification. Since 2.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]