gtkWindowSetDecorated {RGtk2}	R Documentation

gtkWindowSetDecorated

Description

By default, windows are decorated with a title bar, resize controls, etc. Some window managers allow GTK+ to disable these decorations, creating a borderless window. If you set the decorated property to FALSE using this function, GTK+ will do its best to convince the window manager not to decorate the window. Depending on the system, this function may not have any effect when called on a window that is already visible, so you should call it before calling gtkWindowShow().

Usage

gtkWindowSetDecorated(object, setting)

Arguments

object
a GtkWindow

setting
TRUE to decorate the window

Details

On Windows, this function always works, since there's no window manager policy involved.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]