gtkWindowSetDefault {RGtk2}	R Documentation

gtkWindowSetDefault

Description

The default widget is the widget that's activated when the user presses Enter in a dialog (for example). This function sets or unsets the default widget for a GtkWindow about. When setting (rather than unsetting) the default widget it's generally easier to call gtkWidgetGrabFocus on the widget. Before making a widget the default widget, you must set the GTK_CAN_DEFAULT flag on the widget you'd like to make the default using gtkWidgetSetFlags().

Usage

gtkWindowSetDefault(object, default.widget = NULL)

Arguments

object
a GtkWindow

default.widget
widget to be the default, or NULL to unset the default widget for the toplevel. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]