gtkWindowSetDefaultIconList {RGtk2}	R Documentation

gtkWindowSetDefaultIconList

Description

Sets an icon list to be used as fallback for windows that haven't had gtkWindowSetIconList called on them to set up a window-specific icon list. This function allows you to set up the icon for all windows in your app at once.

Usage

gtkWindowSetDefaultIconList(list)

Arguments

list
(element-type GdkPixbuf) (transfer container) a list of GdkPixbuf

Details

See gtkWindowSetIconList for more details.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]