gtkWindowSetDeletable {RGtk2}	R Documentation

gtkWindowSetDeletable

Description

By default, windows have a close button in the window frame. Some window managers allow GTK+ to disable this button. If you set the deletable property to FALSE using this function, GTK+ will do its best to convince the window manager not to show a close button. Depending on the system, this function may not have any effect when called on a window that is already visible, so you should call it before calling gtkWindowShow().

Usage

gtkWindowSetDeletable(object, setting)

Arguments

object
a GtkWindow

setting
TRUE to decorate the window as deletable

Details

On Windows, this function always works, since there's no window manager policy involved. Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]