gtkWindowSetFocus {RGtk2}	R Documentation

gtkWindowSetFocus

Description

If focus is not the current focus widget, and is focusable, sets it as the focus widget for the window. If focus is NULL, unsets the focus widget for this window. To set the focus to a particular widget in the toplevel, it is usually more convenient to use gtkWidgetGrabFocus instead of this function.

Usage

gtkWindowSetFocus(object, focus = NULL)

Arguments

object
a GtkWindow

focus
widget to be the new focus widget, or NULL to unset any focus widget for the toplevel window. [ allow-none ]

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]