gtkWindowSetFocusOnMap {RGtk2}	R Documentation

gtkWindowSetFocusOnMap

Description

Windows may set a hint asking the desktop environment not to receive the input focus when the window is mapped. This function sets this hint.

Usage

gtkWindowSetFocusOnMap(object, setting)

Arguments

object
a GtkWindow

setting
TRUE to let this window receive input focus on map

Details

Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]