gtkWindowSetHasFrame {RGtk2}	R Documentation

gtkWindowSetHasFrame

Description

(Note: this is a special-purpose function for the framebuffer port, that causes GTK+ to draw its own window border. For most applications, you want gtkWindowSetDecorated instead, which tells the window manager whether to draw the window border.)

Usage

gtkWindowSetHasFrame(object, setting)

Arguments

object
a GtkWindow

setting
a boolean

Details

If this function is called on a window with setting of TRUE, before it is realized or showed, it will have a "frame" window around window->window, accessible in window->frame. Using the signal frame_event you can receive all events targeted at the frame.

This function is used by the linux-fb port to implement managed windows, but it could conceivably be used by X-programs that want to do their own window decorations.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]