gtkWindowSetIcon {RGtk2}	R Documentation

gtkWindowSetIcon

Description

Sets up the icon representing a GtkWindow. This icon is used when the window is minimized (also known as iconified). Some window managers or desktop environments may also place it in the window frame, or display it in other contexts.

Usage

gtkWindowSetIcon(object, icon = NULL)

Arguments

object
a GtkWindow

icon
icon image, or NULL. [ allow-none ]

Details

The icon should be provided in whatever size it was naturally drawn; that is, don't scale the image before passing it to GTK+. Scaling is postponed until the last minute, when the desired final size is known, to allow best quality.

If you have your icon hand-drawn in multiple sizes, use gtkWindowSetIconList. Then the best size will be used.

This function is equivalent to calling gtkWindowSetIconList with a 1-element list.

See also gtkWindowSetDefaultIconList to set the icon for all windows in your application in one go.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]