gtkWindowSetIconName {RGtk2}	R Documentation

gtkWindowSetIconName

Description

Sets the icon for the window from a named themed icon. See the docs for GtkIconTheme for more details.

Usage

gtkWindowSetIconName(object, name = NULL)

Arguments

object
a GtkWindow

name
the name of the themed icon. [ allow-none ]

Details

Note that this has nothing to do with the WM_ICON_NAME property which is mentioned in the ICCCM. Since 2.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]