gtkWindowSetOpacity {RGtk2}	R Documentation

gtkWindowSetOpacity

Description

Request the windowing system to make window partially transparent, with opacity 0 being fully transparent and 1 fully opaque. (Values of the opacity parameter are clamped to the [0,1] range.) On X11 this has any effect only on X screens with a compositing manager running. See gtkWidgetIsComposited. On Windows it should work always.

Usage

gtkWindowSetOpacity(object, opacity)

Arguments

object
a GtkWindow

opacity
desired opacity, between 0 and 1

Details

Note that setting a window's opacity after the window has been shown causes it to flicker once on Windows. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]