gtkWindowSetSkipPagerHint {RGtk2}	R Documentation

gtkWindowSetSkipPagerHint

Description

Windows may set a hint asking the desktop environment not to display the window in the pager. This function sets this hint. (A "pager" is any desktop navigation tool such as a workspace switcher that displays a thumbnail representation of the windows on the screen.)

Usage

gtkWindowSetSkipPagerHint(object, setting)

Arguments

object
a GtkWindow

setting
TRUE to keep this window from appearing in the pager

Details

Since 2.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]