gtkWindowSetSkipTaskbarHint {RGtk2}	R Documentation

gtkWindowSetSkipTaskbarHint

Description

Windows may set a hint asking the desktop environment not to display the window in the task bar. This function sets this hint.

Usage

gtkWindowSetSkipTaskbarHint(object, setting)

Arguments

object
a GtkWindow

setting
TRUE to keep this window from appearing in the task bar

Details

Since 2.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]