gtkWindowSetStartupId {RGtk2}	R Documentation

gtkWindowSetStartupId

Description

Startup notification identifiers are used by desktop environment to track application startup, to provide user feedback and other features. This function changes the corresponding property on the underlying GdkWindow. Normally, startup identifier is managed automatically and you should only use this function in special cases like transferring focus from other processes. You should use this function before calling gtkWindowPresent or any equivalent function generating a window map event.

Usage

gtkWindowSetStartupId(object, startup.id)

Arguments

object
a GtkWindow

startup.id
a string with startup-notification identifier

Details

This function is only useful on X11, not with other GTK+ targets. Since 2.12

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]