gtkWindowSetTitle {RGtk2}	R Documentation

gtkWindowSetTitle

Description

Sets the title of the GtkWindow. The title of a window will be displayed in its title bar; on the X Window System, the title bar is rendered by the window manager, so exactly how the title appears to users may vary according to a user's exact configuration. The title should help a user distinguish this window from other windows they may have open. A good title might include the application name and current document filename, for example.

Usage

gtkWindowSetTitle(object, title)

Arguments

object
a GtkWindow

title
title of the window

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]