gtkWindowSetTransientFor {RGtk2}	R Documentation

gtkWindowSetTransientFor

Description

Dialog windows should be set transient for the main application window they were spawned from. This allows window managers to e.g. keep the dialog on top of the main window, or center the dialog over the main window. gtkDialogNewWithButtons and other convenience functions in GTK+ will sometimes call gtkWindowSetTransientFor on your behalf.

Usage

gtkWindowSetTransientFor(object, parent = NULL)

Arguments

object
a GtkWindow

parent
parent window, or NULL. [ allow-none ]

Details

Passing NULL for parent unsets the current transient window.

On Windows, this function puts the child window on top of the parent, much as the window manager would have done on X.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]