gtkWindowSetTypeHint {RGtk2}	R Documentation

gtkWindowSetTypeHint

Description

By setting the type hint for the window, you allow the window manager to decorate and handle the window in a way which is suitable to the function of the window in your application.

Usage

gtkWindowSetTypeHint(object, hint)

Arguments

object
a GtkWindow

hint
the window type

Details

This function should be called before the window becomes visible. gtkDialogNewWithButtons and other convenience functions in GTK+ will sometimes call gtkWindowSetTypeHint on your behalf.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]