gtkWindowSetWmclass {RGtk2}	R Documentation

gtkWindowSetWmclass

Description

Don't use this function. It sets the X Window System "class" and "name" hints for a window. According to the ICCCM, you should always set these to the same value for all windows in an application, and GTK+ sets them to that value by default, so calling this function is sort of pointless. However, you may want to call gtkWindowSetRole on each window in your application, for the benefit of the session manager. Setting the role allows the window manager to restore window positions when loading a saved session.

Usage

gtkWindowSetWmclass(object, wmclass.name, wmclass.class)

Arguments

object
a GtkWindow

wmclass.name
window name hint

wmclass.class
window class hint

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]