gtkObjectGetArgs {RGtk2}	R Documentation

Access properties of a Gtk object

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

These functions allow one to both read and set properties of a Gtk object by name and also in the usual S-like manner. gtkObjectGetArgs retrieves the values of one or more properties by specifying their name as a vector. gtkObjectGetArg retrieves the value of a single property and avoids having to worry about whether the result is a value or a list of values of length 1. gtkObjectSetArgs allows one to set one or more properties by name in the form gtkObjectSetArgs(obj, x=1, y="a"). [.GtkObject and [[<-.GtkObject provide S-like accessors.

Usage

gtkObjectGetArgs(obj, argNames)
gtkObjectGetArg(obj, argName)
gtkObjectSetArgs(obj, ..., .vals)

Arguments

obj
the Gtk object whose properties are to be accessed

argNames
a character vector giving the names of the properties to retrieve.

argName
a string (i.e. character vector of length 1) giving the name of the property whose value is to be retrieved.

...
for [.GtkObject, this is a character vector giving the names of the properties of interest. For gtkObjectSetArgs, this is a collection of name=value pairs where name is the name of the property to set and value is the value to which it is to be set.

.vals
a named list similar to ... in gtkObjectSetArgs

Note

THIS STUFF IS VERY OLD AND DEPRECATED (compatibility wrappers for RGtk 1)

Author(s)

Duncan Temple Lang <duncan@research.bell-labs.com>

[Package RGtk2 version 2.20.31 Index]