pango-Coverage-Maps {RGtk2}	R Documentation

Coverage Maps

Description

Unicode character range coverage storage

Methods and Functions

pangoCoverageNew()
pangoCoverageCopy(object)
pangoCoverageGet(object, index)
pangoCoverageMax(object, other)
pangoCoverageSet(object, index, level)
pangoCoverageToBytes(object)
pangoCoverageFromBytes(bytes)

Detailed Description

It is often necessary in Pango to determine if a particular font can represent a particular character, and also how well it can represent that character. The PangoCoverage is a data structure that is used to represent that information.

Structures

PangoCoverage

The PangoCoverage structure represents a map from Unicode characters to PangoCoverageLevel. It is an opaque structure with no public fields.

Enums and Flags

PangoCoverageLevel

Used to indicate how well a font can represent a particular Unicode character point for a particular script.

none

The character is not representable with the font.

fallback

The character is represented in a way that may be comprehensible but is not the correct graphical form. For instance, a Hangul character represented as a a sequence of Jamos, or a Latin transliteration of a Cyrillic word.

approximate

The character is represented as basically the correct graphical form, but with a stylistic variant inappropriate for the current script.

exact

The character is represented as the correct graphical form.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//pango/pango-Coverage-Maps.html

[Package RGtk2 version 2.20.31 Index]