pango-Scripts-and-Languages {RGtk2}	R Documentation

Scripts and Languages

Description

Identifying writing systems and languages

Methods and Functions

pangoScriptForUnichar(ch)
pangoScriptGetSampleLanguage(script)
pangoScriptIterNew(text, length)
pangoScriptIterGetRange(object)
pangoScriptIterNext(object)
pangoLanguageFromString(language)
pangoLanguageToString(object)
pangoLanguageMatches(object, range.list)
pangoLanguageIncludesScript(object, script)
pangoLanguageGetScripts(object)
pangoLanguageGetDefault()
pangoLanguageGetSampleString(object)

Detailed Description

The functions in this section are used to identify the writing system, or script of individual characters and of ranges within a larger text string.

Structures

PangoScriptIter

A PangoScriptIter is used to iterate through a string and identify ranges in different scripts.

PangoLanguage

The PangoLanguage structure is used to represent a language. PangoLanguage pointers can be efficiently copied and compared with each other.

Enums and Flags

PangoScript

The PangoScript enumeration identifies different writing systems. The values correspond to the names as defined in the Unicode standard. Note that new types may be added in the future. Applications should be ready to handle unknown values. This enumeration is interchangeable with GUnicodeScript. See Unicode Standard Annex #24: Script names (http://www.unicode.org/reports/tr24/).

invalid-code

a value never returned from pangoScriptForUnichar

common

a character used by multiple different scripts

inherited

a mark glyph that takes its script from the base glyph to which it is attached

arabic

Arabic

armenian

Armenian

bengali

Bengali

bopomofo

Bopomofo

cherokee

Cherokee

coptic

Coptic

cyrillic

Cyrillic

deseret

Deseret

devanagari

Devanagari

ethiopic

Ethiopic

georgian

Georgian

gothic

Gothic

greek

Greek

gujarati

Gujarati

gurmukhi

Gurmukhi

han

Han

hangul

Hangul

hebrew

Hebrew

hiragana

Hiragana

kannada

Kannada

katakana

Katakana

khmer

Khmer

lao

Lao

latin

Latin

malayalam

Malayalam

mongolian

Mongolian

myanmar

Myanmar

ogham

Ogham

old-italic

Old Italic

oriya

Oriya

runic

Runic

sinhala

Sinhala

syriac

Syriac

tamil

Tamil

telugu

Telugu

thaana

Thaana

thai

Thai

tibetan

Tibetan

canadian-aboriginal

Canadian Aboriginal

yi

Yi

tagalog

Tagalog

hanunoo

Hanunoo

buhid

Buhid

tagbanwa

Tagbanwa

braille

Braille

cypriot

Cypriot

limbu

Limbu

osmanya

Osmanya

shavian

Shavian

linear-b

Linear B

tai-le

Tai Le

ugaritic

Ugaritic

new-tai-lue

New Tai Lue. Since 1.10

buginese

Buginese. Since 1.10

glagolitic

Glagolitic. Since 1.10

tifinagh

Tifinagh. Since 1.10

syloti-nagri

Syloti Nagri. Since 1.10

old-persian

Old Persian. Since 1.10

kharoshthi

Kharoshthi. Since 1.10

unknown

an unassigned code point. Since 1.14

balinese

Balinese. Since 1.14

cuneiform

Cuneiform. Since 1.14

phoenician

Phoenician. Since 1.14

phags-pa

Phags-pa. Since 1.14

nko

N'Ko. Since 1.14

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//pango/pango-Scripts-and-Languages.html

[Package RGtk2 version 2.20.31 Index]