pango-Tab-Stops {RGtk2}	R Documentation

Tab Stops

Description

Structures for storing tab stops

Methods and Functions

pangoTabArrayNew(initial.size, positions.in.pixels)
pangoTabArrayNewWithPositions(size, positions.in.pixels, ...)
pangoTabArrayCopy(object)
pangoTabArrayGetSize(object)
pangoTabArrayResize(object, new.size)
pangoTabArraySetTab(object, tab.index, alignment, location)
pangoTabArrayGetTab(object, tab.index)
pangoTabArrayGetTabs(object)
pangoTabArrayGetPositionsInPixels(object)

Detailed Description

Functions in this section are used to deal with PangoTabArray objects that can be used to set tab stop positions in a PangoLayout.

Structures

PangoTabArray

A PangoTabArray struct contains a list of tab stops. Each tab stop has an alignment and a position.

Enums and Flags

PangoTabAlign

A PangoTabAlign specifies where a tab stop appears relative to the text.

left

the tab stop appears to the left of the text.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//pango/pango-Tab-Stops.html

[Package RGtk2 version 2.20.31 Index]