pango-Version-Checking {RGtk2}	R Documentation

Version Checking

Description

Tools for checking Pango version at compile- and run-time.

Methods and Functions

pangoVersion()
pangoVersionString()
pangoVersionCheck(required.major, required.minor, required.micro)

Detailed Description

The capital-letter functions defined here can be used to check the version of Pango at compile-time, and to encode Pango versions into integers. The functions can be used to check the version of the linked Pango library at run-time.

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//pango/pango-Version-Checking.html

[Package RGtk2 version 2.20.31 Index]