pangoAttrIteratorGet {RGtk2}	R Documentation

pangoAttrIteratorGet

Description

Find the current attribute of a particular type at the iterator location. When multiple attributes of the same type overlap, the attribute whose range starts closest to the current location is used.

Usage

pangoAttrIteratorGet(object, type)

Arguments

object
[PangoAttrIterator] a PangoAttrIterator

type
[PangoAttrType] the type of attribute to find.

Value

[PangoAttribute] the current attribute of the given type, or NULL if no attribute of that type applies to the current location.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]