pangoAttrIteratorGetFont {RGtk2}	R Documentation

pangoAttrIteratorGetFont

Description

Get the font and other attributes at the current iterator position.

Usage

pangoAttrIteratorGetFont(object)

Arguments

object
[PangoAttrIterator] a PangoAttrIterator

Value

A list containing the following elements:

desc
[PangoFontDescription] a PangoFontDescription to fill in with the current values. The family name in this structure will be set using pangoFontDescriptionSetFamilyStatic using values from an attribute in the PangoAttrList associated with the iterator, so if you plan to keep it around, you must call: pango_font_description_set_family (desc, pango_font_description_get_family (desc)).

language
[PangoLanguage] if non-NULL, location to store language tag for item, or NULL if none is found.

extra.attrs
[list] element type Pango.Attribute): (transfer full. element type Pango.Attribute): (transfer full.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]