pangoAttrIteratorRange {RGtk2}	R Documentation

pangoAttrIteratorRange

Description

Get the range of the current segment. Note that the stored return values are signed, not unsigned like the values in PangoAttribute. To deal with this API oversight, stored return values that wouldn't fit into a signed integer are clamped to G_MAXINT.

Usage

pangoAttrIteratorRange(object)

Arguments

object
[PangoAttrIterator] a PangoAttrIterator

Value

A list containing the following elements:

start
[integer] location to store the start of the range

end
[integer] location to store the end of the range

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]