pangoAttrListFilter {RGtk2}	R Documentation

pangoAttrListFilter

Description

Given a PangoAttrList and callback function, removes any elements of list for which func returns TRUE and inserts them into a new list.

Usage

pangoAttrListFilter(object, func, data)

Arguments

object
[PangoAttrList] a PangoAttrList

func
[PangoAttrFilterFunc] callback function; returns TRUE if an attribute should be filtered out.

data
[R object] Data to be passed to func

Details

Since 1.2

Value

[PangoAttrList] the new PangoAttrList or NULL if no attributes of the given types were found.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]