pangoAttrListInsert {RGtk2}	R Documentation

pangoAttrListInsert

Description

Insert the given attribute into the PangoAttrList. It will be inserted after all other attributes with a matching start.index.

Usage

pangoAttrListInsert(object, attr)

Arguments

object
[PangoAttrList] a PangoAttrList

attr
[PangoAttribute] the attribute to insert. Ownership of this value is assumed by the list.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]