pangoAttrListSplice {RGtk2}	R Documentation

pangoAttrListSplice

Description

This function opens up a hole in list, fills it in with attributes from the left, and then merges other on top of the hole.

Usage

pangoAttrListSplice(object, other, pos, len)

Arguments

object
[PangoAttrList] a PangoAttrList

other
[PangoAttrList] another PangoAttrList

pos
[integer] the position in list at which to insert other

len
[integer] the length of the spliced segment. (Note that this must be specified since the attributes in other may only be present at some subsection of this range)

Details

This operation is equivalent to stretching every attribute that applies at position pos in list by an amount len, and then calling pangoAttrListChange with a copy of each attribute in other in sequence (offset in position by pos).

This operation proves useful for, for instance, inserting a pre-edit string in the middle of an edit buffer.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]