pangoAttrShapeNew {RGtk2}	R Documentation

pangoAttrShapeNew

Description

Create a new shape attribute. A shape is used to impose a particular ink and logical rectangle on the result of shaping a particular glyph. This might be used, for instance, for embedding a picture or a widget inside a PangoLayout.

Usage

pangoAttrShapeNew(ink.rect, logical.rect)

Arguments

ink.rect
[PangoRectangle] ink rectangle to assign to each character

logical.rect
[PangoRectangle] logical rectangle to assign to each character

Value

[PangoAttribute] the newly allocated PangoAttribute,

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]