pangoAttrShapeNewWithData {RGtk2}	R Documentation

pangoAttrShapeNewWithData

Description

Like pangoAttrShapeNew, but a user data pointer is also provided; this pointer can be accessed when later rendering the glyph.

Usage

pangoAttrShapeNewWithData(ink.rect, logical.rect, data)

Arguments

ink.rect
[PangoRectangle] ink rectangle to assign to each character

logical.rect
[PangoRectangle] logical rectangle to assign to each character

data
[R object] user data pointer

Details

Since 1.8

Value

[PangoAttribute] the newly allocated PangoAttribute,

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]