pangoAttributeEqual {RGtk2}	R Documentation

pangoAttributeEqual

Description

Compare two attributes for equality. This compares only the actual value of the two attributes and not the ranges that the attributes apply to.

Usage

pangoAttributeEqual(object, attr2)

Arguments

object
[PangoAttribute] a PangoAttribute

attr2
[PangoAttribute] another PangoAttribute

Value

[logical] TRUE if the two attributes have the same value.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]