pangoBidiTypeForUnichar {RGtk2}	R Documentation

pangoBidiTypeForUnichar

Description

Determines the normative bidirectional character type of a character, as specified in the Unicode Character Database.

Usage

pangoBidiTypeForUnichar(ch)

Arguments

ch
[numeric] a Unicode character

Details

A simplified version of this function is available as pangoUnicharGetDirection(). Since 1.22

Value

[PangoBidiType] the bidirectional character type, as used in the Unicode bidirectional algorithm.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]