pangoBreak {RGtk2}	R Documentation

pangoBreak

Description

Determines possible line, word, and character breaks for a string of Unicode text with a single analysis. For most purposes you may want to use pangoGetLogAttrs.

Usage

pangoBreak(text, analysis)

Arguments

text
[character] the text to process

analysis
[PangoAnalysis] PangoAnalysis structure from pangoItemize

Value

A list containing the following elements:

attrs
[PangoLogAttr] a list to store character information in

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]