pangoCairoContextGetShapeRenderer {RGtk2}	R Documentation

pangoCairoContextGetShapeRenderer

Description

Sets callback function for context to use for rendering attributes of type PANGO_ATTR_SHAPE. See PangoCairoShapeRendererFunc for details.

Usage

pangoCairoContextGetShapeRenderer(object)

Arguments

object
[PangoContext] a PangoContext, from a pangocairo font map

Details

Retrieves callback function and associated user data for rendering attributes of type PANGO_ATTR_SHAPE as set by pangoCairoContextSetShapeRenderer, if any. Since 1.18

Value

A list containing the following elements:

retval
[PangoCairoShapeRendererFunc] the shape rendering callback previously set on the context, or NULL if no shape rendering callback have been set.

data
[R object] Pointer to R object to return user data

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]