pangoCairoCreateContext {RGtk2}	R Documentation

pangoCairoCreateContext

Description

Creates a context object set up to match the current transformation and target surface of the Cairo context. This context can then be used to create a layout using pangoLayoutNew.

Usage

pangoCairoCreateContext(cr)

Arguments

cr
[Cairo] a Cairo context

Details

This function is a convenience function that creates a context using the default font map, then updates it to cr. If you just need to create a layout for use with cr and do not need to access PangoContext directly, you can use pangoCairoCreateLayout instead. Since 1.22

Value

[PangoContext] the newly created PangoContext.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]