pangoCairoCreateLayout {RGtk2}	R Documentation

pangoCairoCreateLayout

Description

Creates a layout object set up to match the current transformation and target surface of the Cairo context. This layout can then be used for text measurement with functions like pangoLayoutGetSize or drawing with functions like pangoCairoShowLayout. If you change the transformation or target surface for cr, you need to call pangoCairoUpdateLayout

Usage

pangoCairoCreateLayout(cr)

Arguments

cr
[Cairo] a Cairo context

Details

This function is the most convenient way to use Cairo with Pango, however it is slightly inefficient since it creates a separate PangoContext object for each layout. This might matter in an application that was laying out large amounts of text. Since 1.10

Value

[PangoLayout] the newly created PangoLayout.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]