pangoCairoErrorUnderlinePath {RGtk2}	R Documentation

pangoCairoErrorUnderlinePath

Description

Add a squiggly line to the current path in the specified cairo context that approximately covers the given rectangle in the style of an underline used to indicate a spelling error. (The width of the underline is rounded to an integer number of up/down segments and the resulting rectangle is centered in the original rectangle)

Usage

pangoCairoErrorUnderlinePath(cr, x, y, width, height)

Arguments

cr
[Cairo] a Cairo context

x
[numeric] The X coordinate of one corner of the rectangle

y
[numeric] The Y coordinate of one corner of the rectangle

width
[numeric] Non-negative width of the rectangle

height
[numeric] Non-negative height of the rectangle

Details

Since 1.14

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]