pangoCairoFontMapGetDefault {RGtk2}	R Documentation

pangoCairoFontMapGetDefault

Description

Gets a default PangoCairoFontMap to use with Cairo.

Usage

pangoCairoFontMapGetDefault()

Details

Note that the type of the returned object will depend on the particular font backend Cairo was compiled to use; You generally should only use the PangoFontMap and PangoCairoFontMap interfaces on the returned object.

The default Cairo fontmap can be changed by using pangoCairoFontMapSetDefault. This can be used to change the Cairo font backend that the default fontmap uses for example. Since 1.10

Value

[PangoFontMap] the default Cairo fontmap for Pango.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]