pangoCairoFontMapNew {RGtk2}	R Documentation

pangoCairoFontMapNew

Description

Creates a new PangoCairoFontMap object; a fontmap is used to cache information about available fonts, and holds certain global parameters such as the resolution. In most cases, you can use pangoCairoFontMapGetDefault instead.

Usage

pangoCairoFontMapNew()

Details

Note that the type of the returned object will depend on the particular font backend Cairo was compiled to use; You generally should only use the PangoFontMap and PangoCairoFontMap interfaces on the returned object. Since 1.10

Value

[PangoFontMap] the newly allocated PangoFontMap,

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]