pangoCairoFontMapNewForFontType {RGtk2}	R Documentation

pangoCairoFontMapNewForFontType

Description

Creates a new PangoCairoFontMap object of the type suitable to be used with cairo font backend of type fonttype.

Usage

pangoCairoFontMapNewForFontType(fonttype)

Arguments

fonttype
[CairoFontType] desired CairoFontType

Details

In most cases one should simply use @pangoCairoFontMapNew, or in fact in most of those cases, just use @pangoCairoFontMapGetDefault. Since 1.18

Value

[PangoFontMap] or NULL if the requested cairo font backend is not supported / compiled in.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]