pangoCairoFontMapSetDefault {RGtk2}	R Documentation

pangoCairoFontMapSetDefault

Description

Sets a default PangoCairoFontMap to use with Cairo.

Usage

pangoCairoFontMapSetDefault(fontmap)

Arguments

fontmap
[PangoCairoFontMap] The new default font map, or NULL

Details

This can be used to change the Cairo font backend that the default fontmap uses for example. The old default font map is unreffed and the new font map referenced.

A value of NULL for fontmap will cause the current default font map to be released and a new default font map to be created on demand, using pangoCairoFontMapNew. Since 1.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]