pangoCairoGlyphStringPath {RGtk2}	R Documentation

pangoCairoGlyphStringPath

Description

Adds the glyphs in glyphs to the current path in the specified cairo context. The origin of the glyphs (the left edge of the baseline) will be at the current point of the cairo context.

Usage

pangoCairoGlyphStringPath(cr, font, glyphs)

Arguments

cr
[Cairo] a Cairo context

font
[PangoFont] a PangoFont from a PangoCairoFontMap

glyphs
[PangoGlyphString] a PangoGlyphString

Details

Since 1.10

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]