pangoCairoShowGlyphItem {RGtk2}	R Documentation

pangoCairoShowGlyphItem

Description

Draws the glyphs in glyph.item in the specified cairo context, embedding the text associated with the glyphs in the output if the output format supports it (PDF for example), otherwise it acts similar to pangoCairoShowGlyphString.

Usage

pangoCairoShowGlyphItem(cr, text, glyph.item)

Arguments

cr
[Cairo] a Cairo context

text
[char] the UTF-8 text that glyph.item refers to

glyph.item
[PangoGlyphItem] a PangoGlyphItem

Details

The origin of the glyphs (the left edge of the baseline) will be drawn at the current point of the cairo context.

Note that text is the start of the text for layout, which is then indexed by glyph_item->item->offset. Since 1.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]