pangoColorCopy {RGtk2}	R Documentation

pangoColorCopy

Description

Creates a copy of src, Primarily used by language bindings, not that useful otherwise (since colors can just be copied by assignment in C).

Usage

pangoColorCopy(object)

Arguments

object
[PangoColor] color to copy, may be NULL

Value

[PangoColor] the newly allocated PangoColor, or NULL if src was NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]