pangoColorParse {RGtk2}	R Documentation

pangoColorParse

Description

Fill in the fields of a color from a string specification. The string can either one of a large set of standard names. (Taken from the X11 ‘rgb.txt’ file), or it can be a hex value in the form '#rgb' '#rrggbb' '#rrrgggbbb' or '#rrrrggggbbbb' where 'r', 'g' and 'b' are hex digits of the red, green, and blue components of the color, respectively. (White in the four forms is '#fff' '#ffffff' '#fffffffff' and '#ffffffffffff')

Usage

pangoColorParse(spec)

Arguments

spec
[char] a string specifying the new color

Value

A list containing the following elements:

retval
[logical] TRUE if parsing of the specifier succeeded, otherwise false.

color
[PangoColor] a PangoColor structure in which to store the result, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]