pangoContextGetMetrics {RGtk2}	R Documentation

pangoContextGetMetrics

Description

Get overall metric information for a particular font description. Since the metrics may be substantially different for different scripts, a language tag can be provided to indicate that the metrics should be retrieved that correspond to the script(s) used by that language.

Usage

pangoContextGetMetrics(object, desc, language = NULL)

Arguments

object
[PangoContext] a PangoContext

desc
[PangoFontDescription] a PangoFontDescription structure. NULL means that the font description from the context will be used.

language
[PangoLanguage] language tag used to determine which script to get the metrics for. NULL means that the language tag from the context will be used. If no language tag is set on the context, metrics for the default language (as determined by pangoLanguageGetDefault) will be returned.

Details

The PangoFontDescription is interpreted in the same way as by pangoItemize, and the family name may be a comma separated list of figures. If characters from multiple of these families would be used to render the string, then the returned fonts would be a composite of the metrics for the fonts loaded for the individual families.

Value

[PangoFontMetrics] a PangoFontMetrics object.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]