pangoContextListFamilies {RGtk2}	R Documentation

pangoContextListFamilies

Description

List all families for a context.

Usage

pangoContextListFamilies(object)

Arguments

object
[PangoContext] a PangoContext

Value

A list containing the following elements:

families
[PangoFontFamily] location to store a pointer to a list of PangoFontFamily *. This list should be freed with gFree().

n.families
[integer] location to store the number of elements in descs

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]