pangoContextSetFontMap {RGtk2}	R Documentation

pangoContextSetFontMap

Description

Sets the font map to be searched when fonts are looked-up in this context. This is only for internal use by Pango backends, a PangoContext obtained via one of the recommended methods should already have a suitable font map.

Usage

pangoContextSetFontMap(object, font.map)

Arguments

object
[PangoContext] a PangoContext

font.map
[PangoFontMap] the PangoFontMap to set.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]