pangoContextSetGravityHint {RGtk2}	R Documentation

pangoContextSetGravityHint

Description

Sets the gravity hint for the context.

Usage

pangoContextSetGravityHint(object, hint)

Arguments

object
[PangoContext] a PangoContext

hint
[PangoGravityHint] the new gravity hint

Details

The gravity hint is used in laying vertical text out, and is only relevant if gravity of the context as returned by pangoContextGetGravity is set PANGO_GRAVITY_EAST or PANGO_GRAVITY_WEST. Since 1.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]