pangoContextSetMatrix {RGtk2}	R Documentation

pangoContextSetMatrix

Description

Sets the transformation matrix that will be applied when rendering with this context. Note that reported metrics are in the user space coordinates before the application of the matrix, not device-space coordinates after the application of the matrix. So, they don't scale with the matrix, though they may change slightly for different matrices, depending on how the text is fit to the pixel grid.

Usage

pangoContextSetMatrix(object, matrix)

Arguments

object
[PangoContext] a PangoContext

matrix
[PangoMatrix] a PangoMatrix, or NULL to unset any existing matrix. (No matrix set is the same as setting the identity matrix.)

Details

Since 1.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]