pangoCoverageMax {RGtk2}	R Documentation

pangoCoverageMax

Description

Set the coverage for each index in coverage to be the max (better) value of the current coverage for the index and the coverage for the corresponding index in other.

Usage

pangoCoverageMax(object, other)

Arguments

object
[PangoCoverage] a PangoCoverage

other
[PangoCoverage] another PangoCoverage

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]