pangoExtentsToPixels {RGtk2}	R Documentation

pangoExtentsToPixels

Description

Converts extents from Pango units to device units, dividing by the PANGO_SCALE factor and performing rounding.

Usage

pangoExtentsToPixels(inclusive, nearest)

Arguments

inclusive
[PangoRectangle] rectangle to round to pixels inclusively, or NULL.

nearest
[PangoRectangle] rectangle to round to nearest pixels, or NULL.

Details

The inclusive rectangle is converted by flooring the x/y coordinates and extending width/height, such that the final rectangle completely includes the original rectangle.

The nearest rectangle is converted by rounding the coordinates of the rectangle to the nearest device unit (pixel).

The rule to which argument to use is: if you want the resulting device-space rectangle to completely contain the original rectangle, pass it in as inclusive. If you want two touching-but-not-overlapping rectangles stay touching-but-not-overlapping after rounding to device units, pass them in as nearest. Since 1.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]