pangoFindBaseDir {RGtk2}	R Documentation

pangoFindBaseDir

Description

Searches a string the first character that has a strong direction, according to the Unicode bidirectional algorithm.

Usage

pangoFindBaseDir(text, length = -1)

Arguments

text
[character] the text to process

length
[integer] length of text in bytes (may be -1 if text is nul-terminated)

Details

Since 1.4

Value

[PangoDirection] The direction corresponding to the first strong character. If no such character is found, then PANGO_DIRECTION_NEUTRAL is returned.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]