pangoFindParagraphBoundary {RGtk2}	R Documentation

pangoFindParagraphBoundary

Description

Locates a paragraph boundary in text. A boundary is caused by delimiter characters, such as a newline, carriage return, carriage return-newline pair, or Unicode paragraph separator character. The index of the run of delimiters is returned in paragraph.delimiter.index. The index of the start of the paragraph (index after all delimiters) is stored in next.paragraph.start.

Usage

pangoFindParagraphBoundary(text, length = -1)

Arguments

text
[character] UTF-8 text

length
[integer] length of text in bytes, or -1 if nul-terminated

Details

If no delimiters are found, both paragraph.delimiter.index and next.paragraph.start are filled with the length of text (an index one off the end).

Value

A list containing the following elements:

paragraph.delimiter.index
[integer] return location for index of delimiter

next.paragraph.start
[integer] return location for start of next paragraph

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]