pangoFontDescriptionBetterMatch {RGtk2}	R Documentation

pangoFontDescriptionBetterMatch

Description

Determines if the style attributes of new.match are a closer match for desc than those of old.match are, or if old.match is NULL, determines if new.match is a match at all. Approximate matching is done for weight and style; other style attributes must match exactly. Style attributes are all attributes other than family and size-related attributes. Approximate matching for style considers PANGO_STYLE_OBLIQUE and PANGO_STYLE_ITALIC as matches, but not as good a match as when the styles are equal.

Usage

pangoFontDescriptionBetterMatch(object, old.match = NULL, new.match)

Arguments

object
[PangoFontDescription] a PangoFontDescription

old.match
[PangoFontDescription] a PangoFontDescription, or NULL

new.match
[PangoFontDescription] a PangoFontDescription

Details

Note that old.match must match desc.

Value

[logical] TRUE if new.match is a better match

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]