pangoFontDescriptionGetFamily {RGtk2}	R Documentation

pangoFontDescriptionGetFamily

Description

Gets the family name field of a font description. See pangoFontDescriptionSetFamily.

Usage

pangoFontDescriptionGetFamily(object)

Arguments

object
[PangoFontDescription] a PangoFontDescription.

Value

[char] the family name field for the font description, or NULL if not previously set. This has the same life-time as the font description itself and should not be freed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]