pangoFontDescriptionGetSetFields {RGtk2}	R Documentation

pangoFontDescriptionGetSetFields

Description

Determines which fields in a font description have been set.

Usage

pangoFontDescriptionGetSetFields(object)

Arguments

object
[PangoFontDescription] a PangoFontDescription

Value

[PangoFontMask] a bitmask with bits set corresponding to the fields in desc that have been set.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]