pangoFontDescriptionGetSizeIsAbsolute {RGtk2}	R Documentation

pangoFontDescriptionGetSizeIsAbsolute

Description

Determines whether the size of the font is in points (not absolute) or device units (absolute). See pangoFontDescriptionSetSize and pangoFontDescriptionSetAbsoluteSize.

Usage

pangoFontDescriptionGetSizeIsAbsolute(object)

Arguments

object
[PangoFontDescription] a PangoFontDescription

Details

Since 1.8

Value

[logical] whether the size for the font description is in points or device units. Use pangoFontDescriptionGetSetFields to find out if the size field of the font description was explicitly set or not.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]