pangoFontDescriptionMerge {RGtk2}	R Documentation

pangoFontDescriptionMerge

Description

Merges the fields that are set in desc.to.merge into the fields in desc. If replace.existing is FALSE, only fields in desc that are not already set are affected. If TRUE, then fields that are already set will be replaced as well.

Usage

pangoFontDescriptionMerge(object, desc.to.merge, replace.existing)

Arguments

object
[PangoFontDescription] a PangoFontDescription

desc.to.merge
[PangoFontDescription] the PangoFontDescription to merge from, or NULL

replace.existing
[logical] if TRUE, replace fields in desc with the corresponding values from desc.to.merge, even if they are already exist.

Details

If desc.to.merge is NULL, this function performs nothing.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]