pangoFontDescriptionSetAbsoluteSize {RGtk2}	R Documentation

pangoFontDescriptionSetAbsoluteSize

Description

Sets the size field of a font description, in device units. This is mutually exclusive with pangoFontDescriptionSetSize which sets the font size in points.

Usage

pangoFontDescriptionSetAbsoluteSize(object, size)

Arguments

object
[PangoFontDescription] a PangoFontDescription

size
[numeric] the new size, in Pango units. There are PANGO_SCALE Pango units in one device unit. For an output backend where a device unit is a pixel, a size value of 10 * PANGO_SCALE gives a 10 pixel font.

Details

Since 1.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]