pangoFontDescriptionSetFamily {RGtk2}	R Documentation

pangoFontDescriptionSetFamily

Description

Sets the family name field of a font description. The family name represents a family of related font styles, and will resolve to a particular PangoFontFamily. In some uses of PangoFontDescription, it is also possible to use a comma separated list of family names for this field.

Usage

pangoFontDescriptionSetFamily(object, family)

Arguments

object
[PangoFontDescription] a PangoFontDescription.

family
[char] a string representing the family name.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]