pangoFontDescriptionSetFamilyStatic {RGtk2}	R Documentation

pangoFontDescriptionSetFamilyStatic

Description

Like pangoFontDescriptionSetFamily, except that no copy of family is made. The caller must make sure that the string passed in stays around until desc has been freed or the name is set again. This function can be used if family is a static string such as a C string literal, or if desc is only needed temporarily.

Usage

pangoFontDescriptionSetFamilyStatic(object, family)

Arguments

object
[PangoFontDescription] a PangoFontDescription

family
[char] a string representing the family name.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]