pangoFontDescriptionSetGravity {RGtk2}	R Documentation

pangoFontDescriptionSetGravity

Description

Sets the gravity field of a font description. The gravity field specifies how the glyphs should be rotated. If gravity is PANGO_GRAVITY_AUTO, this actually unsets the gravity mask on the font description.

Usage

pangoFontDescriptionSetGravity(object, gravity)

Arguments

object
[PangoFontDescription] a PangoFontDescription

gravity
[PangoGravity] the gravity for the font description.

Details

This function is seldom useful to the user. Gravity should normally be set on a PangoContext. Since 1.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]