pangoFontDescriptionSetSize {RGtk2}	R Documentation

pangoFontDescriptionSetSize

Description

Sets the size field of a font description in fractional points. This is mutually exclusive with pangoFontDescriptionSetAbsoluteSize.

Usage

pangoFontDescriptionSetSize(object, size)

Arguments

object
[PangoFontDescription] a PangoFontDescription

size
[integer] the size of the font in points, scaled by PANGO_SCALE. (That is, a size value of 10 * PANGO_SCALE is a 10 point font. The conversion factor between points and device units depends on system configuration and the output device. For screen display, a logical DPI of 96 is common, in which case a 10 point font corresponds to a 10 * (96 / 72) = 13.3 pixel font. Use pangoFontDescriptionSetAbsoluteSize if you need a particular size in device units.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]