pangoFontDescriptionSetStyle {RGtk2}	R Documentation

pangoFontDescriptionSetStyle

Description

Sets the style field of a PangoFontDescription. The PangoStyle enumeration describes whether the font is slanted and the manner in which it is slanted; it can be either PANGO_STYLE_NORMAL, PANGO_STYLE_ITALIC, or PANGO_STYLE_OBLIQUE. Most fonts will either have a italic style or an oblique style, but not both, and font matching in Pango will match italic specifications with oblique fonts and vice-versa if an exact match is not found.

Usage

pangoFontDescriptionSetStyle(object, style)

Arguments

object
[PangoFontDescription] a PangoFontDescription

style
[PangoStyle] the style for the font description

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]