pangoFontDescriptionToString {RGtk2}	R Documentation

pangoFontDescriptionToString

Description

Creates a string representation of a font description. See pangoFontDescriptionFromString for a description of the format of the string representation. The family list in the string description will only have a terminating comma if the last word of the list is a valid style option.

Usage

pangoFontDescriptionToString(object)

Arguments

object
[PangoFontDescription] a PangoFontDescription

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]