pangoFontFaceDescribe {RGtk2}	R Documentation

pangoFontFaceDescribe

Description

Returns the family, style, variant, weight and stretch of a PangoFontFace. The size field of the resulting font description will be unset.

Usage

pangoFontFaceDescribe(object)

Arguments

object
[PangoFontFace] a PangoFontFace

Value

[PangoFontDescription] a newly-created PangoFontDescription structure holding the description of the face.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]