pangoFontFaceGetFaceName {RGtk2}	R Documentation

pangoFontFaceGetFaceName

Description

Gets a name representing the style of this face among the different faces in the PangoFontFamily for the face. This name is unique among all faces in the family and is suitable for displaying to users.

Usage

pangoFontFaceGetFaceName(object)

Arguments

object
[PangoFontFace] a PangoFontFace.

Value

[char] the face name for the face.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]