pangoFontFaceListSizes {RGtk2}	R Documentation

pangoFontFaceListSizes

Description

List the available sizes for a font. This is only applicable to bitmap fonts. For scalable fonts, stores NULL at the location pointed to by sizes and 0 at the location pointed to by n.sizes. The sizes returned are in Pango units and are sorted in ascending order.

Usage

pangoFontFaceListSizes(object)

Arguments

object
[PangoFontFace] a PangoFontFace.

Details

Since 1.4

Value

A list containing the following elements:

sizes
[integer] location to store a pointer to a list of int. This list should be freed with gFree().

n.sizes
[integer] location to store the number of elements in sizes

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]