pangoFontFamilyIsMonospace {RGtk2}	R Documentation

pangoFontFamilyIsMonospace

Description

A monospace font is a font designed for text display where the the characters form a regular grid. For Western languages this would mean that the advance width of all characters are the same, but this categorization also includes Asian fonts which include double-width characters: characters that occupy two grid cells. gUnicharIswide() returns a result that indicates whether a character is typically double-width in a monospace font.

Usage

pangoFontFamilyIsMonospace(object)

Arguments

object
[PangoFontFamily] a PangoFontFamily

Details

The best way to find out the grid-cell size is to call pangoFontMetricsGetApproximateDigitWidth, since the results of pangoFontMetricsGetApproximateCharWidth may be affected by double-width characters. Since 1.4

Value

[logical] TRUE if the family is monospace.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]