pangoFontFamilyListFaces {RGtk2}	R Documentation

pangoFontFamilyListFaces

Description

Lists the different font faces that make up family. The faces in a family share a common design, but differ in slant, weight, width and other aspects.

Usage

pangoFontFamilyListFaces(object)

Arguments

object
[PangoFontFamily] a PangoFontFamily

Value

A list containing the following elements:

faces
[PangoFontFace] location to store a list of pointers to PangoFontFace objects, or NULL. This list should be freed with gFree() when it is no longer needed.

n.faces
[integer] location to store number of elements in faces.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]