pangoFontGetFontMap {RGtk2}	R Documentation

pangoFontGetFontMap

Description

Gets the font map for which the font was created.

Usage

pangoFontGetFontMap(object)

Arguments

object
[PangoFont] a PangoFont, or NULL

Details

Note that the font maintains a weak reference to the font map, so if all references to font map are dropped, the font map will be finalized even if there are fonts created with the font map that are still alive. In that case this function will return NULL. It is the responsibility of the user to ensure that the font map is kept alive. In most uses this is not an issue as a PangoContext holds a reference to the font map. Since 1.10

Value

[PangoFontMap] the PangoFontMap for the font, or NULL if font is NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]