pangoFontGetGlyphExtents {RGtk2}	R Documentation

pangoFontGetGlyphExtents

Description

Gets the logical and ink extents of a glyph within a font. The coordinate system for each rectangle has its origin at the base line and horizontal origin of the character with increasing coordinates extending to the right and down. The functions pangoAscent(), pangoDescent(), pangoLbearing(), and pangoRbearing() can be used to convert from the extents rectangle to more traditional font metrics. The units of the rectangles are in 1/PANGO_SCALE of a device unit.

Usage

pangoFontGetGlyphExtents(object, glyph)

Arguments

object
[PangoFont] a PangoFont

glyph
[numeric] the glyph index

Details

If font is NULL, this function gracefully sets some sane values in the output variables and returns.

Value

A list containing the following elements:

ink.rect
[PangoRectangle] rectangle used to store the extents of the glyph as drawn or NULL to indicate that the result is not needed.

logical.rect
[PangoRectangle] rectangle used to store the logical extents of the glyph or NULL to indicate that the result is not needed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]