pangoFontGetMetrics {RGtk2}	R Documentation

pangoFontGetMetrics

Description

Gets overall metric information for a font. Since the metrics may be substantially different for different scripts, a language tag can be provided to indicate that the metrics should be retrieved that correspond to the script(s) used by that language.

Usage

pangoFontGetMetrics(object, language = NULL)

Arguments

object
[PangoFont] a PangoFont

language
[PangoLanguage] language tag used to determine which script to get the metrics for, or NULL to indicate to get the metrics for the entire font.

Details

If font is NULL, this function gracefully sets some sane values in the output variables and returns.

Value

[PangoFontMetrics] a PangoFontMetrics object.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]