pangoFontMetricsGetApproximateDigitWidth {RGtk2}	R Documentation

pangoFontMetricsGetApproximateDigitWidth

Description

Gets the approximate digit width for a font metrics structure. This is merely a representative value useful, for example, for determining the initial size for a window. Actual digits in text can be wider or narrower than this, though this value is generally somewhat more accurate than the result of pangoFontMetricsGetApproximateCharWidth for digits.

Usage

pangoFontMetricsGetApproximateDigitWidth(object)

Arguments

object
[PangoFontMetrics] a PangoFontMetrics structure

Value

[integer] the digit width, in Pango units.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]