pangoGetLogAttrs {RGtk2}	R Documentation

pangoGetLogAttrs

Description

Computes a PangoLogAttr for each character in text. The log.attrs array must have one PangoLogAttr for each position in text; if text contains N characters, it has N+1 positions, including the last position at the end of the text. text should be an entire paragraph; logical attributes can't be computed without context (for example you need to see spaces on either side of a word to know the word is a word).

Usage

pangoGetLogAttrs(text, level, language)

Arguments

text
[char] text to process

level
[integer] embedding level, or -1 if unknown

language
[PangoLanguage] language tag

Value

A list containing the following elements:

log.attrs
[PangoLogAttr] list with one PangoLogAttr per character in text, plus one extra, to be filled in

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]