pangoGetMirrorChar {RGtk2}	R Documentation

pangoGetMirrorChar

Description

If ch has the Unicode mirrored property and there is another Unicode character that typically has a glyph that is the mirror image of ch's glyph, puts that character in the address pointed to by mirrored.ch. WARNING: pango_get_mirror_char is deprecated and should not be used in newly-written code.

Usage

pangoGetMirrorChar(ch)

Arguments

ch
[numeric] a Unicode character

Details

Use gUnicharGetMirrorChar() instead; the docs for that function provide full details.

Value

A list containing the following elements:

retval
[logical] TRUE if ch has a mirrored character and mirrored.ch is filled in, FALSE otherwise

mirrored.ch
[numeric] location to store the mirrored character

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]