pangoGlyphItemApplyAttrs {RGtk2}	R Documentation

pangoGlyphItemApplyAttrs

Description

Splits a shaped item (PangoGlyphItem) into multiple items based on an attribute list. The idea is that if you have attributes that don't affect shaping, such as color or underline, to avoid affecting shaping, you filter them out (pangoAttrListFilter), apply the shaping process and then reapply them to the result using this function.

Usage

pangoGlyphItemApplyAttrs(glyph.item, text, list)

Arguments

glyph.item
[PangoGlyphItem] a shaped item

text
[char] text that list applies to

list
[PangoAttrList] a PangoAttrList

Details

All attributes that start or end inside a cluster are applied to that cluster; for instance, if half of a cluster is underlined and the other-half strikethrough, then the cluster will end up with both underline and strikethrough attributes. In these cases, it may happen that item->extra_attrs for some of the result items can have multiple attributes of the same type.

This function takes ownership of glyph.item; it will be reused as one of the elements in the list. Since 1.2

Value

[list] a list of glyph items resulting from splitting glyph.item.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]