pangoGlyphItemIterInitEnd {RGtk2}	R Documentation

pangoGlyphItemIterInitEnd

Description

Initializes a PangoGlyphItemIter structure to point to the last cluster in a glyph item. See PangoGlyphItemIter for details of cluster orders.

Usage

pangoGlyphItemIterInitEnd(object, glyph.item, text)

Arguments

object
[PangoGlyphItemIter] a PangoGlyphItemIter

glyph.item
[PangoGlyphItem] the glyph item to iterate over

text
[char] text corresponding to the glyph item

Details

Since 1.22

Value

[logical] FALSE if there are no clusters in the glyph item

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]