pangoGlyphItemIterNextCluster {RGtk2}	R Documentation

pangoGlyphItemIterNextCluster

Description

Advances the iterator to the next cluster in the glyph item. See PangoGlyphItemIter for details of cluster orders.

Usage

pangoGlyphItemIterNextCluster(object)

Arguments

object
[PangoGlyphItemIter] a PangoGlyphItemIter

Details

Since 1.22

Value

[logical] TRUE if the iterator was advanced, FALSE if we were already on the last cluster.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]