pangoGlyphItemLetterSpace {RGtk2}	R Documentation

pangoGlyphItemLetterSpace

Description

Adds spacing between the graphemes of glyph.item to give the effect of typographic letter spacing.

Usage

pangoGlyphItemLetterSpace(glyph.item, text, log.attrs)

Arguments

glyph.item
[PangoGlyphItem] a PangoGlyphItem

text
[char] text that glyph.item corresponds to (glyph_item->item->offset is an offset from the start of text)

log.attrs
[PangoLogAttr] logical attributes for the item (the first logical attribute refers to the position before the first character in the item)

Details

Since 1.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]