pangoGlyphItemSplit {RGtk2}	R Documentation

pangoGlyphItemSplit

Description

Modifies orig to cover only the text after split.index, and returns a new item that covers the text before split.index that used to be in orig. You can think of split.index as the length of the returned item. split.index may not be 0, and it may not be greater than or equal to the length of orig (that is, there must be at least one byte assigned to each item, you can't create a zero-length item).

Usage

pangoGlyphItemSplit(orig, text, split.index)

Arguments

orig
[PangoGlyphItem] a PangoItem

text
[char] text to which positions in orig apply

split.index
[integer] byte index of position to split item, relative to the start of the item

Details

This function is similar in function to pangoItemSplit (and uses it internally.) Since 1.2

Value

[PangoGlyphItem] the newly allocated item representing text before split.index,

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]