pangoGlyphStringGetLogicalWidths {RGtk2}	R Documentation

pangoGlyphStringGetLogicalWidths

Description

Given a PangoGlyphString resulting from pangoShape and the corresponding text, determine the screen width corresponding to each character. When multiple characters compose a single cluster, the width of the entire cluster is divided equally among the characters.

Usage

pangoGlyphStringGetLogicalWidths(object, text, embedding.level)

Arguments

object
[PangoGlyphString] a PangoGlyphString

text
[char] the text corresponding to the glyphs

embedding.level
[integer] the embedding level of the string

Details

See also pangoGlyphItemGetLogicalWidths.

Value

A list containing the following elements:

logical.widths
[integer] a list whose length is the number of characters in text (equal to g_utf8_strlen (text, length) unless text has NUL bytes) to be filled in with the resulting character widths.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]