pangoGlyphStringGetWidth {RGtk2}	R Documentation

pangoGlyphStringGetWidth

Description

Computes the logical width of the glyph string as can also be computed using pangoGlyphStringExtents. However, since this only computes the width, it's much faster. This is in fact only a convenience function that computes the sum of geometry.width for each glyph in the glyphs.

Usage

pangoGlyphStringGetWidth(object)

Arguments

object
[PangoGlyphString] a PangoGlyphString

Details

Since 1.14

Value

[integer] the logical width of the glyph string.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]