pangoGlyphStringIndexToX {RGtk2}	R Documentation

pangoGlyphStringIndexToX

Description

Converts from character position to x position. (X position is measured from the left edge of the run). Character positions are computed by dividing up each cluster into equal portions.

Usage

pangoGlyphStringIndexToX(object, text, analysis, index, trailing)

Arguments

object
[PangoGlyphString] the glyphs return from pangoShape

text
[char] the text for the run

analysis
[PangoAnalysis] the analysis information return from pangoItemize

index
[integer] the byte index within text

trailing
[logical] whether we should compute the result for the beginning (FALSE) or end (TRUE) of the character.

Value

A list containing the following elements:

x.pos
[integer] location to store result

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]