pangoGravityGetForScript {RGtk2}	R Documentation

pangoGravityGetForScript

Description

Based on the script, base gravity, and hint, returns actual gravity to use in laying out a single PangoItem.

Usage

pangoGravityGetForScript(script, base.gravity, hint)

Arguments

script
[PangoScript] PangoScript to query

base.gravity
[PangoGravity] base gravity of the paragraph

hint
[PangoGravityHint] orientation hint

Details

If base.gravity is PANGO_GRAVITY_AUTO, it is first replaced with the preferred gravity of script. To get the preferred gravity of a script, pass PANGO_GRAVITY_AUTO and PANGO_GRAVITY_HINT_STRONG in. Since 1.16

Value

[PangoGravity] resolved gravity suitable to use for a run of text with script.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]