pangoGravityGetForScriptAndWidth {RGtk2}	R Documentation

pangoGravityGetForScriptAndWidth

Description

Based on the script, East Asian width, base gravity, and hint, returns actual gravity to use in laying out a single character or PangoItem.

Usage

pangoGravityGetForScriptAndWidth(script, wide, base.gravity, 
    hint)

Arguments

script
[PangoScript] PangoScript to query

wide
[logical] TRUE for wide characters as returned by gUnicharIswide()

base.gravity
[PangoGravity] base gravity of the paragraph

hint
[PangoGravityHint] orientation hint

Details

This function is similar to pangoGravityGetForScript except that this function makes a distinction between narrow/half-width and wide/full-width characters also. Wide/full-width characters always stand upright, that is, they always take the base gravity, whereas narrow/full-width characters are always rotated in vertical context.

If base.gravity is PANGO_GRAVITY_AUTO, it is first replaced with the preferred gravity of script. Since 1.26

Value

[PangoGravity] resolved gravity suitable to use for a run of text with script and wide.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]