pangoItemize {RGtk2}	R Documentation

pangoItemize

Description

Breaks a piece of text into segments with consistent directional level and shaping engine. Each byte of text will be contained in exactly one of the items in the returned list; the generated list of items will be in logical order (the start offsets of the items are ascending).

Usage

pangoItemize(context, text, start.index, length, attrs, cached.iter = NULL)

Arguments

context
[PangoContext] a structure holding information that affects the itemization process.

text
[char] the text to itemize.

start.index
[integer] first byte in text to process

length
[integer] the number of bytes (not characters) to process after start.index. This must be >= 0.

attrs
[PangoAttrList] the set of attributes that apply to text.

cached.iter
[PangoAttrIterator] Cached attribute iterator, or NULL

Details

cached.iter should be an iterator over attrs currently positioned at a range before or containing start.index; cached.iter will be advanced to the range covering the position just after start.index + length. (i.e. if itemizing in a loop, just keep passing in the same cached.iter).

Value

[list] a list of PangoItem structures.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]