pangoItemizeWithBaseDir {RGtk2}	R Documentation

pangoItemizeWithBaseDir

Description

Like pangoItemize, but the base direction to use when computing bidirectional levels (see pangoContextSetBaseDir), is specified explicitly rather than gotten from the PangoContext.

Usage

pangoItemizeWithBaseDir(context, base.dir, text, start.index, 
    length, attrs, cached.iter = NULL)

Arguments

context
[PangoContext] a structure holding information that affects the itemization process.

base.dir
[PangoDirection] base direction to use for bidirectional processing

text
[char] the text to itemize.

start.index
[integer] first byte in text to process

length
[integer] the number of bytes (not characters) to process after start.index. This must be >= 0.

attrs
[PangoAttrList] the set of attributes that apply to text.

cached.iter
[PangoAttrIterator] Cached attribute iterator, or NULL

Details

Since 1.4

Value

[list] a list of PangoItem structures.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]