pangoLanguageFromString {RGtk2}	R Documentation

pangoLanguageFromString

Description

Take a RFC-3066 format language tag as a string and convert it to a PangoLanguage pointer that can be efficiently copied (copy the pointer) and compared with other language tags (compare the pointer.)

Usage

pangoLanguageFromString(language)

Arguments

language
[char] a string representing a language tag, or NULL

Details

This function first canonicalizes the string by converting it to lowercase, mapping '_' to '-', and stripping all characters other than letters and '-'.

Use pangoLanguageGetDefault if you want to get the PangoLanguage for the current locale of the process.

Value

[PangoLanguage] an opaque pointer to a PangoLanguage structure, or NULL if language was NULL. The returned pointer will be valid forever after, and should not be freed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]