pangoLanguageGetSampleString {RGtk2}	R Documentation

pangoLanguageGetSampleString

Description

Get a string that is representative of the characters needed to render a particular language.

Usage

pangoLanguageGetSampleString(object)

Arguments

object
[PangoLanguage] a PangoLanguage, or NULL

Details

The sample text may be a pangram, but is not necessarily. It is chosen to be demonstrative of normal text in the language, as well as exposing font feature requirements unique to the language. It is suitable for use as sample text in a font selection dialog.

If language is NULL, the default language as found by pangoLanguageGetDefault is used.

If Pango does not have a sample string for language, the classic "The quick brown fox..." is returned. This can be detected by comparing the returned pointer value to that returned for (non-existent) language code "xx". That is, compare to:

pangoLanguageFromString("xx")$getSampleString()

Value

[char] the sample string.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]