pangoLanguageGetScripts {RGtk2}	R Documentation

pangoLanguageGetScripts

Description

Determines the scripts used to to write language. If nothing is known about the language tag language, or if language is NULL, then NULL is returned. The list of scripts returned starts with the script that the language uses most and continues to the one it uses least.

Usage

pangoLanguageGetScripts(object)

Arguments

object
[PangoLanguage] a PangoLanguage, or NULL

Details

The value num.script points at will be set to the number of scripts in the returned list (or zero if NULL is returned).

Most languages use only one script for writing, but there are some that use two (Latin and Cyrillic for example), and a few use three (Japanese for example). Applications should not make any assumptions on the maximum number of scripts returned though, except that it is positive if the return value is not NULL, and it is a small number.

The pangoLanguageIncludesScript function uses this function internally. Since 1.22

Value

A list containing the following elements:

retval
[PangoScript] An list of PangoScript values, with the number of entries in the list stored in num.scripts, or NULL if Pango does not have any information about this particular language tag (also the case if language is NULL).

num.scripts
[integer] location to return number of scripts, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]