pangoLanguageIncludesScript {RGtk2}	R Documentation

pangoLanguageIncludesScript

Description

Determines if script is one of the scripts used to write language. The returned value is conservative; if nothing is known about the language tag language, TRUE will be returned, since, as far as Pango knows, script might be used to write language.

Usage

pangoLanguageIncludesScript(object, script)

Arguments

object
[PangoLanguage] a PangoLanguage, or NULL

script
[PangoScript] a PangoScript

Details

This routine is used in Pango's itemization process when determining if a supplied language tag is relevant to a particular section of text. It probably is not useful for applications in most circumstances.

This function uses pangoLanguageGetScripts internally. Since 1.4

Value

[logical] TRUE if script is one of the scripts used to write language or if nothing is known about language (including the case that language is NULL), FALSE otherwise.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]