pangoLanguageMatches {RGtk2}	R Documentation

pangoLanguageMatches

Description

Checks if a language tag matches one of the elements in a list of language ranges. A language tag is considered to match a range in the list if the range is '*', the range is exactly the tag, or the range is a prefix of the tag, and the character after it in the tag is '-'.

Usage

pangoLanguageMatches(object, range.list)

Arguments

object
[PangoLanguage] a language tag (see pangoLanguageFromString), NULL is allowed and matches nothing but '*'

range.list
[char] a list of language ranges, separated by ';', ':', ',', or space characters. Each element must either be '*', or a RFC 3066 language range canonicalized as by pangoLanguageFromString

Value

[logical] TRUE if a match was found.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]