pangoLayoutGetContext {RGtk2}	R Documentation

pangoLayoutGetContext

Description

Retrieves the PangoContext used for this layout.

Usage

pangoLayoutGetContext(object)

Arguments

object
[PangoLayout] a PangoLayout

Value

[PangoContext] the PangoContext for the layout. This does not have an additional refcount added, so if you want to keep a copy of this around, you must reference it yourself.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]