pangoLayoutGetCursorPos {RGtk2}	R Documentation

pangoLayoutGetCursorPos

Description

Given an index within a layout, determines the positions that of the strong and weak cursors if the insertion point is at that index. The position of each cursor is stored as a zero-width rectangle. The strong cursor location is the location where characters of the directionality equal to the base direction of the layout are inserted. The weak cursor location is the location where characters of the directionality opposite to the base direction of the layout are inserted.

Usage

pangoLayoutGetCursorPos(object, index)

Arguments

object
[PangoLayout] a PangoLayout

index
[integer] the byte index of the cursor

Value

A list containing the following elements:

strong.pos
[PangoRectangle] location to store the strong cursor position (may be NULL)

weak.pos
[PangoRectangle] location to store the weak cursor position (may be NULL)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]