pangoLayoutGetExtents {RGtk2}	R Documentation

pangoLayoutGetExtents

Description

Computes the logical and ink extents of layout. Logical extents are usually what you want for positioning things. Note that both extents may have non-zero x and y. You may want to use those to offset where you render the layout. Not doing that is a very typical bug that shows up as right-to-left layouts not being correctly positioned in a layout with a set width.

Usage

pangoLayoutGetExtents(object)

Arguments

object
[PangoLayout] a PangoLayout

Details

The extents are given in layout coordinates and in Pango units; layout coordinates begin at the top left corner of the layout.

Value

A list containing the following elements:

ink.rect
[PangoRectangle] rectangle used to store the extents of the layout as drawn or NULL to indicate that the result is not needed.

logical.rect
[PangoRectangle] rectangle used to store the logical extents of the layout or NULL to indicate that the result is not needed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]