pangoLayoutGetLineReadonly {RGtk2}	R Documentation

pangoLayoutGetLineReadonly

Description

Retrieves a particular line from a PangoLayout.

Usage

pangoLayoutGetLineReadonly(object, line)

Arguments

object
[PangoLayout] a PangoLayout

line
[integer] the index of a line, which must be between 0 and pango_layout_get_line_count(layout) - 1, inclusive.

Details

This is a faster alternative to pangoLayoutGetLine, but the user is not expected to modify the contents of the line (glyphs, glyph widths, etc.). Since 1.16

Value

[PangoLayoutLine] the requested PangoLayoutLine, or NULL if the index is out of range. This layout line can be ref'ed and retained, but will become invalid if changes are made to the PangoLayout. No changes should be made to the line.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]