pangoLayoutGetLines {RGtk2}	R Documentation

pangoLayoutGetLines

Description

Returns the lines of the layout as a list.

Usage

pangoLayoutGetLines(object)

Arguments

object
[PangoLayout] a PangoLayout

Details

Use the faster pangoLayoutGetLinesReadonly if you do not plan to modify the contents of the lines (glyphs, glyph widths, etc.).

Value

[list] element-type Pango.LayoutLine): (transfer none. element-type Pango.LayoutLine): (transfer none.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]