pangoLayoutGetTabs {RGtk2}	R Documentation

pangoLayoutGetTabs

Description

Gets the current PangoTabArray used by this layout. If no PangoTabArray has been set, then the default tabs are in use and NULL is returned. Default tabs are every 8 spaces.

Usage

pangoLayoutGetTabs(object)

Arguments

object
[PangoLayout] a PangoLayout

Value

[PangoTabArray] a copy of the tabs for this layout, or NULL.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]