pangoLayoutGetUnknownGlyphsCount {RGtk2}	R Documentation

pangoLayoutGetUnknownGlyphsCount

Description

Counts the number unknown glyphs in layout. That is, zero if glyphs for all characters in the layout text were found, or more than zero otherwise.

Usage

pangoLayoutGetUnknownGlyphsCount(object)

Arguments

object
[PangoLayout] a PangoLayout

Details

This function can be used to determine if there are any fonts available to render all characters in a certain string, or when used in combination with PANGO_ATTR_FALLBACK, to check if a certain font supports all the characters in the string. Since 1.16

Value

[integer] The number of unknown glyphs in layout.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]