pangoLayoutIndexToLineX {RGtk2}	R Documentation

pangoLayoutIndexToLineX

Description

Converts from byte index. within the layout to line and X position. (X position is measured from the left edge of the line)

Usage

pangoLayoutIndexToLineX(object, index., trailing)

Arguments

object
[PangoLayout] a PangoLayout

index.
[integer] the byte index of a grapheme within the layout.

trailing
[logical] an integer indicating the edge of the grapheme to retrieve the position of. If 0, the trailing edge of the grapheme, if > 0, the leading of the grapheme.

Value

A list containing the following elements:

line
[integer] location to store resulting line index. (which will between 0 and pango_layout_get_line_count(layout) - 1)

x.pos
[integer] location to store resulting position within line (PANGO_SCALE units per device unit)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]