pangoLayoutIndexToPos {RGtk2}	R Documentation

pangoLayoutIndexToPos

Description

Converts from an index within a PangoLayout to the onscreen position corresponding to the grapheme at that index, which is represented as rectangle. Note that pos->x is always the leading edge of the grapheme and pos->x + pos->width the trailing edge of the grapheme. If the directionality of the grapheme is right-to-left, then pos->width will be negative.

Usage

pangoLayoutIndexToPos(object, index, pos)

Arguments

object
[PangoLayout] a PangoLayout

index
[integer] byte index within layout

pos
[PangoRectangle] rectangle in which to store the position of the grapheme

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]