pangoLayoutIterGetIndex {RGtk2}	R Documentation

pangoLayoutIterGetIndex

Description

Gets the current byte index. Note that iterating forward by char moves in visual order, not logical order, so indexes may not be sequential. Also, the index may be equal to the length of the text in the layout, if on the NULL run (see pangoLayoutIterGetRun).

Usage

pangoLayoutIterGetIndex(object)

Arguments

object
[PangoLayoutIter] a PangoLayoutIter

Value

[integer] current byte index.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]