pangoLayoutIterGetLineYrange {RGtk2}	R Documentation

pangoLayoutIterGetLineYrange

Description

Divides the vertical space in the PangoLayout being iterated over between the lines in the layout, and returns the space belonging to the current line. A line's range includes the line's logical extents, plus half of the spacing above and below the line, if pangoLayoutSetSpacing has been called to set layout spacing. The Y positions are in layout coordinates (origin at top left of the entire layout).

Usage

pangoLayoutIterGetLineYrange(object)

Arguments

object
[PangoLayoutIter] a PangoLayoutIter

Value

A list containing the following elements:

y0
[integer] start of line

y1
[integer] end of line

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]