pangoLayoutIterGetRun {RGtk2}	R Documentation

pangoLayoutIterGetRun

Description

Gets the current run. When iterating by run, at the end of each line, there's a position with a NULL run, so this function can return NULL. The NULL run at the end of each line ensures that all lines have at least one run, even lines consisting of only a newline.

Usage

pangoLayoutIterGetRun(object)

Arguments

object
[PangoLayoutIter] a PangoLayoutIter

Details

Use the faster pangoLayoutIterGetRunReadonly if you do not plan to modify the contents of the run (glyphs, glyph widths, etc.).

Value

[PangoLayoutRun] the current run.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]