pangoLayoutLineGetPixelExtents {RGtk2}	R Documentation

pangoLayoutLineGetPixelExtents

Description

Computes the logical and ink extents of layout.line in device units. This function just calls pangoLayoutLineGetExtents followed by two pangoExtentsToPixels calls, rounding ink.rect and logical.rect such that the rounded rectangles fully contain the unrounded one (that is, passes them as first argument to pangoExtentsToPixels).

Usage

pangoLayoutLineGetPixelExtents(object)

Arguments

object
[PangoLayoutLine] a PangoLayoutLine

Value

A list containing the following elements:

ink.rect
[PangoRectangle] rectangle used to store the extents of the glyph string as drawn, or NULL

logical.rect
[PangoRectangle] rectangle used to store the logical extents of the glyph string, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]