pangoLayoutLineXToIndex {RGtk2}	R Documentation

pangoLayoutLineXToIndex

Description

Converts from x offset to the byte index of the corresponding character within the text of the layout. If x.pos is outside the line, index. and trailing will point to the very first or very last position in the line. This determination is based on the resolved direction of the paragraph; for example, if the resolved direction is right-to-left, then an X position to the right of the line (after it) results in 0 being stored in index. and trailing. An X position to the left of the line results in index. pointing to the (logical) last grapheme in the line and trailing being set to the number of characters in that grapheme. The reverse is true for a left-to-right line.

Usage

pangoLayoutLineXToIndex(object, x.pos)

Arguments

object
[PangoLayoutLine] a PangoLayoutLine

x.pos
[integer] the X offset (in Pango units) from the left edge of the line.

Value

A list containing the following elements:

retval
[logical] FALSE if x.pos was outside the line, TRUE if inside

index
[integer] location to store calculated byte index for the grapheme in which the user clicked.

trailing
[integer] location to store an integer indicating where in the grapheme the user clicked. It will either be zero, or the number of characters in the grapheme. 0 represents the leading edge of the grapheme.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]