pangoLayoutSetAutoDir {RGtk2}	R Documentation

pangoLayoutSetAutoDir

Description

Sets whether to calculate the bidirectional base direction for the layout according to the contents of the layout; when this flag is on (the default), then paragraphs in layout that begin with strong right-to-left characters (Arabic and Hebrew principally), will have right-to-left layout, paragraphs with letters from other scripts will have left-to-right layout. Paragraphs with only neutral characters get their direction from the surrounding paragraphs.

Usage

pangoLayoutSetAutoDir(object, auto.dir)

Arguments

object
[PangoLayout] a PangoLayout

auto.dir
[logical] if TRUE, compute the bidirectional base direction from the layout's contents.

Details

When FALSE, the choice between left-to-right and right-to-left layout is done according to the base direction of the layout's PangoContext. (See pangoContextSetBaseDir).

When the auto-computed direction of a paragraph differs from the base direction of the context, the interpretation of PANGO_ALIGN_LEFT and PANGO_ALIGN_RIGHT are swapped. Since 1.4

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]