pangoLayoutSetEllipsize {RGtk2}	R Documentation

pangoLayoutSetEllipsize

Description

Sets the type of ellipsization being performed for layout. Depending on the ellipsization mode ellipsize text is removed from the start, middle, or end of text so they fit within the width and height of layout set with pangoLayoutSetWidth and pangoLayoutSetHeight.

Usage

pangoLayoutSetEllipsize(object, ellipsize)

Arguments

object
[PangoLayout] a PangoLayout

ellipsize
[PangoEllipsizeMode] the new ellipsization mode for layout

Details

If the layout contains characters such as newlines that force it to be layed out in multiple paragraphs, then whether each paragraph is ellipsized separately or the entire layout is ellipsized as a whole depends on the set height of the layout. See pangoLayoutSetHeight for details. Since 1.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]