pangoLayoutSetJustify {RGtk2}	R Documentation

pangoLayoutSetJustify

Description

Sets whether each complete line should be stretched to fill the entire width of the layout. This stretching is typically done by adding whitespace, but for some scripts (such as Arabic), the justification may be done in more complex ways, like extending the characters.

Usage

pangoLayoutSetJustify(object, justify)

Arguments

object
[PangoLayout] a PangoLayout

justify
[logical] whether the lines in the layout should be justified.

Details

Note that this setting is not implemented and so is ignored in Pango older than 1.18.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]