pangoLayoutSetMarkupWithAccel {RGtk2}	R Documentation

pangoLayoutSetMarkupWithAccel

Description

Sets the layout text and attribute list from marked-up text (see markup format). Replaces the current text and attribute list.

Usage

pangoLayoutSetMarkupWithAccel(object, markup, accel.marker)

Arguments

object
[PangoLayout] a PangoLayout

markup
[char] marked-up text (see markup format)

accel.marker
[numeric] marker for accelerators in the text

Details

If accel.marker is nonzero, the given character will mark the character following it as an accelerator. For example, accel.marker might be an ampersand or underscore. All characters marked as an accelerator will receive a PANGO_UNDERLINE_LOW attribute, and the first character so marked will be returned in accel.char. Two accel.marker characters following each other produce a single literal accel.marker character.

Value

A list containing the following elements:

accel.char
[numeric] return location for first located accelerator, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]