pangoLayoutSetSingleParagraphMode {RGtk2}	R Documentation

pangoLayoutSetSingleParagraphMode

Description

If setting is TRUE, do not treat newlines and similar characters as paragraph separators; instead, keep all text in a single paragraph, and display a glyph for paragraph separator characters. Used when you want to allow editing of newlines on a single text line.

Usage

pangoLayoutSetSingleParagraphMode(object, setting)

Arguments

object
[PangoLayout] a PangoLayout

setting
[logical] new setting

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]