pangoLayoutSetText {RGtk2}	R Documentation

pangoLayoutSetText

Description

Sets the text of the layout.

Usage

pangoLayoutSetText(object, text, length = -1)

Arguments

object
[PangoLayout] a PangoLayout

text
[char] a valid UTF-8 string

length
[integer] maximum length of text, in bytes. -1 indicates that the string is and the length should be calculated. The text will also be truncated on encountering a nul-termination even when length is positive.

Details

Note that if you have used pangoLayoutSetMarkup or pangoLayoutSetMarkupWithAccel on layout before, you may want to call pangoLayoutSetAttributes to clear the attributes set on the layout from the markup as this function does not clear attributes.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]