pangoLayoutSetWidth {RGtk2}	R Documentation

pangoLayoutSetWidth

Description

Sets the width to which the lines of the PangoLayout should wrap or ellipsized. The default value is -1: no width set.

Usage

pangoLayoutSetWidth(object, width)

Arguments

object
[PangoLayout] a PangoLayout.

width
[integer] the desired width in Pango units, or -1 to indicate that no wrapping or ellipsization should be performed.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]