pangoLayoutSetWrap {RGtk2}	R Documentation

pangoLayoutSetWrap

Description

Sets the wrap mode; the wrap mode only has effect if a width is set on the layout with pangoLayoutSetWidth. To turn off wrapping, set the width to -1.

Usage

pangoLayoutSetWrap(object, wrap)

Arguments

object
[PangoLayout] a PangoLayout

wrap
[PangoWrapMode] the wrap mode

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]