pangoMatrixRotate {RGtk2}	R Documentation

pangoMatrixRotate

Description

Changes the transformation represented by matrix to be the transformation given by first rotating by degrees degrees counter-clockwise then applying the original transformation.

Usage

pangoMatrixRotate(object, degrees)

Arguments

object
[PangoMatrix] a PangoMatrix

degrees
[numeric] degrees to rotate counter-clockwise

Details

Since 1.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]