pangoMatrixScale {RGtk2}	R Documentation

pangoMatrixScale

Description

Changes the transformation represented by matrix to be the transformation given by first scaling by sx in the X direction and sy in the Y direction then applying the original transformation.

Usage

pangoMatrixScale(object, scale.x, scale.y)

Arguments

object
[PangoMatrix] a PangoMatrix

scale.x
[numeric] amount to scale by in X direction

scale.y
[numeric] amount to scale by in Y direction

Details

Since 1.6

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]