pangoMatrixTransformDistance {RGtk2}	R Documentation

pangoMatrixTransformDistance

Description

Transforms the distance vector (dx,dy) by matrix. This is similar to pangoMatrixTransformPoint except that the translation components of the transformation are ignored. The calculation of the returned vector is as follows:

Usage

pangoMatrixTransformDistance(object, dx, dy)

Arguments

object
[PangoMatrix] a PangoMatrix, or NULL

dx
[numeric] in/out X component of a distance vector

dy
[numeric] yn/out Y component of a distance vector

Details

dx2 = dx1 * xx + dy1 * xy;
dy2 = dx1 * yx + dy1 * yy;

Affine transformations are position invariant, so the same vector always transforms to the same vector. If (x1,y1) transforms to (x2,y2) then (x1+dx1,y1+dy1) will transform to (x1+dx2,y1+dy2) for all values of x1 and x2. Since 1.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]