pangoMatrixTransformPixelRectangle {RGtk2}	R Documentation

pangoMatrixTransformPixelRectangle

Description

First transforms the rect using matrix, then calculates the bounding box of the transformed rectangle. The rectangle should be in device units (pixels).

Usage

pangoMatrixTransformPixelRectangle(object, rect)

Arguments

object
[PangoMatrix] a PangoMatrix, or NULL

rect
[PangoRectangle] in/out bounding box in device units, or NULL

Details

This function is useful for example when you want to draw a rotated PangoLayout to an image buffer, and want to know how large the image should be and how much you should shift the layout when rendering.

For better accuracy, you should use pangoMatrixTransformRectangle on original rectangle in Pango units and convert to pixels afterward using pangoExtentsToPixels's first argument. Since 1.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]