pangoMatrixTransformRectangle {RGtk2}	R Documentation

pangoMatrixTransformRectangle

Description

First transforms rect using matrix, then calculates the bounding box of the transformed rectangle. The rectangle should be in Pango units.

Usage

pangoMatrixTransformRectangle(object, rect)

Arguments

object
[PangoMatrix] a PangoMatrix, or NULL

rect
[PangoRectangle] in/out bounding box in Pango units, or NULL

Details

This function is useful for example when you want to draw a rotated PangoLayout to an image buffer, and want to know how large the image should be and how much you should shift the layout when rendering.

If you have a rectangle in device units (pixels), use pangoMatrixTransformPixelRectangle.

If you have the rectangle in Pango units and want to convert to transformed pixel bounding box, it is more accurate to transform it first (using this function) and pass the result to pangoExtentsToPixels, first argument, for an inclusive rounded rectangle. However, there are valid reasons that you may want to convert to pixels first and then transform, for example when the transformed coordinates may overflow in Pango units (large matrix translation for example). Since 1.16

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]