pangoParseMarkup {RGtk2}	R Documentation

pangoParseMarkup

Description

Parses marked-up text (see markup format) to create a plain-text string and an attribute list.

Usage

pangoParseMarkup(markup.text, accel.marker, .errwarn = TRUE)

Arguments

markup.text
[char] markup to parse (see markup format)

accel.marker
[numeric] character that precedes an accelerator, or 0 for none

.errwarn
Whether to issue a warning on error or fail silently

Details

If accel.marker is nonzero, the given character will mark the character following it as an accelerator. For example, accel.marker might be an ampersand or underscore. All characters marked as an accelerator will receive a PANGO_UNDERLINE_LOW attribute, and the first character so marked will be returned in accel.char. Two accel.marker characters following each other produce a single literal accel.marker character.

If any error happens, none of the output arguments are touched except for error.

Value

A list containing the following elements:

retval
[logical] FALSE if error is set, otherwise TRUE

attr.list
[PangoAttrList] return location for a PangoAttrList, or NULL

text
[char] return location for text with tags stripped, or NULL

accel.char
[numeric] return location for accelerator char, or NULL

error
[GError] return location for errors, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]