pangoRendererDrawGlyph {RGtk2}	R Documentation

pangoRendererDrawGlyph

Description

Draws a single glyph with coordinates in device space.

Usage

pangoRendererDrawGlyph(object, font, glyph, x, y)

Arguments

object
[PangoRenderer] a PangoRenderer

font
[PangoFont] a PangoFont

glyph
[numeric] the glyph index of a single glyph

x
[numeric] X coordinate of left edge of baseline of glyph

y
[numeric] Y coordinate of left edge of baseline of glyph

Details

Since 1.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]