pangoRendererDrawGlyphItem {RGtk2}	R Documentation

pangoRendererDrawGlyphItem

Description

Draws the glyphs in glyph.item with the specified PangoRenderer, embedding the text associated with the glyphs in the output if the output format supports it (PDF for example).

Usage

pangoRendererDrawGlyphItem(object, text, glyph.item, x, y)

Arguments

object
[PangoRenderer] a PangoRenderer

text
[char] the UTF-8 text that glyph.item refers to, or NULL

glyph.item
[PangoGlyphItem] a PangoGlyphItem

x
[integer] X position of left edge of baseline, in user space coordinates in Pango units.

y
[integer] Y position of left edge of baseline, in user space coordinates in Pango units.

Details

Note that text is the start of the text for layout, which is then indexed by glyph_item->item->offset.

If text is NULL, this simply calls pangoRendererDrawGlyphs.

The default implementation of this method simply falls back to pangoRendererDrawGlyphs. Since 1.22

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]