pangoRendererDrawGlyphs {RGtk2}	R Documentation

pangoRendererDrawGlyphs

Description

Draws the glyphs in glyphs with the specified PangoRenderer.

Usage

pangoRendererDrawGlyphs(object, font, glyphs, x, y)

Arguments

object
[PangoRenderer] a PangoRenderer

font
[PangoFont] a PangoFont

glyphs
[PangoGlyphString] a PangoGlyphString

x
[integer] X position of left edge of baseline, in user space coordinates in Pango units.

y
[integer] Y position of left edge of baseline, in user space coordinates in Pango units.

Details

Since 1.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]