pangoRendererDrawLayout {RGtk2}	R Documentation

pangoRendererDrawLayout

Description

Draws layout with the specified PangoRenderer.

Usage

pangoRendererDrawLayout(object, layout, x, y)

Arguments

object
[PangoRenderer] a PangoRenderer

layout
[PangoLayout] a PangoLayout

x
[integer] X position of left edge of baseline, in user space coordinates in Pango units.

y
[integer] Y position of left edge of baseline, in user space coordinates in Pango units.

Details

Since 1.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]