pangoRendererDrawRectangle {RGtk2}	R Documentation

pangoRendererDrawRectangle

Description

Draws an axis-aligned rectangle in user space coordinates with the specified PangoRenderer.

Usage

pangoRendererDrawRectangle(object, part, x, y, width, height)

Arguments

object
[PangoRenderer] a PangoRenderer

part
[PangoRenderPart] type of object this rectangle is part of

x
[integer] X position at which to draw rectangle, in user space coordinates in Pango units

y
[integer] Y position at which to draw rectangle, in user space coordinates in Pango units

width
[integer] width of rectangle in Pango units in user space coordinates

height
[integer] height of rectangle in Pango units in user space coordinates

Details

This should be called while renderer is already active. Use pangoRendererActivate to activate a renderer. Since 1.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]