pangoRendererGetLayout {RGtk2}	R Documentation

pangoRendererGetLayout

Description

Gets the layout currently being rendered using renderer. Calling this function only makes sense from inside a subclass's methods, like in its draw_shape() for example.

Usage

pangoRendererGetLayout(renderer)

Arguments

renderer
[PangoRenderer] a PangoRenderer

Details

The returned layout should not be modified while still being rendered. Since 1.20

Value

[PangoLayout] the layout, or NULL if no layout is being rendered using renderer at this time.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]