pangoRendererPartChanged {RGtk2}	R Documentation

pangoRendererPartChanged

Description

Informs Pango that the way that the rendering is done for part has changed in a way that would prevent multiple pieces being joined together into one drawing call. For instance, if a subclass of PangoRenderer was to add a stipple option for drawing underlines, it needs to call

Usage

pangoRendererPartChanged(object, part)

Arguments

object
[PangoRenderer] a PangoRenderer

part
[PangoRenderPart] the part for which rendering has changed.

Details

pango_renderer_part_changed (render, PANGO_RENDER_PART_UNDERLINE);

When the stipple changes or underlines with different stipples might be joined together. Pango automatically calls this for changes to colors. (See pangoRendererSetColor) Since 1.8

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]