pangoReorderItems {RGtk2}	R Documentation

pangoReorderItems

Description

From a list of items in logical order and the associated directional levels, produce a list in visual order. The original list is unmodified.

Usage

pangoReorderItems(logical.items)

Arguments

logical.items
[list] a list of PangoItem in logical order.

Value

[list] a list of PangoItem structures in visual order. (Please open a bug if you use this function. It is not a particularly convenient interface, and the code is duplicated elsewhere in Pango for that reason.)

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]