pangoScriptForUnichar {RGtk2}	R Documentation

pangoScriptForUnichar

Description

Looks up the PangoScript for a particular character (as defined by Unicode Standard Annex 24). No check is made for ch being a valid Unicode character; if you pass in invalid character, the result is undefined.

Usage

pangoScriptForUnichar(ch)

Arguments

ch
[numeric] a Unicode character

Details

As of Pango 1.18, this function simply returns the return value of gUnicharGetScript(). Since 1.4

Value

[PangoScript] the PangoScript for the character.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]