pangoScriptIterGetRange {RGtk2}	R Documentation

pangoScriptIterGetRange

Description

Gets information about the range to which iter currently points. The range is the set of locations p where *start <= p < *end. (That is, it doesn't include the character stored at *end)

Usage

pangoScriptIterGetRange(object)

Arguments

object
[PangoScriptIter] a PangoScriptIter

Details

Since 1.4

Value

A list containing the following elements:

start
[char] location to store start position of the range, or NULL

end
[char] location to store end position of the range, or NULL

script
[PangoScript] location to store script for range, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]