pangoScriptIterNew {RGtk2}	R Documentation

pangoScriptIterNew

Description

Create a new PangoScriptIter, used to break a string of Unicode into runs by text. No copy is made of text,

Usage

pangoScriptIterNew(text, length)

Arguments

text
[char] a UTF-8 string

length
[integer] length of text, or -1 if text is nul-terminated.

Details

Since 1.4

Value

[PangoScriptIter] the new script iterator, initialized to point at the first range in the text, If the string is empty, it will point at an empty range.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]