pangoShape {RGtk2}	R Documentation

pangoShape

Description

Given a segment of text and the corresponding PangoAnalysis structure returned from pangoItemize, convert the characters into glyphs. You may also pass in only a substring of the item from pangoItemize.

Usage

pangoShape(text, analysis, glyphs)

Arguments

text
[character] the text to process

analysis
[PangoAnalysis] PangoAnalysis structure from pangoItemize

glyphs
[PangoGlyphString] glyph string in which to store results

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]