pangoTabArrayNewWithPositions {RGtk2}	R Documentation

pangoTabArrayNewWithPositions

Description

This is a convenience function that creates a PangoTabArray and allows you to specify the alignment and position of each tab stop. You must provide an alignment and position for size tab stops.

Usage

pangoTabArrayNewWithPositions(size, positions.in.pixels, ...)

Arguments

size
[integer] number of tab stops in the list

positions.in.pixels
[logical] whether positions are in pixel units

...
additional alignment/position pairs

Value

[PangoTabArray] the newly allocated PangoTabArray,

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]