pangoVersion {RGtk2}	R Documentation

pangoVersion

Description

This is similar to the function PANGO_VERSION except that it returns the encoded version of Pango available at run-time, as opposed to the version available at compile-time.

Usage

pangoVersion()

Details

A version number can be encoded into an integer using pangoVersionEncode(). Since 1.16

Value

[integer] The encoded version of Pango library available at run time.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]