pangoVersionCheck {RGtk2}	R Documentation

pangoVersionCheck

Description

Checks that the Pango library in use is compatible with the given version. Generally you would pass in the constants PANGO_VERSION_MAJOR, PANGO_VERSION_MINOR, PANGO_VERSION_MICRO as the three arguments to this function; that produces a check that the library in use at run-time is compatible with the version of Pango the application or module was compiled against.

Usage

pangoVersionCheck(required.major, required.minor, required.micro)

Arguments

required.major
[integer] the required major version.

required.minor
[integer] the required minor version.

required.micro
[integer] the required major version.

Details

Compatibility is defined by two things: first the version of the running library is newer than the version required.major.required.minor.required.micro. Second the running library must be binary compatible with the version required.major.required.minor.required.micro (same major version.)

For compile-time version checking use pangoVersionCheck(). Since 1.16

Value

[char] NULL if the Pango library is compatible with the given version, or a string describing the version mismatch.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]