gtkAddTimeout {RGtk2}	R Documentation

Control a periodic/timer function

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

These functions allow one to register and remove a function that is called periodically at a fixed interval.

Usage

gtkAddTimeout(interval, f, data=NULL)
gtkRemoveTimeout(id)

Arguments

interval
the number of milliseconds to wait before invoking the call.

f
the function should take one or zero arguments depending on whether the argument data is given. The function should return a logical value. If it returns FALSE, the timer is removed. If it returns TRUE, the timer is re-registered and will be called after interval milliseconds.

data
a value, which if specified, will be passed to the function f when it is invoked. This can be used to parameterize the function to have different functions. The same effect can be obtained using closures.

id
the object identifying the timer function in Gtk that was returned by a call to gtkAddTimeout.

Value

gtkAddTimeout returns an object of class "GtkTimeoutId". This is an integer giving the identifier returned by the low-level Gtk interface.

Note

THIS STUFF IS VERY OLD AND DEPRECATED (compatibility wrappers for RGtk 1)

This does not currently work when running directly in R. However, when running R inside another application such as GGobi, Gnumeric, etc. it and other event-loop facilities will work.

If one wanted to reset the timeout but with a different interval, one should register a new timeout within the function and return FALSE.

Author(s)

Duncan Temple Lang <duncan@research.bell-labs.com>

See Also

gtkAddIdle gtkRemoveIdle gtkAddCallback

[Package RGtk2 version 2.20.31 Index]