transparent-type {RGtk2}	R Documentation

Transparent types in RGtk2

Description

A transparent type in RGtk2 is a non-primitive type passed between the user and the API as an ordinary R object (usually a list with a defined structure).

Details

The RGtk2 documentation defines the public structure of every object. Some of these have been tagged as being transparent, indicating that the R programmer need not obtain an external object but rather simply create the analagous structure in R. Transparent types are usually simple types that would be created inline in C code for convenience, instead of invoking a function with a large number of arguments. RGtk2 emulates this in R.

Usually these structures are constructed as R lists, with optionally named elements. The lists elements are matched up to structure fields according to the same logic as function calls to function definitions (see match.call).

Author(s)

Michael Lawrence

See Also

GParamSpec GtkFileFilterInfo GtkTargetEntry AtkAttribute

GtkSettingsValue GClosure GType

GtkStockItem GtkItemFactoryEntry GtkAllocation GdkAtom GTimeVal

PangoRectangle GdkRectangle AtkAttributeSet GdkRgbCmap GdkKeymapKey GdkGCValues GdkGeometry

GdkPoint GdkSegment GdkColor GdkNativeWindow GError GdkWindowAttr GdkTrapezoid

GtkActionEntry GtkToggleActionEntry GtkRadioActionEntry CairoPath CairoGlyph

CairoPathData AtkTextRectangle AtkTextRange GdkSpan GdkTimeCoord

[Package RGtk2 version 2.20.31 Index]