gtkObjectGetClasses {RGtk2}	R Documentation

Dynamically Computes class information

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

These functions dynamically compute information about the class hierarchy for a given Gtk object or a type. The class hierarchy is defined by the C-level Gtk code and is “reflected” to S via these functions. Most S-level Gtk objects will be created in the RGtk package with the appropriate class. In cases where this is not true, gtkObjectGetClass can be used to compute and assign the appropriate class so that the regular S interface for that object will work.

Usage

gtkObjectGetClasses(w, check=TRUE)
gtkObjectGetTypeName(w)
gtkTypeGetClasses(type)

Arguments

w
the Gtk object whose class information is to be computed.

type
an S object identifying the Gtk type of interest. This can be either a character string giving the name of the type or an object of class GtkType

check
a logical value that can be used to bypass a check that the argument w is of class GtkObject. If the object has not class information but the caller knows that it is a pointer to a Gtk object, then to compute the class conveniently, check can be passed as FALSE.

Details

This uses C code to query the internal Gtk class hierarchy. The class information is not (only) S class information, but mirrors the real Gtk class hieararchy.

Value

A character vector. gtkObjectGetTypeName returns the name of the class/type of the given object. gtkObjectGetClasses and gtkTypeGetClasses return the names of the object class/type and its successive parent classes, up to GtkObject.

Note

THIS STUFF IS VERY OLD AND DEPRECATED (compatibility wrappers for RGtk 1)

Type information about a particular class is not available until that class has been initialized. One can call the C routine gtk_<type-name>_get_type to initialize it or, more commonly and completely, create an instance of that class. So .C("gtk_button_get_type", PACKAGE = "RGtk") or gtkButton() should preceed gtkTypeGetClasses("GtkButton")

Author(s)

Duncan Temple Lang <duncan@research.bell-labs.com>

References

Information on the package is available from http://www.omegahat.org/RGtk.

Information on Gtk is available from http://www.gtk.org.

See Also

gtkObjectGetSignals gtkTypeGetSignals

Examples

  if (gtkInit()) {
    b <- gtkButton()
    class(b)

     # Should be true
    class(b) == gtkObjectGetClasses(b)
  }

[Package RGtk2 version 2.20.31 Index]