virtuals {RGtk2}	R Documentation

Virtual Methods

Description

This function is a wrapper for a virtual method in a GObject class structure. The wrapper is normally only useful for calling virtuals in a parent class inside implementations of new GObject classes.

Author(s)

Michael Lawrence

See Also

gClass

[Package RGtk2 version 2.20.31 Index]