gdkGetWindowSize {RGtk2}	R Documentation

Obtain dimensions of a Gtk window

Description

DEPRECATED!! (compatibility wrappers for RGtk 1!)

Returns the width and height of a Gtk window.

Usage

gdkGetWindowSize(w)

Arguments

w
the object of class "GdkWindow" whose dimensions are to be queried.

Value

An integer vector with two elements:

width
the number of pixels the window spans in the horizontal dimension

height
the number of pixels the window spans in the vertical dimension

Note

THIS STUFF IS VERY OLD AND DEPRECATED (compatibility wrappers for RGtk 1)

Author(s)

Duncan Temple Lang <duncan@research.bell-labs.com>

See Also

gtkWidgetSetUsize

[Package RGtk2 version 2.20.31 Index]