GtkLabel {RGtk2}	R Documentation

GtkLabel

Description

A widget that displays a small to medium amount of text

Methods and Functions

gtkLabelNew(str = NULL, show = TRUE)
gtkLabelSetText(object, str)
gtkLabelSetAttributes(object, attrs)
gtkLabelSetMarkup(object, str)
gtkLabelSetMarkupWithMnemonic(object, str)
gtkLabelSetPattern(object, pattern)
gtkLabelSetJustify(object, jtype)
gtkLabelSetEllipsize(object, mode)
gtkLabelSetWidthChars(object, n.chars)
gtkLabelSetMaxWidthChars(object, n.chars)
gtkLabelGet(object)
gtkLabelParseUline(object, string)
gtkLabelSetLineWrap(object, wrap)
gtkLabelSetLineWrapMode(object, wrap.mode)
gtkLabelGetLayoutOffsets(object)
gtkLabelGetMnemonicKeyval(object)
gtkLabelGetSelectable(object)
gtkLabelGetText(object)
gtkLabelNewWithMnemonic(str = NULL, show = TRUE)
gtkLabelSelectRegion(object, start.offset, end.offset)
gtkLabelSetMnemonicWidget(object, widget)
gtkLabelSetSelectable(object, setting)
gtkLabelSetTextWithMnemonic(object, str)
gtkLabelGetAttributes(object)
gtkLabelGetJustify(object)
gtkLabelGetEllipsize(object)
gtkLabelGetWidthChars(object)
gtkLabelGetMaxWidthChars(object)
gtkLabelGetLabel(object)
gtkLabelGetLayout(object)
gtkLabelGetLineWrap(object)
gtkLabelGetLineWrapMode(object)
gtkLabelGetMnemonicWidget(object)
gtkLabelGetSelectionBounds(object)
gtkLabelGetUseMarkup(object)
gtkLabelGetUseUnderline(object)
gtkLabelGetSingleLineMode(object)
gtkLabelGetAngle(object)
gtkLabelSetLabel(object, str)
gtkLabelSetUseMarkup(object, setting)
gtkLabelSetUseUnderline(object, setting)
gtkLabelSetSingleLineMode(object, single.line.mode)
gtkLabelSetAngle(object, angle)
gtkLabelGetCurrentUri(object)
gtkLabelSetTrackVisitedLinks(object, track.links)
gtkLabelGetTrackVisitedLinks(object)
gtkLabel(str = NULL, show = TRUE)

Hierarchy

GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkMisc
                           +----GtkLabel
                                 +----GtkAccelLabel
                                 +----GtkTipsQuery

Interfaces

GtkLabel implements AtkImplementorIface and GtkBuildable.

Detailed Description

The GtkLabel widget displays a small amount of text. As the name implies, most labels are used to label another widget such as a GtkButton, a GtkMenuItem, or a GtkOptionMenu.

GtkLabel as GtkBuildable

The GtkLabel implementation of the GtkBuildable interface supports a custom <attributes> element, which supports any number of <attribute> elements. the <attribute> element has attributes named name, value, start and end and allows you to specify PangoAttribute values for this label.

A UI definition fragment specifying Pango attributes

<object class="GtkLabel">
  <attributes>
    <attribute name="weight" value="PANGO_WEIGHT_BOLD"/>
    <attribute name="background" value="red" start="5" end="10"/>"
  </attributes>
</object>

The start and end attributes specify the range of characters to which the Pango attribute applies. If start and end are not specified, the attribute is applied to the whole text. Note that specifying ranges does not make much sense with translatable attributes. Use markup embedded in the translatable content instead.

Mnemonics

Labels may contain mnemonics. Mnemonics are underlined characters in the label, used for keyboard navigation. Mnemonics are created by providing a string with an underscore before the mnemonic character, such as "_File", to the functions gtkLabelNewWithMnemonic or gtkLabelSetTextWithMnemonic.

Mnemonics automatically activate any activatable widget the label is inside, such as a GtkButton; if the label is not inside the mnemonic's target widget, you have to tell the label about the target using gtkLabelSetMnemonicWidget. Here's a simple example where the label is inside a button:

## Pressing Alt-H will activate this button
button <- gtkButton()
label <- gtkLabelNewWithMnemonic("_Hello")
button$add(label)

There's a convenience function to create buttons with a mnemonic label already inside:

## Pressing Alt+H will activate this button
button <- gtkButtonNewWithMnemonic("_Hello")

To create a mnemonic for a widget alongside the label, such as a GtkEntry, you have to point the label at the entry with gtkLabelSetMnemonicWidget:

## Pressing Alt+H will focus the entry
entry <- gtkEntry()
label <- gtkLabelNewWithMnemonic("_Hello")
label$setMnemonicWidget(entry)

Markup (styled text)

To make it easy to format text in a label (changing colors, fonts, etc.), label text can be provided in a simple markup format. Here's how to create a label with a small font:

label <- gtkLabelNew()
label$setMarkup("<small>Small text</small>")

(See complete documentation of available tags in the Pango manual.)

The markup passed to gtkLabelSetMarkup must be valid; for example, literal </>/& characters must be escaped as &lt;, &gt;, and &amp;. If you pass text obtained from the user, file, or a network to gtkLabelSetMarkup, you'll want to escape it with gMarkupEscapeText() or gMarkupPrintfEscaped().

Markup strings are just a convenient way to set the PangoAttrList on a label; gtkLabelSetAttributes may be a simpler way to set attributes in some cases. Be careful though; PangoAttrList tends to cause internationalization problems, unless you're applying attributes to the entire string (i.e. unless you set the range of each attribute to [0, G_MAXINT)). The reason is that specifying the start_index and end_index for a PangoAttribute requires knowledge of the exact string being displayed, so translations will cause problems.

Selectable labels

Labels can be made selectable with gtkLabelSetSelectable. Selectable labels allow the user to copy the label contents to the clipboard. Only labels that contain useful-to-copy information – such as error messages – should be made selectable.

Text layout

A label can contain any number of paragraphs, but will have performance problems if it contains more than a small number. Paragraphs are separated by newlines or other paragraph separators understood by Pango.

Labels can automatically wrap text if you call gtkLabelSetLineWrap. gtkLabelSetJustify sets how the lines in a label align with one another. If you want to set how the label as a whole aligns in its available space, see gtkMiscSetAlignment.

Links

Since 2.18, GTK+ supports markup for clickable hyperlinks in addition to regular Pango markup. The markup for links is borrowed from HTML, using the a with href and title attributes. GTK+ renders links similar to the way they appear in web browsers, with colored, underlined text. The title attribute is displayed as a tooltip on the link. An example looks like this:

label$setMarkup("Go to the &lt;a href=\"http://www.gtk.org\" title=\"&amp;lt;i&amp;gt;Our&amp;/i&amp;gt; website\"&gt;GTK+ website&lt;/a&gt; for more...")

It is possible to implement custom handling for links and their tooltips with the "activate-link" signal and the gtkLabelGetCurrentUri function.

Structures

GtkLabel

This should not be accessed directly. Use the accessor functions as described below.

Convenient Construction

gtkLabel is the result of collapsing the constructors of GtkLabel (gtkLabelNew, gtkLabelNewWithMnemonic) and accepts a subset of its arguments matching the required arguments of one of its delegate constructors.

Signals

activate-current-link(label, user.data)

A keybinding signal which gets emitted when the user activates a link in the label.

Applications may also emit the signal with gSignalEmitByName() if they need to control activation of URIs programmatically.

The default bindings for this signal are all forms of the Enter key. Since 2.18

label

The label on which the signal was emitted

user.data

user data set when the signal handler was connected.

activate-link(label, uri, user.data)

The signal which gets emitted to activate a URI. Applications may connect to it to override the default behaviour, which is to call gtkShowUri. Since 2.18

label

The label on which the signal was emitted

uri

the URI that is activated

user.data

user data set when the signal handler was connected.

Returns: [logical] TRUE if the link has been activated

copy-clipboard(label, user.data)

The ::copy-clipboard signal is a keybinding signal which gets emitted to copy the selection to the clipboard.

The default binding for this signal is Ctrl-c.

label

the object which received the signal

user.data

user data set when the signal handler was connected.

move-cursor(entry, step, count, extend.selection, user.data)

The ::move-cursor signal is a keybinding signal which gets emitted when the user initiates a cursor movement. If the cursor is not visible in entry, this signal causes the viewport to be moved instead.

Applications should not connect to it, but may emit it with gSignalEmitByName() if they need to control the cursor programmatically.

The default bindings for this signal come in two variants, the variant with the Shift modifier extends the selection, the variant without the Shift modifer does not. There are too many key combinations to list them all here.

entry

the object which received the signal

step

the granularity of the move, as a GtkMovementStep

count

the number of step units to move

extend.selection

TRUE if the move should extend the selection

user.data

user data set when the signal handler was connected.

populate-popup(label, menu, user.data)

The ::populate-popup signal gets emitted before showing the context menu of the label. Note that only selectable labels have context menus.

If you need to add items to the context menu, connect to this signal and append your menuitems to the menu.

label

The label on which the signal is emitted

menu

the menu that is being populated

user.data

user data set when the signal handler was connected.

Properties

angle [numeric : Read / Write]

The angle that the baseline of the label makes with the horizontal, in degrees, measured counterclockwise. An angle of 90 reads from from bottom to top, an angle of 270, from top to bottom. Ignored if the label is selectable, wrapped, or ellipsized. Allowed values: [0,360] Default value: 0 Since 2.6

attributes [PangoAttrList : * : Read / Write]

A list of style attributes to apply to the text of the label.

cursor-position [integer : Read]

The current position of the insertion cursor in chars. Allowed values: >= 0 Default value: 0

ellipsize [PangoEllipsizeMode : Read / Write]

The preferred place to ellipsize the string, if the label does not have enough room to display the entire string, specified as a PangoEllisizeMode.

Note that setting this property to a value other than PANGO_ELLIPSIZE_NONE has the side-effect that the label requests only enough space to display the ellipsis "...". In particular, this means that ellipsizing labels do not work well in notebook tabs, unless the tab's "tab-expand" property is set to TRUE. Other ways to set a label's width are gtkWidgetSetSizeRequest and gtkLabelSetWidthChars. Default value: PANGO_ELLIPSIZE_NONE Since 2.6

justify [GtkJustification : Read / Write]

The alignment of the lines in the text of the label relative to each other. This does NOT affect the alignment of the label within its allocation. See GtkMisc::xalign for that. Default value: GTK_JUSTIFY_LEFT

label [character : * : Read / Write]

The text of the label. Default value: ""

max-width-chars [integer : Read / Write]

The desired maximum width of the label, in characters. If this property is set to -1, the width will be calculated automatically, otherwise the label will request space for no more than the requested number of characters. If the "width-chars" property is set to a positive value, then the "max-width-chars" property is ignored. Allowed values: >= -1 Default value: -1 Since 2.6

mnemonic-keyval [numeric : Read]

The mnemonic accelerator key for this label. Default value: 16777215

mnemonic-widget [GtkWidget : * : Read / Write]

The widget to be activated when the label's mnemonic key is pressed.

pattern [character : * : Write]

A string with _ characters in positions correspond to characters in the text to underline. Default value: NULL

selectable [logical : Read / Write]

Whether the label text can be selected with the mouse. Default value: FALSE

selection-bound [integer : Read]

The position of the opposite end of the selection from the cursor in chars. Allowed values: >= 0 Default value: 0

single-line-mode [logical : Read / Write]

Whether the label is in single line mode. In single line mode, the height of the label does not depend on the actual text, it is always set to ascent + descent of the font. This can be an advantage in situations where resizing the label because of text changes would be distracting, e.g. in a statusbar. Default value: FALSE Since 2.6

track-visited-links [logical : Read / Write]

Set this property to TRUE to make the label track which links have been clicked. It will then apply the ::visited-link-color color, instead of ::link-color. Default value: TRUE Since 2.18

use-markup [logical : Read / Write]

The text of the label includes XML markup. See pango_parse_markup(). Default value: FALSE

use-underline [logical : Read / Write]

If set, an underline in the text indicates the next character should be used for the mnemonic accelerator key. Default value: FALSE

width-chars [integer : Read / Write]

The desired width of the label, in characters. If this property is set to -1, the width will be calculated automatically, otherwise the label will request either 3 characters or the property value, whichever is greater. If the "width-chars" property is set to a positive value, then the "max-width-chars" property is ignored. Allowed values: >= -1 Default value: -1 Since 2.6

wrap [logical : Read / Write]

If set, wrap lines if the text becomes too wide. Default value: FALSE

wrap-mode [PangoWrapMode : Read / Write]

If line wrapping is on (see the "wrap" property) this controls how the line wrapping is done. The default is PANGO_WRAP_WORD, which means wrap on word boundaries. Default value: PANGO_WRAP_WORD Since 2.10

Author(s)

Derived by RGtkGen from GTK+ documentation

References

http://library.gnome.org/devel//gtk/GtkLabel.html

[Package RGtk2 version 2.20.31 Index]