cairoFillExtents {RGtk2}	R Documentation

cairoFillExtents

Description

Computes a bounding box in user coordinates covering the area that would be affected, (the "inked" area), by a cairoFill operation given the current path and fill parameters. If the current path is empty, returns an empty rectangle ((0,0), (0,0)). Surface dimensions and clipping are not taken into account.

Usage

cairoFillExtents(cr)

Arguments

cr
[Cairo] a cairo context

Details

Contrast with cairoPathExtents, which is similar, but returns non-zero extents for some paths with no inked area, (such as a simple line segment).

Note that cairoFillExtents must necessarily do more work to compute the precise inked areas in light of the fill rule, so cairoPathExtents may be more desirable for sake of performance if the non-inked path extents are desired.

See cairoFill, cairoSetFillRule and cairoFillPreserve.

Value

A list containing the following elements:

x1
[numeric] left of the resulting extents

y1
[numeric] top of the resulting extents

x2
[numeric] right of the resulting extents

y2
[numeric] bottom of the resulting extents

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]