cairoGetDash {RGtk2}	R Documentation

cairoGetDash

Description

Gets the current dash list. If not NULL, dashes should be big enough to hold at least the number of values returned by cairoGetDashCount.

Usage

cairoGetDash(cr)

Arguments

cr
[Cairo] a Cairo

Details

Since 1.4

Value

A list containing the following elements:

dashes
[numeric] return value for the dash list, or NULL

offset
[numeric] return value for the current dash offset, or NULL

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]