cairoGetGroupTarget {RGtk2}	R Documentation

cairoGetGroupTarget

Description

Gets the current destination surface for the context. This is either the original target surface as passed to cairoCreate or the target surface for the current group as started by the most recent call to cairoPushGroup or cairoPushGroupWithContent.

Usage

cairoGetGroupTarget(cr)

Arguments

cr
[Cairo] a cairo context

Details

This function will always return a valid pointer, but the result can be a "nil" surface if cr is already in an error state, (ie. cairoStatus != CAIRO_STATUS_SUCCESS). A nil surface is indicated by cairoSurfaceStatus != CAIRO_STATUS_SUCCESS. Since 1.2

Value

[CairoSurface] the target surface. To keep a reference to it,

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]