cairoImageSurfaceGetData {RGtk2}	R Documentation

cairoImageSurfaceGetData

Description

Get a pointer to the data of the image surface, for direct inspection or modification.

Usage

cairoImageSurfaceGetData(surface)

Arguments

surface
[CairoSurface] a cairo_image_surface_t

Details

Since 1.2

Value

[char] a pointer to the image data of this surface or NULL if surface is not an image surface, or if cairoSurfaceFinish has been called.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]