cairoPatternCreateLinear {RGtk2}	R Documentation

cairoPatternCreateLinear

Description

Create a new linear gradient CairoPattern along the line defined by (x0, y0) and (x1, y1). Before using the gradient pattern, a number of color stops should be defined using cairoPatternAddColorStopRgb or cairoPatternAddColorStopRgba.

Usage

cairoPatternCreateLinear(x0, y0, x1, y1)

Arguments

x0
[numeric] x coordinate of the start point

y0
[numeric] y coordinate of the start point

x1
[numeric] x coordinate of the end point

y1
[numeric] y coordinate of the end point

Details

Note: The coordinates here are in pattern space. For a new pattern, pattern space is identical to user space, but the relationship between the spaces can be changed with cairoPatternSetMatrix.

Value

[CairoPattern] the newly created CairoPattern if successful, or an error pattern in case of no memory. This function will always return a valid pointer, but if an error occurred the pattern status will be set to an error. To inspect the status of a pattern use cairoPatternStatus.

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]