cairoPopGroupToSource {RGtk2}	R Documentation

cairoPopGroupToSource

Description

Terminates the redirection begun by a call to cairoPushGroup or cairoPushGroupWithContent and installs the resulting pattern as the source pattern in the given cairo context.

Usage

cairoPopGroupToSource(cr)

Arguments

cr
[Cairo] a cairo context

Details

The behavior of this function is equivalent to the sequence of operations:

group <- cr$popGroup()
cr$setSource(group)

but is more convenient as their is no need for a variable to store the short-lived pointer to the pattern.

The cairoPopGroup function calls cairoRestore, (balancing a call to cairoSave by the push_group function), so that any changes to the graphics state will not be visible outside the group. Since 1.2

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]