cairoSurfaceGetDeviceOffset {RGtk2}	R Documentation

cairoSurfaceGetDeviceOffset

Description

This function returns the previous device offset set by cairoSurfaceSetDeviceOffset.

Usage

cairoSurfaceGetDeviceOffset(surface)

Arguments

surface
[CairoSurface] a CairoSurface

Details

Since 1.2

Value

A list containing the following elements:

x.offset
[numeric] the offset in the X direction, in device units

y.offset
[numeric] the offset in the Y direction, in device units

Author(s)

Derived by RGtkGen from GTK+ documentation

[Package RGtk2 version 2.20.31 Index]