custom-tree-models {RGtk2}	R Documentation

Custom GtkTreeModel implementations

Description

Functions that allow one to implement a custom GtkTreeModel

Usage

  gtkTreeIter(id, stamp)
  gtkTreeIterGetId(iter)
  gtkTreeIterSetId(iter, id)
  gtkTreeIterGetStamp(iter)
  gtkTreeIterSetStamp(iter, stamp)

Arguments

iter
The GtkTreeIter of a custom model

id
The integer code identifying iter

stamp
The integer code for tracking the version of iter

Details

These functions allow one to create and access GtkTreeIter structures when implementing a GtkTreeModel.

gtkTreeIter creates an iter from scratch, given an id and stamp.

gtkTreeIterGetId and gtkTreeIterSetId access the integer that identifies the data element referred to by iter.

gtkTreeIterGetStamp and gtkTreeIterSetStamp access the integer that serves as a version stamp. After the model changes, the model version should be incremented, so that all existing iters are invalidated, as evidenced by their stamp.

Value

For gtkTreeIter, an external pointer to the underlying C structure.

For gtkTreeIterGetId, the integer code identifying the element referred to by iter.

For gtkTreeIterGetStamp, the integer code identifying the version of iter.

Note

These functions are for implementing GtkTreeModels only! Most of the time, one can use one of the implementations included with GTK+ (GtkListStore or GtkTreeStore) or RGtkDataFrame.

Author(s)

Michael Lawrence

See Also

gClass

[Package RGtk2 version 2.20.31 Index]